Module qt_core::abstract_event_dispatcher::overloading

Traits

AbstractEventDispatcherConnectArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::connect method.

AbstractEventDispatcherConnectStaticArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::connect_static method.

AbstractEventDispatcherDestroyedArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::destroyed method.

AbstractEventDispatcherDisconnectArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::disconnect method.

AbstractEventDispatcherDisconnectStaticArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::disconnect_static method.

AbstractEventDispatcherDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::dynamic_property_names method.

AbstractEventDispatcherInstanceArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::instance method.

AbstractEventDispatcherObjectNameArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::object_name method.

AbstractEventDispatcherPropertyArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::property method.

AbstractEventDispatcherRegisterTimerArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::register_timer method.

AbstractEventDispatcherRegisteredTimersArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::registered_timers method.

AbstractEventDispatcherStartTimerArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::start_timer method.

AbstractEventDispatcherTrArgs
This trait represents a set of arguments accepted by AbstractEventDispatcher::tr method.

AbstractEventDispatcherTrUtf8Args
This trait represents a set of arguments accepted by AbstractEventDispatcher::tr_utf8 method.

TimerInfoNewArgs
This trait represents a set of arguments accepted by TimerInfo::new method.