Struct qt_core::abstract_native_event_filter::AbstractNativeEventFilter

pub struct AbstractNativeEventFilter {
    // some fields omitted
}

C++ type: QAbstractNativeEventFilter

C++ documentation:

The QAbstractNativeEventFilter class provides an interface for receiving native events, such as MSG or XCB event structs.

Methods

impl AbstractNativeEventFilter

fn native_event_filter(&mut self, event_type: &ByteArray, message: *mut c_void, result: *mut c_long) -> bool

C++ method: pure virtual bool QAbstractNativeEventFilter::nativeEventFilter(const QByteArray& eventType, void* message, long* result)

C++ documentation:

This method is called for every native event.

Note: The filter function here receives native messages, for example, MSG or XCB event structs.

It is called by the QPA platform plugin. On Windows, it is called by the event dispatcher.

The type of event eventType is specific to the platform plugin chosen at run-time, and can be used to cast message to the right type.

On X11, eventType is set to "xcb_generic_event_t", and the message can be casted to a xcb_generic_event_t pointer.

On Windows, eventType is set to "windows_generic_MSG" for messages sent to toplevel windows, and "windows_dispatcher_MSG" for system-wide messages such as messages from a registered hot key. In both cases, the message can be casted to a MSG pointer. The result pointer is only used on Windows, and corresponds to the LRESULT pointer.

On Mac, eventType is set to "mac_generic_NSEvent", and the message can be casted to an EventRef.

On Blackberry (not plain QNX) eventType is set to "bps_event_t", and the message can be casted to a bps_event_t pointer.

In your reimplementation of this function, if you want to filter the message out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MyXcbEventFilter : public QAbstractNativeEventFilter
{
public:
    virtual bool nativeEventFilter(const QByteArray &eventType, void message, long ) Q_DECL_OVERRIDE
    {
        if (eventType == "xcb_generic_event_t") {
            xcb_generic_event_t ev = static_cast<xcb_generic_event_t >(message);
            // ...
        }
        return false;
    }
};