Module qt_core::abstract_state::overloading

Traits

AbstractStateConnectArgs
This trait represents a set of arguments accepted by AbstractState::connect method.

AbstractStateConnectStaticArgs
This trait represents a set of arguments accepted by AbstractState::connect_static method.

AbstractStateDestroyedArgs
This trait represents a set of arguments accepted by AbstractState::destroyed method.

AbstractStateDisconnectArgs
This trait represents a set of arguments accepted by AbstractState::disconnect method.

AbstractStateDisconnectStaticArgs
This trait represents a set of arguments accepted by AbstractState::disconnect_static method.

AbstractStateDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by AbstractState::dynamic_property_names method.

AbstractStateObjectNameArgs
This trait represents a set of arguments accepted by AbstractState::object_name method.

AbstractStatePropertyArgs
This trait represents a set of arguments accepted by AbstractState::property method.

AbstractStateStartTimerArgs
This trait represents a set of arguments accepted by AbstractState::start_timer method.

AbstractStateTrArgs
This trait represents a set of arguments accepted by AbstractState::tr method.

AbstractStateTrUtf8Args
This trait represents a set of arguments accepted by AbstractState::tr_utf8 method.