Module qt_core::abstract_transition::overloading

Traits

AbstractTransitionAnimationsArgs
This trait represents a set of arguments accepted by AbstractTransition::animations method.

AbstractTransitionConnectArgs
This trait represents a set of arguments accepted by AbstractTransition::connect method.

AbstractTransitionConnectStaticArgs
This trait represents a set of arguments accepted by AbstractTransition::connect_static method.

AbstractTransitionDestroyedArgs
This trait represents a set of arguments accepted by AbstractTransition::destroyed method.

AbstractTransitionDisconnectArgs
This trait represents a set of arguments accepted by AbstractTransition::disconnect method.

AbstractTransitionDisconnectStaticArgs
This trait represents a set of arguments accepted by AbstractTransition::disconnect_static method.

AbstractTransitionDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by AbstractTransition::dynamic_property_names method.

AbstractTransitionObjectNameArgs
This trait represents a set of arguments accepted by AbstractTransition::object_name method.

AbstractTransitionPropertyArgs
This trait represents a set of arguments accepted by AbstractTransition::property method.

AbstractTransitionStartTimerArgs
This trait represents a set of arguments accepted by AbstractTransition::start_timer method.

AbstractTransitionTargetStatesArgs
This trait represents a set of arguments accepted by AbstractTransition::target_states method.

AbstractTransitionTrArgs
This trait represents a set of arguments accepted by AbstractTransition::tr method.

AbstractTransitionTrUtf8Args
This trait represents a set of arguments accepted by AbstractTransition::tr_utf8 method.