Module qt_core::animation_driver::overloading

Traits

AnimationDriverConnectArgs
This trait represents a set of arguments accepted by AnimationDriver::connect method.

AnimationDriverConnectStaticArgs
This trait represents a set of arguments accepted by AnimationDriver::connect_static method.

AnimationDriverDestroyedArgs
This trait represents a set of arguments accepted by AnimationDriver::destroyed method.

AnimationDriverDisconnectArgs
This trait represents a set of arguments accepted by AnimationDriver::disconnect method.

AnimationDriverDisconnectStaticArgs
This trait represents a set of arguments accepted by AnimationDriver::disconnect_static method.

AnimationDriverDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by AnimationDriver::dynamic_property_names method.

AnimationDriverNewArgs
This trait represents a set of arguments accepted by AnimationDriver::new method.

AnimationDriverObjectNameArgs
This trait represents a set of arguments accepted by AnimationDriver::object_name method.

AnimationDriverPropertyArgs
This trait represents a set of arguments accepted by AnimationDriver::property method.

AnimationDriverStartTimerArgs
This trait represents a set of arguments accepted by AnimationDriver::start_timer method.

AnimationDriverTrArgs
This trait represents a set of arguments accepted by AnimationDriver::tr method.

AnimationDriverTrUtf8Args
This trait represents a set of arguments accepted by AnimationDriver::tr_utf8 method.