Struct qt_core::associative_iterable::AssociativeIterable

pub struct AssociativeIterable {
    // some fields omitted
}

C++ type: QAssociativeIterable

C++ documentation:

The QAssociativeIterable class is an iterable interface for an associative container in a QVariant.

This class allows several methods of accessing the elements of an associative container held within a QVariant. An instance of QAssociativeIterable can be extracted from a QVariant if it can be converted to a QVariantHash or QVariantMap.

QHash<int, QString> mapping;
mapping.insert(7, "Seven");
mapping.insert(11, "Eleven");
mapping.insert(42, "Forty-two");

QVariant variant = QVariant::fromValue(mapping); if (variant.canConvert<QVariantHash>()) { QAssociativeIterable iterable = variant.value<QAssociativeIterable>(); // Can use foreach over the values: foreach (const QVariant &v, iterable) { qDebug() << v; } // Can use C++11 range-for over the values: for (const QVariant &v : iterable) { qDebug() << v; } // Can use iterators: QAssociativeIterable::const_iterator it = iterable.begin(); const QAssociativeIterable::const_iterator end = iterable.end(); for ( ; it != end; ++it) { qDebug() << *it; // The current value qDebug() << it.key(); qDebug() << it.value(); } }
The container itself is not copied before iterating over it.

Methods

impl AssociativeIterable

fn begin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AssociativeIterableBeginArgs<'largs>

C++ method: QAssociativeIterable::begin

This is an overloaded function. Available variants:

Rust arguments:
1) fn begin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::associative_iterable::ConstIterator>
2) fn begin(&self, cpp_utils::AsStruct) -> ::associative_iterable::ConstIterator
C++ method: QAssociativeIterable::const_iterator QAssociativeIterable::begin() const

C++ documentation:

Returns a QAssociativeIterable::const_iterator for the beginning of the container. This can be used in stl-style iteration.

See also end().

fn end<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AssociativeIterableEndArgs<'largs>

C++ method: QAssociativeIterable::end

This is an overloaded function. Available variants:

Rust arguments:
1) fn end(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::associative_iterable::ConstIterator>
2) fn end(&self, cpp_utils::AsStruct) -> ::associative_iterable::ConstIterator
C++ method: QAssociativeIterable::const_iterator QAssociativeIterable::end() const

C++ documentation:

Returns a QAssociativeIterable::const_iterator for the end of the container. This can be used in stl-style iteration.

See also begin().

fn find<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AssociativeIterableFindArgs<'largs>

C++ method: QAssociativeIterable::find

This is an overloaded function. Available variants:

Rust arguments:
1) fn find(&self, (&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::associative_iterable::ConstIterator>
2) fn find(&self, (&::variant::Variant, cpp_utils::AsStruct)) -> ::associative_iterable::ConstIterator
C++ method: QAssociativeIterable::const_iterator QAssociativeIterable::find(const QVariant& key) const

C++ documentation:

Returns a QAssociativeIterable::const_iterator for the given key key in the container, if the types are convertible.

If the key is not found, returns end().

This can be used in stl-style iteration.

This function was introduced in Qt 5.5.

See also begin(), end(), and value().

fn size(&self) -> c_int

C++ method: int QAssociativeIterable::size() const

C++ documentation:

Returns the number of elements in the container.

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: AssociativeIterableValueArgs<'largs>

C++ method: QAssociativeIterable::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, (&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn value(&self, (&::variant::Variant, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QAssociativeIterable::value(const QVariant& key) const

C++ documentation:

Returns the value for the given key in the container, if the types are convertible.

See also find().