Struct qt_core::basic_timer::BasicTimer

pub struct BasicTimer {
    // some fields omitted
}

C++ type: QBasicTimer

C++ documentation:

The QBasicTimer class provides timer events for objects.

This is a fast, lightweight, and low-level class used by Qt internally. We recommend using the higher-level QTimer class rather than this class if you want to use timers in your applications. Note that this timer is a repeating timer that will send subsequent timer events unless the stop() function is called.

To use this class, create a QBasicTimer, and call its start() function with a timeout interval and with a pointer to a QObject subclass. When the timer times out it will send a timer event to the QObject subclass. The timer can be stopped at any time using stop(). isActive() returns true for a timer that is running; i.e. it has been started, has not reached the timeout time, and has not been stopped. The timer's ID can be retrieved using timerId().

The Wiggly example uses QBasicTimer to repaint a widget at regular intervals.

Methods

impl BasicTimer

fn is_active(&self) -> bool

C++ method: bool QBasicTimer::isActive() const

C++ documentation:

Returns true if the timer is running and has not been stopped; otherwise returns false.

See also start() and stop().

fn new<Args>(args: Args) -> Args::ReturnType where Args: BasicTimerNewArgs

C++ method: QBasicTimer::QBasicTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::basic_timer::BasicTimer
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::basic_timer::BasicTimer>
C++ method: [constructor] void QBasicTimer::QBasicTimer()

fn start<'largs, Args>(&'largs mut self, args: Args) where Args: BasicTimerStartArgs<'largs>

C++ method: QBasicTimer::start

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn start(&mut self, (libc::c_int, *mut ::object::Object)) -> ()
C++ method: void QBasicTimer::start(int msec, QObject* obj)

C++ documentation:

Starts (or restarts) the timer with a msec milliseconds timeout. The timer will be a Qt::CoarseTimer. See Qt::TimerType for information on the different timer types.

The given object will receive timer events.

See also stop(), isActive(), QObject::timerEvent(), and Qt::CoarseTimer.

Variant 2

Rust arguments: fn start(&mut self, (libc::c_int, ::qt::TimerType, *mut ::object::Object)) -> ()
C++ method: void QBasicTimer::start(int msec, Qt::TimerType timerType, QObject* obj)

C++ documentation:

This is an overloaded function.

Starts (or restarts) the timer with a msec milliseconds timeout and the given timerType. See Qt::TimerType for information on the different timer types.

obj will receive timer events.

See also stop(), isActive(), QObject::timerEvent(), and Qt::TimerType.

fn stop(&mut self)

C++ method: void QBasicTimer::stop()

C++ documentation:

Stops the timer.

See also start() and isActive().

fn timer_id(&self) -> c_int

C++ method: int QBasicTimer::timerId() const

C++ documentation:

Returns the timer's ID.

See also QTimerEvent::timerId().