Struct qt_core::bit_array::BitArray

pub struct BitArray {
    // some fields omitted
}

C++ type: QBitArray

C++ documentation:

The QBitArray class provides an array of bits.

A QBitArray is an array that gives access to individual bits and provides operators (AND, OR, XOR, and NOT) that work on entire arrays of bits. It uses implicit sharing (copy-on-write) to reduce memory usage and to avoid the needless copying of data.

The following code constructs a QBitArray containing 200 bits initialized to false (0):

QBitArray ba(200);

To initialize the bits to true, either pass true as second argument to the constructor, or call fill() later on.

QBitArray uses 0-based indexes, just like C++ arrays. To access the bit at a particular index position, you can use operator. On non-const bit arrays, operator returns a reference to a bit that can be used on the left side of an assignment. For example:

QBitArray ba;
ba.resize(3);
ba[0] = true;
ba[1] = false;
ba[2] = true;

For technical reasons, it is more efficient to use testBit() and setBit() to access bits in the array than operator. For example:

QBitArray ba(3);
ba.setBit(0, true);
ba.setBit(1, false);
ba.setBit(2, true);

QBitArray supports & (AND), | (OR), (XOR), ~ (NOT), as well as &=, |=, and =. These operators work in the same way as the built-in C++ bitwise operators of the same name. For example:

QBitArray x(5);
x.setBit(3, true);
// x: [ 0, 0, 0, 1, 0 ]

QBitArray y(5); y.setBit(4, true); // y: [ 0, 0, 0, 0, 1 ]

x |= y; // x: [ 0, 0, 0, 1, 1 ]
For historical reasons, QBitArray distinguishes between a null bit array and an empty bit array. A null bit array is a bit array that is initialized using QBitArray's default constructor. An empty bit array is any bit array with size 0. A null bit array is always empty, but an empty bit array isn't necessarily null:

QBitArray().isNull();           // returns true
QBitArray().isEmpty();          // returns true

QBitArray(0).isNull(); // returns false QBitArray(0).isEmpty(); // returns true

QBitArray(3).isNull(); // returns false QBitArray(3).isEmpty(); // returns false
All functions except isNull() treat null bit arrays the same as empty bit arrays; for example, QBitArray() compares equal to QBitArray(0). We recommend that you always use isEmpty() and avoid isNull().

Methods

impl BitArray

fn at(&self, i: c_int) -> bool

C++ method: bool QBitArray::at(int i) const

C++ documentation:

Returns the value of the bit at index position i.

i must be a valid index position in the bit array (i.e., 0 <= i < size()).

See also operator[]().

fn clear(&mut self)

C++ method: void QBitArray::clear()

C++ documentation:

Clears the contents of the bit array and makes it empty.

See also resize() and isEmpty().

fn clear_bit(&mut self, i: c_int)

C++ method: void QBitArray::clearBit(int i)

C++ documentation:

Sets the bit at index position i to 0.

i must be a valid index position in the bit array (i.e., 0 <= i < size()).

See also setBit() and toggleBit().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: BitArrayCountArgs<'largs>

C++ method: QBitArray::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QBitArray::count() const

C++ documentation:

Same as size().

Variant 2

Rust arguments: fn count(&self, bool) -> libc::c_int
C++ method: int QBitArray::count(bool on) const

C++ documentation:

If on is true, this function returns the number of 1-bits stored in the bit array; otherwise the number of 0-bits is returned.

fn data_ptr<'l0>(&'l0 mut self) -> &'l0 mut *mut TypedArrayDataCChar

C++ method: QTypedArrayData<char>*& QBitArray::data_ptr()

fn detach(&mut self)

C++ method: void QBitArray::detach()

fn fill<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: BitArrayFillArgs<'largs>

C++ method: QBitArray::fill

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn fill(&mut self, bool) -> bool
2) fn fill(&mut self, (bool, libc::c_int)) -> bool
C++ method: bool QBitArray::fill(bool val, int size = ?)

C++ documentation:

Sets every bit in the bit array to value, returning true if successful; otherwise returns false. If size is different from -1 (the default), the bit array is resized to size beforehand.

Example:

QBitArray ba(8);
ba.fill(true);
// ba: [ 1, 1, 1, 1, 1, 1, 1, 1 ]

ba.fill(false, 2); // ba: [ 0, 0 ]
See also resize().

Variant 2

Rust arguments: fn fill(&mut self, (bool, libc::c_int, libc::c_int)) -> ()
C++ method: void QBitArray::fill(bool val, int first, int last)

C++ documentation:

This is an overloaded function.

Sets bits at index positions begin up to (but not including) end to value.

begin must be a valid index position in the bit array (0 <= begin < size()).

end must be either a valid index position or equal to size(), in which case the fill operation runs until the end of the array (0 <= end <= size()).

Example:

QBitArray ba(4);
ba.fill(true, 1, 2);            // ba: [ 0, 1, 0, 0 ]
ba.fill(true, 1, 3);            // ba: [ 0, 1, 1, 0 ]
ba.fill(true, 1, 4);            // ba: [ 0, 1, 1, 1 ]

fn is_detached(&self) -> bool

C++ method: bool QBitArray::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QBitArray::isEmpty() const

C++ documentation:

Returns true if this bit array has size 0; otherwise returns false.

See also size().

fn is_null(&self) -> bool

C++ method: bool QBitArray::isNull() const

C++ documentation:

Returns true if this bit array is null; otherwise returns false.

Example:

QBitArray().isNull();           // returns true
QBitArray(0).isNull();          // returns false
QBitArray(3).isNull();          // returns false

Qt makes a distinction between null bit arrays and empty bit arrays for historical reasons. For most applications, what matters is whether or not a bit array contains any data, and this can be determined using isEmpty().

See also isEmpty().

fn new<Args>(args: Args) -> Args::ReturnType where Args: BitArrayNewArgs

C++ method: QBitArray::QBitArray

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::bit_array::BitArray
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::bit_array::BitArray>
C++ method: [constructor] void QBitArray::QBitArray()

Variant 2

Rust arguments:
1) fn new((&::bit_array::BitArray, cpp_utils::AsStruct)) -> ::bit_array::BitArray
2) fn new((&::bit_array::BitArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::bit_array::BitArray>
C++ method: [constructor] void QBitArray::QBitArray(const QBitArray& other)

Variant 3

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::bit_array::BitArray
2) fn new((libc::c_int, bool, cpp_utils::AsStruct)) -> ::bit_array::BitArray
3) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::bit_array::BitArray>
4) fn new((libc::c_int, bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::bit_array::BitArray>
C++ method: [constructor] void QBitArray::QBitArray(int size, bool val = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 BitArray) -> &'l0 mut BitArray

C++ method: QBitArray& QBitArray::operator=(const QBitArray& other)

C++ documentation:

Assigns other to this bit array and returns a reference to this bit array.

fn op_bit_and_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 BitArray) -> &'l0 mut BitArray

C++ method: QBitArray& QBitArray::operator&=(const QBitArray& arg1)

C++ documentation:

Performs the AND operation between all bits in this bit array and other. Assigns the result to this bit array, and returns a reference to it.

The result has the length of the longest of the two bit arrays, with any missing bits (if one array is shorter than the other) taken to be 0.

Example:

QBitArray a(3);
QBitArray b(2);
a[0] = 1; a[1] = 0; a[2] = 1;   // a: [ 1, 0, 1 ]
b[0] = 1; b[1] = 1;             // b: [ 1, 1 ]
a &= b;                         // a: [ 1, 0, 0 ]

See also operator&(), operator|=(), operator=(), and operator~().

fn op_bit_not<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: BitArrayOpBitNotArgs<'largs>

C++ method: QBitArray::operator~

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_bit_not(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::bit_array::BitArray>
2) fn op_bit_not(&self, cpp_utils::AsStruct) -> ::bit_array::BitArray
C++ method: QBitArray QBitArray::operator~() const

C++ documentation:

Returns a bit array that contains the inverted bits of this bit array.

Example:

QBitArray a(3);
QBitArray b;
a[0] = 1; a[1] = 0; a[2] = 1;   // a: [ 1, 0, 1 ]
b = ~a;                         // b: [ 0, 1, 0 ]

See also operator&(), operator|(), and operator().

fn op_bit_or_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 BitArray) -> &'l0 mut BitArray

C++ method: QBitArray& QBitArray::operator|=(const QBitArray& arg1)

C++ documentation:

Performs the OR operation between all bits in this bit array and other. Assigns the result to this bit array, and returns a reference to it.

The result has the length of the longest of the two bit arrays, with any missing bits (if one array is shorter than the other) taken to be 0.

Example:

QBitArray a(3);
QBitArray b(2);
a[0] = 1; a[1] = 0; a[2] = 1;   // a: [ 1, 0, 1 ]
b[0] = 1; b[1] = 1;             // b: [ 1, 1 ]
a |= b;                         // a: [ 1, 1, 1 ]

See also operator|(), operator&=(), operator=(), and operator~().

fn op_bit_xor_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 BitArray) -> &'l0 mut BitArray

C++ method: QBitArray& QBitArray::operator^=(const QBitArray& arg1)

C++ documentation:

Performs the XOR operation between all bits in this bit array and other. Assigns the result to this bit array, and returns a reference to it.

The result has the length of the longest of the two bit arrays, with any missing bits (if one array is shorter than the other) taken to be 0.

Example:

QBitArray a(3);
QBitArray b(2);
a[0] = 1; a[1] = 0; a[2] = 1;   // a: [ 1, 0, 1 ]
b[0] = 1; b[1] = 1;             // b: [ 1, 1 ]
a <span class="operator">= b;                         // a: [ 0, 1, 1 ]

See also operator(), operator&=(), operator|=(), and operator~().

fn op_eq(&self, other: &BitArray) -> bool

C++ method: bool QBitArray::operator==(const QBitArray& other) const

C++ documentation:

Returns true if other is equal to this bit array; otherwise returns false.

See also operator!=().

fn op_index<'largs, Args>(&'largs self, args: Args) -> bool where Args: BitArrayOpIndexArgs<'largs>

C++ method: QBitArray::operator[]

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_index(&self, libc::c_int) -> bool
C++ method: bool QBitArray::operator[](int i) const

C++ documentation:

This is an overloaded function.

Variant 2

Rust arguments: fn op_index(&self, libc::c_uint) -> bool
C++ method: bool QBitArray::operator[](unsigned int i) const

C++ documentation:

This is an overloaded function.

fn op_index_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: BitArrayOpIndexMutArgs<'largs>

C++ method: QBitArray::operator[]

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn op_index_mut(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::bit_ref::BitRef>
2) fn op_index_mut(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::bit_ref::BitRef
C++ method: QBitRef QBitArray::operator[](int i)

C++ documentation:

Returns the bit at index position i as a modifiable reference.

i must be a valid index position in the bit array (i.e., 0 <= i < size()).

Example:

QBitArray a(3);
a[0] = false;
a[1] = true;
a[2] = a[0]  a[1];

The return value is of type QBitRef, a helper class for QBitArray. When you get an object of type QBitRef, you can assign to it, and the assignment will apply to the bit in the QBitArray from which you got the reference.

The functions testBit(), setBit(), and clearBit() are slightly faster.

See also at(), testBit(), setBit(), and clearBit().

Variant 2

Rust arguments:
1) fn op_index_mut(&mut self, (libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::bit_ref::BitRef>
2) fn op_index_mut(&mut self, (libc::c_uint, cpp_utils::AsStruct)) -> ::bit_ref::BitRef
C++ method: QBitRef QBitArray::operator[](unsigned int i)

C++ documentation:

This is an overloaded function.

fn op_neq(&self, other: &BitArray) -> bool

C++ method: bool QBitArray::operator!=(const QBitArray& other) const

C++ documentation:

Returns true if other is not equal to this bit array; otherwise returns false.

See also operator==().

fn resize(&mut self, size: c_int)

C++ method: void QBitArray::resize(int size)

C++ documentation:

Resizes the bit array to size bits.

If size is greater than the current size, the bit array is extended to make it size bits with the extra bits added to the end. The new bits are initialized to false (0).

If size is less than the current size, bits are removed from the end.

See also size().

fn set_bit<'largs, Args>(&'largs mut self, args: Args) where Args: BitArraySetBitArgs<'largs>

C++ method: QBitArray::setBit

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_bit(&mut self, libc::c_int) -> ()
C++ method: void QBitArray::setBit(int i)

C++ documentation:

Sets the bit at index position i to 1.

i must be a valid index position in the bit array (i.e., 0 <= i < size()).

See also clearBit() and toggleBit().

Variant 2

Rust arguments: fn set_bit(&mut self, (libc::c_int, bool)) -> ()
C++ method: void QBitArray::setBit(int i, bool val)

C++ documentation:

This is an overloaded function.

Sets the bit at index position i to value.

fn size(&self) -> c_int

C++ method: int QBitArray::size() const

C++ documentation:

Returns the number of bits stored in the bit array.

See also resize().

fn swap(&mut self, other: &mut BitArray)

C++ method: void QBitArray::swap(QBitArray& other)

C++ documentation:

Swaps bit array other with this bit array. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn test_bit(&self, i: c_int) -> bool

C++ method: bool QBitArray::testBit(int i) const

C++ documentation:

Returns true if the bit at index position i is 1; otherwise returns false.

i must be a valid index position in the bit array (i.e., 0 <= i < size()).

See also setBit() and clearBit().

fn toggle_bit(&mut self, i: c_int) -> bool

C++ method: bool QBitArray::toggleBit(int i)

C++ documentation:

Inverts the value of the bit at index position i, returning the previous value of that bit as either true (if it was set) or false (if it was unset).

If the previous value was 0, the new value will be 1. If the previous value was 1, the new value will be 0.

i must be a valid index position in the bit array (i.e., 0 <= i < size()).

See also setBit() and clearBit().

fn truncate(&mut self, pos: c_int)

C++ method: void QBitArray::truncate(int pos)

C++ documentation:

Truncates the bit array at index position pos.

If pos is beyond the end of the array, nothing happens.

See also resize().