Module qt_core::buffer::overloading

Traits

BufferConnectArgs
This trait represents a set of arguments accepted by Buffer::connect method.

BufferConnectStaticArgs
This trait represents a set of arguments accepted by Buffer::connect_static method.

BufferDestroyedArgs
This trait represents a set of arguments accepted by Buffer::destroyed method.

BufferDisconnectArgs
This trait represents a set of arguments accepted by Buffer::disconnect method.

BufferDisconnectStaticArgs
This trait represents a set of arguments accepted by Buffer::disconnect_static method.

BufferDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Buffer::dynamic_property_names method.

BufferErrorStringArgs
This trait represents a set of arguments accepted by Buffer::error_string method.

BufferNewArgs
This trait represents a set of arguments accepted by Buffer::new method.

BufferObjectNameArgs
This trait represents a set of arguments accepted by Buffer::object_name method.

BufferPeekArgs
This trait represents a set of arguments accepted by Buffer::peek method.

BufferPropertyArgs
This trait represents a set of arguments accepted by Buffer::property method.

BufferReadAllArgs
This trait represents a set of arguments accepted by Buffer::read_all method.

BufferReadArgs
This trait represents a set of arguments accepted by Buffer::read method.

BufferReadLineArgs
This trait represents a set of arguments accepted by Buffer::read_line method.

BufferSetDataArgs
This trait represents a set of arguments accepted by Buffer::set_data method.

BufferStartTimerArgs
This trait represents a set of arguments accepted by Buffer::start_timer method.

BufferTrArgs
This trait represents a set of arguments accepted by Buffer::tr method.

BufferTrUtf8Args
This trait represents a set of arguments accepted by Buffer::tr_utf8 method.

BufferWriteArgs
This trait represents a set of arguments accepted by Buffer::write method.