Function qt_core::byte_array::compress

pub fn compress<Args>(args: Args) -> Args::ReturnType where Args: CompressArgs

C++ method: qCompress

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn compress((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn compress((&::byte_array::ByteArray, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
3) fn compress((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
4) fn compress((&::byte_array::ByteArray, libc::c_int, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray qCompress(const QByteArray& data, int compressionLevel = ?)

C++ documentation:

Compresses the data byte array and returns the compressed data in a new byte array.

The compressionLevel parameter specifies how much compression should be used. Valid values are between 0 and 9, with 9 corresponding to the greatest compression (i.e. smaller compressed data) at the cost of using a slower algorithm. Smaller values (8, 7, ..., 1) provide successively less compression at slightly faster speeds. The value 0 corresponds to no compression at all. The default value is -1, which specifies zlib's default compression.

See also qUncompress().

Variant 2

Rust arguments:
1) fn compress((*const libc::c_uchar, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn compress((*const libc::c_uchar, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
3) fn compress((*const libc::c_uchar, libc::c_int, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
4) fn compress((*const libc::c_uchar, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray qCompress(const unsigned char* data, int nbytes, int compressionLevel = ?)

C++ documentation:

This is an overloaded function.

Compresses the first nbytes of data and returns the compressed data in a new byte array.