Function qt_core::byte_array::op_le

pub fn op_le<Args>(args: Args) -> bool where Args: OpLeArgs

C++ method: operator<=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_le((&::byte_array::ByteArray, &::byte_array::ByteArray)) -> bool
C++ method: bool operator<=(const QByteArray& a1, const QByteArray& a2)

Warning: no exact match found in C++ documentation.Below is the C++ documentation for bool operator<=(const QUuid & lhs, const QUuid & rhs):

Returns true if lhs has the same variant field as rhs and is lexicographically not after rhs. If rhs has a different variant field, the return value is determined by comparing the two variants.

This function was introduced in Qt 5.5.

See also variant().

Variant 2

Rust arguments: fn op_le((&::byte_array::ByteArray, *const libc::c_char)) -> bool
C++ method: bool operator<=(const QByteArray& a1, const char* a2)

Variant 3

Rust arguments: fn op_le((*const libc::c_char, &::byte_array::ByteArray)) -> bool
C++ method: bool operator<=(const char* a1, const QByteArray& a2)