Function qt_core::byte_array::qstrcpy

pub fn qstrcpy(dst: *mut c_char, src: *const c_char) -> *mut c_char

C++ method: char* qstrcpy(char* dst, const char* src)

C++ documentation:

Copies all the characters up to and including the '\0' from src into dst and returns a pointer to dst. If src is 0, it immediately returns 0.

This function assumes that dst is large enough to hold the contents of src.

See also qstrncpy().