Struct qt_core::byte_array_matcher::ByteArrayMatcher

pub struct ByteArrayMatcher {
    // some fields omitted
}

C++ type: QByteArrayMatcher

C++ documentation:

The QByteArrayMatcher class holds a sequence of bytes that can be quickly matched in a byte array.

This class is useful when you have a sequence of bytes that you want to repeatedly match against some byte arrays (perhaps in a loop), or when you want to search for the same sequence of bytes multiple times in the same byte array. Using a matcher object and indexIn() is faster than matching a plain QByteArray with QByteArray::indexOf() if repeated matching takes place. This class offers no benefit if you are doing one-off byte array matches.

Create the QByteArrayMatcher with the QByteArray you want to search for. Then call indexIn() on the QByteArray that you want to search.

Methods

impl ByteArrayMatcher

fn index_in<'largs, Args>(&'largs self, args: Args) -> c_int where Args: ByteArrayMatcherIndexInArgs<'largs>

C++ method: QByteArrayMatcher::indexIn

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn index_in(&self, &::byte_array::ByteArray) -> libc::c_int
2) fn index_in(&self, (&::byte_array::ByteArray, libc::c_int)) -> libc::c_int
C++ method: int QByteArrayMatcher::indexIn(const QByteArray& ba, int from = ?) const

C++ documentation:

Searches the byte array ba, from byte position from (default 0, i.e. from the first byte), for the byte array pattern() that was set in the constructor or in the most recent call to setPattern(). Returns the position where the pattern() matched in ba, or -1 if no match was found.

Variant 2

Rust arguments:
1) fn index_in(&self, (*const libc::c_char, libc::c_int)) -> libc::c_int
2) fn index_in(&self, (*const libc::c_char, libc::c_int, libc::c_int)) -> libc::c_int
C++ method: int QByteArrayMatcher::indexIn(const char* str, int len, int from = ?) const

C++ documentation:

Searches the char string str, which has length len, from byte position from (default 0, i.e. from the first byte), for the byte array pattern() that was set in the constructor or in the most recent call to setPattern(). Returns the position where the pattern() matched in str, or -1 if no match was found.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ByteArrayMatcherNewArgs

C++ method: QByteArrayMatcher::QByteArrayMatcher

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::byte_array_matcher::ByteArrayMatcher
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array_matcher::ByteArrayMatcher>
C++ method: [constructor] void QByteArrayMatcher::QByteArrayMatcher()

Variant 2

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::byte_array_matcher::ByteArrayMatcher
2) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array_matcher::ByteArrayMatcher>
C++ method: [constructor] void QByteArrayMatcher::QByteArrayMatcher(const QByteArray& pattern)

Variant 3

Rust arguments:
1) fn new((&::byte_array_matcher::ByteArrayMatcher, cpp_utils::AsStruct)) -> ::byte_array_matcher::ByteArrayMatcher
2) fn new((&::byte_array_matcher::ByteArrayMatcher, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array_matcher::ByteArrayMatcher>
C++ method: [constructor] void QByteArrayMatcher::QByteArrayMatcher(const QByteArrayMatcher& other)

Variant 4

Rust arguments:
1) fn new((*const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::byte_array_matcher::ByteArrayMatcher
2) fn new((*const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array_matcher::ByteArrayMatcher>
C++ method: [constructor] void QByteArrayMatcher::QByteArrayMatcher(const char* pattern, int length)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 ByteArrayMatcher) -> &'l0 mut ByteArrayMatcher

C++ method: QByteArrayMatcher& QByteArrayMatcher::operator=(const QByteArrayMatcher& other)

C++ documentation:

Assigns the other byte array matcher to this byte array matcher.

fn pattern<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ByteArrayMatcherPatternArgs<'largs>

C++ method: QByteArrayMatcher::pattern

This is an overloaded function. Available variants:

Rust arguments:
1) fn pattern(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn pattern(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QByteArrayMatcher::pattern() const

C++ documentation:

Returns the byte array pattern that this byte array matcher will search for.

See also setPattern().

fn set_pattern(&mut self, pattern: &ByteArray)

C++ method: void QByteArrayMatcher::setPattern(const QByteArray& pattern)

C++ documentation:

Sets the byte array that this byte array matcher will search for to pattern.

See also pattern() and indexIn().