Enum qt_core::char::Script

pub enum Script {
    Unknown,
    Inherited,
    Common,
    Latin,
    Greek,
    Cyrillic,
    Armenian,
    Hebrew,
    Arabic,
    Syriac,
    Thaana,
    Devanagari,
    Bengali,
    Gurmukhi,
    Gujarati,
    Oriya,
    Tamil,
    Telugu,
    Kannada,
    Malayalam,
    Sinhala,
    Thai,
    Lao,
    Tibetan,
    Myanmar,
    Georgian,
    Hangul,
    Ethiopic,
    Cherokee,
    CanadianAboriginal,
    Ogham,
    Runic,
    Khmer,
    Mongolian,
    Hiragana,
    Katakana,
    Bopomofo,
    Han,
    Yi,
    OldItalic,
    Gothic,
    Deseret,
    Tagalog,
    Hanunoo,
    Buhid,
    Tagbanwa,
    Coptic,
    Limbu,
    TaiLe,
    LinearB,
    Ugaritic,
    Shavian,
    Osmanya,
    Cypriot,
    Braille,
    Buginese,
    NewTaiLue,
    Glagolitic,
    Tifinagh,
    SylotiNagri,
    OldPersian,
    Kharoshthi,
    Balinese,
    Cuneiform,
    Phoenician,
    PhagsPa,
    Nko,
    Sundanese,
    Lepcha,
    OlChiki,
    Vai,
    Saurashtra,
    KayahLi,
    Rejang,
    Lycian,
    Carian,
    Lydian,
    Cham,
    TaiTham,
    TaiViet,
    Avestan,
    EgyptianHieroglyphs,
    Samaritan,
    Lisu,
    Bamum,
    Javanese,
    MeeteiMayek,
    ImperialAramaic,
    OldSouthArabian,
    InscriptionalParthian,
    InscriptionalPahlavi,
    OldTurkic,
    Kaithi,
    Batak,
    Brahmi,
    Mandaic,
    Chakma,
    MeroiticCursive,
    MeroiticHieroglyphs,
    Miao,
    Sharada,
    SoraSompeng,
    Takri,
    CaucasianAlbanian,
    BassaVah,
    Duployan,
    Elbasan,
    Grantha,
    PahawhHmong,
    Khojki,
    LinearA,
    Mahajani,
    Manichaean,
    MendeKikakui,
    Modi,
    Mro,
    OldNorthArabian,
    Nabataean,
    Palmyrene,
    PauCinHau,
    OldPermic,
    PsalterPahlavi,
    Siddham,
    Khudawadi,
    Tirhuta,
    WarangCiti,
    Count,
}

C++ type: QChar::Script

C++ documentation:

This enum type defines the Unicode script property values.

For details about the Unicode script property values see Unicode Standard Annex #24.

In order to conform to C/C++ naming conventions "Script_" is prepended to the codes used in the Unicode Standard.

This enum was introduced or modified in Qt 5.1.

See also script().

Variants
Unknown

For unassigned, private-use, noncharacter, and surrogate code points. (C++ enum variant: Script_Unknown = 0)

Inherited

For characters that may be used with multiple scripts and that inherit their script from the preceding characters. These include nonspacing marks, enclosing marks, and zero width joiner/non-joiner characters. (C++ enum variant: Script_Inherited = 1)

Common

For characters that may be used with multiple scripts and that do not inherit their script from the preceding characters. (C++ enum variant: Script_Common = 2)

Latin

(C++ enum variant: Script_Latin = 3)

Greek

(C++ enum variant: Script_Greek = 4)

Cyrillic

(C++ enum variant: Script_Cyrillic = 5)

Armenian

(C++ enum variant: Script_Armenian = 6)

Hebrew

(C++ enum variant: Script_Hebrew = 7)

Arabic

(C++ enum variant: Script_Arabic = 8)

Syriac

(C++ enum variant: Script_Syriac = 9)

Thaana

(C++ enum variant: Script_Thaana = 10)

Devanagari

(C++ enum variant: Script_Devanagari = 11)

Bengali

(C++ enum variant: Script_Bengali = 12)

Gurmukhi

(C++ enum variant: Script_Gurmukhi = 13)

Gujarati

(C++ enum variant: Script_Gujarati = 14)

Oriya

(C++ enum variant: Script_Oriya = 15)

Tamil

(C++ enum variant: Script_Tamil = 16)

Telugu

(C++ enum variant: Script_Telugu = 17)

Kannada

(C++ enum variant: Script_Kannada = 18)

Malayalam

(C++ enum variant: Script_Malayalam = 19)

Sinhala

(C++ enum variant: Script_Sinhala = 20)

Thai

(C++ enum variant: Script_Thai = 21)

Lao

(C++ enum variant: Script_Lao = 22)

Tibetan

(C++ enum variant: Script_Tibetan = 23)

Myanmar

(C++ enum variant: Script_Myanmar = 24)

Georgian

(C++ enum variant: Script_Georgian = 25)

Hangul

(C++ enum variant: Script_Hangul = 26)

Ethiopic

(C++ enum variant: Script_Ethiopic = 27)

Cherokee

(C++ enum variant: Script_Cherokee = 28)

CanadianAboriginal

(C++ enum variant: Script_CanadianAboriginal = 29)

Ogham

(C++ enum variant: Script_Ogham = 30)

Runic

(C++ enum variant: Script_Runic = 31)

Khmer

(C++ enum variant: Script_Khmer = 32)

Mongolian

(C++ enum variant: Script_Mongolian = 33)

Hiragana

(C++ enum variant: Script_Hiragana = 34)

Katakana

(C++ enum variant: Script_Katakana = 35)

Bopomofo

(C++ enum variant: Script_Bopomofo = 36)

Han

(C++ enum variant: Script_Han = 37)

Yi

(C++ enum variant: Script_Yi = 38)

OldItalic

(C++ enum variant: Script_OldItalic = 39)

Gothic

(C++ enum variant: Script_Gothic = 40)

Deseret

(C++ enum variant: Script_Deseret = 41)

Tagalog

(C++ enum variant: Script_Tagalog = 42)

Hanunoo

(C++ enum variant: Script_Hanunoo = 43)

Buhid

(C++ enum variant: Script_Buhid = 44)

Tagbanwa

(C++ enum variant: Script_Tagbanwa = 45)

Coptic

(C++ enum variant: Script_Coptic = 46)

Limbu

(C++ enum variant: Script_Limbu = 47)

TaiLe

(C++ enum variant: Script_TaiLe = 48)

LinearB

(C++ enum variant: Script_LinearB = 49)

Ugaritic

(C++ enum variant: Script_Ugaritic = 50)

Shavian

(C++ enum variant: Script_Shavian = 51)

Osmanya

(C++ enum variant: Script_Osmanya = 52)

Cypriot

(C++ enum variant: Script_Cypriot = 53)

Braille

(C++ enum variant: Script_Braille = 54)

Buginese

(C++ enum variant: Script_Buginese = 55)

NewTaiLue

(C++ enum variant: Script_NewTaiLue = 56)

Glagolitic

(C++ enum variant: Script_Glagolitic = 57)

Tifinagh

(C++ enum variant: Script_Tifinagh = 58)

SylotiNagri

(C++ enum variant: Script_SylotiNagri = 59)

OldPersian

(C++ enum variant: Script_OldPersian = 60)

Kharoshthi

(C++ enum variant: Script_Kharoshthi = 61)

Balinese

(C++ enum variant: Script_Balinese = 62)

Cuneiform

(C++ enum variant: Script_Cuneiform = 63)

Phoenician

(C++ enum variant: Script_Phoenician = 64)

PhagsPa

(C++ enum variant: Script_PhagsPa = 65)

Nko

(C++ enum variant: Script_Nko = 66)

Sundanese

(C++ enum variant: Script_Sundanese = 67)

Lepcha

(C++ enum variant: Script_Lepcha = 68)

OlChiki

(C++ enum variant: Script_OlChiki = 69)

Vai

(C++ enum variant: Script_Vai = 70)

Saurashtra

(C++ enum variant: Script_Saurashtra = 71)

KayahLi

(C++ enum variant: Script_KayahLi = 72)

Rejang

(C++ enum variant: Script_Rejang = 73)

Lycian

(C++ enum variant: Script_Lycian = 74)

Carian

(C++ enum variant: Script_Carian = 75)

Lydian

(C++ enum variant: Script_Lydian = 76)

Cham

(C++ enum variant: Script_Cham = 77)

TaiTham

(C++ enum variant: Script_TaiTham = 78)

TaiViet

(C++ enum variant: Script_TaiViet = 79)

Avestan

(C++ enum variant: Script_Avestan = 80)

EgyptianHieroglyphs

(C++ enum variant: Script_EgyptianHieroglyphs = 81)

Samaritan

(C++ enum variant: Script_Samaritan = 82)

Lisu

(C++ enum variant: Script_Lisu = 83)

Bamum

(C++ enum variant: Script_Bamum = 84)

Javanese

(C++ enum variant: Script_Javanese = 85)

MeeteiMayek

(C++ enum variant: Script_MeeteiMayek = 86)

ImperialAramaic

(C++ enum variant: Script_ImperialAramaic = 87)

OldSouthArabian

(C++ enum variant: Script_OldSouthArabian = 88)

InscriptionalParthian

(C++ enum variant: Script_InscriptionalParthian = 89)

InscriptionalPahlavi

(C++ enum variant: Script_InscriptionalPahlavi = 90)

OldTurkic

(C++ enum variant: Script_OldTurkic = 91)

Kaithi

(C++ enum variant: Script_Kaithi = 92)

Batak

(C++ enum variant: Script_Batak = 93)

Brahmi

(C++ enum variant: Script_Brahmi = 94)

Mandaic

(C++ enum variant: Script_Mandaic = 95)

Chakma

(C++ enum variant: Script_Chakma = 96)

MeroiticCursive

(C++ enum variant: Script_MeroiticCursive = 97)

MeroiticHieroglyphs

(C++ enum variant: Script_MeroiticHieroglyphs = 98)

Miao

(C++ enum variant: Script_Miao = 99)

Sharada

(C++ enum variant: Script_Sharada = 100)

SoraSompeng

(C++ enum variant: Script_SoraSompeng = 101)

Takri

(C++ enum variant: Script_Takri = 102)

CaucasianAlbanian

(C++ enum variant: Script_CaucasianAlbanian = 103)

BassaVah

(C++ enum variant: Script_BassaVah = 104)

Duployan

(C++ enum variant: Script_Duployan = 105)

Elbasan

(C++ enum variant: Script_Elbasan = 106)

Grantha

(C++ enum variant: Script_Grantha = 107)

PahawhHmong

(C++ enum variant: Script_PahawhHmong = 108)

Khojki

(C++ enum variant: Script_Khojki = 109)

LinearA

(C++ enum variant: Script_LinearA = 110)

Mahajani

(C++ enum variant: Script_Mahajani = 111)

Manichaean

(C++ enum variant: Script_Manichaean = 112)

MendeKikakui

(C++ enum variant: Script_MendeKikakui = 113)

Modi

(C++ enum variant: Script_Modi = 114)

Mro

(C++ enum variant: Script_Mro = 115)

OldNorthArabian

(C++ enum variant: Script_OldNorthArabian = 116)

Nabataean

(C++ enum variant: Script_Nabataean = 117)

Palmyrene

(C++ enum variant: Script_Palmyrene = 118)

PauCinHau

(C++ enum variant: Script_PauCinHau = 119)

OldPermic

(C++ enum variant: Script_OldPermic = 120)

PsalterPahlavi

(C++ enum variant: Script_PsalterPahlavi = 121)

Siddham

(C++ enum variant: Script_Siddham = 122)

Khudawadi

(C++ enum variant: Script_Khudawadi = 123)

Tirhuta

(C++ enum variant: Script_Tirhuta = 124)

WarangCiti

(C++ enum variant: Script_WarangCiti = 125)

Count

C++ enum variant: ScriptCount = 126