Enum qt_core::char::SpecialCharacter

pub enum SpecialCharacter {
    Null,
    Tabulation,
    LineFeed,
    CarriageReturn,
    Space,
    Nbsp,
    SoftHyphen,
    LineSeparator,
    ParagraphSeparator,
    ByteOrderMark,
    ObjectReplacementCharacter,
    ReplacementCharacter,
    ByteOrderSwapped,
    LastValidCodePoint,
}

C++ type: QChar::SpecialCharacter

C++ documentation:

Variants
Null

A QChar with this value isNull(). (C++ enum variant: Null = 0)

Tabulation

Character tabulation. (C++ enum variant: Tabulation = 9)

LineFeed

(C++ enum variant: LineFeed = 10)

CarriageReturn

(C++ enum variant: CarriageReturn = 13)

Space

(C++ enum variant: Space = 32)

Nbsp

Non-breaking space. (C++ enum variant: Nbsp = 160)

SoftHyphen

(C++ enum variant: SoftHyphen = 173)

LineSeparator

(C++ enum variant: LineSeparator = 8232)

ParagraphSeparator

(C++ enum variant: ParagraphSeparator = 8233)

ByteOrderMark

(C++ enum variant: ByteOrderMark = 65279)

ObjectReplacementCharacter

Used to represent an object such as an image when such objects cannot be presented. (C++ enum variant: ObjectReplacementCharacter = 65532)

ReplacementCharacter

The character shown when a font has no glyph for a certain codepoint. A special question mark character is often used. Codecs use this codepoint when input data cannot be represented in Unicode. (C++ enum variant: ReplacementCharacter = 65533)

ByteOrderSwapped

(C++ enum variant: ByteOrderSwapped = 65534)

LastValidCodePoint

(C++ enum variant: LastValidCodePoint = 1114111)