Enum qt_core::char::UnicodeVersion

pub enum UnicodeVersion {
    UnicodeUnassigned,
    Unicode11,
    Unicode20,
    Unicode212,
    Unicode30,
    Unicode31,
    Unicode32,
    Unicode40,
    Unicode41,
    Unicode50,
    Unicode51,
    Unicode52,
    Unicode60,
    Unicode61,
    Unicode62,
    Unicode63,
    Unicode70,
}

C++ type: QChar::UnicodeVersion

C++ documentation:

Specifies which version of the Unicode standard introduced a certain character.

See also unicodeVersion() and currentUnicodeVersion().

Variants
UnicodeUnassigned

The value is not assigned to any character in version 6.3 of Unicode. (C++ enum variant: Unicode_Unassigned = 0)

Unicode11

Version 1.1 (C++ enum variant: Unicode_1_1 = 1)

Unicode20

Version 2.0 (C++ enum variant: Unicode_2_0 = 2)

Unicode212

Version 2.1.2 (C++ enum variant: Unicode_2_1_2 = 3)

Unicode30

Version 3.0 (C++ enum variant: Unicode_3_0 = 4)

Unicode31

Version 3.1 (C++ enum variant: Unicode_3_1 = 5)

Unicode32

Version 3.2 (C++ enum variant: Unicode_3_2 = 6)

Unicode40

Version 4.0 (C++ enum variant: Unicode_4_0 = 7)

Unicode41

Version 4.1 (C++ enum variant: Unicode_4_1 = 8)

Unicode50

Version 5.0 (C++ enum variant: Unicode_5_0 = 9)

Unicode51

Version 5.1 (C++ enum variant: Unicode_5_1 = 10)

Unicode52

Version 5.2 (C++ enum variant: Unicode_5_2 = 11)

Unicode60

Version 6.0 (C++ enum variant: Unicode_6_0 = 12)

Unicode61

Version 6.1 (C++ enum variant: Unicode_6_1 = 13)

Unicode62

Version 6.2 (C++ enum variant: Unicode_6_2 = 14)

Unicode63

Version 6.3 Since Qt 5.3 (C++ enum variant: Unicode_6_3 = 15)

Unicode70

C++ enum variant: Unicode_7_0 = 16