Struct qt_core::char::Char

pub struct Char {
    // some fields omitted
}

C++ type: QChar

C++ documentation:

The QChar class provides a 16-bit Unicode character.

In Qt, Unicode characters are 16-bit entities without any markup or structure. This class represents such an entity. It is lightweight, so it can be used everywhere. Most compilers treat it like a unsigned short.

QChar provides a full complement of testing/classification functions, converting to and from other formats, converting from composed to decomposed Unicode, and trying to compare and case-convert if you ask it to.

The classification functions include functions like those in the standard C++ header <cctype> (formerly <ctype.h>), but operating on the full range of Unicode characters, not just for the ASCII range. They all return true if the character is a certain type of character; otherwise they return false. These classification functions are isNull() (returns true if the character is '\0'), isPrint() (true if the character is any sort of printable character, including whitespace), isPunct() (any sort of punctation), isMark() (Unicode Mark), isLetter() (a letter), isNumber() (any sort of numeric character, not just 0-9), isLetterOrNumber(), and isDigit() (decimal digits). All of these are wrappers around category() which return the Unicode-defined category of each character. Some of these also calculate the derived properties (for example isSpace() returns true if the character is of category Separator_* or an exceptional code point from Other_Control category).

QChar also provides direction(), which indicates the "natural" writing direction of this character. The joiningType() function indicates how the character joins with it's neighbors (needed mostly for Arabic or Syriac) and finally hasMirrored(), which indicates whether the character needs to be mirrored when it is printed in it's "unnatural" writing direction.

Composed Unicode characters (like ring) can be converted to decomposed Unicode ("a" followed by "ring above") by using decomposition().

In Unicode, comparison is not necessarily possible and case conversion is very difficult at best. Unicode, covering the "entire" world, also includes most of the world's case and sorting problems. operator==() and friends will do comparison based purely on the numeric Unicode value (code point) of the characters, and toUpper() and toLower() will do case changes when the character has a well-defined uppercase/lowercase equivalent. For locale-dependent comparisons, use QString::localeAwareCompare().

The conversion functions include unicode() (to a scalar), toLatin1() (to scalar, but converts all non-Latin-1 characters to 0), row() (gives the Unicode row), cell() (gives the Unicode cell), digitValue() (gives the integer value of any of the numerous digit characters), and a host of constructors.

QChar provides constructors and cast operators that make it easy to convert to and from traditional 8-bit chars. If you defined QT_NO_CAST_FROM_ASCII and QT_NO_CAST_TO_ASCII, as explained in the QString documentation, you will need to explicitly call fromLatin1(), or use QLatin1Char, to construct a QChar from an 8-bit char, and you will need to call toLatin1() to get the 8-bit value back.

For more information see "About the Unicode Character Database".

Methods

impl Char

fn category(&self) -> Category

C++ method: QChar::Category QChar::category() const

C++ documentation:

Returns the character's category.

fn category_static(ucs4: c_uint) -> Category

C++ method: static QChar::Category QChar::category(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the category of the UCS-4-encoded character specified by ucs4.

fn cell(&self) -> c_uchar

C++ method: unsigned char QChar::cell() const

C++ documentation:

Returns the cell (least significant byte) of the Unicode character.

See also row().

fn combining_class(&self) -> c_uchar

C++ method: unsigned char QChar::combiningClass() const

C++ documentation:

Returns the combining class for the character as defined in the Unicode standard. This is mainly useful as a positioning hint for marks attached to a base character.

The Qt text rendering engine uses this information to correctly position non-spacing marks around a base character.

fn combining_class_static(ucs4: c_uint) -> c_uchar

C++ method: static unsigned char QChar::combiningClass(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the combining class for the UCS-4-encoded character specified by ucs4, as defined in the Unicode standard.

fn current_unicode_version() -> UnicodeVersion

C++ method: static QChar::UnicodeVersion QChar::currentUnicodeVersion()

C++ documentation:

Returns the most recent supported Unicode version.

fn decomposition<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CharDecompositionArgs<'largs>

C++ method: QChar::decomposition

This is an overloaded function. Available variants:

Rust arguments:
1) fn decomposition(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn decomposition(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QChar::decomposition() const

C++ documentation:

Decomposes a character into it's constituent parts. Returns an empty string if no decomposition exists.

fn decomposition_static<Args>(args: Args) -> Args::ReturnType where Args: CharDecompositionStaticArgs

C++ method: QChar::decomposition

This is an overloaded function. Available variants:

Rust arguments:
1) fn decomposition_static((libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn decomposition_static((libc::c_uint, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QChar::decomposition(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Decomposes the UCS-4-encoded character specified by ucs4 into it's constituent parts. Returns an empty string if no decomposition exists.

fn decomposition_tag(&self) -> Decomposition

C++ method: QChar::Decomposition QChar::decompositionTag() const

C++ documentation:

Returns the tag defining the composition of the character. Returns QChar::NoDecomposition if no decomposition exists.

fn decomposition_tag_static(ucs4: c_uint) -> Decomposition

C++ method: static QChar::Decomposition QChar::decompositionTag(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the tag defining the composition of the UCS-4-encoded character specified by ucs4. Returns QChar::NoDecomposition if no decomposition exists.

fn digit_value(&self) -> c_int

C++ method: int QChar::digitValue() const

C++ documentation:

Returns the numeric value of the digit, or -1 if the character is not a digit.

fn digit_value_static(ucs4: c_uint) -> c_int

C++ method: static int QChar::digitValue(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the numeric value of the digit specified by the UCS-4-encoded character, ucs4, or -1 if the character is not a digit.

fn direction(&self) -> Direction

C++ method: QChar::Direction QChar::direction() const

C++ documentation:

Returns the character's direction.

fn direction_static(ucs4: c_uint) -> Direction

C++ method: static QChar::Direction QChar::direction(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the direction of the UCS-4-encoded character specified by ucs4.

fn from_latin1<Args>(args: Args) -> Args::ReturnType where Args: CharFromLatin1Args

C++ method: QChar::fromLatin1

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_latin1((libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
2) fn from_latin1((libc::c_char, cpp_utils::AsStruct)) -> ::char::Char
C++ method: static QChar QChar::fromLatin1(char c)

C++ documentation:

Converts the Latin-1 character c to its equivalent QChar. This is mainly useful for non-internationalized software.

An alternative is to use QLatin1Char.

See also toLatin1() and unicode().

fn has_mirrored(&self) -> bool

C++ method: bool QChar::hasMirrored() const

C++ documentation:

Returns true if the character should be reversed if the text direction is reversed; otherwise returns false.

A bit faster equivalent of (ch.mirroredChar() != ch).

See also mirroredChar().

fn has_mirrored_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::hasMirrored(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 should be reversed if the text direction is reversed; otherwise returns false.

A bit faster equivalent of (QChar::mirroredChar(ucs4) != ucs4).

This function was introduced in Qt 5.0.

See also mirroredChar().

fn high_surrogate(ucs4: c_uint) -> c_ushort

C++ method: static unsigned short QChar::highSurrogate(unsigned int ucs4)

C++ documentation:

Returns the high surrogate part of a UCS-4-encoded code point. The returned result is undefined if ucs4 is smaller than 0x10000.

fn is_digit(&self) -> bool

C++ method: bool QChar::isDigit() const

C++ documentation:

Returns true if the character is a decimal digit (Number_DecimalDigit); otherwise returns false.

See also isNumber().

fn is_digit_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isDigit(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a decimal digit (Number_DecimalDigit); otherwise returns false.

This function was introduced in Qt 5.0.

See also isNumber().

fn is_high_surrogate(&self) -> bool

C++ method: bool QChar::isHighSurrogate() const

C++ documentation:

Returns true if the QChar is the high part of a UTF16 surrogate (for example if its code point is in range [0xd800..0xdbff]); false otherwise.

fn is_high_surrogate_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isHighSurrogate(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is the high part of a UTF16 surrogate (for example if its code point is in range [0xd800..0xdbff]); false otherwise.

fn is_letter(&self) -> bool

C++ method: bool QChar::isLetter() const

C++ documentation:

Returns true if the character is a letter (Letter_* categories); otherwise returns false.

fn is_letter_or_number(&self) -> bool

C++ method: bool QChar::isLetterOrNumber() const

C++ documentation:

Returns true if the character is a letter or number (Letter_* or Number_* categories); otherwise returns false.

fn is_letter_or_number_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isLetterOrNumber(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a letter or number (Letter_* or Number_* categories); otherwise returns false.

This function was introduced in Qt 5.0.

fn is_letter_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isLetter(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a letter (Letter_* categories); otherwise returns false.

This function was introduced in Qt 5.0.

fn is_low_surrogate(&self) -> bool

C++ method: bool QChar::isLowSurrogate() const

C++ documentation:

Returns true if the QChar is the low part of a UTF16 surrogate (for example if its code point is in range [0xdc00..0xdfff]); false otherwise.

fn is_low_surrogate_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isLowSurrogate(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is the low part of a UTF16 surrogate (for example if its code point is in range [0xdc00..0xdfff]); false otherwise.

fn is_lower(&self) -> bool

C++ method: bool QChar::isLower() const

C++ documentation:

Returns true if the character is a lowercase letter, for example category() is Letter_Lowercase.

See also isUpper(), toLower(), and toUpper().

fn is_lower_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isLower(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a lowercase letter, for example category() is Letter_Lowercase.

This function was introduced in Qt 5.0.

See also isUpper(), toLower(), and toUpper().

fn is_mark(&self) -> bool

C++ method: bool QChar::isMark() const

C++ documentation:

Returns true if the character is a mark (Mark_* categories); otherwise returns false.

See QChar::Category for more information regarding marks.

fn is_mark_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isMark(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a mark (Mark_* categories); otherwise returns false.

This function was introduced in Qt 5.0.

fn is_non_character(&self) -> bool

C++ method: bool QChar::isNonCharacter() const

C++ documentation:

Returns true if the QChar is a non-character; false otherwise.

Unicode has a certain number of code points that are classified as "non-characters:" that is, they can be used for internal purposes in applications but cannot be used for text interchange. Those are the last two entries each Unicode Plane ([0xfffe..0xffff], [0x1fffe..0x1ffff], etc.) as well as the entries in range [0xfdd0..0xfdef].

This function was introduced in Qt 5.0.

fn is_non_character_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isNonCharacter(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a non-character; false otherwise.

Unicode has a certain number of code points that are classified as "non-characters:" that is, they can be used for internal purposes in applications but cannot be used for text interchange. Those are the last two entries each Unicode Plane ([0xfffe..0xffff], [0x1fffe..0x1ffff], etc.) as well as the entries in range [0xfdd0..0xfdef].

This function was introduced in Qt 5.0.

fn is_null(&self) -> bool

C++ method: bool QChar::isNull() const

C++ documentation:

Returns true if the character is the Unicode character 0x0000 ('\0'); otherwise returns false.

fn is_number(&self) -> bool

C++ method: bool QChar::isNumber() const

C++ documentation:

Returns true if the character is a number (Number_* categories, not just 0-9); otherwise returns false.

See also isDigit().

fn is_number_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isNumber(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a number (Number_* categories, not just 0-9); otherwise returns false.

This function was introduced in Qt 5.0.

See also isDigit().

fn is_print(&self) -> bool

C++ method: bool QChar::isPrint() const

C++ documentation:

Returns true if the character is a printable character; otherwise returns false. This is any character not of category Other_*.

Note that this gives no indication of whether the character is available in a particular font.

fn is_print_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isPrint(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a printable character; otherwise returns false. This is any character not of category Other_*.

Note that this gives no indication of whether the character is available in a particular font.

This function was introduced in Qt 5.0.

fn is_punct(&self) -> bool

C++ method: bool QChar::isPunct() const

C++ documentation:

Returns true if the character is a punctuation mark (Punctuation_* categories); otherwise returns false.

fn is_punct_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isPunct(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a punctuation mark (Punctuation_* categories); otherwise returns false.

This function was introduced in Qt 5.0.

fn is_space(&self) -> bool

C++ method: bool QChar::isSpace() const

C++ documentation:

Returns true if the character is a separator character (Separator_* categories or certain code points from Other_Control category); otherwise returns false.

fn is_space_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isSpace(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a separator character (Separator_* categories or certain code points from Other_Control category); otherwise returns false.

This function was introduced in Qt 5.0.

fn is_surrogate(&self) -> bool

C++ method: bool QChar::isSurrogate() const

C++ documentation:

Returns true if the QChar contains a code point that is in either the high or the low part of the UTF-16 surrogate range (for example if its code point is in range [0xd800..0xdfff]); false otherwise.

This function was introduced in Qt 5.0.

fn is_surrogate_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isSurrogate(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 contains a code point that is in either the high or the low part of the UTF-16 surrogate range (for example if its code point is in range [0xd800..0xdfff]); false otherwise.

This function was introduced in Qt 5.0.

fn is_symbol(&self) -> bool

C++ method: bool QChar::isSymbol() const

C++ documentation:

Returns true if the character is a symbol (Symbol_* categories); otherwise returns false.

fn is_symbol_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isSymbol(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a symbol (Symbol_* categories); otherwise returns false.

This function was introduced in Qt 5.0.

fn is_title_case(&self) -> bool

C++ method: bool QChar::isTitleCase() const

C++ documentation:

Returns true if the character is a titlecase letter, for example category() is Letter_Titlecase.

See also isLower(), toUpper(), toLower(), and toTitleCase().

fn is_title_case_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isTitleCase(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is a titlecase letter, for example category() is Letter_Titlecase.

This function was introduced in Qt 5.0.

See also isLower(), toUpper(), toLower(), and toTitleCase().

fn is_upper(&self) -> bool

C++ method: bool QChar::isUpper() const

C++ documentation:

Returns true if the character is an uppercase letter, for example category() is Letter_Uppercase.

See also isLower(), toUpper(), and toLower().

fn is_upper_static(ucs4: c_uint) -> bool

C++ method: static bool QChar::isUpper(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns true if the UCS-4-encoded character specified by ucs4 is an uppercase letter, for example category() is Letter_Uppercase.

This function was introduced in Qt 5.0.

See also isLower(), toUpper(), and toLower().

fn joining(&self) -> Joining

C++ method: QChar::Joining QChar::joining() const

C++ documentation:

Returns information about the joining properties of the character (needed for certain languages such as Arabic).

fn joining_static(ucs4: c_uint) -> Joining

C++ method: static QChar::Joining QChar::joining(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns information about the joining properties of the UCS-4-encoded character specified by ucs4 (needed for certain languages such as Arabic).

fn joining_type(&self) -> JoiningType

C++ method: QChar::JoiningType QChar::joiningType() const

C++ documentation:

Returns information about the joining type attributes of the character (needed for certain languages such as Arabic or Syriac).

This function was introduced in Qt 5.3.

fn joining_type_static(ucs4: c_uint) -> JoiningType

C++ method: static QChar::JoiningType QChar::joiningType(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns information about the joining type attributes of the UCS-4-encoded character specified by ucs4 (needed for certain languages such as Arabic or Syriac).

This function was introduced in Qt 5.3.

fn low_surrogate(ucs4: c_uint) -> c_ushort

C++ method: static unsigned short QChar::lowSurrogate(unsigned int ucs4)

C++ documentation:

Returns the low surrogate part of a UCS-4-encoded code point. The returned result is undefined if ucs4 is smaller than 0x10000.

fn mirrored_char<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CharMirroredCharArgs<'largs>

C++ method: QChar::mirroredChar

This is an overloaded function. Available variants:

Rust arguments:
1) fn mirrored_char(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn mirrored_char(&self, cpp_utils::AsStruct) -> ::char::Char
C++ method: QChar QChar::mirroredChar() const

C++ documentation:

Returns the mirrored character if this character is a mirrored character; otherwise returns the character itself.

See also hasMirrored().

fn mirrored_char_static(ucs4: c_uint) -> c_uint

C++ method: static unsigned int QChar::mirroredChar(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the mirrored character if the UCS-4-encoded character specified by ucs4 is a mirrored character; otherwise returns the character itself.

See also hasMirrored().

fn new<Args>(args: Args) -> Args::ReturnType where Args: CharNewArgs

C++ method: QChar::QChar

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::char::Char
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar()

Variant 2

Rust arguments:
1) fn new((::char::SpecialCharacter, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((::char::SpecialCharacter, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(QChar::SpecialCharacter s)

Variant 3

Rust arguments:
1) fn new((&::latin1_char::Latin1Char, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((&::latin1_char::Latin1Char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(QLatin1Char ch)

Variant 4

Rust arguments:
1) fn new((libc::c_char, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(char c)

Variant 5

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(int rc)

Variant 6

Rust arguments:
1) fn new((libc::c_short, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((libc::c_short, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(short rc)

Variant 7

Rust arguments:
1) fn new((libc::c_uchar, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((libc::c_uchar, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(unsigned char c)

Variant 8

Rust arguments:
1) fn new((libc::c_uchar, libc::c_uchar, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((libc::c_uchar, libc::c_uchar, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(unsigned char c, unsigned char r)

Variant 9

Rust arguments:
1) fn new((libc::c_uint, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(unsigned int rc)

Variant 10

Rust arguments:
1) fn new((libc::c_ushort, cpp_utils::AsStruct)) -> ::char::Char
2) fn new((libc::c_ushort, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
C++ method: [constructor] void QChar::QChar(unsigned short rc)

fn requires_surrogates(ucs4: c_uint) -> bool

C++ method: static bool QChar::requiresSurrogates(unsigned int ucs4)

C++ documentation:

Returns true if the UCS-4-encoded character specified by ucs4 can be split into the high and low parts of a UTF16 surrogate (for example if its code point is greater than or equals to 0x10000); false otherwise.

fn row(&self) -> c_uchar

C++ method: unsigned char QChar::row() const

C++ documentation:

Returns the row (most significant byte) of the Unicode character.

See also cell().

fn script(&self) -> Script

C++ method: QChar::Script QChar::script() const

C++ documentation:

Returns the Unicode script property value for this character.

This function was introduced in Qt 5.1.

fn script_static(ucs4: c_uint) -> Script

C++ method: static QChar::Script QChar::script(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the Unicode script property value for the character specified in its UCS-4-encoded form as ucs4.

This function was introduced in Qt 5.1.

fn set_cell(&mut self, cell: c_uchar)

C++ method: void QChar::setCell(unsigned char cell)

fn set_row(&mut self, row: c_uchar)

C++ method: void QChar::setRow(unsigned char row)

fn surrogate_to_ucs4<Args>(args: Args) -> c_uint where Args: CharSurrogateToUcs4Args

C++ method: QChar::surrogateToUcs4

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn surrogate_to_ucs4((&::char::Char, &::char::Char)) -> libc::c_uint
C++ method: static unsigned int QChar::surrogateToUcs4(QChar high, QChar low)

C++ documentation:

This is an overloaded function.

Converts a UTF16 surrogate pair (high, low) to it's UCS-4-encoded code point.

Variant 2

Rust arguments: fn surrogate_to_ucs4((libc::c_ushort, libc::c_ushort)) -> libc::c_uint
C++ method: static unsigned int QChar::surrogateToUcs4(unsigned short high, unsigned short low)

C++ documentation:

Converts a UTF16 surrogate pair with the given high and low values to it's UCS-4-encoded code point.

fn to_case_folded<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CharToCaseFoldedArgs<'largs>

C++ method: QChar::toCaseFolded

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_case_folded(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn to_case_folded(&self, cpp_utils::AsStruct) -> ::char::Char
C++ method: QChar QChar::toCaseFolded() const

C++ documentation:

Returns the case folded equivalent of the character. For most Unicode characters this is the same as toLower().

fn to_case_folded_static(ucs4: c_uint) -> c_uint

C++ method: static unsigned int QChar::toCaseFolded(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the case folded equivalent of the UCS-4-encoded character specified by ucs4. For most Unicode characters this is the same as toLower().

fn to_latin1(&self) -> c_char

C++ method: char QChar::toLatin1() const

C++ documentation:

Returns the Latin-1 character equivalent to the QChar, or 0. This is mainly useful for non-internationalized software.

Note: It is not possible to distinguish a non-Latin-1 character from a Latin-1 0 (NUL) character. Prefer to use unicode(), which does not have this ambiguity.

See also unicode().

fn to_lower<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CharToLowerArgs<'largs>

C++ method: QChar::toLower

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_lower(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn to_lower(&self, cpp_utils::AsStruct) -> ::char::Char
C++ method: QChar QChar::toLower() const

C++ documentation:

Returns the lowercase equivalent if the character is uppercase or titlecase; otherwise returns the character itself.

fn to_lower_static(ucs4: c_uint) -> c_uint

C++ method: static unsigned int QChar::toLower(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the lowercase equivalent of the UCS-4-encoded character specified by ucs4 if the character is uppercase or titlecase; otherwise returns the character itself.

fn to_title_case<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CharToTitleCaseArgs<'largs>

C++ method: QChar::toTitleCase

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_title_case(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn to_title_case(&self, cpp_utils::AsStruct) -> ::char::Char
C++ method: QChar QChar::toTitleCase() const

C++ documentation:

Returns the title case equivalent if the character is lowercase or uppercase; otherwise returns the character itself.

fn to_title_case_static(ucs4: c_uint) -> c_uint

C++ method: static unsigned int QChar::toTitleCase(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the title case equivalent of the UCS-4-encoded character specified by ucs4 if the character is lowercase or uppercase; otherwise returns the character itself.

fn to_upper<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CharToUpperArgs<'largs>

C++ method: QChar::toUpper

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_upper(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn to_upper(&self, cpp_utils::AsStruct) -> ::char::Char
C++ method: QChar QChar::toUpper() const

C++ documentation:

Returns the uppercase equivalent if the character is lowercase or titlecase; otherwise returns the character itself.

fn to_upper_static(ucs4: c_uint) -> c_uint

C++ method: static unsigned int QChar::toUpper(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the uppercase equivalent of the UCS-4-encoded character specified by ucs4 if the character is lowercase or titlecase; otherwise returns the character itself.

fn unicode(&self) -> c_ushort

C++ method: unsigned short QChar::unicode() const

C++ documentation:

Returns the numeric Unicode value of the QChar.

fn unicode_mut<'l0>(&'l0 mut self) -> &'l0 mut c_ushort

C++ method: unsigned short& QChar::unicode()

C++ documentation:

Returns a reference to the numeric Unicode value of the QChar.

fn unicode_version(&self) -> UnicodeVersion

C++ method: QChar::UnicodeVersion QChar::unicodeVersion() const

C++ documentation:

Returns the Unicode version that introduced this character.

fn unicode_version_static(ucs4: c_uint) -> UnicodeVersion

C++ method: static QChar::UnicodeVersion QChar::unicodeVersion(unsigned int ucs4)

C++ documentation:

This is an overloaded function.

Returns the Unicode version that introduced the character specified in its UCS-4-encoded form as ucs4.