Module qt_core::char_ref::overloading

Traits

CharRefAsQCharArgs
This trait represents a set of arguments accepted by CharRef::as_q_char method.

CharRefDecompositionArgs
This trait represents a set of arguments accepted by CharRef::decomposition method.

CharRefMirroredCharArgs
This trait represents a set of arguments accepted by CharRef::mirrored_char method.

CharRefOpAssignArgs
This trait represents a set of arguments accepted by CharRef::op_assign method.

CharRefToLowerArgs
This trait represents a set of arguments accepted by CharRef::to_lower method.

CharRefToTitleCaseArgs
This trait represents a set of arguments accepted by CharRef::to_title_case method.

CharRefToUpperArgs
This trait represents a set of arguments accepted by CharRef::to_upper method.