Struct qt_core::collator::Collator

pub struct Collator {
    // some fields omitted
}

C++ type: QCollator

C++ documentation:

The QCollator class compares strings according to a localized collation algorithm.

QCollator is initialized with a QLocale and an optional collation strategy. It tries to initialize the collator with the specified values. The collator can then be used to compare and sort strings in a locale dependent fashion.

A QCollator object can be used together with template based sorting algorithms such as std::sort to sort a list of QStrings.

In addition to the locale and collation strategy, several optional flags can be set that influence the result of the collation.

Methods

impl Collator

fn case_sensitivity(&self) -> CaseSensitivity

C++ method: Qt::CaseSensitivity QCollator::caseSensitivity() const

C++ documentation:

Returns case sensitivity of the collator.

See also setCaseSensitivity().

fn compare<'largs, Args>(&'largs self, args: Args) -> c_int where Args: CollatorCompareArgs<'largs>

C++ method: QCollator::compare

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn compare(&self, (*const ::char::Char, libc::c_int, *const ::char::Char, libc::c_int)) -> libc::c_int
C++ method: int QCollator::compare(const QChar* s1, int len1, const QChar* s2, int len2) const

C++ documentation:

This is an overloaded function.

Compares s1 with s2. len1 and len2 specify the length of the QChar arrays pointer to by s1 and s2.

Returns an integer less than, equal to, or greater than zero depending on whether s1 is smaller, equal or larger than s2.

Variant 2

Rust arguments: fn compare(&self, (&::string::String, &::string::String)) -> libc::c_int
C++ method: int QCollator::compare(const QString& s1, const QString& s2) const

C++ documentation:

Compares s1 with s2. Returns an integer less than, equal to, or greater than zero depending on whether s1 is smaller, equal or larger than s2.

Variant 3

Rust arguments: fn compare(&self, (&::string_ref::StringRef, &::string_ref::StringRef)) -> libc::c_int
C++ method: int QCollator::compare(const QStringRef& s1, const QStringRef& s2) const

C++ documentation:

This is an overloaded function.

Compares s1 with s2. Returns an integer less than, equal to, or greater than zero depending on whether s1 is smaller, equal or larger than s2.

fn ignore_punctuation(&self) -> bool

C++ method: bool QCollator::ignorePunctuation() const

C++ documentation:

Returns true if punctuation characters and symbols are ignored when determining sort order.

See also setIgnorePunctuation().

fn locale<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CollatorLocaleArgs<'largs>

C++ method: QCollator::locale

This is an overloaded function. Available variants:

Rust arguments:
1) fn locale(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::locale::Locale>
2) fn locale(&self, cpp_utils::AsStruct) -> ::locale::Locale
C++ method: QLocale QCollator::locale() const

C++ documentation:

Returns the locale of the collator.

See also setLocale().

fn new<Args>(args: Args) -> Args::ReturnType where Args: CollatorNewArgs

C++ method: QCollator::QCollator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::collator::Collator, cpp_utils::AsStruct)) -> ::collator::Collator
2) fn new((&::collator::Collator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::collator::Collator>
C++ method: [constructor] void QCollator::QCollator(const QCollator& arg1)

Variant 2

Rust arguments:
1) fn new((&::locale::Locale, cpp_utils::AsStruct)) -> ::collator::Collator
2) fn new(cpp_utils::AsStruct) -> ::collator::Collator
3) fn new((&::locale::Locale, cpp_utils::AsBox)) -> cpp_utils::CppBox<::collator::Collator>
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::collator::Collator>
C++ method: [constructor] void QCollator::QCollator(const QLocale& locale = ?)

fn numeric_mode(&self) -> bool

C++ method: bool QCollator::numericMode() const

C++ documentation:

Returns true if numeric sorting is enabled, false otherwise.

See also setNumericMode().

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 Collator) -> &'l0 mut Collator

C++ method: QCollator& QCollator::operator=(const QCollator& arg1)

C++ documentation:

Assigns other to this collator.

fn op_call(&self, s1: &String, s2: &String) -> bool

C++ method: bool QCollator::operator()(const QString& s1, const QString& s2) const

fn set_case_sensitivity(&mut self, cs: CaseSensitivity)

C++ method: void QCollator::setCaseSensitivity(Qt::CaseSensitivity cs)

C++ documentation:

Sets the case sensitivity of the collator.

See also caseSensitivity().

fn set_ignore_punctuation(&mut self, on: bool)

C++ method: void QCollator::setIgnorePunctuation(bool on)

C++ documentation:

If on is set to true, punctuation characters and symbols are ignored when determining sort order.

The default is locale dependent.

See also ignorePunctuation().

fn set_locale(&mut self, locale: &Locale)

C++ method: void QCollator::setLocale(const QLocale& locale)

C++ documentation:

Sets the locale of the collator to locale.

See also locale().

fn set_numeric_mode(&mut self, on: bool)

C++ method: void QCollator::setNumericMode(bool on)

C++ documentation:

Enables numeric sorting mode when on is set to true.

This will enable proper sorting of numeric digits, so that e.g. 100 sorts after 99.

By default this mode is off.

Note: On Windows, this functionality makes use of the ICU library. If Qt was compiled without ICU support, it falls back to code using native Windows API, which only works from Windows 7 onwards. On older versions of Windows, it will not work and a warning will be emitted at runtime.

See also numericMode().

fn sort_key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CollatorSortKeyArgs<'largs>

C++ method: QCollator::sortKey

This is an overloaded function. Available variants:

Rust arguments:
1) fn sort_key(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::collator_sort_key::CollatorSortKey>
2) fn sort_key(&self, (&::string::String, cpp_utils::AsStruct)) -> ::collator_sort_key::CollatorSortKey
C++ method: QCollatorSortKey QCollator::sortKey(const QString& string) const

C++ documentation:

Returns a sortKey for string.

Creating the sort key is usually somewhat slower, than using the compare() methods directly. But if the string is compared repeatedly (e.g. when sorting a whole list of strings), it's usually faster to create the sort keys for each string and then sort using the keys.

fn swap(&mut self, other: &mut Collator)

C++ method: void QCollator::swap(QCollator& other)

C++ documentation:

Swaps this collator with other. This function is very fast and never fails.