Struct qt_core::collator_sort_key::CollatorSortKey

pub struct CollatorSortKey {
    // some fields omitted
}

C++ type: QCollatorSortKey

C++ documentation:

The QCollatorSortKey class can be used to speed up string collation.

The QCollatorSortKey class is always created by QCollator::sortKey() and is used for fast strings collation, for example when collating many strings.

Methods

impl CollatorSortKey

fn compare(&self, key: &CollatorSortKey) -> c_int

C++ method: int QCollatorSortKey::compare(const QCollatorSortKey& key) const

C++ documentation:

Compares the key to otherKey. Returns a negative value if the key is less than otherKey, 0 if the key is equal to otherKey or a positive value if the key is greater than otherKey.

See also operator<().

fn new<Args>(args: Args) -> Args::ReturnType where Args: CollatorSortKeyNewArgs

C++ method: QCollatorSortKey::QCollatorSortKey

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::collator_sort_key::CollatorSortKey, cpp_utils::AsStruct)) -> ::collator_sort_key::CollatorSortKey
2) fn new((&::collator_sort_key::CollatorSortKey, cpp_utils::AsBox)) -> cpp_utils::CppBox<::collator_sort_key::CollatorSortKey>
C++ method: [constructor] void QCollatorSortKey::QCollatorSortKey(const QCollatorSortKey& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 CollatorSortKey) -> &'l0 mut CollatorSortKey

C++ method: QCollatorSortKey& QCollatorSortKey::operator=(const QCollatorSortKey& other)

C++ documentation:

Assigns other to this collator key.

fn swap(&mut self, other: &mut CollatorSortKey)

C++ method: void QCollatorSortKey::swap(QCollatorSortKey& other)

C++ documentation:

Swaps this collator key with other.