Struct qt_core::command_line_option::CommandLineOption

pub struct CommandLineOption {
    // some fields omitted
}

C++ type: QCommandLineOption

C++ documentation:

The QCommandLineOption class defines a possible command-line option.

This class is used to describe an option on the command line. It allows different ways of defining the same option with multiple aliases possible. It is also used to describe how the option is used - it may be a flag (e.g. -v) or take a value (e.g. -o file).

Examples:

QCommandLineOption verboseOption("verbose", "Verbose mode. Prints out more information.");
QCommandLineOption outputOption(QStringList() << "o" << "output", "Write generated data into <file>.", "file");

Methods

impl CommandLineOption

fn default_values<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CommandLineOptionDefaultValuesArgs<'largs>

C++ method: QCommandLineOption::defaultValues

This is an overloaded function. Available variants:

Rust arguments:
1) fn default_values(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn default_values(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QCommandLineOption::defaultValues() const

C++ documentation:

Returns the default values set for this option.

See also setDefaultValues().

fn description<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CommandLineOptionDescriptionArgs<'largs>

C++ method: QCommandLineOption::description

This is an overloaded function. Available variants:

Rust arguments:
1) fn description(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn description(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QCommandLineOption::description() const

C++ documentation:

Returns the description set for this option.

See also setDescription().

fn names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CommandLineOptionNamesArgs<'largs>

C++ method: QCommandLineOption::names

This is an overloaded function. Available variants:

Rust arguments:
1) fn names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn names(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QCommandLineOption::names() const

C++ documentation:

Returns the names set for this option.

fn new<Args>(args: Args) -> Args::ReturnType where Args: CommandLineOptionNewArgs

C++ method: QCommandLineOption::QCommandLineOption

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::command_line_option::CommandLineOption, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
2) fn new((&::command_line_option::CommandLineOption, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
C++ method: [constructor] void QCommandLineOption::QCommandLineOption(const QCommandLineOption& other)

Variant 2

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
C++ method: [constructor] void QCommandLineOption::QCommandLineOption(const QString& name)

Variant 3

Rust arguments:
1) fn new((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
2) fn new((&::string::String, &::string::String, &::string::String, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
3) fn new((&::string::String, &::string::String, &::string::String, &::string::String, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
4) fn new((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
5) fn new((&::string::String, &::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
6) fn new((&::string::String, &::string::String, &::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
C++ method: [constructor] void QCommandLineOption::QCommandLineOption(const QString& name, const QString& description, const QString& valueName = ?, const QString& defaultValue = ?)

Variant 4

Rust arguments:
1) fn new((&::string_list::StringList, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
2) fn new((&::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
C++ method: [constructor] void QCommandLineOption::QCommandLineOption(const QStringList& names)

Variant 5

Rust arguments:
1) fn new((&::string_list::StringList, &::string::String, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
2) fn new((&::string_list::StringList, &::string::String, &::string::String, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
3) fn new((&::string_list::StringList, &::string::String, &::string::String, &::string::String, cpp_utils::AsStruct)) -> ::command_line_option::CommandLineOption
4) fn new((&::string_list::StringList, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
5) fn new((&::string_list::StringList, &::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
6) fn new((&::string_list::StringList, &::string::String, &::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::command_line_option::CommandLineOption>
C++ method: [constructor] void QCommandLineOption::QCommandLineOption(const QStringList& names, const QString& description, const QString& valueName = ?, const QString& defaultValue = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 CommandLineOption) -> &'l0 mut CommandLineOption

C++ method: QCommandLineOption& QCommandLineOption::operator=(const QCommandLineOption& other)

C++ documentation:

Makes a copy of the other object and assigns it to this QCommandLineOption object.

fn set_default_value(&mut self, default_value: &String)

C++ method: void QCommandLineOption::setDefaultValue(const QString& defaultValue)

C++ documentation:

Sets the default value used for this option to defaultValue.

The default value is used if the user of the application does not specify the option on the command line.

If defaultValue is empty, the option has no default values.

See also defaultValues() and setDefaultValues().

fn set_default_values(&mut self, default_values: &StringList)

C++ method: void QCommandLineOption::setDefaultValues(const QStringList& defaultValues)

C++ documentation:

Sets the list of default values used for this option to defaultValues.

The default values are used if the user of the application does not specify the option on the command line.

See also defaultValues() and setDefaultValue().

fn set_description(&mut self, description: &String)

C++ method: void QCommandLineOption::setDescription(const QString& description)

C++ documentation:

Sets the description used for this option to description.

It is customary to add a "." at the end of the description.

The description is used by QCommandLineParser::showHelp().

See also description().

fn set_value_name(&mut self, name: &String)

C++ method: void QCommandLineOption::setValueName(const QString& name)

C++ documentation:

Sets the name of the expected value, for the documentation, to valueName.

Options without a value assigned have a boolean-like behavior: either the user specifies --option or they don't.

Options with a value assigned need to set a name for the expected value, for the documentation of the option in the help output. An option with names o and output, and a value name of file will appear as -o, --output <file>.

Call QCommandLineParser::value() if you expect the option to be present only once, and QCommandLineParser::values() if you expect that option to be present multiple times.

See also valueName().

fn swap(&mut self, other: &mut CommandLineOption)

C++ method: void QCommandLineOption::swap(QCommandLineOption& other)

C++ documentation:

Swaps option other with this option. This operation is very fast and never fails.

fn value_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CommandLineOptionValueNameArgs<'largs>

C++ method: QCommandLineOption::valueName

This is an overloaded function. Available variants:

Rust arguments:
1) fn value_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn value_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QCommandLineOption::valueName() const

C++ documentation:

Returns the name of the expected value.

If empty, the option doesn't take a value.

See also setValueName().