Module qt_core::command_line_parser::overloading

Traits

CommandLineParserAddHelpOptionArgs
This trait represents a set of arguments accepted by CommandLineParser::add_help_option method.

CommandLineParserAddPositionalArgumentArgs
This trait represents a set of arguments accepted by CommandLineParser::add_positional_argument method.

CommandLineParserAddVersionOptionArgs
This trait represents a set of arguments accepted by CommandLineParser::add_version_option method.

CommandLineParserApplicationDescriptionArgs
This trait represents a set of arguments accepted by CommandLineParser::application_description method.

CommandLineParserErrorTextArgs
This trait represents a set of arguments accepted by CommandLineParser::error_text method.

CommandLineParserHelpTextArgs
This trait represents a set of arguments accepted by CommandLineParser::help_text method.

CommandLineParserIsSetArgs
This trait represents a set of arguments accepted by CommandLineParser::is_set method.

CommandLineParserNewArgs
This trait represents a set of arguments accepted by CommandLineParser::new method.

CommandLineParserOptionNamesArgs
This trait represents a set of arguments accepted by CommandLineParser::option_names method.

CommandLineParserPositionalArgumentsArgs
This trait represents a set of arguments accepted by CommandLineParser::positional_arguments method.

CommandLineParserProcessArgs
This trait represents a set of arguments accepted by CommandLineParser::process method.

CommandLineParserShowHelpArgs
This trait represents a set of arguments accepted by CommandLineParser::show_help method.

CommandLineParserTrArgs
This trait represents a set of arguments accepted by CommandLineParser::tr method.

CommandLineParserTrUtf8Args
This trait represents a set of arguments accepted by CommandLineParser::tr_utf8 method.

CommandLineParserUnknownOptionNamesArgs
This trait represents a set of arguments accepted by CommandLineParser::unknown_option_names method.

CommandLineParserValueArgs
This trait represents a set of arguments accepted by CommandLineParser::value method.

CommandLineParserValuesArgs
This trait represents a set of arguments accepted by CommandLineParser::values method.