Module qt_core::core_application::overloading

Traits

CoreApplicationApplicationDirPathArgs
This trait represents a set of arguments accepted by CoreApplication::application_dir_path method.

CoreApplicationApplicationFilePathArgs
This trait represents a set of arguments accepted by CoreApplication::application_file_path method.

CoreApplicationApplicationNameArgs
This trait represents a set of arguments accepted by CoreApplication::application_name method.

CoreApplicationApplicationVersionArgs
This trait represents a set of arguments accepted by CoreApplication::application_version method.

CoreApplicationArgumentsArgs
This trait represents a set of arguments accepted by CoreApplication::arguments method.

CoreApplicationConnectArgs
This trait represents a set of arguments accepted by CoreApplication::connect method.

CoreApplicationConnectStaticArgs
This trait represents a set of arguments accepted by CoreApplication::connect_static method.

CoreApplicationDestroyedArgs
This trait represents a set of arguments accepted by CoreApplication::destroyed method.

CoreApplicationDisconnectArgs
This trait represents a set of arguments accepted by CoreApplication::disconnect method.

CoreApplicationDisconnectStaticArgs
This trait represents a set of arguments accepted by CoreApplication::disconnect_static method.

CoreApplicationDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by CoreApplication::dynamic_property_names method.

CoreApplicationExitArgs
This trait represents a set of arguments accepted by CoreApplication::exit method.

CoreApplicationLibraryPathsArgs
This trait represents a set of arguments accepted by CoreApplication::library_paths method.

CoreApplicationNewArgs
This trait represents a set of arguments accepted by CoreApplication::new method.

CoreApplicationObjectNameArgs
This trait represents a set of arguments accepted by CoreApplication::object_name method.

CoreApplicationOrganizationDomainArgs
This trait represents a set of arguments accepted by CoreApplication::organization_domain method.

CoreApplicationOrganizationNameArgs
This trait represents a set of arguments accepted by CoreApplication::organization_name method.

CoreApplicationPostEventArgs
This trait represents a set of arguments accepted by CoreApplication::post_event method.

CoreApplicationProcessEventsArgs
This trait represents a set of arguments accepted by CoreApplication::process_events method.

CoreApplicationPropertyArgs
This trait represents a set of arguments accepted by CoreApplication::property method.

CoreApplicationRemovePostedEventsArgs
This trait represents a set of arguments accepted by CoreApplication::remove_posted_events method.

CoreApplicationSendPostedEventsArgs
This trait represents a set of arguments accepted by CoreApplication::send_posted_events method.

CoreApplicationSetAttributeArgs
This trait represents a set of arguments accepted by CoreApplication::set_attribute method.

CoreApplicationStartTimerArgs
This trait represents a set of arguments accepted by CoreApplication::start_timer method.

CoreApplicationTrArgs
This trait represents a set of arguments accepted by CoreApplication::tr method.

CoreApplicationTrUtf8Args
This trait represents a set of arguments accepted by CoreApplication::tr_utf8 method.

CoreApplicationTranslateArgs
This trait represents a set of arguments accepted by CoreApplication::translate method.