Struct qt_core::cryptographic_hash::CryptographicHash

pub struct CryptographicHash {
    // some fields omitted
}

C++ type: QCryptographicHash

C++ documentation:

The QCryptographicHash class provides a way to generate cryptographic hashes.

QCryptographicHash can be used to generate cryptographic hashes of binary or text data.

Currently MD4, MD5, SHA-1, SHA-224, SHA-256, SHA-384, and SHA-512 are supported.

Methods

impl CryptographicHash

fn add_data<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: CryptographicHashAddDataArgs<'largs>

C++ method: QCryptographicHash::addData

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn add_data(&mut self, *mut ::qiod_evice::QiodEvice) -> bool
C++ method: bool QCryptographicHash::addData(QIODevice* device)

C++ documentation:

Reads the data from the open QIODevice device until it ends and hashes it. Returns true if reading was successful.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn add_data(&mut self, &::byte_array::ByteArray) -> ()
C++ method: void QCryptographicHash::addData(const QByteArray& data)

C++ documentation:

This function overloads addData().

Variant 3

Rust arguments: fn add_data(&mut self, (*const libc::c_char, libc::c_int)) -> ()
C++ method: void QCryptographicHash::addData(const char* data, int length)

C++ documentation:

Adds the first length chars of data to the cryptographic hash.

fn hash<Args>(args: Args) -> Args::ReturnType where Args: CryptographicHashHashArgs

C++ method: QCryptographicHash::hash

This is an overloaded function. Available variants:

Rust arguments:
1) fn hash((&::byte_array::ByteArray, ::cryptographic_hash::Algorithm, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn hash((&::byte_array::ByteArray, ::cryptographic_hash::Algorithm, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: static QByteArray QCryptographicHash::hash(const QByteArray& data, QCryptographicHash::Algorithm method)

C++ documentation:

Returns the hash of data using method.

fn new<Args>(args: Args) -> Args::ReturnType where Args: CryptographicHashNewArgs

C++ method: QCryptographicHash::QCryptographicHash

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((::cryptographic_hash::Algorithm, cpp_utils::AsStruct)) -> ::cryptographic_hash::CryptographicHash
2) fn new((::cryptographic_hash::Algorithm, cpp_utils::AsBox)) -> cpp_utils::CppBox<::cryptographic_hash::CryptographicHash>
C++ method: [constructor] void QCryptographicHash::QCryptographicHash(QCryptographicHash::Algorithm method)

fn reset(&mut self)

C++ method: void QCryptographicHash::reset()

C++ documentation:

Resets the object.

fn result<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: CryptographicHashResultArgs<'largs>

C++ method: QCryptographicHash::result

This is an overloaded function. Available variants:

Rust arguments:
1) fn result(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn result(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QCryptographicHash::result() const

C++ documentation:

Returns the final hash value.

See also QByteArray::toHex().