Struct qt_core::data_stream::DataStream

pub struct DataStream {
    // some fields omitted
}

C++ type: QDataStream

C++ documentation:

The QDataStream class provides serialization of binary data to a QIODevice.

A data stream is a binary stream of encoded information which is 100% independent of the host computer's operating system, CPU or byte order. For example, a data stream that is written by a PC under Windows can be read by a Sun SPARC running Solaris.

You can also use a data stream to read/write raw unencoded binary data. If you want a "parsing" input stream, see QTextStream.

The QDataStream class implements the serialization of C++'s basic data types, like char, short, int, char *, etc. Serialization of more complex data is accomplished by breaking up the data into primitive units.

A data stream cooperates closely with a QIODevice. A QIODevice represents an input/output medium one can read data from and write data to. The QFile class is an example of an I/O device.

Example (write binary data to a stream):

QFile file("file.dat");
file.open(QIODevice::WriteOnly);
QDataStream out(&file);   // we will serialize the data into the file
out << QString("the answer is");   // serialize a string
out << (qint32)42;        // serialize an integer

Example (read binary data from a stream):

QFile file("file.dat");
file.open(QIODevice::ReadOnly);
QDataStream in(&file);    // read the data serialized from the file
QString str;
qint32 a;
in >> str >> a;           // extract "the answer is" and 42

Each item written to the stream is written in a predefined binary format that varies depending on the item's type. Supported Qt types include QBrush, QColor, QDateTime, QFont, QPixmap, QString, QVariant and many others. For the complete list of all Qt types supporting data streaming see Serializing Qt Data Types.

For integers it is best to always cast to a Qt integer type for writing, and to read back into the same Qt integer type. This ensures that you get integers of the size you want and insulates you from compiler and platform differences.

To take one example, a char * string is written as a 32-bit integer equal to the length of the string including the '\0' byte, followed by all the characters of the string including the '\0' byte. When reading a char * string, 4 bytes are read to create the 32-bit length value, then that many characters for the char * string including the '\0' terminator are read.

The initial I/O device is usually set in the constructor, but can be changed with setDevice(). If you've reached the end of the data (or if there is no I/O device set) atEnd() will return true.

Methods

impl DataStream

fn at_end(&self) -> bool

C++ method: bool QDataStream::atEnd() const

C++ documentation:

Returns true if the I/O device has reached the end position (end of the stream or file) or if there is no I/O device set; otherwise returns false.

See also QIODevice::atEnd().

fn byte_order(&self) -> ByteOrder

C++ method: QDataStream::ByteOrder QDataStream::byteOrder() const

C++ documentation:

Returns the current byte order setting -- either BigEndian or LittleEndian.

See also setByteOrder().

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QDataStream::device() const

C++ documentation:

Returns the I/O device currently set, or 0 if no device is currently set.

See also setDevice().

fn floating_point_precision(&self) -> FloatingPointPrecision

C++ method: QDataStream::FloatingPointPrecision QDataStream::floatingPointPrecision() const

C++ documentation:

Returns the floating point precision of the data stream.

This function was introduced in Qt 4.6.

See also FloatingPointPrecision and setFloatingPointPrecision().

fn new<Args>(args: Args) -> Args::ReturnType where Args: DataStreamNewArgs

C++ method: QDataStream::QDataStream

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::data_stream::DataStream
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::data_stream::DataStream>
C++ method: [constructor] void QDataStream::QDataStream()

Variant 2

Rust arguments:
1) fn new((*mut ::byte_array::ByteArray, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsStruct)) -> ::data_stream::DataStream
2) fn new((*mut ::byte_array::ByteArray, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::data_stream::DataStream>
C++ method: [constructor] void QDataStream::QDataStream(QByteArray* arg1, QFlags<QIODevice::OpenModeFlag> flags)

Variant 3

Rust arguments:
1) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::data_stream::DataStream
2) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::data_stream::DataStream>
C++ method: [constructor] void QDataStream::QDataStream(QIODevice* arg1)

Variant 4

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::data_stream::DataStream
2) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::data_stream::DataStream>
C++ method: [constructor] void QDataStream::QDataStream(const QByteArray& arg1)

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut DataStream where Args: DataStreamOpShlArgs<'largs>

C++ method: QDataStream::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, bool) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(bool i)

C++ documentation:

Writes a boolean value, i, to the stream. Returns a reference to the stream.

Variant 2

Rust arguments: fn op_shl(&mut self, *const libc::c_char) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(const char* str)

C++ documentation:

This is an overloaded function.

Writes the '\0'-terminated string s to the stream and returns a reference to the stream.

The string is serialized using writeBytes().

See also writeBytes() and writeRawData().

Variant 3

Rust arguments: fn op_shl(&mut self, libc::c_double) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(double f)

C++ documentation:

This is an overloaded function.

Writes a floating point number, f, to the stream using the standard IEEE 754 format. Returns a reference to the stream.

See also setFloatingPointPrecision().

Variant 4

Rust arguments: fn op_shl(&mut self, libc::c_float) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(float f)

C++ documentation:

This is an overloaded function.

Writes a floating point number, f, to the stream using the standard IEEE 754 format. Returns a reference to the stream.

See also setFloatingPointPrecision().

Variant 5

Rust arguments: fn op_shl(&mut self, i16) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(qint16 i)

C++ documentation:

This is an overloaded function.

Writes a signed 16-bit integer, i, to the stream and returns a reference to the stream.

Variant 6

Rust arguments: fn op_shl(&mut self, i32) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(qint32 i)

C++ documentation:

This is an overloaded function.

Writes a signed 32-bit integer, i, to the stream and returns a reference to the stream.

Variant 7

Rust arguments: fn op_shl(&mut self, i64) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(qint64 i)

C++ documentation:

This is an overloaded function.

Writes a signed 64-bit integer, i, to the stream and returns a reference to the stream.

Variant 8

Rust arguments: fn op_shl(&mut self, i8) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(qint8 i)

C++ documentation:

Writes a signed byte, i, to the stream and returns a reference to the stream.

Variant 9

Rust arguments: fn op_shl(&mut self, u16) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(quint16 i)

C++ documentation:

This is an overloaded function.

Writes an unsigned 16-bit integer, i, to the stream and returns a reference to the stream.

Variant 10

Rust arguments: fn op_shl(&mut self, u32) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(quint32 i)

C++ documentation:

This is an overloaded function.

Writes an unsigned integer, i, to the stream as a 32-bit unsigned integer (quint32). Returns a reference to the stream.

Variant 11

Rust arguments: fn op_shl(&mut self, u64) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(quint64 i)

C++ documentation:

This is an overloaded function.

Writes an unsigned 64-bit integer, i, to the stream and returns a reference to the stream.

Variant 12

Rust arguments: fn op_shl(&mut self, u8) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator<<(quint8 i)

C++ documentation:

This is an overloaded function.

Writes an unsigned byte, i, to the stream and returns a reference to the stream.

fn op_shr<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut DataStream where Args: DataStreamOpShrArgs<'largs>

C++ method: QDataStream::operator>>

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shr(&mut self, &'l1 mut bool) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(bool& i)

C++ documentation:

Reads a boolean value from the stream into i. Returns a reference to the stream.

Variant 2

Rust arguments: fn op_shr(&mut self, &'l1 mut *mut libc::c_char) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(char*& str)

C++ documentation:

This is an overloaded function.

Reads the '\0'-terminated string s from the stream and returns a reference to the stream.

The string is deserialized using readBytes().

Space for the string is allocated using new [] -- the caller must destroy it with delete [].

See also readBytes() and readRawData().

Variant 3

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_double) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(double& f)

C++ documentation:

This is an overloaded function.

Reads a floating point number from the stream into f, using the standard IEEE 754 format. Returns a reference to the stream.

See also setFloatingPointPrecision().

Variant 4

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_float) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(float& f)

C++ documentation:

This is an overloaded function.

Reads a floating point number from the stream into f, using the standard IEEE 754 format. Returns a reference to the stream.

See also setFloatingPointPrecision().

Variant 5

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_int) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(int& i)

C++ documentation:

This is an overloaded function.

Reads a signed 32-bit integer from the stream into i, and returns a reference to the stream.

Variant 6

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_longlong) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(long long& i)

C++ documentation:

This is an overloaded function.

Reads a signed 64-bit integer from the stream into i, and returns a reference to the stream.

Variant 7

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_short) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(short& i)

C++ documentation:

This is an overloaded function.

Reads a signed 16-bit integer from the stream into i, and returns a reference to the stream.

Variant 8

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_schar) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(signed char& i)

C++ documentation:

Reads a signed byte from the stream into i, and returns a reference to the stream.

Variant 9

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_uchar) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(unsigned char& i)

C++ documentation:

This is an overloaded function.

Reads an unsigned byte from the stream into i, and returns a reference to the stream.

Variant 10

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_uint) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(unsigned int& i)

C++ documentation:

This is an overloaded function.

Reads an unsigned 32-bit integer from the stream into i, and returns a reference to the stream.

Variant 11

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_ulonglong) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(unsigned long long& i)

C++ documentation:

This is an overloaded function.

Reads an unsigned 64-bit integer from the stream, into i, and returns a reference to the stream.

Variant 12

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_ushort) -> &'l0 mut ::data_stream::DataStream
C++ method: QDataStream& QDataStream::operator>>(unsigned short& i)

C++ documentation:

This is an overloaded function.

Reads an unsigned 16-bit integer from the stream into i, and returns a reference to the stream.

fn read_bytes<'l0, 'l1, 'l2>(&'l0 mut self, arg1: &'l1 mut *mut c_char, len: &'l2 mut c_uint) -> &'l0 mut DataStream

C++ method: QDataStream& QDataStream::readBytes(char*& arg1, unsigned int& len)

C++ documentation:

Reads the buffer s from the stream and returns a reference to the stream.

The buffer s is allocated using new []. Destroy it with the delete [] operator.

The l parameter is set to the length of the buffer. If the string read is empty, l is set to 0 and s is set to a null pointer.

The serialization format is a quint32 length specifier first, then l bytes of data.

See also readRawData() and writeBytes().

fn read_raw_data(&mut self, arg1: *mut c_char, len: c_int) -> c_int

C++ method: int QDataStream::readRawData(char* arg1, int len)

C++ documentation:

Reads at most len bytes from the stream into s and returns the number of bytes read. If an error occurs, this function returns -1.

The buffer s must be preallocated. The data is not encoded.

See also readBytes(), QIODevice::read(), and writeRawData().

fn reset_status(&mut self)

C++ method: void QDataStream::resetStatus()

C++ documentation:

Resets the status of the data stream.

See also Status, status(), and setStatus().

fn set_byte_order(&mut self, arg1: ByteOrder)

C++ method: void QDataStream::setByteOrder(QDataStream::ByteOrder arg1)

C++ documentation:

Sets the serialization byte order to bo.

The bo parameter can be QDataStream::BigEndian or QDataStream::LittleEndian.

The default setting is big endian. We recommend leaving this setting unless you have special requirements.

See also byteOrder().

fn set_device(&mut self, arg1: *mut QiodEvice)

C++ method: void QDataStream::setDevice(QIODevice* arg1)

C++ documentation:

void QDataStream::setDevice(QIODevice *d)

Sets the I/O device to d, which can be 0 to unset to current I/O device.

See also device().

fn set_floating_point_precision(&mut self, precision: FloatingPointPrecision)

C++ method: void QDataStream::setFloatingPointPrecision(QDataStream::FloatingPointPrecision precision)

C++ documentation:

Sets the floating point precision of the data stream to precision. If the floating point precision is DoublePrecision and the version of the data stream is Qt_4_6 or higher, all floating point numbers will be written and read with 64-bit precision. If the floating point precision is SinglePrecision and the version is Qt_4_6 or higher, all floating point numbers will be written and read with 32-bit precision.

For versions prior to Qt_4_6, the precision of floating point numbers in the data stream depends on the stream operator called.

The default is DoublePrecision.

Warning: This property must be set to the same value on the object that writes and the object that reads the data stream.

This function was introduced in Qt 4.6.

See also floatingPointPrecision().

fn set_status(&mut self, status: Status)

C++ method: void QDataStream::setStatus(QDataStream::Status status)

C++ documentation:

Sets the status of the data stream to the status given.

Subsequent calls to setStatus() are ignored until resetStatus() is called.

See also Status, status(), and resetStatus().

fn set_version(&mut self, arg1: c_int)

C++ method: void QDataStream::setVersion(int arg1)

C++ documentation:

Sets the version number of the data serialization format to v, a value of the Version enum.

You don't have to set a version if you are using the current version of Qt, but for your own custom binary formats we recommend that you do; see Versioning in the Detailed Description.

To accommodate new functionality, the datastream serialization format of some Qt classes has changed in some versions of Qt. If you want to read data that was created by an earlier version of Qt, or write data that can be read by a program that was compiled with an earlier version of Qt, use this function to modify the serialization format used by QDataStream.

The Version enum provides symbolic constants for the different versions of Qt. For example:

QDataStream out(file);
out.setVersion(QDataStream::Qt_4_0);

See also version() and Version.

fn skip_raw_data(&mut self, len: c_int) -> c_int

C++ method: int QDataStream::skipRawData(int len)

C++ documentation:

Skips len bytes from the device. Returns the number of bytes actually skipped, or -1 on error.

This is equivalent to calling readRawData() on a buffer of length len and ignoring the buffer.

This function was introduced in Qt 4.1.

See also QIODevice::seek().

fn status(&self) -> Status

C++ method: QDataStream::Status QDataStream::status() const

C++ documentation:

Returns the status of the data stream.

See also Status, setStatus(), and resetStatus().

fn unset_device(&mut self)

C++ method: void QDataStream::unsetDevice()

C++ documentation:

Unsets the I/O device. Use setDevice(0) instead.

fn version(&self) -> c_int

C++ method: int QDataStream::version() const

C++ documentation:

Returns the version number of the data serialization format.

See also setVersion() and Version.

fn write_bytes<'l0>(&'l0 mut self, arg1: *const c_char, len: c_uint) -> &'l0 mut DataStream

C++ method: QDataStream& QDataStream::writeBytes(const char* arg1, unsigned int len)

C++ documentation:

Writes the length specifier len and the buffer s to the stream and returns a reference to the stream.

The len is serialized as a quint32, followed by len bytes from s. Note that the data is not encoded.

See also writeRawData() and readBytes().

fn write_raw_data(&mut self, arg1: *const c_char, len: c_int) -> c_int

C++ method: int QDataStream::writeRawData(const char* arg1, int len)

C++ documentation:

Writes len bytes from s to the stream. Returns the number of bytes actually written, or -1 on error. The data is not encoded.

See also writeBytes(), QIODevice::write(), and readRawData().