Struct qt_core::date::Date

pub struct Date {
    // some fields omitted
}

C++ type: QDate

C++ documentation:

The QDate class provides date functions.

A QDate object contains a calendar date, i.e. year, month, and day numbers, in the Gregorian calendar. It can read the current date from the system clock. It provides functions for comparing dates, and for manipulating dates. For example, it is possible to add and subtract days, months, and years to dates.

A QDate object is typically created by giving the year, month, and day numbers explicitly. Note that QDate interprets two digit years as is, i.e., years 0 - 99. A QDate can also be constructed with the static function currentDate(), which creates a QDate object containing the system clock's date. An explicit date can also be set using setDate(). The fromString() function returns a QDate given a string and a date format which is used to interpret the date within the string.

The year(), month(), and day() functions provide access to the year, month, and day numbers. Also, dayOfWeek() and dayOfYear() functions are provided. The same information is provided in textual format by the toString(), shortDayName(), longDayName(), shortMonthName(), and longMonthName() functions.

QDate provides a full set of operators to compare two QDate objects where smaller means earlier, and larger means later.

You can increment (or decrement) a date by a given number of days using addDays(). Similarly you can use addMonths() and addYears(). The daysTo() function returns the number of days between two dates.

The daysInMonth() and daysInYear() functions return how many days there are in this date's month and year, respectively. The isLeapYear() function indicates whether a date is in a leap year.

No Year 0

There is no year 0. Dates in that year are considered invalid. The year -1 is the year "1 before Christ" or "1 before current era." The day before 1 January 1 CE is 31 December 1 BCE.

Range of Valid Dates

Dates are stored internally as a Julian Day number, an integer count of every day in a contiguous range, with 24 November 4714 BCE in the Gregorian calendar being Julian Day 0 (1 January 4713 BCE in the Julian calendar). As well as being an efficient and accurate way of storing an absolute date, it is suitable for converting a Date into other calendar systems such as Hebrew, Islamic or Chinese. The Julian Day number can be obtained using QDate::toJulianDay() and can be set using QDate::fromJulianDay().

The range of dates able to be stored by QDate as a Julian Day number is for technical reasons limited to between -784350574879 and 784354017364, which means from before 2 billion BCE to after 2 billion CE.

Methods

impl Date

fn add_days<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateAddDaysArgs<'largs>

C++ method: QDate::addDays

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_days(&self, (i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
2) fn add_days(&self, (i64, cpp_utils::AsStruct)) -> ::date::Date
C++ method: QDate QDate::addDays(qint64 days) const

C++ documentation:

Returns a QDate object containing a date ndays later than the date of this object (or earlier if ndays is negative).

Returns a null date if the current date is invalid or the new date is out of range.

See also addMonths(), addYears(), and daysTo().

fn add_months<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateAddMonthsArgs<'largs>

C++ method: QDate::addMonths

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_months(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
2) fn add_months(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::date::Date
C++ method: QDate QDate::addMonths(int months) const

C++ documentation:

Returns a QDate object containing a date nmonths later than the date of this object (or earlier if nmonths is negative).

Note: If the ending day/month combination does not exist in the resulting month/year, this function will return a date that is the latest valid date.

See also addDays() and addYears().

fn add_years<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateAddYearsArgs<'largs>

C++ method: QDate::addYears

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_years(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
2) fn add_years(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::date::Date
C++ method: QDate QDate::addYears(int years) const

C++ documentation:

Returns a QDate object containing a date nyears later than the date of this object (or earlier if nyears is negative).

Note: If the ending day/month combination does not exist in the resulting year (i.e., if the date was Feb 29 and the final year is not a leap year), this function will return a date that is the latest valid date (that is, Feb 28).

See also addDays() and addMonths().

fn current_date<Args>(args: Args) -> Args::ReturnType where Args: DateCurrentDateArgs

C++ method: QDate::currentDate

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_date(cpp_utils::AsBox) -> cpp_utils::CppBox<::date::Date>
2) fn current_date(cpp_utils::AsStruct) -> ::date::Date
C++ method: static QDate QDate::currentDate()

C++ documentation:

Returns the current date, as reported by the system clock.

See also QTime::currentTime() and QDateTime::currentDateTime().

fn day(&self) -> c_int

C++ method: int QDate::day() const

C++ documentation:

Returns the day of the month (1 to 31) of this date.

Returns 0 if the date is invalid.

See also year(), month(), and dayOfWeek().

fn day_of_week(&self) -> c_int

C++ method: int QDate::dayOfWeek() const

C++ documentation:

Returns the weekday (1 = Monday to 7 = Sunday) for this date.

Returns 0 if the date is invalid.

See also day(), dayOfYear(), and Qt::DayOfWeek.

fn day_of_year(&self) -> c_int

C++ method: int QDate::dayOfYear() const

C++ documentation:

Returns the day of the year (1 to 365 or 366 on leap years) for this date.

Returns 0 if the date is invalid.

See also day() and dayOfWeek().

fn days_in_month(&self) -> c_int

C++ method: int QDate::daysInMonth() const

C++ documentation:

Returns the number of days in the month (28 to 31) for this date.

Returns 0 if the date is invalid.

See also day() and daysInYear().

fn days_in_year(&self) -> c_int

C++ method: int QDate::daysInYear() const

C++ documentation:

Returns the number of days in the year (365 or 366) for this date.

Returns 0 if the date is invalid.

See also day() and daysInMonth().

fn days_to(&self, arg1: &Date) -> i64

C++ method: qint64 QDate::daysTo(const QDate& arg1) const

C++ documentation:

Returns the number of days from this date to d (which is negative if d is earlier than this date).

Returns 0 if either date is invalid.

Example:

QDate d1(1995, 5, 17);  // May 17, 1995
QDate d2(1995, 5, 20);  // May 20, 1995
d1.daysTo(d2);          // returns 3
d2.daysTo(d1);          // returns -3

See also addDays().

fn from_julian_day<Args>(args: Args) -> Args::ReturnType where Args: DateFromJulianDayArgs

C++ method: QDate::fromJulianDay

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_julian_day((i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
2) fn from_julian_day((i64, cpp_utils::AsStruct)) -> ::date::Date
C++ method: static QDate QDate::fromJulianDay(qint64 jd)

C++ documentation:

Converts the Julian day jd to a QDate.

See also toJulianDay().

fn from_string<Args>(args: Args) -> Args::ReturnType where Args: DateFromStringArgs

C++ method: QDate::fromString

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_string((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
2) fn from_string((&::string::String, ::qt::DateFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
3) fn from_string((&::string::String, cpp_utils::AsStruct)) -> ::date::Date
4) fn from_string((&::string::String, ::qt::DateFormat, cpp_utils::AsStruct)) -> ::date::Date
C++ method: static QDate QDate::fromString(const QString& s, Qt::DateFormat f = ?)

C++ documentation:

Returns the QDate represented by the string, using the format given, or an invalid date if the string cannot be parsed.

Note for Qt::TextDate: It is recommended that you use the English short month names (e.g. "Jan"). Although localized month names can also be used, they depend on the user's locale settings.

Variant 2

Rust arguments:
1) fn from_string((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
2) fn from_string((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::date::Date
C++ method: static QDate QDate::fromString(const QString& s, const QString& format)

C++ documentation:

Returns the QDate represented by the string, using the format given, or an invalid date if the string cannot be parsed.

These expressions may be used for the format:

Expression	Output

d	The day as a number without a leading zero (1 to 31)

dd	The day as a number with a leading zero (01 to 31)

ddd	The abbreviated localized day name (e.g. 'Mon' to 'Sun'). Uses the system locale to localize the name, i.e. QLocale::system().

dddd	The long localized day name (e.g. 'Monday' to 'Sunday'). Uses the system locale to localize the name, i.e. QLocale::system().

M	The month as a number without a leading zero (1 to 12)

MM	The month as a number with a leading zero (01 to 12)

MMM	The abbreviated localized month name (e.g. 'Jan' to 'Dec'). Uses the system locale to localize the name, i.e. QLocale::system().

MMMM	The long localized month name (e.g. 'January' to 'December'). Uses the system locale to localize the name, i.e. QLocale::system().

yy	The year as two digit number (00 to 99)

yyyy	The year as four digit number. If the year is negative, a minus sign is prepended in addition.

All other input characters will be treated as text. Any sequence of characters that are enclosed in single quotes will also be treated as text and will not be used as an expression. For example:

QDate date = QDate::fromString("1MM12car2003", "d'MM'MMcaryyyy");
// date is 1 December 2003

If the format is not satisfied, an invalid QDate is returned. The expressions that don't expect leading zeroes (d, M) will be greedy. This means that they will use two digits even if this will put them outside the accepted range of values and leaves too few digits for other sections. For example, the following format string could have meant January 30 but the M will grab two digits, resulting in an invalid date:

QDate date = QDate::fromString("130", "Md"); // invalid

For any field that is not represented in the format the following defaults are used:

Field	Default value

Year	1900

Month	1

Day	1

The following examples demonstrate the default values:

QDate::fromString("1.30", "M.d");           // January 30 1900
QDate::fromString("20000110", "yyyyMMdd");  // January 10, 2000
QDate::fromString("20000110", "yyyyMd");    // January 10, 2000

See also QDateTime::fromString(), QTime::fromString(), QDate::toString(), QDateTime::toString(), and QTime::toString().

fn get_date(&mut self, year: *mut c_int, month: *mut c_int, day: *mut c_int)

C++ method: void QDate::getDate(int* year, int* month, int* day)

C++ documentation:

Extracts the date's year, month, and day, and assigns them to *year, *month, and *day. The pointers may be null.

Returns 0 if the date is invalid.

This function was introduced in Qt 4.5.

See also year(), month(), day(), and isValid().

fn is_leap_year(year: c_int) -> bool

C++ method: static bool QDate::isLeapYear(int year)

C++ documentation:

Returns true if the specified year is a leap year; otherwise returns false.

fn is_null(&self) -> bool

C++ method: bool QDate::isNull() const

C++ documentation:

Returns true if the date is null; otherwise returns false. A null date is invalid.

Note: The behavior of this function is equivalent to isValid().

See also isValid().

fn is_valid(&self) -> bool

C++ method: bool QDate::isValid() const

C++ documentation:

Returns true if this date is valid; otherwise returns false.

See also isNull().

fn is_valid_static(y: c_int, m: c_int, d: c_int) -> bool

C++ method: static bool QDate::isValid(int y, int m, int d)

C++ documentation:

This is an overloaded function.

Returns true if the specified date (year, month, and day) is valid; otherwise returns false.

Example:

QDate::isValid(2002, 5, 17);  // true
QDate::isValid(2002, 2, 30);  // false (Feb 30 does not exist)
QDate::isValid(2004, 2, 29);  // true (2004 is a leap year)
QDate::isValid(2000, 2, 29);  // true (2000 is a leap year)
QDate::isValid(2006, 2, 29);  // false (2006 is not a leap year)
QDate::isValid(2100, 2, 29);  // false (2100 is not a leap year)
QDate::isValid(1202, 6, 6);   // true (even though 1202 is pre-Gregorian)

See also isNull() and setDate().

fn long_day_name<Args>(args: Args) -> Args::ReturnType where Args: DateLongDayNameArgs

C++ method: QDate::longDayName

This is an overloaded function. Available variants:

Rust arguments:
1) fn long_day_name((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn long_day_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn long_day_name((libc::c_int, cpp_utils::AsStruct)) -> ::string::String
4) fn long_day_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QDate::longDayName(int weekday, QDate::MonthNameType type = ?)

C++ documentation:

Returns the long name of the weekday for the representation specified by type.

The days are enumerated using the following convention:

1 = "Monday"

2 = "Tuesday"

3 = "Wednesday"

4 = "Thursday"

5 = "Friday"

6 = "Saturday"

7 = "Sunday"

The day names will be localized according to the system's locale settings, i.e. using QLocale::system().

Returns an empty string if the date is invalid.

This function was introduced in Qt 4.5.

See also toString(), shortDayName(), shortMonthName(), and longMonthName().

fn long_month_name<Args>(args: Args) -> Args::ReturnType where Args: DateLongMonthNameArgs

C++ method: QDate::longMonthName

This is an overloaded function. Available variants:

Rust arguments:
1) fn long_month_name((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn long_month_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn long_month_name((libc::c_int, cpp_utils::AsStruct)) -> ::string::String
4) fn long_month_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QDate::longMonthName(int month, QDate::MonthNameType type = ?)

C++ documentation:

Returns the long name of the month for the representation specified by type.

The months are enumerated using the following convention:

1 = "January"

2 = "February"

3 = "March"

4 = "April"

5 = "May"

6 = "June"

7 = "July"

8 = "August"

9 = "September"

10 = "October"

11 = "November"

12 = "December"

The month names will be localized according to the system's locale settings, i.e. using QLocale::system().

Returns an empty string if the date is invalid.

This function was introduced in Qt 4.5.

See also toString(), shortMonthName(), shortDayName(), and longDayName().

fn month(&self) -> c_int

C++ method: int QDate::month() const

C++ documentation:

Returns the number corresponding to the month of this date, using the following convention:

1 = "January"

2 = "February"

3 = "March"

4 = "April"

5 = "May"

6 = "June"

7 = "July"

8 = "August"

9 = "September"

10 = "October"

11 = "November"

12 = "December"

Returns 0 if the date is invalid.

See also year() and day().

fn new<Args>(args: Args) -> Args::ReturnType where Args: DateNewArgs

C++ method: QDate::QDate

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::date::Date
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::date::Date>
C++ method: [constructor] void QDate::QDate()

Variant 2

Rust arguments:
1) fn new((libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::date::Date
2) fn new((libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date::Date>
C++ method: [constructor] void QDate::QDate(int y, int m, int d)

fn op_eq(&self, other: &Date) -> bool

C++ method: bool QDate::operator==(const QDate& other) const

C++ documentation:

Returns true if this date is equal to d; otherwise returns false.

fn op_ge(&self, other: &Date) -> bool

C++ method: bool QDate::operator>=(const QDate& other) const

C++ documentation:

Returns true if this date is later than or equal to d; otherwise returns false.

fn op_gt(&self, other: &Date) -> bool

C++ method: bool QDate::operator>(const QDate& other) const

C++ documentation:

Returns true if this date is later than d; otherwise returns false.

fn op_le(&self, other: &Date) -> bool

C++ method: bool QDate::operator<=(const QDate& other) const

C++ documentation:

Returns true if this date is earlier than or equal to d; otherwise returns false.

fn op_lt(&self, other: &Date) -> bool

C++ method: bool QDate::operator<(const QDate& other) const

C++ documentation:

Returns true if this date is earlier than d; otherwise returns false.

fn op_neq(&self, other: &Date) -> bool

C++ method: bool QDate::operator!=(const QDate& other) const

C++ documentation:

Returns true if this date is different from d; otherwise returns false.

fn set_date(&mut self, year: c_int, month: c_int, day: c_int) -> bool

C++ method: bool QDate::setDate(int year, int month, int day)

C++ documentation:

Sets the date's year, month, and day. Returns true if the date is valid; otherwise returns false.

If the specified date is invalid, the QDate object is set to be invalid.

This function was introduced in Qt 4.2.

See also isValid().

fn short_day_name<Args>(args: Args) -> Args::ReturnType where Args: DateShortDayNameArgs

C++ method: QDate::shortDayName

This is an overloaded function. Available variants:

Rust arguments:
1) fn short_day_name((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn short_day_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn short_day_name((libc::c_int, cpp_utils::AsStruct)) -> ::string::String
4) fn short_day_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QDate::shortDayName(int weekday, QDate::MonthNameType type = ?)

C++ documentation:

Returns the short name of the weekday for the representation specified by type.

The days are enumerated using the following convention:

1 = "Mon"

2 = "Tue"

3 = "Wed"

4 = "Thu"

5 = "Fri"

6 = "Sat"

7 = "Sun"

The day names will be localized according to the system's locale settings, i.e. using QLocale::system().

Returns an empty string if the date is invalid.

This function was introduced in Qt 4.5.

See also toString(), shortMonthName(), longMonthName(), and longDayName().

fn short_month_name<Args>(args: Args) -> Args::ReturnType where Args: DateShortMonthNameArgs

C++ method: QDate::shortMonthName

This is an overloaded function. Available variants:

Rust arguments:
1) fn short_month_name((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn short_month_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn short_month_name((libc::c_int, cpp_utils::AsStruct)) -> ::string::String
4) fn short_month_name((libc::c_int, ::date::MonthNameType, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QDate::shortMonthName(int month, QDate::MonthNameType type = ?)

C++ documentation:

Returns the short name of the month for the representation specified by type.

The months are enumerated using the following convention:

1 = "Jan"

2 = "Feb"

3 = "Mar"

4 = "Apr"

5 = "May"

6 = "Jun"

7 = "Jul"

8 = "Aug"

9 = "Sep"

10 = "Oct"

11 = "Nov"

12 = "Dec"

The month names will be localized according to the system's locale settings, i.e. using QLocale::system().

Returns an empty string if the date is invalid.

This function was introduced in Qt 4.5.

See also toString(), longMonthName(), shortDayName(), and longDayName().

fn to_julian_day(&self) -> i64

C++ method: qint64 QDate::toJulianDay() const

C++ documentation:

Converts the date to a Julian day.

See also fromJulianDay().

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateToStringArgs<'largs>

C++ method: QDate::toString

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_string(&self, (::qt::DateFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
3) fn to_string(&self, (::qt::DateFormat, cpp_utils::AsStruct)) -> ::string::String
4) fn to_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDate::toString(Qt::DateFormat f = ?) const

C++ documentation:

This is an overloaded function.

Returns the date as a string. The format parameter determines the format of the string.

If the format is Qt::TextDate, the string is formatted in the default way. QDate::shortDayName() and QDate::shortMonthName() are used to generate the string, so the day and month names will be localized names using the system locale, i.e. QLocale::system(). An example of this formatting is "Sat May 20 1995".

If the format is Qt::ISODate, the string format corresponds to the ISO 8601 extended specification for representations of dates and times, taking the form YYYY-MM-DD, where YYYY is the year, MM is the month of the year (between 01 and 12), and DD is the day of the month between 01 and 31.

If the format is Qt::SystemLocaleShortDate or Qt::SystemLocaleLongDate, the string format depends on the locale settings of the system. Identical to calling QLocale::system().toString(date, QLocale::ShortFormat) or QLocale::system().toString(date, QLocale::LongFormat).

If the format is Qt::DefaultLocaleShortDate or Qt::DefaultLocaleLongDate, the string format depends on the default application locale. This is the locale set with QLocale::setDefault(), or the system locale if no default locale has been set. Identical to calling QLocale().toString(date, QLocale::ShortFormat) or QLocale().toString(date, QLocale::LongFormat).

If the format is Qt::RFC2822Date, the string is formatted in an RFC 2822 compatible way. An example of this formatting is "20 May 1995".

If the date is invalid, an empty string will be returned.

Warning: The Qt::ISODate format is only valid for years in the range 0 to 9999. This restriction may apply to locale-aware formats as well, depending on the locale settings.

See also shortDayName() and shortMonthName().

Variant 2

Rust arguments:
1) fn to_string(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QDate::toString(const QString& format) const

C++ documentation:

Returns the date as a string. The format parameter determines the format of the result string.

These expressions may be used:

Expression	Output

d	the day as number without a leading zero (1 to 31)

dd	the day as number with a leading zero (01 to 31)

ddd	the abbreviated localized day name (e.g. 'Mon' to 'Sun'). Uses the system locale to localize the name, i.e. QLocale::system().

dddd	the long localized day name (e.g. 'Monday' to 'Sunday'). Uses the system locale to localize the name, i.e. QLocale::system().

M	the month as number without a leading zero (1 to 12)

MM	the month as number with a leading zero (01 to 12)

MMM	the abbreviated localized month name (e.g. 'Jan' to 'Dec'). Uses the system locale to localize the name, i.e. QLocale::system().

MMMM	the long localized month name (e.g. 'January' to 'December'). Uses the system locale to localize the name, i.e. QLocale::system().

yy	the year as two digit number (00 to 99)

yyyy	the year as four digit number. If the year is negative, a minus sign is prepended in addition.

All other input characters will be ignored. Any sequence of characters that are enclosed in single quotes will be treated as text and not be used as an expression. Two consecutive single quotes ("''") are replaced by a singlequote in the output. Formats without separators (e.g. "ddMM") are currently not supported.

Example format strings (assuming that the QDate is the 20 July 1969):

Format	Result

dd.MM.yyyy	20.07.1969

ddd MMMM d yy	Sun July 20 69

'The day is' dddd	The day is Sunday

If the datetime is invalid, an empty string will be returned.

See also QDateTime::toString(), QTime::toString(), and QLocale::toString().

fn week_number<'largs, Args>(&'largs self, args: Args) -> c_int where Args: DateWeekNumberArgs<'largs>

C++ method: QDate::weekNumber

This is an overloaded function. Available variants:

Rust arguments:
1) fn week_number(&self, ()) -> libc::c_int
2) fn week_number(&self, *mut libc::c_int) -> libc::c_int
C++ method: int QDate::weekNumber(int* yearNum = ?) const

C++ documentation:

Returns the week number (1 to 53), and stores the year in yearNumber unless yearNumber is null (the default).

Returns 0 if the date is invalid.

In accordance with ISO 8601, weeks start on Monday and the first Thursday of a year is always in week 1 of that year. Most years have 52 weeks, but some have 53.

yearNumber is not always the same as year(). For example, 1 January 2000 has week number 52 in the year 1999, and 31 December 2002 has week number 1 in the year 2003.

See also isValid().

fn year(&self) -> c_int

C++ method: int QDate::year() const

C++ documentation:

Returns the year of this date. Negative numbers indicate years before 1 CE, such that year -44 is 44 BCE.

Returns 0 if the date is invalid.

See also month() and day().