Function qt_core::date_time::hash

pub fn hash<Args>(args: Args) -> c_uint where Args: HashArgs

C++ method: qHash

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn hash(&::date::Date) -> libc::c_uint
2) fn hash((&::date::Date, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QDate& key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments:
1) fn hash(&::date_time::DateTime) -> libc::c_uint
2) fn hash((&::date_time::DateTime, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QDateTime& key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 3

Rust arguments:
1) fn hash(&::time::Time) -> libc::c_uint
2) fn hash((&::time::Time, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QTime& key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.