Module qt_core::date_time::overloading

Traits

DateTimeAddDaysArgs
This trait represents a set of arguments accepted by DateTime::add_days method.

DateTimeAddMSecsArgs
This trait represents a set of arguments accepted by DateTime::add_m_secs method.

DateTimeAddMonthsArgs
This trait represents a set of arguments accepted by DateTime::add_months method.

DateTimeAddSecsArgs
This trait represents a set of arguments accepted by DateTime::add_secs method.

DateTimeAddYearsArgs
This trait represents a set of arguments accepted by DateTime::add_years method.

DateTimeCurrentDateTimeArgs
This trait represents a set of arguments accepted by DateTime::current_date_time method.

DateTimeCurrentDateTimeUtcArgs
This trait represents a set of arguments accepted by DateTime::current_date_time_utc method.

DateTimeDateArgs
This trait represents a set of arguments accepted by DateTime::date method.

DateTimeFromMSecsSinceEpochArgs
This trait represents a set of arguments accepted by DateTime::from_m_secs_since_epoch method.

DateTimeFromStringArgs
This trait represents a set of arguments accepted by DateTime::from_string method.

DateTimeFromTimeTArgs
This trait represents a set of arguments accepted by DateTime::from_time_t method.

DateTimeNewArgs
This trait represents a set of arguments accepted by DateTime::new method.

DateTimeTimeArgs
This trait represents a set of arguments accepted by DateTime::time method.

DateTimeTimeZoneAbbreviationArgs
This trait represents a set of arguments accepted by DateTime::time_zone_abbreviation method.

DateTimeTimeZoneArgs
This trait represents a set of arguments accepted by DateTime::time_zone method.

DateTimeToLocalTimeArgs
This trait represents a set of arguments accepted by DateTime::to_local_time method.

DateTimeToOffsetFromUtcArgs
This trait represents a set of arguments accepted by DateTime::to_offset_from_utc method.

DateTimeToStringArgs
This trait represents a set of arguments accepted by DateTime::to_string method.

DateTimeToTimeSpecArgs
This trait represents a set of arguments accepted by DateTime::to_time_spec method.

DateTimeToTimeZoneArgs
This trait represents a set of arguments accepted by DateTime::to_time_zone method.

DateTimeToUtcArgs
This trait represents a set of arguments accepted by DateTime::to_utc method.

HashArgs
This trait represents a set of arguments accepted by hash method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.