Struct qt_core::date_time::DateTime

pub struct DateTime {
    // some fields omitted
}

C++ type: QDateTime

C++ documentation:

The QDateTime class provides date and time functions.

A QDateTime object contains a calendar date and a clock time (a "datetime"). It is a combination of the QDate and QTime classes. It can read the current datetime from the system clock. It provides functions for comparing datetimes and for manipulating a datetime by adding a number of seconds, days, months, or years.

A QDateTime object is typically created either by giving a date and time explicitly in the constructor, or by using the static function currentDateTime() that returns a QDateTime object set to the system clock's time. The date and time can be changed with setDate() and setTime(). A datetime can also be set using the setTime_t() function that takes a POSIX-standard "number of seconds since 00:00:00 on January 1, 1970" value. The fromString() function returns a QDateTime, given a string and a date format used to interpret the date within the string.

The date() and time() functions provide access to the date and time parts of the datetime. The same information is provided in textual format by the toString() function.

QDateTime provides a full set of operators to compare two QDateTime objects, where smaller means earlier and larger means later.

You can increment (or decrement) a datetime by a given number of milliseconds using addMSecs(), seconds using addSecs(), or days using addDays(). Similarly, you can use addMonths() and addYears(). The daysTo() function returns the number of days between two datetimes, secsTo() returns the number of seconds between two datetimes, and msecsTo() returns the number of milliseconds between two datetimes.

QDateTime can store datetimes as local time or as UTC. QDateTime::currentDateTime() returns a QDateTime expressed as local time; use toUTC() to convert it to UTC. You can also use timeSpec() to find out if a QDateTime object stores a UTC time or a local time. Operations such as addSecs() and secsTo() are aware of daylight saving time (DST).

Note: QDateTime does not account for leap seconds.

No Year 0

There is no year 0. Dates in that year are considered invalid. The year -1 is the year "1 before Christ" or "1 before current era." The day before 1 January 1 CE is 31 December 1 BCE.

Range of Valid Dates

The range of valid values able to be stored in QDateTime is dependent on the internal storage implementation. QDateTime is currently stored in a qint64 as a serial msecs value encoding the date and time. This restricts the date range to about +/- 292 million years, compared to the QDate range of +/- 2 billion years. Care must be taken when creating a QDateTime with extreme values that you do not overflow the storage. The exact range of supported values varies depending on the Qt::TimeSpec and time zone.

Use of System Timezone

QDateTime uses the system's time zone information to determine the offset of local time from UTC. If the system is not configured correctly or not up-to-date, QDateTime will give wrong results as well.

Daylight Savings Time (DST)

QDateTime takes into account the system's time zone information when dealing with DST. On modern Unix systems, this means it applies the correct historical DST data whenever possible. On Windows and Windows CE, where the system doesn't support historical DST data, historical accuracy is not maintained with respect to DST.

The range of valid dates taking DST into account is 1970-01-01 to the present, and rules are in place for handling DST correctly until 2037-12-31, but these could change. For dates falling outside that range, QDateTime makes a best guess using the rules for year 1970 or 2037, but we can't guarantee accuracy. This means QDateTime doesn't take into account changes in a locale's time zone before 1970, even if the system's time zone database supports that information.

QDateTime takes into consideration the Standard Time to Daylight Time transition. For example if the transition is at 2am and the clock goes forward to 3am, then there is a "missing" hour from 02:00:00 to 02:59:59.999 which QDateTime considers to be invalid. Any date maths performed will take this missing hour into account and return a valid result.

Offset From UTC

A Qt::TimeSpec of Qt::OffsetFromUTC is also supported. This allows you to define a QDateTime relative to UTC at a fixed offset of a given number of seconds from UTC. For example, an offset of +3600 seconds is one hour ahead of UTC and is usually written in ISO standard notation as "UTC+01:00". Daylight Savings Time never applies with this TimeSpec.

There is no explicit size restriction to the offset seconds, but there is an implicit limit imposed when using the toString() and fromString() methods which use a format of [+|-]hh:mm, effectively limiting the range to +/- 99 hours and 59 minutes and whole minutes only. Note that currently no time zone lies outside the range of +/- 14 hours.

Time Zone Support

A Qt::TimeSpec of Qt::TimeZone is also supported in conjunction with the QTimeZone class. This allows you to define a datetime in a named time zone adhering to a consistent set of daylight savings transition rules. For example a time zone of "Europe/Berlin" will apply the daylight savings rules as used in Germany since 1970. Note that the transition rules applied depend on the platform support. See the QTimeZone documentation for more details.

Methods

impl DateTime

fn add_days<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeAddDaysArgs<'largs>

C++ method: QDateTime::addDays

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_days(&self, (i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn add_days(&self, (i64, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::addDays(qint64 days) const

C++ documentation:

Returns a QDateTime object containing a datetime ndays days later than the datetime of this object (or earlier if ndays is negative).

If the timeSpec() is Qt::LocalTime and the resulting date and time fall in the Standard Time to Daylight Time transition hour then the result will be adjusted accordingly, i.e. if the transition is at 2am and the clock goes forward to 3am and the result falls between 2am and 3am then the result will be adjusted to fall after 3am.

See also daysTo(), addMonths(), addYears(), and addSecs().

fn add_m_secs<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeAddMSecsArgs<'largs>

C++ method: QDateTime::addMSecs

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_m_secs(&self, (i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn add_m_secs(&self, (i64, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::addMSecs(qint64 msecs) const

C++ documentation:

Returns a QDateTime object containing a datetime msecs miliseconds later than the datetime of this object (or earlier if msecs is negative).

If this datetime is invalid, an invalid datetime will be returned.

See also addSecs(), msecsTo(), addDays(), addMonths(), and addYears().

fn add_months<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeAddMonthsArgs<'largs>

C++ method: QDateTime::addMonths

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_months(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn add_months(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::addMonths(int months) const

C++ documentation:

Returns a QDateTime object containing a datetime nmonths months later than the datetime of this object (or earlier if nmonths is negative).

If the timeSpec() is Qt::LocalTime and the resulting date and time fall in the Standard Time to Daylight Time transition hour then the result will be adjusted accordingly, i.e. if the transition is at 2am and the clock goes forward to 3am and the result falls between 2am and 3am then the result will be adjusted to fall after 3am.

See also daysTo(), addDays(), addYears(), and addSecs().

fn add_secs<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeAddSecsArgs<'largs>

C++ method: QDateTime::addSecs

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_secs(&self, (i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn add_secs(&self, (i64, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::addSecs(qint64 secs) const

C++ documentation:

Returns a QDateTime object containing a datetime s seconds later than the datetime of this object (or earlier if s is negative).

If this datetime is invalid, an invalid datetime will be returned.

See also addMSecs(), secsTo(), addDays(), addMonths(), and addYears().

fn add_years<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeAddYearsArgs<'largs>

C++ method: QDateTime::addYears

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_years(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn add_years(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::addYears(int years) const

C++ documentation:

Returns a QDateTime object containing a datetime nyears years later than the datetime of this object (or earlier if nyears is negative).

If the timeSpec() is Qt::LocalTime and the resulting date and time fall in the Standard Time to Daylight Time transition hour then the result will be adjusted accordingly, i.e. if the transition is at 2am and the clock goes forward to 3am and the result falls between 2am and 3am then the result will be adjusted to fall after 3am.

See also daysTo(), addDays(), addMonths(), and addSecs().

fn current_date_time<Args>(args: Args) -> Args::ReturnType where Args: DateTimeCurrentDateTimeArgs

C++ method: QDateTime::currentDateTime

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_date_time(cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn current_date_time(cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::currentDateTime()

C++ documentation:

Returns the current datetime, as reported by the system clock, in the local time zone.

See also currentDateTimeUtc(), QDate::currentDate(), QTime::currentTime(), and toTimeSpec().

fn current_date_time_utc<Args>(args: Args) -> Args::ReturnType where Args: DateTimeCurrentDateTimeUtcArgs

C++ method: QDateTime::currentDateTimeUtc

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_date_time_utc(cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn current_date_time_utc(cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::currentDateTimeUtc()

C++ documentation:

Returns the current datetime, as reported by the system clock, in UTC.

This function was introduced in Qt 4.7.

See also currentDateTime(), QDate::currentDate(), QTime::currentTime(), and toTimeSpec().

fn current_m_secs_since_epoch() -> i64

C++ method: static qint64 QDateTime::currentMSecsSinceEpoch()

C++ documentation:

Returns the number of milliseconds since 1970-01-01T00:00:00 Universal Coordinated Time. This number is like the POSIX time_t variable, but expressed in milliseconds instead.

This function was introduced in Qt 4.7.

See also currentDateTime(), currentDateTimeUtc(), toTime_t(), and toTimeSpec().

fn date<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeDateArgs<'largs>

C++ method: QDateTime::date

This is an overloaded function. Available variants:

Rust arguments:
1) fn date(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date::Date>
2) fn date(&self, cpp_utils::AsStruct) -> ::date::Date
C++ method: QDate QDateTime::date() const

C++ documentation:

Returns the date part of the datetime.

See also setDate(), time(), and timeSpec().

fn days_to(&self, arg1: &DateTime) -> i64

C++ method: qint64 QDateTime::daysTo(const QDateTime& arg1) const

C++ documentation:

Returns the number of days from this datetime to the other datetime. The number of days is counted as the number of times midnight is reached between this datetime to the other datetime. This means that a 10 minute difference from 23:55 to 0:05 the next day counts as one day.

If the other datetime is earlier than this datetime, the value returned is negative.

Example:

QDateTime startDate(QDate(2012, 7, 6), QTime(8, 30, 0));
QDateTime endDate(QDate(2012, 7, 7), QTime(16, 30, 0));
qDebug() << "Days from startDate to endDate: " << startDate.daysTo(endDate);

startDate = QDateTime(QDate(2012, 7, 6), QTime(23, 55, 0)); endDate = QDateTime(QDate(2012, 7, 7), QTime(0, 5, 0)); qDebug() << "Days from startDate to endDate: " << startDate.daysTo(endDate);

qSwap(startDate, endDate); // Make endDate before startDate. qDebug() << "Days from startDate to endDate: " << startDate.daysTo(endDate);
See also addDays(), secsTo(), and msecsTo().

fn from_m_secs_since_epoch<Args>(args: Args) -> Args::ReturnType where Args: DateTimeFromMSecsSinceEpochArgs

C++ method: QDateTime::fromMSecsSinceEpoch

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_m_secs_since_epoch((i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_m_secs_since_epoch((i64, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromMSecsSinceEpoch(qint64 msecs)

C++ documentation:

Returns a datetime whose date and time are the number of milliseconds, msecs, that have passed since 1970-01-01T00:00:00.000, Coordinated Universal Time (Qt::UTC), and converted to Qt::LocalTime. On systems that do not support time zones, the time will be set as if local time were Qt::UTC.

Note that there are possible values for msecs that lie outside the valid range of QDateTime, both negative and positive. The behavior of this function is undefined for those values.

This function was introduced in Qt 4.7.

See also toTime_t() and setTime_t().

Variant 2

Rust arguments:
1) fn from_m_secs_since_epoch((i64, ::qt::TimeSpec, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_m_secs_since_epoch((i64, ::qt::TimeSpec, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
3) fn from_m_secs_since_epoch((i64, ::qt::TimeSpec, cpp_utils::AsStruct)) -> ::date_time::DateTime
4) fn from_m_secs_since_epoch((i64, ::qt::TimeSpec, libc::c_int, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromMSecsSinceEpoch(qint64 msecs, Qt::TimeSpec spec, int offsetFromUtc = ?)

C++ documentation:

Returns a datetime whose date and time are the number of milliseconds msecs that have passed since 1970-01-01T00:00:00.000, Coordinated Universal Time (Qt::UTC) and converted to the given spec.

Note that there are possible values for msecs that lie outside the valid range of QDateTime, both negative and positive. The behavior of this function is undefined for those values.

If the spec is not Qt::OffsetFromUTC then the offsetSeconds will be ignored. If the spec is Qt::OffsetFromUTC and the offsetSeconds is 0 then the spec will be set to Qt::UTC, i.e. an offset of 0 seconds.

If spec is Qt::TimeZone then the spec will be set to Qt::LocalTime, i.e. the current system time zone.

This function was introduced in Qt 5.2.

See also fromTime_t().

Variant 3

Rust arguments:
1) fn from_m_secs_since_epoch((i64, &::time_zone::TimeZone, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_m_secs_since_epoch((i64, &::time_zone::TimeZone, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromMSecsSinceEpoch(qint64 msecs, const QTimeZone& timeZone)

C++ documentation:

Returns a datetime whose date and time are the number of milliseconds msecs that have passed since 1970-01-01T00:00:00.000, Coordinated Universal Time (Qt::UTC) and with the given timeZone.

This function was introduced in Qt 5.2.

See also fromTime_t().

fn from_string<Args>(args: Args) -> Args::ReturnType where Args: DateTimeFromStringArgs

C++ method: QDateTime::fromString

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_string((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_string((&::string::String, ::qt::DateFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
3) fn from_string((&::string::String, cpp_utils::AsStruct)) -> ::date_time::DateTime
4) fn from_string((&::string::String, ::qt::DateFormat, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromString(const QString& s, Qt::DateFormat f = ?)

C++ documentation:

Returns the QDateTime represented by the string, using the format given, or an invalid datetime if this is not possible.

Note for Qt::TextDate: It is recommended that you use the English short month names (e.g. "Jan"). Although localized month names can also be used, they depend on the user's locale settings.

Variant 2

Rust arguments:
1) fn from_string((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_string((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromString(const QString& s, const QString& format)

C++ documentation:

Returns the QDateTime represented by the string, using the format given, or an invalid datetime if the string cannot be parsed.

These expressions may be used for the date part of the format string:

Expression	Output

d	the day as number without a leading zero (1 to 31)

dd	the day as number with a leading zero (01 to 31)

ddd	the abbreviated localized day name (e.g. 'Mon' to 'Sun'). Uses QDate::shortDayName().

dddd	the long localized day name (e.g. 'Monday' to 'Sunday'). Uses QDate::longDayName().

M	the month as number without a leading zero (1-12)

MM	the month as number with a leading zero (01-12)

MMM	the abbreviated localized month name (e.g. 'Jan' to 'Dec'). Uses QDate::shortMonthName().

MMMM	the long localized month name (e.g. 'January' to 'December'). Uses QDate::longMonthName().

yy	the year as two digit number (00-99)

yyyy	the year as four digit number

Note: Unlike the other version of this function, day and month names must be given in the user's local language. It is only possible to use the English names if the user's language is English.

These expressions may be used for the time part of the format string:

Expression	Output

h	the hour without a leading zero (0 to 23 or 1 to 12 if AM/PM display)

hh	the hour with a leading zero (00 to 23 or 01 to 12 if AM/PM display)

H	the hour without a leading zero (0 to 23, even with AM/PM display)

HH	the hour with a leading zero (00 to 23, even with AM/PM display)

m	the minute without a leading zero (0 to 59)

mm	the minute with a leading zero (00 to 59)

s	the second without a leading zero (0 to 59)

ss	the second with a leading zero (00 to 59)

z	the milliseconds without leading zeroes (0 to 999)

zzz	the milliseconds with leading zeroes (000 to 999)

AP or A	interpret as an AM/PM time. AP must be either "AM" or "PM".

ap or a	Interpret as an AM/PM time. ap must be either "am" or "pm".

All other input characters will be treated as text. Any sequence of characters that are enclosed in single quotes will also be treated as text and not be used as an expression.

QTime time1 = QTime::fromString("131", "HHh");
// time1 is 13:00:00
QTime time1 = QTime::fromString("1apA", "1amAM");
// time1 is 01:00:00

QDateTime dateTime2 = QDateTime::fromString("M1d1y9800:01:02", "'M'M'd'd'y'yyhh:mm:ss"); // dateTime is 1 January 1998 00:01:02
If the format is not satisfied, an invalid QDateTime is returned. The expressions that don't have leading zeroes (d, M, h, m, s, z) will be greedy. This means that they will use two digits even if this will put them outside the range and/or leave too few digits for other sections.

QDateTime dateTime = QDateTime::fromString("130", "Mm"); // invalid

This could have meant 1 January 00:30.00 but the M will grab two digits.

Incorrectly specified fields of the string will cause an invalid QDateTime to be returned. For example, consider the following code, where the two digit year 12 is read as 1912 (see the table below for all field defaults); the resulting datetime is invalid because 23 April 1912 was a Tuesday, not a Monday:

QString string = "Monday, 23 April 12 22:51:41";
QString format = "dddd, d MMMM yy hh:mm:ss";
QDateTime invalid = QDateTime::fromString(string, format);

The correct code is:

QString string = "Tuesday, 23 April 12 22:51:41";
QString format = "dddd, d MMMM yy hh:mm:ss";
QDateTime valid = QDateTime::fromString(string, format);

For any field that is not represented in the format, the following defaults are used:

Field	Default value

Year	1900

Month	1 (January)

Day	1

Hour	0

Minute	0

Second	0

For example:

QDateTime dateTime = QDateTime::fromString("1.30.1", "M.d.s");
// dateTime is January 30 in 1900 at 00:00:01.
dateTime = QDateTime::fromString("12", "yy");
// dateTime is January 1 in 1912 at 00:00:00.

See also QDate::fromString(), QTime::fromString(), QDate::toString(), QDateTime::toString(), and QTime::toString().

fn from_time_t<Args>(args: Args) -> Args::ReturnType where Args: DateTimeFromTimeTArgs

C++ method: QDateTime::fromTime_t

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_time_t((libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_time_t((libc::c_uint, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromTime_t(unsigned int secsSince1Jan1970UTC)

C++ documentation:

Returns a datetime whose date and time are the number of seconds that have passed since 1970-01-01T00:00:00, Coordinated Universal Time (Qt::UTC) and converted to Qt::LocalTime. On systems that do not support time zones, the time will be set as if local time were Qt::UTC.

This function was introduced in Qt 4.2.

See also toTime_t() and setTime_t().

Variant 2

Rust arguments:
1) fn from_time_t((libc::c_uint, ::qt::TimeSpec, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_time_t((libc::c_uint, ::qt::TimeSpec, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
3) fn from_time_t((libc::c_uint, ::qt::TimeSpec, cpp_utils::AsStruct)) -> ::date_time::DateTime
4) fn from_time_t((libc::c_uint, ::qt::TimeSpec, libc::c_int, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromTime_t(unsigned int secsSince1Jan1970UTC, Qt::TimeSpec spec, int offsetFromUtc = ?)

C++ documentation:

Returns a datetime whose date and time are the number of seconds that have passed since 1970-01-01T00:00:00, Coordinated Universal Time (Qt::UTC) and converted to the given spec.

If the spec is not Qt::OffsetFromUTC then the offsetSeconds will be ignored. If the spec is Qt::OffsetFromUTC and the offsetSeconds is 0 then the spec will be set to Qt::UTC, i.e. an offset of 0 seconds.

This function was introduced in Qt 5.2.

See also toTime_t() and setTime_t().

Variant 3

Rust arguments:
1) fn from_time_t((libc::c_uint, &::time_zone::TimeZone, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn from_time_t((libc::c_uint, &::time_zone::TimeZone, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: static QDateTime QDateTime::fromTime_t(unsigned int secsSince1Jan1970UTC, const QTimeZone& timeZone)

C++ documentation:

Returns a datetime whose date and time are the number of seconds that have passed since 1970-01-01T00:00:00, Coordinated Universal Time (Qt::UTC) and with the given timeZone.

This function was introduced in Qt 5.2.

See also toTime_t() and setTime_t().

fn is_daylight_time(&self) -> bool

C++ method: bool QDateTime::isDaylightTime() const

C++ documentation:

Returns if this datetime falls in Daylight Savings Time.

If the Qt::TimeSpec is not Qt::LocalTime or Qt::TimeZone then will always return false.

This function was introduced in Qt 5.2.

See also timeSpec().

fn is_null(&self) -> bool

C++ method: bool QDateTime::isNull() const

C++ documentation:

Returns true if both the date and the time are null; otherwise returns false. A null datetime is invalid.

See also QDate::isNull(), QTime::isNull(), and isValid().

fn is_valid(&self) -> bool

C++ method: bool QDateTime::isValid() const

C++ documentation:

Returns true if both the date and the time are valid and they are valid in the current Qt::TimeSpec, otherwise returns false.

If the timeSpec() is Qt::LocalTime or Qt::TimeZone then the date and time are checked to see if they fall in the Standard Time to Daylight Time transition hour, i.e. if the transition is at 2am and the clock goes forward to 3am then the time from 02:00:00 to 02:59:59.999 is considered to be invalid.

See also QDate::isValid() and QTime::isValid().

fn msecs_to(&self, arg1: &DateTime) -> i64

C++ method: qint64 QDateTime::msecsTo(const QDateTime& arg1) const

C++ documentation:

Returns the number of milliseconds from this datetime to the other datetime. If the other datetime is earlier than this datetime, the value returned is negative.

Before performing the comparison, the two datetimes are converted to Qt::UTC to ensure that the result is correct if one of the two datetimes has daylight saving time (DST) and the other doesn't.

Returns 0 if either datetime is invalid.

See also addMSecs(), daysTo(), and QTime::msecsTo().

fn new<Args>(args: Args) -> Args::ReturnType where Args: DateTimeNewArgs

C++ method: QDateTime::QDateTime

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::date_time::DateTime
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
C++ method: [constructor] void QDateTime::QDateTime()

Variant 2

Rust arguments:
1) fn new((&::date::Date, cpp_utils::AsStruct)) -> ::date_time::DateTime
2) fn new((&::date::Date, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
C++ method: [constructor] void QDateTime::QDateTime(const QDate& arg1)

Variant 3

Rust arguments:
1) fn new((&::date::Date, &::time::Time, cpp_utils::AsStruct)) -> ::date_time::DateTime
2) fn new((&::date::Date, &::time::Time, ::qt::TimeSpec, cpp_utils::AsStruct)) -> ::date_time::DateTime
3) fn new((&::date::Date, &::time::Time, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
4) fn new((&::date::Date, &::time::Time, ::qt::TimeSpec, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
C++ method: [constructor] void QDateTime::QDateTime(const QDate& arg1, const QTime& arg2, Qt::TimeSpec spec = ?)

Variant 4

Rust arguments:
1) fn new((&::date::Date, &::time::Time, ::qt::TimeSpec, libc::c_int, cpp_utils::AsStruct)) -> ::date_time::DateTime
2) fn new((&::date::Date, &::time::Time, ::qt::TimeSpec, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
C++ method: [constructor] void QDateTime::QDateTime(const QDate& date, const QTime& time, Qt::TimeSpec spec, int offsetSeconds)

Variant 5

Rust arguments:
1) fn new((&::date::Date, &::time::Time, &::time_zone::TimeZone, cpp_utils::AsStruct)) -> ::date_time::DateTime
2) fn new((&::date::Date, &::time::Time, &::time_zone::TimeZone, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
C++ method: [constructor] void QDateTime::QDateTime(const QDate& date, const QTime& time, const QTimeZone& timeZone)

Variant 6

Rust arguments:
1) fn new((&::date_time::DateTime, cpp_utils::AsStruct)) -> ::date_time::DateTime
2) fn new((&::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
C++ method: [constructor] void QDateTime::QDateTime(const QDateTime& other)

fn offset_from_utc(&self) -> c_int

C++ method: int QDateTime::offsetFromUtc() const

C++ documentation:

Returns the current Offset From UTC in seconds.

If the timeSpec() is Qt::OffsetFromUTC this will be the value originally set.

If the timeSpec() is Qt::TimeZone this will be the offset effective in the Time Zone including any Daylight Saving Offset.

If the timeSpec() is Qt::LocalTime this will be the difference between the Local Time and UTC including any Daylight Saving Offset.

If the timeSpec() is Qt::UTC this will be 0.

This function was introduced in Qt 5.2.

See also setOffsetFromUtc().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 DateTime) -> &'l0 mut DateTime

C++ method: QDateTime& QDateTime::operator=(const QDateTime& other)

C++ documentation:

Makes a copy of the other datetime and returns a reference to the copy.

fn op_eq(&self, other: &DateTime) -> bool

C++ method: bool QDateTime::operator==(const QDateTime& other) const

C++ documentation:

Returns true if this datetime is equal to the other datetime; otherwise returns false.

See also operator!=().

fn op_ge(&self, other: &DateTime) -> bool

C++ method: bool QDateTime::operator>=(const QDateTime& other) const

C++ documentation:

Returns true if this datetime is later than or equal to the other datetime; otherwise returns false.

fn op_gt(&self, other: &DateTime) -> bool

C++ method: bool QDateTime::operator>(const QDateTime& other) const

C++ documentation:

Returns true if this datetime is later than the other datetime; otherwise returns false.

fn op_le(&self, other: &DateTime) -> bool

C++ method: bool QDateTime::operator<=(const QDateTime& other) const

C++ documentation:

Returns true if this datetime is earlier than or equal to the other datetime; otherwise returns false.

fn op_lt(&self, other: &DateTime) -> bool

C++ method: bool QDateTime::operator<(const QDateTime& other) const

C++ documentation:

Returns true if this datetime is earlier than the other datetime; otherwise returns false.

fn op_neq(&self, other: &DateTime) -> bool

C++ method: bool QDateTime::operator!=(const QDateTime& other) const

C++ documentation:

Returns true if this datetime is different from the other datetime; otherwise returns false.

Two datetimes are different if either the date, the time, or the time zone components are different.

See also operator==().

fn secs_to(&self, arg1: &DateTime) -> i64

C++ method: qint64 QDateTime::secsTo(const QDateTime& arg1) const

C++ documentation:

Returns the number of seconds from this datetime to the other datetime. If the other datetime is earlier than this datetime, the value returned is negative.

Before performing the comparison, the two datetimes are converted to Qt::UTC to ensure that the result is correct if one of the two datetimes has daylight saving time (DST) and the other doesn't.

Returns 0 if either datetime is invalid.

Example:

QDateTime now = QDateTime::currentDateTime();
QDateTime xmas(QDate(now.date().year(), 12, 25), QTime(0, 0));
qDebug("There are %d seconds to Christmas", now.secsTo(xmas));

See also addSecs(), daysTo(), and QTime::secsTo().

fn set_date(&mut self, date: &Date)

C++ method: void QDateTime::setDate(const QDate& date)

C++ documentation:

Sets the date part of this datetime to date. If no time is set yet, it is set to midnight. If date is invalid, this QDateTime becomes invalid.

See also date(), setTime(), and setTimeSpec().

fn set_m_secs_since_epoch(&mut self, msecs: i64)

C++ method: void QDateTime::setMSecsSinceEpoch(qint64 msecs)

C++ documentation:

Sets the date and time given the number of milliseconds msecs that have passed since 1970-01-01T00:00:00.000, Coordinated Universal Time (Qt::UTC). On systems that do not support time zones this function will behave as if local time were Qt::UTC.

Note that passing the minimum of qint64 (std::numeric_limits<qint64>::min()) to msecs will result in undefined behavior.

This function was introduced in Qt 4.7.

See also toMSecsSinceEpoch() and setTime_t().

fn set_offset_from_utc(&mut self, offset_seconds: c_int)

C++ method: void QDateTime::setOffsetFromUtc(int offsetSeconds)

C++ documentation:

Sets the timeSpec() to Qt::OffsetFromUTC and the offset to offsetSeconds. The datetime will refer to a different point in time.

The maximum and minimum offset is 14 positive or negative hours. If offsetSeconds is larger or smaller than that, then the result is undefined.

If offsetSeconds is 0 then the timeSpec() will be set to Qt::UTC.

This function was introduced in Qt 5.2.

See also isValid() and offsetFromUtc().

fn set_time(&mut self, time: &Time)

C++ method: void QDateTime::setTime(const QTime& time)

C++ documentation:

Sets the time part of this datetime to time. If time is not valid, this function sets it to midnight. Therefore, it's possible to clear any set time in a QDateTime by setting it to a default QTime:

QDateTime dt = QDateTime::currentDateTime();
dt.setTime(QTime());

See also time(), setDate(), and setTimeSpec().

fn set_time_spec(&mut self, spec: TimeSpec)

C++ method: void QDateTime::setTimeSpec(Qt::TimeSpec spec)

C++ documentation:

Sets the time specification used in this datetime to spec. The datetime will refer to a different point in time.

If spec is Qt::OffsetFromUTC then the timeSpec() will be set to Qt::UTC, i.e. an effective offset of 0.

If spec is Qt::TimeZone then the spec will be set to Qt::LocalTime, i.e. the current system time zone.

Example:

QDateTime local(QDateTime::currentDateTime());
qDebug() << "Local time is:" << local;

QDateTime UTC(local); UTC.setTimeSpec(Qt::UTC); qDebug() << "UTC time is:" << UTC;

qDebug() << "There are" << local.secsTo(UTC) << "seconds difference between the datetimes.";
See also timeSpec(), setDate(), setTime(), setTimeZone(), and Qt::TimeSpec.

fn set_time_t(&mut self, secs_since1_jan1970_u_t_c: c_uint)

C++ method: void QDateTime::setTime_t(unsigned int secsSince1Jan1970UTC)

C++ documentation:

Sets the date and time given the number of seconds that have passed since 1970-01-01T00:00:00, Coordinated Universal Time (Qt::UTC). On systems that do not support time zones this function will behave as if local time were Qt::UTC.

See also toTime_t().

fn set_time_zone(&mut self, to_zone: &TimeZone)

C++ method: void QDateTime::setTimeZone(const QTimeZone& toZone)

C++ documentation:

Sets the time zone used in this datetime to toZone. The datetime will refer to a different point in time.

If toZone is invalid then the datetime will be invalid.

This function was introduced in Qt 5.2.

See also timeZone() and Qt::TimeSpec.

fn set_utc_offset(&mut self, seconds: c_int)

C++ method: void QDateTime::setUtcOffset(int seconds)

fn swap(&mut self, other: &mut DateTime)

C++ method: void QDateTime::swap(QDateTime& other)

C++ documentation:

Swaps this datetime with other. This operation is very fast and never fails.

This function was introduced in Qt 5.0.

fn time<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeTimeArgs<'largs>

C++ method: QDateTime::time

This is an overloaded function. Available variants:

Rust arguments:
1) fn time(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::time::Time>
2) fn time(&self, cpp_utils::AsStruct) -> ::time::Time
C++ method: QTime QDateTime::time() const

C++ documentation:

Returns the time part of the datetime.

See also setTime(), date(), and timeSpec().

fn time_spec(&self) -> TimeSpec

C++ method: Qt::TimeSpec QDateTime::timeSpec() const

C++ documentation:

Returns the time specification of the datetime.

See also setTimeSpec(), date(), time(), and Qt::TimeSpec.

fn time_zone<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeTimeZoneArgs<'largs>

C++ method: QDateTime::timeZone

This is an overloaded function. Available variants:

Rust arguments:
1) fn time_zone(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::time_zone::TimeZone>
2) fn time_zone(&self, cpp_utils::AsStruct) -> ::time_zone::TimeZone
C++ method: QTimeZone QDateTime::timeZone() const

C++ documentation:

Returns the time zone of the datetime.

If the timeSpec() is Qt::LocalTime then an instance of the current system time zone will be returned. Note however that if you copy this time zone the instance will not remain in sync if the system time zone changes.

This function was introduced in Qt 5.2.

See also setTimeZone() and Qt::TimeSpec.

fn time_zone_abbreviation<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeTimeZoneAbbreviationArgs<'largs>

C++ method: QDateTime::timeZoneAbbreviation

This is an overloaded function. Available variants:

Rust arguments:
1) fn time_zone_abbreviation(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn time_zone_abbreviation(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDateTime::timeZoneAbbreviation() const

C++ documentation:

Returns the Time Zone Abbreviation for the datetime.

If the timeSpec() is Qt::UTC this will be "UTC".

If the timeSpec() is Qt::OffsetFromUTC this will be in the format "UTC[+-]00:00".

If the timeSpec() is Qt::LocalTime then the host system is queried for the correct abbreviation.

Note that abbreviations may or may not be localized.

Note too that the abbreviation is not guaranteed to be a unique value, i.e. different time zones may have the same abbreviation.

This function was introduced in Qt 5.2.

See also timeSpec().

fn to_local_time<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeToLocalTimeArgs<'largs>

C++ method: QDateTime::toLocalTime

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_local_time(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn to_local_time(&self, cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::toLocalTime() const

C++ documentation:

Returns a datetime containing the date and time information in this datetime, but specified using the Qt::LocalTime definition.

Example:

QDateTime UTC(QDateTime::currentDateTimeUtc());
QDateTime local(UTC.toLocalTime());
qDebug() << "UTC time is:" << UTC;
qDebug() << "Local time is:" << local;
qDebug() << "No difference between times:" << UTC.secsTo(local);

See also toTimeSpec().

fn to_m_secs_since_epoch(&self) -> i64

C++ method: qint64 QDateTime::toMSecsSinceEpoch() const

C++ documentation:

Returns the datetime as the number of milliseconds that have passed since 1970-01-01T00:00:00.000, Coordinated Universal Time (Qt::UTC).

On systems that do not support time zones, this function will behave as if local time were Qt::UTC.

The behavior for this function is undefined if the datetime stored in this object is not valid. However, for all valid dates, this function returns a unique value.

This function was introduced in Qt 4.7.

See also toTime_t() and setMSecsSinceEpoch().

fn to_offset_from_utc<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeToOffsetFromUtcArgs<'largs>

C++ method: QDateTime::toOffsetFromUtc

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_offset_from_utc(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn to_offset_from_utc(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::toOffsetFromUtc(int offsetSeconds) const

C++ documentation:

Returns a copy of this datetime converted to a spec of Qt::OffsetFromUTC with the given offsetSeconds.

If the offsetSeconds equals 0 then a UTC datetime will be returned

This function was introduced in Qt 5.2.

See also setOffsetFromUtc(), offsetFromUtc(), and toTimeSpec().

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeToStringArgs<'largs>

C++ method: QDateTime::toString

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_string(&self, (::qt::DateFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
3) fn to_string(&self, (::qt::DateFormat, cpp_utils::AsStruct)) -> ::string::String
4) fn to_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDateTime::toString(Qt::DateFormat f = ?) const

C++ documentation:

This is an overloaded function.

Returns the datetime as a string in the format given.

If the format is Qt::TextDate, the string is formatted in the default way. QDate::shortDayName(), QDate::shortMonthName(), and QTime::toString() are used to generate the string, so the day and month names will be localized names using the system locale, i.e. QLocale::system(). An example of this formatting is "Wed May 20 03:40:13 1998".

If the format is Qt::ISODate, the string format corresponds to the ISO 8601 extended specification for representations of dates and times, taking the form YYYY-MM-DDTHH:mm:ss[Z|[+|-]HH:mm], depending on the timeSpec() of the QDateTime. If the timeSpec() is Qt::UTC, Z will be appended to the string; if the timeSpec() is Qt::OffsetFromUTC, the offset in hours and minutes from UTC will be appended to the string.

If the format is Qt::SystemLocaleShortDate or Qt::SystemLocaleLongDate, the string format depends on the locale settings of the system. Identical to calling QLocale::system().toString(datetime, QLocale::ShortFormat) or QLocale::system().toString(datetime, QLocale::LongFormat).

If the format is Qt::DefaultLocaleShortDate or Qt::DefaultLocaleLongDate, the string format depends on the default application locale. This is the locale set with QLocale::setDefault(), or the system locale if no default locale has been set. Identical to calling QLocale().toString(datetime, QLocale::ShortFormat) or QLocale().toString(datetime, QLocale::LongFormat).

If the format is Qt::RFC2822Date, the string is formatted following RFC 2822.

If the datetime is invalid, an empty string will be returned.

Warning: The Qt::ISODate format is only valid for years in the range 0 to 9999. This restriction may apply to locale-aware formats as well, depending on the locale settings.

See also QDate::toString(), QTime::toString(), and Qt::DateFormat.

Variant 2

Rust arguments:
1) fn to_string(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QDateTime::toString(const QString& format) const

C++ documentation:

Returns the datetime as a string. The format parameter determines the format of the result string.

These expressions may be used for the date:

Expression	Output

d	the day as number without a leading zero (1 to 31)

dd	the day as number with a leading zero (01 to 31)

ddd	the abbreviated localized day name (e.g. 'Mon' to 'Sun'). Uses the system locale to localize the name, i.e. QLocale::system().

dddd	the long localized day name (e.g. 'Monday' to 'Qt::Sunday'). Uses the system locale to localize the name, i.e. QLocale::system().

M	the month as number without a leading zero (1-12)

MM	the month as number with a leading zero (01-12)

MMM	the abbreviated localized month name (e.g. 'Jan' to 'Dec'). Uses the system locale to localize the name, i.e. QLocale::system().

MMMM	the long localized month name (e.g. 'January' to 'December'). Uses the system locale to localize the name, i.e. QLocale::system().

yy	the year as two digit number (00-99)

yyyy	the year as four digit number

These expressions may be used for the time:

Expression	Output

h	the hour without a leading zero (0 to 23 or 1 to 12 if AM/PM display)

hh	the hour with a leading zero (00 to 23 or 01 to 12 if AM/PM display)

H	the hour without a leading zero (0 to 23, even with AM/PM display)

HH	the hour with a leading zero (00 to 23, even with AM/PM display)

m	the minute without a leading zero (0 to 59)

mm	the minute with a leading zero (00 to 59)

s	the second without a leading zero (0 to 59)

ss	the second with a leading zero (00 to 59)

z	the milliseconds without leading zeroes (0 to 999)

zzz	the milliseconds with leading zeroes (000 to 999)

AP or A	use AM/PM display. A/AP will be replaced by either "AM" or "PM".

ap or a	use am/pm display. a/ap will be replaced by either "am" or "pm".

t	the timezone (for example "CEST")

All other input characters will be ignored. Any sequence of characters that are enclosed in single quotes will be treated as text and not be used as an expression. Two consecutive single quotes ("''") are replaced by a singlequote in the output. Formats without separators (e.g. "HHmm") are currently not supported.

Example format strings (assumed that the QDateTime is 21 May 2001 14:13:09):

Format	Result

dd.MM.yyyy	21.05.2001

ddd MMMM d yy	Tue May 21 01

hh:mm:ss.zzz	14:13:09.042

h:m:s ap	2:13:9 pm

If the datetime is invalid, an empty string will be returned.

See also QDate::toString(), QTime::toString(), and QLocale::toString().

fn to_time_spec<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeToTimeSpecArgs<'largs>

C++ method: QDateTime::toTimeSpec

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_time_spec(&self, (::qt::TimeSpec, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn to_time_spec(&self, (::qt::TimeSpec, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::toTimeSpec(Qt::TimeSpec spec) const

C++ documentation:

Returns a copy of this datetime converted to the given time spec.

If spec is Qt::OffsetFromUTC then it is set to Qt::UTC. To set to a spec of Qt::OffsetFromUTC use toOffsetFromUtc().

If spec is Qt::TimeZone then it is set to Qt::LocalTime, i.e. the local Time Zone.

Example:

QDateTime local(QDateTime::currentDateTime());
QDateTime UTC(local.toTimeSpec(Qt::UTC));
qDebug() << "Local time is:" << local;
qDebug() << "UTC time is:" << UTC;
qDebug() << "No difference between times:" << local.secsTo(UTC);

See also timeSpec(), toTimeZone(), toUTC(), and toLocalTime().

fn to_time_t(&self) -> c_uint

C++ method: unsigned int QDateTime::toTime_t() const

C++ documentation:

Returns the datetime as the number of seconds that have passed since 1970-01-01T00:00:00, Coordinated Universal Time (Qt::UTC).

On systems that do not support time zones, this function will behave as if local time were Qt::UTC.

Note: This function returns a 32-bit unsigned integer, so it does not support dates before 1970, but it does support dates after 2038-01-19T03:14:06, which may not be valid time_t values. Be careful when passing those time_t values to system functions, which could interpret them as negative dates.

If the date is outside the range 1970-01-01T00:00:00 to 2106-02-07T06:28:14, this function returns -1 cast to an unsigned integer (i.e., 0xFFFFFFFF).

To get an extended range, use toMSecsSinceEpoch().

See also toMSecsSinceEpoch() and setTime_t().

fn to_time_zone<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeToTimeZoneArgs<'largs>

C++ method: QDateTime::toTimeZone

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_time_zone(&self, (&::time_zone::TimeZone, cpp_utils::AsBox)) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn to_time_zone(&self, (&::time_zone::TimeZone, cpp_utils::AsStruct)) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::toTimeZone(const QTimeZone& toZone) const

C++ documentation:

Returns a copy of this datetime converted to the given timeZone

This function was introduced in Qt 5.2.

See also timeZone() and toTimeSpec().

fn to_utc<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DateTimeToUtcArgs<'largs>

C++ method: QDateTime::toUTC

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_utc(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn to_utc(&self, cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: QDateTime QDateTime::toUTC() const

C++ documentation:

Returns a datetime containing the date and time information in this datetime, but specified using the Qt::UTC definition.

Example:

QDateTime local(QDateTime::currentDateTime());
QDateTime UTC(local.toUTC());
qDebug() << "Local time is:" << local;
qDebug() << "UTC time is:" << UTC;
qDebug() << "No difference between times:" << local.secsTo(UTC);

See also toTimeSpec().

fn utc_offset(&self) -> c_int

C++ method: int QDateTime::utcOffset() const