Struct qt_core::debug::Debug

pub struct Debug {
    // some fields omitted
}

C++ type: QDebug

C++ documentation:

The QDebug class provides an output stream for debugging information.

QDebug is used whenever the developer needs to write out debugging or tracing information to a device, file, string or console.

Methods

impl Debug

fn auto_insert_spaces(&self) -> bool

C++ method: bool QDebug::autoInsertSpaces() const

C++ documentation:

Returns true if this QDebug instance will automatically insert spaces between writes.

This function was introduced in Qt 5.0.

See also setAutoInsertSpaces() and QDebugStateSaver.

fn maybe_quote<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Debug where Args: DebugMaybeQuoteArgs<'largs>

C++ method: QDebug::maybeQuote

This is an overloaded function. Available variants:

Rust arguments:
1) fn maybe_quote(&mut self, libc::c_char) -> &'l0 mut ::debug::Debug
2) fn maybe_quote(&mut self, ()) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::maybeQuote(char c = ?)

C++ documentation:

Writes a character c to the debug stream, depending on the current setting for automatic insertion of quotes, and returns a reference to the stream.

The default character is a double quote ".

This function was introduced in Qt 5.4.

See also quote() and noquote().

fn maybe_space<'l0>(&'l0 mut self) -> &'l0 mut Debug

C++ method: QDebug& QDebug::maybeSpace()

C++ documentation:

Writes a space character to the debug stream, depending on the current setting for automatic insertion of spaces, and returns a reference to the stream.

See also space() and nospace().

fn new<Args>(args: Args) -> Args::ReturnType where Args: DebugNewArgs

C++ method: QDebug::QDebug

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::debug::Debug
2) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
C++ method: [constructor] void QDebug::QDebug(QIODevice* device)

Variant 2

Rust arguments:
1) fn new((*mut ::string::String, cpp_utils::AsStruct)) -> ::debug::Debug
2) fn new((*mut ::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
C++ method: [constructor] void QDebug::QDebug(QString* string)

Variant 3

Rust arguments:
1) fn new((::message_logger::MsgType, cpp_utils::AsStruct)) -> ::debug::Debug
2) fn new((::message_logger::MsgType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
C++ method: [constructor] void QDebug::QDebug(QtMsgType t)

Variant 4

Rust arguments:
1) fn new((&::debug::Debug, cpp_utils::AsStruct)) -> ::debug::Debug
2) fn new((&::debug::Debug, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
C++ method: [constructor] void QDebug::QDebug(const QDebug& o)

fn noquote<'l0>(&'l0 mut self) -> &'l0 mut Debug

C++ method: QDebug& QDebug::noquote()

C++ documentation:

Disables automatic insertion of quotation characters around QChar, QString and QByteArray contents and returns a reference to the stream.

This function was introduced in Qt 5.4.

See also quote() and maybeQuote().

fn nospace<'l0>(&'l0 mut self) -> &'l0 mut Debug

C++ method: QDebug& QDebug::nospace()

C++ documentation:

Disables automatic insertion of spaces and returns a reference to the stream.

See also space() and maybeSpace().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 Debug) -> &'l0 mut Debug

C++ method: QDebug& QDebug::operator=(const QDebug& other)

C++ documentation:

Assigns the other debug stream to this stream and returns a reference to this stream.

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Debug where Args: DebugOpShlArgs<'largs>

C++ method: QDebug::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::char::Char) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(QChar t)

C++ documentation:

Writes the character, t, to the stream and returns a reference to the stream.

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::latin1_string::Latin1String) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(QLatin1String t)

C++ documentation:

Writes the Latin1-encoded string, s, to the stream and returns a reference to the stream.

Variant 3

Rust arguments: fn op_shl(&mut self, &'l1 ::text_stream_manipulator::TextStreamManipulator) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(QTextStreamManipulator m)

Variant 4

Rust arguments: fn op_shl(&mut self, bool) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(bool t)

C++ documentation:

Writes the boolean value, t, to the stream and returns a reference to the stream.

Variant 5

Rust arguments: fn op_shl(&mut self, libc::c_char) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(char t)

C++ documentation:

Writes the character, t, to the stream and returns a reference to the stream.

Variant 6

Rust arguments: fn op_shl(&mut self, &'l1 ::byte_array::ByteArray) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(const QByteArray& t)

C++ documentation:

Writes the byte array, b, to the stream and returns a reference to the stream.

Variant 7

Rust arguments: fn op_shl(&mut self, &'l1 ::string::String) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(const QString& t)

C++ documentation:

Writes the string, s, to the stream and returns a reference to the stream.

Variant 8

Rust arguments: fn op_shl(&mut self, &'l1 ::string_ref::StringRef) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(const QStringRef& t)

C++ documentation:

Writes the string reference, s, to the stream and returns a reference to the stream.

Variant 9

Rust arguments: fn op_shl(&mut self, *const libc::c_char) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(const char* t)

C++ documentation:

Writes the '\0'-terminated string, s, to the stream and returns a reference to the stream.

Variant 10

Rust arguments: fn op_shl(&mut self, *const libc::c_void) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(const void* t)

C++ documentation:

Writes a pointer, p, to the stream and returns a reference to the stream.

Variant 11

Rust arguments: fn op_shl(&mut self, libc::c_double) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(double t)

C++ documentation:

Writes the 64-bit floating point number, f, to the stream and returns a reference to the stream.

Variant 12

Rust arguments: fn op_shl(&mut self, libc::c_float) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(float t)

C++ documentation:

Writes the 32-bit floating point number, f, to the stream and returns a reference to the stream.

Variant 13

Rust arguments: fn op_shl(&mut self, libc::c_int) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(int t)

C++ documentation:

Writes the signed integer, i, to the stream and returns a reference to the stream.

Variant 14

Rust arguments: fn op_shl(&mut self, libc::c_long) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(long t)

C++ documentation:

Writes the signed long integer, l, to the stream and returns a reference to the stream.

Variant 15

Rust arguments: fn op_shl(&mut self, u64) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(quint64 t)

C++ documentation:

Writes then unsigned 64-bit integer, i, to the stream and returns a reference to the stream.

Variant 16

Rust arguments: fn op_shl(&mut self, libc::c_short) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(short t)

C++ documentation:

Writes the signed short integer, i, to the stream and returns a reference to the stream.

Variant 17

Rust arguments: fn op_shl(&mut self, libc::c_uint) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(unsigned int t)

C++ documentation:

Writes then unsigned integer, i, to the stream and returns a reference to the stream.

Variant 18

Rust arguments: fn op_shl(&mut self, libc::c_ushort) -> &'l0 mut ::debug::Debug
C++ method: QDebug& QDebug::operator<<(unsigned short t)

C++ documentation:

Writes then unsigned short integer, i, to the stream and returns a reference to the stream.

fn quote<'l0>(&'l0 mut self) -> &'l0 mut Debug

C++ method: QDebug& QDebug::quote()

C++ documentation:

Enables automatic insertion of quotation characters around QChar, QString and QByteArray contents and returns a reference to the stream.

Quoting is enabled by default.

This function was introduced in Qt 5.4.

See also noquote() and maybeQuote().

fn reset_format<'l0>(&'l0 mut self) -> &'l0 mut Debug

C++ method: QDebug& QDebug::resetFormat()

C++ documentation:

Resets the stream formatting options, bringing it back to its original constructed state.

This function was introduced in Qt 5.4.

See also space() and quote().

fn set_auto_insert_spaces(&mut self, b: bool)

C++ method: void QDebug::setAutoInsertSpaces(bool b)

C++ documentation:

Enables automatic insertion of spaces between writes if b is true; otherwise automatic insertion of spaces is disabled.

This function was introduced in Qt 5.0.

See also autoInsertSpaces() and QDebugStateSaver.

fn space<'l0>(&'l0 mut self) -> &'l0 mut Debug

C++ method: QDebug& QDebug::space()

C++ documentation:

Writes a space character to the debug stream and returns a reference to the stream.

The stream remembers that automatic insertion of spaces is enabled for future writes.

See also nospace() and maybeSpace().

fn swap(&mut self, other: &mut Debug)

C++ method: void QDebug::swap(QDebug& other)

C++ documentation:

Swaps this debug stream instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.