Struct qt_core::debug_state_saver::DebugStateSaver

pub struct DebugStateSaver {
    // some fields omitted
}

C++ type: QDebugStateSaver

C++ documentation:

Convenience class for custom QDebug operators

Saves the settings used by QDebug, and restores them upon destruction, then calls maybeSpace(), to separate arguments with a space if autoInsertSpaces() was true at the time of constructing the QDebugStateSaver.

The automatic insertion of spaces between writes is one of the settings that QDebugStateSaver stores for the duration of the current block.

The settings of the internal QTextStream are also saved and restored, so that using << hex in a QDebug operator doesn't affect other QDebug operators.

Methods

impl DebugStateSaver

fn new<Args>(args: Args) -> Args::ReturnType where Args: DebugStateSaverNewArgs

C++ method: QDebugStateSaver::QDebugStateSaver

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&mut ::debug::Debug, cpp_utils::AsStruct)) -> ::debug_state_saver::DebugStateSaver
2) fn new((&mut ::debug::Debug, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug_state_saver::DebugStateSaver>
C++ method: [constructor] void QDebugStateSaver::QDebugStateSaver(QDebug& dbg)