Enum qt_core::dir::Filter

pub enum Filter {
    NoFilter,
    Dirs,
    Files,
    Drives,
    AllEntries,
    NoSymLinks,
    TypeMask,
    Readable,
    Writable,
    Executable,
    PermissionMask,
    Modified,
    Hidden,
    System,
    AccessMask,
    AllDirs,
    CaseSensitive,
    NoDot,
    NoDotDot,
    NoDotAndDotDot,
}

C++ type: QDir::Filter

C++ documentation:

This enum describes the filtering options available to QDir; e.g. for entryList() and entryInfoList(). The filter value is specified by combining values from the following list using the bitwise OR operator:

Functions that use Filter enum values to filter lists of files and directories will include symbolic links to files and directories unless you set the NoSymLinks value.

A default constructed QDir will not filter out files based on their permissions, so entryList() and entryInfoList() will return all files that are readable, writable, executable, or any combination of the three. This makes the default easy to write, and at the same time useful.

For example, setting the Readable, Writable, and Files flags allows all files to be listed for which the application has read access, write access or both. If the Dirs and Drives flags are also included in this combination then all drives, directories, all files that the application can read, write, or execute, and symlinks to such files/directories can be listed.

To retrieve the permissons for a directory, use the entryInfoList() function to get the associated QFileInfo objects and then use the QFileInfo::permissons() to obtain the permissions and ownership for each file.

The Filters type is a typedef for QFlags<Filter>. It stores an OR combination of Filter values.

Variants
NoFilter

C++ enum variant: NoFilter = -1

Dirs

List directories that match the filters. (C++ enum variant: Dirs = 1)

Files

List files. (C++ enum variant: Files = 2)

Drives

List disk drives (ignored under Unix). (C++ enum variant: Drives = 4)

AllEntries

List directories, files, drives and symlinks (this does not list broken symlinks unless you specify System). (C++ enum variant: AllEntries = 7)

NoSymLinks

Do not list symbolic links (ignored by operating systems that don't support symbolic links). (C++ enum variant: NoSymLinks = 8)

TypeMask

C++ enum variant: TypeMask = 15

Readable

List files for which the application has read access. The Readable value needs to be combined with Dirs or Files. (C++ enum variant: Readable = 16)

Writable

List files for which the application has write access. The Writable value needs to be combined with Dirs or Files. (C++ enum variant: Writable = 32)

Executable

List files for which the application has execute access. The Executable value needs to be combined with Dirs or Files. (C++ enum variant: Executable = 64)

PermissionMask

C++ enum variant: PermissionMask = 112

Modified

Only list files that have been modified (ignored on Unix). (C++ enum variant: Modified = 128)

Hidden

List hidden files (on Unix, files starting with a "."). (C++ enum variant: Hidden = 256)

System

List system files (on Unix, FIFOs, sockets and device files are included; on Windows, .lnk files are included) (C++ enum variant: System = 512)

AccessMask

C++ enum variant: AccessMask = 1008

AllDirs

List all directories; i.e. don't apply the filters to directory names. (C++ enum variant: AllDirs = 1024)

CaseSensitive

The filter should be case sensitive. (C++ enum variant: CaseSensitive = 2048)

NoDot

Do not list the special entry ".". (C++ enum variant: NoDot = 8192)

NoDotDot

Do not list the special entry "..". (C++ enum variant: NoDotDot = 16384)

NoDotAndDotDot

Do not list the special entries "." and "..". (C++ enum variant: NoDotAndDotDot = 24576)