Enum qt_core::dir::SortFlag

pub enum SortFlag {
    NoSort,
    Name,
    Time,
    Size,
    Unsorted,
    DirsFirst,
    Reversed,
    IgnoreCase,
    DirsLast,
    LocaleAware,
    Type,
}

C++ type: QDir::SortFlag

C++ documentation:

This enum describes the sort options available to QDir, e.g. for entryList() and entryInfoList(). The sort value is specified by OR-ing together values from the following list:

You can only specify one of the first four.

If you specify both DirsFirst and Reversed, directories are still put first, but in reverse order; the files will be listed after the directories, again in reverse order.

The SortFlags type is a typedef for QFlags<SortFlag>. It stores an OR combination of SortFlag values.

Variants
NoSort

Not sorted by default. (C++ enum variant: NoSort = -1)

Name

Sort by name. (C++ enum variant: Name = 0)

Time

Sort by time (modification time). (C++ enum variant: Time = 1)

Size

Sort by file size. (C++ enum variant: Size = 2)

Unsorted

This variant corresponds to multiple C++ enum variants with the same value:

Unsorted = 3: Do not sort.

SortByMask = 3

DirsFirst

Put the directories first, then the files. (C++ enum variant: DirsFirst = 4)

Reversed

Reverse the sort order. (C++ enum variant: Reversed = 8)

IgnoreCase

Sort case-insensitively. (C++ enum variant: IgnoreCase = 16)

DirsLast

Put the files first, then the directories. (C++ enum variant: DirsLast = 32)

LocaleAware

Sort items appropriately using the current locale settings. (C++ enum variant: LocaleAware = 64)

Type

Sort by file type (extension). (C++ enum variant: Type = 128)