Struct qt_core::dir::Dir

pub struct Dir {
    // some fields omitted
}

C++ type: QDir

C++ documentation:

The QDir class provides access to directory structures and their contents.

A QDir is used to manipulate path names, access information regarding paths and files, and manipulate the underlying file system. It can also be used to access Qt's resource system.

Qt uses "/" as a universal directory separator in the same way that "/" is used as a path separator in URLs. If you always use "/" as a directory separator, Qt will translate your paths to conform to the underlying operating system.

A QDir can point to a file using either a relative or an absolute path. Absolute paths begin with the directory separator (optionally preceded by a drive specification under Windows). Relative file names begin with a directory name or a file name and specify a path relative to the current directory.

Examples of absolute paths:

QDir("/home/user/Documents")
QDir("C:/Documents and Settings")

On Windows, the second example above will be translated to C:\Documents and Settings when used to access files.

Examples of relative paths:

QDir("images/landscape.png")

You can use the isRelative() or isAbsolute() functions to check if a QDir is using a relative or an absolute file path. Call makeAbsolute() to convert a relative QDir to an absolute one.

Methods

impl Dir

fn absolute_file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirAbsoluteFilePathArgs<'largs>

C++ method: QDir::absoluteFilePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn absolute_file_path(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn absolute_file_path(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QDir::absoluteFilePath(const QString& fileName) const

C++ documentation:

Returns the absolute path name of a file in the directory. Does not check if the file actually exists in the directory; but see exists(). Redundant multiple separators or "." and ".." directories in fileName are not removed (see cleanPath()).

See also relativeFilePath(), filePath(), and canonicalPath().

fn absolute_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirAbsolutePathArgs<'largs>

C++ method: QDir::absolutePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn absolute_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn absolute_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDir::absolutePath() const

C++ documentation:

Returns the absolute path (a path that starts with "/" or with a drive specification), which may contain symbolic links, but never contains redundant ".", ".." or multiple separators.

See also setPath(), canonicalPath(), exists(), cleanPath(), dirName(), and absoluteFilePath().

fn add_resource_search_path(path: &String)

C++ method: static void QDir::addResourceSearchPath(const QString& path)

C++ documentation:

Use QDir::addSearchPath() with a prefix instead.

Adds path to the search paths searched in to find resources that are not specified with an absolute path. The default search path is to search only in the root (:/).

See also The Qt Resource System.

fn add_search_path(prefix: &String, path: &String)

C++ method: static void QDir::addSearchPath(const QString& prefix, const QString& path)

C++ documentation:

Adds path to the search path for prefix.

This function was introduced in Qt 4.3.

See also setSearchPaths().

fn canonical_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirCanonicalPathArgs<'largs>

C++ method: QDir::canonicalPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn canonical_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn canonical_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDir::canonicalPath() const

C++ documentation:

Returns the canonical path, i.e. a path without symbolic links or redundant "." or ".." elements.

On systems that do not have symbolic links this function will always return the same string that absolutePath() returns. If the canonical path does not exist (normally due to dangling symbolic links) canonicalPath() returns an empty string.

Example:

QString bin = "/local/bin";         // where /local/bin is a symlink to /usr/bin
QDir binDir(bin);
QString canonicalBin = binDir.canonicalPath();
// canonicalBin now equals "/usr/bin"

QString ls = "/local/bin/ls"; // where ls is the executable "ls" QDir lsDir(ls); QString canonicalLs = lsDir.canonicalPath(); // canonicalLS now equals "/usr/bin/ls".
See also path(), absolutePath(), exists(), cleanPath(), dirName(), and absoluteFilePath().

fn cd(&mut self, dir_name: &String) -> bool

C++ method: bool QDir::cd(const QString& dirName)

C++ documentation:

Changes the QDir's directory to dirName.

Returns true if the new directory exists; otherwise returns false. Note that the logical cd() operation is not performed if the new directory does not exist.

Calling cd("..") is equivalent to calling cdUp().

See also cdUp(), isReadable(), exists(), and path().

fn cd_up(&mut self) -> bool

C++ method: bool QDir::cdUp()

C++ documentation:

Changes directory by moving one directory up from the QDir's current directory.

Returns true if the new directory exists; otherwise returns false. Note that the logical cdUp() operation is not performed if the new directory does not exist.

See also cd(), isReadable(), exists(), and path().

fn clean_path<Args>(args: Args) -> Args::ReturnType where Args: DirCleanPathArgs

C++ method: QDir::cleanPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn clean_path((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn clean_path((&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QDir::cleanPath(const QString& path)

C++ documentation:

Returns path with directory separators normalized (converted to "/") and redundant ones removed, and "."s and ".."s resolved (as far as possible).

Symbolic links are kept. This function does not return the canonical path, but rather the simplest version of the input. For example, "./local" becomes "local", "local/../bin" becomes "bin" and "/local/usr/../bin" becomes "/local/bin".

See also absolutePath() and canonicalPath().

fn count(&self) -> c_uint

C++ method: unsigned int QDir::count() const

C++ documentation:

Returns the total number of directories and files in the directory.

Equivalent to entryList().count().

See also operator[]() and entryList().

fn current<Args>(args: Args) -> Args::ReturnType where Args: DirCurrentArgs

C++ method: QDir::current

This is an overloaded function. Available variants:

Rust arguments:
1) fn current(cpp_utils::AsBox) -> cpp_utils::CppBox<::dir::Dir>
2) fn current(cpp_utils::AsStruct) -> ::dir::Dir
C++ method: static QDir QDir::current()

C++ documentation:

Returns the application's current directory.

The directory is constructed using the absolute path of the current directory, ensuring that its path() will be the same as its absolutePath().

See also currentPath(), setCurrent(), home(), root(), and temp().

fn current_path<Args>(args: Args) -> Args::ReturnType where Args: DirCurrentPathArgs

C++ method: QDir::currentPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_path(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn current_path(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QDir::currentPath()

C++ documentation:

Returns the absolute path of the application's current directory. The current directory is the last directory set with QDir::setCurrent() or, if that was never called, the directory at which this application was started at by the parent process.

See also current(), setCurrent(), homePath(), rootPath(), tempPath(), and QCoreApplication::applicationDirPath().

fn dir_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirDirNameArgs<'largs>

C++ method: QDir::dirName

This is an overloaded function. Available variants:

Rust arguments:
1) fn dir_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn dir_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDir::dirName() const

C++ documentation:

Returns the name of the directory; this is not the same as the path, e.g. a directory with the name "mail", might have the path "/var/spool/mail". If the directory has no name (e.g. it is the root directory) an empty string is returned.

No check is made to ensure that a directory with this name actually exists; but see exists().

See also path(), filePath(), absolutePath(), and absoluteFilePath().

fn drives<Args>(args: Args) -> Args::ReturnType where Args: DirDrivesArgs

C++ method: QDir::drives

This is an overloaded function. Available variants:

Rust arguments:
1) fn drives(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListFileInfoRef>
2) fn drives(cpp_utils::AsStruct) -> ::list::ListFileInfoRef
C++ method: static QList<QFileInfo> QDir::drives()

C++ documentation:

Returns a list of the root directories on this system.

On Windows this returns a list of QFileInfo objects containing "C:/", "D:/", etc. On other operating systems, it returns a list containing just one root directory (i.e. "/").

See also root() and rootPath().

fn entry_info_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirEntryInfoListArgs<'largs>

C++ method: QDir::entryInfoList

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn entry_info_list(&self, (::flags::Flags<::dir::Filter>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListFileInfoRef>
2) fn entry_info_list(&self, (::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListFileInfoRef>
3) fn entry_info_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListFileInfoRef>
4) fn entry_info_list(&self, (::flags::Flags<::dir::Filter>, cpp_utils::AsStruct)) -> ::list::ListFileInfoRef
5) fn entry_info_list(&self, (::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsStruct)) -> ::list::ListFileInfoRef
6) fn entry_info_list(&self, cpp_utils::AsStruct) -> ::list::ListFileInfoRef
C++ method: QList<QFileInfo> QDir::entryInfoList(QFlags<QDir::Filter> filters = ?, QFlags<QDir::SortFlag> sort = ?) const

C++ documentation:

This is an overloaded function.

Returns a list of QFileInfo objects for all the files and directories in the directory, ordered according to the name and attribute filters previously set with setNameFilters() and setFilter(), and sorted according to the flags set with setSorting().

The attribute filter and sorting specifications can be overridden using the filters and sort arguments.

Returns an empty list if the directory is unreadable, does not exist, or if nothing matches the specification.

See also entryList(), setNameFilters(), setSorting(), setFilter(), isReadable(), and exists().

Variant 2

Rust arguments:
1) fn entry_info_list(&self, (&::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListFileInfoRef>
2) fn entry_info_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListFileInfoRef>
3) fn entry_info_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListFileInfoRef>
4) fn entry_info_list(&self, (&::string_list::StringList, cpp_utils::AsStruct)) -> ::list::ListFileInfoRef
5) fn entry_info_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, cpp_utils::AsStruct)) -> ::list::ListFileInfoRef
6) fn entry_info_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsStruct)) -> ::list::ListFileInfoRef
C++ method: QList<QFileInfo> QDir::entryInfoList(const QStringList& nameFilters, QFlags<QDir::Filter> filters = ?, QFlags<QDir::SortFlag> sort = ?) const

C++ documentation:

Returns a list of QFileInfo objects for all the files and directories in the directory, ordered according to the name and attribute filters previously set with setNameFilters() and setFilter(), and sorted according to the flags set with setSorting().

The name filter, file attribute filter, and sorting specification can be overridden using the nameFilters, filters, and sort arguments.

Returns an empty list if the directory is unreadable, does not exist, or if nothing matches the specification.

See also entryList(), setNameFilters(), setSorting(), setFilter(), isReadable(), and exists().

fn entry_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirEntryListArgs<'largs>

C++ method: QDir::entryList

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn entry_list(&self, (::flags::Flags<::dir::Filter>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn entry_list(&self, (::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
3) fn entry_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
4) fn entry_list(&self, (::flags::Flags<::dir::Filter>, cpp_utils::AsStruct)) -> ::string_list::StringList
5) fn entry_list(&self, (::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsStruct)) -> ::string_list::StringList
6) fn entry_list(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QDir::entryList(QFlags<QDir::Filter> filters = ?, QFlags<QDir::SortFlag> sort = ?) const

C++ documentation:

This is an overloaded function.

Returns a list of the names of all the files and directories in the directory, ordered according to the name and attribute filters previously set with setNameFilters() and setFilter(), and sorted according to the flags set with setSorting().

The attribute filter and sorting specifications can be overridden using the filters and sort arguments.

Returns an empty list if the directory is unreadable, does not exist, or if nothing matches the specification.

Note: To list symlinks that point to non existing files, System must be passed to the filter.

See also entryInfoList(), setNameFilters(), setSorting(), and setFilter().

Variant 2

Rust arguments:
1) fn entry_list(&self, (&::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn entry_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
3) fn entry_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
4) fn entry_list(&self, (&::string_list::StringList, cpp_utils::AsStruct)) -> ::string_list::StringList
5) fn entry_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, cpp_utils::AsStruct)) -> ::string_list::StringList
6) fn entry_list(&self, (&::string_list::StringList, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsStruct)) -> ::string_list::StringList
C++ method: QStringList QDir::entryList(const QStringList& nameFilters, QFlags<QDir::Filter> filters = ?, QFlags<QDir::SortFlag> sort = ?) const

C++ documentation:

Returns a list of the names of all the files and directories in the directory, ordered according to the name and attribute filters previously set with setNameFilters() and setFilter(), and sorted according to the flags set with setSorting().

The name filter, file attribute filter, and sorting specification can be overridden using the nameFilters, filters, and sort arguments.

Returns an empty list if the directory is unreadable, does not exist, or if nothing matches the specification.

See also entryInfoList(), setNameFilters(), setSorting(), and setFilter().

fn exists<'largs, Args>(&'largs self, args: Args) -> bool where Args: DirExistsArgs<'largs>

C++ method: QDir::exists

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn exists(&self, ()) -> bool
C++ method: bool QDir::exists() const

C++ documentation:

This is an overloaded function.

Returns true if the directory exists; otherwise returns false. (If a file with the same name is found this function will return false).

The overload of this function that accepts an argument is used to test for the presence of files and directories within a directory.

See also QFileInfo::exists() and QFile::exists().

Variant 2

Rust arguments: fn exists(&self, &::string::String) -> bool
C++ method: bool QDir::exists(const QString& name) const

C++ documentation:

Returns true if the file called name exists; otherwise returns false.

Unless name contains an absolute file path, the file name is assumed to be relative to the directory itself, so this function is typically used to check for the presence of files within a directory.

See also QFileInfo::exists() and QFile::exists().

fn file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirFilePathArgs<'largs>

C++ method: QDir::filePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_path(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn file_path(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QDir::filePath(const QString& fileName) const

C++ documentation:

Returns the path name of a file in the directory. Does not check if the file actually exists in the directory; but see exists(). If the QDir is relative the returned path name will also be relative. Redundant multiple separators or "." and ".." directories in fileName are not removed (see cleanPath()).

See also dirName(), absoluteFilePath(), isRelative(), and canonicalPath().

fn filter(&self) -> Flags<Filter>

C++ method: QFlags<QDir::Filter> QDir::filter() const

C++ documentation:

Returns the value set by setFilter()

See also setFilter().

fn from_native_separators<Args>(args: Args) -> Args::ReturnType where Args: DirFromNativeSeparatorsArgs

C++ method: QDir::fromNativeSeparators

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_native_separators((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn from_native_separators((&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QDir::fromNativeSeparators(const QString& pathName)

C++ documentation:

Returns pathName using '/' as file separator. On Windows, for instance, fromNativeSeparators("c:\winnt\system32") returns "c:/winnt/system32".

The returned string may be the same as the argument on some operating systems, for example on Unix.

This function was introduced in Qt 4.2.

See also toNativeSeparators() and separator().

fn home<Args>(args: Args) -> Args::ReturnType where Args: DirHomeArgs

C++ method: QDir::home

This is an overloaded function. Available variants:

Rust arguments:
1) fn home(cpp_utils::AsBox) -> cpp_utils::CppBox<::dir::Dir>
2) fn home(cpp_utils::AsStruct) -> ::dir::Dir
C++ method: static QDir QDir::home()

C++ documentation:

Returns the user's home directory.

The directory is constructed using the absolute path of the home directory, ensuring that its path() will be the same as its absolutePath().

See homePath() for details.

See also drives(), current(), root(), and temp().

fn home_path<Args>(args: Args) -> Args::ReturnType where Args: DirHomePathArgs

C++ method: QDir::homePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn home_path(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn home_path(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QDir::homePath()

C++ documentation:

Returns the absolute path of the user's home directory.

Under Windows this function will return the directory of the current user's profile. Typically, this is:

C:/Documents and Settings/Username

Use the toNativeSeparators() function to convert the separators to the ones that are appropriate for the underlying operating system.

If the directory of the current user's profile does not exist or cannot be retrieved, the following alternatives will be checked (in the given order) until an existing and available path is found:

The path specified by the USERPROFILE environment variable.

The path formed by concatenating the HOMEDRIVE and HOMEPATH environment variables.

The path specified by the HOME environment variable.

The path returned by the rootPath() function (which uses the SystemDrive environment variable)

The C:/ directory.

Under non-Windows operating systems the HOME environment variable is used if it exists, otherwise the path returned by the rootPath().

See also home(), currentPath(), rootPath(), and tempPath().

fn is_absolute(&self) -> bool

C++ method: bool QDir::isAbsolute() const

C++ documentation:

Returns true if the directory's path is absolute; otherwise returns false. See isAbsolutePath().

See also isRelative(), makeAbsolute(), and cleanPath().

fn is_absolute_path(path: &String) -> bool

C++ method: static bool QDir::isAbsolutePath(const QString& path)

C++ documentation:

Returns true if path is absolute; returns false if it is relative.

See also isAbsolute(), isRelativePath(), makeAbsolute(), and cleanPath().

fn is_readable(&self) -> bool

C++ method: bool QDir::isReadable() const

C++ documentation:

Returns true if the directory is readable and we can open files by name; otherwise returns false.

Warning: A false value from this function is not a guarantee that files in the directory are not accessible.

See also QFileInfo::isReadable().

fn is_relative(&self) -> bool

C++ method: bool QDir::isRelative() const

C++ documentation:

Returns true if the directory path is relative; otherwise returns false. (Under Unix a path is relative if it does not start with a "/").

See also makeAbsolute(), isAbsolute(), isAbsolutePath(), and cleanPath().

fn is_relative_path(path: &String) -> bool

C++ method: static bool QDir::isRelativePath(const QString& path)

C++ documentation:

Returns true if path is relative; returns false if it is absolute.

See also isRelative(), isAbsolutePath(), and makeAbsolute().

fn is_root(&self) -> bool

C++ method: bool QDir::isRoot() const

C++ documentation:

Returns true if the directory is the root directory; otherwise returns false.

Note: If the directory is a symbolic link to the root directory this function returns false. If you want to test for this use canonicalPath(), e.g.

QDir dir("/tmp/root_link");
dir = dir.canonicalPath();
if (dir.isRoot())
    qWarning("It is a root link");

See also root() and rootPath().

fn make_absolute(&mut self) -> bool

C++ method: bool QDir::makeAbsolute()

C++ documentation:

Converts the directory path to an absolute path. If it is already absolute nothing happens. Returns true if the conversion succeeded; otherwise returns false.

See also isAbsolute(), isAbsolutePath(), isRelative(), and cleanPath().

fn match_<Args>(args: Args) -> bool where Args: DirMatchArgs

C++ method: QDir::match

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn match_((&::string::String, &::string::String)) -> bool
C++ method: static bool QDir::match(const QString& filter, const QString& fileName)

C++ documentation:

Returns true if the fileName matches the wildcard (glob) pattern filter; otherwise returns false. The filter may contain multiple patterns separated by spaces or semicolons. The matching is case insensitive.

See also QRegExp wildcard matching, QRegExp::exactMatch(), entryList(), and entryInfoList().

Variant 2

Rust arguments: fn match_((&::string_list::StringList, &::string::String)) -> bool
C++ method: static bool QDir::match(const QStringList& filters, const QString& fileName)

C++ documentation:

This is an overloaded function.

Returns true if the fileName matches any of the wildcard (glob) patterns in the list of filters; otherwise returns false. The matching is case insensitive.

See also QRegExp wildcard matching, QRegExp::exactMatch(), entryList(), and entryInfoList().

fn mkdir(&self, dir_name: &String) -> bool

C++ method: bool QDir::mkdir(const QString& dirName) const

C++ documentation:

Creates a sub-directory called dirName.

Returns true on success; otherwise returns false.

If the directory already exists when this function is called, it will return false.

See also rmdir().

fn mkpath(&self, dir_path: &String) -> bool

C++ method: bool QDir::mkpath(const QString& dirPath) const

C++ documentation:

Creates the directory path dirPath.

The function will create all parent directories necessary to create the directory.

Returns true if successful; otherwise returns false.

If the path already exists when this function is called, it will return true.

See also rmpath().

fn name_filters<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirNameFiltersArgs<'largs>

C++ method: QDir::nameFilters

This is an overloaded function. Available variants:

Rust arguments:
1) fn name_filters(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn name_filters(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QDir::nameFilters() const

C++ documentation:

Returns the string list set by setNameFilters()

See also setNameFilters().

fn name_filters_from_string<Args>(args: Args) -> Args::ReturnType where Args: DirNameFiltersFromStringArgs

C++ method: QDir::nameFiltersFromString

This is an overloaded function. Available variants:

Rust arguments:
1) fn name_filters_from_string((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn name_filters_from_string((&::string::String, cpp_utils::AsStruct)) -> ::string_list::StringList
C++ method: static QStringList QDir::nameFiltersFromString(const QString& nameFilter)

fn new<Args>(args: Args) -> Args::ReturnType where Args: DirNewArgs

C++ method: QDir::QDir

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::dir::Dir, cpp_utils::AsStruct)) -> ::dir::Dir
2) fn new((&::dir::Dir, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir::Dir>
C++ method: [constructor] void QDir::QDir(const QDir& arg1)

Variant 2

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::dir::Dir
2) fn new((&::string::String, cpp_utils::AsStruct)) -> ::dir::Dir
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::dir::Dir>
4) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir::Dir>
C++ method: [constructor] void QDir::QDir(const QString& path = ?)

Variant 3

Rust arguments:
1) fn new((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::dir::Dir
2) fn new((&::string::String, &::string::String, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsStruct)) -> ::dir::Dir
3) fn new((&::string::String, &::string::String, ::flags::Flags<::dir::SortFlag>, ::flags::Flags<::dir::Filter>, cpp_utils::AsStruct)) -> ::dir::Dir
4) fn new((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir::Dir>
5) fn new((&::string::String, &::string::String, ::flags::Flags<::dir::SortFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir::Dir>
6) fn new((&::string::String, &::string::String, ::flags::Flags<::dir::SortFlag>, ::flags::Flags<::dir::Filter>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir::Dir>
C++ method: [constructor] void QDir::QDir(const QString& path, const QString& nameFilter, QFlags<QDir::SortFlag> sort = ?, QFlags<QDir::Filter> filter = ?)

fn op_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Dir where Args: DirOpAssignArgs<'largs>

C++ method: QDir::operator=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_assign(&mut self, &'l1 ::dir::Dir) -> &'l0 mut ::dir::Dir
C++ method: QDir& QDir::operator=(const QDir& arg1)

C++ documentation:

Makes a copy of the dir object and assigns it to this QDir object.

Variant 2

Rust arguments: fn op_assign(&mut self, &'l1 ::string::String) -> &'l0 mut ::dir::Dir
C++ method: QDir& QDir::operator=(const QString& path)

fn op_eq(&self, dir: &Dir) -> bool

C++ method: bool QDir::operator==(const QDir& dir) const

C++ documentation:

Returns true if directory dir and this directory have the same path and their sort and filter settings are the same; otherwise returns false.

Example:

// The current directory is "/usr/local"
QDir d1("/usr/local/bin");
QDir d2("bin");
if (d1 == d2)
    qDebug("They're the same");

fn op_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirOpIndexArgs<'largs>

C++ method: QDir::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn op_index(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QDir::operator[](int arg1) const

C++ documentation:

Returns the file name at position pos in the list of file names. Equivalent to entryList().at(index). pos must be a valid index position in the list (i.e., 0 <= pos < count()).

See also count() and entryList().

fn op_neq(&self, dir: &Dir) -> bool

C++ method: bool QDir::operator!=(const QDir& dir) const

C++ documentation:

Returns true if directory dir and this directory have different paths or different sort or filter settings; otherwise returns false.

Example:

// The current directory is "/usr/local"
QDir d1("/usr/local/bin");
d1.setFilter(QDir::Executable);
QDir d2("bin");
if (d1 != d2)
    qDebug("They differ");

fn path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirPathArgs<'largs>

C++ method: QDir::path

This is an overloaded function. Available variants:

Rust arguments:
1) fn path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDir::path() const

C++ documentation:

Returns the path. This may contain symbolic links, but never contains redundant ".", ".." or multiple separators.

The returned path can be either absolute or relative (see setPath()).

See also setPath(), absolutePath(), exists(), cleanPath(), dirName(), absoluteFilePath(), toNativeSeparators(), and makeAbsolute().

fn refresh(&self)

C++ method: void QDir::refresh() const

C++ documentation:

Refreshes the directory information.

fn relative_file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirRelativeFilePathArgs<'largs>

C++ method: QDir::relativeFilePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn relative_file_path(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn relative_file_path(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QDir::relativeFilePath(const QString& fileName) const

C++ documentation:

Returns the path to fileName relative to the directory.

QDir dir("/home/bob");
QString s;

s = dir.relativeFilePath("images/file.jpg"); // s is "images/file.jpg" s = dir.relativeFilePath("/home/mary/file.txt"); // s is "../mary/file.txt"
See also absoluteFilePath(), filePath(), and canonicalPath().

fn remove(&mut self, file_name: &String) -> bool

C++ method: bool QDir::remove(const QString& fileName)

C++ documentation:

Removes the file, fileName.

Returns true if the file is removed successfully; otherwise returns false.

fn remove_recursively(&mut self) -> bool

C++ method: bool QDir::removeRecursively()

C++ documentation:

Removes the directory, including all its contents.

Returns true if successful, otherwise false.

If a file or directory cannot be removed, removeRecursively() keeps going and attempts to delete as many files and sub-directories as possible, then returns false.

If the directory was already removed, the method returns true (expected result already reached).

Note: this function is meant for removing a small application-internal directory (such as a temporary directory), but not user-visible directories. For user-visible operations, it is rather recommended to report errors more precisely to the user, to offer solutions in case of errors, to show progress during the deletion since it could take several minutes, etc.

This function was introduced in Qt 5.0.

fn rename(&mut self, old_name: &String, new_name: &String) -> bool

C++ method: bool QDir::rename(const QString& oldName, const QString& newName)

C++ documentation:

Renames a file or directory from oldName to newName, and returns true if successful; otherwise returns false.

On most file systems, rename() fails only if oldName does not exist, or if a file with the new name already exists. However, there are also other reasons why rename() can fail. For example, on at least one file system rename() fails if newName points to an open file.

If oldName is a file (not a directory) that can't be renamed right away, Qt will try to copy oldName to newName and remove oldName.

See also QFile::rename().

fn rmdir(&self, dir_name: &String) -> bool

C++ method: bool QDir::rmdir(const QString& dirName) const

C++ documentation:

Removes the directory specified by dirName.

The directory must be empty for rmdir() to succeed.

Returns true if successful; otherwise returns false.

See also mkdir().

fn rmpath(&self, dir_path: &String) -> bool

C++ method: bool QDir::rmpath(const QString& dirPath) const

C++ documentation:

Removes the directory path dirPath.

The function will remove all parent directories in dirPath, provided that they are empty. This is the opposite of mkpath(dirPath).

Returns true if successful; otherwise returns false.

See also mkpath().

fn root<Args>(args: Args) -> Args::ReturnType where Args: DirRootArgs

C++ method: QDir::root

This is an overloaded function. Available variants:

Rust arguments:
1) fn root(cpp_utils::AsBox) -> cpp_utils::CppBox<::dir::Dir>
2) fn root(cpp_utils::AsStruct) -> ::dir::Dir
C++ method: static QDir QDir::root()

C++ documentation:

Returns the root directory.

The directory is constructed using the absolute path of the root directory, ensuring that its path() will be the same as its absolutePath().

See rootPath() for details.

See also drives(), current(), home(), and temp().

fn root_path<Args>(args: Args) -> Args::ReturnType where Args: DirRootPathArgs

C++ method: QDir::rootPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn root_path(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn root_path(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QDir::rootPath()

C++ documentation:

Returns the absolute path of the root directory.

For Unix operating systems this returns "/". For Windows file systems this normally returns "c:/".

See also root(), drives(), currentPath(), homePath(), and tempPath().

fn search_paths<Args>(args: Args) -> Args::ReturnType where Args: DirSearchPathsArgs

C++ method: QDir::searchPaths

This is an overloaded function. Available variants:

Rust arguments:
1) fn search_paths((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn search_paths((&::string::String, cpp_utils::AsStruct)) -> ::string_list::StringList
C++ method: static QStringList QDir::searchPaths(const QString& prefix)

C++ documentation:

Returns the search paths for prefix.

This function was introduced in Qt 4.3.

See also setSearchPaths() and addSearchPath().

fn separator<Args>(args: Args) -> Args::ReturnType where Args: DirSeparatorArgs

C++ method: QDir::separator

This is an overloaded function. Available variants:

Rust arguments:
1) fn separator(cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn separator(cpp_utils::AsStruct) -> ::char::Char
C++ method: static QChar QDir::separator()

C++ documentation:

Returns the native directory separator: "/" under Unix and "" under Windows.

You do not need to use this function to build file paths. If you always use "/", Qt will translate your paths to conform to the underlying operating system. If you want to display paths to the user using their operating system's separator use toNativeSeparators().

fn set_current(path: &String) -> bool

C++ method: static bool QDir::setCurrent(const QString& path)

C++ documentation:

Sets the application's current working directory to path. Returns true if the directory was successfully changed; otherwise returns false.

See also current(), currentPath(), home(), root(), and temp().

fn set_filter(&mut self, filter: Flags<Filter>)

C++ method: void QDir::setFilter(QFlags<QDir::Filter> filter)

C++ documentation:

Sets the filter used by entryList() and entryInfoList() to filters. The filter is used to specify the kind of files that should be returned by entryList() and entryInfoList(). See QDir::Filter.

See also filter() and setNameFilters().

fn set_name_filters(&mut self, name_filters: &StringList)

C++ method: void QDir::setNameFilters(const QStringList& nameFilters)

C++ documentation:

Sets the name filters used by entryList() and entryInfoList() to the list of filters specified by nameFilters.

Each name filter is a wildcard (globbing) filter that understands and ? wildcards. (See QRegExp wildcard matching.)

For example, the following code sets three name filters on a QDir to ensure that only files with extensions typically used for C++ source files are listed:

    QStringList filters;
    filters << ".cpp" << "*.cxx" << "*.cc";
    dir.setNameFilters(filters);

See also nameFilters() and setFilter().

fn set_path(&mut self, path: &String)

C++ method: void QDir::setPath(const QString& path)

C++ documentation:

Sets the path of the directory to path. The path is cleaned of redundant ".", ".." and of multiple separators. No check is made to see whether a directory with this path actually exists; but you can check for yourself using exists().

The path can be either absolute or relative. Absolute paths begin with the directory separator "/" (optionally preceded by a drive specification under Windows). Relative file names begin with a directory name or a file name and specify a path relative to the current directory. An example of an absolute path is the string "/tmp/quartz", a relative path might look like "src/fatlib".

See also path(), absolutePath(), exists(), cleanPath(), dirName(), absoluteFilePath(), isRelative(), and makeAbsolute().

fn set_search_paths(prefix: &String, search_paths: &StringList)

C++ method: static void QDir::setSearchPaths(const QString& prefix, const QStringList& searchPaths)

C++ documentation:

Sets or replaces Qt's search paths for file names with the prefix prefix to searchPaths.

To specify a prefix for a file name, prepend the prefix followed by a single colon (e.g., "images:undo.png", "xmldocs:books.xml"). prefix can only contain letters or numbers (e.g., it cannot contain a colon, nor a slash).

Qt uses this search path to locate files with a known prefix. The search path entries are tested in order, starting with the first entry.

QDir::setSearchPaths("icons", QStringList(QDir::homePath() + "/images"));
QDir::setSearchPaths("docs", QStringList(":/embeddedDocuments"));
...
QPixmap pixmap("icons:undo.png"); // will look for undo.png in QDir::homePath() + "/images"
QFile file("docs:design.odf"); // will look in the :/embeddedDocuments resource path

File name prefix must be at least 2 characters long to avoid conflicts with Windows drive letters.

Search paths may contain paths to The Qt Resource System.

This function was introduced in Qt 4.3.

See also searchPaths().

fn set_sorting(&mut self, sort: Flags<SortFlag>)

C++ method: void QDir::setSorting(QFlags<QDir::SortFlag> sort)

C++ documentation:

Sets the sort order used by entryList() and entryInfoList().

The sort is specified by OR-ing values from the enum QDir::SortFlag.

See also sorting() and SortFlag.

fn sorting(&self) -> Flags<SortFlag>

C++ method: QFlags<QDir::SortFlag> QDir::sorting() const

C++ documentation:

Returns the value set by setSorting()

See also setSorting() and SortFlag.

fn swap(&mut self, other: &mut Dir)

C++ method: void QDir::swap(QDir& other)

C++ documentation:

Swaps this QDir instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn temp<Args>(args: Args) -> Args::ReturnType where Args: DirTempArgs

C++ method: QDir::temp

This is an overloaded function. Available variants:

Rust arguments:
1) fn temp(cpp_utils::AsBox) -> cpp_utils::CppBox<::dir::Dir>
2) fn temp(cpp_utils::AsStruct) -> ::dir::Dir
C++ method: static QDir QDir::temp()

C++ documentation:

Returns the system's temporary directory.

The directory is constructed using the absolute path of the temporary directory, ensuring that its path() will be the same as its absolutePath().

See tempPath() for details.

See also drives(), current(), home(), and root().

fn temp_path<Args>(args: Args) -> Args::ReturnType where Args: DirTempPathArgs

C++ method: QDir::tempPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn temp_path(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn temp_path(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QDir::tempPath()

C++ documentation:

Returns the absolute path of the system's temporary directory.

On Unix/Linux systems this is the path in the TMPDIR environment variable or /tmp if TMPDIR is not defined. On Windows this is usually the path in the TEMP or TMP environment variable. The path returned by this method doesn't end with a directory separator unless it is the root directory (of a drive).

See also temp(), currentPath(), homePath(), and rootPath().

fn to_native_separators<Args>(args: Args) -> Args::ReturnType where Args: DirToNativeSeparatorsArgs

C++ method: QDir::toNativeSeparators

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_native_separators((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_native_separators((&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QDir::toNativeSeparators(const QString& pathName)

C++ documentation:

Returns pathName with the '/' separators converted to separators that are appropriate for the underlying operating system.

On Windows, toNativeSeparators("c:/winnt/system32") returns "c:\winnt\system32".

The returned string may be the same as the argument on some operating systems, for example on Unix.

This function was introduced in Qt 4.2.

See also fromNativeSeparators() and separator().