Struct qt_core::dir_iterator::DirIterator

pub struct DirIterator {
    // some fields omitted
}

C++ type: QDirIterator

C++ documentation:

The QDirIterator class provides an iterator for directory entrylists.

You can use QDirIterator to navigate entries of a directory one at a time. It is similar to QDir::entryList() and QDir::entryInfoList(), but because it lists entries one at a time instead of all at once, it scales better and is more suitable for large directories. It also supports listing directory contents recursively, and following symbolic links. Unlike QDir::entryList(), QDirIterator does not support sorting.

The QDirIterator constructor takes a QDir or a directory as argument. After construction, the iterator is located before the first directory entry. Here's how to iterate over all the entries sequentially:

QDirIterator it("/etc", QDirIterator::Subdirectories);
while (it.hasNext()) {
    qDebug() << it.next();

// /etc/. // /etc/.. // /etc/X11 // /etc/X11/fs // ... }
The next() function returns the path to the next directory entry and advances the iterator. You can also call filePath() to get the current file path without advancing the iterator. The fileName() function returns only the name of the file, similar to how QDir::entryList() works. You can also call fileInfo() to get a QFileInfo for the current entry.

Unlike Qt's container iterators, QDirIterator is uni-directional (i.e., you cannot iterate directories in reverse order) and does not allow random access.

Methods

impl DirIterator

fn file_info<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirIteratorFileInfoArgs<'largs>

C++ method: QDirIterator::fileInfo

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_info(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::file_info::FileInfo>
2) fn file_info(&self, cpp_utils::AsStruct) -> ::file_info::FileInfo
C++ method: QFileInfo QDirIterator::fileInfo() const

C++ documentation:

Returns a QFileInfo for the current directory entry.

See also filePath() and fileName().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirIteratorFileNameArgs<'largs>

C++ method: QDirIterator::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDirIterator::fileName() const

C++ documentation:

Returns the file name for the current directory entry, without the path prepended.

This function is convenient when iterating a single directory. When using the QDirIterator::Subdirectories flag, you can use filePath() to get the full path.

See also filePath() and fileInfo().

fn file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirIteratorFilePathArgs<'largs>

C++ method: QDirIterator::filePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn file_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDirIterator::filePath() const

C++ documentation:

Returns the full file path for the current directory entry.

See also fileInfo() and fileName().

fn has_next(&self) -> bool

C++ method: bool QDirIterator::hasNext() const

C++ documentation:

Returns true if there is at least one more entry in the directory; otherwise, false is returned.

See also next(), fileName(), filePath(), and fileInfo().

fn new<Args>(args: Args) -> Args::ReturnType where Args: DirIteratorNewArgs

C++ method: QDirIterator::QDirIterator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::dir::Dir, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
2) fn new((&::dir::Dir, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
3) fn new((&::dir::Dir, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
4) fn new((&::dir::Dir, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
C++ method: [constructor] void QDirIterator::QDirIterator(const QDir& dir, QFlags<QDirIterator::IteratorFlag> flags = ?)

Variant 2

Rust arguments:
1) fn new((&::string::String, ::flags::Flags<::dir::Filter>, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
2) fn new((&::string::String, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
3) fn new((&::string::String, ::flags::Flags<::dir::Filter>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
4) fn new((&::string::String, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
C++ method: [constructor] void QDirIterator::QDirIterator(const QString& path, QFlags<QDir::Filter> filter, QFlags<QDirIterator::IteratorFlag> flags = ?)

Variant 3

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
2) fn new((&::string::String, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
3) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
4) fn new((&::string::String, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
C++ method: [constructor] void QDirIterator::QDirIterator(const QString& path, QFlags<QDirIterator::IteratorFlag> flags = ?)

Variant 4

Rust arguments:
1) fn new((&::string::String, &::string_list::StringList, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
2) fn new((&::string::String, &::string_list::StringList, ::flags::Flags<::dir::Filter>, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
3) fn new((&::string::String, &::string_list::StringList, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsStruct)) -> ::dir_iterator::DirIterator
4) fn new((&::string::String, &::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
5) fn new((&::string::String, &::string_list::StringList, ::flags::Flags<::dir::Filter>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
6) fn new((&::string::String, &::string_list::StringList, ::flags::Flags<::dir::Filter>, ::flags::Flags<::dir_iterator::IteratorFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::dir_iterator::DirIterator>
C++ method: [constructor] void QDirIterator::QDirIterator(const QString& path, const QStringList& nameFilters, QFlags<QDir::Filter> filters = ?, QFlags<QDirIterator::IteratorFlag> flags = ?)

fn next<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: DirIteratorNextArgs<'largs>

C++ method: QDirIterator::next

This is an overloaded function. Available variants:

Rust arguments:
1) fn next(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn next(&mut self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDirIterator::next()

C++ documentation:

Advances the iterator to the next entry, and returns the file path of this new entry. If hasNext() returns false, this function does nothing, and returns an empty QString.

You can call fileName() or filePath() to get the current entry file name or path, or fileInfo() to get a QFileInfo for the current entry.

See also hasNext(), fileName(), filePath(), and fileInfo().

fn path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: DirIteratorPathArgs<'largs>

C++ method: QDirIterator::path

This is an overloaded function. Available variants:

Rust arguments:
1) fn path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QDirIterator::path() const

C++ documentation:

Returns the base directory of the iterator.