Enum qt_core::event::Type

pub enum Type {
    None,
    Timer,
    MouseButtonPress,
    MouseButtonRelease,
    MouseButtonDblClick,
    MouseMove,
    KeyPress,
    KeyRelease,
    FocusIn,
    FocusOut,
    Enter,
    Leave,
    Paint,
    Move,
    Resize,
    Create,
    Destroy,
    Show,
    Hide,
    Close,
    Quit,
    ParentChange,
    ThreadChange,
    FocusAboutToChange,
    WindowActivate,
    WindowDeactivate,
    ShowToParent,
    HideToParent,
    Wheel,
    WindowTitleChange,
    WindowIconChange,
    ApplicationWindowIconChange,
    ApplicationFontChange,
    ApplicationLayoutDirectionChange,
    ApplicationPaletteChange,
    PaletteChange,
    Clipboard,
    Speech,
    MetaCall,
    SockAct,
    ShortcutOverride,
    DeferredDelete,
    DragEnter,
    DragMove,
    DragLeave,
    Drop,
    DragResponse,
    ChildAdded,
    ChildPolished,
    ChildRemoved,
    ShowWindowRequest,
    PolishRequest,
    Polish,
    LayoutRequest,
    UpdateRequest,
    UpdateLater,
    EmbeddingControl,
    ActivateControl,
    DeactivateControl,
    ContextMenu,
    InputMethod,
    TabletMove,
    LocaleChange,
    LanguageChange,
    LayoutDirectionChange,
    Style,
    TabletPress,
    TabletRelease,
    OkRequest,
    HelpRequest,
    IconDrag,
    FontChange,
    EnabledChange,
    ActivationChange,
    StyleChange,
    IconTextChange,
    ModifiedChange,
    WindowBlocked,
    WindowUnblocked,
    WindowStateChange,
    ReadOnlyChange,
    MouseTrackingChange,
    ToolTip,
    WhatsThis,
    StatusTip,
    ActionChanged,
    ActionAdded,
    ActionRemoved,
    FileOpen,
    Shortcut,
    WhatsThisClicked,
    ToolBarChange,
    ApplicationActivate,
    ApplicationDeactivate,
    QueryWhatsThis,
    EnterWhatsThisMode,
    LeaveWhatsThisMode,
    ZOrderChange,
    HoverEnter,
    HoverLeave,
    HoverMove,
    ParentAboutToChange,
    WinEventAct,
    AcceptDropsChange,
    ZeroTimerEvent,
    GraphicsSceneMouseMove,
    GraphicsSceneMousePress,
    GraphicsSceneMouseRelease,
    GraphicsSceneMouseDoubleClick,
    GraphicsSceneContextMenu,
    GraphicsSceneHoverEnter,
    GraphicsSceneHoverMove,
    GraphicsSceneHoverLeave,
    GraphicsSceneHelp,
    GraphicsSceneDragEnter,
    GraphicsSceneDragMove,
    GraphicsSceneDragLeave,
    GraphicsSceneDrop,
    GraphicsSceneWheel,
    KeyboardLayoutChange,
    DynamicPropertyChange,
    TabletEnterProximity,
    TabletLeaveProximity,
    NonClientAreaMouseMove,
    NonClientAreaMouseButtonPress,
    NonClientAreaMouseButtonRelease,
    NonClientAreaMouseButtonDblClick,
    MacSizeChange,
    ContentsRectChange,
    MacGLWindowChange,
    FutureCallOut,
    GraphicsSceneResize,
    GraphicsSceneMove,
    CursorChange,
    ToolTipChange,
    NetworkReplyUpdated,
    GrabMouse,
    UngrabMouse,
    GrabKeyboard,
    UngrabKeyboard,
    MacGLClearDrawable,
    StateMachineSignal,
    StateMachineWrapped,
    TouchBegin,
    TouchUpdate,
    TouchEnd,
    NativeGesture,
    Gesture,
    RequestSoftwareInputPanel,
    CloseSoftwareInputPanel,
    GestureOverride,
    WinIdChange,
    ScrollPrepare,
    Scroll,
    Expose,
    InputMethodQuery,
    OrientationChange,
    TouchCancel,
    ThemeChange,
    SockClose,
    PlatformPanel,
    StyleAnimationUpdate,
    ApplicationStateChange,
    WindowChangeInternal,
    ScreenChangeInternal,
    PlatformSurface,
    User,
    MaxUser,
}

C++ type: QEvent::Type

C++ documentation:

This enum type defines the valid event types in Qt. The event types and the specialized classes for each type are as follows:

User events should have values between User and MaxUser:

For convenience, you can use the registerEventType() function to register and reserve a custom event type for your application. Doing so will allow you to avoid accidentally re-using a custom event type already in use elsewhere in your application.

Variants
None

Not an event. (C++ enum variant: None = 0)

Timer

Regular timer events (QTimerEvent). (C++ enum variant: Timer = 1)

MouseButtonPress

Mouse press (QMouseEvent). (C++ enum variant: MouseButtonPress = 2)

MouseButtonRelease

Mouse release (QMouseEvent). (C++ enum variant: MouseButtonRelease = 3)

MouseButtonDblClick

Mouse press again (QMouseEvent). (C++ enum variant: MouseButtonDblClick = 4)

MouseMove

Mouse move (QMouseEvent). (C++ enum variant: MouseMove = 5)

KeyPress

Key press (QKeyEvent). (C++ enum variant: KeyPress = 6)

KeyRelease

Key release (QKeyEvent). (C++ enum variant: KeyRelease = 7)

FocusIn

Widget or Window gains keyboard focus (QFocusEvent). (C++ enum variant: FocusIn = 8)

FocusOut

Widget or Window loses keyboard focus (QFocusEvent). (C++ enum variant: FocusOut = 9)

Enter

Mouse enters widget's boundaries (QEnterEvent). (C++ enum variant: Enter = 10)

Leave

Mouse leaves widget's boundaries. (C++ enum variant: Leave = 11)

Paint

Screen update necessary (QPaintEvent). (C++ enum variant: Paint = 12)

Move

Widget's position changed (QMoveEvent). (C++ enum variant: Move = 13)

Resize

Widget's size changed (QResizeEvent). (C++ enum variant: Resize = 14)

Create

C++ enum variant: Create = 15

Destroy

C++ enum variant: Destroy = 16

Show

Widget was shown on screen (QShowEvent). (C++ enum variant: Show = 17)

Hide

Widget was hidden (QHideEvent). (C++ enum variant: Hide = 18)

Close

Widget was closed (QCloseEvent). (C++ enum variant: Close = 19)

Quit

C++ enum variant: Quit = 20

ParentChange

The widget parent has changed. (C++ enum variant: ParentChange = 21)

ThreadChange

The object is moved to another thread. This is the last event sent to this object in the previous thread. See QObject::moveToThread(). (C++ enum variant: ThreadChange = 22)

FocusAboutToChange

Widget or Window focus is about to change (QFocusEvent) (C++ enum variant: FocusAboutToChange = 23)

WindowActivate

Window was activated. (C++ enum variant: WindowActivate = 24)

WindowDeactivate

Window was deactivated. (C++ enum variant: WindowDeactivate = 25)

ShowToParent

A child widget has been shown. (C++ enum variant: ShowToParent = 26)

HideToParent

A child widget has been hidden. (C++ enum variant: HideToParent = 27)

Wheel

Mouse wheel rolled (QWheelEvent). (C++ enum variant: Wheel = 31)

WindowTitleChange

The window title has changed. (C++ enum variant: WindowTitleChange = 33)

WindowIconChange

The window's icon has changed. (C++ enum variant: WindowIconChange = 34)

ApplicationWindowIconChange

The application's icon has changed. (C++ enum variant: ApplicationWindowIconChange = 35)

ApplicationFontChange

The default application font has changed. (C++ enum variant: ApplicationFontChange = 36)

ApplicationLayoutDirectionChange

The default application layout direction has changed. (C++ enum variant: ApplicationLayoutDirectionChange = 37)

ApplicationPaletteChange

The default application palette has changed. (C++ enum variant: ApplicationPaletteChange = 38)

PaletteChange

Palette of the widget changed. (C++ enum variant: PaletteChange = 39)

Clipboard

The clipboard contents have changed. (C++ enum variant: Clipboard = 40)

Speech

C++ enum variant: Speech = 42

MetaCall

An asynchronous method invocation via QMetaObject::invokeMethod(). (C++ enum variant: MetaCall = 43)

SockAct

Socket activated, used to implement QSocketNotifier. (C++ enum variant: SockAct = 50)

ShortcutOverride

Key press in child, for overriding shortcut key handling (QKeyEvent). (C++ enum variant: ShortcutOverride = 51)

DeferredDelete

The object will be deleted after it has cleaned up (QDeferredDeleteEvent) (C++ enum variant: DeferredDelete = 52)

DragEnter

The cursor enters a widget during a drag and drop operation (QDragEnterEvent). (C++ enum variant: DragEnter = 60)

DragMove

A drag and drop operation is in progress (QDragMoveEvent). (C++ enum variant: DragMove = 61)

DragLeave

The cursor leaves a widget during a drag and drop operation (QDragLeaveEvent). (C++ enum variant: DragLeave = 62)

Drop

A drag and drop operation is completed (QDropEvent). (C++ enum variant: Drop = 63)

DragResponse

C++ enum variant: DragResponse = 64

ChildAdded

An object gets a child (QChildEvent). (C++ enum variant: ChildAdded = 68)

ChildPolished

A widget child gets polished (QChildEvent). (C++ enum variant: ChildPolished = 69)

ChildRemoved

An object loses a child (QChildEvent). (C++ enum variant: ChildRemoved = 71)

ShowWindowRequest

C++ enum variant: ShowWindowRequest = 73

PolishRequest

The widget should be polished. (C++ enum variant: PolishRequest = 74)

Polish

The widget is polished. (C++ enum variant: Polish = 75)

LayoutRequest

Widget layout needs to be redone. (C++ enum variant: LayoutRequest = 76)

UpdateRequest

The widget should be repainted. (C++ enum variant: UpdateRequest = 77)

UpdateLater

The widget should be queued to be repainted at a later time. (C++ enum variant: UpdateLater = 78)

EmbeddingControl

C++ enum variant: EmbeddingControl = 79

ActivateControl

C++ enum variant: ActivateControl = 80

DeactivateControl

C++ enum variant: DeactivateControl = 81

ContextMenu

Context popup menu (QContextMenuEvent). (C++ enum variant: ContextMenu = 82)

InputMethod

An input method is being used (QInputMethodEvent). (C++ enum variant: InputMethod = 83)

TabletMove

Wacom tablet move (QTabletEvent). (C++ enum variant: TabletMove = 87)

LocaleChange

The system locale has changed. (C++ enum variant: LocaleChange = 88)

LanguageChange

The application translation changed. (C++ enum variant: LanguageChange = 89)

LayoutDirectionChange

The direction of layouts changed. (C++ enum variant: LayoutDirectionChange = 90)

Style

C++ enum variant: Style = 91

TabletPress

Wacom tablet press (QTabletEvent). (C++ enum variant: TabletPress = 92)

TabletRelease

Wacom tablet release (QTabletEvent). (C++ enum variant: TabletRelease = 93)

OkRequest

Ok button in decoration pressed. Supported only for Windows CE. (C++ enum variant: OkRequest = 94)

HelpRequest

C++ enum variant: HelpRequest = 95

IconDrag

The main icon of a window has been dragged away (QIconDragEvent). (C++ enum variant: IconDrag = 96)

FontChange

Widget's font has changed. (C++ enum variant: FontChange = 97)

EnabledChange

Widget's enabled state has changed. (C++ enum variant: EnabledChange = 98)

ActivationChange

A widget's top-level window activation state has changed. (C++ enum variant: ActivationChange = 99)

StyleChange

Widget's style has been changed. (C++ enum variant: StyleChange = 100)

IconTextChange

Widget's icon text has been changed. (C++ enum variant: IconTextChange = 101)

ModifiedChange

Widgets modification state has been changed. (C++ enum variant: ModifiedChange = 102)

WindowBlocked

The window is blocked by a modal dialog. (C++ enum variant: WindowBlocked = 103)

WindowUnblocked

The window is unblocked after a modal dialog exited. (C++ enum variant: WindowUnblocked = 104)

WindowStateChange

The window's state (minimized, maximized or full-screen) has changed (QWindowStateChangeEvent). (C++ enum variant: WindowStateChange = 105)

ReadOnlyChange

Widget's read-only state has changed (since Qt 5.4). (C++ enum variant: ReadOnlyChange = 106)

MouseTrackingChange

The mouse tracking state has changed. (C++ enum variant: MouseTrackingChange = 109)

ToolTip

A tooltip was requested (QHelpEvent). (C++ enum variant: ToolTip = 110)

WhatsThis

The widget should reveal "What's This?" help (QHelpEvent). (C++ enum variant: WhatsThis = 111)

StatusTip

A status tip is requested (QStatusTipEvent). (C++ enum variant: StatusTip = 112)

ActionChanged

An action has been changed (QActionEvent). (C++ enum variant: ActionChanged = 113)

ActionAdded

A new action has been added (QActionEvent). (C++ enum variant: ActionAdded = 114)

ActionRemoved

An action has been removed (QActionEvent). (C++ enum variant: ActionRemoved = 115)

FileOpen

File open request (QFileOpenEvent). (C++ enum variant: FileOpen = 116)

Shortcut

Key press in child for shortcut key handling (QShortcutEvent). (C++ enum variant: Shortcut = 117)

WhatsThisClicked

A link in a widget's "What's This?" help was clicked. (C++ enum variant: WhatsThisClicked = 118)

ToolBarChange

The toolbar button is toggled on OS X. (C++ enum variant: ToolBarChange = 120)

ApplicationActivate

This variant corresponds to multiple C++ enum variants with the same value:

ApplicationActivate = 121: This enum has been deprecated. Use ApplicationStateChange instead.

ApplicationActivated = 121: This enum has been deprecated. Use ApplicationStateChange instead.

ApplicationDeactivate

This variant corresponds to multiple C++ enum variants with the same value:

ApplicationDeactivate = 122: This enum has been deprecated. Use ApplicationStateChange instead.

ApplicationDeactivated = 122

QueryWhatsThis

The widget should accept the event if it has "What's This?" help. (C++ enum variant: QueryWhatsThis = 123)

EnterWhatsThisMode

Send to toplevel widgets when the application enters "What's This?" mode. (C++ enum variant: EnterWhatsThisMode = 124)

LeaveWhatsThisMode

Send to toplevel widgets when the application leaves "What's This?" mode. (C++ enum variant: LeaveWhatsThisMode = 125)

ZOrderChange

The widget's z-order has changed. This event is never sent to top level windows. (C++ enum variant: ZOrderChange = 126)

HoverEnter

The mouse cursor enters a hover widget (QHoverEvent). (C++ enum variant: HoverEnter = 127)

HoverLeave

The mouse cursor leaves a hover widget (QHoverEvent). (C++ enum variant: HoverLeave = 128)

HoverMove

The mouse cursor moves inside a hover widget (QHoverEvent). (C++ enum variant: HoverMove = 129)

ParentAboutToChange

The widget parent is about to change. (C++ enum variant: ParentAboutToChange = 131)

WinEventAct

A Windows-specific activation event has occurred. (C++ enum variant: WinEventAct = 132)

AcceptDropsChange

C++ enum variant: AcceptDropsChange = 152

ZeroTimerEvent

C++ enum variant: ZeroTimerEvent = 154

GraphicsSceneMouseMove

Move mouse in a graphics scene (QGraphicsSceneMouseEvent). (C++ enum variant: GraphicsSceneMouseMove = 155)

GraphicsSceneMousePress

Mouse press in a graphics scene (QGraphicsSceneMouseEvent). (C++ enum variant: GraphicsSceneMousePress = 156)

GraphicsSceneMouseRelease

Mouse release in a graphics scene (QGraphicsSceneMouseEvent). (C++ enum variant: GraphicsSceneMouseRelease = 157)

GraphicsSceneMouseDoubleClick

Mouse press again (double click) in a graphics scene (QGraphicsSceneMouseEvent). (C++ enum variant: GraphicsSceneMouseDoubleClick = 158)

GraphicsSceneContextMenu

Context popup menu over a graphics scene (QGraphicsSceneContextMenuEvent). (C++ enum variant: GraphicsSceneContextMenu = 159)

GraphicsSceneHoverEnter

The mouse cursor enters a hover item in a graphics scene (QGraphicsSceneHoverEvent). (C++ enum variant: GraphicsSceneHoverEnter = 160)

GraphicsSceneHoverMove

The mouse cursor moves inside a hover item in a graphics scene (QGraphicsSceneHoverEvent). (C++ enum variant: GraphicsSceneHoverMove = 161)

GraphicsSceneHoverLeave

The mouse cursor leaves a hover item in a graphics scene (QGraphicsSceneHoverEvent). (C++ enum variant: GraphicsSceneHoverLeave = 162)

GraphicsSceneHelp

The user requests help for a graphics scene (QHelpEvent). (C++ enum variant: GraphicsSceneHelp = 163)

GraphicsSceneDragEnter

The cursor enters a graphics scene during a drag and drop operation (QGraphicsSceneDragDropEvent). (C++ enum variant: GraphicsSceneDragEnter = 164)

GraphicsSceneDragMove

A drag and drop operation is in progress over a scene (QGraphicsSceneDragDropEvent). (C++ enum variant: GraphicsSceneDragMove = 165)

GraphicsSceneDragLeave

The cursor leaves a graphics scene during a drag and drop operation (QGraphicsSceneDragDropEvent). (C++ enum variant: GraphicsSceneDragLeave = 166)

GraphicsSceneDrop

A drag and drop operation is completed over a scene (QGraphicsSceneDragDropEvent). (C++ enum variant: GraphicsSceneDrop = 167)

GraphicsSceneWheel

Mouse wheel rolled in a graphics scene (QGraphicsSceneWheelEvent). (C++ enum variant: GraphicsSceneWheel = 168)

KeyboardLayoutChange

The keyboard layout has changed. (C++ enum variant: KeyboardLayoutChange = 169)

DynamicPropertyChange

A dynamic property was added, changed, or removed from the object. (C++ enum variant: DynamicPropertyChange = 170)

TabletEnterProximity

Wacom tablet enter proximity event (QTabletEvent), sent to QApplication. (C++ enum variant: TabletEnterProximity = 171)

TabletLeaveProximity

Wacom tablet leave proximity event (QTabletEvent), sent to QApplication. (C++ enum variant: TabletLeaveProximity = 172)

NonClientAreaMouseMove

A mouse move occurred outside the client area. (C++ enum variant: NonClientAreaMouseMove = 173)

NonClientAreaMouseButtonPress

A mouse button press occurred outside the client area. (C++ enum variant: NonClientAreaMouseButtonPress = 174)

NonClientAreaMouseButtonRelease

A mouse button release occurred outside the client area. (C++ enum variant: NonClientAreaMouseButtonRelease = 175)

NonClientAreaMouseButtonDblClick

A mouse double click occurred outside the client area. (C++ enum variant: NonClientAreaMouseButtonDblClick = 176)

MacSizeChange

The user changed his widget sizes (OS X only). (C++ enum variant: MacSizeChange = 177)

ContentsRectChange

The margins of the widget's content rect changed. (C++ enum variant: ContentsRectChange = 178)

MacGLWindowChange

C++ enum variant: MacGLWindowChange = 179

FutureCallOut

C++ enum variant: FutureCallOut = 180

GraphicsSceneResize

Widget was resized (QGraphicsSceneResizeEvent). (C++ enum variant: GraphicsSceneResize = 181)

GraphicsSceneMove

Widget was moved (QGraphicsSceneMoveEvent). (C++ enum variant: GraphicsSceneMove = 182)

CursorChange

The widget's cursor has changed. (C++ enum variant: CursorChange = 183)

ToolTipChange

The widget's tooltip has changed. (C++ enum variant: ToolTipChange = 184)

NetworkReplyUpdated

C++ enum variant: NetworkReplyUpdated = 185

GrabMouse

Item gains mouse grab (QGraphicsItem only). (C++ enum variant: GrabMouse = 186)

UngrabMouse

Item loses mouse grab (QGraphicsItem only). (C++ enum variant: UngrabMouse = 187)

GrabKeyboard

Item gains keyboard grab (QGraphicsItem only). (C++ enum variant: GrabKeyboard = 188)

UngrabKeyboard

Item loses keyboard grab (QGraphicsItem only). (C++ enum variant: UngrabKeyboard = 189)

MacGLClearDrawable

C++ enum variant: MacGLClearDrawable = 191

StateMachineSignal

A signal delivered to a state machine (QStateMachine::SignalEvent). (C++ enum variant: StateMachineSignal = 192)

StateMachineWrapped

The event is a wrapper for, i.e., contains, another event (QStateMachine::WrappedEvent). (C++ enum variant: StateMachineWrapped = 193)

TouchBegin

Beginning of a sequence of touch-screen or track-pad events (QTouchEvent). (C++ enum variant: TouchBegin = 194)

TouchUpdate

Touch-screen event (QTouchEvent). (C++ enum variant: TouchUpdate = 195)

TouchEnd

End of touch-event sequence (QTouchEvent). (C++ enum variant: TouchEnd = 196)

NativeGesture

The system has detected a gesture (QNativeGestureEvent). (C++ enum variant: NativeGesture = 197)

Gesture

A gesture was triggered (QGestureEvent). (C++ enum variant: Gesture = 198)

RequestSoftwareInputPanel

A widget wants to open a software input panel (SIP). (C++ enum variant: RequestSoftwareInputPanel = 199)

CloseSoftwareInputPanel

A widget wants to close the software input panel (SIP). (C++ enum variant: CloseSoftwareInputPanel = 200)

GestureOverride

A gesture override was triggered (QGestureEvent). (C++ enum variant: GestureOverride = 202)

WinIdChange

The window system identifer for this native widget has changed. (C++ enum variant: WinIdChange = 203)

ScrollPrepare

The object needs to fill in its geometry information (QScrollPrepareEvent). (C++ enum variant: ScrollPrepare = 204)

Scroll

The object needs to scroll to the supplied position (QScrollEvent). (C++ enum variant: Scroll = 205)

Expose

Sent to a window when its on-screen contents are invalidated and need to be flushed from the backing store. (C++ enum variant: Expose = 206)

InputMethodQuery

A input method query event (QInputMethodQueryEvent) (C++ enum variant: InputMethodQuery = 207)

OrientationChange

The screens orientation has changes (QScreenOrientationChangeEvent). (C++ enum variant: OrientationChange = 208)

TouchCancel

Cancellation of touch-event sequence (QTouchEvent). (C++ enum variant: TouchCancel = 209)

ThemeChange

C++ enum variant: ThemeChange = 210

SockClose

C++ enum variant: SockClose = 211

PlatformPanel

A platform specific panel has been requested. (C++ enum variant: PlatformPanel = 212)

StyleAnimationUpdate

C++ enum variant: StyleAnimationUpdate = 213

ApplicationStateChange

The state of the application has changed. (C++ enum variant: ApplicationStateChange = 214)

WindowChangeInternal

C++ enum variant: WindowChangeInternal = 215

ScreenChangeInternal

C++ enum variant: ScreenChangeInternal = 216

PlatformSurface

A native platform surface has been created or is about to be destroyed (QPlatformSurfaceEvent). (C++ enum variant: PlatformSurface = 217)

User

User-defined event. (C++ enum variant: User = 1000)

MaxUser

Last user event ID. (C++ enum variant: MaxUser = 65535)