Enum qt_core::event_loop::ProcessEventsFlag

pub enum ProcessEventsFlag {
    AllEvents,
    ExcludeUserInputEvents,
    ExcludeSocketNotifiers,
    WaitForMoreEvents,
    X11ExcludeTimers,
    EventLoopExec,
    DialogExec,
}

C++ type: QEventLoop::ProcessEventsFlag

C++ documentation:

This enum controls the types of events processed by the processEvents() functions.

The ProcessEventsFlags type is a typedef for QFlags<ProcessEventsFlag>. It stores an OR combination of ProcessEventsFlag values.

See also processEvents().

Variants
AllEvents

All events. Note that DeferredDelete events are processed specially. See QObject::deleteLater() for more details. (C++ enum variant: AllEvents = 0)

ExcludeUserInputEvents

Do not process user input events, such as ButtonPress and KeyPress. Note that the events are not discarded; they will be delivered the next time processEvents() is called without the ExcludeUserInputEvents flag. (C++ enum variant: ExcludeUserInputEvents = 1)

ExcludeSocketNotifiers

Do not process socket notifier events. Note that the events are not discarded; they will be delivered the next time processEvents() is called without the ExcludeSocketNotifiers flag. (C++ enum variant: ExcludeSocketNotifiers = 2)

WaitForMoreEvents

Wait for events if no pending events are available. (C++ enum variant: WaitForMoreEvents = 4)

X11ExcludeTimers

C++ enum variant: X11ExcludeTimers = 8

EventLoopExec

C++ enum variant: EventLoopExec = 32

DialogExec

C++ enum variant: DialogExec = 64