Module qt_core::event_transition::overloading

Traits

EventTransitionAnimationsArgs
This trait represents a set of arguments accepted by EventTransition::animations method.

EventTransitionConnectArgs
This trait represents a set of arguments accepted by EventTransition::connect method.

EventTransitionConnectStaticArgs
This trait represents a set of arguments accepted by EventTransition::connect_static method.

EventTransitionDestroyedArgs
This trait represents a set of arguments accepted by EventTransition::destroyed method.

EventTransitionDisconnectArgs
This trait represents a set of arguments accepted by EventTransition::disconnect method.

EventTransitionDisconnectStaticArgs
This trait represents a set of arguments accepted by EventTransition::disconnect_static method.

EventTransitionDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by EventTransition::dynamic_property_names method.

EventTransitionNewArgs
This trait represents a set of arguments accepted by EventTransition::new method.

EventTransitionObjectNameArgs
This trait represents a set of arguments accepted by EventTransition::object_name method.

EventTransitionPropertyArgs
This trait represents a set of arguments accepted by EventTransition::property method.

EventTransitionStartTimerArgs
This trait represents a set of arguments accepted by EventTransition::start_timer method.

EventTransitionTargetStatesArgs
This trait represents a set of arguments accepted by EventTransition::target_states method.

EventTransitionTrArgs
This trait represents a set of arguments accepted by EventTransition::tr method.

EventTransitionTrUtf8Args
This trait represents a set of arguments accepted by EventTransition::tr_utf8 method.