Struct qt_core::exception::Exception

pub struct Exception {
    // some fields omitted
}

C++ type: QException

C++ documentation:

The QException class provides a base class for exceptions that can transferred across threads.

Qt Concurrent supports throwing and catching exceptions across thread boundaries, provided that the exception inherit from QException and implement two helper functions:

class MyException : public QException
{
public:
    void raise() const { throw this; }
    MyException clone() const { return new MyException(*this); }
};

QException subclasses must be thrown by value and caught by reference:

try  {
    QtConcurrent::blockingMap(list, throwFunction); // throwFunction throws MyException
} catch (MyException &e) {
    // handle exception
}

If you throw an exception that is not a subclass of QException, the Qt functions will throw a QUnhandledException in the receiver thread.

When using QFuture, transferred exceptions will be thrown when calling the following functions:

QFuture::waitForFinished()

QFuture::result()

QFuture::resultAt()

QFuture::results()

Methods

impl Exception

fn clone(&self) -> *mut Exception

C++ method: virtual QException* QException::clone() const

C++ documentation:

In your QException subclass, reimplement clone() like this:

MyException MyException::clone() const { return new MyException(this); }

fn raise(&self)

C++ method: virtual void QException::raise() const

C++ documentation:

In your QException subclass, reimplement raise() like this:

void MyException::raise() const { throw *this; }