Module qt_core::file::overloading

Traits

FileConnectArgs
This trait represents a set of arguments accepted by File::connect method.

FileConnectStaticArgs
This trait represents a set of arguments accepted by File::connect_static method.

FileDecodeNameArgs
This trait represents a set of arguments accepted by File::decode_name method.

FileDestroyedArgs
This trait represents a set of arguments accepted by File::destroyed method.

FileDisconnectArgs
This trait represents a set of arguments accepted by File::disconnect method.

FileDisconnectStaticArgs
This trait represents a set of arguments accepted by File::disconnect_static method.

FileDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by File::dynamic_property_names method.

FileEncodeNameArgs
This trait represents a set of arguments accepted by File::encode_name method.

FileErrorStringArgs
This trait represents a set of arguments accepted by File::error_string method.

FileFileNameArgs
This trait represents a set of arguments accepted by File::file_name method.

FileMapArgs
This trait represents a set of arguments accepted by File::map method.

FileNewArgs
This trait represents a set of arguments accepted by File::new method.

FileObjectNameArgs
This trait represents a set of arguments accepted by File::object_name method.

FileOpenArgs
This trait represents a set of arguments accepted by File::open method.

FilePeekArgs
This trait represents a set of arguments accepted by File::peek method.

FilePropertyArgs
This trait represents a set of arguments accepted by File::property method.

FileReadAllArgs
This trait represents a set of arguments accepted by File::read_all method.

FileReadArgs
This trait represents a set of arguments accepted by File::read method.

FileReadLineArgs
This trait represents a set of arguments accepted by File::read_line method.

FileReadLinkArgs
This trait represents a set of arguments accepted by File::read_link method.

FileReadLinkStaticArgs
This trait represents a set of arguments accepted by File::read_link_static method.

FileStartTimerArgs
This trait represents a set of arguments accepted by File::start_timer method.

FileSymLinkTargetArgs
This trait represents a set of arguments accepted by File::sym_link_target method.

FileSymLinkTargetStaticArgs
This trait represents a set of arguments accepted by File::sym_link_target_static method.

FileTrArgs
This trait represents a set of arguments accepted by File::tr method.

FileTrUtf8Args
This trait represents a set of arguments accepted by File::tr_utf8 method.

FileWriteArgs
This trait represents a set of arguments accepted by File::write method.