Enum qt_core::file_device::FileError

pub enum FileError {
    No,
    Read,
    Write,
    Fatal,
    Resource,
    Open,
    Abort,
    TimeOut,
    Unspecified,
    Remove,
    Rename,
    Position,
    Resize,
    Permissions,
    Copy,
}

C++ type: QFileDevice::FileError

C++ documentation:

This enum describes the errors that may be returned by the error() function.

Variants
No

No error occurred. (C++ enum variant: NoError = 0)

Read

An error occurred when reading from the file. (C++ enum variant: ReadError = 1)

Write

An error occurred when writing to the file. (C++ enum variant: WriteError = 2)

Fatal

A fatal error occurred. (C++ enum variant: FatalError = 3)

Resource

Out of resources (e.g., too many open files, out of memory, etc.) (C++ enum variant: ResourceError = 4)

Open

The file could not be opened. (C++ enum variant: OpenError = 5)

Abort

The operation was aborted. (C++ enum variant: AbortError = 6)

TimeOut

A timeout occurred. (C++ enum variant: TimeOutError = 7)

Unspecified

An unspecified error occurred. (C++ enum variant: UnspecifiedError = 8)

Remove

The file could not be removed. (C++ enum variant: RemoveError = 9)

Rename

The file could not be renamed. (C++ enum variant: RenameError = 10)

Position

The position in the file could not be changed. (C++ enum variant: PositionError = 11)

Resize

The file could not be resized. (C++ enum variant: ResizeError = 12)

Permissions

The file could not be accessed. (C++ enum variant: PermissionsError = 13)

Copy

The file could not be copied. (C++ enum variant: CopyError = 14)