Enum qt_core::file_device::Permission

pub enum Permission {
    ExeOther,
    WriteOther,
    ReadOther,
    ExeGroup,
    WriteGroup,
    ReadGroup,
    ExeUser,
    WriteUser,
    ReadUser,
    ExeOwner,
    WriteOwner,
    ReadOwner,
}

C++ type: QFileDevice::Permission

C++ documentation:

This enum is used by the permission() function to report the permissions and ownership of a file. The values may be OR-ed together to test multiple permissions and ownership values.

Warning: Because of differences in the platforms supported by Qt, the semantics of ReadUser, WriteUser and ExeUser are platform-dependent: On Unix, the rights of the owner of the file are returned and on Windows the rights of the current user are returned. This behavior might change in a future Qt version.

Note: On NTFS file systems, ownership and permissions checking is disabled by default for performance reasons. To enable it, include the following line:

extern Q_CORE_EXPORT int qt_ntfs_permission_lookup;

Permission checking is then turned on and off by incrementing and decrementing qt_ntfs_permission_lookup by 1.

qt_ntfs_permission_lookup++; // turn checking on
qt_ntfs_permission_lookup--; // turn it off again

The Permissions type is a typedef for QFlags<Permission>. It stores an OR combination of Permission values.

Variants
ExeOther

The file is executable by anyone. (C++ enum variant: ExeOther = 1)

WriteOther

The file is writable by anyone. (C++ enum variant: WriteOther = 2)

ReadOther

The file is readable by anyone. (C++ enum variant: ReadOther = 4)

ExeGroup

The file is executable by the group. (C++ enum variant: ExeGroup = 16)

WriteGroup

The file is writable by the group. (C++ enum variant: WriteGroup = 32)

ReadGroup

The file is readable by the group. (C++ enum variant: ReadGroup = 64)

ExeUser

The file is executable by the user. (C++ enum variant: ExeUser = 256)

WriteUser

The file is writable by the user. (C++ enum variant: WriteUser = 512)

ReadUser

The file is readable by the user. (C++ enum variant: ReadUser = 1024)

ExeOwner

The file is executable by the owner of the file. (C++ enum variant: ExeOwner = 4096)

WriteOwner

The file is writable by the owner of the file. (C++ enum variant: WriteOwner = 8192)

ReadOwner

The file is readable by the owner of the file. (C++ enum variant: ReadOwner = 16384)