Module qt_core::file_device::overloading

Traits

FileDeviceConnectArgs
This trait represents a set of arguments accepted by FileDevice::connect method.

FileDeviceConnectStaticArgs
This trait represents a set of arguments accepted by FileDevice::connect_static method.

FileDeviceDestroyedArgs
This trait represents a set of arguments accepted by FileDevice::destroyed method.

FileDeviceDisconnectArgs
This trait represents a set of arguments accepted by FileDevice::disconnect method.

FileDeviceDisconnectStaticArgs
This trait represents a set of arguments accepted by FileDevice::disconnect_static method.

FileDeviceDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by FileDevice::dynamic_property_names method.

FileDeviceErrorStringArgs
This trait represents a set of arguments accepted by FileDevice::error_string method.

FileDeviceFileNameArgs
This trait represents a set of arguments accepted by FileDevice::file_name method.

FileDeviceMapArgs
This trait represents a set of arguments accepted by FileDevice::map method.

FileDeviceObjectNameArgs
This trait represents a set of arguments accepted by FileDevice::object_name method.

FileDevicePeekArgs
This trait represents a set of arguments accepted by FileDevice::peek method.

FileDevicePropertyArgs
This trait represents a set of arguments accepted by FileDevice::property method.

FileDeviceReadAllArgs
This trait represents a set of arguments accepted by FileDevice::read_all method.

FileDeviceReadArgs
This trait represents a set of arguments accepted by FileDevice::read method.

FileDeviceReadLineArgs
This trait represents a set of arguments accepted by FileDevice::read_line method.

FileDeviceStartTimerArgs
This trait represents a set of arguments accepted by FileDevice::start_timer method.

FileDeviceTrArgs
This trait represents a set of arguments accepted by FileDevice::tr method.

FileDeviceTrUtf8Args
This trait represents a set of arguments accepted by FileDevice::tr_utf8 method.

FileDeviceWriteArgs
This trait represents a set of arguments accepted by FileDevice::write method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.