Module qt_core::file_info::overloading

Traits

FileInfoAbsoluteDirArgs
This trait represents a set of arguments accepted by FileInfo::absolute_dir method.

FileInfoAbsoluteFilePathArgs
This trait represents a set of arguments accepted by FileInfo::absolute_file_path method.

FileInfoAbsolutePathArgs
This trait represents a set of arguments accepted by FileInfo::absolute_path method.

FileInfoBaseNameArgs
This trait represents a set of arguments accepted by FileInfo::base_name method.

FileInfoBundleNameArgs
This trait represents a set of arguments accepted by FileInfo::bundle_name method.

FileInfoCanonicalFilePathArgs
This trait represents a set of arguments accepted by FileInfo::canonical_file_path method.

FileInfoCanonicalPathArgs
This trait represents a set of arguments accepted by FileInfo::canonical_path method.

FileInfoCompleteBaseNameArgs
This trait represents a set of arguments accepted by FileInfo::complete_base_name method.

FileInfoCompleteSuffixArgs
This trait represents a set of arguments accepted by FileInfo::complete_suffix method.

FileInfoCreatedArgs
This trait represents a set of arguments accepted by FileInfo::created method.

FileInfoDirArgs
This trait represents a set of arguments accepted by FileInfo::dir method.

FileInfoFileNameArgs
This trait represents a set of arguments accepted by FileInfo::file_name method.

FileInfoFilePathArgs
This trait represents a set of arguments accepted by FileInfo::file_path method.

FileInfoGroupArgs
This trait represents a set of arguments accepted by FileInfo::group method.

FileInfoLastModifiedArgs
This trait represents a set of arguments accepted by FileInfo::last_modified method.

FileInfoLastReadArgs
This trait represents a set of arguments accepted by FileInfo::last_read method.

FileInfoNewArgs
This trait represents a set of arguments accepted by FileInfo::new method.

FileInfoOwnerArgs
This trait represents a set of arguments accepted by FileInfo::owner method.

FileInfoPathArgs
This trait represents a set of arguments accepted by FileInfo::path method.

FileInfoReadLinkArgs
This trait represents a set of arguments accepted by FileInfo::read_link method.

FileInfoSetFileArgs
This trait represents a set of arguments accepted by FileInfo::set_file method.

FileInfoSuffixArgs
This trait represents a set of arguments accepted by FileInfo::suffix method.

FileInfoSymLinkTargetArgs
This trait represents a set of arguments accepted by FileInfo::sym_link_target method.