Struct qt_core::file_info::FileInfo

pub struct FileInfo {
    // some fields omitted
}

C++ type: QFileInfo

C++ documentation:

The QFileInfo class provides system-independent file information.

QFileInfo provides information about a file's name and position (path) in the file system, its access rights and whether it is a directory or symbolic link, etc. The file's size and last modified/read times are also available. QFileInfo can also be used to obtain information about a Qt resource.

A QFileInfo can point to a file with either a relative or an absolute file path. Absolute file paths begin with the directory separator "/" (or with a drive specification on Windows). Relative file names begin with a directory name or a file name and specify a path relative to the current working directory. An example of an absolute path is the string "/tmp/quartz". A relative path might look like "src/fatlib". You can use the function isRelative() to check whether a QFileInfo is using a relative or an absolute file path. You can call the function makeAbsolute() to convert a relative QFileInfo's path to an absolute path.

The file that the QFileInfo works on is set in the constructor or later with setFile(). Use exists() to see if the file exists and size() to get its size.

The file's type is obtained with isFile(), isDir() and isSymLink(). The symLinkTarget() function provides the name of the file the symlink points to.

On Unix (including OS X and iOS), the symlink has the same size() has the file it points to, because Unix handles symlinks transparently; similarly, opening a symlink using QFile effectively opens the link's target. For example:

#ifdef Q_OS_UNIX

QFileInfo info1("/home/bob/bin/untabify"); info1.isSymLink(); // returns true info1.absoluteFilePath(); // returns "/home/bob/bin/untabify" info1.size(); // returns 56201 info1.symLinkTarget(); // returns "/opt/pretty++/bin/untabify"

QFileInfo info2(info1.symLinkTarget()); info2.isSymLink(); // returns false info2.absoluteFilePath(); // returns "/opt/pretty++/bin/untabify" info2.size(); // returns 56201

#endif
On Windows, symlinks (shortcuts) are .lnk files. The reported size() is that of the symlink (not the link's target), and opening a symlink using QFile opens the .lnk file. For example:

#ifdef Q_OS_WIN

QFileInfo info1("C:\Documents and Settings\Bob\untabify.lnk"); info1.isSymLink(); // returns true info1.absoluteFilePath(); // returns "C:/Documents and Settings/Bob/untabify.lnk" info1.size(); // returns 743 info1.symLinkTarget(); // returns "C:/Pretty++/untabify"

QFileInfo info2(info1.symLinkTarget()); info2.isSymLink(); // returns false info2.absoluteFilePath(); // returns "C:/Pretty++/untabify" info2.size(); // returns 63942

#endif
Elements of the file's name can be extracted with path() and fileName(). The fileName()'s parts can be extracted with baseName(), suffix() or completeSuffix(). QFileInfo objects to directories created by Qt classes will not have a trailing file separator. If you wish to use trailing separators in your own file info objects, just append one to the file name given to the constructors or setFile().

The file's dates are returned by created(), lastModified() and lastRead(). Information about the file's access permissions is obtained with isReadable(), isWritable() and isExecutable(). The file's ownership is available from owner(), ownerId(), group() and groupId(). You can examine a file's permissions and ownership in a single statement using the permission() function.

Note: On NTFS file systems, ownership and permissions checking is disabled by default for performance reasons. To enable it, include the following line:

extern Q_CORE_EXPORT int qt_ntfs_permission_lookup;

Permission checking is then turned on and off by incrementing and decrementing qt_ntfs_permission_lookup by 1.

qt_ntfs_permission_lookup++; // turn checking on
qt_ntfs_permission_lookup--; // turn it off again

Methods

impl FileInfo

fn absolute_dir<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoAbsoluteDirArgs<'largs>

C++ method: QFileInfo::absoluteDir

This is an overloaded function. Available variants:

Rust arguments:
1) fn absolute_dir(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::dir::Dir>
2) fn absolute_dir(&self, cpp_utils::AsStruct) -> ::dir::Dir
C++ method: QDir QFileInfo::absoluteDir() const

C++ documentation:

Returns the file's absolute path as a QDir object.

See also dir(), filePath(), fileName(), and isRelative().

fn absolute_file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoAbsoluteFilePathArgs<'largs>

C++ method: QFileInfo::absoluteFilePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn absolute_file_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn absolute_file_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::absoluteFilePath() const

C++ documentation:

Returns an absolute path including the file name.

The absolute path name consists of the full path and the file name. On Unix this will always begin with the root, '/', directory. On Windows this will always begin 'D:/' where D is a drive letter, except for network shares that are not mapped to a drive letter, in which case the path will begin '//sharename/'. QFileInfo will uppercase drive letters. Note that QDir does not do this. The code snippet below shows this.

    QFileInfo fi("c:/temp/foo"); => fi.absoluteFilePath() => "C:/temp/foo"

This function returns the same as filePath(), unless isRelative() is true. In contrast to canonicalFilePath(), symbolic links or redundant "." or ".." elements are not necessarily removed.

Warning: If filePath() is empty the behavior of this function is undefined.

See also filePath(), canonicalFilePath(), and isRelative().

fn absolute_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoAbsolutePathArgs<'largs>

C++ method: QFileInfo::absolutePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn absolute_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn absolute_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::absolutePath() const

C++ documentation:

Returns a file's path absolute path. This doesn't include the file name.

On Unix the absolute path will always begin with the root, '/', directory. On Windows this will always begin 'D:/' where D is a drive letter, except for network shares that are not mapped to a drive letter, in which case the path will begin '//sharename/'.

In contrast to canonicalPath() symbolic links or redundant "." or ".." elements are not necessarily removed.

Warning: If filePath() is empty the behavior of this function is undefined.

See also absoluteFilePath(), path(), canonicalPath(), fileName(), and isRelative().

fn base_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoBaseNameArgs<'largs>

C++ method: QFileInfo::baseName

This is an overloaded function. Available variants:

Rust arguments:
1) fn base_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn base_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::baseName() const

C++ documentation:

Returns the base name of the file without the path.

The base name consists of all characters in the file up to (but not including) the first '.' character.

Example:

QFileInfo fi("/tmp/archive.tar.gz");
QString base = fi.baseName();  // base = "archive"

The base name of a file is computed equally on all platforms, independent of file naming conventions (e.g., ".bashrc" on Unix has an empty base name, and the suffix is "bashrc").

See also fileName(), suffix(), completeSuffix(), and completeBaseName().

fn bundle_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoBundleNameArgs<'largs>

C++ method: QFileInfo::bundleName

This is an overloaded function. Available variants:

Rust arguments:
1) fn bundle_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn bundle_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::bundleName() const

C++ documentation:

Returns the name of the bundle.

On OS X and iOS this returns the proper localized name for a bundle if the path isBundle(). On all other platforms an empty QString is returned.

Example:

QFileInfo fi("/Applications/Safari.app");
QString bundle = fi.bundleName();                // name = "Safari"

This function was introduced in Qt 4.3.

See also isBundle(), filePath(), baseName(), and suffix().

fn caching(&self) -> bool

C++ method: bool QFileInfo::caching() const

C++ documentation:

Returns true if caching is enabled; otherwise returns false.

See also setCaching() and refresh().

fn canonical_file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoCanonicalFilePathArgs<'largs>

C++ method: QFileInfo::canonicalFilePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn canonical_file_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn canonical_file_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::canonicalFilePath() const

C++ documentation:

Returns the canonical path including the file name, i.e. an absolute path without symbolic links or redundant "." or ".." elements.

If the file does not exist, canonicalFilePath() returns an empty string.

See also filePath(), absoluteFilePath(), and dir().

fn canonical_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoCanonicalPathArgs<'largs>

C++ method: QFileInfo::canonicalPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn canonical_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn canonical_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::canonicalPath() const

C++ documentation:

Returns the file's path canonical path (excluding the file name), i.e. an absolute path without symbolic links or redundant "." or ".." elements.

If the file does not exist, canonicalPath() returns an empty string.

See also path() and absolutePath().

fn complete_base_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoCompleteBaseNameArgs<'largs>

C++ method: QFileInfo::completeBaseName

This is an overloaded function. Available variants:

Rust arguments:
1) fn complete_base_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn complete_base_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::completeBaseName() const

C++ documentation:

Returns the complete base name of the file without the path.

The complete base name consists of all characters in the file up to (but not including) the last '.' character.

Example:

QFileInfo fi("/tmp/archive.tar.gz");
QString base = fi.completeBaseName();  // base = "archive.tar"

See also fileName(), suffix(), completeSuffix(), and baseName().

fn complete_suffix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoCompleteSuffixArgs<'largs>

C++ method: QFileInfo::completeSuffix

This is an overloaded function. Available variants:

Rust arguments:
1) fn complete_suffix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn complete_suffix(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::completeSuffix() const

C++ documentation:

Returns the complete suffix of the file.

The complete suffix consists of all characters in the file after (but not including) the first '.'.

Example:

QFileInfo fi("/tmp/archive.tar.gz");
QString ext = fi.completeSuffix();  // ext = "tar.gz"

See also fileName(), suffix(), baseName(), and completeBaseName().

fn created<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoCreatedArgs<'largs>

C++ method: QFileInfo::created

This is an overloaded function. Available variants:

Rust arguments:
1) fn created(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn created(&self, cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: QDateTime QFileInfo::created() const

C++ documentation:

Returns the date and time when the file was created.

On most Unix systems, this function returns the time of the last status change. A status change occurs when the file is created, but it also occurs whenever the user writes or sets inode information (for example, changing the file permissions).

If neither creation time nor "last status change" time are not available, returns the same as lastModified().

See also lastModified() and lastRead().

fn dir<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoDirArgs<'largs>

C++ method: QFileInfo::dir

This is an overloaded function. Available variants:

Rust arguments:
1) fn dir(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::dir::Dir>
2) fn dir(&self, cpp_utils::AsStruct) -> ::dir::Dir
C++ method: QDir QFileInfo::dir() const

C++ documentation:

Returns the path of the object's parent directory as a QDir object.

Note: The QDir returned always corresponds to the object's parent directory, even if the QFileInfo represents a directory.

For each of the following, dir() returns a QDir for "~/examples/191697".

    QFileInfo fileInfo1("~/examples/191697/.");
    QFileInfo fileInfo2("~/examples/191697/..");
    QFileInfo fileInfo3("~/examples/191697/main.cpp");

For each of the following, dir() returns a QDir for ".".

    QFileInfo fileInfo4(".");
    QFileInfo fileInfo5("..");
    QFileInfo fileInfo6("main.cpp");

See also absolutePath(), filePath(), fileName(), isRelative(), and absoluteDir().

fn exists(&self) -> bool

C++ method: bool QFileInfo::exists() const

C++ documentation:

Returns true if the file exists; otherwise returns false.

Note: If the file is a symlink that points to a non-existing file, false is returned.

fn exists_static(file: &String) -> bool

C++ method: static bool QFileInfo::exists(const QString& file)

C++ documentation:

Returns true if the file exists; otherwise returns false.

Note: If file is a symlink that points to a non-existing file, false is returned.

Note: Using this function is faster than using QFileInfo(file).exists() for file system access.

This function was introduced in Qt 5.2.

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoFileNameArgs<'largs>

C++ method: QFileInfo::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::fileName() const

C++ documentation:

Returns the name of the file, excluding the path.

Example:

QFileInfo fi("/tmp/archive.tar.gz");
QString name = fi.fileName();                // name = "archive.tar.gz"

Note that, if this QFileInfo object is given a path ending in a slash, the name of the file is considered empty.

See also isRelative(), filePath(), baseName(), and suffix().

fn file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoFilePathArgs<'largs>

C++ method: QFileInfo::filePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn file_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::filePath() const

C++ documentation:

Returns the file name, including the path (which may be absolute or relative).

See also absoluteFilePath(), canonicalFilePath(), and isRelative().

fn group<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoGroupArgs<'largs>

C++ method: QFileInfo::group

This is an overloaded function. Available variants:

Rust arguments:
1) fn group(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn group(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::group() const

C++ documentation:

Returns the group of the file. On Windows, on systems where files do not have groups, or if an error occurs, an empty string is returned.

This function can be time consuming under Unix (in the order of milliseconds).

See also groupId(), owner(), and ownerId().

fn group_id(&self) -> c_uint

C++ method: unsigned int QFileInfo::groupId() const

C++ documentation:

Returns the id of the group the file belongs to.

On Windows and on systems where files do not have groups this function always returns (uint) -2.

See also group(), owner(), and ownerId().

fn is_absolute(&self) -> bool

C++ method: bool QFileInfo::isAbsolute() const

C++ documentation:

Returns true if the file path name is absolute, otherwise returns false if the path is relative.

See also isRelative().

fn is_bundle(&self) -> bool

C++ method: bool QFileInfo::isBundle() const

C++ documentation:

Returns true if this object points to a bundle or to a symbolic link to a bundle on OS X and iOS; otherwise returns false.

This function was introduced in Qt 4.3.

See also isDir(), isSymLink(), and isFile().

fn is_dir(&self) -> bool

C++ method: bool QFileInfo::isDir() const

C++ documentation:

Returns true if this object points to a directory or to a symbolic link to a directory; otherwise returns false.

See also isFile(), isSymLink(), and isBundle().

fn is_executable(&self) -> bool

C++ method: bool QFileInfo::isExecutable() const

C++ documentation:

Returns true if the file is executable; otherwise returns false.

See also isReadable(), isWritable(), and permission().

fn is_file(&self) -> bool

C++ method: bool QFileInfo::isFile() const

C++ documentation:

Returns true if this object points to a file or to a symbolic link to a file. Returns false if the object points to something which isn't a file, such as a directory.

See also isDir(), isSymLink(), and isBundle().

fn is_hidden(&self) -> bool

C++ method: bool QFileInfo::isHidden() const

C++ documentation:

Returns true if this is a `hidden' file; otherwise returns false.

Note: This function returns true for the special entries "." and ".." on Unix, even though QDir::entryList threats them as shown.

fn is_native_path(&self) -> bool

C++ method: bool QFileInfo::isNativePath() const

C++ documentation:

Returns true if the file path can be used directly with native APIs. Returns false if the file is otherwise supported by a virtual file system inside Qt, such as the Qt Resource System.

Note: Native paths may still require conversion of path separators and character encoding, depending on platform and input requirements of the native API.

This function was introduced in Qt 5.0.

See also QDir::toNativeSeparators(), QFile::encodeName(), filePath(), absoluteFilePath(), and canonicalFilePath().

fn is_readable(&self) -> bool

C++ method: bool QFileInfo::isReadable() const

C++ documentation:

Returns true if the user can read the file; otherwise returns false.

See also isWritable(), isExecutable(), and permission().

fn is_relative(&self) -> bool

C++ method: bool QFileInfo::isRelative() const

C++ documentation:

Returns true if the file path name is relative, otherwise returns false if the path is absolute (e.g. under Unix a path is absolute if it begins with a "/").

See also isAbsolute().

fn is_root(&self) -> bool

C++ method: bool QFileInfo::isRoot() const

C++ documentation:

Returns true if the object points to a directory or to a symbolic link to a directory, and that directory is the root directory; otherwise returns false.

fn is_sym_link(&self) -> bool

C++ method: bool QFileInfo::isSymLink() const

C++ documentation:

Returns true if this object points to a symbolic link (or to a shortcut on Windows); otherwise returns false.

On Unix (including OS X and iOS), opening a symlink effectively opens the link's target. On Windows, it opens the .lnk file itself.

Example:

QFileInfo info(fileName);
if (info.isSymLink())
    fileName = info.symLinkTarget();

Note: If the symlink points to a non existing file, exists() returns false.

See also isFile(), isDir(), and symLinkTarget().

fn is_writable(&self) -> bool

C++ method: bool QFileInfo::isWritable() const

C++ documentation:

Returns true if the user can write to the file; otherwise returns false.

See also isReadable(), isExecutable(), and permission().

fn last_modified<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoLastModifiedArgs<'largs>

C++ method: QFileInfo::lastModified

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_modified(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn last_modified(&self, cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: QDateTime QFileInfo::lastModified() const

C++ documentation:

Returns the date and time when the file was last modified.

See also created() and lastRead().

fn last_read<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoLastReadArgs<'largs>

C++ method: QFileInfo::lastRead

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_read(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn last_read(&self, cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: QDateTime QFileInfo::lastRead() const

C++ documentation:

Returns the date and time when the file was last read (accessed).

On platforms where this information is not available, returns the same as lastModified().

See also created() and lastModified().

fn make_absolute(&mut self) -> bool

C++ method: bool QFileInfo::makeAbsolute()

C++ documentation:

Converts the file's path to an absolute path if it is not already in that form. Returns true to indicate that the path was converted; otherwise returns false to indicate that the path was already absolute.

See also filePath() and isRelative().

fn new<Args>(args: Args) -> Args::ReturnType where Args: FileInfoNewArgs

C++ method: QFileInfo::QFileInfo

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::file_info::FileInfo
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::file_info::FileInfo>
C++ method: [constructor] void QFileInfo::QFileInfo()

Variant 2

Rust arguments:
1) fn new((&::dir::Dir, &::string::String, cpp_utils::AsStruct)) -> ::file_info::FileInfo
2) fn new((&::dir::Dir, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::file_info::FileInfo>
C++ method: [constructor] void QFileInfo::QFileInfo(const QDir& dir, const QString& file)

Variant 3

Rust arguments:
1) fn new((&::file::File, cpp_utils::AsStruct)) -> ::file_info::FileInfo
2) fn new((&::file::File, cpp_utils::AsBox)) -> cpp_utils::CppBox<::file_info::FileInfo>
C++ method: [constructor] void QFileInfo::QFileInfo(const QFile& file)

Variant 4

Rust arguments:
1) fn new((&::file_info::FileInfo, cpp_utils::AsStruct)) -> ::file_info::FileInfo
2) fn new((&::file_info::FileInfo, cpp_utils::AsBox)) -> cpp_utils::CppBox<::file_info::FileInfo>
C++ method: [constructor] void QFileInfo::QFileInfo(const QFileInfo& fileinfo)

Variant 5

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::file_info::FileInfo
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::file_info::FileInfo>
C++ method: [constructor] void QFileInfo::QFileInfo(const QString& file)

fn op_assign<'l0, 'l1>(&'l0 mut self, fileinfo: &'l1 FileInfo) -> &'l0 mut FileInfo

C++ method: QFileInfo& QFileInfo::operator=(const QFileInfo& fileinfo)

C++ documentation:

Makes a copy of the given fileinfo and assigns it to this QFileInfo.

fn op_eq(&self, fileinfo: &FileInfo) -> bool

C++ method: bool QFileInfo::operator==(const QFileInfo& fileinfo) const

C++ documentation:

Returns true if this QFileInfo object refers to a file in the same location as fileinfo; otherwise returns false.

Note that the result of comparing two empty QFileInfo objects, containing no file references (file paths that do not exist or are empty), is undefined.

Warning: This will not compare two different symbolic links pointing to the same file.

Warning: Long and short file names that refer to the same file on Windows are treated as if they referred to different files.

See also operator!=().

fn op_neq(&self, fileinfo: &FileInfo) -> bool

C++ method: bool QFileInfo::operator!=(const QFileInfo& fileinfo) const

C++ documentation:

Returns true if this QFileInfo object refers to a different file than the one specified by fileinfo; otherwise returns false.

See also operator==().

fn owner<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoOwnerArgs<'largs>

C++ method: QFileInfo::owner

This is an overloaded function. Available variants:

Rust arguments:
1) fn owner(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn owner(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::owner() const

C++ documentation:

Returns the owner of the file. On systems where files do not have owners, or if an error occurs, an empty string is returned.

This function can be time consuming under Unix (in the order of milliseconds).

See also ownerId(), group(), and groupId().

fn owner_id(&self) -> c_uint

C++ method: unsigned int QFileInfo::ownerId() const

C++ documentation:

Returns the id of the owner of the file.

On Windows and on systems where files do not have owners this function returns ((uint) -2).

See also owner(), group(), and groupId().

fn path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoPathArgs<'largs>

C++ method: QFileInfo::path

This is an overloaded function. Available variants:

Rust arguments:
1) fn path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::path() const

C++ documentation:

Returns the file's path. This doesn't include the file name.

Note that, if this QFileInfo object is given a path ending in a slash, the name of the file is considered empty and this function will return the entire path.

See also filePath(), absolutePath(), canonicalPath(), dir(), fileName(), and isRelative().

fn permission(&self, permissions: Flags<Permission>) -> bool

C++ method: bool QFileInfo::permission(QFlags<QFileDevice::Permission> permissions) const

C++ documentation:

Tests for file permissions. The permissions argument can be several flags of type QFile::Permissions OR-ed together to check for permission combinations.

On systems where files do not have permissions this function always returns true.

Example:

QFileInfo fi("/tmp/archive.tar.gz");
if (fi.permission(QFile::WriteUser | QFile::ReadGroup))
    qWarning("I can change the file; my group can read the file");
if (fi.permission(QFile::WriteGroup | QFile::WriteOther))
    qWarning("The group or others can change the file");

See also isReadable(), isWritable(), and isExecutable().

fn permissions(&self) -> Flags<Permission>

C++ method: QFlags<QFileDevice::Permission> QFileInfo::permissions() const

C++ documentation:

Returns the complete OR-ed together combination of QFile::Permissions for the file.

fn read_link<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoReadLinkArgs<'largs>

C++ method: QFileInfo::readLink

This is an overloaded function. Available variants:

Rust arguments:
1) fn read_link(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn read_link(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::readLink() const

C++ documentation:

Use symLinkTarget() instead.

fn refresh(&mut self)

C++ method: void QFileInfo::refresh()

C++ documentation:

Refreshes the information about the file, i.e. reads in information from the file system the next time a cached property is fetched.

Note: On Windows CE, there might be a delay for the file system driver to detect changes on the file.

fn set_caching(&mut self, on: bool)

C++ method: void QFileInfo::setCaching(bool on)

C++ documentation:

If enable is true, enables caching of file information. If enable is false caching is disabled.

When caching is enabled, QFileInfo reads the file information from the file system the first time it's needed, but generally not later.

Caching is enabled by default.

See also refresh() and caching().

fn set_file<'largs, Args>(&'largs mut self, args: Args) where Args: FileInfoSetFileArgs<'largs>

C++ method: QFileInfo::setFile

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_file(&mut self, (&::dir::Dir, &::string::String)) -> ()
C++ method: void QFileInfo::setFile(const QDir& dir, const QString& file)

C++ documentation:

This is an overloaded function.

Sets the file that the QFileInfo provides information about to file in directory dir.

If file includes a relative path, the QFileInfo will also have a relative path.

See also isRelative().

Variant 2

Rust arguments: fn set_file(&mut self, &::file::File) -> ()
C++ method: void QFileInfo::setFile(const QFile& file)

C++ documentation:

This is an overloaded function.

Sets the file that the QFileInfo provides information about to file.

If file includes a relative path, the QFileInfo will also have a relative path.

See also isRelative().

Variant 3

Rust arguments: fn set_file(&mut self, &::string::String) -> ()
C++ method: void QFileInfo::setFile(const QString& file)

C++ documentation:

Sets the file that the QFileInfo provides information about to file.

The file can also include an absolute or relative file path. Absolute paths begin with the directory separator (e.g. "/" under Unix) or a drive specification (under Windows). Relative file names begin with a directory name or a file name and specify a path relative to the current directory.

Example:

QString absolute = "/local/bin";
QString relative = "local/bin";
QFileInfo absFile(absolute);
QFileInfo relFile(relative);

QDir::setCurrent(QDir::rootPath()); // absFile and relFile now point to the same file

QDir::setCurrent("/tmp"); // absFile now points to "/local/bin", // while relFile points to "/tmp/local/bin"
See also isFile(), isRelative(), QDir::setCurrent(), and QDir::isRelativePath().

fn size(&self) -> i64

C++ method: qint64 QFileInfo::size() const

C++ documentation:

Returns the file size in bytes. If the file does not exist or cannot be fetched, 0 is returned.

See also exists().

fn suffix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoSuffixArgs<'largs>

C++ method: QFileInfo::suffix

This is an overloaded function. Available variants:

Rust arguments:
1) fn suffix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn suffix(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::suffix() const

C++ documentation:

Returns the suffix of the file.

The suffix consists of all characters in the file after (but not including) the last '.'.

Example:

QFileInfo fi("/tmp/archive.tar.gz");
QString ext = fi.suffix();  // ext = "gz"

The suffix of a file is computed equally on all platforms, independent of file naming conventions (e.g., ".bashrc" on Unix has an empty base name, and the suffix is "bashrc").

See also fileName(), completeSuffix(), baseName(), and completeBaseName().

fn swap(&mut self, other: &mut FileInfo)

C++ method: void QFileInfo::swap(QFileInfo& other)

C++ documentation:

Swaps this file info with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn sym_link_target<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FileInfoSymLinkTargetArgs<'largs>

C++ method: QFileInfo::symLinkTarget

This is an overloaded function. Available variants:

Rust arguments:
1) fn sym_link_target(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn sym_link_target(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFileInfo::symLinkTarget() const

C++ documentation:

Returns the absolute path to the file or directory a symlink (or shortcut on Windows) points to, or a an empty string if the object isn't a symbolic link.

This name may not represent an existing file; it is only a string. QFileInfo::exists() returns true if the symlink points to an existing file.

This function was introduced in Qt 4.2.

See also exists(), isSymLink(), isDir(), and isFile().