Module qt_core::file_selector::overloading

Traits

FileSelectorAllSelectorsArgs
This trait represents a set of arguments accepted by FileSelector::all_selectors method.

FileSelectorConnectArgs
This trait represents a set of arguments accepted by FileSelector::connect method.

FileSelectorConnectStaticArgs
This trait represents a set of arguments accepted by FileSelector::connect_static method.

FileSelectorDestroyedArgs
This trait represents a set of arguments accepted by FileSelector::destroyed method.

FileSelectorDisconnectArgs
This trait represents a set of arguments accepted by FileSelector::disconnect method.

FileSelectorDisconnectStaticArgs
This trait represents a set of arguments accepted by FileSelector::disconnect_static method.

FileSelectorDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by FileSelector::dynamic_property_names method.

FileSelectorExtraSelectorsArgs
This trait represents a set of arguments accepted by FileSelector::extra_selectors method.

FileSelectorNewArgs
This trait represents a set of arguments accepted by FileSelector::new method.

FileSelectorObjectNameArgs
This trait represents a set of arguments accepted by FileSelector::object_name method.

FileSelectorPropertyArgs
This trait represents a set of arguments accepted by FileSelector::property method.

FileSelectorSelectArgs
This trait represents a set of arguments accepted by FileSelector::select method.

FileSelectorStartTimerArgs
This trait represents a set of arguments accepted by FileSelector::start_timer method.

FileSelectorTrArgs
This trait represents a set of arguments accepted by FileSelector::tr method.

FileSelectorTrUtf8Args
This trait represents a set of arguments accepted by FileSelector::tr_utf8 method.