Module qt_core::file_system_watcher::overloading

Traits

FileSystemWatcherAddPathsArgs
This trait represents a set of arguments accepted by FileSystemWatcher::add_paths method.

FileSystemWatcherConnectArgs
This trait represents a set of arguments accepted by FileSystemWatcher::connect method.

FileSystemWatcherConnectStaticArgs
This trait represents a set of arguments accepted by FileSystemWatcher::connect_static method.

FileSystemWatcherDestroyedArgs
This trait represents a set of arguments accepted by FileSystemWatcher::destroyed method.

FileSystemWatcherDirectoriesArgs
This trait represents a set of arguments accepted by FileSystemWatcher::directories method.

FileSystemWatcherDisconnectArgs
This trait represents a set of arguments accepted by FileSystemWatcher::disconnect method.

FileSystemWatcherDisconnectStaticArgs
This trait represents a set of arguments accepted by FileSystemWatcher::disconnect_static method.

FileSystemWatcherDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by FileSystemWatcher::dynamic_property_names method.

FileSystemWatcherFilesArgs
This trait represents a set of arguments accepted by FileSystemWatcher::files method.

FileSystemWatcherNewArgs
This trait represents a set of arguments accepted by FileSystemWatcher::new method.

FileSystemWatcherObjectNameArgs
This trait represents a set of arguments accepted by FileSystemWatcher::object_name method.

FileSystemWatcherPropertyArgs
This trait represents a set of arguments accepted by FileSystemWatcher::property method.

FileSystemWatcherRemovePathsArgs
This trait represents a set of arguments accepted by FileSystemWatcher::remove_paths method.

FileSystemWatcherStartTimerArgs
This trait represents a set of arguments accepted by FileSystemWatcher::start_timer method.

FileSystemWatcherTrArgs
This trait represents a set of arguments accepted by FileSystemWatcher::tr method.

FileSystemWatcherTrUtf8Args
This trait represents a set of arguments accepted by FileSystemWatcher::tr_utf8 method.