Module qt_core::final_state::overloading

Traits

FinalStateConnectArgs
This trait represents a set of arguments accepted by FinalState::connect method.

FinalStateConnectStaticArgs
This trait represents a set of arguments accepted by FinalState::connect_static method.

FinalStateDestroyedArgs
This trait represents a set of arguments accepted by FinalState::destroyed method.

FinalStateDisconnectArgs
This trait represents a set of arguments accepted by FinalState::disconnect method.

FinalStateDisconnectStaticArgs
This trait represents a set of arguments accepted by FinalState::disconnect_static method.

FinalStateDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by FinalState::dynamic_property_names method.

FinalStateNewArgs
This trait represents a set of arguments accepted by FinalState::new method.

FinalStateObjectNameArgs
This trait represents a set of arguments accepted by FinalState::object_name method.

FinalStatePropertyArgs
This trait represents a set of arguments accepted by FinalState::property method.

FinalStateStartTimerArgs
This trait represents a set of arguments accepted by FinalState::start_timer method.

FinalStateTrArgs
This trait represents a set of arguments accepted by FinalState::tr method.

FinalStateTrUtf8Args
This trait represents a set of arguments accepted by FinalState::tr_utf8 method.