Struct qt_core::future::FutureVoid

pub struct FutureVoid {
    // some fields omitted
}

C++ type: QFuture<void>

Methods

impl FutureVoid

fn cancel(&mut self)

C++ method: void QFuture<void>::cancel()

C++ documentation:

Cancels the asynchronous computation represented by this future. Note that the cancelation is asynchronous. Use waitForFinished() after calling cancel() when you need synchronous cancelation.

Results currently available may still be accessed on a canceled future, but new results will not become available after calling this function. Any QFutureWatcher object that is watching this future will not deliver progress and result ready signals on a canceled future.

Be aware that not all asynchronous computations can be canceled. For example, the future returned by QtConcurrent::run() cannot be canceled; but the future returned by QtConcurrent::mappedReduced() can.

fn is_canceled(&self) -> bool

C++ method: bool QFuture<void>::isCanceled() const

C++ documentation:

Returns true if the asynchronous computation has been canceled with the cancel() function; otherwise returns false.

Be aware that the computation may still be running even though this function returns true. See cancel() for more details.

fn is_finished(&self) -> bool

C++ method: bool QFuture<void>::isFinished() const

C++ documentation:

Returns true if the asynchronous computation represented by this future has finished; otherwise returns false.

fn is_paused(&self) -> bool

C++ method: bool QFuture<void>::isPaused() const

C++ documentation:

Returns true if the asynchronous computation has been paused with the pause() function; otherwise returns false.

Be aware that the computation may still be running even though this function returns true. See setPaused() for more details.

See also setPaused() and togglePaused().

fn is_running(&self) -> bool

C++ method: bool QFuture<void>::isRunning() const

C++ documentation:

Returns true if the asynchronous computation represented by this future is currently running; otherwise returns false.

fn is_started(&self) -> bool

C++ method: bool QFuture<void>::isStarted() const

C++ documentation:

Returns true if the asynchronous computation represented by this future has been started; otherwise returns false.

fn new<Args>(args: Args) -> Args::ReturnType where Args: FutureVoidNewArgs

C++ method: QFuture<void>::QFuture

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::future::FutureVoid
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::future::FutureVoid>
C++ method: [constructor] void QFuture<void>::QFuture()

Variant 2

Rust arguments:
1) fn new((&::future::FutureVoid, cpp_utils::AsStruct)) -> ::future::FutureVoid
2) fn new((&::future::FutureVoid, cpp_utils::AsBox)) -> cpp_utils::CppBox<::future::FutureVoid>
C++ method: [constructor] void QFuture<void>::QFuture(const QFuture<void>& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 FutureVoid) -> &'l0 mut FutureVoid

C++ method: QFuture<void>& QFuture<void>::operator=(const QFuture<void>& other)

C++ documentation:

Assigns other to this future and returns a reference to this future.

fn op_eq(&self, other: &FutureVoid) -> bool

C++ method: bool QFuture<void>::operator==(const QFuture<void>& other) const

C++ documentation:

Returns true if other is a copy of this future; otherwise returns false.

fn op_neq(&self, other: &FutureVoid) -> bool

C++ method: bool QFuture<void>::operator!=(const QFuture<void>& other) const

C++ documentation:

Returns true if other is not a copy of this future; otherwise returns false.

fn pause(&mut self)

C++ method: void QFuture<void>::pause()

C++ documentation:

Pauses the asynchronous computation represented by this future. This is a convenience method that simply calls setPaused(true).

See also resume().

fn progress_maximum(&self) -> c_int

C++ method: int QFuture<void>::progressMaximum() const

C++ documentation:

Returns the maximum progressValue().

See also progressValue() and progressMinimum().

fn progress_minimum(&self) -> c_int

C++ method: int QFuture<void>::progressMinimum() const

C++ documentation:

Returns the minimum progressValue().

See also progressValue() and progressMaximum().

fn progress_text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: FutureVoidProgressTextArgs<'largs>

C++ method: QFuture<void>::progressText

This is an overloaded function. Available variants:

Rust arguments:
1) fn progress_text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn progress_text(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QFuture<void>::progressText() const

C++ documentation:

Returns the (optional) textual representation of the progress as reported by the asynchronous computation.

Be aware that not all computations provide a textual representation of the progress, and as such, this function may return an empty string.

fn progress_value(&self) -> c_int

C++ method: int QFuture<void>::progressValue() const

C++ documentation:

Returns the current progress value, which is between the progressMinimum() and progressMaximum().

See also progressMinimum() and progressMaximum().

fn result_count(&self) -> c_int

C++ method: int QFuture<void>::resultCount() const

C++ documentation:

Returns the number of continuous results available in this future. The real number of results stored might be different from this value, due to gaps in the result set. It is always safe to iterate through the results from 0 to resultCount().

See also result(), resultAt(), and results().

fn resume(&mut self)

C++ method: void QFuture<void>::resume()

C++ documentation:

Resumes the asynchronous computation represented by this future. This is a convenience method that simply calls setPaused(false).

See also pause().

fn set_paused(&mut self, paused: bool)

C++ method: void QFuture<void>::setPaused(bool paused)

C++ documentation:

If paused is true, this function pauses the asynchronous computation represented by the future. If the computation is already paused, this function does nothing. Any QFutureWatcher object that is watching this future will stop delivering progress and result ready signals while the future is paused. Signal delivery will continue once the future is resumed.

If paused is false, this function resumes the asynchronous computation. If the computation was not previously paused, this function does nothing.

Be aware that not all computations can be paused. For example, the future returned by QtConcurrent::run() cannot be paused; but the future returned by QtConcurrent::mappedReduced() can.

See also isPaused(), pause(), resume(), and togglePaused().

fn toggle_paused(&mut self)

C++ method: void QFuture<void>::togglePaused()

C++ documentation:

Toggles the paused state of the asynchronous computation. In other words, if the computation is currently paused, calling this function resumes it; if the computation is running, it is paused. This is a convenience method for calling setPaused(!isPaused()).

See also setPaused(), pause(), and resume().

fn wait_for_finished(&mut self)

C++ method: void QFuture<void>::waitForFinished()

C++ documentation:

Waits for the asynchronous computation to finish (including cancel()ed computations).