Function qt_core::global::fuzzy_compare

pub fn fuzzy_compare<Args>(args: Args) -> bool where Args: FuzzyCompareArgs

C++ method: qFuzzyCompare

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn fuzzy_compare((libc::c_double, libc::c_double)) -> bool
C++ method: bool qFuzzyCompare(double p1, double p2)

C++ documentation:

Compares the floating point value p1 and p2 and returns true if they are considered equal, otherwise false.

Note that comparing values where either p1 or p2 is 0.0 will not work. The solution to this is to compare against values greater than or equal to 1.0.

        // Instead of comparing with 0.0
                qFuzzyCompare(0.0,1.0e-200); // This will return false
        // Compare adding 1 to both values will fix the problem
                qFuzzyCompare(1 + 0.0, 1 + 1.0e-200); // This will return true

The two numbers are compared in a relative way, where the exactness is stronger the smaller the numbers are.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

Variant 2

Rust arguments: fn fuzzy_compare((libc::c_float, libc::c_float)) -> bool
C++ method: bool qFuzzyCompare(float p1, float p2)

C++ documentation:

Compares the floating point value p1 and p2 and returns true if they are considered equal, otherwise false.

The two numbers are compared in a relative way, where the exactness is stronger the smaller the numbers are.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.