Function qt_core::global::qgetenv

pub fn qgetenv<Args>(args: Args) -> Args::ReturnType where Args: QgetenvArgs

C++ method: qgetenv

This is an overloaded function. Available variants:

Rust arguments:
1) fn qgetenv((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn qgetenv((*const libc::c_char, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray qgetenv(const char* varName)

C++ documentation:

Returns the value of the environment variable with name varName. To get the variable string, use QByteArray::constData().

Note: qgetenv() was introduced because getenv() from the standard C library was deprecated in VC2005 (and later versions). qgetenv() uses the new replacement function in VC, and calls the standard C library's implementation on all other platforms.

See also qputenv(), qEnvironmentVariableIsSet(), and qEnvironmentVariableIsEmpty().