Function qt_core::hash::hash

pub fn hash<Args>(args: Args) -> c_uint where Args: HashArgs

C++ method: qHash

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn hash(&::latin1_string::Latin1String) -> libc::c_uint
2) fn hash((&::latin1_string::Latin1String, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(QLatin1String key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments:
1) fn hash(libc::c_char) -> libc::c_uint
2) fn hash((libc::c_char, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(char key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 3

Rust arguments:
1) fn hash(&::bit_array::BitArray) -> libc::c_uint
2) fn hash((&::bit_array::BitArray, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QBitArray& key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 4

Rust arguments:
1) fn hash(&::byte_array::ByteArray) -> libc::c_uint
2) fn hash((&::byte_array::ByteArray, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QByteArray& key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 5

Rust arguments:
1) fn hash(&::char::Char) -> libc::c_uint
2) fn hash((&::char::Char, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QChar key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 6

Rust arguments:
1) fn hash(&::string::String) -> libc::c_uint
2) fn hash((&::string::String, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QString& key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 7

Rust arguments:
1) fn hash(&::string_ref::StringRef) -> libc::c_uint
2) fn hash((&::string_ref::StringRef, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(const QStringRef& key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 8

Rust arguments:
1) fn hash(libc::c_double) -> libc::c_uint
2) fn hash((libc::c_double, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(double key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.3.

Variant 9

Rust arguments:
1) fn hash(libc::c_float) -> libc::c_uint
2) fn hash((libc::c_float, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(float key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.3.

Variant 10

Rust arguments:
1) fn hash(libc::c_int) -> libc::c_uint
2) fn hash((libc::c_int, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(int key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 11

Rust arguments:
1) fn hash(libc::c_long) -> libc::c_uint
2) fn hash((libc::c_long, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(long key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 12

Rust arguments:
1) fn hash(u64) -> libc::c_uint
2) fn hash((u64, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(quint64 key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 13

Rust arguments:
1) fn hash(libc::c_short) -> libc::c_uint
2) fn hash((libc::c_short, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(short key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 14

Rust arguments:
1) fn hash(libc::c_uchar) -> libc::c_uint
2) fn hash((libc::c_uchar, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(unsigned char key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 15

Rust arguments:
1) fn hash(libc::c_uint) -> libc::c_uint
2) fn hash((libc::c_uint, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(unsigned int key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.

Variant 16

Rust arguments:
1) fn hash(libc::c_ushort) -> libc::c_uint
2) fn hash((libc::c_ushort, libc::c_uint)) -> libc::c_uint
C++ method: unsigned int qHash(unsigned short key, unsigned int seed = ?)

C++ documentation:

Returns the hash value for the key, using seed to seed the calculation.

This function was introduced in Qt 5.0.