Struct qt_core::hash::HashStringRefVariantRef

pub struct HashStringRefVariantRef {
    // some fields omitted
}

C++ type: QHash<QString, QVariant>

Methods

impl HashStringRefVariantRef

fn capacity(&self) -> c_int

C++ method: int QHash<QString, QVariant>::capacity() const

C++ documentation:

Returns the number of buckets in the QHash's internal hash table.

The sole purpose of this function is to provide a means of fine tuning QHash's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the hash, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QHash<QString, QVariant>::clear()

C++ documentation:

Removes all items from the hash.

See also remove().

fn contains(&self, key: &String) -> bool

C++ method: bool QHash<QString, QVariant>::contains(const QString& key) const

C++ documentation:

Returns true if the hash contains an item with the key; otherwise returns false.

See also count() and QMultiHash::contains().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: HashStringRefVariantRefCountArgs<'largs>

C++ method: QHash<QString, QVariant>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QHash<QString, QVariant>::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &::string::String) -> libc::c_int
C++ method: int QHash<QString, QVariant>::count(const QString& key) const

C++ documentation:

Returns the number of items associated with the key.

See also contains() and insertMulti().

fn detach(&mut self)

C++ method: void QHash<QString, QVariant>::detach()

fn empty(&self) -> bool

C++ method: bool QHash<QString, QVariant>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the hash is empty; otherwise returns false.

fn is_detached(&self) -> bool

C++ method: bool QHash<QString, QVariant>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QHash<QString, QVariant>::isEmpty() const

C++ documentation:

Returns true if the hash contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &HashStringRefVariantRef) -> bool

C++ method: bool QHash<QString, QVariant>::isSharedWith(const QHash<QString, QVariant>& other) const

fn key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefKeyArgs<'largs>

C++ method: QHash<QString, QVariant>::key

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn key(&self, (&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn key(&self, (&::variant::Variant, cpp_utils::AsStruct)) -> ::string::String
C++ method: const QString QHash<QString, QVariant>::key(const QVariant& value) const

C++ documentation:

Returns the first key mapped to value.

If the hash contains no item with the value, the function returns a default-constructed key.

This function can be slow (linear time), because QHash's internal data structure is optimized for fast lookup by key, not by value.

See also value() and keys().

Variant 2

Rust arguments:
1) fn key(&self, (&::variant::Variant, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn key(&self, (&::variant::Variant, &::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: const QString QHash<QString, QVariant>::key(const QVariant& value, const QString& defaultKey) const

C++ documentation:

This is an overloaded function.

Returns the first key mapped to value, or defaultKey if the hash contains no item mapped to value.

This function can be slow (linear time), because QHash's internal data structure is optimized for fast lookup by key, not by value.

This function was introduced in Qt 4.3.

fn keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefKeysArgs<'largs>

C++ method: QHash<QString, QVariant>::keys

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListStringRef>
2) fn keys(&self, cpp_utils::AsStruct) -> ::list::ListStringRef
C++ method: QList<QString> QHash<QString, QVariant>::keys() const

C++ documentation:

Returns a list containing all the keys in the hash, in an arbitrary order. Keys that occur multiple times in the hash (because items were inserted with insertMulti(), or unite() was used) also occur multiple times in the list.

To obtain a list of unique keys, where each key from the map only occurs once, use uniqueKeys().

The order is guaranteed to be the same as that used by values().

See also uniqueKeys(), values(), and key().

Variant 2

Rust arguments:
1) fn keys(&self, (&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListStringRef>
2) fn keys(&self, (&::variant::Variant, cpp_utils::AsStruct)) -> ::list::ListStringRef
C++ method: QList<QString> QHash<QString, QVariant>::keys(const QVariant& value) const

C++ documentation:

This is an overloaded function.

Returns a list containing all the keys associated with value value, in an arbitrary order.

This function can be slow (linear time), because QHash's internal data structure is optimized for fast lookup by key, not by value.

fn new<Args>(args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefNewArgs

C++ method: QHash<QString, QVariant>::QHash

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::hash::HashStringRefVariantRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::hash::HashStringRefVariantRef>
C++ method: [constructor] void QHash<QString, QVariant>::QHash()

Variant 2

Rust arguments:
1) fn new((&::hash::HashStringRefVariantRef, cpp_utils::AsStruct)) -> ::hash::HashStringRefVariantRef
2) fn new((&::hash::HashStringRefVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::hash::HashStringRefVariantRef>
C++ method: [constructor] void QHash<QString, QVariant>::QHash(const QHash<QString, QVariant>& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 HashStringRefVariantRef) -> &'l0 mut HashStringRefVariantRef

C++ method: QHash<QString, QVariant>& QHash<QString, QVariant>::operator=(const QHash<QString, QVariant>& other)

C++ documentation:

Assigns other to this hash and returns a reference to this hash.

fn op_eq(&self, other: &HashStringRefVariantRef) -> bool

C++ method: bool QHash<QString, QVariant>::operator==(const QHash<QString, QVariant>& other) const

C++ documentation:

Returns true if other is equal to this hash; otherwise returns false.

Two hashes are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator!=().

fn op_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefOpIndexArgs<'largs>

C++ method: QHash<QString, QVariant>::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn op_index(&self, (&::string::String, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: const QVariant QHash<QString, QVariant>::operator[](const QString& key) const

C++ documentation:

This is an overloaded function.

Same as value().

fn op_index_mut<'l0, 'l1>(&'l0 mut self, key: &'l1 String) -> &'l0 mut Variant

C++ method: QVariant& QHash<QString, QVariant>::operator[](const QString& key)

C++ documentation:

Returns the value associated with the key as a modifiable reference.

If the hash contains no item with the key, the function inserts a default-constructed value into the hash with the key, and returns a reference to it. If the hash contains multiple items with the key, this function returns a reference to the most recently inserted value.

See also insert() and value().

fn op_neq(&self, other: &HashStringRefVariantRef) -> bool

C++ method: bool QHash<QString, QVariant>::operator!=(const QHash<QString, QVariant>& other) const

C++ documentation:

Returns true if other is not equal to this hash; otherwise returns false.

Two hashes are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator==().

fn remove(&mut self, key: &String) -> c_int

C++ method: int QHash<QString, QVariant>::remove(const QString& key)

C++ documentation:

Removes all the items that have the key from the hash. Returns the number of items removed which is usually 1 but will be 0 if the key isn't in the hash, or greater than 1 if insertMulti() has been used with the key.

See also clear(), take(), and QMultiHash::remove().

fn reserve(&mut self, size: c_int)

C++ method: void QHash<QString, QVariant>::reserve(int size)

C++ documentation:

Ensures that the QHash's internal hash table consists of at least size buckets.

This function is useful for code that needs to build a huge hash and wants to avoid repeated reallocation. For example:

QHash<QString, int> hash;
hash.reserve(20000);
for (int i = 0; i < 20000; ++i)
    hash.insert(keys[i], values[i]);

Ideally, size should be slightly more than the maximum number of items expected in the hash. size doesn't have to be prime, because QHash will use a prime number internally anyway. If size is an underestimate, the worst that will happen is that the QHash will be a bit slower.

In general, you will rarely ever need to call this function. QHash's internal hash table automatically shrinks or grows to provide good performance without wasting too much memory.

See also squeeze() and capacity().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QHash<QString, QVariant>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QHash<QString, QVariant>::size() const

C++ documentation:

Returns the number of items in the hash.

See also isEmpty() and count().

fn squeeze(&mut self)

C++ method: void QHash<QString, QVariant>::squeeze()

C++ documentation:

Reduces the size of the QHash's internal hash table to save memory.

The sole purpose of this function is to provide a means of fine tuning QHash's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn swap(&mut self, other: &mut HashStringRefVariantRef)

C++ method: void QHash<QString, QVariant>::swap(QHash<QString, QVariant>& other)

C++ documentation:

Swaps hash other with this hash. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefTakeArgs<'largs>

C++ method: QHash<QString, QVariant>::take

This is an overloaded function. Available variants:

Rust arguments:
1) fn take(&mut self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn take(&mut self, (&::string::String, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QHash<QString, QVariant>::take(const QString& key)

C++ documentation:

Removes the item with the key from the hash and returns the value associated with it.

If the item does not exist in the hash, the function simply returns a default-constructed value. If there are multiple items for key in the hash, only the most recently inserted one is removed.

If you don't use the return value, remove() is more efficient.

See also remove().

fn unique_keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefUniqueKeysArgs<'largs>

C++ method: QHash<QString, QVariant>::uniqueKeys

This is an overloaded function. Available variants:

Rust arguments:
1) fn unique_keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListStringRef>
2) fn unique_keys(&self, cpp_utils::AsStruct) -> ::list::ListStringRef
C++ method: QList<QString> QHash<QString, QVariant>::uniqueKeys() const

C++ documentation:

Returns a list containing all the keys in the map. Keys that occur multiple times in the map (because items were inserted with insertMulti(), or unite() was used) occur only once in the returned list.

This function was introduced in Qt 4.2.

See also keys() and values().

fn unite<'l0, 'l1>(&'l0 mut self, other: &'l1 HashStringRefVariantRef) -> &'l0 mut HashStringRefVariantRef

C++ method: QHash<QString, QVariant>& QHash<QString, QVariant>::unite(const QHash<QString, QVariant>& other)

C++ documentation:

Inserts all the items in the other hash into this hash. If a key is common to both hashes, the resulting hash will contain the key multiple times.

See also insertMulti().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefValueArgs<'largs>

C++ method: QHash<QString, QVariant>::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn value(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn value(&self, (&::string::String, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: const QVariant QHash<QString, QVariant>::value(const QString& key) const

C++ documentation:

Returns the value associated with the key.

If the hash contains no item with the key, the function returns a default-constructed value. If there are multiple items for the key in the hash, the value of the most recently inserted one is returned.

See also key(), values(), contains(), and operator[]().

Variant 2

Rust arguments:
1) fn value(&self, (&::string::String, &::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn value(&self, (&::string::String, &::variant::Variant, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: const QVariant QHash<QString, QVariant>::value(const QString& key, const QVariant& defaultValue) const

C++ documentation:

This is an overloaded function.

If the hash contains no item with the given key, the function returns defaultValue.

fn values<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: HashStringRefVariantRefValuesArgs<'largs>

C++ method: QHash<QString, QVariant>::values

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn values(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListVariantRef>
2) fn values(&self, cpp_utils::AsStruct) -> ::list::ListVariantRef
C++ method: QList<QVariant> QHash<QString, QVariant>::values() const

C++ documentation:

Returns a list containing all the values in the hash, in an arbitrary order. If a key is associated with multiple values, all of its values will be in the list, and not just the most recently inserted one.

The order is guaranteed to be the same as that used by keys().

See also keys() and value().

Variant 2

Rust arguments:
1) fn values(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListVariantRef>
2) fn values(&self, (&::string::String, cpp_utils::AsStruct)) -> ::list::ListVariantRef
C++ method: QList<QVariant> QHash<QString, QVariant>::values(const QString& key) const

C++ documentation:

This is an overloaded function.

Returns a list of all the values associated with the key, from the most recently inserted to the least recently inserted.

See also count() and insertMulti().