Enum qt_core::history_state::HistoryType

pub enum HistoryType {
    Shallow,
    Deep,
}

C++ type: QHistoryState::HistoryType

C++ documentation:

This enum specifies the type of history that a QHistoryState records.

Variants
Shallow

Only the immediate child states of the parent state are recorded. In this case a transition with the history state as its target will end up in the immediate child state that the parent was in the last time it was exited. This is the default. (C++ enum variant: ShallowHistory = 0)

Deep

Nested states are recorded. In this case a transition with the history state as its target will end up in the most deeply nested descendant state the parent was in the last time it was exited. (C++ enum variant: DeepHistory = 1)