Module qt_core::item_selection::overloading

Traits

ItemSelectionAppendArgs
This trait represents a set of arguments accepted by ItemSelection::append method.

ItemSelectionCountArgs
This trait represents a set of arguments accepted by ItemSelection::count method.

ItemSelectionIndexOfArgs
This trait represents a set of arguments accepted by ItemSelection::index_of method.

ItemSelectionIndexesArgs
This trait represents a set of arguments accepted by ItemSelection::indexes method.

ItemSelectionLastIndexOfArgs
This trait represents a set of arguments accepted by ItemSelection::last_index_of method.

ItemSelectionMidArgs
This trait represents a set of arguments accepted by ItemSelection::mid method.

ItemSelectionNewArgs
This trait represents a set of arguments accepted by ItemSelection::new method.

ItemSelectionOpAddArgs
This trait represents a set of arguments accepted by ItemSelection::op_add method.

ItemSelectionOpAddAssignArgs
This trait represents a set of arguments accepted by ItemSelection::op_add_assign method.

ItemSelectionOpShlArgs
This trait represents a set of arguments accepted by ItemSelection::op_shl method.

ItemSelectionSwapArgs
This trait represents a set of arguments accepted by ItemSelection::swap method.

ItemSelectionTakeAtArgs
This trait represents a set of arguments accepted by ItemSelection::take_at method.

ItemSelectionTakeFirstArgs
This trait represents a set of arguments accepted by ItemSelection::take_first method.

ItemSelectionTakeLastArgs
This trait represents a set of arguments accepted by ItemSelection::take_last method.

ItemSelectionValueArgs
This trait represents a set of arguments accepted by ItemSelection::value method.