Enum qt_core::item_selection_model::SelectionFlag

pub enum SelectionFlag {
    NoUpdate,
    Clear,
    Select,
    ClearAndSelect,
    Deselect,
    Toggle,
    Current,
    SelectCurrent,
    ToggleCurrent,
    Rows,
    Columns,
}

C++ type: QItemSelectionModel::SelectionFlag

C++ documentation:

This enum describes the way the selection model will be updated.

The SelectionFlags type is a typedef for QFlags<SelectionFlag>. It stores an OR combination of SelectionFlag values.

Variants
NoUpdate

No selection will be made. (C++ enum variant: NoUpdate = 0)

Clear

The complete selection will be cleared. (C++ enum variant: Clear = 1)

Select

All specified indexes will be selected. (C++ enum variant: Select = 2)

ClearAndSelect

A combination of Clear and Select, provided for convenience. (C++ enum variant: ClearAndSelect = 3)

Deselect

All specified indexes will be deselected. (C++ enum variant: Deselect = 4)

Toggle

All specified indexes will be selected or deselected depending on their current state. (C++ enum variant: Toggle = 8)

Current

The current selection will be updated. (C++ enum variant: Current = 16)

SelectCurrent

A combination of Select and Current, provided for convenience. (C++ enum variant: SelectCurrent = 18)

ToggleCurrent

A combination of Toggle and Current, provided for convenience. (C++ enum variant: ToggleCurrent = 24)

Rows

All indexes will be expanded to span rows. (C++ enum variant: Rows = 32)

Columns

All indexes will be expanded to span columns. (C++ enum variant: Columns = 64)