Struct qt_core::item_selection_model::ItemSelectionModel

pub struct ItemSelectionModel {
    // some fields omitted
}

C++ type: QItemSelectionModel

C++ documentation:

The QItemSelectionModel class keeps track of a view's selected items.

A QItemSelectionModel keeps track of the selected items in a view, or in several views onto the same model. It also keeps track of the currently selected item in a view.

The QItemSelectionModel class is one of the Model/View Classes and is part of Qt's model/view framework.

The selected items are stored using ranges. Whenever you want to modify the selected items use select() and provide either a QItemSelection, or a QModelIndex and a QItemSelectionModel::SelectionFlag.

The QItemSelectionModel takes a two layer approach to selection management, dealing with both selected items that have been committed and items that are part of the current selection. The current selected items are part of the current interactive selection (for example with rubber-band selection or keyboard-shift selections).

To update the currently selected items, use the bitwise OR of QItemSelectionModel::Current and any of the other SelectionFlags. If you omit the QItemSelectionModel::Current command, a new current selection will be created, and the previous one added to the whole selection. All functions operate on both layers; for example, selecteditems() will return items from both layers.

Note: Since 5.5, model, hasSelection, and currentIndex are meta-object properties.

Methods

impl ItemSelectionModel

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QItemSelectionModel::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QItemSelectionModel::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn clear(&mut self)

C++ method: virtual void QItemSelectionModel::clear()

C++ documentation:

Clears the selection model. Emits selectionChanged() and currentChanged().

fn clear_current_index(&mut self)

C++ method: virtual void QItemSelectionModel::clearCurrentIndex()

C++ documentation:

Clears the current index. Emits currentChanged().

fn clear_selection(&mut self)

C++ method: void QItemSelectionModel::clearSelection()

C++ documentation:

Clears the selection in the selection model. Emits selectionChanged().

This function was introduced in Qt 4.2.

fn column_intersects_selection(&self, column: c_int, parent: &ModelIndex) -> bool

C++ method: bool QItemSelectionModel::columnIntersectsSelection(int column, const QModelIndex& parent) const

C++ documentation:

Returns true if there are any items selected in the column with the given parent.

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelConnectArgs<'largs>

C++ method: QItemSelectionModel::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: QMetaObject::Connection QItemSelectionModel::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: ItemSelectionModelConnectStaticArgs

C++ method: QItemSelectionModel::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QItemSelectionModel::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QItemSelectionModel::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn current_changed(&mut self, current: &ModelIndex, previous: &ModelIndex)

C++ method: void QItemSelectionModel::currentChanged(const QModelIndex& current, const QModelIndex& previous)

C++ documentation:

This signal is emitted whenever the current item changes. The previous model item index is replaced by the current index as the selection's current item.

Note that this signal will not be emitted when the item model is reset.

See also currentIndex(), setCurrentIndex(), and selectionChanged().

fn current_column_changed(&mut self, current: &ModelIndex, previous: &ModelIndex)

C++ method: void QItemSelectionModel::currentColumnChanged(const QModelIndex& current, const QModelIndex& previous)

C++ documentation:

This signal is emitted if the current item changes and its column is different to the column of the previous current item.

Note that this signal will not be emitted when the item model is reset.

See also currentChanged(), currentRowChanged(), currentIndex(), and setCurrentIndex().

fn current_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelCurrentIndexArgs<'largs>

C++ method: QItemSelectionModel::currentIndex

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_index(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn current_index(&self, cpp_utils::AsStruct) -> ::model_index::ModelIndex
C++ method: QModelIndex QItemSelectionModel::currentIndex() const

C++ documentation:

Returns the model item index for the current item, or an invalid index if there is no current item.

See also setCurrentIndex().

fn current_row_changed(&mut self, current: &ModelIndex, previous: &ModelIndex)

C++ method: void QItemSelectionModel::currentRowChanged(const QModelIndex& current, const QModelIndex& previous)

C++ documentation:

This signal is emitted if the current item changes and its row is different to the row of the previous current item.

Note that this signal will not be emitted when the item model is reset.

See also currentChanged(), currentColumnChanged(), currentIndex(), and setCurrentIndex().

fn delete_later(&mut self)

C++ method: void QItemSelectionModel::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: ItemSelectionModelDestroyedArgs<'largs>

C++ method: QItemSelectionModel::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut ::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QItemSelectionModel::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: ItemSelectionModelDisconnectArgs<'largs>

C++ method: QItemSelectionModel::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const ::object::Object) -> bool
2) fn disconnect(&self, (*const ::object::Object, *const libc::c_char)) -> bool
C++ method: bool QItemSelectionModel::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const ::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QItemSelectionModel::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: ItemSelectionModelDisconnectStaticArgs

C++ method: QItemSelectionModel::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&::meta_object::Connection) -> bool
C++ method: static bool QItemSelectionModel::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod)) -> bool
C++ method: static bool QItemSelectionModel::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QItemSelectionModel::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QItemSelectionModel::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QItemSelectionModel::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelDynamicPropertyNamesArgs<'largs>

C++ method: QItemSelectionModel::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: QList<QByteArray> QItemSelectionModel::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QItemSelectionModel::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QItemSelectionModel::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn has_selection(&self) -> bool

C++ method: bool QItemSelectionModel::hasSelection() const

C++ documentation:

Returns true if the selection model contains any selection ranges; otherwise returns false.

This function was introduced in Qt 4.2.

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QItemSelectionModel::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QItemSelectionModel::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_column_selected(&self, column: c_int, parent: &ModelIndex) -> bool

C++ method: bool QItemSelectionModel::isColumnSelected(int column, const QModelIndex& parent) const

C++ documentation:

Returns true if all items are selected in the column with the given parent.

Note that this function is usually faster than calling isSelected() on all items in the same column and that unselectable items are ignored.

fn is_row_selected(&self, row: c_int, parent: &ModelIndex) -> bool

C++ method: bool QItemSelectionModel::isRowSelected(int row, const QModelIndex& parent) const

C++ documentation:

Returns true if all items are selected in the row with the given parent.

Note that this function is usually faster than calling isSelected() on all items in the same row and that unselectable items are ignored.

fn is_selected(&self, index: &ModelIndex) -> bool

C++ method: bool QItemSelectionModel::isSelected(const QModelIndex& index) const

C++ documentation:

Returns true if the given model item index is selected.

fn is_widget_type(&self) -> bool

C++ method: bool QItemSelectionModel::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QItemSelectionModel::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn kill_timer(&mut self, id: c_int)

C++ method: void QItemSelectionModel::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QItemSelectionModel::metaObject() const

fn model(&self) -> *const AbstractItemModel

C++ method: const QAbstractItemModel* QItemSelectionModel::model() const

C++ documentation:

Returns the item model operated on by the selection model.

See also setModel().

fn model_changed(&mut self, model: *mut AbstractItemModel)

C++ method: void QItemSelectionModel::modelChanged(QAbstractItemModel* model)

C++ documentation:

This signal is emitted when the model is successfully set with setModel().

This function was introduced in Qt 5.5.

See also model() and setModel().

fn model_mut(&mut self) -> *mut AbstractItemModel

C++ method: QAbstractItemModel* QItemSelectionModel::model()

C++ documentation:

Returns the item model operated on by the selection model.

This function was introduced in Qt 5.5.

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QItemSelectionModel::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ItemSelectionModelNewArgs

C++ method: QItemSelectionModel::QItemSelectionModel

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*mut ::abstract_item_model::AbstractItemModel, cpp_utils::AsStruct)) -> ::item_selection_model::ItemSelectionModel
2) fn new(cpp_utils::AsStruct) -> ::item_selection_model::ItemSelectionModel
3) fn new((*mut ::abstract_item_model::AbstractItemModel, cpp_utils::AsBox)) -> cpp_utils::CppBox<::item_selection_model::ItemSelectionModel>
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::item_selection_model::ItemSelectionModel>
C++ method: [constructor] void QItemSelectionModel::QItemSelectionModel(QAbstractItemModel* model = ?)

Variant 2

Rust arguments:
1) fn new((*mut ::abstract_item_model::AbstractItemModel, *mut ::object::Object, cpp_utils::AsStruct)) -> ::item_selection_model::ItemSelectionModel
2) fn new((*mut ::abstract_item_model::AbstractItemModel, *mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::item_selection_model::ItemSelectionModel>
C++ method: [constructor] void QItemSelectionModel::QItemSelectionModel(QAbstractItemModel* model, QObject* parent)

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelObjectNameArgs<'largs>

C++ method: QItemSelectionModel::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QItemSelectionModel::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QItemSelectionModel::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelPropertyArgs<'largs>

C++ method: QItemSelectionModel::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QItemSelectionModel::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QItemSelectionModel::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QItemSelectionModel::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QItemSelectionModel::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QItemSelectionModel::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn reset(&mut self)

C++ method: virtual void QItemSelectionModel::reset()

C++ documentation:

Clears the selection model. Does not emit any signals.

fn row_intersects_selection(&self, row: c_int, parent: &ModelIndex) -> bool

C++ method: bool QItemSelectionModel::rowIntersectsSelection(int row, const QModelIndex& parent) const

C++ documentation:

Returns true if there are any items selected in the row with the given parent.

fn select<'largs, Args>(&'largs mut self, args: Args) where Args: ItemSelectionModelSelectArgs<'largs>

C++ method: QItemSelectionModel::select

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn select(&mut self, (&::item_selection::ItemSelection, ::flags::Flags<::item_selection_model::SelectionFlag>)) -> ()
C++ method: virtual void QItemSelectionModel::select(const QItemSelection& selection, QFlags<QItemSelectionModel::SelectionFlag> command)

C++ documentation:

Selects the item selection using the specified command, and emits selectionChanged().

See also QItemSelectionModel::SelectionFlag.

Variant 2

Rust arguments: fn select(&mut self, (&::model_index::ModelIndex, ::flags::Flags<::item_selection_model::SelectionFlag>)) -> ()
C++ method: virtual void QItemSelectionModel::select(const QModelIndex& index, QFlags<QItemSelectionModel::SelectionFlag> command)

C++ documentation:

Selects the model item index using the specified command, and emits selectionChanged().

See also QItemSelectionModel::SelectionFlags.

fn selected_columns<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelSelectedColumnsArgs<'largs>

C++ method: QItemSelectionModel::selectedColumns

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_columns(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListModelIndexRef>
2) fn selected_columns(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListModelIndexRef>
3) fn selected_columns(&self, cpp_utils::AsStruct) -> ::list::ListModelIndexRef
4) fn selected_columns(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListModelIndexRef
C++ method: QList<QModelIndex> QItemSelectionModel::selectedColumns(int row = ?) const

C++ documentation:

Returns the indexes in the given row for columns where all rows are selected.

This function was introduced in Qt 4.2.

See also selectedIndexes() and selectedRows().

fn selected_indexes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelSelectedIndexesArgs<'largs>

C++ method: QItemSelectionModel::selectedIndexes

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_indexes(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListModelIndexRef>
2) fn selected_indexes(&self, cpp_utils::AsStruct) -> ::list::ListModelIndexRef
C++ method: QList<QModelIndex> QItemSelectionModel::selectedIndexes() const

C++ documentation:

This property was introduced in Qt 5.5.

Access functions:

QModelIndexList 	selectedIndexes() const

Notifier signal:

void 	selectionChanged(const QItemSelection & selected, const QItemSelection & deselected)

fn selected_rows<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelSelectedRowsArgs<'largs>

C++ method: QItemSelectionModel::selectedRows

This is an overloaded function. Available variants:

Rust arguments:
1) fn selected_rows(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListModelIndexRef>
2) fn selected_rows(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListModelIndexRef>
3) fn selected_rows(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListModelIndexRef
4) fn selected_rows(&self, cpp_utils::AsStruct) -> ::list::ListModelIndexRef
C++ method: QList<QModelIndex> QItemSelectionModel::selectedRows(int column = ?) const

C++ documentation:

Returns the indexes in the given column for the rows where all columns are selected.

This function was introduced in Qt 4.2.

See also selectedIndexes() and selectedColumns().

fn selection<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionModelSelectionArgs<'largs>

C++ method: QItemSelectionModel::selection

This is an overloaded function. Available variants:

Rust arguments:
1) fn selection(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::item_selection::ItemSelection>
2) fn selection(&self, cpp_utils::AsStruct) -> ::item_selection::ItemSelection
C++ method: const QItemSelection QItemSelectionModel::selection() const

C++ documentation:

Access functions:

const QItemSelection 	selection() const

Notifier signal:

void 	selectionChanged(const QItemSelection & selected, const QItemSelection & deselected)

fn selection_changed(&mut self, selected: &ItemSelection, deselected: &ItemSelection)

C++ method: void QItemSelectionModel::selectionChanged(const QItemSelection& selected, const QItemSelection& deselected)

C++ documentation:

This property was introduced in Qt 5.5.

Access functions:

QModelIndexList 	selectedIndexes() const

Notifier signal:

void 	selectionChanged(const QItemSelection & selected, const QItemSelection & deselected)

fn set_current_index(&mut self, index: &ModelIndex, command: Flags<SelectionFlag>)

C++ method: virtual void QItemSelectionModel::setCurrentIndex(const QModelIndex& index, QFlags<QItemSelectionModel::SelectionFlag> command)

C++ documentation:

Sets the model item index to be the current item, and emits currentChanged(). The current item is used for keyboard navigation and focus indication; it is independent of any selected items, although a selected item can also be the current item.

Depending on the specified command, the index can also become part of the current selection.

See also currentIndex() and select().

fn set_model(&mut self, model: *mut AbstractItemModel)

C++ method: void QItemSelectionModel::setModel(QAbstractItemModel* model)

C++ documentation:

Sets the model. The modelChanged() signal will be emitted.

This function was introduced in Qt 5.5.

See also model() and modelChanged().

fn set_object_name(&mut self, name: &String)

C++ method: void QItemSelectionModel::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QItemSelectionModel::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QItemSelectionModel::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QItemSelectionModel::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QItemSelectionModel::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: ItemSelectionModelStartTimerArgs<'largs>

C++ method: QItemSelectionModel::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, ::qt::TimerType)) -> libc::c_int
C++ method: int QItemSelectionModel::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn thread(&self) -> *mut Thread

C++ method: QThread* QItemSelectionModel::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: ItemSelectionModelTrArgs

C++ method: QItemSelectionModel::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QItemSelectionModel::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: ItemSelectionModelTrUtf8Args

C++ method: QItemSelectionModel::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QItemSelectionModel::trUtf8(const char* s, const char* c, int n)

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QItemSelectionModel::userData(unsigned int id) const