Struct qt_core::item_selection_range::ItemSelectionRange

pub struct ItemSelectionRange {
    // some fields omitted
}

C++ type: QItemSelectionRange

C++ documentation:

The QItemSelectionRange class manages information about a range of selected items in a model.

A QItemSelectionRange contains information about a range of selected items in a model. A range of items is a contiguous array of model items, extending to cover a number of adjacent rows and columns with a common parent item; this can be visualized as a two-dimensional block of cells in a table. A selection range has a top(), left() a bottom(), right() and a parent().

The QItemSelectionRange class is one of the Model/View Classes and is part of Qt's model/view framework.

The model items contained in the selection range can be obtained using the indexes() function. Use QItemSelectionModel::selectedIndexes() to get a list of all selected items for a view.

You can determine whether a given model item lies within a particular range by using the contains() function. Ranges can also be compared using the overloaded operators for equality and inequality, and the intersects() function allows you to determine whether two ranges overlap.

Methods

impl ItemSelectionRange

fn bottom(&self) -> c_int

C++ method: int QItemSelectionRange::bottom() const

C++ documentation:

Returns the row index corresponding to the lowermost selected row in the selection range.

fn bottom_right<'l0>(&'l0 self) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QItemSelectionRange::bottomRight() const

C++ documentation:

Returns the index for the item located at the bottom-right corner of the selection range.

See also bottom(), right(), and topLeft().

fn contains<'largs, Args>(&'largs self, args: Args) -> bool where Args: ItemSelectionRangeContainsArgs<'largs>

C++ method: QItemSelectionRange::contains

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn contains(&self, &::model_index::ModelIndex) -> bool
C++ method: bool QItemSelectionRange::contains(const QModelIndex& index) const

C++ documentation:

Returns true if the model item specified by the index lies within the range of selected items; otherwise returns false.

Variant 2

Rust arguments: fn contains(&self, (libc::c_int, libc::c_int, &::model_index::ModelIndex)) -> bool
C++ method: bool QItemSelectionRange::contains(int row, int column, const QModelIndex& parentIndex) const

C++ documentation:

This is an overloaded function.

Returns true if the model item specified by (row, column) and with parentIndex as the parent item lies within the range of selected items; otherwise returns false.

fn height(&self) -> c_int

C++ method: int QItemSelectionRange::height() const

C++ documentation:

Returns the number of selected rows in the selection range.

fn indexes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionRangeIndexesArgs<'largs>

C++ method: QItemSelectionRange::indexes

This is an overloaded function. Available variants:

Rust arguments:
1) fn indexes(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListModelIndexRef>
2) fn indexes(&self, cpp_utils::AsStruct) -> ::list::ListModelIndexRef
C++ method: QList<QModelIndex> QItemSelectionRange::indexes() const

C++ documentation:

Returns the list of model index items stored in the selection.

fn intersected<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionRangeIntersectedArgs<'largs>

C++ method: QItemSelectionRange::intersected

This is an overloaded function. Available variants:

Rust arguments:
1) fn intersected(&self, (&::item_selection_range::ItemSelectionRange, cpp_utils::AsBox)) -> cpp_utils::CppBox<::item_selection_range::ItemSelectionRange>
2) fn intersected(&self, (&::item_selection_range::ItemSelectionRange, cpp_utils::AsStruct)) -> ::item_selection_range::ItemSelectionRange
C++ method: QItemSelectionRange QItemSelectionRange::intersected(const QItemSelectionRange& other) const

C++ documentation:

Returns a new selection range containing only the items that are found in both the selection range and the other selection range.

This function was introduced in Qt 4.2.

fn intersects(&self, other: &ItemSelectionRange) -> bool

C++ method: bool QItemSelectionRange::intersects(const QItemSelectionRange& other) const

C++ documentation:

Returns true if this selection range intersects (overlaps with) the other range given; otherwise returns false.

fn is_empty(&self) -> bool

C++ method: bool QItemSelectionRange::isEmpty() const

C++ documentation:

Returns true if the selection range contains no selectable item

This function was introduced in Qt 4.7.

fn is_valid(&self) -> bool

C++ method: bool QItemSelectionRange::isValid() const

C++ documentation:

Returns true if the selection range is valid; otherwise returns false.

fn left(&self) -> c_int

C++ method: int QItemSelectionRange::left() const

C++ documentation:

Returns the column index corresponding to the leftmost selected column in the selection range.

fn model(&self) -> *const AbstractItemModel

C++ method: const QAbstractItemModel* QItemSelectionRange::model() const

C++ documentation:

Returns the model that the items in the selection range belong to.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ItemSelectionRangeNewArgs

C++ method: QItemSelectionRange::QItemSelectionRange

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::item_selection_range::ItemSelectionRange
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::item_selection_range::ItemSelectionRange>
C++ method: [constructor] void QItemSelectionRange::QItemSelectionRange()

Variant 2

Rust arguments:
1) fn new((&::item_selection_range::ItemSelectionRange, cpp_utils::AsStruct)) -> ::item_selection_range::ItemSelectionRange
2) fn new((&::item_selection_range::ItemSelectionRange, cpp_utils::AsBox)) -> cpp_utils::CppBox<::item_selection_range::ItemSelectionRange>
C++ method: [constructor] void QItemSelectionRange::QItemSelectionRange(const QItemSelectionRange& other)

Variant 3

Rust arguments:
1) fn new((&::model_index::ModelIndex, cpp_utils::AsStruct)) -> ::item_selection_range::ItemSelectionRange
2) fn new((&::model_index::ModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::item_selection_range::ItemSelectionRange>
C++ method: [constructor] void QItemSelectionRange::QItemSelectionRange(const QModelIndex& index)

Variant 4

Rust arguments:
1) fn new((&::model_index::ModelIndex, &::model_index::ModelIndex, cpp_utils::AsStruct)) -> ::item_selection_range::ItemSelectionRange
2) fn new((&::model_index::ModelIndex, &::model_index::ModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::item_selection_range::ItemSelectionRange>
C++ method: [constructor] void QItemSelectionRange::QItemSelectionRange(const QModelIndex& topLeft, const QModelIndex& bottomRight)

fn op_eq(&self, other: &ItemSelectionRange) -> bool

C++ method: bool QItemSelectionRange::operator==(const QItemSelectionRange& other) const

C++ documentation:

Returns true if the selection range is exactly the same as the other range given; otherwise returns false.

fn op_lt(&self, other: &ItemSelectionRange) -> bool

C++ method: bool QItemSelectionRange::operator<(const QItemSelectionRange& other) const

C++ documentation:

Returns true if the selection range is less than the other range given; otherwise returns false.

The less than calculation is not directly useful to developers - the way that ranges with different parents compare is not defined. This operator only exists so that the class can be used with QMap.

fn op_neq(&self, other: &ItemSelectionRange) -> bool

C++ method: bool QItemSelectionRange::operator!=(const QItemSelectionRange& other) const

C++ documentation:

Returns true if the selection range differs from the other range given; otherwise returns false.

fn parent<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ItemSelectionRangeParentArgs<'largs>

C++ method: QItemSelectionRange::parent

This is an overloaded function. Available variants:

Rust arguments:
1) fn parent(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn parent(&self, cpp_utils::AsStruct) -> ::model_index::ModelIndex
C++ method: QModelIndex QItemSelectionRange::parent() const

C++ documentation:

Returns the parent model item index of the items in the selection range.

fn right(&self) -> c_int

C++ method: int QItemSelectionRange::right() const

C++ documentation:

Returns the column index corresponding to the rightmost selected column in the selection range.

fn top(&self) -> c_int

C++ method: int QItemSelectionRange::top() const

C++ documentation:

Returns the row index corresponding to the uppermost selected row in the selection range.

fn top_left<'l0>(&'l0 self) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QItemSelectionRange::topLeft() const

C++ documentation:

Returns the index for the item located at the top-left corner of the selection range.

See also top(), left(), and bottomRight().

fn width(&self) -> c_int

C++ method: int QItemSelectionRange::width() const

C++ documentation:

Returns the number of selected columns in the selection range.