Struct qt_core::json_array::ConstIterator

pub struct ConstIterator {
    // some fields omitted
}

C++ type: QJsonArray::const_iterator

C++ documentation:

The QJsonArray::const_iterator class provides an STL-style const iterator for QJsonArray.

QJsonArray::const_iterator allows you to iterate over a QJsonArray. If you want to modify the QJsonArray as you iterate over it, use QJsonArray::iterator instead. It is generally a good practice to use QJsonArray::const_iterator on a non-const QJsonArray as well, unless you need to change the QJsonArray through the iterator. Const iterators are slightly faster and improves code readability.

The default QJsonArray::const_iterator constructor creates an uninitialized iterator. You must initialize it using a QJsonArray function like QJsonArray::constBegin(), QJsonArray::constEnd(), or QJsonArray::insert() before you can start iterating.

Most QJsonArray functions accept an integer index rather than an iterator. For that reason, iterators are rarely useful in connection with QJsonArray. One place where STL-style iterators do make sense is as arguments to generic algorithms.

Multiple iterators can be used on the same array. However, be aware that any non-const function call performed on the QJsonArray will render all existing iterators undefined.

Methods

impl ConstIterator

fn a(&self) -> *const JsonArray

C++ method: const QJsonArray* QJsonArray::const_iterator::a() const

fn i(&self) -> c_int

C++ method: int QJsonArray::const_iterator::i() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: ConstIteratorNewArgs

C++ method: QJsonArray::const_iterator::const_iterator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::json_array::ConstIterator
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::ConstIterator>
C++ method: [constructor] void QJsonArray::const_iterator::const_iterator()

Variant 2

Rust arguments:
1) fn new((*const ::json_array::JsonArray, libc::c_int, cpp_utils::AsStruct)) -> ::json_array::ConstIterator
2) fn new((*const ::json_array::JsonArray, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::ConstIterator>
C++ method: [constructor] void QJsonArray::const_iterator::const_iterator(const QJsonArray* array, int index)

Variant 3

Rust arguments:
1) fn new((&::json_array::ConstIterator, cpp_utils::AsStruct)) -> ::json_array::ConstIterator
2) fn new((&::json_array::ConstIterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::ConstIterator>
C++ method: [constructor] void QJsonArray::const_iterator::const_iterator(const QJsonArray::const_iterator& o)

Variant 4

Rust arguments:
1) fn new((&::json_array::Iterator, cpp_utils::AsStruct)) -> ::json_array::ConstIterator
2) fn new((&::json_array::Iterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::ConstIterator>
C++ method: [constructor] void QJsonArray::const_iterator::const_iterator(const QJsonArray::iterator& o)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpAddArgs<'largs>

C++ method: QJsonArray::const_iterator::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn op_add(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::const_iterator::operator+(int j) const

C++ documentation:

Returns an iterator to the item at j positions forward from this iterator.

See also operator-() and operator+=().

fn op_add_assign<'l0>(&'l0 mut self, j: c_int) -> &'l0 mut ConstIterator

C++ method: QJsonArray::const_iterator& QJsonArray::const_iterator::operator+=(int j)

C++ documentation:

Advances the iterator by j items.

See also operator-=() and operator+().

fn op_dec<'l0>(&'l0 mut self) -> &'l0 mut ConstIterator

C++ method: QJsonArray::const_iterator& QJsonArray::const_iterator::operator--()

C++ documentation:

The prefix -- operator (--it) makes the preceding item current and returns an iterator to the new current item.

Calling this function on QAssociativeIterable::begin() leads to undefined results.

See also operator++().

fn op_dec_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpDecPostfixArgs<'largs>

C++ method: QJsonArray::const_iterator::operator--

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::const_iterator::operator--(int arg1)

C++ documentation:

This is an overloaded function.

The postfix -- operator (it--) makes the preceding item current and returns an iterator to the previously current item.

fn op_eq(&self, o: &ConstIterator) -> bool

C++ method: bool QJsonArray::const_iterator::operator==(const QJsonArray::const_iterator& o) const

C++ documentation:

Returns true if other points to the same item as this iterator; otherwise returns false.

See also operator!=().

fn op_ge(&self, other: &ConstIterator) -> bool

C++ method: bool QJsonArray::const_iterator::operator>=(const QJsonArray::const_iterator& other) const

C++ documentation:

Returns true if the item pointed to by this iterator is greater than or equal to the item pointed to by the other iterator.

fn op_gt(&self, other: &ConstIterator) -> bool

C++ method: bool QJsonArray::const_iterator::operator>(const QJsonArray::const_iterator& other) const

C++ documentation:

Returns true if the item pointed to by this iterator is greater than the item pointed to by the other iterator.

fn op_inc<'l0>(&'l0 mut self) -> &'l0 mut ConstIterator

C++ method: QJsonArray::const_iterator& QJsonArray::const_iterator::operator++()

C++ documentation:

The prefix ++ operator (++it) advances the iterator to the next item in the container and returns an iterator to the new current item.

Calling this function on QAssociativeIterable::end() leads to undefined results.

See also operator--().

fn op_inc_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpIncPostfixArgs<'largs>

C++ method: QJsonArray::const_iterator::operator++

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::const_iterator::operator++(int arg1)

C++ documentation:

This is an overloaded function.

The postfix ++ operator (it++) advances the iterator to the next item in the container and returns an iterator to the previously current item.

fn op_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpIndexArgs<'largs>

C++ method: QJsonArray::const_iterator::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn op_index(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonArray::const_iterator::operator[](int j) const

C++ documentation:

Returns the item at offset j from the item pointed to by this iterator (the item at position *this + j).

This function is provided to make QJsonArray iterators behave like C++ pointers.

See also operator+().

fn op_indirection<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpIndirectionArgs<'largs>

C++ method: QJsonArray::const_iterator::operator*

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_indirection(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn op_indirection(&self, cpp_utils::AsStruct) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonArray::const_iterator::operator*() const

C++ documentation:

Returns the current value, converted to a QVariant.

fn op_le(&self, other: &ConstIterator) -> bool

C++ method: bool QJsonArray::const_iterator::operator<=(const QJsonArray::const_iterator& other) const

C++ documentation:

Returns true if the item pointed to by this iterator is less than or equal to the item pointed to by the other iterator.

fn op_lt(&self, other: &ConstIterator) -> bool

C++ method: bool QJsonArray::const_iterator::operator<(const QJsonArray::const_iterator& other) const

C++ documentation:

Returns true if the item pointed to by this iterator is less than the item pointed to by the other iterator.

fn op_neq(&self, o: &ConstIterator) -> bool

C++ method: bool QJsonArray::const_iterator::operator!=(const QJsonArray::const_iterator& o) const

C++ documentation:

Returns true if other points to a different item than this iterator; otherwise returns false.

See also operator==().

fn op_struct_deref<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpStructDerefArgs<'largs>

C++ method: QJsonArray::const_iterator::operator->

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_struct_deref(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value_ptr::JsonValuePtr>
2) fn op_struct_deref(&self, cpp_utils::AsStruct) -> ::json_value_ptr::JsonValuePtr
C++ method: QJsonValuePtr QJsonArray::const_iterator::operator->() const

C++ documentation:

Returns a pointer to the current result.

fn op_sub<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpSubArgs<'largs>

C++ method: QJsonArray::const_iterator::operator-

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn op_sub(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn op_sub(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::const_iterator::operator-(int j) const

C++ documentation:

Returns an iterator to the item at j positions backward from this iterator.

See also operator+() and operator-=().

Variant 2

Rust arguments: fn op_sub(&self, &::json_array::ConstIterator) -> libc::c_int
C++ method: int QJsonArray::const_iterator::operator-(QJsonArray::const_iterator j) const

fn op_sub_assign<'l0>(&'l0 mut self, j: c_int) -> &'l0 mut ConstIterator

C++ method: QJsonArray::const_iterator& QJsonArray::const_iterator::operator-=(int j)

C++ documentation:

Makes the iterator go back by j items.

See also operator+=() and operator-().

fn set_a(&mut self, value: *const JsonArray)

C++ method: void QJsonArray::const_iterator::set_a(const QJsonArray* value)

fn set_i(&mut self, value: c_int)

C++ method: void QJsonArray::const_iterator::set_i(int value)