Struct qt_core::json_array::JsonArray

pub struct JsonArray {
    // some fields omitted
}

C++ type: QJsonArray

C++ documentation:

The QJsonArray class encapsulates a JSON array.

A JSON array is a list of values. The list can be manipulated by inserting and removing QJsonValue's from the array.

A QJsonArray can be converted to and from a QVariantList. You can query the number of entries with size(), insert(), and removeAt() entries from it and iterate over its content using the standard C++ iterator pattern.

QJsonArray is an implicitly shared class and shares the data with the document it has been created from as long as it is not being modified.

You can convert the array to and from text based JSON through QJsonDocument.

Methods

impl JsonArray

fn append(&mut self, value: &JsonValue)

C++ method: void QJsonArray::append(const QJsonValue& value)

C++ documentation:

Inserts value at the end of the array.

See also prepend() and insert().

fn at<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayAtArgs<'largs>

C++ method: QJsonArray::at

This is an overloaded function. Available variants:

Rust arguments:
1) fn at(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn at(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonArray::at(int i) const

C++ documentation:

Returns a QJsonValue representing the value for index i.

The returned QJsonValue is Undefined, if i is out of bounds.

fn begin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayBeginArgs<'largs>

C++ method: QJsonArray::begin

This is an overloaded function. Available variants:

Rust arguments:
1) fn begin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn begin(&self, cpp_utils::AsStruct) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::begin() const

C++ documentation:

This is an overloaded function.

fn begin_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonArrayBeginMutArgs<'largs>

C++ method: QJsonArray::begin

This is an overloaded function. Available variants:

Rust arguments:
1) fn begin_mut(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::Iterator>
2) fn begin_mut(&mut self, cpp_utils::AsStruct) -> ::json_array::Iterator
C++ method: QJsonArray::iterator QJsonArray::begin()

C++ documentation:

Returns an STL-style iterator pointing to the first item in the array.

See also constBegin() and end().

fn const_begin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayConstBeginArgs<'largs>

C++ method: QJsonArray::constBegin

This is an overloaded function. Available variants:

Rust arguments:
1) fn const_begin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn const_begin(&self, cpp_utils::AsStruct) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::constBegin() const

C++ documentation:

Returns a const STL-style iterator pointing to the first item in the array.

See also begin() and constEnd().

fn const_end<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayConstEndArgs<'largs>

C++ method: QJsonArray::constEnd

This is an overloaded function. Available variants:

Rust arguments:
1) fn const_end(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn const_end(&self, cpp_utils::AsStruct) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::constEnd() const

C++ documentation:

Returns a const STL-style iterator pointing to the imaginary item after the last item in the array.

See also constBegin() and end().

fn contains(&self, element: &JsonValue) -> bool

C++ method: bool QJsonArray::contains(const QJsonValue& element) const

C++ documentation:

Returns true if the array contains an occurrence of value, otherwise false.

See also count().

fn count(&self) -> c_int

C++ method: int QJsonArray::count() const

C++ documentation:

Same as size().

See also size().

fn empty(&self) -> bool

C++ method: bool QJsonArray::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty() and returns true if the array is empty.

fn end<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayEndArgs<'largs>

C++ method: QJsonArray::end

This is an overloaded function. Available variants:

Rust arguments:
1) fn end(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::ConstIterator>
2) fn end(&self, cpp_utils::AsStruct) -> ::json_array::ConstIterator
C++ method: QJsonArray::const_iterator QJsonArray::end() const

C++ documentation:

This is an overloaded function.

fn end_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonArrayEndMutArgs<'largs>

C++ method: QJsonArray::end

This is an overloaded function. Available variants:

Rust arguments:
1) fn end_mut(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::Iterator>
2) fn end_mut(&mut self, cpp_utils::AsStruct) -> ::json_array::Iterator
C++ method: QJsonArray::iterator QJsonArray::end()

C++ documentation:

Returns an STL-style iterator pointing to the imaginary item after the last item in the array.

See also begin() and constEnd().

fn erase<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonArrayEraseArgs<'largs>

C++ method: QJsonArray::erase

This is an overloaded function. Available variants:

Rust arguments:
1) fn erase(&mut self, (&::json_array::Iterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::Iterator>
2) fn erase(&mut self, (&::json_array::Iterator, cpp_utils::AsStruct)) -> ::json_array::Iterator
C++ method: QJsonArray::iterator QJsonArray::erase(QJsonArray::iterator it)

C++ documentation:

Removes the item pointed to by it, and returns an iterator pointing to the next item.

See also removeAt().

fn first<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayFirstArgs<'largs>

C++ method: QJsonArray::first

This is an overloaded function. Available variants:

Rust arguments:
1) fn first(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn first(&self, cpp_utils::AsStruct) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonArray::first() const

C++ documentation:

Returns the first value stored in the array.

Same as at(0).

See also at().

fn from_string_list<Args>(args: Args) -> Args::ReturnType where Args: JsonArrayFromStringListArgs

C++ method: QJsonArray::fromStringList

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_string_list((&::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::JsonArray>
2) fn from_string_list((&::string_list::StringList, cpp_utils::AsStruct)) -> ::json_array::JsonArray
C++ method: static QJsonArray QJsonArray::fromStringList(const QStringList& list)

C++ documentation:

Converts the string list list to a QJsonArray.

The values in list will be converted to JSON values.

See also toVariantList() and QJsonValue::fromVariant().

fn from_variant_list<Args>(args: Args) -> Args::ReturnType where Args: JsonArrayFromVariantListArgs

C++ method: QJsonArray::fromVariantList

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_variant_list((&::list::ListVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::JsonArray>
2) fn from_variant_list((&::list::ListVariantRef, cpp_utils::AsStruct)) -> ::json_array::JsonArray
C++ method: static QJsonArray QJsonArray::fromVariantList(const QList<QVariant>& list)

C++ documentation:

Converts the variant list list to a QJsonArray.

The QVariant values in list will be converted to JSON values.

See also toVariantList() and QJsonValue::fromVariant().

fn insert<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonArrayInsertArgs<'largs>

C++ method: QJsonArray::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn insert(&mut self, (&::json_array::Iterator, &::json_value::JsonValue, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::Iterator>
2) fn insert(&mut self, (&::json_array::Iterator, &::json_value::JsonValue, cpp_utils::AsStruct)) -> ::json_array::Iterator
C++ method: QJsonArray::iterator QJsonArray::insert(QJsonArray::iterator before, const QJsonValue& value)

C++ documentation:

Inserts value before the position pointed to by before, and returns an iterator pointing to the newly inserted item.

See also erase() and insert().

Variant 2

Rust arguments: fn insert(&mut self, (libc::c_int, &::json_value::JsonValue)) -> ()
C++ method: void QJsonArray::insert(int i, const QJsonValue& value)

C++ documentation:

Inserts value at index position i in the array. If i is 0, the value is prepended to the array. If i is size(), the value is appended to the array.

See also append(), prepend(), replace(), and removeAt().

fn is_empty(&self) -> bool

C++ method: bool QJsonArray::isEmpty() const

C++ documentation:

Returns true if the object is empty. This is the same as size() == 0.

See also size().

fn last<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayLastArgs<'largs>

C++ method: QJsonArray::last

This is an overloaded function. Available variants:

Rust arguments:
1) fn last(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn last(&self, cpp_utils::AsStruct) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonArray::last() const

C++ documentation:

Returns the last value stored in the array.

Same as at(size() - 1).

See also at().

fn new<Args>(args: Args) -> Args::ReturnType where Args: JsonArrayNewArgs

C++ method: QJsonArray::QJsonArray

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::json_array::JsonArray
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::JsonArray>
C++ method: [constructor] void QJsonArray::QJsonArray()

Variant 2

Rust arguments:
1) fn new((&::json_array::JsonArray, cpp_utils::AsStruct)) -> ::json_array::JsonArray
2) fn new((&::json_array::JsonArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::JsonArray>
C++ method: [constructor] void QJsonArray::QJsonArray(const QJsonArray& other)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayOpAddArgs<'largs>

C++ method: QJsonArray::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::json_value::JsonValue, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::JsonArray>
2) fn op_add(&self, (&::json_value::JsonValue, cpp_utils::AsStruct)) -> ::json_array::JsonArray
C++ method: QJsonArray QJsonArray::operator+(const QJsonValue& v) const

C++ documentation:

Returns an array that contains all the items in this array followed by the provided value.

This function was introduced in Qt 5.3.

See also operator+=().

fn op_add_assign<'l0, 'l1>(&'l0 mut self, v: &'l1 JsonValue) -> &'l0 mut JsonArray

C++ method: QJsonArray& QJsonArray::operator+=(const QJsonValue& v)

C++ documentation:

Appends value to the array, and returns a reference to the array itself.

This function was introduced in Qt 5.3.

See also append() and operator<<().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 JsonArray) -> &'l0 mut JsonArray

C++ method: QJsonArray& QJsonArray::operator=(const QJsonArray& other)

C++ documentation:

Assigns other to this array.

fn op_eq(&self, other: &JsonArray) -> bool

C++ method: bool QJsonArray::operator==(const QJsonArray& other) const

C++ documentation:

Returns true if this array is equal to other.

fn op_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayOpIndexArgs<'largs>

C++ method: QJsonArray::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn op_index(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonArray::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at().

fn op_index_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonArrayOpIndexMutArgs<'largs>

C++ method: QJsonArray::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index_mut(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value_ref::JsonValueRef>
2) fn op_index_mut(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_value_ref::JsonValueRef
C++ method: QJsonValueRef QJsonArray::operator[](int i)

C++ documentation:

Returns the value at index position i as a modifiable reference. i must be a valid index position in the array (i.e., 0 <= i < size()).

The return value is of type and QJsonObject. When you get an object of type use it as if it were a reference to a QJsonValue. If you assign to it, the assignment will apply to the character in the QJsonArray of QJsonObject from which you got the reference.

See also at().

fn op_neq(&self, other: &JsonArray) -> bool

C++ method: bool QJsonArray::operator!=(const QJsonArray& other) const

C++ documentation:

Returns true if this array is not equal to other.

fn op_shl<'l0, 'l1>(&'l0 mut self, v: &'l1 JsonValue) -> &'l0 mut JsonArray

C++ method: QJsonArray& QJsonArray::operator<<(const QJsonValue& v)

C++ documentation:

Appends value to the array, and returns a reference to the array itself.

This function was introduced in Qt 5.3.

See also operator+=() and append().

fn pop_back(&mut self)

C++ method: void QJsonArray::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast(). The array must not be empty. If the array can be empty, call isEmpty() before calling this function.

fn pop_front(&mut self)

C++ method: void QJsonArray::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst(). The array must not be empty. If the array can be empty, call isEmpty() before calling this function.

fn prepend(&mut self, value: &JsonValue)

C++ method: void QJsonArray::prepend(const QJsonValue& value)

C++ documentation:

Inserts value at the beginning of the array.

This is the same as insert(0, value) and will prepend value to the array.

See also append() and insert().

fn push_back(&mut self, t: &JsonValue)

C++ method: void QJsonArray::push_back(const QJsonValue& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value) and will append value to the array.

fn push_front(&mut self, t: &JsonValue)

C++ method: void QJsonArray::push_front(const QJsonValue& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value) and will prepend value to the array.

fn remove_at(&mut self, i: c_int)

C++ method: void QJsonArray::removeAt(int i)

C++ documentation:

Removes the value at index position i. i must be a valid index position in the array (i.e., 0 <= i < size()).

See also insert() and replace().

fn remove_first(&mut self)

C++ method: void QJsonArray::removeFirst()

C++ documentation:

Removes the first item in the array. Calling this function is equivalent to calling removeAt(0). The array must not be empty. If the array can be empty, call isEmpty() before calling this function.

See also removeAt() and removeLast().

fn remove_last(&mut self)

C++ method: void QJsonArray::removeLast()

C++ documentation:

Removes the last item in the array. Calling this function is equivalent to calling removeAt(size() - 1). The array must not be empty. If the array can be empty, call isEmpty() before calling this function.

See also removeAt() and removeFirst().

fn replace(&mut self, i: c_int, value: &JsonValue)

C++ method: void QJsonArray::replace(int i, const QJsonValue& value)

C++ documentation:

Replaces the item at index position i with value. i must be a valid index position in the array (i.e., 0 <= i < size()).

See also operator[]() and removeAt().

fn size(&self) -> c_int

C++ method: int QJsonArray::size() const

C++ documentation:

Returns the number of values stored in the array.

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonArrayTakeAtArgs<'largs>

C++ method: QJsonArray::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonArray::takeAt(int i)

C++ documentation:

Removes the item at index position i and returns it. i must be a valid index position in the array (i.e., 0 <= i < size()).

If you don't use the return value, removeAt() is more efficient.

See also removeAt().

fn to_variant_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonArrayToVariantListArgs<'largs>

C++ method: QJsonArray::toVariantList

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_variant_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListVariantRef>
2) fn to_variant_list(&self, cpp_utils::AsStruct) -> ::list::ListVariantRef
C++ method: QList<QVariant> QJsonArray::toVariantList() const

C++ documentation:

Converts this object to a QVariantList.

Returns the created map.