Module qt_core::json_document::overloading

Traits

JsonDocumentArrayArgs
This trait represents a set of arguments accepted by JsonDocument::array method.

JsonDocumentFromBinaryDataArgs
This trait represents a set of arguments accepted by JsonDocument::from_binary_data method.

JsonDocumentFromJsonArgs
This trait represents a set of arguments accepted by JsonDocument::from_json method.

JsonDocumentFromRawDataArgs
This trait represents a set of arguments accepted by JsonDocument::from_raw_data method.

JsonDocumentFromVariantArgs
This trait represents a set of arguments accepted by JsonDocument::from_variant method.

JsonDocumentNewArgs
This trait represents a set of arguments accepted by JsonDocument::new method.

JsonDocumentObjectArgs
This trait represents a set of arguments accepted by JsonDocument::object method.

JsonDocumentToBinaryDataArgs
This trait represents a set of arguments accepted by JsonDocument::to_binary_data method.

JsonDocumentToJsonArgs
This trait represents a set of arguments accepted by JsonDocument::to_json method.

JsonDocumentToVariantArgs
This trait represents a set of arguments accepted by JsonDocument::to_variant method.