Struct qt_core::json_document::JsonDocument

pub struct JsonDocument {
    // some fields omitted
}

C++ type: QJsonDocument

C++ documentation:

The QJsonDocument class provides a way to read and write JSON documents.

QJsonDocument is a class that wraps a complete JSON document and can read and write this document both from a UTF-8 encoded text based representation as well as Qt's own binary format.

A JSON document can be converted from its text-based representation to a QJsonDocument using QJsonDocument::fromJson(). toJson() converts it back to text. The parser is very fast and efficient and converts the JSON to the binary representation used by Qt.

Validity of the parsed document can be queried with !isNull()

A document can be queried as to whether it contains an array or an object using isArray() and isObject(). The array or object contained in the document can be retrieved using array() or object() and then read or manipulated.

A document can also be created from a stored binary representation using fromBinaryData() or fromRawData().

Methods

impl JsonDocument

fn array<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonDocumentArrayArgs<'largs>

C++ method: QJsonDocument::array

This is an overloaded function. Available variants:

Rust arguments:
1) fn array(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::JsonArray>
2) fn array(&self, cpp_utils::AsStruct) -> ::json_array::JsonArray
C++ method: QJsonArray QJsonDocument::array() const

C++ documentation:

Returns the QJsonArray contained in the document.

Returns an empty array if the document contains an object.

See also isArray(), object(), and setArray().

fn from_binary_data<Args>(args: Args) -> Args::ReturnType where Args: JsonDocumentFromBinaryDataArgs

C++ method: QJsonDocument::fromBinaryData

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_binary_data((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
2) fn from_binary_data((&::byte_array::ByteArray, ::json_document::DataValidation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
3) fn from_binary_data((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
4) fn from_binary_data((&::byte_array::ByteArray, ::json_document::DataValidation, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
C++ method: static QJsonDocument QJsonDocument::fromBinaryData(const QByteArray& data, QJsonDocument::DataValidation validation = ?)

C++ documentation:

Creates a QJsonDocument from data.

validation decides whether the data is checked for validity before being used. By default the data is validated. If the data is not valid, the method returns a null document.

See also toBinaryData(), fromRawData(), isNull(), and DataValidation.

fn from_json<Args>(args: Args) -> Args::ReturnType where Args: JsonDocumentFromJsonArgs

C++ method: QJsonDocument::fromJson

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_json((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
2) fn from_json((&::byte_array::ByteArray, *mut ::json_parse_error::JsonParseError, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
3) fn from_json((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
4) fn from_json((&::byte_array::ByteArray, *mut ::json_parse_error::JsonParseError, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
C++ method: static QJsonDocument QJsonDocument::fromJson(const QByteArray& json, QJsonParseError* error = ?)

C++ documentation:

Parses a UTF-8 encoded JSON document and creates a QJsonDocument from it.

json contains the json document to be parsed.

The optional error variable can be used to pass in a QJsonParseError data structure that will contain information about possible errors encountered during parsing.

See also toJson() and QJsonParseError.

fn from_raw_data<Args>(args: Args) -> Args::ReturnType where Args: JsonDocumentFromRawDataArgs

C++ method: QJsonDocument::fromRawData

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_raw_data((*const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
2) fn from_raw_data((*const libc::c_char, libc::c_int, ::json_document::DataValidation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
3) fn from_raw_data((*const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
4) fn from_raw_data((*const libc::c_char, libc::c_int, ::json_document::DataValidation, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
C++ method: static QJsonDocument QJsonDocument::fromRawData(const char* data, int size, QJsonDocument::DataValidation validation = ?)

C++ documentation:

Creates a QJsonDocument that uses the first size bytes from data. It assumes data contains a binary encoded JSON document. The created document does not take ownership of data and the caller has to guarantee that data will not be deleted or modified as long as any QJsonDocument, QJsonObject or QJsonArray still references the data.

data has to be aligned to a 4 byte boundary.

validation decides whether the data is checked for validity before being used. By default the data is validated. If the data is not valid, the method returns a null document.

Returns a QJsonDocument representing the data.

See also rawData(), fromBinaryData(), isNull(), and DataValidation.

fn from_variant<Args>(args: Args) -> Args::ReturnType where Args: JsonDocumentFromVariantArgs

C++ method: QJsonDocument::fromVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_variant((&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
2) fn from_variant((&::variant::Variant, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
C++ method: static QJsonDocument QJsonDocument::fromVariant(const QVariant& variant)

C++ documentation:

Creates a QJsonDocument from the QVariant variant.

If the variant contains any other type than a QVariantMap, QVariantList or QStringList, the returned document document is invalid.

See also toVariant().

fn is_array(&self) -> bool

C++ method: bool QJsonDocument::isArray() const

C++ documentation:

Returns true if the document contains an array.

See also array() and isObject().

fn is_empty(&self) -> bool

C++ method: bool QJsonDocument::isEmpty() const

C++ documentation:

Returns true if the document doesn't contain any data.

fn is_null(&self) -> bool

C++ method: bool QJsonDocument::isNull() const

C++ documentation:

returns true if this document is null.

Null documents are documents created through the default constructor.

Documents created from UTF-8 encoded text or the binary format are validated during parsing. If validation fails, the returned document will also be null.

fn is_object(&self) -> bool

C++ method: bool QJsonDocument::isObject() const

C++ documentation:

Returns true if the document contains an object.

See also object() and isArray().

fn new<Args>(args: Args) -> Args::ReturnType where Args: JsonDocumentNewArgs

C++ method: QJsonDocument::QJsonDocument

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::json_document::JsonDocument
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::json_document::JsonDocument>
C++ method: [constructor] void QJsonDocument::QJsonDocument()

Variant 2

Rust arguments:
1) fn new((&::json_array::JsonArray, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
2) fn new((&::json_array::JsonArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
C++ method: [constructor] void QJsonDocument::QJsonDocument(const QJsonArray& array)

Variant 3

Rust arguments:
1) fn new((&::json_document::JsonDocument, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
2) fn new((&::json_document::JsonDocument, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
C++ method: [constructor] void QJsonDocument::QJsonDocument(const QJsonDocument& other)

Variant 4

Rust arguments:
1) fn new((&::json_object::JsonObject, cpp_utils::AsStruct)) -> ::json_document::JsonDocument
2) fn new((&::json_object::JsonObject, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_document::JsonDocument>
C++ method: [constructor] void QJsonDocument::QJsonDocument(const QJsonObject& object)

fn object<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonDocumentObjectArgs<'largs>

C++ method: QJsonDocument::object

This is an overloaded function. Available variants:

Rust arguments:
1) fn object(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::JsonObject>
2) fn object(&self, cpp_utils::AsStruct) -> ::json_object::JsonObject
C++ method: QJsonObject QJsonDocument::object() const

C++ documentation:

Returns the QJsonObject contained in the document.

Returns an empty object if the document contains an array.

See also isObject(), array(), and setObject().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 JsonDocument) -> &'l0 mut JsonDocument

C++ method: QJsonDocument& QJsonDocument::operator=(const QJsonDocument& other)

C++ documentation:

Assigns the other document to this QJsonDocument. Returns a reference to this object.

fn op_eq(&self, other: &JsonDocument) -> bool

C++ method: bool QJsonDocument::operator==(const QJsonDocument& other) const

C++ documentation:

Returns true if the other document is equal to this document.

fn op_neq(&self, other: &JsonDocument) -> bool

C++ method: bool QJsonDocument::operator!=(const QJsonDocument& other) const

C++ documentation:

returns true if other is not equal to this document

fn raw_data(&self, size: *mut c_int) -> *const c_char

C++ method: const char* QJsonDocument::rawData(int* size) const

C++ documentation:

Returns the raw binary representation of the data size will contain the size of the returned data.

This method is useful to e.g. stream the JSON document in it's binary form to a file.

fn set_array(&mut self, array: &JsonArray)

C++ method: void QJsonDocument::setArray(const QJsonArray& array)

C++ documentation:

Sets array as the main object of this document.

See also setObject() and array().

fn set_object(&mut self, object: &JsonObject)

C++ method: void QJsonDocument::setObject(const QJsonObject& object)

C++ documentation:

Sets object as the main object of this document.

See also setArray() and object().

fn to_binary_data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonDocumentToBinaryDataArgs<'largs>

C++ method: QJsonDocument::toBinaryData

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_binary_data(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_binary_data(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QJsonDocument::toBinaryData() const

C++ documentation:

Returns a binary representation of the document.

The binary representation is also the native format used internally in Qt, and is very efficient and fast to convert to and from.

The binary format can be stored on disk and interchanged with other applications or computers. fromBinaryData() can be used to convert it back into a JSON document.

See also fromBinaryData().

fn to_json<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonDocumentToJsonArgs<'largs>

C++ method: QJsonDocument::toJson

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_json(&self, (::json_document::JsonFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_json(&self, (::json_document::JsonFormat, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray QJsonDocument::toJson(QJsonDocument::JsonFormat format) const

C++ documentation:

Converts the QJsonDocument to a UTF-8 encoded JSON document in the provided format.

See also fromJson() and JsonFormat.

Variant 2

Rust arguments:
1) fn to_json(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_json(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QJsonDocument::toJson() const

fn to_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonDocumentToVariantArgs<'largs>

C++ method: QJsonDocument::toVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn to_variant(&self, cpp_utils::AsStruct) -> ::variant::Variant
C++ method: QVariant QJsonDocument::toVariant() const

C++ documentation:

Returns a QVariant representing the Json document.

The returned variant will be a QVariantList if the document is a QJsonArray and a QVariantMap if the document is a QJsonObject.

See also fromVariant() and QJsonValue::toVariant().