Struct qt_core::json_object::Iterator

pub struct Iterator {
    // some fields omitted
}

C++ type: QJsonObject::iterator

C++ documentation:

The QJsonObject::iterator class provides an STL-style non-const iterator for QJsonObject.

QJsonObject::iterator allows you to iterate over a QJsonObject and to modify the value (but not the key) stored under a particular key. If you want to iterate over a const QJsonObject, you should use QJsonObject::const_iterator. It is generally good practice to use QJsonObject::const_iterator on a non-const QJsonObject as well, unless you need to change the QJsonObject through the iterator. Const iterators are slightly faster, and improve code readability.

The default QJsonObject::iterator constructor creates an uninitialized iterator. You must initialize it using a QJsonObject function like QJsonObject::begin(), QJsonObject::end(), or QJsonObject::find() before you can start iterating.

Multiple iterators can be used on the same object. Existing iterators will however become dangling once the object gets modified.

Methods

impl Iterator

fn key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorKeyArgs<'largs>

C++ method: QJsonObject::iterator::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn key(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QJsonObject::iterator::key() const

C++ documentation:

Returns the current item's key as a const reference.

There is no direct way of changing an item's key through an iterator, although it can be done by calling QHash::erase() followed by QHash::insert() or QHash::insertMulti().

See also value().

fn new<Args>(args: Args) -> Args::ReturnType where Args: IteratorNewArgs

C++ method: QJsonObject::iterator::iterator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::json_object::Iterator
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::Iterator>
C++ method: [constructor] void QJsonObject::iterator::iterator()

Variant 2

Rust arguments:
1) fn new((*mut ::json_object::JsonObject, libc::c_int, cpp_utils::AsStruct)) -> ::json_object::Iterator
2) fn new((*mut ::json_object::JsonObject, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
C++ method: [constructor] void QJsonObject::iterator::iterator(QJsonObject* obj, int index)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorOpAddArgs<'largs>

C++ method: QJsonObject::iterator::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn op_add(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::iterator::operator+(int j) const

C++ documentation:

Returns an iterator to the item at j positions forward from this iterator. (If j is negative, the iterator goes backward.)

This operation can be slow for large j values.

See also operator-().

fn op_add_assign<'l0>(&'l0 mut self, j: c_int) -> &'l0 mut Iterator

C++ method: QJsonObject::iterator& QJsonObject::iterator::operator+=(int j)

C++ documentation:

Advances the iterator by j items. (If j is negative, the iterator goes backward.)

See also operator-=() and operator+().

fn op_dec<'l0>(&'l0 mut self) -> &'l0 mut Iterator

C++ method: QJsonObject::iterator& QJsonObject::iterator::operator--()

C++ documentation:

The prefix -- operator (--i) makes the preceding item current and returns an iterator pointing to the new current item.

Calling this function on QHash::begin() leads to undefined results.

See also operator++().

fn op_dec_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IteratorOpDecPostfixArgs<'largs>

C++ method: QJsonObject::iterator::operator--

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::iterator::operator--(int arg1)

C++ documentation:

This is an overloaded function.

The postfix -- operator (i--) makes the preceding item current and returns an iterator pointing to the previously current item.

fn op_eq<'largs, Args>(&'largs self, args: Args) -> bool where Args: IteratorOpEqArgs<'largs>

C++ method: QJsonObject::iterator::operator==

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_eq(&self, &::json_object::ConstIterator) -> bool
C++ method: bool QJsonObject::iterator::operator==(const QJsonObject::const_iterator& other) const

C++ documentation:

Returns true if other points to the same item as this iterator; otherwise returns false.

See also operator!=().

Variant 2

Rust arguments: fn op_eq(&self, &::json_object::Iterator) -> bool
C++ method: bool QJsonObject::iterator::operator==(const QJsonObject::iterator& other) const

C++ documentation:

Returns true if other points to the same item as this iterator; otherwise returns false.

See also operator!=().

fn op_inc<'l0>(&'l0 mut self) -> &'l0 mut Iterator

C++ method: QJsonObject::iterator& QJsonObject::iterator::operator++()

C++ documentation:

The prefix ++ operator (++i) advances the iterator to the next item in the hash and returns an iterator to the new current item.

Calling this function on QHash::end() leads to undefined results.

See also operator--().

fn op_inc_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: IteratorOpIncPostfixArgs<'largs>

C++ method: QJsonObject::iterator::operator++

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::iterator::operator++(int arg1)

C++ documentation:

This is an overloaded function.

The postfix ++ operator (i++) advances the iterator to the next item in the hash and returns an iterator to the previously current item.

fn op_indirection<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorOpIndirectionArgs<'largs>

C++ method: QJsonObject::iterator::operator*

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_indirection(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value_ref::JsonValueRef>
2) fn op_indirection(&self, cpp_utils::AsStruct) -> ::json_value_ref::JsonValueRef
C++ method: QJsonValueRef QJsonObject::iterator::operator*() const

C++ documentation:

Returns a modifiable reference to the current item's value.

Same as value().

See also key().

fn op_neq<'largs, Args>(&'largs self, args: Args) -> bool where Args: IteratorOpNeqArgs<'largs>

C++ method: QJsonObject::iterator::operator!=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_neq(&self, &::json_object::ConstIterator) -> bool
C++ method: bool QJsonObject::iterator::operator!=(const QJsonObject::const_iterator& other) const

C++ documentation:

Returns true if other points to a different item than this iterator; otherwise returns false.

See also operator==().

Variant 2

Rust arguments: fn op_neq(&self, &::json_object::Iterator) -> bool
C++ method: bool QJsonObject::iterator::operator!=(const QJsonObject::iterator& other) const

C++ documentation:

Returns true if other points to a different item than this iterator; otherwise returns false.

See also operator==().

fn op_struct_deref<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorOpStructDerefArgs<'largs>

C++ method: QJsonObject::iterator::operator->

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_struct_deref(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value_ref_ptr::JsonValueRefPtr>
2) fn op_struct_deref(&self, cpp_utils::AsStruct) -> ::json_value_ref_ptr::JsonValueRefPtr
C++ method: QJsonValueRefPtr QJsonObject::iterator::operator->() const

C++ documentation:

Returns a pointer to the current item's value.

See also value().

fn op_sub<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorOpSubArgs<'largs>

C++ method: QJsonObject::iterator::operator-

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_sub(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn op_sub(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::iterator::operator-(int j) const

C++ documentation:

Returns an iterator to the item at j positions backward from this iterator. (If j is negative, the iterator goes forward.)

This operation can be slow for large j values.

See also operator+().

fn op_sub_assign<'l0>(&'l0 mut self, j: c_int) -> &'l0 mut Iterator

C++ method: QJsonObject::iterator& QJsonObject::iterator::operator-=(int j)

C++ documentation:

Makes the iterator go back by j items. (If j is negative, the iterator goes forward.)

See also operator+=() and operator-().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: IteratorValueArgs<'largs>

C++ method: QJsonObject::iterator::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value_ref::JsonValueRef>
2) fn value(&self, cpp_utils::AsStruct) -> ::json_value_ref::JsonValueRef
C++ method: QJsonValueRef QJsonObject::iterator::value() const

C++ documentation:

Returns a modifiable reference to the current item's value.

You can change the value of an item by using value() on the left side of an assignment, for example:

if (i.key() == "Hello")
    i.value() = "Bonjour";

See also key() and operator*().