Struct qt_core::json_object::JsonObject

pub struct JsonObject {
    // some fields omitted
}

C++ type: QJsonObject

C++ documentation:

The QJsonObject class encapsulates a JSON object.

A JSON object is a list of key value pairs, where the keys are unique strings and the values are represented by a QJsonValue.

A QJsonObject can be converted to and from a QVariantMap. You can query the number of (key, value) pairs with size(), insert(), and remove() entries from it and iterate over its content using the standard C++ iterator pattern.

QJsonObject is an implicitly shared class, and shares the data with the document it has been created from as long as it is not being modified.

You can convert the object to and from text based JSON through QJsonDocument.

Methods

impl JsonObject

fn begin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectBeginArgs<'largs>

C++ method: QJsonObject::begin

This is an overloaded function. Available variants:

Rust arguments:
1) fn begin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::ConstIterator>
2) fn begin(&self, cpp_utils::AsStruct) -> ::json_object::ConstIterator
C++ method: QJsonObject::const_iterator QJsonObject::begin() const

C++ documentation:

This is an overloaded function.

fn begin_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonObjectBeginMutArgs<'largs>

C++ method: QJsonObject::begin

This is an overloaded function. Available variants:

Rust arguments:
1) fn begin_mut(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn begin_mut(&mut self, cpp_utils::AsStruct) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::begin()

C++ documentation:

Returns an STL-style iterator pointing to the first item in the object.

See also constBegin() and end().

fn const_begin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectConstBeginArgs<'largs>

C++ method: QJsonObject::constBegin

This is an overloaded function. Available variants:

Rust arguments:
1) fn const_begin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::ConstIterator>
2) fn const_begin(&self, cpp_utils::AsStruct) -> ::json_object::ConstIterator
C++ method: QJsonObject::const_iterator QJsonObject::constBegin() const

C++ documentation:

Returns a const STL-style iterator pointing to the first item in the object.

See also begin() and constEnd().

fn const_end<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectConstEndArgs<'largs>

C++ method: QJsonObject::constEnd

This is an overloaded function. Available variants:

Rust arguments:
1) fn const_end(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::ConstIterator>
2) fn const_end(&self, cpp_utils::AsStruct) -> ::json_object::ConstIterator
C++ method: QJsonObject::const_iterator QJsonObject::constEnd() const

C++ documentation:

Returns a const STL-style iterator pointing to the imaginary item after the last item in the object.

See also constBegin() and end().

fn const_find<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectConstFindArgs<'largs>

C++ method: QJsonObject::constFind

This is an overloaded function. Available variants:

Rust arguments:
1) fn const_find(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::ConstIterator>
2) fn const_find(&self, (&::string::String, cpp_utils::AsStruct)) -> ::json_object::ConstIterator
C++ method: QJsonObject::const_iterator QJsonObject::constFind(const QString& key) const

C++ documentation:

Returns a const iterator pointing to the item with key key in the map.

If the map contains no item with key key, the function returns constEnd().

fn contains(&self, key: &String) -> bool

C++ method: bool QJsonObject::contains(const QString& key) const

C++ documentation:

Returns true if the object contains key key.

See also insert(), remove(), and take().

fn count(&self) -> c_int

C++ method: int QJsonObject::count() const

C++ documentation:

This is an overloaded function.

Same as size().

fn empty(&self) -> bool

C++ method: bool QJsonObject::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the object is empty; otherwise returning false.

fn end<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectEndArgs<'largs>

C++ method: QJsonObject::end

This is an overloaded function. Available variants:

Rust arguments:
1) fn end(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::ConstIterator>
2) fn end(&self, cpp_utils::AsStruct) -> ::json_object::ConstIterator
C++ method: QJsonObject::const_iterator QJsonObject::end() const

C++ documentation:

This is an overloaded function.

fn end_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonObjectEndMutArgs<'largs>

C++ method: QJsonObject::end

This is an overloaded function. Available variants:

Rust arguments:
1) fn end_mut(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn end_mut(&mut self, cpp_utils::AsStruct) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::end()

C++ documentation:

Returns an STL-style iterator pointing to the imaginary item after the last item in the object.

See also begin() and constEnd().

fn erase<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonObjectEraseArgs<'largs>

C++ method: QJsonObject::erase

This is an overloaded function. Available variants:

Rust arguments:
1) fn erase(&mut self, (&::json_object::Iterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn erase(&mut self, (&::json_object::Iterator, cpp_utils::AsStruct)) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::erase(QJsonObject::iterator it)

C++ documentation:

Removes the (key, value) pair pointed to by the iterator it from the map, and returns an iterator to the next item in the map.

See also remove().

fn find<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectFindArgs<'largs>

C++ method: QJsonObject::find

This is an overloaded function. Available variants:

Rust arguments:
1) fn find(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::ConstIterator>
2) fn find(&self, (&::string::String, cpp_utils::AsStruct)) -> ::json_object::ConstIterator
C++ method: QJsonObject::const_iterator QJsonObject::find(const QString& key) const

C++ documentation:

This is an overloaded function.

fn find_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonObjectFindMutArgs<'largs>

C++ method: QJsonObject::find

This is an overloaded function. Available variants:

Rust arguments:
1) fn find_mut(&mut self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn find_mut(&mut self, (&::string::String, cpp_utils::AsStruct)) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::find(const QString& key)

C++ documentation:

Returns an iterator pointing to the item with key key in the map.

If the map contains no item with key key, the function returns end().

fn from_variant_hash<Args>(args: Args) -> Args::ReturnType where Args: JsonObjectFromVariantHashArgs

C++ method: QJsonObject::fromVariantHash

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_variant_hash((&::hash::HashStringRefVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::JsonObject>
2) fn from_variant_hash((&::hash::HashStringRefVariantRef, cpp_utils::AsStruct)) -> ::json_object::JsonObject
C++ method: static QJsonObject QJsonObject::fromVariantHash(const QHash<QString, QVariant>& map)

C++ documentation:

Converts the variant hash hash to a QJsonObject.

The keys in hash will be used as the keys in the JSON object, and the QVariant values will be converted to JSON values.

This function was introduced in Qt 5.5.

See also fromVariantMap(), toVariantHash(), and QJsonValue::fromVariant().

fn from_variant_map<Args>(args: Args) -> Args::ReturnType where Args: JsonObjectFromVariantMapArgs

C++ method: QJsonObject::fromVariantMap

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_variant_map((&::map::MapStringRefVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::JsonObject>
2) fn from_variant_map((&::map::MapStringRefVariantRef, cpp_utils::AsStruct)) -> ::json_object::JsonObject
C++ method: static QJsonObject QJsonObject::fromVariantMap(const QMap<QString, QVariant>& map)

C++ documentation:

Converts the variant map map to a QJsonObject.

The keys in map will be used as the keys in the JSON object, and the QVariant values will be converted to JSON values.

See also fromVariantHash(), toVariantMap(), and QJsonValue::fromVariant().

fn insert<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonObjectInsertArgs<'largs>

C++ method: QJsonObject::insert

This is an overloaded function. Available variants:

Rust arguments:
1) fn insert(&mut self, (&::string::String, &::json_value::JsonValue, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::Iterator>
2) fn insert(&mut self, (&::string::String, &::json_value::JsonValue, cpp_utils::AsStruct)) -> ::json_object::Iterator
C++ method: QJsonObject::iterator QJsonObject::insert(const QString& key, const QJsonValue& value)

C++ documentation:

Inserts a new item with the key key and a value of value.

If there is already an item with the key key, then that item's value is replaced with value.

Returns an iterator pointing to the inserted item.

If the value is QJsonValue::Undefined, it will cause the key to get removed from the object. The returned iterator will then point to end().

See also remove(), take(), QJsonObject::iterator, and end().

fn is_empty(&self) -> bool

C++ method: bool QJsonObject::isEmpty() const

C++ documentation:

Returns true if the object is empty. This is the same as size() == 0.

See also size().

fn keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectKeysArgs<'largs>

C++ method: QJsonObject::keys

This is an overloaded function. Available variants:

Rust arguments:
1) fn keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn keys(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QJsonObject::keys() const

C++ documentation:

Returns a list of all keys in this object.

The list is sorted lexographically.

fn length(&self) -> c_int

C++ method: int QJsonObject::length() const

C++ documentation:

This is an overloaded function.

Same as size().

fn new<Args>(args: Args) -> Args::ReturnType where Args: JsonObjectNewArgs

C++ method: QJsonObject::QJsonObject

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::json_object::JsonObject
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::JsonObject>
C++ method: [constructor] void QJsonObject::QJsonObject()

Variant 2

Rust arguments:
1) fn new((&::json_object::JsonObject, cpp_utils::AsStruct)) -> ::json_object::JsonObject
2) fn new((&::json_object::JsonObject, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::JsonObject>
C++ method: [constructor] void QJsonObject::QJsonObject(const QJsonObject& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 JsonObject) -> &'l0 mut JsonObject

C++ method: QJsonObject& QJsonObject::operator=(const QJsonObject& other)

C++ documentation:

Assigns other to this object.

fn op_eq(&self, other: &JsonObject) -> bool

C++ method: bool QJsonObject::operator==(const QJsonObject& other) const

C++ documentation:

Returns true if other is equal to this object.

fn op_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectOpIndexArgs<'largs>

C++ method: QJsonObject::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn op_index(&self, (&::string::String, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonObject::operator[](const QString& key) const

C++ documentation:

Returns a QJsonValue representing the value for the key key.

This does the same as value().

The returned QJsonValue is QJsonValue::Undefined if the key does not exist.

See also value(), QJsonValue, and QJsonValue::isUndefined().

fn op_index_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonObjectOpIndexMutArgs<'largs>

C++ method: QJsonObject::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index_mut(&mut self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value_ref::JsonValueRef>
2) fn op_index_mut(&mut self, (&::string::String, cpp_utils::AsStruct)) -> ::json_value_ref::JsonValueRef
C++ method: QJsonValueRef QJsonObject::operator[](const QString& key)

C++ documentation:

Returns a reference to the value for key.

The return value is of type QJsonValueRef, a helper class for QJsonArray and QJsonObject. When you get an object of type QJsonValueRef, you can use it as if it were a reference to a QJsonValue. If you assign to it, the assignment will apply to the element in the QJsonArray or QJsonObject from which you got the reference.

See also value().

fn op_neq(&self, other: &JsonObject) -> bool

C++ method: bool QJsonObject::operator!=(const QJsonObject& other) const

C++ documentation:

Returns true if other is not equal to this object.

fn remove(&mut self, key: &String)

C++ method: void QJsonObject::remove(const QString& key)

C++ documentation:

Removes key from the object.

See also insert() and take().

fn size(&self) -> c_int

C++ method: int QJsonObject::size() const

C++ documentation:

Returns the number of (key, value) pairs stored in the object.

fn take<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: JsonObjectTakeArgs<'largs>

C++ method: QJsonObject::take

This is an overloaded function. Available variants:

Rust arguments:
1) fn take(&mut self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn take(&mut self, (&::string::String, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonObject::take(const QString& key)

C++ documentation:

Removes key from the object.

Returns a QJsonValue containing the value referenced by key. If key was not contained in the object, the returned QJsonValue is QJsonValue::Undefined.

See also insert(), remove(), and QJsonValue.

fn to_variant_hash<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectToVariantHashArgs<'largs>

C++ method: QJsonObject::toVariantHash

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_variant_hash(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::hash::HashStringRefVariantRef>
2) fn to_variant_hash(&self, cpp_utils::AsStruct) -> ::hash::HashStringRefVariantRef
C++ method: QHash<QString, QVariant> QJsonObject::toVariantHash() const

C++ documentation:

Converts this object to a QVariantHash.

Returns the created hash.

This function was introduced in Qt 5.5.

See also toVariantMap().

fn to_variant_map<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectToVariantMapArgs<'largs>

C++ method: QJsonObject::toVariantMap

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_variant_map(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::map::MapStringRefVariantRef>
2) fn to_variant_map(&self, cpp_utils::AsStruct) -> ::map::MapStringRefVariantRef
C++ method: QMap<QString, QVariant> QJsonObject::toVariantMap() const

C++ documentation:

Converts this object to a QVariantMap.

Returns the created map.

See also toVariantHash().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonObjectValueArgs<'largs>

C++ method: QJsonObject::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn value(&self, (&::string::String, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: QJsonValue QJsonObject::value(const QString& key) const

C++ documentation:

Returns a QJsonValue representing the value for the key key.

The returned QJsonValue is QJsonValue::Undefined if the key does not exist.

See also QJsonValue and QJsonValue::isUndefined().