Enum qt_core::json_parse_error::ParseError

pub enum ParseError {
    NoError,
    UnterminatedObject,
    MissingNameSeparator,
    UnterminatedArray,
    MissingValueSeparator,
    IllegalValue,
    TerminationByNumber,
    IllegalNumber,
    IllegalEscapeSequence,
    IllegalUTF8String,
    UnterminatedString,
    MissingObject,
    DeepNesting,
    DocumentTooLarge,
    GarbageAtEnd,
}

C++ type: QJsonParseError::ParseError

C++ documentation:

This enum describes the type of error that occurred during the parsing of a JSON document.

Variants
NoError

No error occurred (C++ enum variant: NoError = 0)

UnterminatedObject

An object is not correctly terminated with a closing curly bracket (C++ enum variant: UnterminatedObject = 1)

MissingNameSeparator

A comma separating different items is missing (C++ enum variant: MissingNameSeparator = 2)

UnterminatedArray

The array is not correctly terminated with a closing square bracket (C++ enum variant: UnterminatedArray = 3)

MissingValueSeparator

A colon separating keys from values inside objects is missing (C++ enum variant: MissingValueSeparator = 4)

IllegalValue

The value is illegal (C++ enum variant: IllegalValue = 5)

TerminationByNumber

The input stream ended while parsing a number (C++ enum variant: TerminationByNumber = 6)

IllegalNumber

The number is not well formed (C++ enum variant: IllegalNumber = 7)

IllegalEscapeSequence

An illegal escape sequence occurred in the input (C++ enum variant: IllegalEscapeSequence = 8)

IllegalUTF8String

An illegal UTF8 sequence occurred in the input (C++ enum variant: IllegalUTF8String = 9)

UnterminatedString

A string wasn't terminated with a quote (C++ enum variant: UnterminatedString = 10)

MissingObject

An object was expected but couldn't be found (C++ enum variant: MissingObject = 11)

DeepNesting

The JSON document is too deeply nested for the parser to parse it (C++ enum variant: DeepNesting = 12)

DocumentTooLarge

The JSON document is too large for the parser to parse it (C++ enum variant: DocumentTooLarge = 13)

GarbageAtEnd

The parsed document contains additional garbage characters at the end (C++ enum variant: GarbageAtEnd = 14)