Module qt_core::json_value::overloading

Traits

JsonValueFromVariantArgs
This trait represents a set of arguments accepted by JsonValue::from_variant method.

JsonValueNewArgs
This trait represents a set of arguments accepted by JsonValue::new method.

JsonValueToArrayArgs
This trait represents a set of arguments accepted by JsonValue::to_array method.

JsonValueToBoolArgs
This trait represents a set of arguments accepted by JsonValue::to_bool method.

JsonValueToDoubleArgs
This trait represents a set of arguments accepted by JsonValue::to_double method.

JsonValueToIntArgs
This trait represents a set of arguments accepted by JsonValue::to_int method.

JsonValueToObjectArgs
This trait represents a set of arguments accepted by JsonValue::to_object method.

JsonValueToStringArgs
This trait represents a set of arguments accepted by JsonValue::to_string method.

JsonValueToVariantArgs
This trait represents a set of arguments accepted by JsonValue::to_variant method.