Struct qt_core::json_value::JsonValue

pub struct JsonValue {
    // some fields omitted
}

C++ type: QJsonValue

C++ documentation:

The QJsonValue class encapsulates a value in JSON.

A value in JSON can be one of 6 basic types:

JSON is a format to store structured data. It has 6 basic data types:

bool QJsonValue::Bool

double QJsonValue::Double

string QJsonValue::String

array QJsonValue::Array

object QJsonValue::Object

null QJsonValue::Null

A value can represent any of the above data types. In addition, QJsonValue has one special flag to represent undefined values. This can be queried with isUndefined().

The type of the value can be queried with type() or accessors like isBool(), isString(), and so on. Likewise, the value can be converted to the type stored in it using the toBool(), toString() and so on.

Values are strictly typed internally and contrary to QVariant will not attempt to do any implicit type conversions. This implies that converting to a type that is not stored in the value will return a default constructed return value.

Methods

impl JsonValue

fn from_variant<Args>(args: Args) -> Args::ReturnType where Args: JsonValueFromVariantArgs

C++ method: QJsonValue::fromVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_variant((&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn from_variant((&::variant::Variant, cpp_utils::AsStruct)) -> ::json_value::JsonValue
C++ method: static QJsonValue QJsonValue::fromVariant(const QVariant& variant)

C++ documentation:

Converts variant to a QJsonValue and returns it.

The conversion will convert QVariant types as follows:

Source type	Destination type

QMetaType::Bool

QJsonValue::Bool

QMetaType::Int

QMetaType::UInt

QMetaType::LongLong

QMetaType::ULongLong

QMetaType::Float

QMetaType::Double

QJsonValue::Double

QMetaType::QString

QJsonValue::String

QMetaType::QStringList

QMetaType::QVariantList

QJsonValue::Array

QMetaType::QVariantMap

QMetaType::QVariantHash

QJsonValue::Object

For all other QVariant types a conversion to a QString will be attempted. If the returned string is empty, a Null QJsonValue will be stored, otherwise a String value using the returned QString.

See also toVariant().

fn is_array(&self) -> bool

C++ method: bool QJsonValue::isArray() const

C++ documentation:

Returns true if the value contains an array.

See also toArray().

fn is_bool(&self) -> bool

C++ method: bool QJsonValue::isBool() const

C++ documentation:

Returns true if the value contains a boolean.

See also toBool().

fn is_double(&self) -> bool

C++ method: bool QJsonValue::isDouble() const

C++ documentation:

Returns true if the value contains a double.

See also toDouble().

fn is_null(&self) -> bool

C++ method: bool QJsonValue::isNull() const

C++ documentation:

Returns true if the value is null.

fn is_object(&self) -> bool

C++ method: bool QJsonValue::isObject() const

C++ documentation:

Returns true if the value contains an object.

See also toObject().

fn is_string(&self) -> bool

C++ method: bool QJsonValue::isString() const

C++ documentation:

Returns true if the value contains a string.

See also toString().

fn is_undefined(&self) -> bool

C++ method: bool QJsonValue::isUndefined() const

C++ documentation:

Returns true if the value is undefined. This can happen in certain error cases as e.g. accessing a non existing key in a QJsonObject.

fn new<Args>(args: Args) -> Args::ReturnType where Args: JsonValueNewArgs

C++ method: QJsonValue::QJsonValue

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((::json_value::Type, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new(cpp_utils::AsStruct) -> ::json_value::JsonValue
3) fn new((::json_value::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(QJsonValue::Type arg1 = ?)

Variant 2

Rust arguments:
1) fn new((&::latin1_string::Latin1String, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((&::latin1_string::Latin1String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(QLatin1String s)

Variant 3

Rust arguments:
1) fn new((bool, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(bool b)

Variant 4

Rust arguments:
1) fn new((&::json_array::JsonArray, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((&::json_array::JsonArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(const QJsonArray& a)

Variant 5

Rust arguments:
1) fn new((&::json_object::JsonObject, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((&::json_object::JsonObject, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(const QJsonObject& o)

Variant 6

Rust arguments:
1) fn new((&::json_value::JsonValue, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((&::json_value::JsonValue, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(const QJsonValue& other)

Variant 7

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(const QString& s)

Variant 8

Rust arguments:
1) fn new((*const libc::c_char, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(const char* s)

Variant 9

Rust arguments:
1) fn new((libc::c_double, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(double n)

Variant 10

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(int n)

Variant 11

Rust arguments:
1) fn new((i64, cpp_utils::AsStruct)) -> ::json_value::JsonValue
2) fn new((i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_value::JsonValue>
C++ method: [constructor] void QJsonValue::QJsonValue(qint64 n)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 JsonValue) -> &'l0 mut JsonValue

C++ method: QJsonValue& QJsonValue::operator=(const QJsonValue& other)

C++ documentation:

Assigns the value stored in other to this object.

fn op_eq(&self, other: &JsonValue) -> bool

C++ method: bool QJsonValue::operator==(const QJsonValue& other) const

C++ documentation:

Returns true if the value is equal to other.

fn op_neq(&self, other: &JsonValue) -> bool

C++ method: bool QJsonValue::operator!=(const QJsonValue& other) const

C++ documentation:

Returns true if the value is not equal to other.

fn to_array<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonValueToArrayArgs<'largs>

C++ method: QJsonValue::toArray

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_array(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::JsonArray>
2) fn to_array(&self, cpp_utils::AsStruct) -> ::json_array::JsonArray
C++ method: QJsonArray QJsonValue::toArray() const

C++ documentation:

This is an overloaded function.

Converts the value to an array and returns it.

If type() is not Array, a QJsonArray() will be returned.

Variant 2

Rust arguments:
1) fn to_array(&self, (&::json_array::JsonArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_array::JsonArray>
2) fn to_array(&self, (&::json_array::JsonArray, cpp_utils::AsStruct)) -> ::json_array::JsonArray
C++ method: QJsonArray QJsonValue::toArray(const QJsonArray& defaultValue) const

C++ documentation:

Converts the value to an array and returns it.

If type() is not Array, the defaultValue will be returned.

fn to_bool<'largs, Args>(&'largs self, args: Args) -> bool where Args: JsonValueToBoolArgs<'largs>

C++ method: QJsonValue::toBool

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_bool(&self, bool) -> bool
2) fn to_bool(&self, ()) -> bool
C++ method: bool QJsonValue::toBool(bool defaultValue = ?) const

C++ documentation:

Converts the value to a bool and returns it.

If type() is not bool, the defaultValue will be returned.

fn to_double<'largs, Args>(&'largs self, args: Args) -> c_double where Args: JsonValueToDoubleArgs<'largs>

C++ method: QJsonValue::toDouble

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_double(&self, libc::c_double) -> libc::c_double
2) fn to_double(&self, ()) -> libc::c_double
C++ method: double QJsonValue::toDouble(double defaultValue = ?) const

C++ documentation:

Converts the value to a double and returns it.

If type() is not Double, the defaultValue will be returned.

fn to_int<'largs, Args>(&'largs self, args: Args) -> c_int where Args: JsonValueToIntArgs<'largs>

C++ method: QJsonValue::toInt

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_int(&self, libc::c_int) -> libc::c_int
2) fn to_int(&self, ()) -> libc::c_int
C++ method: int QJsonValue::toInt(int defaultValue = ?) const

C++ documentation:

Converts the value to an int and returns it.

If type() is not Double or the value is not a whole number, the defaultValue will be returned.

fn to_object<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonValueToObjectArgs<'largs>

C++ method: QJsonValue::toObject

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_object(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::JsonObject>
2) fn to_object(&self, cpp_utils::AsStruct) -> ::json_object::JsonObject
C++ method: QJsonObject QJsonValue::toObject() const

C++ documentation:

This is an overloaded function.

Converts the value to an object and returns it.

If type() is not Object, the QJsonObject() will be returned.

Variant 2

Rust arguments:
1) fn to_object(&self, (&::json_object::JsonObject, cpp_utils::AsBox)) -> cpp_utils::CppBox<::json_object::JsonObject>
2) fn to_object(&self, (&::json_object::JsonObject, cpp_utils::AsStruct)) -> ::json_object::JsonObject
C++ method: QJsonObject QJsonValue::toObject(const QJsonObject& defaultValue) const

C++ documentation:

Converts the value to an object and returns it.

If type() is not Object, the defaultValue will be returned.

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonValueToStringArgs<'largs>

C++ method: QJsonValue::toString

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
3) fn to_string(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
4) fn to_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QJsonValue::toString(const QString& defaultValue = ?) const

C++ documentation:

Converts the value to a QString and returns it.

If type() is not String, the defaultValue will be returned.

fn to_variant<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: JsonValueToVariantArgs<'largs>

C++ method: QJsonValue::toVariant

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_variant(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn to_variant(&self, cpp_utils::AsStruct) -> ::variant::Variant
C++ method: QVariant QJsonValue::toVariant() const

C++ documentation:

Converts the value to a QVariant().

The QJsonValue types will be converted as follows:

See also fromVariant().

fn type_(&self) -> Type

C++ method: QJsonValue::Type QJsonValue::type() const

C++ documentation:

Returns the type of the value.

See also QJsonValue::Type.