Module qt_core::json_value_ref::overloading

Traits

JsonValueRefAsQJsonValueArgs
This trait represents a set of arguments accepted by JsonValueRef::as_q_json_value method.

JsonValueRefNewArgs
This trait represents a set of arguments accepted by JsonValueRef::new method.

JsonValueRefOpAssignArgs
This trait represents a set of arguments accepted by JsonValueRef::op_assign method.

JsonValueRefToArrayArgs
This trait represents a set of arguments accepted by JsonValueRef::to_array method.

JsonValueRefToBoolArgs
This trait represents a set of arguments accepted by JsonValueRef::to_bool method.

JsonValueRefToDoubleArgs
This trait represents a set of arguments accepted by JsonValueRef::to_double method.

JsonValueRefToIntArgs
This trait represents a set of arguments accepted by JsonValueRef::to_int method.

JsonValueRefToObjectArgs
This trait represents a set of arguments accepted by JsonValueRef::to_object method.

JsonValueRefToStringArgs
This trait represents a set of arguments accepted by JsonValueRef::to_string method.

JsonValueRefToVariantArgs
This trait represents a set of arguments accepted by JsonValueRef::to_variant method.