Struct qt_core::latin1_string::Latin1String

pub struct Latin1String {
    // some fields omitted
}

C++ type: QLatin1String

C++ documentation:

The QLatin1String class provides a thin wrapper around an US-ASCII/Latin-1 encoded string literal.

Many of QString's member functions are overloaded to accept const char * instead of QString. This includes the copy constructor, the assignment operator, the comparison operators, and various other functions such as insert(), replace(), and indexOf(). These functions are usually optimized to avoid constructing a QString object for the const char * data. For example, assuming str is a QString,

if (str == "auto" || str == "extern"
        || str == "static" || str == "register") {
    ...
}

is much faster than

if (str == QString("auto") || str == QString("extern")
        || str == QString("static") || str == QString("register")) {
    ...
}

because it doesn't construct four temporary QString objects and make a deep copy of the character data.

Applications that define QT_NO_CAST_FROM_ASCII (as explained in the QString documentation) don't have access to QString's const char * API. To provide an efficient way of specifying constant Latin-1 strings, Qt provides the QLatin1String, which is just a very thin wrapper around a const char . Using QLatin1String, the example code above becomes

if (str == QLatin1String("auto")
        || str == QLatin1String("extern")
        || str == QLatin1String("static")
        || str == QLatin1String("register") {
    ...
}

This is a bit longer to type, but it provides exactly the same benefits as the first version of the code, and is faster than converting the Latin-1 strings using QString::fromLatin1().

Thanks to the QString(QLatin1String) constructor, QLatin1String can be used everywhere a QString is expected. For example:

QLabel label = new QLabel(QLatin1String("MOD"), this);

Methods

impl Latin1String

fn data(&self) -> *const c_char

C++ method: const char* QLatin1String::data() const

C++ documentation:

Returns the Latin-1 string stored in this object.

fn latin1(&self) -> *const c_char

C++ method: const char* QLatin1String::latin1() const

C++ documentation:

Returns the Latin-1 string stored in this object.

fn new<Args>(args: Args) -> Args::ReturnType where Args: Latin1StringNewArgs

C++ method: QLatin1String::QLatin1String

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::latin1_string::Latin1String
2) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::latin1_string::Latin1String>
C++ method: [constructor] void QLatin1String::QLatin1String(const QByteArray& s)

Variant 2

Rust arguments:
1) fn new((*const libc::c_char, cpp_utils::AsStruct)) -> ::latin1_string::Latin1String
2) fn new((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::latin1_string::Latin1String>
C++ method: [constructor] void QLatin1String::QLatin1String(const char* s)

Variant 3

Rust arguments:
1) fn new((*const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::latin1_string::Latin1String
2) fn new((*const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::latin1_string::Latin1String>
C++ method: [constructor] void QLatin1String::QLatin1String(const char* s, int sz)

fn op_eq<'largs, Args>(&'largs self, args: Args) -> bool where Args: Latin1StringOpEqArgs<'largs>

C++ method: QLatin1String::operator==

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_eq(&self, &::byte_array::ByteArray) -> bool
C++ method: bool QLatin1String::operator==(const QByteArray& s) const

C++ documentation:

This is an overloaded function.

The other byte array is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn op_eq(&self, &::string::String) -> bool
C++ method: bool QLatin1String::operator==(const QString& s) const

C++ documentation:

Returns true if this string is equal to string other; otherwise returns false.

The comparison is based exclusively on the numeric Unicode values of the characters and is very fast, but is not what a human would expect. Consider sorting user-interface strings with QString::localeAwareCompare().

Variant 3

Rust arguments: fn op_eq(&self, *const libc::c_char) -> bool
C++ method: bool QLatin1String::operator==(const char* s) const

C++ documentation:

This is an overloaded function.

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 4.3.

fn op_ge<'largs, Args>(&'largs self, args: Args) -> bool where Args: Latin1StringOpGeArgs<'largs>

C++ method: QLatin1String::operator>=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_ge(&self, &::byte_array::ByteArray) -> bool
C++ method: bool QLatin1String::operator>=(const QByteArray& s) const

C++ documentation:

This is an overloaded function.

The other array is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn op_ge(&self, &::string::String) -> bool
C++ method: bool QLatin1String::operator>=(const QString& s) const

C++ documentation:

Returns true if this string is lexically greater than or equal to string other; otherwise returns false.

The comparison is based exclusively on the numeric Unicode values of the characters and is very fast, but is not what a human would expect. Consider sorting user-interface strings with QString::localeAwareCompare().

Variant 3

Rust arguments: fn op_ge(&self, *const libc::c_char) -> bool
C++ method: bool QLatin1String::operator>=(const char* s) const

C++ documentation:

This is an overloaded function.

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 4.3.

fn op_gt<'largs, Args>(&'largs self, args: Args) -> bool where Args: Latin1StringOpGtArgs<'largs>

C++ method: QLatin1String::operator>

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_gt(&self, &::byte_array::ByteArray) -> bool
C++ method: bool QLatin1String::operator>(const QByteArray& s) const

C++ documentation:

This is an overloaded function.

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn op_gt(&self, &::string::String) -> bool
C++ method: bool QLatin1String::operator>(const QString& s) const

C++ documentation:

Returns true if this string is lexically greater than string other; otherwise returns false.

The comparison is based exclusively on the numeric Unicode values of the characters and is very fast, but is not what a human would expect. Consider sorting user-interface strings with QString::localeAwareCompare().

Variant 3

Rust arguments: fn op_gt(&self, *const libc::c_char) -> bool
C++ method: bool QLatin1String::operator>(const char* s) const

C++ documentation:

This is an overloaded function.

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 4.3.

fn op_le<'largs, Args>(&'largs self, args: Args) -> bool where Args: Latin1StringOpLeArgs<'largs>

C++ method: QLatin1String::operator<=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_le(&self, &::byte_array::ByteArray) -> bool
C++ method: bool QLatin1String::operator<=(const QByteArray& s) const

C++ documentation:

This is an overloaded function.

The other array is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn op_le(&self, &::string::String) -> bool
C++ method: bool QLatin1String::operator<=(const QString& s) const

C++ documentation:

Returns true if this string is lexically less than or equal to string other; otherwise returns false.

The comparison is based exclusively on the numeric Unicode values of the characters and is very fast, but is not what a human would expect. Consider sorting user-interface strings with QString::localeAwareCompare().

Variant 3

Rust arguments: fn op_le(&self, *const libc::c_char) -> bool
C++ method: bool QLatin1String::operator<=(const char* s) const

C++ documentation:

This is an overloaded function.

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 4.3.

fn op_lt<'largs, Args>(&'largs self, args: Args) -> bool where Args: Latin1StringOpLtArgs<'largs>

C++ method: QLatin1String::operator<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_lt(&self, &::byte_array::ByteArray) -> bool
C++ method: bool QLatin1String::operator<(const QByteArray& s) const

C++ documentation:

This is an overloaded function.

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn op_lt(&self, &::string::String) -> bool
C++ method: bool QLatin1String::operator<(const QString& s) const

C++ documentation:

Returns true if this string is lexically less than the other string; otherwise returns false.

The comparison is based exclusively on the numeric Unicode values of the characters and is very fast, but is not what a human would expect. Consider sorting user-interface strings using the QString::localeAwareCompare() function.

Variant 3

Rust arguments: fn op_lt(&self, *const libc::c_char) -> bool
C++ method: bool QLatin1String::operator<(const char* s) const

C++ documentation:

This is an overloaded function.

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 4.3.

fn op_neq<'largs, Args>(&'largs self, args: Args) -> bool where Args: Latin1StringOpNeqArgs<'largs>

C++ method: QLatin1String::operator!=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_neq(&self, &::byte_array::ByteArray) -> bool
C++ method: bool QLatin1String::operator!=(const QByteArray& s) const

C++ documentation:

This function overloads operator!=().

The other byte array is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn op_neq(&self, &::string::String) -> bool
C++ method: bool QLatin1String::operator!=(const QString& s) const

C++ documentation:

Returns true if this string is not equal to string other; otherwise returns false.

The comparison is based exclusively on the numeric Unicode values of the characters and is very fast, but is not what a human would expect. Consider sorting user-interface strings with QString::localeAwareCompare().

Variant 3

Rust arguments: fn op_neq(&self, *const libc::c_char) -> bool
C++ method: bool QLatin1String::operator!=(const char* s) const

C++ documentation:

This function overloads operator!=().

The other const char pointer is converted to a QString using the QString::fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

This function was introduced in Qt 4.3.

fn size(&self) -> c_int

C++ method: int QLatin1String::size() const

C++ documentation:

Returns the size of the Latin-1 string stored in this object.