Module qt_core::library::overloading

Traits

LibraryConnectArgs
This trait represents a set of arguments accepted by Library::connect method.

LibraryConnectStaticArgs
This trait represents a set of arguments accepted by Library::connect_static method.

LibraryDestroyedArgs
This trait represents a set of arguments accepted by Library::destroyed method.

LibraryDisconnectArgs
This trait represents a set of arguments accepted by Library::disconnect method.

LibraryDisconnectStaticArgs
This trait represents a set of arguments accepted by Library::disconnect_static method.

LibraryDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Library::dynamic_property_names method.

LibraryErrorStringArgs
This trait represents a set of arguments accepted by Library::error_string method.

LibraryFileNameArgs
This trait represents a set of arguments accepted by Library::file_name method.

LibraryNewArgs
This trait represents a set of arguments accepted by Library::new method.

LibraryObjectNameArgs
This trait represents a set of arguments accepted by Library::object_name method.

LibraryPropertyArgs
This trait represents a set of arguments accepted by Library::property method.

LibraryResolveStaticArgs
This trait represents a set of arguments accepted by Library::resolve_static method.

LibrarySetFileNameAndVersionArgs
This trait represents a set of arguments accepted by Library::set_file_name_and_version method.

LibraryStartTimerArgs
This trait represents a set of arguments accepted by Library::start_timer method.

LibraryTrArgs
This trait represents a set of arguments accepted by Library::tr method.

LibraryTrUtf8Args
This trait represents a set of arguments accepted by Library::tr_utf8 method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.