Struct qt_core::library::Library

pub struct Library {
    // some fields omitted
}

C++ type: QLibrary

C++ documentation:

The QLibrary class loads shared libraries at runtime.

An instance of a QLibrary object operates on a single shared object file (which we call a "library", but is also known as a "DLL"). A QLibrary provides access to the functionality in the library in a platform independent way. You can either pass a file name in the constructor, or set it explicitly with setFileName(). When loading the library, QLibrary searches in all the system-specific library locations (e.g. LD_LIBRARY_PATH on Unix), unless the file name has an absolute path.

If the file name is an absolute path then an attempt is made to load this path first. If the file cannot be found, QLibrary tries the name with different platform-specific file prefixes, like "lib" on Unix and Mac, and suffixes, like ".so" on Unix, ".dylib" on the Mac, or ".dll" on Windows.

If the file path is not absolute then QLibrary modifies the search order to try the system-specific prefixes and suffixes first, followed by the file path specified.

This makes it possible to specify shared libraries that are only identified by their basename (i.e. without their suffix), so the same code will work on different operating systems yet still minimise the number of attempts to find the library.

The most important functions are load() to dynamically load the library file, isLoaded() to check whether loading was successful, and resolve() to resolve a symbol in the library. The resolve() function implicitly tries to load the library if it has not been loaded yet. Multiple instances of QLibrary can be used to access the same physical library. Once loaded, libraries remain in memory until the application terminates. You can attempt to unload a library using unload(), but if other instances of QLibrary are using the same library, the call will fail, and unloading will only happen when every instance has called unload().

A typical use of QLibrary is to resolve an exported symbol in a library, and to call the C function that this symbol represents. This is called "explicit linking" in contrast to "implicit linking", which is done by the link step in the build process when linking an executable against a library.

The following code snippet loads a library, resolves the symbol "mysymbol", and calls the function if everything succeeded. If something goes wrong, e.g. the library file does not exist or the symbol is not defined, the function pointer will be 0 and won't be called.

QLibrary myLib("mylib");
typedef void (MyPrototype)();
MyPrototype myFunction = (MyPrototype) myLib.resolve("mysymbol");
if (myFunction)
    myFunction();

The symbol must be exported as a C function from the library for resolve() to work. This means that the function must be wrapped in an extern "C" block if the library is compiled with a C++ compiler. On Windows, this also requires the use of a dllexport macro; see resolve() for the details of how this is done. For convenience, there is a static resolve() function which you can use if you just want to call a function in a library without explicitly loading the library first:

typedef void (MyPrototype)();
MyPrototype myFunction =
        (MyPrototype) QLibrary::resolve("mylib", "mysymbol");
if (myFunction)
    myFunction();

Methods

impl Library

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QLibrary::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QLibrary::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LibraryConnectArgs<'largs>

C++ method: QLibrary::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: QMetaObject::Connection QLibrary::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: LibraryConnectStaticArgs

C++ method: QLibrary::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QLibrary::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QLibrary::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn delete_later(&mut self)

C++ method: void QLibrary::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: LibraryDestroyedArgs<'largs>

C++ method: QLibrary::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut ::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QLibrary::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: LibraryDisconnectArgs<'largs>

C++ method: QLibrary::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const ::object::Object) -> bool
2) fn disconnect(&self, (*const ::object::Object, *const libc::c_char)) -> bool
C++ method: bool QLibrary::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const ::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QLibrary::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: LibraryDisconnectStaticArgs

C++ method: QLibrary::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&::meta_object::Connection) -> bool
C++ method: static bool QLibrary::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod)) -> bool
C++ method: static bool QLibrary::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QLibrary::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QLibrary::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QLibrary::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LibraryDynamicPropertyNamesArgs<'largs>

C++ method: QLibrary::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: QList<QByteArray> QLibrary::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn error_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LibraryErrorStringArgs<'largs>

C++ method: QLibrary::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn error_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QLibrary::errorString() const

C++ documentation:

Returns a text string with the description of the last error that occurred. Currently, errorString will only be set if load(), unload() or resolve() for some reason fails.

This function was introduced in Qt 4.2.

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QLibrary::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QLibrary::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LibraryFileNameArgs<'largs>

C++ method: QLibrary::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QLibrary::fileName() const

C++ documentation:

This property holds the file name of the library.

We recommend omitting the file's suffix in the file name, since QLibrary will automatically look for the file with the appropriate suffix (see isLibrary()).

When loading the library, QLibrary searches in all system-specific library locations (for example, LD_LIBRARY_PATH on Unix), unless the file name has an absolute path. After loading the library successfully, fileName() returns the fully-qualified file name of the library, including the full path to the library if one was given in the constructor or passed to setFileName().

For example, after successfully loading the "GL" library on Unix platforms, fileName() will return "libGL.so". If the file name was originally passed as "/usr/lib/libGL", fileName() will return "/usr/lib/libGL.so".

Access functions:

QString 	fileName() const

void 	setFileName(const QString & fileName)

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QLibrary::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QLibrary::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_library(file_name: &String) -> bool

C++ method: static bool QLibrary::isLibrary(const QString& fileName)

C++ documentation:

Returns true if fileName has a valid suffix for a loadable library; otherwise returns false.

Platform	Valid suffixes

Windows	.dll, .DLL

Unix/Linux	.so

AIX	.a

HP-UX	.sl, .so (HP-UXi)

OS X and iOS	.dylib, .bundle, .so

Trailing versioning numbers on Unix are ignored.

fn is_loaded(&self) -> bool

C++ method: bool QLibrary::isLoaded() const

C++ documentation:

Returns true if the library is loaded; otherwise returns false.

See also load().

fn is_widget_type(&self) -> bool

C++ method: bool QLibrary::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QLibrary::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn kill_timer(&mut self, id: c_int)

C++ method: void QLibrary::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn load(&mut self) -> bool

C++ method: bool QLibrary::load()

C++ documentation:

Loads the library and returns true if the library was loaded successfully; otherwise returns false. Since resolve() always calls this function before resolving any symbols it is not necessary to call it explicitly. In some situations you might want the library loaded in advance, in which case you would use this function.

See also unload().

fn load_hints(&self) -> Flags<LoadHint>

C++ method: QFlags<QLibrary::LoadHint> QLibrary::loadHints() const

C++ documentation:

This property holds give the load() function some hints on how it should behave.

You can give some hints on how the symbols are resolved. Usually, the symbols are not resolved at load time, but resolved lazily, (that is, when resolve() is called). If you set the loadHints to ResolveAllSymbolsHint, then all symbols will be resolved at load time if the platform supports it.

Setting ExportExternalSymbolsHint will make the external symbols in the library available for resolution in subsequent loaded libraries.

If LoadArchiveMemberHint is set, the file name is composed of two components: A path which is a reference to an archive file followed by the second component which is the reference to the archive member. For instance, the fileName libGL.a(shr_64.o) will refer to the library shr_64.o in the archive file named libGL.a. This is only supported on the AIX platform.

Setting PreventUnloadHint will only apply on Unix platforms.

The interpretation of the load hints is platform dependent, and if you use it you are probably making some assumptions on which platform you are compiling for, so use them only if you understand the consequences of them.

By default, none of these flags are set, so libraries will be loaded with lazy symbol resolution, and will not export external symbols for resolution in other dynamically-loaded libraries.

Note: Setting this property after the library has been loaded has no effect and loadHints() will not reflect those changes.

Note: This property is shared among all QLibrary instances that refer to the same library.

Access functions:

LoadHints 	loadHints() const

void 	setLoadHints(LoadHints hints)

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QLibrary::metaObject() const

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QLibrary::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn new<Args>(args: Args) -> Args::ReturnType where Args: LibraryNewArgs

C++ method: QLibrary::QLibrary

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::library::Library
2) fn new((*mut ::object::Object, cpp_utils::AsStruct)) -> ::library::Library
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::library::Library>
4) fn new((*mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::library::Library>
C++ method: [constructor] void QLibrary::QLibrary(QObject* parent = ?)

Variant 2

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::library::Library
2) fn new((&::string::String, *mut ::object::Object, cpp_utils::AsStruct)) -> ::library::Library
3) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::library::Library>
4) fn new((&::string::String, *mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::library::Library>
C++ method: [constructor] void QLibrary::QLibrary(const QString& fileName, QObject* parent = ?)

Variant 3

Rust arguments:
1) fn new((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::library::Library
2) fn new((&::string::String, &::string::String, *mut ::object::Object, cpp_utils::AsStruct)) -> ::library::Library
3) fn new((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::library::Library>
4) fn new((&::string::String, &::string::String, *mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::library::Library>
C++ method: [constructor] void QLibrary::QLibrary(const QString& fileName, const QString& version, QObject* parent = ?)

Variant 4

Rust arguments:
1) fn new((&::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::library::Library
2) fn new((&::string::String, libc::c_int, *mut ::object::Object, cpp_utils::AsStruct)) -> ::library::Library
3) fn new((&::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::library::Library>
4) fn new((&::string::String, libc::c_int, *mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::library::Library>
C++ method: [constructor] void QLibrary::QLibrary(const QString& fileName, int verNum, QObject* parent = ?)

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LibraryObjectNameArgs<'largs>

C++ method: QLibrary::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QLibrary::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QLibrary::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: LibraryPropertyArgs<'largs>

C++ method: QLibrary::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QLibrary::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QLibrary::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QLibrary::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QLibrary::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QLibrary::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn resolve(&mut self, symbol: *const c_char) -> extern fn()

C++ method: void (*FN_PTR)() QLibrary::resolve(const char* symbol)

C++ documentation:

Returns the address of the exported symbol symbol. The library is loaded if necessary. The function returns 0 if the symbol could not be resolved or if the library could not be loaded.

Example:

typedef int (*AvgFunction)(int, int);

AvgFunction avg = (AvgFunction) library->resolve("avg"); if (avg) return avg(5, 8); else return -1;
The symbol must be exported as a C function from the library. This means that the function must be wrapped in an extern "C" if the library is compiled with a C++ compiler. On Windows you must also explicitly export the function from the DLL using the __declspec(dllexport) compiler directive, for example:

extern "C" MY_EXPORT int avg(int a, int b)
{
    return (a + b) / 2;
}

with MY_EXPORT defined as

#ifdef Q_OS_WIN
#define MY_EXPORT __declspec(dllexport)
#else
#define MY_EXPORT
#endif

fn resolve_static<Args>(args: Args) -> extern fn() where Args: LibraryResolveStaticArgs

C++ method: QLibrary::resolve

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn resolve_static((&::string::String, &::string::String, *const libc::c_char)) -> extern "C" fn()
C++ method: static void (*FN_PTR)() QLibrary::resolve(const QString& fileName, const QString& version, const char* symbol)

C++ documentation:

This is an overloaded function.

Loads the library fileName with full version number version and returns the address of the exported symbol symbol. Note that fileName should not include the platform-specific file suffix; (see fileName). The library remains loaded until the application exits. version is ignored on Windows.

The function returns 0 if the symbol could not be resolved or if the library could not be loaded.

This function was introduced in Qt 4.4.

See also resolve().

Variant 2

Rust arguments: fn resolve_static((&::string::String, *const libc::c_char)) -> extern "C" fn()
C++ method: static void (*FN_PTR)() QLibrary::resolve(const QString& fileName, const char* symbol)

C++ documentation:

This is an overloaded function.

Loads the library fileName and returns the address of the exported symbol symbol. Note that fileName should not include the platform-specific file suffix; (see fileName). The library remains loaded until the application exits.

The function returns 0 if the symbol could not be resolved or if the library could not be loaded.

See also resolve().

Variant 3

Rust arguments: fn resolve_static((&::string::String, libc::c_int, *const libc::c_char)) -> extern "C" fn()
C++ method: static void (*FN_PTR)() QLibrary::resolve(const QString& fileName, int verNum, const char* symbol)

C++ documentation:

This is an overloaded function.

Loads the library fileName with major version number verNum and returns the address of the exported symbol symbol. Note that fileName should not include the platform-specific file suffix; (see fileName). The library remains loaded until the application exits. verNum is ignored on Windows.

The function returns 0 if the symbol could not be resolved or if the library could not be loaded.

See also resolve().

fn set_file_name(&mut self, file_name: &String)

C++ method: void QLibrary::setFileName(const QString& fileName)

C++ documentation:

This property holds the file name of the library.

We recommend omitting the file's suffix in the file name, since QLibrary will automatically look for the file with the appropriate suffix (see isLibrary()).

When loading the library, QLibrary searches in all system-specific library locations (for example, LD_LIBRARY_PATH on Unix), unless the file name has an absolute path. After loading the library successfully, fileName() returns the fully-qualified file name of the library, including the full path to the library if one was given in the constructor or passed to setFileName().

For example, after successfully loading the "GL" library on Unix platforms, fileName() will return "libGL.so". If the file name was originally passed as "/usr/lib/libGL", fileName() will return "/usr/lib/libGL.so".

Access functions:

QString 	fileName() const

void 	setFileName(const QString & fileName)

fn set_file_name_and_version<'largs, Args>(&'largs mut self, args: Args) where Args: LibrarySetFileNameAndVersionArgs<'largs>

C++ method: QLibrary::setFileNameAndVersion

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_file_name_and_version(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QLibrary::setFileNameAndVersion(const QString& fileName, const QString& version)

C++ documentation:

Sets the fileName property and full version number to fileName and version respectively. The version parameter is ignored on Windows.

This function was introduced in Qt 4.4.

See also setFileName().

Variant 2

Rust arguments: fn set_file_name_and_version(&mut self, (&::string::String, libc::c_int)) -> ()
C++ method: void QLibrary::setFileNameAndVersion(const QString& fileName, int verNum)

C++ documentation:

Sets the fileName property and major version number to fileName and versionNumber respectively. The versionNumber is ignored on Windows.

See also setFileName().

fn set_load_hints(&mut self, hints: Flags<LoadHint>)

C++ method: void QLibrary::setLoadHints(QFlags<QLibrary::LoadHint> hints)

C++ documentation:

This property holds give the load() function some hints on how it should behave.

You can give some hints on how the symbols are resolved. Usually, the symbols are not resolved at load time, but resolved lazily, (that is, when resolve() is called). If you set the loadHints to ResolveAllSymbolsHint, then all symbols will be resolved at load time if the platform supports it.

Setting ExportExternalSymbolsHint will make the external symbols in the library available for resolution in subsequent loaded libraries.

If LoadArchiveMemberHint is set, the file name is composed of two components: A path which is a reference to an archive file followed by the second component which is the reference to the archive member. For instance, the fileName libGL.a(shr_64.o) will refer to the library shr_64.o in the archive file named libGL.a. This is only supported on the AIX platform.

Setting PreventUnloadHint will only apply on Unix platforms.

The interpretation of the load hints is platform dependent, and if you use it you are probably making some assumptions on which platform you are compiling for, so use them only if you understand the consequences of them.

By default, none of these flags are set, so libraries will be loaded with lazy symbol resolution, and will not export external symbols for resolution in other dynamically-loaded libraries.

Note: Setting this property after the library has been loaded has no effect and loadHints() will not reflect those changes.

Note: This property is shared among all QLibrary instances that refer to the same library.

Access functions:

LoadHints 	loadHints() const

void 	setLoadHints(LoadHints hints)

fn set_object_name(&mut self, name: &String)

C++ method: void QLibrary::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QLibrary::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QLibrary::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QLibrary::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QLibrary::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: LibraryStartTimerArgs<'largs>

C++ method: QLibrary::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, ::qt::TimerType)) -> libc::c_int
C++ method: int QLibrary::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn thread(&self) -> *mut Thread

C++ method: QThread* QLibrary::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: LibraryTrArgs

C++ method: QLibrary::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QLibrary::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: LibraryTrUtf8Args

C++ method: QLibrary::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QLibrary::trUtf8(const char* s, const char* c, int n)

fn unload(&mut self) -> bool

C++ method: bool QLibrary::unload()

C++ documentation:

Unloads the library and returns true if the library could be unloaded; otherwise returns false.

This happens automatically on application termination, so you shouldn't normally need to call this function.

If other instances of QLibrary are using the same library, the call will fail, and unloading will only happen when every instance has called unload().

Note that on Mac OS X 10.3 (Panther), dynamic libraries cannot be unloaded.

See also resolve() and load().

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QLibrary::userData(unsigned int id) const