Module qt_core::library_info::overloading

Traits

LibraryInfoBuildDateArgs
This trait represents a set of arguments accepted by LibraryInfo::build_date method.

LibraryInfoLicensedProductsArgs
This trait represents a set of arguments accepted by LibraryInfo::licensed_products method.

LibraryInfoLicenseeArgs
This trait represents a set of arguments accepted by LibraryInfo::licensee method.

LibraryInfoLocationArgs
This trait represents a set of arguments accepted by LibraryInfo::location method.

LibraryInfoPlatformPluginArgumentsArgs
This trait represents a set of arguments accepted by LibraryInfo::platform_plugin_arguments method.