Struct qt_core::library_info::LibraryInfo

pub struct LibraryInfo {
    // some fields omitted
}

C++ type: QLibraryInfo

C++ documentation:

The QLibraryInfo class provides information about the Qt library.

Many pieces of information are established when Qt is configured and built. This class provides an abstraction for accessing that information. By using the static functions of this class, an application can obtain information about the instance of the Qt library which the application is using at run-time.

You can also use a qt.conf file to override the hard-coded paths that are compiled into the Qt library. For more information, see the Using qt.conf documentation.

Methods

impl LibraryInfo

fn build() -> *const c_char

C++ method: static const char* QLibraryInfo::build()

fn build_date<Args>(args: Args) -> Args::ReturnType where Args: LibraryInfoBuildDateArgs

C++ method: QLibraryInfo::buildDate

This is an overloaded function. Available variants:

Rust arguments:
1) fn build_date(cpp_utils::AsBox) -> cpp_utils::CppBox<::date::Date>
2) fn build_date(cpp_utils::AsStruct) -> ::date::Date
C++ method: static QDate QLibraryInfo::buildDate()

C++ documentation:

This function used to return the installation date for this build of Qt, but now returns a constant date.

This function was introduced in Qt 4.6.

fn is_debug_build() -> bool

C++ method: static bool QLibraryInfo::isDebugBuild()

C++ documentation:

Returns true if this build of Qt was built with debugging enabled, or false if it was built in release mode.

This function was introduced in Qt 5.0.

fn licensed_products<Args>(args: Args) -> Args::ReturnType where Args: LibraryInfoLicensedProductsArgs

C++ method: QLibraryInfo::licensedProducts

This is an overloaded function. Available variants:

Rust arguments:
1) fn licensed_products(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn licensed_products(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QLibraryInfo::licensedProducts()

C++ documentation:

Returns the products that the license for this build of Qt has access to.

See also licensee().

fn licensee<Args>(args: Args) -> Args::ReturnType where Args: LibraryInfoLicenseeArgs

C++ method: QLibraryInfo::licensee

This is an overloaded function. Available variants:

Rust arguments:
1) fn licensee(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn licensee(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QLibraryInfo::licensee()

C++ documentation:

Returns the person to whom this build of Qt is licensed.

See also licensedProducts().

fn location<Args>(args: Args) -> Args::ReturnType where Args: LibraryInfoLocationArgs

C++ method: QLibraryInfo::location

This is an overloaded function. Available variants:

Rust arguments:
1) fn location((::library_info::LibraryLocation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn location((::library_info::LibraryLocation, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QLibraryInfo::location(QLibraryInfo::LibraryLocation arg1)

C++ documentation:

Returns the location specified by loc.

fn platform_plugin_arguments<Args>(args: Args) -> Args::ReturnType where Args: LibraryInfoPlatformPluginArgumentsArgs

C++ method: QLibraryInfo::platformPluginArguments

This is an overloaded function. Available variants:

Rust arguments:
1) fn platform_plugin_arguments((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn platform_plugin_arguments((&::string::String, cpp_utils::AsStruct)) -> ::string_list::StringList
C++ method: static QStringList QLibraryInfo::platformPluginArguments(const QString& platformName)