Struct qt_core::list::ListPersistentModelIndexRef

pub struct ListPersistentModelIndexRef {
    // some fields omitted
}

C++ type: QList<QPersistentModelIndex>

Methods

impl ListPersistentModelIndexRef

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: ListPersistentModelIndexRefAppendArgs<'largs>

C++ method: QList<QPersistentModelIndex>::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &::list::ListPersistentModelIndexRef) -> ()
C++ method: void QList<QPersistentModelIndex>::append(const QList<QPersistentModelIndex>& t)

C++ documentation:

This is an overloaded function.

Appends the items of the value list to this list.

This function was introduced in Qt 4.5.

See also operator<<() and operator+=().

Variant 2

Rust arguments: fn append(&mut self, &::persistent_model_index::PersistentModelIndex) -> ()
C++ method: void QList<QPersistentModelIndex>::append(const QPersistentModelIndex& t)

C++ documentation:

Inserts value at the end of the list.

Example:

QList<QString> list;
list.append("one");
list.append("two");
list.append("three");
// list: ["one", "two", "three"]

This is the same as list.insert(size(), value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also operator<<(), prepend(), and insert().

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QList<QPersistentModelIndex>::at(int i) const

C++ documentation:

Returns the item at index position i in the list. i must be a valid index position in the list (i.e., 0 <= i < size()).

This function is very fast (constant time).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QList<QPersistentModelIndex>::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut PersistentModelIndex

C++ method: QPersistentModelIndex& QList<QPersistentModelIndex>::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn clear(&mut self)

C++ method: void QList<QPersistentModelIndex>::clear()

C++ documentation:

Removes all items from the list.

See also removeAll().

fn contains(&self, t: &PersistentModelIndex) -> bool

C++ method: bool QList<QPersistentModelIndex>::contains(const QPersistentModelIndex& t) const

C++ documentation:

Returns true if the list contains an occurrence of value; otherwise returns false.

This function requires the value type to have an implementation of operator==().

See also indexOf() and count().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: ListPersistentModelIndexRefCountArgs<'largs>

C++ method: QList<QPersistentModelIndex>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QList<QPersistentModelIndex>::count() const

C++ documentation:

Returns the number of items in the list. This is effectively the same as size().

Variant 2

Rust arguments: fn count(&self, &::persistent_model_index::PersistentModelIndex) -> libc::c_int
C++ method: int QList<QPersistentModelIndex>::count(const QPersistentModelIndex& t) const

C++ documentation:

Returns the number of occurrences of value in the list.

This function requires the value type to have an implementation of operator==().

See also contains() and indexOf().

fn detach(&mut self)

C++ method: void QList<QPersistentModelIndex>::detach()

fn detach_shared(&mut self)

C++ method: void QList<QPersistentModelIndex>::detachShared()

fn empty(&self) -> bool

C++ method: bool QList<QPersistentModelIndex>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty() and returns true if the list is empty.

fn ends_with(&self, t: &PersistentModelIndex) -> bool

C++ method: bool QList<QPersistentModelIndex>::endsWith(const QPersistentModelIndex& t) const

C++ documentation:

Returns true if this list is not empty and its last item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and contains().

fn first<'l0>(&'l0 self) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QList<QPersistentModelIndex>::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut PersistentModelIndex

C++ method: QPersistentModelIndex& QList<QPersistentModelIndex>::first()

C++ documentation:

Returns a reference to the first item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QList<QPersistentModelIndex>::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut PersistentModelIndex

C++ method: QPersistentModelIndex& QList<QPersistentModelIndex>::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: ListPersistentModelIndexRefIndexOfArgs<'largs>

C++ method: QList<QPersistentModelIndex>::indexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn index_of(&self, &::persistent_model_index::PersistentModelIndex) -> libc::c_int
2) fn index_of(&self, (&::persistent_model_index::PersistentModelIndex, libc::c_int)) -> libc::c_int
C++ method: int QList<QPersistentModelIndex>::indexOf(const QPersistentModelIndex& t, int from = ?) const

C++ documentation:

Returns the index position of the first occurrence of value in the list, searching forward from index position from. Returns -1 if no item matched.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "B" << "A";
list.indexOf("B");          // returns 1
list.indexOf("B", 1);       // returns 1
list.indexOf("B", 2);       // returns 3
list.indexOf("X");          // returns -1

This function requires the value type to have an implementation of operator==().

Note that QList uses 0-based indexes, just like C++ arrays. Negative indexes are not supported with the exception of the value mentioned above.

See also lastIndexOf() and contains().

fn insert(&mut self, i: c_int, t: &PersistentModelIndex)

C++ method: void QList<QPersistentModelIndex>::insert(int i, const QPersistentModelIndex& t)

C++ documentation:

Inserts value at index position i in the list. If i is 0, the value is prepended to the list. If i is size(), the value is appended to the list.

Example:

QList<QString> list;
list << "alpha" << "beta" << "delta";
list.insert(2, "gamma");
// list: ["alpha", "beta", "gamma", "delta"]

See also append(), prepend(), replace(), and removeAt().

fn is_detached(&self) -> bool

C++ method: bool QList<QPersistentModelIndex>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QList<QPersistentModelIndex>::isEmpty() const

C++ documentation:

Returns true if the list contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &ListPersistentModelIndexRef) -> bool

C++ method: bool QList<QPersistentModelIndex>::isSharedWith(const QList<QPersistentModelIndex>& other) const

fn last<'l0>(&'l0 self) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QList<QPersistentModelIndex>::last() const

C++ documentation:

This is an overloaded function.

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: ListPersistentModelIndexRefLastIndexOfArgs<'largs>

C++ method: QList<QPersistentModelIndex>::lastIndexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_index_of(&self, &::persistent_model_index::PersistentModelIndex) -> libc::c_int
2) fn last_index_of(&self, (&::persistent_model_index::PersistentModelIndex, libc::c_int)) -> libc::c_int
C++ method: int QList<QPersistentModelIndex>::lastIndexOf(const QPersistentModelIndex& t, int from = ?) const

C++ documentation:

Returns the index position of the last occurrence of value in the list, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "B" << "A";
list.lastIndexOf("B");      // returns 3
list.lastIndexOf("B", 3);   // returns 3
list.lastIndexOf("B", 2);   // returns 1
list.lastIndexOf("X");      // returns -1

This function requires the value type to have an implementation of operator==().

Note that QList uses 0-based indexes, just like C++ arrays. Negative indexes are not supported with the exception of the value mentioned above.

See also indexOf().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut PersistentModelIndex

C++ method: QPersistentModelIndex& QList<QPersistentModelIndex>::last()

C++ documentation:

Returns a reference to the last item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QList<QPersistentModelIndex>::length() const

C++ documentation:

This function is identical to count().

This function was introduced in Qt 4.5.

See also count().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListPersistentModelIndexRefMidArgs<'largs>

C++ method: QList<QPersistentModelIndex>::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListPersistentModelIndexRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListPersistentModelIndexRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListPersistentModelIndexRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::list::ListPersistentModelIndexRef
C++ method: QList<QPersistentModelIndex> QList<QPersistentModelIndex>::mid(int pos, int length = ?) const

C++ documentation:

Returns a sub-list which includes elements from this list, starting at position pos. If length is -1 (the default), all elements from pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn move_(&mut self, from: c_int, to: c_int)

C++ method: void QList<QPersistentModelIndex>::move(int from, int to)

C++ documentation:

Moves the item at index position from to index position to.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.move(1, 4);
// list: ["A", "C", "D", "E", "B", "F"]

This is the same as insert(to, takeAt(from)).This function assumes that both from and to are at least 0 but less than size(). To avoid failure, test that both from and to are at least 0 and less than size().

See also swap(), insert(), and takeAt().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ListPersistentModelIndexRefNewArgs

C++ method: QList<QPersistentModelIndex>::QList

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::list::ListPersistentModelIndexRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListPersistentModelIndexRef>
C++ method: [constructor] void QList<QPersistentModelIndex>::QList()

Variant 2

Rust arguments:
1) fn new((&::list::ListPersistentModelIndexRef, cpp_utils::AsStruct)) -> ::list::ListPersistentModelIndexRef
2) fn new((&::list::ListPersistentModelIndexRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListPersistentModelIndexRef>
C++ method: [constructor] void QList<QPersistentModelIndex>::QList(const QList<QPersistentModelIndex>& l)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListPersistentModelIndexRefOpAddArgs<'largs>

C++ method: QList<QPersistentModelIndex>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::list::ListPersistentModelIndexRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListPersistentModelIndexRef>
2) fn op_add(&self, (&::list::ListPersistentModelIndexRef, cpp_utils::AsStruct)) -> ::list::ListPersistentModelIndexRef
C++ method: QList<QPersistentModelIndex> QList<QPersistentModelIndex>::operator+(const QList<QPersistentModelIndex>& l) const

C++ documentation:

Returns a list that contains all the items in this list followed by all the items in the other list.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut ListPersistentModelIndexRef where Args: ListPersistentModelIndexRefOpAddAssignArgs<'largs>

C++ method: QList<QPersistentModelIndex>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::list::ListPersistentModelIndexRef) -> &'l0 mut ::list::ListPersistentModelIndexRef
C++ method: QList<QPersistentModelIndex>& QList<QPersistentModelIndex>::operator+=(const QList<QPersistentModelIndex>& l)

C++ documentation:

Appends the items of the other list to this list and returns a reference to this list.

See also operator+() and append().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::persistent_model_index::PersistentModelIndex) -> &'l0 mut ::list::ListPersistentModelIndexRef
C++ method: QList<QPersistentModelIndex>& QList<QPersistentModelIndex>::operator+=(const QPersistentModelIndex& t)

C++ documentation:

This is an overloaded function.

Appends value to the list.

See also append() and operator<<().

fn op_assign<'l0, 'l1>(&'l0 mut self, l: &'l1 ListPersistentModelIndexRef) -> &'l0 mut ListPersistentModelIndexRef

C++ method: QList<QPersistentModelIndex>& QList<QPersistentModelIndex>::operator=(const QList<QPersistentModelIndex>& l)

C++ documentation:

Assigns other to this list and returns a reference to this list.

fn op_eq(&self, l: &ListPersistentModelIndexRef) -> bool

C++ method: bool QList<QPersistentModelIndex>::operator==(const QList<QPersistentModelIndex>& l) const

C++ documentation:

Returns true if other is equal to this list; otherwise returns false.

Two lists are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator!=().

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 PersistentModelIndex

C++ method: const QPersistentModelIndex& QList<QPersistentModelIndex>::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(). This function runs in constant time.

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut PersistentModelIndex

C++ method: QPersistentModelIndex& QList<QPersistentModelIndex>::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference. i must be a valid index position in the list (i.e., 0 <= i < size()).

If this function is called on a list that is currently being shared, it will trigger a copy of all elements. Otherwise, this function runs in constant time. If you do not want to modify the list you should use QList::at().

See also at() and value().

fn op_neq(&self, l: &ListPersistentModelIndexRef) -> bool

C++ method: bool QList<QPersistentModelIndex>::operator!=(const QList<QPersistentModelIndex>& l) const

C++ documentation:

Returns true if other is not equal to this list; otherwise returns false.

Two lists are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator==().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut ListPersistentModelIndexRef where Args: ListPersistentModelIndexRefOpShlArgs<'largs>

C++ method: QList<QPersistentModelIndex>::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::list::ListPersistentModelIndexRef) -> &'l0 mut ::list::ListPersistentModelIndexRef
C++ method: QList<QPersistentModelIndex>& QList<QPersistentModelIndex>::operator<<(const QList<QPersistentModelIndex>& l)

C++ documentation:

Appends the items of the other list to this list and returns a reference to this list.

See also operator+=() and append().

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::persistent_model_index::PersistentModelIndex) -> &'l0 mut ::list::ListPersistentModelIndexRef
C++ method: QList<QPersistentModelIndex>& QList<QPersistentModelIndex>::operator<<(const QPersistentModelIndex& t)

C++ documentation:

This is an overloaded function.

Appends value to the list.

fn pop_back(&mut self)

C++ method: void QList<QPersistentModelIndex>::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn pop_front(&mut self)

C++ method: void QList<QPersistentModelIndex>::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn prepend(&mut self, t: &PersistentModelIndex)

C++ method: void QList<QPersistentModelIndex>::prepend(const QPersistentModelIndex& t)

C++ documentation:

Inserts value at the beginning of the list.

Example:

QList<QString> list;
list.prepend("one");
list.prepend("two");
list.prepend("three");
// list: ["three", "two", "one"]

This is the same as list.insert(0, value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also append() and insert().

fn push_back(&mut self, t: &PersistentModelIndex)

C++ method: void QList<QPersistentModelIndex>::push_back(const QPersistentModelIndex& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &PersistentModelIndex)

C++ method: void QList<QPersistentModelIndex>::push_front(const QPersistentModelIndex& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove_all(&mut self, t: &PersistentModelIndex) -> c_int

C++ method: int QList<QPersistentModelIndex>::removeAll(const QPersistentModelIndex& t)

C++ documentation:

Removes all occurrences of value in the list and returns the number of entries removed.

Example:

QList<QString> list;
list << "sun" << "cloud" << "sun" << "rain";
list.removeAll("sun");
// list: ["cloud", "rain"]

This function requires the value type to have an implementation of operator==().

See also removeOne(), removeAt(), takeAt(), and replace().

fn remove_at(&mut self, i: c_int)

C++ method: void QList<QPersistentModelIndex>::removeAt(int i)

C++ documentation:

Removes the item at index position i. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also takeAt(), removeFirst(), removeLast(), and removeOne().

fn remove_first(&mut self)

C++ method: void QList<QPersistentModelIndex>::removeFirst()

C++ documentation:

Removes the first item in the list. Calling this function is equivalent to calling removeAt(0). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeFirst().

fn remove_last(&mut self)

C++ method: void QList<QPersistentModelIndex>::removeLast()

C++ documentation:

Removes the last item in the list. Calling this function is equivalent to calling removeAt(size() - 1). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeLast().

fn remove_one(&mut self, t: &PersistentModelIndex) -> bool

C++ method: bool QList<QPersistentModelIndex>::removeOne(const QPersistentModelIndex& t)

C++ documentation:

Removes the first occurrence of value in the list and returns true on success; otherwise returns false.

Example:

QList<QString> list;
list << "sun" << "cloud" << "sun" << "rain";
list.removeOne("sun");
// list: ["cloud", ,"sun", "rain"]

This function requires the value type to have an implementation of operator==().

This function was introduced in Qt 4.4.

See also removeAll(), removeAt(), takeAt(), and replace().

fn replace(&mut self, i: c_int, t: &PersistentModelIndex)

C++ method: void QList<QPersistentModelIndex>::replace(int i, const QPersistentModelIndex& t)

C++ documentation:

Replaces the item at index position i with value. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also operator[]() and removeAt().

fn reserve(&mut self, size: c_int)

C++ method: void QList<QPersistentModelIndex>::reserve(int size)

C++ documentation:

Reserve space for alloc elements.

If alloc is smaller than the current size of the list, nothing will happen.

Use this function to avoid repetetive reallocation of QList's internal data if you can predict how many elements will be appended. Note that the reservation applies only to the internal pointer array.

This function was introduced in Qt 4.7.

fn set_sharable(&mut self, sharable: bool)

C++ method: void QList<QPersistentModelIndex>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QList<QPersistentModelIndex>::size() const

C++ documentation:

Returns the number of items in the list.

See also isEmpty() and count().

fn starts_with(&self, t: &PersistentModelIndex) -> bool

C++ method: bool QList<QPersistentModelIndex>::startsWith(const QPersistentModelIndex& t) const

C++ documentation:

Returns true if this list is not empty and its first item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and contains().

fn swap<'largs, Args>(&'largs mut self, args: Args) where Args: ListPersistentModelIndexRefSwapArgs<'largs>

C++ method: QList<QPersistentModelIndex>::swap

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn swap(&mut self, &mut ::list::ListPersistentModelIndexRef) -> ()
C++ method: void QList<QPersistentModelIndex>::swap(QList<QPersistentModelIndex>& other)

C++ documentation:

Swaps list other with this list. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

Variant 2

Rust arguments: fn swap(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QList<QPersistentModelIndex>::swap(int i, int j)

C++ documentation:

Exchange the item at index position i with the item at index position j. This function assumes that both i and j are at least 0 but less than size(). To avoid failure, test that both i and j are at least 0 and less than size().

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.swap(1, 4);
// list: ["A", "E", "C", "D", "B", "F"]

See also move().

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ListPersistentModelIndexRefTakeAtArgs<'largs>

C++ method: QList<QPersistentModelIndex>::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex QList<QPersistentModelIndex>::takeAt(int i)

C++ documentation:

Removes the item at index position i and returns it. i must be a valid index position in the list (i.e., 0 <= i < size()).

If you don't use the return value, removeAt() is more efficient.

See also removeAt(), takeFirst(), and takeLast().

fn take_first<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ListPersistentModelIndexRefTakeFirstArgs<'largs>

C++ method: QList<QPersistentModelIndex>::takeFirst

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_first(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
2) fn take_first(&mut self, cpp_utils::AsStruct) -> ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex QList<QPersistentModelIndex>::takeFirst()

C++ documentation:

Removes the first item in the list and returns it. This is the same as takeAt(0). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeFirst() is more efficient.

See also takeLast(), takeAt(), and removeFirst().

fn take_last<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ListPersistentModelIndexRefTakeLastArgs<'largs>

C++ method: QList<QPersistentModelIndex>::takeLast

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_last(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
2) fn take_last(&mut self, cpp_utils::AsStruct) -> ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex QList<QPersistentModelIndex>::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This is the same as takeAt(size() - 1). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeLast() is more efficient.

See also takeFirst(), takeAt(), and removeLast().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListPersistentModelIndexRefValueArgs<'largs>

C++ method: QList<QPersistentModelIndex>::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn value(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
2) fn value(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex QList<QPersistentModelIndex>::value(int i) const

C++ documentation:

Returns the value at index position i in the list.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that the index is going to be within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments:
1) fn value(&self, (libc::c_int, &::persistent_model_index::PersistentModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
2) fn value(&self, (libc::c_int, &::persistent_model_index::PersistentModelIndex, cpp_utils::AsStruct)) -> ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex QList<QPersistentModelIndex>::value(int i, const QPersistentModelIndex& defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.