Struct qt_core::list::ListTimerInfoRef

pub struct ListTimerInfoRef {
    // some fields omitted
}

C++ type: QList<QAbstractEventDispatcher::TimerInfo>

Methods

impl ListTimerInfoRef

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: ListTimerInfoRefAppendArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &::abstract_event_dispatcher::TimerInfo) -> ()
C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::append(const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

Inserts value at the end of the list.

Example:

QList<QString> list;
list.append("one");
list.append("two");
list.append("three");
// list: ["one", "two", "three"]

This is the same as list.insert(size(), value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also operator<<(), prepend(), and insert().

Variant 2

Rust arguments: fn append(&mut self, &::list::ListTimerInfoRef) -> ()
C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::append(const QList<QAbstractEventDispatcher::TimerInfo>& t)

C++ documentation:

This is an overloaded function.

Appends the items of the value list to this list.

This function was introduced in Qt 4.5.

See also operator<<() and operator+=().

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 TimerInfo

C++ method: const QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::at(int i) const

C++ documentation:

Returns the item at index position i in the list. i must be a valid index position in the list (i.e., 0 <= i < size()).

This function is very fast (constant time).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 TimerInfo

C++ method: const QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut TimerInfo

C++ method: QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn clear(&mut self)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::clear()

C++ documentation:

Removes all items from the list.

See also removeAll().

fn count(&self) -> c_int

C++ method: int QList<QAbstractEventDispatcher::TimerInfo>::count() const

C++ documentation:

Returns the number of items in the list. This is effectively the same as size().

fn detach(&mut self)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::detach()

fn detach_shared(&mut self)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::detachShared()

fn empty(&self) -> bool

C++ method: bool QList<QAbstractEventDispatcher::TimerInfo>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty() and returns true if the list is empty.

fn first<'l0>(&'l0 self) -> &'l0 TimerInfo

C++ method: const QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut TimerInfo

C++ method: QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::first()

C++ documentation:

Returns a reference to the first item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 TimerInfo

C++ method: const QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut TimerInfo

C++ method: QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn insert(&mut self, i: c_int, t: &TimerInfo)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::insert(int i, const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

Inserts value at index position i in the list. If i is 0, the value is prepended to the list. If i is size(), the value is appended to the list.

Example:

QList<QString> list;
list << "alpha" << "beta" << "delta";
list.insert(2, "gamma");
// list: ["alpha", "beta", "gamma", "delta"]

See also append(), prepend(), replace(), and removeAt().

fn is_detached(&self) -> bool

C++ method: bool QList<QAbstractEventDispatcher::TimerInfo>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QList<QAbstractEventDispatcher::TimerInfo>::isEmpty() const

C++ documentation:

Returns true if the list contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &ListTimerInfoRef) -> bool

C++ method: bool QList<QAbstractEventDispatcher::TimerInfo>::isSharedWith(const QList<QAbstractEventDispatcher::TimerInfo>& other) const

fn last<'l0>(&'l0 self) -> &'l0 TimerInfo

C++ method: const QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::last() const

C++ documentation:

This is an overloaded function.

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut TimerInfo

C++ method: QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::last()

C++ documentation:

Returns a reference to the last item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QList<QAbstractEventDispatcher::TimerInfo>::length() const

C++ documentation:

This function is identical to count().

This function was introduced in Qt 4.5.

See also count().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListTimerInfoRefMidArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListTimerInfoRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListTimerInfoRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListTimerInfoRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::list::ListTimerInfoRef
C++ method: QList<QAbstractEventDispatcher::TimerInfo> QList<QAbstractEventDispatcher::TimerInfo>::mid(int pos, int length = ?) const

C++ documentation:

Returns a sub-list which includes elements from this list, starting at position pos. If length is -1 (the default), all elements from pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn move_(&mut self, from: c_int, to: c_int)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::move(int from, int to)

C++ documentation:

Moves the item at index position from to index position to.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.move(1, 4);
// list: ["A", "C", "D", "E", "B", "F"]

This is the same as insert(to, takeAt(from)).This function assumes that both from and to are at least 0 but less than size(). To avoid failure, test that both from and to are at least 0 and less than size().

See also swap(), insert(), and takeAt().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ListTimerInfoRefNewArgs

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::QList

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::list::ListTimerInfoRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListTimerInfoRef>
C++ method: [constructor] void QList<QAbstractEventDispatcher::TimerInfo>::QList()

Variant 2

Rust arguments:
1) fn new((&::list::ListTimerInfoRef, cpp_utils::AsStruct)) -> ::list::ListTimerInfoRef
2) fn new((&::list::ListTimerInfoRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListTimerInfoRef>
C++ method: [constructor] void QList<QAbstractEventDispatcher::TimerInfo>::QList(const QList<QAbstractEventDispatcher::TimerInfo>& l)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ListTimerInfoRefOpAddArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::list::ListTimerInfoRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListTimerInfoRef>
2) fn op_add(&self, (&::list::ListTimerInfoRef, cpp_utils::AsStruct)) -> ::list::ListTimerInfoRef
C++ method: QList<QAbstractEventDispatcher::TimerInfo> QList<QAbstractEventDispatcher::TimerInfo>::operator+(const QList<QAbstractEventDispatcher::TimerInfo>& l) const

C++ documentation:

Returns a list that contains all the items in this list followed by all the items in the other list.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut ListTimerInfoRef where Args: ListTimerInfoRefOpAddAssignArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::abstract_event_dispatcher::TimerInfo) -> &'l0 mut ::list::ListTimerInfoRef
C++ method: QList<QAbstractEventDispatcher::TimerInfo>& QList<QAbstractEventDispatcher::TimerInfo>::operator+=(const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

This is an overloaded function.

Appends value to the list.

See also append() and operator<<().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::list::ListTimerInfoRef) -> &'l0 mut ::list::ListTimerInfoRef
C++ method: QList<QAbstractEventDispatcher::TimerInfo>& QList<QAbstractEventDispatcher::TimerInfo>::operator+=(const QList<QAbstractEventDispatcher::TimerInfo>& l)

C++ documentation:

Appends the items of the other list to this list and returns a reference to this list.

See also operator+() and append().

fn op_assign<'l0, 'l1>(&'l0 mut self, l: &'l1 ListTimerInfoRef) -> &'l0 mut ListTimerInfoRef

C++ method: QList<QAbstractEventDispatcher::TimerInfo>& QList<QAbstractEventDispatcher::TimerInfo>::operator=(const QList<QAbstractEventDispatcher::TimerInfo>& l)

C++ documentation:

Assigns other to this list and returns a reference to this list.

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 TimerInfo

C++ method: const QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(). This function runs in constant time.

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut TimerInfo

C++ method: QAbstractEventDispatcher::TimerInfo& QList<QAbstractEventDispatcher::TimerInfo>::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference. i must be a valid index position in the list (i.e., 0 <= i < size()).

If this function is called on a list that is currently being shared, it will trigger a copy of all elements. Otherwise, this function runs in constant time. If you do not want to modify the list you should use QList::at().

See also at() and value().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut ListTimerInfoRef where Args: ListTimerInfoRefOpShlArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::abstract_event_dispatcher::TimerInfo) -> &'l0 mut ::list::ListTimerInfoRef
C++ method: QList<QAbstractEventDispatcher::TimerInfo>& QList<QAbstractEventDispatcher::TimerInfo>::operator<<(const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

This is an overloaded function.

Appends value to the list.

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::list::ListTimerInfoRef) -> &'l0 mut ::list::ListTimerInfoRef
C++ method: QList<QAbstractEventDispatcher::TimerInfo>& QList<QAbstractEventDispatcher::TimerInfo>::operator<<(const QList<QAbstractEventDispatcher::TimerInfo>& l)

C++ documentation:

Appends the items of the other list to this list and returns a reference to this list.

See also operator+=() and append().

fn pop_back(&mut self)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn pop_front(&mut self)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn prepend(&mut self, t: &TimerInfo)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::prepend(const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

Inserts value at the beginning of the list.

Example:

QList<QString> list;
list.prepend("one");
list.prepend("two");
list.prepend("three");
// list: ["three", "two", "one"]

This is the same as list.insert(0, value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also append() and insert().

fn push_back(&mut self, t: &TimerInfo)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::push_back(const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &TimerInfo)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::push_front(const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove_at(&mut self, i: c_int)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::removeAt(int i)

C++ documentation:

Removes the item at index position i. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also takeAt(), removeFirst(), removeLast(), and removeOne().

fn remove_first(&mut self)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::removeFirst()

C++ documentation:

Removes the first item in the list. Calling this function is equivalent to calling removeAt(0). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeFirst().

fn remove_last(&mut self)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::removeLast()

C++ documentation:

Removes the last item in the list. Calling this function is equivalent to calling removeAt(size() - 1). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeLast().

fn replace(&mut self, i: c_int, t: &TimerInfo)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::replace(int i, const QAbstractEventDispatcher::TimerInfo& t)

C++ documentation:

Replaces the item at index position i with value. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also operator[]() and removeAt().

fn reserve(&mut self, size: c_int)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::reserve(int size)

C++ documentation:

Reserve space for alloc elements.

If alloc is smaller than the current size of the list, nothing will happen.

Use this function to avoid repetetive reallocation of QList's internal data if you can predict how many elements will be appended. Note that the reservation applies only to the internal pointer array.

This function was introduced in Qt 4.7.

fn set_sharable(&mut self, sharable: bool)

C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QList<QAbstractEventDispatcher::TimerInfo>::size() const

C++ documentation:

Returns the number of items in the list.

See also isEmpty() and count().

fn swap<'largs, Args>(&'largs mut self, args: Args) where Args: ListTimerInfoRefSwapArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::swap

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn swap(&mut self, &mut ::list::ListTimerInfoRef) -> ()
C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::swap(QList<QAbstractEventDispatcher::TimerInfo>& other)

C++ documentation:

Swaps list other with this list. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

Variant 2

Rust arguments: fn swap(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QList<QAbstractEventDispatcher::TimerInfo>::swap(int i, int j)

C++ documentation:

Exchange the item at index position i with the item at index position j. This function assumes that both i and j are at least 0 but less than size(). To avoid failure, test that both i and j are at least 0 and less than size().

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.swap(1, 4);
// list: ["A", "E", "C", "D", "B", "F"]

See also move().

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ListTimerInfoRefTakeAtArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::abstract_event_dispatcher::TimerInfo>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::abstract_event_dispatcher::TimerInfo
C++ method: QAbstractEventDispatcher::TimerInfo QList<QAbstractEventDispatcher::TimerInfo>::takeAt(int i)

C++ documentation:

Removes the item at index position i and returns it. i must be a valid index position in the list (i.e., 0 <= i < size()).

If you don't use the return value, removeAt() is more efficient.

See also removeAt(), takeFirst(), and takeLast().

fn take_first<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ListTimerInfoRefTakeFirstArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::takeFirst

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_first(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::abstract_event_dispatcher::TimerInfo>
2) fn take_first(&mut self, cpp_utils::AsStruct) -> ::abstract_event_dispatcher::TimerInfo
C++ method: QAbstractEventDispatcher::TimerInfo QList<QAbstractEventDispatcher::TimerInfo>::takeFirst()

C++ documentation:

Removes the first item in the list and returns it. This is the same as takeAt(0). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeFirst() is more efficient.

See also takeLast(), takeAt(), and removeFirst().

fn take_last<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ListTimerInfoRefTakeLastArgs<'largs>

C++ method: QList<QAbstractEventDispatcher::TimerInfo>::takeLast

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_last(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::abstract_event_dispatcher::TimerInfo>
2) fn take_last(&mut self, cpp_utils::AsStruct) -> ::abstract_event_dispatcher::TimerInfo
C++ method: QAbstractEventDispatcher::TimerInfo QList<QAbstractEventDispatcher::TimerInfo>::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This is the same as takeAt(size() - 1). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeLast() is more efficient.

See also takeFirst(), takeAt(), and removeLast().