Module qt_core::locale::overloading

Traits

LocaleAmTextArgs
This trait represents a set of arguments accepted by Locale::am_text method.

LocaleBcp47NameArgs
This trait represents a set of arguments accepted by Locale::bcp47_name method.

LocaleCArgs
This trait represents a set of arguments accepted by Locale::c method.

LocaleCountriesForLanguageArgs
This trait represents a set of arguments accepted by Locale::countries_for_language method.

LocaleCountryToStringArgs
This trait represents a set of arguments accepted by Locale::country_to_string method.

LocaleCreateSeparatedListArgs
This trait represents a set of arguments accepted by Locale::create_separated_list method.

LocaleCurrencySymbolArgs
This trait represents a set of arguments accepted by Locale::currency_symbol method.

LocaleDateFormatArgs
This trait represents a set of arguments accepted by Locale::date_format method.

LocaleDateTimeFormatArgs
This trait represents a set of arguments accepted by Locale::date_time_format method.

LocaleDayNameArgs
This trait represents a set of arguments accepted by Locale::day_name method.

LocaleDecimalPointArgs
This trait represents a set of arguments accepted by Locale::decimal_point method.

LocaleExponentialArgs
This trait represents a set of arguments accepted by Locale::exponential method.

LocaleGroupSeparatorArgs
This trait represents a set of arguments accepted by Locale::group_separator method.

LocaleLanguageToStringArgs
This trait represents a set of arguments accepted by Locale::language_to_string method.

LocaleMatchingLocalesArgs
This trait represents a set of arguments accepted by Locale::matching_locales method.

LocaleMonthNameArgs
This trait represents a set of arguments accepted by Locale::month_name method.

LocaleNameArgs
This trait represents a set of arguments accepted by Locale::name method.

LocaleNativeCountryNameArgs
This trait represents a set of arguments accepted by Locale::native_country_name method.

LocaleNativeLanguageNameArgs
This trait represents a set of arguments accepted by Locale::native_language_name method.

LocaleNegativeSignArgs
This trait represents a set of arguments accepted by Locale::negative_sign method.

LocaleNewArgs
This trait represents a set of arguments accepted by Locale::new method.

LocalePercentArgs
This trait represents a set of arguments accepted by Locale::percent method.

LocalePmTextArgs
This trait represents a set of arguments accepted by Locale::pm_text method.

LocalePositiveSignArgs
This trait represents a set of arguments accepted by Locale::positive_sign method.

LocaleQuoteStringArgs
This trait represents a set of arguments accepted by Locale::quote_string method.

LocaleScriptToStringArgs
This trait represents a set of arguments accepted by Locale::script_to_string method.

LocaleStandaloneDayNameArgs
This trait represents a set of arguments accepted by Locale::standalone_day_name method.

LocaleStandaloneMonthNameArgs
This trait represents a set of arguments accepted by Locale::standalone_month_name method.

LocaleSystemArgs
This trait represents a set of arguments accepted by Locale::system method.

LocaleTimeFormatArgs
This trait represents a set of arguments accepted by Locale::time_format method.

LocaleToCurrencyStringArgs
This trait represents a set of arguments accepted by Locale::to_currency_string method.

LocaleToDateArgs
This trait represents a set of arguments accepted by Locale::to_date method.

LocaleToDateTimeArgs
This trait represents a set of arguments accepted by Locale::to_date_time method.

LocaleToDoubleArgs
This trait represents a set of arguments accepted by Locale::to_double method.

LocaleToFloatArgs
This trait represents a set of arguments accepted by Locale::to_float method.

LocaleToIntArgs
This trait represents a set of arguments accepted by Locale::to_int method.

LocaleToLongLongArgs
This trait represents a set of arguments accepted by Locale::to_long_long method.

LocaleToLowerArgs
This trait represents a set of arguments accepted by Locale::to_lower method.

LocaleToShortArgs
This trait represents a set of arguments accepted by Locale::to_short method.

LocaleToStringArgs
This trait represents a set of arguments accepted by Locale::to_string method.

LocaleToTimeArgs
This trait represents a set of arguments accepted by Locale::to_time method.

LocaleToUIntArgs
This trait represents a set of arguments accepted by Locale::to_u_int method.

LocaleToULongLongArgs
This trait represents a set of arguments accepted by Locale::to_u_long_long method.

LocaleToUShortArgs
This trait represents a set of arguments accepted by Locale::to_u_short method.

LocaleToUpperArgs
This trait represents a set of arguments accepted by Locale::to_upper method.

LocaleUiLanguagesArgs
This trait represents a set of arguments accepted by Locale::ui_languages method.

LocaleWeekdaysArgs
This trait represents a set of arguments accepted by Locale::weekdays method.

LocaleZeroDigitArgs
This trait represents a set of arguments accepted by Locale::zero_digit method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.