Struct qt_core::lock_file::LockFile

pub struct LockFile {
    // some fields omitted
}

C++ type: QLockFile

C++ documentation:

The QLockFile class provides locking between processes using a file.

A lock file can be used to prevent multiple processes from accessing concurrently the same resource. For instance, a configuration file on disk, or a socket, a port, a region of shared memory...

Serialization is only guaranteed if all processes that access the shared resource use QLockFile, with the same file path.

QLockFile supports two use cases: to protect a resource for a short-term operation (e.g. verifying if a configuration file has changed before saving new settings), and for long-lived protection of a resource (e.g. a document opened by a user in an editor) for an indefinite amount of time.

When protecting for a short-term operation, it is acceptable to call lock() and wait until any running operation finishes. When protecting a resource over a long time, however, the application should always call setStaleLockTime(0) and then tryLock() with a short timeout, in order to warn the user that the resource is locked.

If the process holding the lock crashes, the lock file stays on disk and can prevent any other process from accessing the shared resource, ever. For this reason, QLockFile tries to detect such a "stale" lock file, based on the process ID written into the file. To cover the situation that the process ID got reused meanwhile, the current process name is compared to the name of the process that corresponds to the process ID from the lock file. If the process names differ, the lock file is considered stale. Additionally, the last modification time of the lock file (30s by default, for the use case of a short-lived operation) is taken into account. If the lock file is found to be stale, it will be deleted.

For the use case of protecting a resource over a long time, you should therefore call setStaleLockTime(0), and when tryLock() returns LockFailedError, inform the user that the document is locked, possibly using getLockInfo() for more details.

Methods

impl LockFile

fn error(&self) -> LockError

C++ method: QLockFile::LockError QLockFile::error() const

C++ documentation:

Returns the lock file error status.

If tryLock() returns false, this function can be called to find out the reason why the locking failed.

fn get_lock_info(&self, pid: *mut c_longlong, hostname: *mut String, appname: *mut String) -> bool

C++ method: bool QLockFile::getLockInfo(long long* pid, QString* hostname, QString* appname) const

C++ documentation:

Retrieves information about the current owner of the lock file.

If tryLock() returns false, and error() returns LockFailedError, this function can be called to find out more information about the existing lock file:

the PID of the application (returned in pid)

the hostname it's running on (useful in case of networked filesystems),

the name of the application which created it (returned in appname),

Note that tryLock() automatically deleted the file if there is no running application with this PID, so LockFailedError can only happen if there is an application with this PID (it could be unrelated though).

This can be used to inform users about the existing lock file and give them the choice to delete it. After removing the file using removeStaleLockFile(), the application can call tryLock() again.

This function returns true if the information could be successfully retrieved, false if the lock file doesn't exist or doesn't contain the expected data. This can happen if the lock file was deleted between the time where tryLock() failed and the call to this function. Simply call tryLock() again if this happens.

fn is_locked(&self) -> bool

C++ method: bool QLockFile::isLocked() const

C++ documentation:

Returns true if the lock was acquired by this QLockFile instance, otherwise returns false.

See also lock(), unlock(), and tryLock().

fn lock(&mut self) -> bool

C++ method: bool QLockFile::lock()

C++ documentation:

Creates the lock file.

If another process (or another thread) has created the lock file already, this function will block until that process (or thread) releases it.

Calling this function multiple times on the same lock from the same thread without unlocking first is not allowed. This function will dead-lock when the file is locked recursively.

Returns true if the lock was acquired, false if it could not be acquired due to an unrecoverable error, such as no permissions in the parent directory.

See also unlock() and tryLock().

fn new<Args>(args: Args) -> Args::ReturnType where Args: LockFileNewArgs

C++ method: QLockFile::QLockFile

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::lock_file::LockFile
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::lock_file::LockFile>
C++ method: [constructor] void QLockFile::QLockFile(const QString& fileName)

fn remove_stale_lock_file(&mut self) -> bool

C++ method: bool QLockFile::removeStaleLockFile()

C++ documentation:

Attempts to forcefully remove an existing lock file.

Calling this is not recommended when protecting a short-lived operation: QLockFile already takes care of removing lock files after they are older than staleLockTime().

This method should only be called when protecting a resource for a long time, i.e. with staleLockTime(0), and after tryLock() returned LockFailedError, and the user agreed on removing the lock file.

Returns true on success, false if the lock file couldn't be removed. This happens on Windows, when the application owning the lock is still running.

fn set_stale_lock_time(&mut self, arg1: c_int)

C++ method: void QLockFile::setStaleLockTime(int arg1)

C++ documentation:

Sets staleLockTime to be the time in milliseconds after which a lock file is considered stale. The default value is 30000, i.e. 30 seconds. If your application typically keeps the file locked for more than 30 seconds (for instance while saving megabytes of data for 2 minutes), you should set a bigger value using setStaleLockTime().

The value of staleLockTime is used by lock() and tryLock() in order to determine when an existing lock file is considered stale, i.e. left over by a crashed process. This is useful for the case where the PID got reused meanwhile, so one way to detect a stale lock file is by the fact that it has been around for a long time.

See also staleLockTime().

fn stale_lock_time(&self) -> c_int

C++ method: int QLockFile::staleLockTime() const

C++ documentation:

Returns the time in milliseconds after which a lock file is considered stale.

See also setStaleLockTime().

fn try_lock<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: LockFileTryLockArgs<'largs>

C++ method: QLockFile::tryLock

This is an overloaded function. Available variants:

Rust arguments:
1) fn try_lock(&mut self, ()) -> bool
2) fn try_lock(&mut self, libc::c_int) -> bool
C++ method: bool QLockFile::tryLock(int timeout = ?)

C++ documentation:

Attempts to create the lock file. This function returns true if the lock was obtained; otherwise it returns false. If another process (or another thread) has created the lock file already, this function will wait for at most timeout milliseconds for the lock file to become available.

Note: Passing a negative number as the timeout is equivalent to calling lock(), i.e. this function will wait forever until the lock file can be locked if timeout is negative.

If the lock was obtained, it must be released with unlock() before another process (or thread) can successfully lock it.

Calling this function multiple times on the same lock from the same thread without unlocking first is not allowed, this function will always return false when attempting to lock the file recursively.

See also lock() and unlock().

fn unlock(&mut self)

C++ method: void QLockFile::unlock()

C++ documentation:

Releases the lock, by deleting the lock file.

Calling unlock() without locking the file first, does nothing.

See also lock() and tryLock().