Struct qt_core::logging_category::LoggingCategory

pub struct LoggingCategory {
    // some fields omitted
}

C++ type: QLoggingCategory

C++ documentation:

The QLoggingCategory class represents a category, or 'area' in the logging infrastructure.

QLoggingCategory represents a certain logging category - identified by a string - at runtime. A category can be configured to enable or disable logging of messages per message type. Whether a message type is enabled or not can be checked with the isDebugEnabled(), isInfoEnabled(), isWarningEnabled(), and isCriticalEnabled() methods.

All objects are meant to be configured by a common registry (see also Configuring Categories). Different objects can also represent the same category. It is therefore not recommended to export objects across module boundaries, nor to manipulate the objects directly, nor to inherit from QLoggingCategory.

Methods

impl LoggingCategory

fn category_name(&self) -> *const c_char

C++ method: const char* QLoggingCategory::categoryName() const

C++ documentation:

Returns the name of the category.

fn default_category() -> *mut LoggingCategory

C++ method: static QLoggingCategory* QLoggingCategory::defaultCategory()

C++ documentation:

Returns a pointer to the global category "default" that is used e.g. by qDebug(), qInfo(), qWarning(), qCritical(), qFatal().

Note: The returned pointer may be null during destruction of static objects.

Note: Ownership of the category is not transferred, do not delete the returned pointer.

fn install_filter(arg1: extern fn(*mut LoggingCategory)) -> extern fn(*mut LoggingCategory)

C++ method: static void (*FN_PTR)(QLoggingCategory*) QLoggingCategory::installFilter(void (*FN_PTR)(QLoggingCategory*) arg1)

C++ documentation:

Installs a function filter that is used to determine which categories and message types should be enabled. Returns a pointer to the previous installed filter.

Every QLoggingCategory object created is passed to the filter, and the filter is free to change the respective category configuration with setEnabled().

The filter might be called concurrently from different threads, and therefore has to be reentrant.

Example:

QLoggingCategory::CategoryFilter oldCategoryFilter;

void myCategoryFilter(QLoggingCategory *category) { // configure driver.usb category here, otherwise forward to to default filter. if (qstrcmp(category->categoryName(), "driver.usb") == 0) category->setEnabled(QtDebugMsg, true); else oldCategoryFilter(category); }
An alternative way of configuring the default filter is via setFilterRules().

fn is_critical_enabled(&self) -> bool

C++ method: bool QLoggingCategory::isCriticalEnabled() const

C++ documentation:

Returns true if critical messages should be shown for this category. Returns false otherwise.

Note: The qCCritical() macro already does this check before executing any code. However, calling this method may be useful to avoid expensive generation of data that is only used for debug output.

fn is_debug_enabled(&self) -> bool

C++ method: bool QLoggingCategory::isDebugEnabled() const

C++ documentation:

Returns true if debug messages should be shown for this category. Returns false otherwise.

Note: The qCDebug() macro already does this check before executing any code. However, calling this method may be useful to avoid expensive generation of data that is only used for debug output.

fn is_enabled(&self, type_: MsgType) -> bool

C++ method: bool QLoggingCategory::isEnabled(QtMsgType type) const

C++ documentation:

Returns true if a message of type msgtype for the category should be shown. Returns false otherwise.

fn is_info_enabled(&self) -> bool

C++ method: bool QLoggingCategory::isInfoEnabled() const

C++ documentation:

Returns true if informational messages should be shown for this category. Returns false otherwise.

Note: The qCInfo() macro already does this check before executing any code. However, calling this method may be useful to avoid expensive generation of data that is only used for debug output.

This function was introduced in Qt 5.5.

fn is_warning_enabled(&self) -> bool

C++ method: bool QLoggingCategory::isWarningEnabled() const

C++ documentation:

Returns true if warning messages should be shown for this category. Returns false otherwise.

Note: The qCWarning() macro already does this check before executing any code. However, calling this method may be useful to avoid expensive generation of data that is only used for debug output.

fn new<Args>(args: Args) -> Args::ReturnType where Args: LoggingCategoryNewArgs

C++ method: QLoggingCategory::QLoggingCategory

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*const libc::c_char, cpp_utils::AsStruct)) -> ::logging_category::LoggingCategory
2) fn new((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::logging_category::LoggingCategory>
C++ method: [constructor] void QLoggingCategory::QLoggingCategory(const char* category)

Variant 2

Rust arguments:
1) fn new((*const libc::c_char, ::message_logger::MsgType, cpp_utils::AsStruct)) -> ::logging_category::LoggingCategory
2) fn new((*const libc::c_char, ::message_logger::MsgType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::logging_category::LoggingCategory>
C++ method: [constructor] void QLoggingCategory::QLoggingCategory(const char* category, QtMsgType severityLevel)

fn op_call<'l0>(&'l0 self) -> &'l0 LoggingCategory

C++ method: const QLoggingCategory& QLoggingCategory::operator()() const

C++ documentation:

Returns the object itself. This allows both a QLoggingCategory variable, and a factory method returning a QLoggingCategory, to be used in qCDebug(), qCWarning(), qCCritical() macros.

fn op_call_mut<'l0>(&'l0 mut self) -> &'l0 mut LoggingCategory

C++ method: QLoggingCategory& QLoggingCategory::operator()()

C++ documentation:

Returns the object itself. This allows both a QLoggingCategory variable, and a factory method returning a QLoggingCategory, to be used in qCDebug(), qCWarning(), qCCritical() macros.

fn set_enabled(&mut self, type_: MsgType, enable: bool)

C++ method: void QLoggingCategory::setEnabled(QtMsgType type, bool enable)

C++ documentation:

Changes the message type type for the category to enable.

Note: Changes only affect the current QLoggingCategory object, and won't change the settings of other objects for the same category name. Use either setFilterRules() or installFilter() to change the configuration globally.

Note: QtFatalMsg cannot be changed. It will always return true.

See also isEnabled().

fn set_filter_rules(rules: &String)

C++ method: static void QLoggingCategory::setFilterRules(const QString& rules)

C++ documentation:

Configures which categories and message types should be enabled through a a set of rules.

Example:

    QLoggingCategory::setFilterRules(QStringLiteral("driver.usb.debug=true"));

Note: The rules might be ignored if a custom category filter is installed with installFilter(), or if the user defined QT_LOGGING_CONF or QT_LOGGING_RULES environment variable.