Struct qt_core::map::MapStringRefVariantRef

pub struct MapStringRefVariantRef {
    // some fields omitted
}

C++ type: QMap<QString, QVariant>

Methods

impl MapStringRefVariantRef

fn clear(&mut self)

C++ method: void QMap<QString, QVariant>::clear()

C++ documentation:

Removes all items from the map.

See also remove().

fn contains(&self, key: &String) -> bool

C++ method: bool QMap<QString, QVariant>::contains(const QString& key) const

C++ documentation:

Returns true if the map contains an item with key key; otherwise returns false.

See also count() and QMultiMap::contains().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: MapStringRefVariantRefCountArgs<'largs>

C++ method: QMap<QString, QVariant>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QMap<QString, QVariant>::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &::string::String) -> libc::c_int
C++ method: int QMap<QString, QVariant>::count(const QString& key) const

C++ documentation:

Returns the number of items associated with key key.

See also contains(), insertMulti(), and QMultiMap::count().

fn detach(&mut self)

C++ method: void QMap<QString, QVariant>::detach()

fn empty(&self) -> bool

C++ method: bool QMap<QString, QVariant>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the map is empty; otherwise returning false.

fn first<'l0>(&'l0 self) -> &'l0 Variant

C++ method: const QVariant& QMap<QString, QVariant>::first() const

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.2.

fn first_key<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QMap<QString, QVariant>::firstKey() const

C++ documentation:

Returns a reference to the smallest key in the map. This function assumes that the map is not empty.

This executes in constant time.

This function was introduced in Qt 5.2.

See also lastKey(), first(), and isEmpty().

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut Variant

C++ method: QVariant& QMap<QString, QVariant>::first()

C++ documentation:

Returns a reference to the first value in the map, that is the value mapped to the smallest key. This function assumes that the map is not empty.

When unshared (or const version is called), this executes in constant time.

This function was introduced in Qt 5.2.

See also last(), firstKey(), and isEmpty().

fn is_detached(&self) -> bool

C++ method: bool QMap<QString, QVariant>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QMap<QString, QVariant>::isEmpty() const

C++ documentation:

Returns true if the map contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &MapStringRefVariantRef) -> bool

C++ method: bool QMap<QString, QVariant>::isSharedWith(const QMap<QString, QVariant>& other) const

fn key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefKeyArgs<'largs>

C++ method: QMap<QString, QVariant>::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key(&self, (&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn key(&self, (&::variant::Variant, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn key(&self, (&::variant::Variant, cpp_utils::AsStruct)) -> ::string::String
4) fn key(&self, (&::variant::Variant, &::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: const QString QMap<QString, QVariant>::key(const QVariant& value, const QString& defaultKey = ?) const

C++ documentation:

This is an overloaded function.

Returns the first key with value value, or defaultKey if the map contains no item with value value. If no defaultKey is provided the function returns a default-constructed key.

This function can be slow (linear time), because QMap's internal data structure is optimized for fast lookup by key, not by value.

This function was introduced in Qt 4.3.

See also value() and keys().

fn keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefKeysArgs<'largs>

C++ method: QMap<QString, QVariant>::keys

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListStringRef>
2) fn keys(&self, cpp_utils::AsStruct) -> ::list::ListStringRef
C++ method: QList<QString> QMap<QString, QVariant>::keys() const

C++ documentation:

Returns a list containing all the keys in the map in ascending order. Keys that occur multiple times in the map (because items were inserted with insertMulti(), or unite() was used) also occur multiple times in the list.

To obtain a list of unique keys, where each key from the map only occurs once, use uniqueKeys().

The order is guaranteed to be the same as that used by values().

See also uniqueKeys(), values(), and key().

Variant 2

Rust arguments:
1) fn keys(&self, (&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListStringRef>
2) fn keys(&self, (&::variant::Variant, cpp_utils::AsStruct)) -> ::list::ListStringRef
C++ method: QList<QString> QMap<QString, QVariant>::keys(const QVariant& value) const

C++ documentation:

This is an overloaded function.

Returns a list containing all the keys associated with value value in ascending order.

This function can be slow (linear time), because QMap's internal data structure is optimized for fast lookup by key, not by value.

fn last<'l0>(&'l0 self) -> &'l0 Variant

C++ method: const QVariant& QMap<QString, QVariant>::last() const

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.2.

fn last_key<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QMap<QString, QVariant>::lastKey() const

C++ documentation:

Returns a reference to the largest key in the map. This function assumes that the map is not empty.

This executes in logarithmic time.

This function was introduced in Qt 5.2.

See also firstKey(), last(), and isEmpty().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut Variant

C++ method: QVariant& QMap<QString, QVariant>::last()

C++ documentation:

Returns a reference to the last value in the map, that is the value mapped to the largest key. This function assumes that the map is not empty.

When unshared (or const version is called), this executes in logarithmic time.

This function was introduced in Qt 5.2.

See also first(), lastKey(), and isEmpty().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefNewArgs

C++ method: QMap<QString, QVariant>::QMap

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::map::MapStringRefVariantRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::map::MapStringRefVariantRef>
C++ method: [constructor] void QMap<QString, QVariant>::QMap()

Variant 2

Rust arguments:
1) fn new((&::map::MapStringRefVariantRef, cpp_utils::AsStruct)) -> ::map::MapStringRefVariantRef
2) fn new((&::map::MapStringRefVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::map::MapStringRefVariantRef>
C++ method: [constructor] void QMap<QString, QVariant>::QMap(const QMap<QString, QVariant>& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 MapStringRefVariantRef) -> &'l0 mut MapStringRefVariantRef

C++ method: QMap<QString, QVariant>& QMap<QString, QVariant>::operator=(const QMap<QString, QVariant>& other)

C++ documentation:

Assigns other to this map and returns a reference to this map.

fn op_eq(&self, other: &MapStringRefVariantRef) -> bool

C++ method: bool QMap<QString, QVariant>::operator==(const QMap<QString, QVariant>& other) const

C++ documentation:

Returns true if other is equal to this map; otherwise returns false.

Two maps are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator!=().

fn op_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefOpIndexArgs<'largs>

C++ method: QMap<QString, QVariant>::operator[]

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_index(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn op_index(&self, (&::string::String, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: const QVariant QMap<QString, QVariant>::operator[](const QString& key) const

C++ documentation:

This is an overloaded function.

Same as value().

fn op_index_mut<'l0, 'l1>(&'l0 mut self, key: &'l1 String) -> &'l0 mut Variant

C++ method: QVariant& QMap<QString, QVariant>::operator[](const QString& key)

C++ documentation:

Returns the value associated with the key key as a modifiable reference.

If the map contains no item with key key, the function inserts a default-constructed value into the map with key key, and returns a reference to it. If the map contains multiple items with key key, this function returns a reference to the most recently inserted value.

See also insert() and value().

fn op_neq(&self, other: &MapStringRefVariantRef) -> bool

C++ method: bool QMap<QString, QVariant>::operator!=(const QMap<QString, QVariant>& other) const

C++ documentation:

Returns true if other is not equal to this map; otherwise returns false.

Two maps are considered equal if they contain the same (key, value) pairs.

This function requires the value type to implement operator==().

See also operator==().

fn remove(&mut self, key: &String) -> c_int

C++ method: int QMap<QString, QVariant>::remove(const QString& key)

C++ documentation:

Removes all the items that have the key key from the map. Returns the number of items removed which is usually 1 but will be 0 if the key isn't in the map, or > 1 if insertMulti() has been used with the key.

See also clear(), take(), and QMultiMap::remove().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QMap<QString, QVariant>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QMap<QString, QVariant>::size() const

C++ documentation:

Returns the number of (key, value) pairs in the map.

See also isEmpty() and count().

fn swap(&mut self, other: &mut MapStringRefVariantRef)

C++ method: void QMap<QString, QVariant>::swap(QMap<QString, QVariant>& other)

C++ documentation:

Swaps map other with this map. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefTakeArgs<'largs>

C++ method: QMap<QString, QVariant>::take

This is an overloaded function. Available variants:

Rust arguments:
1) fn take(&mut self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn take(&mut self, (&::string::String, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QMap<QString, QVariant>::take(const QString& key)

C++ documentation:

Removes the item with the key key from the map and returns the value associated with it.

If the item does not exist in the map, the function simply returns a default-constructed value. If there are multiple items for key in the map, only the most recently inserted one is removed and returned.

If you don't use the return value, remove() is more efficient.

See also remove().

fn unique_keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefUniqueKeysArgs<'largs>

C++ method: QMap<QString, QVariant>::uniqueKeys

This is an overloaded function. Available variants:

Rust arguments:
1) fn unique_keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListStringRef>
2) fn unique_keys(&self, cpp_utils::AsStruct) -> ::list::ListStringRef
C++ method: QList<QString> QMap<QString, QVariant>::uniqueKeys() const

C++ documentation:

Returns a list containing all the keys in the map in ascending order. Keys that occur multiple times in the map (because items were inserted with insertMulti(), or unite() was used) occur only once in the returned list.

This function was introduced in Qt 4.2.

See also keys() and values().

fn unite<'l0, 'l1>(&'l0 mut self, other: &'l1 MapStringRefVariantRef) -> &'l0 mut MapStringRefVariantRef

C++ method: QMap<QString, QVariant>& QMap<QString, QVariant>::unite(const QMap<QString, QVariant>& other)

C++ documentation:

Inserts all the items in the other map into this map. If a key is common to both maps, the resulting map will contain the key multiple times.

See also insertMulti().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefValueArgs<'largs>

C++ method: QMap<QString, QVariant>::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn value(&self, (&::string::String, &::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
3) fn value(&self, (&::string::String, cpp_utils::AsStruct)) -> ::variant::Variant
4) fn value(&self, (&::string::String, &::variant::Variant, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: const QVariant QMap<QString, QVariant>::value(const QString& key, const QVariant& defaultValue = ?) const

C++ documentation:

Returns the value associated with the key key.

If the map contains no item with key key, the function returns defaultValue. If no defaultValue is specified, the function returns a default-constructed value. If there are multiple items for key in the map, the value of the most recently inserted one is returned.

See also key(), values(), contains(), and operator[]().

fn values<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MapStringRefVariantRefValuesArgs<'largs>

C++ method: QMap<QString, QVariant>::values

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn values(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListVariantRef>
2) fn values(&self, cpp_utils::AsStruct) -> ::list::ListVariantRef
C++ method: QList<QVariant> QMap<QString, QVariant>::values() const

C++ documentation:

Returns a list containing all the values in the map, in ascending order of their keys. If a key is associated with multiple values, all of its values will be in the list, and not just the most recently inserted one.

See also keys() and value().

Variant 2

Rust arguments:
1) fn values(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListVariantRef>
2) fn values(&self, (&::string::String, cpp_utils::AsStruct)) -> ::list::ListVariantRef
C++ method: QList<QVariant> QMap<QString, QVariant>::values(const QString& key) const

C++ documentation:

This is an overloaded function.

Returns a list containing all the values associated with key key, from the most recently inserted to the least recently inserted one.

See also count() and insertMulti().