Module qt_core::margins::overloading

Traits

MarginsNewArgs
This trait represents a set of arguments accepted by Margins::new method.

MarginsOpAddAssignArgs
This trait represents a set of arguments accepted by Margins::op_add_assign method.

MarginsOpDivAssignArgs
This trait represents a set of arguments accepted by Margins::op_div_assign method.

MarginsOpMulAssignArgs
This trait represents a set of arguments accepted by Margins::op_mul_assign method.

MarginsOpSubAssignArgs
This trait represents a set of arguments accepted by Margins::op_sub_assign method.

OpAddArgs
This trait represents a set of arguments accepted by op_add method.

OpDivArgs
This trait represents a set of arguments accepted by op_div method.

OpMulArgs
This trait represents a set of arguments accepted by op_mul method.

OpNegArgs
This trait represents a set of arguments accepted by op_neg method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

OpShrArgs
This trait represents a set of arguments accepted by op_shr method.

OpSubArgs
This trait represents a set of arguments accepted by op_sub method.

OpUnaryPlusArgs
This trait represents a set of arguments accepted by op_unary_plus method.