Struct qt_core::margins::Margins

pub struct Margins {
    // some fields omitted
}

C++ type: QMargins

C++ documentation:

The QMargins class defines the four margins of a rectangle.

QMargin defines a set of four margins; left, top, right and bottom, that describe the size of the borders surrounding a rectangle.

The isNull() function returns true only if all margins are set to zero.

QMargin objects can be streamed as well as compared.

Methods

impl Margins

fn bottom(&self) -> c_int

C++ method: int QMargins::bottom() const

C++ documentation:

Returns the bottom margin.

See also setBottom().

fn is_null(&self) -> bool

C++ method: bool QMargins::isNull() const

C++ documentation:

Returns true if all margins are is 0; otherwise returns false.

fn left(&self) -> c_int

C++ method: int QMargins::left() const

C++ documentation:

Returns the left margin.

See also setLeft().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MarginsNewArgs

C++ method: QMargins::QMargins

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::margins::Margins
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::margins::Margins>
C++ method: [constructor] void QMargins::QMargins()

Variant 2

Rust arguments:
1) fn new((libc::c_int, libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::margins::Margins
2) fn new((libc::c_int, libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::margins::Margins>
C++ method: [constructor] void QMargins::QMargins(int left, int top, int right, int bottom)

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Margins where Args: MarginsOpAddAssignArgs<'largs>

C++ method: QMargins::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::margins::Margins) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator+=(const QMargins& margins)

C++ documentation:

Add each component of margins to the respective component of this object and returns a reference to it.

This function was introduced in Qt 5.1.

See also operator-=().

Variant 2

Rust arguments: fn op_add_assign(&mut self, libc::c_int) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator+=(int arg1)

C++ documentation:

This is an overloaded function.

Adds the addend to each component of this object and returns a reference to it.

See also operator-=().

fn op_div_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Margins where Args: MarginsOpDivAssignArgs<'largs>

C++ method: QMargins::operator/=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_div_assign(&mut self, libc::c_double) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator/=(double arg1)

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.1.

See also operator*=().

Variant 2

Rust arguments: fn op_div_assign(&mut self, libc::c_int) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator/=(int arg1)

C++ documentation:

Divides each component of this object by divisor and returns a reference to it.

This function was introduced in Qt 5.1.

See also operator*=().

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Margins where Args: MarginsOpMulAssignArgs<'largs>

C++ method: QMargins::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, libc::c_double) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator*=(double arg1)

C++ documentation:

This is an overloaded function.

Multiplies each component of this object by factor and returns a reference to it.

This function was introduced in Qt 5.1.

See also operator/=().

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_int) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator*=(int arg1)

C++ documentation:

Multiplies each component of this object by factor and returns a reference to it.

This function was introduced in Qt 5.1.

See also operator/=().

fn op_sub_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Margins where Args: MarginsOpSubAssignArgs<'largs>

C++ method: QMargins::operator-=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_sub_assign(&mut self, &'l1 ::margins::Margins) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator-=(const QMargins& margins)

C++ documentation:

Subtract each component of margins from the respective component of this object and returns a reference to it.

This function was introduced in Qt 5.1.

See also operator+=().

Variant 2

Rust arguments: fn op_sub_assign(&mut self, libc::c_int) -> &'l0 mut ::margins::Margins
C++ method: QMargins& QMargins::operator-=(int arg1)

C++ documentation:

This is an overloaded function.

Subtracts the subtrahend from each component of this object and returns a reference to it.

See also operator+=().

fn right(&self) -> c_int

C++ method: int QMargins::right() const

C++ documentation:

Returns the right margin.

See also setRight().

fn set_bottom(&mut self, bottom: c_int)

C++ method: void QMargins::setBottom(int bottom)

C++ documentation:

Sets the bottom margin to bottom.

See also bottom().

fn set_left(&mut self, left: c_int)

C++ method: void QMargins::setLeft(int left)

C++ documentation:

Sets the left margin to left.

See also left().

fn set_right(&mut self, right: c_int)

C++ method: void QMargins::setRight(int right)

C++ documentation:

Sets the right margin to right.

See also right().

fn set_top(&mut self, top: c_int)

C++ method: void QMargins::setTop(int top)

C++ documentation:

Sets the Top margin to Top.

See also top().

fn top(&self) -> c_int

C++ method: int QMargins::top() const

C++ documentation:

Returns the top margin.

See also setTop().