Struct qt_core::margins_f::MarginsF

pub struct MarginsF {
    // some fields omitted
}

C++ type: QMarginsF

C++ documentation:

The QMarginsF class defines the four margins of a rectangle.

QMarginsF defines a set of four margins; left, top, right and bottom, that describe the size of the borders surrounding a rectangle.

The isNull() function returns true only if all margins are set to zero.

QMarginsF objects can be streamed as well as compared.

Methods

impl MarginsF

fn bottom(&self) -> c_double

C++ method: double QMarginsF::bottom() const

C++ documentation:

Returns the bottom margin.

See also setBottom().

fn is_null(&self) -> bool

C++ method: bool QMarginsF::isNull() const

C++ documentation:

Returns true if all margins are 0; otherwise returns false.

fn left(&self) -> c_double

C++ method: double QMarginsF::left() const

C++ documentation:

Returns the left margin.

See also setLeft().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MarginsFNewArgs

C++ method: QMarginsF::QMarginsF

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::margins_f::MarginsF
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::margins_f::MarginsF>
C++ method: [constructor] void QMarginsF::QMarginsF()

Variant 2

Rust arguments:
1) fn new((&::margins::Margins, cpp_utils::AsStruct)) -> ::margins_f::MarginsF
2) fn new((&::margins::Margins, cpp_utils::AsBox)) -> cpp_utils::CppBox<::margins_f::MarginsF>
C++ method: [constructor] void QMarginsF::QMarginsF(const QMargins& margins)

Variant 3

Rust arguments:
1) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::margins_f::MarginsF
2) fn new((libc::c_double, libc::c_double, libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::margins_f::MarginsF>
C++ method: [constructor] void QMarginsF::QMarginsF(double left, double top, double right, double bottom)

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut MarginsF where Args: MarginsFOpAddAssignArgs<'largs>

C++ method: QMarginsF::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::margins_f::MarginsF) -> &'l0 mut ::margins_f::MarginsF
C++ method: QMarginsF& QMarginsF::operator+=(const QMarginsF& margins)

C++ documentation:

Add each component of margins to the respective component of this object and returns a reference to it.

See also operator-=().

Variant 2

Rust arguments: fn op_add_assign(&mut self, libc::c_double) -> &'l0 mut ::margins_f::MarginsF
C++ method: QMarginsF& QMarginsF::operator+=(double addend)

C++ documentation:

This is an overloaded function.

Adds the addend to each component of this object and returns a reference to it.

See also operator-=().

fn op_div_assign<'l0>(&'l0 mut self, divisor: c_double) -> &'l0 mut MarginsF

C++ method: QMarginsF& QMarginsF::operator/=(double divisor)

C++ documentation:

Divides each component of this object by divisor and returns a reference to it.

See also operator*=().

fn op_mul_assign<'l0>(&'l0 mut self, factor: c_double) -> &'l0 mut MarginsF

C++ method: QMarginsF& QMarginsF::operator*=(double factor)

C++ documentation:

Multiplies each component of this object by factor and returns a reference to it.

See also operator/=().

fn op_sub_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut MarginsF where Args: MarginsFOpSubAssignArgs<'largs>

C++ method: QMarginsF::operator-=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_sub_assign(&mut self, &'l1 ::margins_f::MarginsF) -> &'l0 mut ::margins_f::MarginsF
C++ method: QMarginsF& QMarginsF::operator-=(const QMarginsF& margins)

C++ documentation:

Subtract each component of margins from the respective component of this object and returns a reference to it.

See also operator+=().

Variant 2

Rust arguments: fn op_sub_assign(&mut self, libc::c_double) -> &'l0 mut ::margins_f::MarginsF
C++ method: QMarginsF& QMarginsF::operator-=(double subtrahend)

C++ documentation:

This is an overloaded function.

Subtracts the subtrahend from each component of this object and returns a reference to it.

See also operator+=().

fn right(&self) -> c_double

C++ method: double QMarginsF::right() const

C++ documentation:

Returns the right margin.

See also setRight().

fn set_bottom(&mut self, bottom: c_double)

C++ method: void QMarginsF::setBottom(double bottom)

C++ documentation:

Sets the bottom margin to bottom.

See also bottom().

fn set_left(&mut self, left: c_double)

C++ method: void QMarginsF::setLeft(double left)

C++ documentation:

Sets the left margin to left.

See also left().

fn set_right(&mut self, right: c_double)

C++ method: void QMarginsF::setRight(double right)

C++ documentation:

Sets the right margin to right.

See also right().

fn set_top(&mut self, top: c_double)

C++ method: void QMarginsF::setTop(double top)

C++ documentation:

Sets the Top margin to Top.

See also top().

fn to_margins<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MarginsFToMarginsArgs<'largs>

C++ method: QMarginsF::toMargins

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_margins(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::margins::Margins>
2) fn to_margins(&self, cpp_utils::AsStruct) -> ::margins::Margins
C++ method: QMargins QMarginsF::toMargins() const

C++ documentation:

Returns an integer based copy of this margins object.

Note that the components in the returned margins will be rounded to the nearest integer.

See also QMarginsF().

fn top(&self) -> c_double

C++ method: double QMarginsF::top() const

C++ documentation:

Returns the top margin.

See also setTop().