Function qt_core::math::next_power_of_two

pub fn next_power_of_two<Args>(args: Args) -> Args::ReturnType where Args: NextPowerOfTwoArgs

C++ method: qNextPowerOfTwo

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn next_power_of_two(i32) -> u32
C++ method: quint32 qNextPowerOfTwo(qint32 v)

C++ documentation:

This is an overloaded function.

This function returns the nearest power of two greater than value. For negative values it returns 0.

This function was introduced in Qt 5.4.

Variant 2

Rust arguments: fn next_power_of_two(u32) -> u32
C++ method: quint32 qNextPowerOfTwo(quint32 v)

C++ documentation:

This function returns the nearest power of two greater than value. For 0 it returns 1, and for values larger than or equal to 231 it returns 0.

This function was introduced in Qt 5.4.

Variant 3

Rust arguments: fn next_power_of_two(i64) -> u64
C++ method: quint64 qNextPowerOfTwo(qint64 v)

C++ documentation:

This is an overloaded function.

This function returns the nearest power of two greater than value. For negative values it returns 0.

This function was introduced in Qt 5.4.

Variant 4

Rust arguments: fn next_power_of_two(u64) -> u64
C++ method: quint64 qNextPowerOfTwo(quint64 v)

C++ documentation:

This function returns the nearest power of two greater than value. For 0 it returns 1, and for values larger than or equal to 263 it returns 0.

This function was introduced in Qt 5.4.