Struct qt_core::message_authentication_code::MessageAuthenticationCode

pub struct MessageAuthenticationCode {
    // some fields omitted
}

C++ type: QMessageAuthenticationCode

C++ documentation:

The QMessageAuthenticationCode class provides a way to generate hash-based message authentication codes.

QMessageAuthenticationCode supports all cryptographic hashes which are supported by QCryptographicHash.

To generate message authentication code, pass hash algorithm QCryptographicHash::Algorithm to constructor, then set key and message by setKey() and addData() functions. Result can be acquired by result() function.

    QByteArray key = "key";
    QByteArray message = "The quick brown fox jumps over the lazy dog";
    ...
    QMessageAuthenticationCode code(QCryptographicHash::Sha1);
    code.setKey(key);
    code.addData(message);
    code.result().toHex();      // returns "de7c9b85b8b78aa6bc8a7a36f70a90701c9db4d9"

Alternatively, this effect can be achieved by providing message, key and method to hash() method.

    QMessageAuthenticationCode::hash(message, key, QCryptographicHash::Sha1).toHex();

Methods

impl MessageAuthenticationCode

fn add_data<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: MessageAuthenticationCodeAddDataArgs<'largs>

C++ method: QMessageAuthenticationCode::addData

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn add_data(&mut self, *mut ::qiod_evice::QiodEvice) -> bool
C++ method: bool QMessageAuthenticationCode::addData(QIODevice* device)

C++ documentation:

Reads the data from the open QIODevice device until it ends and adds it to message. Returns true if reading was successful.

Note: device must be already opened.

Variant 2

Rust arguments: fn add_data(&mut self, &::byte_array::ByteArray) -> ()
C++ method: void QMessageAuthenticationCode::addData(const QByteArray& data)

C++ documentation:

This function overloads addData().

Variant 3

Rust arguments: fn add_data(&mut self, (*const libc::c_char, libc::c_int)) -> ()
C++ method: void QMessageAuthenticationCode::addData(const char* data, int length)

C++ documentation:

Adds the first length chars of data to the message.

fn hash<Args>(args: Args) -> Args::ReturnType where Args: MessageAuthenticationCodeHashArgs

C++ method: QMessageAuthenticationCode::hash

This is an overloaded function. Available variants:

Rust arguments:
1) fn hash((&::byte_array::ByteArray, &::byte_array::ByteArray, ::cryptographic_hash::Algorithm, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn hash((&::byte_array::ByteArray, &::byte_array::ByteArray, ::cryptographic_hash::Algorithm, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: static QByteArray QMessageAuthenticationCode::hash(const QByteArray& message, const QByteArray& key, QCryptographicHash::Algorithm method)

C++ documentation:

Returns the authentication code for the message message using the key key and the method method.

fn new<Args>(args: Args) -> Args::ReturnType where Args: MessageAuthenticationCodeNewArgs

C++ method: QMessageAuthenticationCode::QMessageAuthenticationCode

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((::cryptographic_hash::Algorithm, cpp_utils::AsStruct)) -> ::message_authentication_code::MessageAuthenticationCode
2) fn new((::cryptographic_hash::Algorithm, &::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::message_authentication_code::MessageAuthenticationCode
3) fn new((::cryptographic_hash::Algorithm, cpp_utils::AsBox)) -> cpp_utils::CppBox<::message_authentication_code::MessageAuthenticationCode>
4) fn new((::cryptographic_hash::Algorithm, &::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::message_authentication_code::MessageAuthenticationCode>
C++ method: [constructor] void QMessageAuthenticationCode::QMessageAuthenticationCode(QCryptographicHash::Algorithm method, const QByteArray& key = ?)

fn reset(&mut self)

C++ method: void QMessageAuthenticationCode::reset()

C++ documentation:

Resets message data. Calling this method doesn't affect the key.

fn result<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MessageAuthenticationCodeResultArgs<'largs>

C++ method: QMessageAuthenticationCode::result

This is an overloaded function. Available variants:

Rust arguments:
1) fn result(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn result(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QMessageAuthenticationCode::result() const

C++ documentation:

Returns the final authentication code.

See also QByteArray::toHex().

fn set_key(&mut self, key: &ByteArray)

C++ method: void QMessageAuthenticationCode::setKey(const QByteArray& key)

C++ documentation:

Sets secret key. Calling this method automatically resets the object state.