Module qt_core::message_logger::overloading

Traits

FormatLogMessageArgs
This trait represents a set of arguments accepted by format_log_message method.

MessageLoggerCriticalArgs
This trait represents a set of arguments accepted by MessageLogger::critical method.

MessageLoggerDebugArgs
This trait represents a set of arguments accepted by MessageLogger::debug method.

MessageLoggerInfoArgs
This trait represents a set of arguments accepted by MessageLogger::info method.

MessageLoggerNewArgs
This trait represents a set of arguments accepted by MessageLogger::new method.

MessageLoggerNoDebugArgs
This trait represents a set of arguments accepted by MessageLogger::no_debug method.

MessageLoggerWarningArgs
This trait represents a set of arguments accepted by MessageLogger::warning method.