Struct qt_core::message_logger::MessageLogger

pub struct MessageLogger {
    // some fields omitted
}

C++ type: QMessageLogger

C++ documentation:

The QMessageLogger class generates log messages.

QMessageLogger is used to generate messages for the Qt logging framework. Usually one uses it through qDebug(), qInfo(), qWarning(), qCritical, or qFatal() functions, which are actually macros: For example qDebug() expands to QMessageLogger(FILE, LINE, Q_FUNC_INFO).debug() for debug builds, and QMessageLogger(0, 0, 0).debug() for release builds.

One example of direct use is to forward errors that stem from a scripting language, e.g. QML:

    void statusChanged(QDeclarativeComponent::Status status) {
        if (status == QDeclarativeComponent::Error) {
            foreach (const QDeclarativeError &error, component->errors()) {
                const QByteArray file = error.url().toEncoded();
                QMessageLogger(file.constData(), error.line(), 0).debug() << error.description();
            }
        }
    }

Methods

impl MessageLogger

fn critical<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MessageLoggerCriticalArgs<'largs>

C++ method: QMessageLogger::critical

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn critical(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::debug::Debug>
2) fn critical(&self, cpp_utils::AsStruct) -> ::debug::Debug
C++ method: QDebug QMessageLogger::critical() const

C++ documentation:

Logs a critical message using a QDebug stream

See also qCritical() and QDebug.

Variant 2

Rust arguments:
1) fn critical(&self, (&::logging_category::LoggingCategory, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
2) fn critical(&self, (&::logging_category::LoggingCategory, cpp_utils::AsStruct)) -> ::debug::Debug
C++ method: QDebug QMessageLogger::critical(const QLoggingCategory& cat) const

C++ documentation:

Logs a critical message into category cat using a QDebug stream.

This function was introduced in Qt 5.3.

See also qCCritical() and QDebug.

fn debug<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MessageLoggerDebugArgs<'largs>

C++ method: QMessageLogger::debug

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn debug(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::debug::Debug>
2) fn debug(&self, cpp_utils::AsStruct) -> ::debug::Debug
C++ method: QDebug QMessageLogger::debug() const

C++ documentation:

Logs a debug message using a QDebug stream

See also qDebug() and QDebug.

Variant 2

Rust arguments:
1) fn debug(&self, (&::logging_category::LoggingCategory, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
2) fn debug(&self, (&::logging_category::LoggingCategory, cpp_utils::AsStruct)) -> ::debug::Debug
C++ method: QDebug QMessageLogger::debug(const QLoggingCategory& cat) const

C++ documentation:

Logs a debug message into category cat using a QDebug stream.

This function was introduced in Qt 5.3.

See also qCDebug() and QDebug.

fn info<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MessageLoggerInfoArgs<'largs>

C++ method: QMessageLogger::info

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn info(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::debug::Debug>
2) fn info(&self, cpp_utils::AsStruct) -> ::debug::Debug
C++ method: QDebug QMessageLogger::info() const

C++ documentation:

Logs an informational message using a QDebug stream.

This function was introduced in Qt 5.5.

See also qInfo() and QDebug.

Variant 2

Rust arguments:
1) fn info(&self, (&::logging_category::LoggingCategory, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
2) fn info(&self, (&::logging_category::LoggingCategory, cpp_utils::AsStruct)) -> ::debug::Debug
C++ method: QDebug QMessageLogger::info(const QLoggingCategory& cat) const

C++ documentation:

Logs an informational message into the category cat using a QDebug stream.

This function was introduced in Qt 5.5.

See also qCInfo() and QDebug.

fn new<Args>(args: Args) -> Args::ReturnType where Args: MessageLoggerNewArgs

C++ method: QMessageLogger::QMessageLogger

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::message_logger::MessageLogger
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::message_logger::MessageLogger>
C++ method: [constructor] void QMessageLogger::QMessageLogger()

Variant 2

Rust arguments:
1) fn new((*const libc::c_char, libc::c_int, *const libc::c_char, cpp_utils::AsStruct)) -> ::message_logger::MessageLogger
2) fn new((*const libc::c_char, libc::c_int, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::message_logger::MessageLogger>
C++ method: [constructor] void QMessageLogger::QMessageLogger(const char* file, int line, const char* function)

Variant 3

Rust arguments:
1) fn new((*const libc::c_char, libc::c_int, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> ::message_logger::MessageLogger
2) fn new((*const libc::c_char, libc::c_int, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::message_logger::MessageLogger>
C++ method: [constructor] void QMessageLogger::QMessageLogger(const char* file, int line, const char* function, const char* category)

fn no_debug<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MessageLoggerNoDebugArgs<'largs>

C++ method: QMessageLogger::noDebug

This is an overloaded function. Available variants:

Rust arguments:
1) fn no_debug(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::no_debug::NoDebug>
2) fn no_debug(&self, cpp_utils::AsStruct) -> ::no_debug::NoDebug
C++ method: QNoDebug QMessageLogger::noDebug() const

fn warning<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MessageLoggerWarningArgs<'largs>

C++ method: QMessageLogger::warning

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn warning(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::debug::Debug>
2) fn warning(&self, cpp_utils::AsStruct) -> ::debug::Debug
C++ method: QDebug QMessageLogger::warning() const

C++ documentation:

Logs a warning message using a QDebug stream

See also qWarning() and QDebug.

Variant 2

Rust arguments:
1) fn warning(&self, (&::logging_category::LoggingCategory, cpp_utils::AsBox)) -> cpp_utils::CppBox<::debug::Debug>
2) fn warning(&self, (&::logging_category::LoggingCategory, cpp_utils::AsStruct)) -> ::debug::Debug
C++ method: QDebug QMessageLogger::warning(const QLoggingCategory& cat) const

C++ documentation:

Logs a warning message into category cat using a QDebug stream.

See also qCWarning() and QDebug.