Struct qt_core::meta_enum::MetaEnum

pub struct MetaEnum {
    // some fields omitted
}

C++ type: QMetaEnum

C++ documentation:

The QMetaEnum class provides meta-data about an enumerator.

Use name() for the enumerator's name. The enumerator's keys (names of each enumerated item) are returned by key(); use keyCount() to find the number of keys. isFlag() returns whether the enumerator is meant to be used as a flag, meaning that its values can be combined using the OR operator.

The conversion functions keyToValue(), valueToKey(), keysToValue(), and valueToKeys() allow conversion between the integer representation of an enumeration or set value and its literal representation. The scope() function returns the class scope this enumerator was declared in.

Methods

impl MetaEnum

fn enclosing_meta_object(&self) -> *const MetaObject

C++ method: const QMetaObject* QMetaEnum::enclosingMetaObject() const

fn is_flag(&self) -> bool

C++ method: bool QMetaEnum::isFlag() const

C++ documentation:

Returns true if this enumerator is used as a flag; otherwise returns false.

When used as flags, enumerators can be combined using the OR operator.

See also keysToValue() and valueToKeys().

fn is_valid(&self) -> bool

C++ method: bool QMetaEnum::isValid() const

C++ documentation:

Returns true if this enum is valid (has a name); otherwise returns false.

See also name().

fn key(&self, index: c_int) -> *const c_char

C++ method: const char* QMetaEnum::key(int index) const

C++ documentation:

Returns the key with the given index, or 0 if no such key exists.

See also keyCount(), value(), and valueToKey().

fn key_count(&self) -> c_int

C++ method: int QMetaEnum::keyCount() const

C++ documentation:

Returns the number of keys.

See also key().

fn key_to_value<'largs, Args>(&'largs self, args: Args) -> c_int where Args: MetaEnumKeyToValueArgs<'largs>

C++ method: QMetaEnum::keyToValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_to_value(&self, *const libc::c_char) -> libc::c_int
2) fn key_to_value(&self, (*const libc::c_char, *mut bool)) -> libc::c_int
C++ method: int QMetaEnum::keyToValue(const char* key, bool* ok = ?) const

C++ documentation:

Returns the integer value of the given enumeration key, or -1 if key is not defined.

If key is not defined, *ok is set to false; otherwise *ok is set to true.

For flag types, use keysToValue().

See also valueToKey(), isFlag(), and keysToValue().

fn keys_to_value<'largs, Args>(&'largs self, args: Args) -> c_int where Args: MetaEnumKeysToValueArgs<'largs>

C++ method: QMetaEnum::keysToValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn keys_to_value(&self, *const libc::c_char) -> libc::c_int
2) fn keys_to_value(&self, (*const libc::c_char, *mut bool)) -> libc::c_int
C++ method: int QMetaEnum::keysToValue(const char* keys, bool* ok = ?) const

C++ documentation:

Returns the value derived from combining together the values of the keys using the OR operator, or -1 if keys is not defined. Note that the strings in keys must be '|'-separated.

If keys is not defined, *ok is set to false; otherwise *ok is set to true.

See also isFlag(), valueToKey(), and valueToKeys().

fn name(&self) -> *const c_char

C++ method: const char* QMetaEnum::name() const

C++ documentation:

Returns the name of the enumerator (without the scope).

For example, the Qt::AlignmentFlag enumeration has AlignmentFlag as the name and Qt as the scope.

See also isValid() and scope().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MetaEnumNewArgs

C++ method: QMetaEnum::QMetaEnum

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::meta_enum::MetaEnum
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::meta_enum::MetaEnum>
C++ method: [constructor] void QMetaEnum::QMetaEnum()

fn scope(&self) -> *const c_char

C++ method: const char* QMetaEnum::scope() const

C++ documentation:

Returns the scope this enumerator was declared in.

For example, the Qt::AlignmentFlag enumeration has Qt as the scope and AlignmentFlag as the name.

See also name().

fn value(&self, index: c_int) -> c_int

C++ method: int QMetaEnum::value(int index) const

C++ documentation:

Returns the value with the given index; or returns -1 if there is no such value.

See also keyCount(), key(), and keyToValue().

fn value_to_key(&self, value: c_int) -> *const c_char

C++ method: const char* QMetaEnum::valueToKey(int value) const

C++ documentation:

Returns the string that is used as the name of the given enumeration value, or 0 if value is not defined.

For flag types, use valueToKeys().

See also isFlag() and valueToKeys().

fn value_to_keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaEnumValueToKeysArgs<'largs>

C++ method: QMetaEnum::valueToKeys

This is an overloaded function. Available variants:

Rust arguments:
1) fn value_to_keys(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn value_to_keys(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray QMetaEnum::valueToKeys(int value) const

C++ documentation:

Returns a byte array of '|'-separated keys that represents the given value.

See also isFlag(), valueToKey(), and keysToValue().