Struct qt_core::meta_method::MetaMethod

pub struct MetaMethod {
    // some fields omitted
}

C++ type: QMetaMethod

C++ documentation:

The QMetaMethod class provides meta-data about a member function.

A QMetaMethod has a methodType(), a methodSignature(), a list of parameterTypes() and parameterNames(), a return typeName(), a tag(), and an access() specifier. You can use invoke() to invoke the method on an arbitrary QObject.

Methods

impl MetaMethod

fn access(&self) -> Access

C++ method: QMetaMethod::Access QMetaMethod::access() const

C++ documentation:

Returns the access specification of this method (private, protected, or public).

Signals are always protected, meaning that you can only emit them from the class or from a subclass.

See also methodType().

fn attributes(&self) -> c_int

C++ method: int QMetaMethod::attributes() const

fn enclosing_meta_object(&self) -> *const MetaObject

C++ method: const QMetaObject* QMetaMethod::enclosingMetaObject() const

fn get_parameter_types(&self, types: *mut c_int)

C++ method: void QMetaMethod::getParameterTypes(int* types) const

fn invoke<'largs, Args>(&'largs self, args: Args) -> bool where Args: MetaMethodInvokeArgs<'largs>

C++ method: QMetaMethod::invoke

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn invoke(&self, *mut ::object::Object) -> bool
2) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument)) -> bool
3) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke(&self, (*mut ::object::Object, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: bool QMetaMethod::invoke(QObject* object, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?) const

C++ documentation:

This function overloads invoke().

This overload invokes this method using the connection type Qt::AutoConnection and ignores return values.

Variant 2

Rust arguments:
1) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument)) -> bool
2) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument)) -> bool
3) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke(&self, (*mut ::object::Object, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: bool QMetaMethod::invoke(QObject* object, QGenericReturnArgument returnValue, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?) const

C++ documentation:

This function overloads invoke().

This overload always invokes this method using the connection type Qt::AutoConnection.

Variant 3

Rust arguments:
1) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType)) -> bool
2) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument)) -> bool
3) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: bool QMetaMethod::invoke(QObject* object, Qt::ConnectionType connectionType, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?) const

C++ documentation:

This function overloads invoke().

This overload can be used if the return value of the member is of no interest.

Variant 4

Rust arguments:
1) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument)) -> bool
2) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument)) -> bool
3) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke(&self, (*mut ::object::Object, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: bool QMetaMethod::invoke(QObject* object, Qt::ConnectionType connectionType, QGenericReturnArgument returnValue, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?) const

C++ documentation:

Invokes this method on the object object. Returns true if the member could be invoked. Returns false if there is no such member or the parameters did not match.

The invocation can be either synchronous or asynchronous, depending on the connectionType:

If connectionType is Qt::DirectConnection, the member will be invoked immediately.

If connectionType is Qt::QueuedConnection, a QEvent will be posted and the member is invoked as soon as the application enters the main event loop.

If connectionType is Qt::AutoConnection, the member is invoked synchronously if object lives in the same thread as the caller; otherwise it will invoke the member asynchronously.

The return value of this method call is placed in returnValue. If the invocation is asynchronous, the return value cannot be evaluated. You can pass up to ten arguments (val0, val1, val2, val3, val4, val5, val6, val7, val8, and val9) to this method call.

QGenericArgument and QGenericReturnArgument are internal helper classes. Because signals and slots can be dynamically invoked, you must enclose the arguments using the Q_ARG() and Q_RETURN_ARG() macros. Q_ARG() takes a type name and a const reference of that type; Q_RETURN_ARG() takes a type name and a non-const reference.

To asynchronously invoke the animateClick() slot on a QPushButton:

int methodIndex = pushButton->metaObject()->indexOfMethod("animateClick()");
QMetaMethod method = metaObject->method(methodIndex);
method.invoke(pushButton, Qt::QueuedConnection);

With asynchronous method invocations, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QMetaMethod::invoke: Unable to handle unregistered datatype 'MyType'

call qRegisterMetaType() to register the data type before you call QMetaMethod::invoke().

To synchronously invoke the compute(QString, int, double) slot on some arbitrary object obj retrieve its return value:

QString retVal;
QByteArray normalizedSignature = QMetaObject::normalizedSignature("compute(QString, int, double)");
int methodIndex = obj->metaObject()->indexOfMethod(normalizedSignature);
QMetaMethod method = obj->metaObject()->method(methodIndex);
method.invoke(obj,
              Qt::DirectConnection,
              Q_RETURN_ARG(QString, retVal),
              Q_ARG(QString, "sqrt"),
              Q_ARG(int, 42),
              Q_ARG(double, 9.7));

QMetaObject::normalizedSignature() is used here to ensure that the format of the signature is what invoke() expects. E.g. extra whitespace is removed.

If the "compute" slot does not take exactly one QString, one int and one double in the specified order, the call will fail.

Warning: this method will not test the validity of the arguments: object must be an instance of the class of the QMetaObject of which this QMetaMethod has been constructed with. The arguments must have the same type as the ones expected by the method, else, the behaviour is undefined.

See also Q_ARG(), Q_RETURN_ARG(), qRegisterMetaType(), and QMetaObject::invokeMethod().

fn invoke_on_gadget<'largs, Args>(&'largs self, args: Args) -> bool where Args: MetaMethodInvokeOnGadgetArgs<'largs>

C++ method: QMetaMethod::invokeOnGadget

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn invoke_on_gadget(&self, *mut libc::c_void) -> bool
2) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument)) -> bool
3) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: bool QMetaMethod::invokeOnGadget(void* gadget, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?) const

C++ documentation:

This is an overloaded function.

This overload invokes this method for a gadget and ignores return values.

This function was introduced in Qt 5.5.

Variant 2

Rust arguments:
1) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument)) -> bool
2) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument)) -> bool
3) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke_on_gadget(&self, (*mut libc::c_void, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: bool QMetaMethod::invokeOnGadget(void* gadget, QGenericReturnArgument returnValue, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?) const

C++ documentation:

Invokes this method on a Q_GADGET. Returns true if the member could be invoked. Returns false if there is no such member or the parameters did not match.

The pointer gadget must point to an instance of the gadget class.

The invocation is always synchronous.

The return value of this method call is placed in returnValue. You can pass up to ten arguments (val0, val1, val2, val3, val4, val5, val6, val7, val8, and val9) to this method call.

Warning: this method will not test the validity of the arguments: gadget must be an instance of the class of the QMetaObject of which this QMetaMethod has been constructed with. The arguments must have the same type as the ones expected by the method, else, the behavior is undefined.

This function was introduced in Qt 5.5.

See also Q_ARG(), Q_RETURN_ARG(), qRegisterMetaType(), and QMetaObject::invokeMethod().

fn is_valid(&self) -> bool

C++ method: bool QMetaMethod::isValid() const

C++ documentation:

Returns true if this method is valid (can be introspected and invoked), otherwise returns false.

This function was introduced in Qt 5.0.

fn method_index(&self) -> c_int

C++ method: int QMetaMethod::methodIndex() const

C++ documentation:

Returns this method's index.

This function was introduced in Qt 4.6.

fn method_signature<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaMethodMethodSignatureArgs<'largs>

C++ method: QMetaMethod::methodSignature

This is an overloaded function. Available variants:

Rust arguments:
1) fn method_signature(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn method_signature(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QMetaMethod::methodSignature() const

C++ documentation:

Returns the signature of this method (e.g., setValue(double)).

This function was introduced in Qt 5.0.

See also parameterTypes() and parameterNames().

fn method_type(&self) -> MethodType

C++ method: QMetaMethod::MethodType QMetaMethod::methodType() const

C++ documentation:

Returns the type of this method (signal, slot, or method).

See also access().

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaMethodNameArgs<'largs>

C++ method: QMetaMethod::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn name(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QMetaMethod::name() const

C++ documentation:

Returns the name of this method.

This function was introduced in Qt 5.0.

See also methodSignature() and parameterCount().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MetaMethodNewArgs

C++ method: QMetaMethod::QMetaMethod

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::meta_method::MetaMethod
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::meta_method::MetaMethod>
C++ method: [constructor] void QMetaMethod::QMetaMethod()

fn parameter_count(&self) -> c_int

C++ method: int QMetaMethod::parameterCount() const

C++ documentation:

Returns the number of parameters of this method.

This function was introduced in Qt 5.0.

See also parameterType() and parameterNames().

fn parameter_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaMethodParameterNamesArgs<'largs>

C++ method: QMetaMethod::parameterNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn parameter_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn parameter_names(&self, cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: QList<QByteArray> QMetaMethod::parameterNames() const

C++ documentation:

Returns a list of parameter names.

See also parameterTypes() and methodSignature().

fn parameter_type(&self, index: c_int) -> c_int

C++ method: int QMetaMethod::parameterType(int index) const

C++ documentation:

Returns the type of the parameter at the given index.

The return value is one of the types that are registered with QMetaType, or QMetaType::UnknownType if the type is not registered.

This function was introduced in Qt 5.0.

See also parameterCount(), returnType(), and QMetaType.

fn parameter_types<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaMethodParameterTypesArgs<'largs>

C++ method: QMetaMethod::parameterTypes

This is an overloaded function. Available variants:

Rust arguments:
1) fn parameter_types(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn parameter_types(&self, cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: QList<QByteArray> QMetaMethod::parameterTypes() const

C++ documentation:

Returns a list of parameter types.

See also parameterNames() and methodSignature().

fn return_type(&self) -> c_int

C++ method: int QMetaMethod::returnType() const

C++ documentation:

Returns the return type of this method.

The return value is one of the types that are registered with QMetaType, or QMetaType::UnknownType if the type is not registered.

This function was introduced in Qt 5.0.

See also parameterType(), QMetaType, and typeName().

fn revision(&self) -> c_int

C++ method: int QMetaMethod::revision() const

C++ documentation:

Returns the method revision if one was specified by Q_REVISION, otherwise returns 0.

This function was introduced in Qt 5.1.

fn tag(&self) -> *const c_char

C++ method: const char* QMetaMethod::tag() const

C++ documentation:

Returns the tag associated with this method.

Tags are special macros recognized by moc that make it possible to add extra information about a method.

Tag information can be added in the following way in the function declaration:

#ifndef Q_MOC_RUN
// define the tag text
#  define THISISTESTTAG
#endif
...
private slots:
    THISISTESTTAG void testFunc();

and the information can be accessed by using:

MainWindow win;
win.show();

int functionIndex = win.metaObject()->indexOfSlot("testFunc()"); QMetaMethod mm = metaObject()->method(functionIndex); qDebug() << mm.tag(); // prints THISISTESTTAG
For the moment, moc will extract and record all tags, but it will not handle any of them specially.

Note: Since Qt 5.0, moc expands preprocessor macros, so it is necessary to surround the definition with #ifndef Q_MOC_RUN, as shown in the example above. This was not required in Qt 4. The code as shown above works with Qt 4 too.

fn type_name(&self) -> *const c_char

C++ method: const char* QMetaMethod::typeName() const

C++ documentation:

Returns the return type name of this method.

See also returnType() and QMetaType::type().