Module qt_core::meta_object::overloading

Traits

ConnectionNewArgs
This trait represents a set of arguments accepted by Connection::new method.

MetaObjectActivateArgs
This trait represents a set of arguments accepted by MetaObject::activate method.

MetaObjectCastArgs
This trait represents a set of arguments accepted by MetaObject::cast method.

MetaObjectCheckConnectArgsArgs
This trait represents a set of arguments accepted by MetaObject::check_connect_args method.

MetaObjectClassInfoArgs
This trait represents a set of arguments accepted by MetaObject::class_info method.

MetaObjectConnectArgs
This trait represents a set of arguments accepted by MetaObject::connect method.

MetaObjectConstructorArgs
This trait represents a set of arguments accepted by MetaObject::constructor method.

MetaObjectEnumeratorArgs
This trait represents a set of arguments accepted by MetaObject::enumerator method.

MetaObjectInvokeMethodArgs
This trait represents a set of arguments accepted by MetaObject::invoke_method method.

MetaObjectMethodArgs
This trait represents a set of arguments accepted by MetaObject::method method.

MetaObjectNewInstanceArgs
This trait represents a set of arguments accepted by MetaObject::new_instance method.

MetaObjectNormalizedSignatureArgs
This trait represents a set of arguments accepted by MetaObject::normalized_signature method.

MetaObjectNormalizedTypeArgs
This trait represents a set of arguments accepted by MetaObject::normalized_type method.

MetaObjectPropertyArgs
This trait represents a set of arguments accepted by MetaObject::property method.

MetaObjectTrArgs
This trait represents a set of arguments accepted by MetaObject::tr method.

MetaObjectUserPropertyArgs
This trait represents a set of arguments accepted by MetaObject::user_property method.