Struct qt_core::meta_object::MetaObject

pub struct MetaObject {
    // some fields omitted
}

C++ type: QMetaObject

C++ documentation:

The QMetaObject class contains meta-information about Qt objects.

The Qt Meta-Object System in Qt is responsible for the signals and slots inter-object communication mechanism, runtime type information, and the Qt property system. A single QMetaObject instance is created for each QObject subclass that is used in an application, and this instance stores all the meta-information for the QObject subclass. This object is available as QObject::metaObject().

This class is not normally required for application programming, but it is useful if you write meta-applications, such as scripting engines or GUI builders.

The functions you are most likely to find useful are these:

className() returns the name of a class.

superClass() returns the superclass's meta-object.

method() and methodCount() provide information about a class's meta-methods (signals, slots and other invokable member functions).

enumerator() and enumeratorCount() and provide information about a class's enumerators.

propertyCount() and property() provide information about a class's properties.

constructor() and constructorCount() provide information about a class's meta-constructors.

The index functions indexOfConstructor(), indexOfMethod(), indexOfEnumerator(), and indexOfProperty() map names of constructors, member functions, enumerators, or properties to indexes in the meta-object. For example, Qt uses indexOfMethod() internally when you connect a signal to a slot.

Classes can also have a list of name--value pairs of additional class information, stored in QMetaClassInfo objects. The number of pairs is returned by classInfoCount(), single pairs are returned by classInfo(), and you can search for pairs with indexOfClassInfo().

Methods

impl MetaObject

fn activate<Args>(args: Args) where Args: MetaObjectActivateArgs

C++ method: QMetaObject::activate

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn activate((*mut ::object::Object, *const ::meta_object::MetaObject, libc::c_int, *mut *mut libc::c_void)) -> ()
C++ method: static void QMetaObject::activate(QObject* sender, const QMetaObject* arg2, int local_signal_index, void** argv)

Variant 2

Rust arguments: fn activate((*mut ::object::Object, libc::c_int, *mut *mut libc::c_void)) -> ()
C++ method: static void QMetaObject::activate(QObject* sender, int signal_index, void** argv)

Variant 3

Rust arguments: fn activate((*mut ::object::Object, libc::c_int, libc::c_int, *mut *mut libc::c_void)) -> ()
C++ method: static void QMetaObject::activate(QObject* sender, int signal_offset, int local_signal_index, void** argv)

fn cast<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectCastArgs<'largs>

C++ method: QMetaObject::cast

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn cast(&self, *mut ::object::Object) -> *mut ::object::Object
C++ method: QObject* QMetaObject::cast(QObject* obj) const

Variant 2

Rust arguments: fn cast(&self, *const ::object::Object) -> *const ::object::Object
C++ method: const QObject* QMetaObject::cast(const QObject* obj) const

fn check_connect_args<Args>(args: Args) -> bool where Args: MetaObjectCheckConnectArgsArgs

C++ method: QMetaObject::checkConnectArgs

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn check_connect_args((&::meta_method::MetaMethod, &::meta_method::MetaMethod)) -> bool
C++ method: static bool QMetaObject::checkConnectArgs(const QMetaMethod& signal, const QMetaMethod& method)

C++ documentation:

This is an overloaded function.

Returns true if the signal and method arguments are compatible; otherwise returns false.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn check_connect_args((*const libc::c_char, *const libc::c_char)) -> bool
C++ method: static bool QMetaObject::checkConnectArgs(const char* signal, const char* method)

C++ documentation:

Returns true if the signal and method arguments are compatible; otherwise returns false.

Both signal and method are expected to be normalized.

See also normalizedSignature().

fn class_info<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectClassInfoArgs<'largs>

C++ method: QMetaObject::classInfo

This is an overloaded function. Available variants:

Rust arguments:
1) fn class_info(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_class_info::MetaClassInfo>
2) fn class_info(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::meta_class_info::MetaClassInfo
C++ method: QMetaClassInfo QMetaObject::classInfo(int index) const

C++ documentation:

Returns the meta-data for the item of class information with the given index.

Example:

class MyClass : public QObject
{
    Q_OBJECT
    Q_CLASSINFO("author", "Sabrina Schweinsteiger")
    Q_CLASSINFO("url", "http://doc.moosesoft.co.uk/1.0/")

public: ... };
See also classInfoCount(), classInfoOffset(), and indexOfClassInfo().

fn class_info_count(&self) -> c_int

C++ method: int QMetaObject::classInfoCount() const

C++ documentation:

Returns the number of items of class information in this class.

See also classInfo(), classInfoOffset(), and indexOfClassInfo().

fn class_info_offset(&self) -> c_int

C++ method: int QMetaObject::classInfoOffset() const

C++ documentation:

Returns the class information offset for this class; i.e. the index position of this class's first class information item.

If the class has no superclasses with class information, the offset is 0; otherwise the offset is the sum of all the class information items in the class's superclasses.

See also classInfo(), classInfoCount(), and indexOfClassInfo().

fn class_name(&self) -> *const c_char

C++ method: const char* QMetaObject::className() const

C++ documentation:

Returns the class name.

See also superClass().

fn connect<Args>(args: Args) -> Args::ReturnType where Args: MetaObjectConnectArgs

C++ method: QMetaObject::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect((*const ::object::Object, libc::c_int, *const ::object::Object, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect((*const ::object::Object, libc::c_int, *const ::object::Object, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect((*const ::object::Object, libc::c_int, *const ::object::Object, libc::c_int, libc::c_int, *mut libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
4) fn connect((*const ::object::Object, libc::c_int, *const ::object::Object, libc::c_int, cpp_utils::AsStruct)) -> ::meta_object::Connection
5) fn connect((*const ::object::Object, libc::c_int, *const ::object::Object, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::meta_object::Connection
6) fn connect((*const ::object::Object, libc::c_int, *const ::object::Object, libc::c_int, libc::c_int, *mut libc::c_int, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QMetaObject::connect(const QObject* sender, int signal_index, const QObject* receiver, int method_index, int type = ?, int* types = ?)

fn connect_slots_by_name(o: *mut Object)

C++ method: static void QMetaObject::connectSlotsByName(QObject* o)

C++ documentation:

Searches recursively for all child objects of the given object, and connects matching signals from them to slots of object that follow the following form:

void on_<object name>_<signal name>(<signal parameters>);

Let's assume our object has a child object of type QPushButton with the object name button1. The slot to catch the button's clicked() signal would be:

void on_button1_clicked();

If object itself has a properly set object name, its own signals are also connected to its respective slots.

See also QObject::setObjectName().

fn constructor<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectConstructorArgs<'largs>

C++ method: QMetaObject::constructor

This is an overloaded function. Available variants:

Rust arguments:
1) fn constructor(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_method::MetaMethod>
2) fn constructor(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::meta_method::MetaMethod
C++ method: QMetaMethod QMetaObject::constructor(int index) const

C++ documentation:

Returns the meta-data for the constructor with the given index.

This function was introduced in Qt 4.5.

See also constructorCount() and newInstance().

fn constructor_count(&self) -> c_int

C++ method: int QMetaObject::constructorCount() const

C++ documentation:

Returns the number of constructors in this class.

This function was introduced in Qt 4.5.

See also constructor() and indexOfConstructor().

fn disconnect(sender: *const Object, signal_index: c_int, receiver: *const Object, method_index: c_int) -> bool

C++ method: static bool QMetaObject::disconnect(const QObject* sender, int signal_index, const QObject* receiver, int method_index)

fn disconnect_one(sender: *const Object, signal_index: c_int, receiver: *const Object, method_index: c_int) -> bool

C++ method: static bool QMetaObject::disconnectOne(const QObject* sender, int signal_index, const QObject* receiver, int method_index)

fn enumerator<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectEnumeratorArgs<'largs>

C++ method: QMetaObject::enumerator

This is an overloaded function. Available variants:

Rust arguments:
1) fn enumerator(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_enum::MetaEnum>
2) fn enumerator(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::meta_enum::MetaEnum
C++ method: QMetaEnum QMetaObject::enumerator(int index) const

C++ documentation:

Returns the meta-data for the enumerator with the given index.

See also enumeratorCount(), enumeratorOffset(), and indexOfEnumerator().

fn enumerator_count(&self) -> c_int

C++ method: int QMetaObject::enumeratorCount() const

C++ documentation:

Returns the number of enumerators in this class.

See also enumerator(), enumeratorOffset(), and indexOfEnumerator().

fn enumerator_offset(&self) -> c_int

C++ method: int QMetaObject::enumeratorOffset() const

C++ documentation:

Returns the enumerator offset for this class; i.e. the index position of this class's first enumerator.

If the class has no superclasses with enumerators, the offset is 0; otherwise the offset is the sum of all the enumerators in the class's superclasses.

See also enumerator(), enumeratorCount(), and indexOfEnumerator().

fn index_of_class_info(&self, name: *const c_char) -> c_int

C++ method: int QMetaObject::indexOfClassInfo(const char* name) const

C++ documentation:

Finds class information item name and returns its index; otherwise returns -1.

See also classInfo(), classInfoCount(), and classInfoOffset().

fn index_of_constructor(&self, constructor: *const c_char) -> c_int

C++ method: int QMetaObject::indexOfConstructor(const char* constructor) const

C++ documentation:

Finds constructor and returns its index; otherwise returns -1.

Note that the constructor has to be in normalized form, as returned by normalizedSignature().

This function was introduced in Qt 4.5.

See also constructor(), constructorCount(), and normalizedSignature().

fn index_of_enumerator(&self, name: *const c_char) -> c_int

C++ method: int QMetaObject::indexOfEnumerator(const char* name) const

C++ documentation:

Finds enumerator name and returns its index; otherwise returns -1.

See also enumerator(), enumeratorCount(), and enumeratorOffset().

fn index_of_method(&self, method: *const c_char) -> c_int

C++ method: int QMetaObject::indexOfMethod(const char* method) const

C++ documentation:

Finds method and returns its index; otherwise returns -1.

Note that the method has to be in normalized form, as returned by normalizedSignature().

See also method(), methodCount(), methodOffset(), and normalizedSignature().

fn index_of_property(&self, name: *const c_char) -> c_int

C++ method: int QMetaObject::indexOfProperty(const char* name) const

C++ documentation:

Finds property name and returns its index; otherwise returns -1.

See also property(), propertyCount(), and propertyOffset().

fn index_of_signal(&self, signal: *const c_char) -> c_int

C++ method: int QMetaObject::indexOfSignal(const char* signal) const

C++ documentation:

Finds signal and returns its index; otherwise returns -1.

This is the same as indexOfMethod(), except that it will return -1 if the method exists but isn't a signal.

Note that the signal has to be in normalized form, as returned by normalizedSignature().

See also indexOfMethod(), normalizedSignature(), method(), methodCount(), and methodOffset().

fn index_of_slot(&self, slot: *const c_char) -> c_int

C++ method: int QMetaObject::indexOfSlot(const char* slot) const

C++ documentation:

Finds slot and returns its index; otherwise returns -1.

This is the same as indexOfMethod(), except that it will return -1 if the method exists but isn't a slot.

See also indexOfMethod(), method(), methodCount(), and methodOffset().

fn invoke_method<Args>(args: Args) -> bool where Args: MetaObjectInvokeMethodArgs

C++ method: QMetaObject::invokeMethod

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn invoke_method((*mut ::object::Object, *const libc::c_char)) -> bool
2) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument)) -> bool
3) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: static bool QMetaObject::invokeMethod(QObject* obj, const char* member, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?)

C++ documentation:

This function overloads invokeMethod().

This overload invokes the member using the connection type Qt::AutoConnection and ignores return values.

Variant 2

Rust arguments:
1) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument)) -> bool
2) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument)) -> bool
3) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke_method((*mut ::object::Object, *const libc::c_char, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: static bool QMetaObject::invokeMethod(QObject* obj, const char* member, QGenericReturnArgument ret, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?)

C++ documentation:

This function overloads invokeMethod().

This overload always invokes the member using the connection type Qt::AutoConnection.

Variant 3

Rust arguments:
1) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument)) -> bool
2) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument)) -> bool
3) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_return_argument::GenericReturnArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: static bool QMetaObject::invokeMethod(QObject* obj, const char* member, Qt::ConnectionType arg3, QGenericReturnArgument ret, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?)

C++ documentation:

Invokes the member (a signal or a slot name) on the object obj. Returns true if the member could be invoked. Returns false if there is no such member or the parameters did not match.

The invocation can be either synchronous or asynchronous, depending on type:

If type is Qt::DirectConnection, the member will be invoked immediately.

If type is Qt::QueuedConnection, a QEvent will be sent and the member is invoked as soon as the application enters the main event loop.

If type is Qt::BlockingQueuedConnection, the method will be invoked in the same way as for Qt::QueuedConnection, except that the current thread will block until the event is delivered. Using this connection type to communicate between objects in the same thread will lead to deadlocks.

If type is Qt::AutoConnection, the member is invoked synchronously if obj lives in the same thread as the caller; otherwise it will invoke the member asynchronously.

The return value of the member function call is placed in ret. If the invocation is asynchronous, the return value cannot be evaluated. You can pass up to ten arguments (val0, val1, val2, val3, val4, val5, val6, val7, val8, and val9) to the member function.

QGenericArgument and QGenericReturnArgument are internal helper classes. Because signals and slots can be dynamically invoked, you must enclose the arguments using the Q_ARG() and Q_RETURN_ARG() macros. Q_ARG() takes a type name and a const reference of that type; Q_RETURN_ARG() takes a type name and a non-const reference.

You only need to pass the name of the signal or slot to this function, not the entire signature. For example, to asynchronously invoke the quit() slot on a QThread, use the following code:

QMetaObject::invokeMethod(thread, "quit",
                          Qt::QueuedConnection);

With asynchronous method invocations, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QMetaObject::invokeMethod: Unable to handle unregistered datatype 'MyType'

call qRegisterMetaType() to register the data type before you call invokeMethod().

To synchronously invoke the compute(QString, int, double) slot on some arbitrary object obj retrieve its return value:

QString retVal;
QMetaObject::invokeMethod(obj, "compute", Qt::DirectConnection,
                          Q_RETURN_ARG(QString, retVal),
                          Q_ARG(QString, "sqrt"),
                          Q_ARG(int, 42),
                          Q_ARG(double, 9.7));

If the "compute" slot does not take exactly one QString, one int and one double in the specified order, the call will fail.

See also Q_ARG(), Q_RETURN_ARG(), qRegisterMetaType(), and QMetaMethod::invoke().

Variant 4

Rust arguments:
1) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType)) -> bool
2) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument)) -> bool
3) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
4) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
5) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
6) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
7) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
8) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
9) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
10) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
11) fn invoke_method((*mut ::object::Object, *const libc::c_char, ::qt::ConnectionType, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> bool
C++ method: static bool QMetaObject::invokeMethod(QObject* obj, const char* member, Qt::ConnectionType type, QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?)

C++ documentation:

This function overloads invokeMethod().

This overload can be used if the return value of the member is of no interest.

fn metacall(arg1: *mut Object, arg2: Call, arg3: c_int, arg4: *mut *mut c_void) -> c_int

C++ method: static int QMetaObject::metacall(QObject* arg1, QMetaObject::Call arg2, int arg3, void** arg4)

fn method<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectMethodArgs<'largs>

C++ method: QMetaObject::method

This is an overloaded function. Available variants:

Rust arguments:
1) fn method(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_method::MetaMethod>
2) fn method(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::meta_method::MetaMethod
C++ method: QMetaMethod QMetaObject::method(int index) const

C++ documentation:

Returns the meta-data for the method with the given index.

See also methodCount(), methodOffset(), and indexOfMethod().

fn method_count(&self) -> c_int

C++ method: int QMetaObject::methodCount() const

C++ documentation:

Returns the number of methods in this class, including the number of methods provided by each base class. These include signals and slots as well as normal member functions.

Use code like the following to obtain a QStringList containing the methods specific to a given class:

const QMetaObject* metaObject = obj->metaObject();
QStringList methods;
for(int i = metaObject->methodOffset(); i < metaObject->methodCount(); ++i)
    methods << QString::fromLatin1(metaObject->method(i).methodSignature());

See also method(), methodOffset(), and indexOfMethod().

fn method_offset(&self) -> c_int

C++ method: int QMetaObject::methodOffset() const

C++ documentation:

Returns the method offset for this class; i.e. the index position of this class's first member function.

The offset is the sum of all the methods in the class's superclasses (which is always positive since QObject has the deleteLater() slot and a destroyed() signal).

See also method(), methodCount(), and indexOfMethod().

fn new_instance<'largs, Args>(&'largs self, args: Args) -> *mut Object where Args: MetaObjectNewInstanceArgs<'largs>

C++ method: QMetaObject::newInstance

This is an overloaded function. Available variants:

Rust arguments:
1) fn new_instance(&self, ()) -> *mut ::object::Object
2) fn new_instance(&self, &::generic_argument::GenericArgument) -> *mut ::object::Object
3) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
4) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
5) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
6) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
7) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
8) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
9) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
10) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
11) fn new_instance(&self, (&::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument, &::generic_argument::GenericArgument)) -> *mut ::object::Object
C++ method: QObject* QMetaObject::newInstance(QGenericArgument val0 = ?, QGenericArgument val1 = ?, QGenericArgument val2 = ?, QGenericArgument val3 = ?, QGenericArgument val4 = ?, QGenericArgument val5 = ?, QGenericArgument val6 = ?, QGenericArgument val7 = ?, QGenericArgument val8 = ?, QGenericArgument val9 = ?) const

C++ documentation:

Constructs a new instance of this class. You can pass up to ten arguments (val0, val1, val2, val3, val4, val5, val6, val7, val8, and val9) to the constructor. Returns the new object, or 0 if no suitable constructor is available.

Note that only constructors that are declared with the Q_INVOKABLE modifier are made available through the meta-object system.

This function was introduced in Qt 4.5.

See also Q_ARG() and constructor().

fn normalized_signature<Args>(args: Args) -> Args::ReturnType where Args: MetaObjectNormalizedSignatureArgs

C++ method: QMetaObject::normalizedSignature

This is an overloaded function. Available variants:

Rust arguments:
1) fn normalized_signature((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn normalized_signature((*const libc::c_char, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: static QByteArray QMetaObject::normalizedSignature(const char* method)

C++ documentation:

Normalizes the signature of the given method.

Qt uses normalized signatures to decide whether two given signals and slots are compatible. Normalization reduces whitespace to a minimum, moves 'const' to the front where appropriate, removes 'const' from value types and replaces const references with values.

See also checkConnectArgs() and normalizedType().

fn normalized_type<Args>(args: Args) -> Args::ReturnType where Args: MetaObjectNormalizedTypeArgs

C++ method: QMetaObject::normalizedType

This is an overloaded function. Available variants:

Rust arguments:
1) fn normalized_type((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn normalized_type((*const libc::c_char, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: static QByteArray QMetaObject::normalizedType(const char* type)

C++ documentation:

Normalizes a type.

See QMetaObject::normalizedSignature() for a description on how Qt normalizes.

Example:

QByteArray normType = QMetaObject::normalizedType(" int    const  ");
// normType is now "const int"

This function was introduced in Qt 4.2.

See also normalizedSignature().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectPropertyArgs<'largs>

C++ method: QMetaObject::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_property::MetaProperty>
2) fn property(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::meta_property::MetaProperty
C++ method: QMetaProperty QMetaObject::property(int index) const

C++ documentation:

Returns the meta-data for the property with the given index. If no such property exists, a null QMetaProperty is returned.

See also propertyCount(), propertyOffset(), and indexOfProperty().

fn property_count(&self) -> c_int

C++ method: int QMetaObject::propertyCount() const

C++ documentation:

Returns the number of properties in this class, including the number of properties provided by each base class.

Use code like the following to obtain a QStringList containing the properties specific to a given class:

const QMetaObject* metaObject = obj->metaObject();
QStringList properties;
for(int i = metaObject->propertyOffset(); i < metaObject->propertyCount(); ++i)
    properties << QString::fromLatin1(metaObject->property(i).name());

See also property(), propertyOffset(), and indexOfProperty().

fn property_offset(&self) -> c_int

C++ method: int QMetaObject::propertyOffset() const

C++ documentation:

Returns the property offset for this class; i.e. the index position of this class's first property.

The offset is the sum of all the properties in the class's superclasses (which is always positive since QObject has the name() property).

See also property(), propertyCount(), and indexOfProperty().

fn static_metacall(&self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: int QMetaObject::static_metacall(QMetaObject::Call arg1, int arg2, void** arg3) const

fn super_class(&self) -> *const MetaObject

C++ method: const QMetaObject* QMetaObject::superClass() const

C++ documentation:

Returns the meta-object of the superclass, or 0 if there is no such object.

See also className().

fn tr<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectTrArgs<'largs>

C++ method: QMetaObject::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr(&self, (*const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr(&self, (*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn tr(&self, (*const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> ::string::String
4) fn tr(&self, (*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QMetaObject::tr(const char* s, const char* c, int n = ?) const

fn user_property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaObjectUserPropertyArgs<'largs>

C++ method: QMetaObject::userProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn user_property(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::meta_property::MetaProperty>
2) fn user_property(&self, cpp_utils::AsStruct) -> ::meta_property::MetaProperty
C++ method: QMetaProperty QMetaObject::userProperty() const

C++ documentation:

Returns the property that has the USER flag set to true.

This function was introduced in Qt 4.2.

See also QMetaProperty::isUser().