Struct qt_core::meta_property::MetaProperty

pub struct MetaProperty {
    // some fields omitted
}

C++ type: QMetaProperty

C++ documentation:

The QMetaProperty class provides meta-data about a property.

Property meta-data is obtained from an object's meta-object. See QMetaObject::property() and QMetaObject::propertyCount() for details.

Methods

impl MetaProperty

fn enclosing_meta_object(&self) -> *const MetaObject

C++ method: const QMetaObject* QMetaProperty::enclosingMetaObject() const

fn enumerator<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaPropertyEnumeratorArgs<'largs>

C++ method: QMetaProperty::enumerator

This is an overloaded function. Available variants:

Rust arguments:
1) fn enumerator(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::meta_enum::MetaEnum>
2) fn enumerator(&self, cpp_utils::AsStruct) -> ::meta_enum::MetaEnum
C++ method: QMetaEnum QMetaProperty::enumerator() const

C++ documentation:

Returns the enumerator if this property's type is an enumerator type; otherwise the returned value is undefined.

See also isEnumType() and isFlagType().

fn has_notify_signal(&self) -> bool

C++ method: bool QMetaProperty::hasNotifySignal() const

C++ documentation:

Returns true if this property has a corresponding change notify signal; otherwise returns false.

See also notifySignal().

fn has_std_cpp_set(&self) -> bool

C++ method: bool QMetaProperty::hasStdCppSet() const

fn is_constant(&self) -> bool

C++ method: bool QMetaProperty::isConstant() const

C++ documentation:

Returns true if the property is constant; otherwise returns false.

A property is constant if the Q_PROPERTY()'s CONSTANT attribute is set.

This function was introduced in Qt 4.6.

fn is_designable<'largs, Args>(&'largs self, args: Args) -> bool where Args: MetaPropertyIsDesignableArgs<'largs>

C++ method: QMetaProperty::isDesignable

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_designable(&self, ()) -> bool
2) fn is_designable(&self, *const ::object::Object) -> bool
C++ method: bool QMetaProperty::isDesignable(const QObject* obj = ?) const

C++ documentation:

Returns true if this property is designable for the given object; otherwise returns false.

If no object is given, the function returns false if the Q_PROPERTY()'s DESIGNABLE attribute is false; otherwise returns true (if the attribute is true or is a function or expression).

See also isScriptable() and isStored().

fn is_editable<'largs, Args>(&'largs self, args: Args) -> bool where Args: MetaPropertyIsEditableArgs<'largs>

C++ method: QMetaProperty::isEditable

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_editable(&self, ()) -> bool
2) fn is_editable(&self, *const ::object::Object) -> bool
C++ method: bool QMetaProperty::isEditable(const QObject* obj = ?) const

C++ documentation:

Returns true if the property is editable for the given object; otherwise returns false.

If no object is given, the function returns false if the Q_PROPERTY()'s EDITABLE attribute is false; otherwise returns true (if the attribute is true or is a function or expression).

See also isDesignable(), isScriptable(), and isStored().

fn is_enum_type(&self) -> bool

C++ method: bool QMetaProperty::isEnumType() const

C++ documentation:

Returns true if the property's type is an enumeration value; otherwise returns false.

See also enumerator() and isFlagType().

fn is_final(&self) -> bool

C++ method: bool QMetaProperty::isFinal() const

C++ documentation:

Returns true if the property is final; otherwise returns false.

A property is final if the Q_PROPERTY()'s FINAL attribute is set.

This function was introduced in Qt 4.6.

fn is_flag_type(&self) -> bool

C++ method: bool QMetaProperty::isFlagType() const

C++ documentation:

Returns true if the property's type is an enumeration value that is used as a flag; otherwise returns false.

Flags can be combined using the OR operator. A flag type is implicitly also an enum type.

See also isEnumType(), enumerator(), and QMetaEnum::isFlag().

fn is_readable(&self) -> bool

C++ method: bool QMetaProperty::isReadable() const

C++ documentation:

Returns true if this property is readable; otherwise returns false.

See also isWritable(), read(), and isValid().

fn is_resettable(&self) -> bool

C++ method: bool QMetaProperty::isResettable() const

C++ documentation:

Returns true if this property can be reset to a default value; otherwise returns false.

See also reset().

fn is_scriptable<'largs, Args>(&'largs self, args: Args) -> bool where Args: MetaPropertyIsScriptableArgs<'largs>

C++ method: QMetaProperty::isScriptable

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_scriptable(&self, ()) -> bool
2) fn is_scriptable(&self, *const ::object::Object) -> bool
C++ method: bool QMetaProperty::isScriptable(const QObject* obj = ?) const

C++ documentation:

Returns true if the property is scriptable for the given object; otherwise returns false.

If no object is given, the function returns false if the Q_PROPERTY()'s SCRIPTABLE attribute is false; otherwise returns true (if the attribute is true or is a function or expression).

See also isDesignable() and isStored().

fn is_stored<'largs, Args>(&'largs self, args: Args) -> bool where Args: MetaPropertyIsStoredArgs<'largs>

C++ method: QMetaProperty::isStored

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_stored(&self, ()) -> bool
2) fn is_stored(&self, *const ::object::Object) -> bool
C++ method: bool QMetaProperty::isStored(const QObject* obj = ?) const

C++ documentation:

Returns true if the property is stored for object; otherwise returns false.

If no object is given, the function returns false if the Q_PROPERTY()'s STORED attribute is false; otherwise returns true (if the attribute is true or is a function or expression).

See also isDesignable() and isScriptable().

fn is_user<'largs, Args>(&'largs self, args: Args) -> bool where Args: MetaPropertyIsUserArgs<'largs>

C++ method: QMetaProperty::isUser

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_user(&self, ()) -> bool
2) fn is_user(&self, *const ::object::Object) -> bool
C++ method: bool QMetaProperty::isUser(const QObject* obj = ?) const

C++ documentation:

Returns true if this property is designated as the USER property, i.e., the one that the user can edit for object or that is significant in some other way. Otherwise it returns false. e.g., the text property is the USER editable property of a QLineEdit.

If object is null, the function returns false if the Q_PROPERTY()'s USER attribute is false. Otherwise it returns true.

See also QMetaObject::userProperty(), isDesignable(), and isScriptable().

fn is_valid(&self) -> bool

C++ method: bool QMetaProperty::isValid() const

C++ documentation:

Returns true if this property is valid (readable); otherwise returns false.

See also isReadable().

fn is_writable(&self) -> bool

C++ method: bool QMetaProperty::isWritable() const

C++ documentation:

Returns true if this property is writable; otherwise returns false.

See also isReadable() and write().

fn name(&self) -> *const c_char

C++ method: const char* QMetaProperty::name() const

C++ documentation:

Returns this property's name.

See also type() and typeName().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MetaPropertyNewArgs

C++ method: QMetaProperty::QMetaProperty

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::meta_property::MetaProperty
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::meta_property::MetaProperty>
C++ method: [constructor] void QMetaProperty::QMetaProperty()

fn notify_signal<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaPropertyNotifySignalArgs<'largs>

C++ method: QMetaProperty::notifySignal

This is an overloaded function. Available variants:

Rust arguments:
1) fn notify_signal(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::meta_method::MetaMethod>
2) fn notify_signal(&self, cpp_utils::AsStruct) -> ::meta_method::MetaMethod
C++ method: QMetaMethod QMetaProperty::notifySignal() const

C++ documentation:

Returns the QMetaMethod instance of the property change notifying signal if one was specified, otherwise returns an invalid QMetaMethod.

This function was introduced in Qt 4.5.

See also hasNotifySignal().

fn notify_signal_index(&self) -> c_int

C++ method: int QMetaProperty::notifySignalIndex() const

C++ documentation:

Returns the index of the property change notifying signal if one was specified, otherwise returns -1.

This function was introduced in Qt 4.6.

See also hasNotifySignal().

fn property_index(&self) -> c_int

C++ method: int QMetaProperty::propertyIndex() const

C++ documentation:

Returns this property's index.

This function was introduced in Qt 4.6.

fn read<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaPropertyReadArgs<'largs>

C++ method: QMetaProperty::read

This is an overloaded function. Available variants:

Rust arguments:
1) fn read(&self, (*const ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn read(&self, (*const ::object::Object, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QMetaProperty::read(const QObject* obj) const

C++ documentation:

Reads the property's value from the given object. Returns the value if it was able to read it; otherwise returns an invalid variant.

See also write(), reset(), and isReadable().

fn read_on_gadget<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MetaPropertyReadOnGadgetArgs<'largs>

C++ method: QMetaProperty::readOnGadget

This is an overloaded function. Available variants:

Rust arguments:
1) fn read_on_gadget(&self, (*const libc::c_void, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn read_on_gadget(&self, (*const libc::c_void, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QMetaProperty::readOnGadget(const void* gadget) const

C++ documentation:

Reads the property's value from the given gadget. Returns the value if it was able to read it; otherwise returns an invalid variant.

This function should only be used if this is a property of a Q_GADGET

This function was introduced in Qt 5.5.

fn reset(&self, obj: *mut Object) -> bool

C++ method: bool QMetaProperty::reset(QObject* obj) const

C++ documentation:

Resets the property for the given object with a reset method. Returns true if the reset worked; otherwise returns false.

Reset methods are optional; only a few properties support them.

See also read() and write().

fn reset_on_gadget(&self, gadget: *mut c_void) -> bool

C++ method: bool QMetaProperty::resetOnGadget(void* gadget) const

C++ documentation:

Resets the property for the given gadget with a reset method. Returns true if the reset worked; otherwise returns false.

Reset methods are optional; only a few properties support them.

This function should only be used if this is a property of a Q_GADGET

This function was introduced in Qt 5.5.

fn revision(&self) -> c_int

C++ method: int QMetaProperty::revision() const

C++ documentation:

Returns the property revision if one was specified by REVISION, otherwise returns 0.

This function was introduced in Qt 5.1.

fn type_(&self) -> Type

C++ method: QVariant::Type QMetaProperty::type() const

C++ documentation:

Returns this property's type. The return value is one of the values of the QVariant::Type enumeration.

See also userType(), typeName(), and name().

fn type_name(&self) -> *const c_char

C++ method: const char* QMetaProperty::typeName() const

C++ documentation:

Returns the name of this property's type.

See also type() and name().

fn user_type(&self) -> c_int

C++ method: int QMetaProperty::userType() const

C++ documentation:

Returns this property's user type. The return value is one of the values that are registered with QMetaType, or QMetaType::UnknownType if the type is not registered.

This function was introduced in Qt 4.2.

See also type(), QMetaType, and typeName().

fn write(&self, obj: *mut Object, value: &Variant) -> bool

C++ method: bool QMetaProperty::write(QObject* obj, const QVariant& value) const

C++ documentation:

Writes value as the property's value to the given object. Returns true if the write succeeded; otherwise returns false.

See also read(), reset(), and isWritable().

fn write_on_gadget(&self, gadget: *mut c_void, value: &Variant) -> bool

C++ method: bool QMetaProperty::writeOnGadget(void* gadget, const QVariant& value) const

C++ documentation:

Writes value as the property's value to the given gadget. Returns true if the write succeeded; otherwise returns false.

This function should only be used if this is a property of a Q_GADGET

This function was introduced in Qt 5.5.