Enum qt_core::meta_type::TypeFlag

pub enum TypeFlag {
    NeedsConstruction,
    NeedsDestruction,
    MovableType,
    PointerToQObject,
    IsEnumeration,
    SharedPointerToQObject,
    WeakPointerToQObject,
    TrackingPointerToQObject,
    WasDeclaredAsMetaType,
    IsGadget,
}

C++ type: QMetaType::TypeFlag

C++ documentation:

The enum describes attributes of a type supported by QMetaType.

The TypeFlags type is a typedef for QFlags<TypeFlag>. It stores an OR combination of TypeFlag values.

Variants
NeedsConstruction

This type has non-trivial constructors. If the flag is not set instances can be safely initialized with memset to 0. (C++ enum variant: NeedsConstruction = 1)

NeedsDestruction

This type has a non-trivial destructor. If the flag is not set calls to the destructor are not necessary before discarding objects. (C++ enum variant: NeedsDestruction = 2)

MovableType

An instance of a type having this attribute can be safely moved by memcpy. (C++ enum variant: MovableType = 4)

PointerToQObject

This type is a pointer to a derived of QObject This type is a Q_GADGET and it's corresponding QMetaObject can be accessed with QMetaType::metaObject Since 5.5. (C++ enum variant: PointerToQObject = 8)

IsEnumeration

This type is an enumeration (C++ enum variant: IsEnumeration = 16)

SharedPointerToQObject

C++ enum variant: SharedPointerToQObject = 32

WeakPointerToQObject

C++ enum variant: WeakPointerToQObject = 64

TrackingPointerToQObject

C++ enum variant: TrackingPointerToQObject = 128

WasDeclaredAsMetaType

C++ enum variant: WasDeclaredAsMetaType = 256

IsGadget

C++ enum variant: IsGadget = 512