Module qt_core::meta_type::overloading

Traits

MetaTypeConstructArgs
This trait represents a set of arguments accepted by MetaType::construct method.

MetaTypeCreateArgs
This trait represents a set of arguments accepted by MetaType::create method.

MetaTypeCreateStaticArgs
This trait represents a set of arguments accepted by MetaType::create_static method.

MetaTypeNewArgs
This trait represents a set of arguments accepted by MetaType::new method.

MetaTypeRegisterNormalizedTypeArgs
This trait represents a set of arguments accepted by MetaType::register_normalized_type method.

MetaTypeRegisterTypeArgs
This trait represents a set of arguments accepted by MetaType::register_type method.

MetaTypeTypeArgs
This trait represents a set of arguments accepted by MetaType::type_ method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.