Struct qt_core::meta_type::MetaType

pub struct MetaType {
    // some fields omitted
}

C++ type: QMetaType

C++ documentation:

The QMetaType class manages named types in the meta-object system.

The class is used as a helper to marshall types in QVariant and in queued signals and slots connections. It associates a type name to a type so that it can be created and destructed dynamically at run-time. Declare new types with Q_DECLARE_METATYPE() to make them available to QVariant and other template-based functions. Call qRegisterMetaType() to make types available to non-template based functions, such as the queued signal and slot connections.

Any class or struct that has a public default constructor, a public copy constructor, and a public destructor can be registered.

The following code allocates and destructs an instance of MyClass:

int id = QMetaType::type("MyClass");
if (id != QMetaType::UnknownType) {
    void *myClassPtr = QMetaType::create(id);
    ...
    QMetaType::destroy(id, myClassPtr);
    myClassPtr = 0;
}

If we want the stream operators operator<<() and operator>>() to work on QVariant objects that store custom types, the custom type must provide operator<<() and operator>>() operators.

Methods

impl MetaType

fn compare(lhs: *const c_void, rhs: *const c_void, type_id: c_int, result: *mut c_int) -> bool

C++ method: static bool QMetaType::compare(const void* lhs, const void* rhs, int typeId, int* result)

C++ documentation:

Compares the objects at lhs and rhs. Both objects need to be of type typeId. result is set to less than, equal to or greater than zero, if lhs is less than, equal to or greater than rhs. Returns true, if the comparison succeeded, otherwise false.

This function was introduced in Qt 5.2.

fn construct<'largs, Args>(&'largs self, args: Args) -> *mut c_void where Args: MetaTypeConstructArgs<'largs>

C++ method: QMetaType::construct

This is an overloaded function. Available variants:

Rust arguments:
1) fn construct(&self, *mut libc::c_void) -> *mut libc::c_void
2) fn construct(&self, (*mut libc::c_void, *const libc::c_void)) -> *mut libc::c_void
C++ method: void* QMetaType::construct(void* where, const void* copy = ?) const

C++ documentation:

Constructs a value of the type that this QMetaType instance was constructed for in the existing memory addressed by where, that is a copy of copy, and returns where. If copy is zero, the value is default constructed.

This is a low-level function for explicitly managing the memory used to store the type. Consider calling create() if you don't need this level of control (that is, use "new" rather than "placement new").

You must ensure that where points to a location where the new value can be stored and that where is suitably aligned. The type's size can be queried by calling sizeOf().

The rule of thumb for alignment is that a type is aligned to its natural boundary, which is the smallest power of 2 that is bigger than the type, unless that alignment is larger than the maximum useful alignment for the platform. For practical purposes, alignment larger than 2 * sizeof(void*) is only necessary for special hardware instructions (e.g., aligned SSE loads and stores on x86).

This function was introduced in Qt 5.0.

fn construct_static(type_: c_int, where_: *mut c_void, copy: *const c_void) -> *mut c_void

C++ method: static void* QMetaType::construct(int type, void* where, const void* copy)

C++ documentation:

Constructs a value of the given type in the existing memory addressed by where, that is a copy of copy, and returns where. If copy is zero, the value is default constructed.

This is a low-level function for explicitly managing the memory used to store the type. Consider calling create() if you don't need this level of control (that is, use "new" rather than "placement new").

You must ensure that where points to a location that can store a value of type type, and that where is suitably aligned. The type's size can be queried by calling sizeOf().

The rule of thumb for alignment is that a type is aligned to its natural boundary, which is the smallest power of 2 that is bigger than the type, unless that alignment is larger than the maximum useful alignment for the platform. For practical purposes, alignment larger than 2 * sizeof(void*) is only necessary for special hardware instructions (e.g., aligned SSE loads and stores on x86).

This function was introduced in Qt 5.0.

See also destruct() and sizeOf().

fn convert(from: *const c_void, from_type_id: c_int, to: *mut c_void, to_type_id: c_int) -> bool

C++ method: static bool QMetaType::convert(const void* from, int fromTypeId, void* to, int toTypeId)

C++ documentation:

Converts the object at from from fromTypeId to the preallocated space at to typed toTypeId. Returns true, if the conversion succeeded, otherwise false.

This function was introduced in Qt 5.2.

fn create<'largs, Args>(&'largs self, args: Args) -> *mut c_void where Args: MetaTypeCreateArgs<'largs>

C++ method: QMetaType::create

This is an overloaded function. Available variants:

Rust arguments:
1) fn create(&self, *const libc::c_void) -> *mut libc::c_void
2) fn create(&self, ()) -> *mut libc::c_void
C++ method: void* QMetaType::create(const void* copy = ?) const

C++ documentation:

Returns a copy of copy, assuming it is of the type that this QMetaType instance was created for. If copy is null, creates a default constructed instance.

This function was introduced in Qt 5.0.

See also QMetaType::destroy().

fn create_static<Args>(args: Args) -> *mut c_void where Args: MetaTypeCreateStaticArgs

C++ method: QMetaType::create

This is an overloaded function. Available variants:

Rust arguments:
1) fn create_static(libc::c_int) -> *mut libc::c_void
2) fn create_static((libc::c_int, *const libc::c_void)) -> *mut libc::c_void
C++ method: static void* QMetaType::create(int type, const void* copy = ?)

C++ documentation:

Returns a copy of copy, assuming it is of type type. If copy is zero, creates a default constructed instance.

See also destroy(), isRegistered(), and Type.

fn debug_stream(dbg: &mut Debug, rhs: *const c_void, type_id: c_int) -> bool

C++ method: static bool QMetaType::debugStream(QDebug& dbg, const void* rhs, int typeId)

C++ documentation:

Streams the object at rhs of type typeId to the debug stream dbg. Returns true on success, otherwise false.

This function was introduced in Qt 5.2.

fn destroy(&self, data: *mut c_void)

C++ method: void QMetaType::destroy(void* data) const

C++ documentation:

Destroys the data, assuming it is of the type that this QMetaType instance was created for.

This function was introduced in Qt 5.0.

See also QMetaType::create().

fn destroy_static(type_: c_int, data: *mut c_void)

C++ method: static void QMetaType::destroy(int type, void* data)

C++ documentation:

Destroys the data, assuming it is of the type given.

See also create(), isRegistered(), and Type.

fn destruct(&self, data: *mut c_void)

C++ method: void QMetaType::destruct(void* data) const

C++ documentation:

Destructs the value, located at data, assuming that it is of the type for which this QMetaType instance was constructed for.

Unlike destroy(), this function only invokes the type's destructor, it doesn't invoke the delete operator.

This function was introduced in Qt 5.0.

See also QMetaType::construct().

fn destruct_static(type_: c_int, where_: *mut c_void)

C++ method: static void QMetaType::destruct(int type, void* where)

C++ documentation:

Destructs the value of the given type, located at where.

Unlike destroy(), this function only invokes the type's destructor, it doesn't invoke the delete operator.

This function was introduced in Qt 5.0.

See also construct().

fn equals(lhs: *const c_void, rhs: *const c_void, type_id: c_int, result: *mut c_int) -> bool

C++ method: static bool QMetaType::equals(const void* lhs, const void* rhs, int typeId, int* result)

C++ documentation:

Compares the objects at lhs and rhs. Both objects need to be of type typeId. result is set to zero, if lhs equals to rhs. Returns true, if the comparison succeeded, otherwise false.

This function was introduced in Qt 5.5.

fn flags(&self) -> Flags<TypeFlag>

C++ method: QFlags<QMetaType::TypeFlag> QMetaType::flags() const

C++ documentation:

Returns flags of the type for which this QMetaType instance was constructed.

This function was introduced in Qt 5.0.

See also QMetaType::TypeFlags and QMetaType::typeFlags().

fn has_registered_comparators(type_id: c_int) -> bool

C++ method: static bool QMetaType::hasRegisteredComparators(int typeId)

C++ documentation:

Returns true, if the meta type system has registered comparators for type id typeId.

This function was introduced in Qt 5.2.

fn has_registered_converter_function(from_type_id: c_int, to_type_id: c_int) -> bool

C++ method: static bool QMetaType::hasRegisteredConverterFunction(int fromTypeId, int toTypeId)

C++ documentation:

Returns true, if the meta type system has a registered conversion from meta type id fromTypeId to toTypeId

This function was introduced in Qt 5.2.

fn has_registered_debug_stream_operator(type_id: c_int) -> bool

C++ method: static bool QMetaType::hasRegisteredDebugStreamOperator(int typeId)

C++ documentation:

Returns true, if the meta type system has a registered debug stream operator for type id typeId.

This function was introduced in Qt 5.2.

fn is_registered(&self) -> bool

C++ method: bool QMetaType::isRegistered() const

C++ documentation:

Returns true if this QMetaType object contains valid information about a type, false otherwise.

This function was introduced in Qt 5.0.

fn is_registered_static(type_: c_int) -> bool

C++ method: static bool QMetaType::isRegistered(int type)

C++ documentation:

Returns true if the datatype with ID type is registered; otherwise returns false.

See also type(), typeName(), and Type.

fn is_valid(&self) -> bool

C++ method: bool QMetaType::isValid() const

C++ documentation:

Returns true if this QMetaType object contains valid information about a type, false otherwise.

This function was introduced in Qt 5.0.

fn load(stream: &mut DataStream, type_: c_int, data: *mut c_void) -> bool

C++ method: static bool QMetaType::load(QDataStream& stream, int type, void* data)

C++ documentation:

Reads the object of the specified type from the given stream into data. Returns true if the object is loaded successfully; otherwise returns false.

The type must have been registered with qRegisterMetaType() and qRegisterMetaTypeStreamOperators() beforehand.

Normally, you should not need to call this function directly. Instead, use QVariant's operator>>(), which relies on load() to stream custom types.

See also save() and qRegisterMetaTypeStreamOperators().

fn meta_object(&self) -> *const MetaObject

C++ method: const QMetaObject* QMetaType::metaObject() const

C++ documentation:

return a QMetaObject relative to this type.

If the type is a pointer type to a subclass of QObject, flags() contains QMetaType::PointerToQObject and this function returns the corresponding QMetaObject. This can be used to in combinaison with QMetaObject::construct to create QObject of this type.

If the type is a Q_GADGET, flags() contains QMetaType::IsGadget, and this function returns its QMetaObject. This can be used to retrieve QMetaMethod and QMetaProperty and use them on a pointer of this type. (given by QVariant::data for example)

If the type is an enumeration, flags() contains QMetaType::IsEnumeration, and this function returns the QMetaObject of the enclosing object if the enum was registered as a Q_ENUM or 0 otherwise

This function was introduced in Qt 5.5.

See also QMetaType::metaObjectForType() and QMetaType::flags().

fn meta_object_for_type(type_: c_int) -> *const MetaObject

C++ method: static const QMetaObject* QMetaType::metaObjectForType(int type)

C++ documentation:

returns QMetaType::metaObject for type

This function was introduced in Qt 5.0.

See also metaObject().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MetaTypeNewArgs

C++ method: QMetaType::QMetaType

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::meta_type::MetaType
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_type::MetaType>
C++ method: [constructor] void QMetaType::QMetaType(const int type)

fn register_normalized_type<Args>(args: Args) -> c_int where Args: MetaTypeRegisterNormalizedTypeArgs

C++ method: QMetaType::registerNormalizedType

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn register_normalized_type((&::byte_array::ByteArray, extern "C" fn(*mut libc::c_void), extern "C" fn(*const libc::c_void) -> *mut libc::c_void, extern "C" fn(*mut libc::c_void), extern "C" fn(*mut libc::c_void, *const libc::c_void) -> *mut libc::c_void, libc::c_int, ::flags::Flags<::meta_type::TypeFlag>, *const ::meta_object::MetaObject)) -> libc::c_int
C++ method: static int QMetaType::registerNormalizedType(const QByteArray& normalizedTypeName, void (*FN_PTR)(void*) deleter, void* (*FN_PTR)(const void*) creator, void (*FN_PTR)(void*) destructor, void* (*FN_PTR)(void*, const void*) constructor, int size, QFlags<QMetaType::TypeFlag> flags, const QMetaObject* metaObject)

Variant 2

Rust arguments: fn register_normalized_type((&::byte_array::ByteArray, extern "C" fn(*mut libc::c_void), extern "C" fn(*mut libc::c_void, *const libc::c_void) -> *mut libc::c_void, libc::c_int, ::flags::Flags<::meta_type::TypeFlag>, *const ::meta_object::MetaObject)) -> libc::c_int
C++ method: static int QMetaType::registerNormalizedType(const QByteArray& normalizedTypeName, void (*FN_PTR)(void*) destructor, void* (*FN_PTR)(void*, const void*) constructor, int size, QFlags<QMetaType::TypeFlag> flags, const QMetaObject* metaObject)

fn register_normalized_typedef(normalized_type_name: &ByteArray, alias_id: c_int) -> c_int

C++ method: static int QMetaType::registerNormalizedTypedef(const QByteArray& normalizedTypeName, int aliasId)

fn register_type<Args>(args: Args) -> c_int where Args: MetaTypeRegisterTypeArgs

C++ method: QMetaType::registerType

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn register_type((*const libc::c_char, extern "C" fn(*mut libc::c_void), extern "C" fn(*const libc::c_void) -> *mut libc::c_void)) -> libc::c_int
C++ method: static int QMetaType::registerType(const char* typeName, void (*FN_PTR)(void*) deleter, void* (*FN_PTR)(const void*) creator)

Variant 2

Rust arguments: fn register_type((*const libc::c_char, extern "C" fn(*mut libc::c_void), extern "C" fn(*const libc::c_void) -> *mut libc::c_void, extern "C" fn(*mut libc::c_void), extern "C" fn(*mut libc::c_void, *const libc::c_void) -> *mut libc::c_void, libc::c_int, ::flags::Flags<::meta_type::TypeFlag>, *const ::meta_object::MetaObject)) -> libc::c_int
C++ method: static int QMetaType::registerType(const char* typeName, void (*FN_PTR)(void*) deleter, void* (*FN_PTR)(const void*) creator, void (*FN_PTR)(void*) destructor, void* (*FN_PTR)(void*, const void*) constructor, int size, QFlags<QMetaType::TypeFlag> flags, const QMetaObject* metaObject)

fn register_typedef(type_name: *const c_char, alias_id: c_int) -> c_int

C++ method: static int QMetaType::registerTypedef(const char* typeName, int aliasId)

fn save(stream: &mut DataStream, type_: c_int, data: *const c_void) -> bool

C++ method: static bool QMetaType::save(QDataStream& stream, int type, const void* data)

C++ documentation:

Writes the object pointed to by data with the ID type to the given stream. Returns true if the object is saved successfully; otherwise returns false.

The type must have been registered with qRegisterMetaType() and qRegisterMetaTypeStreamOperators() beforehand.

Normally, you should not need to call this function directly. Instead, use QVariant's operator<<(), which relies on save() to stream custom types.

See also load() and qRegisterMetaTypeStreamOperators().

fn size_of(&self) -> c_int

C++ method: int QMetaType::sizeOf() const

C++ documentation:

Returns the size of the type in bytes (i.e. sizeof(T), where T is the actual type for which this QMetaType instance was constructed for).

This function is typically used together with construct() to perform low-level management of the memory used by a type.

This function was introduced in Qt 5.0.

See also QMetaType::construct() and QMetaType::sizeOf().

fn size_of_static(type_: c_int) -> c_int

C++ method: static int QMetaType::sizeOf(int type)

C++ documentation:

Returns the size of the given type in bytes (i.e. sizeof(T), where T is the actual type identified by the type argument).

This function is typically used together with construct() to perform low-level management of the memory used by a type.

This function was introduced in Qt 5.0.

See also construct().

fn type_<Args>(args: Args) -> c_int where Args: MetaTypeTypeArgs

C++ method: QMetaType::type

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn type_(&::byte_array::ByteArray) -> libc::c_int
C++ method: static int QMetaType::type(const QByteArray& typeName)

C++ documentation:

This is an overloaded function.

Returns a handle to the type called typeName, or 0 if there is no such type.

This function was introduced in Qt 5.5.

See also isRegistered() and typeName().

Variant 2

Rust arguments: fn type_(*const libc::c_char) -> libc::c_int
C++ method: static int QMetaType::type(const char* typeName)

C++ documentation:

Returns a handle to the type called typeName, or QMetaType::UnknownType if there is no such type.

See also isRegistered(), typeName(), and Type.

fn type_flags(type_: c_int) -> Flags<TypeFlag>

C++ method: static QFlags<QMetaType::TypeFlag> QMetaType::typeFlags(int type)

C++ documentation:

Returns flags of the given type.

This function was introduced in Qt 5.0.

See also QMetaType::TypeFlags.

fn type_name(type_: c_int) -> *const c_char

C++ method: static const char* QMetaType::typeName(int type)

C++ documentation:

Returns the type name associated with the given typeId, or 0 if no matching type was found. The returned pointer must not be deleted.

See also type(), isRegistered(), and Type.

fn unregister_type(type_: c_int) -> bool

C++ method: static bool QMetaType::unregisterType(int type)