Module qt_core::mime_data::overloading

Traits

MimeDataColorDataArgs
This trait represents a set of arguments accepted by MimeData::color_data method.

MimeDataConnectArgs
This trait represents a set of arguments accepted by MimeData::connect method.

MimeDataConnectStaticArgs
This trait represents a set of arguments accepted by MimeData::connect_static method.

MimeDataDataArgs
This trait represents a set of arguments accepted by MimeData::data method.

MimeDataDestroyedArgs
This trait represents a set of arguments accepted by MimeData::destroyed method.

MimeDataDisconnectArgs
This trait represents a set of arguments accepted by MimeData::disconnect method.

MimeDataDisconnectStaticArgs
This trait represents a set of arguments accepted by MimeData::disconnect_static method.

MimeDataDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by MimeData::dynamic_property_names method.

MimeDataFormatsArgs
This trait represents a set of arguments accepted by MimeData::formats method.

MimeDataHtmlArgs
This trait represents a set of arguments accepted by MimeData::html method.

MimeDataImageDataArgs
This trait represents a set of arguments accepted by MimeData::image_data method.

MimeDataNewArgs
This trait represents a set of arguments accepted by MimeData::new method.

MimeDataObjectNameArgs
This trait represents a set of arguments accepted by MimeData::object_name method.

MimeDataPropertyArgs
This trait represents a set of arguments accepted by MimeData::property method.

MimeDataStartTimerArgs
This trait represents a set of arguments accepted by MimeData::start_timer method.

MimeDataTextArgs
This trait represents a set of arguments accepted by MimeData::text method.

MimeDataTrArgs
This trait represents a set of arguments accepted by MimeData::tr method.

MimeDataTrUtf8Args
This trait represents a set of arguments accepted by MimeData::tr_utf8 method.

MimeDataUrlsArgs
This trait represents a set of arguments accepted by MimeData::urls method.