Module qt_core::mime_database::overloading

Traits

MimeDatabaseAllMimeTypesArgs
This trait represents a set of arguments accepted by MimeDatabase::all_mime_types method.

MimeDatabaseMimeTypeForDataArgs
This trait represents a set of arguments accepted by MimeDatabase::mime_type_for_data method.

MimeDatabaseMimeTypeForFileArgs
This trait represents a set of arguments accepted by MimeDatabase::mime_type_for_file method.

MimeDatabaseMimeTypeForFileNameAndDataArgs
This trait represents a set of arguments accepted by MimeDatabase::mime_type_for_file_name_and_data method.

MimeDatabaseMimeTypeForNameArgs
This trait represents a set of arguments accepted by MimeDatabase::mime_type_for_name method.

MimeDatabaseMimeTypeForUrlArgs
This trait represents a set of arguments accepted by MimeDatabase::mime_type_for_url method.

MimeDatabaseMimeTypesForFileNameArgs
This trait represents a set of arguments accepted by MimeDatabase::mime_types_for_file_name method.

MimeDatabaseNewArgs
This trait represents a set of arguments accepted by MimeDatabase::new method.

MimeDatabaseSuffixForFileNameArgs
This trait represents a set of arguments accepted by MimeDatabase::suffix_for_file_name method.