Struct qt_core::mime_database::MimeDatabase

pub struct MimeDatabase {
    // some fields omitted
}

C++ type: QMimeDatabase

C++ documentation:

The QMimeDatabase class maintains a database of MIME types.

The MIME type database is provided by the freedesktop.org shared-mime-info project. If the MIME type database cannot be found on the system, as is the case on most Windows, OS X, and iOS systems, Qt will use its own copy of it.

Applications which want to define custom MIME types need to install an XML file into the locations searched for MIME definitions. These locations can be queried with

QStandardPaths::locateAll(QStandardPaths::GenericDataLocation, QLatin1String("mime/packages"),
                          QStandardPaths::LocateDirectory);

On a typical Unix system, this will be /usr/share/mime/packages/, but it is also possible to extend the list of directories by setting the environment variable XDG_DATA_DIRS. For instance adding /opt/myapp/share to XDG_DATA_DIRS will result in /opt/myapp/share/mime/packages/ being searched for MIME definitions.

Here is an example of MIME XML:

<?xml version="1.0" encoding="UTF-8"?>
<mime-info xmlns="http://www.freedesktop.org/standards/shared-mime-info">
  <mime-type type="application/vnd.nokia.qt.qmakeprofile">
    <comment xml:lang="en">Qt qmake Profile</comment>
    <glob pattern="*.pro" weight="50"/>
  </mime-type>
</mime-info>

For more details about the syntax of XML MIME definitions, including defining "magic" in order to detect MIME types based on data as well, read the Shared Mime Info specification at http://standards.freedesktop.org/shared-mime-info-spec/shared-mime-info-spec-latest.html

On Unix systems, a binary cache is used for more performance. This cache is generated by the command "update-mime-database path", where path would be /opt/myapp/share/mime in the above example. Make sure to run this command when installing the MIME type definition file.

QMimeDatabase db;
QMimeType mime = db.mimeTypeForFile(fileName);
if (mime.inherits("text/plain")) {
    // The file is plain text, we can display it in a QTextEdit
}

Methods

impl MimeDatabase

fn all_mime_types<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseAllMimeTypesArgs<'largs>

C++ method: QMimeDatabase::allMimeTypes

This is an overloaded function. Available variants:

Rust arguments:
1) fn all_mime_types(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListMimeTypeRef>
2) fn all_mime_types(&self, cpp_utils::AsStruct) -> ::list::ListMimeTypeRef
C++ method: QList<QMimeType> QMimeDatabase::allMimeTypes() const

C++ documentation:

Returns the list of all available MIME types.

This can be useful for showing all MIME types to the user, for instance in a MIME type editor. Do not use unless really necessary in other cases though, prefer using the mimeTypeForXxx() methods for performance reasons.

fn mime_type_for_data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseMimeTypeForDataArgs<'largs>

C++ method: QMimeDatabase::mimeTypeForData

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn mime_type_for_data(&self, (*mut ::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_data(&self, (*mut ::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForData(QIODevice* device) const

C++ documentation:

Returns a MIME type for the data in device.

A valid MIME type is always returned. If the data in device doesn't match any known MIME type data, the default MIME type (application/octet-stream) is returned.

Variant 2

Rust arguments:
1) fn mime_type_for_data(&self, (&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_data(&self, (&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForData(const QByteArray& data) const

C++ documentation:

Returns a MIME type for data.

A valid MIME type is always returned. If data doesn't match any known MIME type data, the default MIME type (application/octet-stream) is returned.

fn mime_type_for_file<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseMimeTypeForFileArgs<'largs>

C++ method: QMimeDatabase::mimeTypeForFile

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn mime_type_for_file(&self, (&::file_info::FileInfo, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_file(&self, (&::file_info::FileInfo, ::mime_database::MatchMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
3) fn mime_type_for_file(&self, (&::file_info::FileInfo, cpp_utils::AsStruct)) -> ::mime_type::MimeType
4) fn mime_type_for_file(&self, (&::file_info::FileInfo, ::mime_database::MatchMode, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForFile(const QFileInfo& fileInfo, QMimeDatabase::MatchMode mode = ?) const

C++ documentation:

Returns a MIME type for fileInfo.

A valid MIME type is always returned.

The default matching algorithm looks at both the file name and the file contents, if necessary. The file extension has priority over the contents, but the contents will be used if the file extension is unknown, or matches multiple MIME types. If fileInfo is a Unix symbolic link, the file that it refers to will be used instead. If the file doesn't match any known pattern or data, the default MIME type (application/octet-stream) is returned.

When mode is set to MatchExtension, only the file name is used, not the file contents. The file doesn't even have to exist. If the file name doesn't match any known pattern, the default MIME type (application/octet-stream) is returned. If multiple MIME types match this file, the first one (alphabetically) is returned.

When mode is set to MatchContent, and the file is readable, only the file contents are used to determine the MIME type. This is equivalent to calling mimeTypeForData with a QFile as input device.

fileInfo may refer to an absolute or relative path.

See also QMimeType::isDefault() and mimeTypeForData().

Variant 2

Rust arguments:
1) fn mime_type_for_file(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_file(&self, (&::string::String, ::mime_database::MatchMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
3) fn mime_type_for_file(&self, (&::string::String, cpp_utils::AsStruct)) -> ::mime_type::MimeType
4) fn mime_type_for_file(&self, (&::string::String, ::mime_database::MatchMode, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForFile(const QString& fileName, QMimeDatabase::MatchMode mode = ?) const

C++ documentation:

Returns a MIME type for the file named fileName using mode.

This is an overloaded function.

fn mime_type_for_file_name_and_data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseMimeTypeForFileNameAndDataArgs<'largs>

C++ method: QMimeDatabase::mimeTypeForFileNameAndData

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn mime_type_for_file_name_and_data(&self, (&::string::String, *mut ::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_file_name_and_data(&self, (&::string::String, *mut ::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForFileNameAndData(const QString& fileName, QIODevice* device) const

C++ documentation:

Returns a MIME type for the given fileName and device data.

This overload can be useful when the file is remote, and we started to download some of its data in a device. This allows to do full MIME type matching for remote files as well.

If the device is not open, it will be opened by this function, and closed after the MIME type detection is completed.

A valid MIME type is always returned. If device data doesn't match any known MIME type data, the default MIME type (application/octet-stream) is returned.

This method looks at both the file name and the file contents, if necessary. The file extension has priority over the contents, but the contents will be used if the file extension is unknown, or matches multiple MIME types.

Variant 2

Rust arguments:
1) fn mime_type_for_file_name_and_data(&self, (&::string::String, &::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_file_name_and_data(&self, (&::string::String, &::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForFileNameAndData(const QString& fileName, const QByteArray& data) const

C++ documentation:

Returns a MIME type for the given fileName and device data.

This overload can be useful when the file is remote, and we started to download some of its data. This allows to do full MIME type matching for remote files as well.

A valid MIME type is always returned. If data doesn't match any known MIME type data, the default MIME type (application/octet-stream) is returned.

This method looks at both the file name and the file contents, if necessary. The file extension has priority over the contents, but the contents will be used if the file extension is unknown, or matches multiple MIME types.

fn mime_type_for_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseMimeTypeForNameArgs<'largs>

C++ method: QMimeDatabase::mimeTypeForName

This is an overloaded function. Available variants:

Rust arguments:
1) fn mime_type_for_name(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_name(&self, (&::string::String, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForName(const QString& nameOrAlias) const

C++ documentation:

Returns a MIME type for nameOrAlias or an invalid one if none found.

fn mime_type_for_url<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseMimeTypeForUrlArgs<'largs>

C++ method: QMimeDatabase::mimeTypeForUrl

This is an overloaded function. Available variants:

Rust arguments:
1) fn mime_type_for_url(&self, (&::url::Url, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
2) fn mime_type_for_url(&self, (&::url::Url, cpp_utils::AsStruct)) -> ::mime_type::MimeType
C++ method: QMimeType QMimeDatabase::mimeTypeForUrl(const QUrl& url) const

C++ documentation:

Returns a MIME type for url.

If the URL is a local file, this calls mimeTypeForFile.

Otherwise the matching is done based on the file name only, except for schemes where file names don't mean much, like HTTP. This method always returns the default mimetype for HTTP URLs, use QNetworkAccessManager to handle HTTP URLs properly.

A valid MIME type is always returned. If url doesn't match any known MIME type data, the default MIME type (application/octet-stream) is returned.

fn mime_types_for_file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseMimeTypesForFileNameArgs<'largs>

C++ method: QMimeDatabase::mimeTypesForFileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn mime_types_for_file_name(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListMimeTypeRef>
2) fn mime_types_for_file_name(&self, (&::string::String, cpp_utils::AsStruct)) -> ::list::ListMimeTypeRef
C++ method: QList<QMimeType> QMimeDatabase::mimeTypesForFileName(const QString& fileName) const

C++ documentation:

Returns the MIME types for the file name fileName.

If the file name doesn't match any known pattern, an empty list is returned. If multiple MIME types match this file, they are all returned.

This function does not try to open the file. To also use the content when determining the MIME type, use mimeTypeForFile() or mimeTypeForFileNameAndData() instead.

See also mimeTypeForFile().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MimeDatabaseNewArgs

C++ method: QMimeDatabase::QMimeDatabase

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::mime_database::MimeDatabase
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::mime_database::MimeDatabase>
C++ method: [constructor] void QMimeDatabase::QMimeDatabase()

fn suffix_for_file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeDatabaseSuffixForFileNameArgs<'largs>

C++ method: QMimeDatabase::suffixForFileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn suffix_for_file_name(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn suffix_for_file_name(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QMimeDatabase::suffixForFileName(const QString& fileName) const

C++ documentation:

Returns the suffix for the file fileName, as known by the MIME database.

This allows to pre-select "tar.bz2" for foo.tar.bz2, but still only "txt" for my.file.with.dots.txt.