Module qt_core::mime_type::overloading

Traits

MimeTypeAliasesArgs
This trait represents a set of arguments accepted by MimeType::aliases method.

MimeTypeAllAncestorsArgs
This trait represents a set of arguments accepted by MimeType::all_ancestors method.

MimeTypeCommentArgs
This trait represents a set of arguments accepted by MimeType::comment method.

MimeTypeFilterStringArgs
This trait represents a set of arguments accepted by MimeType::filter_string method.

MimeTypeGenericIconNameArgs
This trait represents a set of arguments accepted by MimeType::generic_icon_name method.

MimeTypeGlobPatternsArgs
This trait represents a set of arguments accepted by MimeType::glob_patterns method.

MimeTypeIconNameArgs
This trait represents a set of arguments accepted by MimeType::icon_name method.

MimeTypeNameArgs
This trait represents a set of arguments accepted by MimeType::name method.

MimeTypeNewArgs
This trait represents a set of arguments accepted by MimeType::new method.

MimeTypeParentMimeTypesArgs
This trait represents a set of arguments accepted by MimeType::parent_mime_types method.

MimeTypePreferredSuffixArgs
This trait represents a set of arguments accepted by MimeType::preferred_suffix method.

MimeTypeSuffixesArgs
This trait represents a set of arguments accepted by MimeType::suffixes method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.