Struct qt_core::mime_type::MimeType

pub struct MimeType {
    // some fields omitted
}

C++ type: QMimeType

C++ documentation:

The QMimeType class describes types of file or data, represented by a MIME type string.

For instance a file named "readme.txt" has the MIME type "text/plain". The MIME type can be determined from the file name, or from the file contents, or from both. MIME type determination can also be done on buffers of data not coming from files.

Determining the MIME type of a file can be useful to make sure your application supports it. It is also useful in file-manager-like applications or widgets, in order to display an appropriate icon for the file, or even the descriptive comment in detailed views.

To check if a file has the expected MIME type, you should use inherits() rather than a simple string comparison based on the name(). This is because MIME types can inherit from each other: for instance a C source file is a specific type of plain text file, so text/x-csrc inherits text/plain.

Methods

impl MimeType

fn aliases<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeAliasesArgs<'largs>

C++ method: QMimeType::aliases

This is an overloaded function. Available variants:

Rust arguments:
1) fn aliases(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn aliases(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QMimeType::aliases() const

C++ documentation:

Return the list of aliases of this mimetype.

For instance, for text/csv, the returned list would be: text/x-csv, text/x-comma-separated-values.

Note that all QMimeType instances refer to proper mimetypes, never to aliases directly.

The order of the aliases in the list is undefined.

fn all_ancestors<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeAllAncestorsArgs<'largs>

C++ method: QMimeType::allAncestors

This is an overloaded function. Available variants:

Rust arguments:
1) fn all_ancestors(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn all_ancestors(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QMimeType::allAncestors() const

C++ documentation:

Return all the parent mimetypes of this mimetype, direct and indirect. This includes the parent(s) of its parent(s), etc.

For instance, for image/svg+xml the list would be: application/xml, text/plain, application/octet-stream.

Note that application/octet-stream is the ultimate parent for all types of files (but not directories).

fn comment<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeCommentArgs<'largs>

C++ method: QMimeType::comment

This is an overloaded function. Available variants:

Rust arguments:
1) fn comment(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn comment(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QMimeType::comment() const

C++ documentation:

Returns the description of the MIME type to be displayed on user interfaces.

The system language (QLocale::system().name()) is used to select the appropriate translation.

fn filter_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeFilterStringArgs<'largs>

C++ method: QMimeType::filterString

This is an overloaded function. Available variants:

Rust arguments:
1) fn filter_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn filter_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QMimeType::filterString() const

C++ documentation:

Returns a filter string usable for a file dialog.

fn generic_icon_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeGenericIconNameArgs<'largs>

C++ method: QMimeType::genericIconName

This is an overloaded function. Available variants:

Rust arguments:
1) fn generic_icon_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn generic_icon_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QMimeType::genericIconName() const

C++ documentation:

Returns the file name of a generic icon that represents the MIME type.

This should be used if the icon returned by iconName() cannot be found on the system. It is used for categories of similar types (like spreadsheets or archives) that can use a common icon. The freedesktop.org Icon Naming Specification lists a set of such icon names.

The icon name can be given to QIcon::fromTheme() in order to load the icon.

fn glob_patterns<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeGlobPatternsArgs<'largs>

C++ method: QMimeType::globPatterns

This is an overloaded function. Available variants:

Rust arguments:
1) fn glob_patterns(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn glob_patterns(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QMimeType::globPatterns() const

C++ documentation:

Returns the list of glob matching patterns.

fn icon_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeIconNameArgs<'largs>

C++ method: QMimeType::iconName

This is an overloaded function. Available variants:

Rust arguments:
1) fn icon_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn icon_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QMimeType::iconName() const

C++ documentation:

Returns the file name of an icon image that represents the MIME type.

The icon name can be given to QIcon::fromTheme() in order to load the icon.

fn inherits(&self, mime_type_name: &String) -> bool

C++ method: bool QMimeType::inherits(const QString& mimeTypeName) const

C++ documentation:

Returns true if this mimetype is mimeTypeName, or inherits mimeTypeName (see parentMimeTypes()), or mimeTypeName is an alias for this mimetype.

fn is_default(&self) -> bool

C++ method: bool QMimeType::isDefault() const

C++ documentation:

Returns true if this MIME type is the default MIME type which applies to all files: application/octet-stream.

fn is_valid(&self) -> bool

C++ method: bool QMimeType::isValid() const

C++ documentation:

Returns true if the QMimeType object contains valid data, otherwise returns false. A valid MIME type has a non-empty name(). The invalid MIME type is the default-constructed QMimeType.

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeNameArgs<'largs>

C++ method: QMimeType::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QMimeType::name() const

C++ documentation:

Returns the name of the MIME type.

fn new<Args>(args: Args) -> Args::ReturnType where Args: MimeTypeNewArgs

C++ method: QMimeType::QMimeType

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::mime_type::MimeType
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::mime_type::MimeType>
C++ method: [constructor] void QMimeType::QMimeType()

Variant 2

Rust arguments:
1) fn new((&::mime_type::MimeType, cpp_utils::AsStruct)) -> ::mime_type::MimeType
2) fn new((&::mime_type::MimeType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mime_type::MimeType>
C++ method: [constructor] void QMimeType::QMimeType(const QMimeType& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 MimeType) -> &'l0 mut MimeType

C++ method: QMimeType& QMimeType::operator=(const QMimeType& other)

C++ documentation:

Assigns the data of other to this QMimeType object, and returns a reference to this object.

fn op_eq(&self, other: &MimeType) -> bool

C++ method: bool QMimeType::operator==(const QMimeType& other) const

C++ documentation:

Returns true if other equals this QMimeType object, otherwise returns false. The name is the unique identifier for a mimetype, so two mimetypes with the same name, are equal.

fn op_neq(&self, other: &MimeType) -> bool

C++ method: bool QMimeType::operator!=(const QMimeType& other) const

C++ documentation:

Returns true if other does not equal this QMimeType object, otherwise returns false.

fn parent_mime_types<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeParentMimeTypesArgs<'largs>

C++ method: QMimeType::parentMimeTypes

This is an overloaded function. Available variants:

Rust arguments:
1) fn parent_mime_types(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn parent_mime_types(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QMimeType::parentMimeTypes() const

C++ documentation:

A type is a subclass of another type if any instance of the first type is also an instance of the second. For example, all image/svg+xml files are also text/xml, text/plain and application/octet-stream files. Subclassing is about the format, rather than the category of the data (for example, there is no 'generic spreadsheet' class that all spreadsheets inherit from). Conversely, the parent mimetype of image/svg+xml is text/xml.

A mimetype can have multiple parents. For instance application/x-perl has two parents: application/x-executable and text/plain. This makes it possible to both execute perl scripts, and to open them in text editors.

fn preferred_suffix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypePreferredSuffixArgs<'largs>

C++ method: QMimeType::preferredSuffix

This is an overloaded function. Available variants:

Rust arguments:
1) fn preferred_suffix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn preferred_suffix(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QMimeType::preferredSuffix() const

C++ documentation:

Returns the preferred suffix for the MIME type. No leading dot is included, so for instance this would return "pdf" for application/pdf. The return value can be empty, for mime types which do not have any suffixes associated.

fn suffixes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: MimeTypeSuffixesArgs<'largs>

C++ method: QMimeType::suffixes

This is an overloaded function. Available variants:

Rust arguments:
1) fn suffixes(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn suffixes(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QMimeType::suffixes() const

C++ documentation:

Returns the known suffixes for the MIME type. No leading dot is included, so for instance this would return "jpg", "jpeg" for image/jpeg.

fn swap(&mut self, other: &mut MimeType)

C++ method: void QMimeType::swap(QMimeType& other)

C++ documentation:

Swaps QMimeType other with this QMimeType object.

This operation is very fast and never fails.

The swap() method helps with the implementation of assignment operators in an exception-safe way. For more information consult More C++ Idioms - Copy-and-swap.