Struct qt_core::model_index::ModelIndex

pub struct ModelIndex {
    // some fields omitted
}

C++ type: QModelIndex

C++ documentation:

The QModelIndex class is used to locate data in a data model.

This class is used as an index into item models derived from QAbstractItemModel. The index is used by item views, delegates, and selection models to locate an item in the model.

New QModelIndex objects are created by the model using the QAbstractItemModel::createIndex() function. An invalid model index can be constructed with the QModelIndex constructor. Invalid indexes are often used as parent indexes when referring to top-level items in a model.

Model indexes refer to items in models, and contain all the information required to specify their locations in those models. Each index is located in a given row and column, and may have a parent index; use row(), column(), and parent() to obtain this information. Each top-level item in a model is represented by a model index that does not have a parent index - in this case, parent() will return an invalid model index, equivalent to an index constructed with the zero argument form of the QModelIndex() constructor.

To obtain a model index that refers to an existing item in a model, call QAbstractItemModel::index() with the required row and column values, and the model index of the parent. When referring to top-level items in a model, supply QModelIndex() as the parent index.

The model() function returns the model that the index references as a QAbstractItemModel. The child() function is used to examine items held under the index in the model. The sibling() function allows you to traverse items in the model on the same level as the index.

Note: Model indexes should be used immediately and then discarded. You should not rely on indexes to remain valid after calling model functions that change the structure of the model or delete items. If you need to keep a model index over time use a QPersistentModelIndex.

Methods

impl ModelIndex

fn child<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ModelIndexChildArgs<'largs>

C++ method: QModelIndex::child

This is an overloaded function. Available variants:

Rust arguments:
1) fn child(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn child(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::model_index::ModelIndex
C++ method: QModelIndex QModelIndex::child(int row, int column) const

C++ documentation:

Returns the child of the model index that is stored in the given row and column.

Note: This function does not work for an invalid model index which is often used as the root index.

See also parent() and sibling().

fn column(&self) -> c_int

C++ method: int QModelIndex::column() const

C++ documentation:

Returns the column this model index refers to.

fn data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ModelIndexDataArgs<'largs>

C++ method: QModelIndex::data

This is an overloaded function. Available variants:

Rust arguments:
1) fn data(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn data(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
3) fn data(&self, cpp_utils::AsStruct) -> ::variant::Variant
4) fn data(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QModelIndex::data(int role = ?) const

C++ documentation:

Returns the data for the given role for the item referred to by the index.

fn flags(&self) -> Flags<ItemFlag>

C++ method: QFlags<Qt::ItemFlag> QModelIndex::flags() const

C++ documentation:

Returns the flags for the item referred to by the index.

This function was introduced in Qt 4.2.

fn internal_id(&self) -> usize

C++ method: quintptr QModelIndex::internalId() const

C++ documentation:

Returns a quintptr used by the model to associate the index with the internal data structure.

See also QAbstractItemModel::createIndex().

fn internal_pointer(&self) -> *mut c_void

C++ method: void* QModelIndex::internalPointer() const

C++ documentation:

Returns a void * pointer used by the model to associate the index with the internal data structure.

See also QAbstractItemModel::createIndex().

fn is_valid(&self) -> bool

C++ method: bool QModelIndex::isValid() const

C++ documentation:

Returns true if this model index is valid; otherwise returns false.

A valid index belongs to a model, and has non-negative row and column numbers.

See also model(), row(), and column().

fn model(&self) -> *const AbstractItemModel

C++ method: const QAbstractItemModel* QModelIndex::model() const

C++ documentation:

Returns a pointer to the model containing the item that this index refers to.

A const pointer to the model is returned because calls to non-const functions of the model might invalidate the model index and possibly crash your application.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ModelIndexNewArgs

C++ method: QModelIndex::QModelIndex

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::model_index::ModelIndex
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::model_index::ModelIndex>
C++ method: [constructor] void QModelIndex::QModelIndex()

fn op_eq(&self, other: &ModelIndex) -> bool

C++ method: bool QModelIndex::operator==(const QModelIndex& other) const

C++ documentation:

Returns true if this model index refers to the same location as the other model index; otherwise returns false.

All values in the model index are used when comparing with another model index.

fn op_lt(&self, other: &ModelIndex) -> bool

C++ method: bool QModelIndex::operator<(const QModelIndex& other) const

C++ documentation:

Returns true if this model index is smaller than the other model index; otherwise returns false.

The less than calculation is not directly useful to developers - the way that indexes with different parents compare is not defined. This operator only exists so that the class can be used with QMap.

This function was introduced in Qt 4.1.

fn op_neq(&self, other: &ModelIndex) -> bool

C++ method: bool QModelIndex::operator!=(const QModelIndex& other) const

C++ documentation:

Returns true if this model index does not refer to the same location as the other model index; otherwise returns false.

fn parent<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ModelIndexParentArgs<'largs>

C++ method: QModelIndex::parent

This is an overloaded function. Available variants:

Rust arguments:
1) fn parent(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn parent(&self, cpp_utils::AsStruct) -> ::model_index::ModelIndex
C++ method: QModelIndex QModelIndex::parent() const

C++ documentation:

Returns the parent of the model index, or QModelIndex() if it has no parent.

See also child(), sibling(), and model().

fn row(&self) -> c_int

C++ method: int QModelIndex::row() const

C++ documentation:

Returns the row this model index refers to.

fn sibling<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ModelIndexSiblingArgs<'largs>

C++ method: QModelIndex::sibling

This is an overloaded function. Available variants:

Rust arguments:
1) fn sibling(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn sibling(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::model_index::ModelIndex
C++ method: QModelIndex QModelIndex::sibling(int row, int column) const

C++ documentation:

Returns the sibling at row and column. If there is no sibling at this position, an invalid QModelIndex is returned.

See also parent() and child().