Struct qt_core::mutex::Mutex

pub struct Mutex {
    // some fields omitted
}

C++ type: QMutex

C++ documentation:

The QMutex class provides access serialization between threads.

The purpose of a QMutex is to protect an object, data structure or section of code so that only one thread can access it at a time (this is similar to the Java synchronized keyword). It is usually best to use a mutex with a QMutexLocker since this makes it easy to ensure that locking and unlocking are performed consistently.

For example, say there is a method that prints a message to the user on two lines:

int number = 6;

void method1() { number *= 5; number /= 4; }

void method2() { number = 3; number /= 2; }
If these two methods are called in succession, the following happens:

// method1()
number = 5;        // number is now 30
number /= 4;        // number is now 7

// method2() number = 3; // number is now 21 number /= 2; // number is now 10
If these two methods are called simultaneously from two threads then the following sequence could result:

// Thread 1 calls method1()
number = 5;        // number is now 30

// Thread 2 calls method2(). // // Most likely Thread 1 has been put to sleep by the operating // system to allow Thread 2 to run. number *= 3; // number is now 90 number /= 2; // number is now 45

// Thread 1 finishes executing. number /= 4; // number is now 11, instead of 10
If we add a mutex, we should get the result we want:

QMutex mutex;
int number = 6;

void method1() { mutex.lock(); number *= 5; number /= 4; mutex.unlock(); }

void method2() { mutex.lock(); number *= 3; number /= 2; mutex.unlock(); }
Then only one thread can modify number at any given time and the result is correct. This is a trivial example, of course, but applies to any other case where things need to happen in a particular sequence.

When you call lock() in a thread, other threads that try to call lock() in the same place will block until the thread that got the lock calls unlock(). A non-blocking alternative to lock() is tryLock().

Methods

impl Mutex

fn is_recursive(&mut self) -> bool

C++ method: bool QMutex::isRecursive()

fn lock(&mut self)

C++ method: void QMutex::lock()

C++ documentation:

Locks the mutex. If another thread has locked the mutex then this call will block until that thread has unlocked it.

Calling this function multiple times on the same mutex from the same thread is allowed if this mutex is a recursive mutex. If this mutex is a non-recursive mutex, this function will dead-lock when the mutex is locked recursively.

See also unlock().

fn new<Args>(args: Args) -> Args::ReturnType where Args: MutexNewArgs

C++ method: QMutex::QMutex

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((::mutex::RecursionMode, cpp_utils::AsStruct)) -> ::mutex::Mutex
2) fn new(cpp_utils::AsStruct) -> ::mutex::Mutex
3) fn new((::mutex::RecursionMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::mutex::Mutex>
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::mutex::Mutex>
C++ method: [constructor] void QMutex::QMutex(QMutex::RecursionMode mode = ?)

fn try_lock<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: MutexTryLockArgs<'largs>

C++ method: QMutex::tryLock

This is an overloaded function. Available variants:

Rust arguments:
1) fn try_lock(&mut self, ()) -> bool
2) fn try_lock(&mut self, libc::c_int) -> bool
C++ method: bool QMutex::tryLock(int timeout = ?)

C++ documentation:

Attempts to lock the mutex. This function returns true if the lock was obtained; otherwise it returns false. If another thread has locked the mutex, this function will wait for at most timeout milliseconds for the mutex to become available.

Note: Passing a negative number as the timeout is equivalent to calling lock(), i.e. this function will wait forever until mutex can be locked if timeout is negative.

If the lock was obtained, the mutex must be unlocked with unlock() before another thread can successfully lock it.

Calling this function multiple times on the same mutex from the same thread is allowed if this mutex is a recursive mutex. If this mutex is a non-recursive mutex, this function will always return false when attempting to lock the mutex recursively.

See also lock() and unlock().

fn unlock(&mut self)

C++ method: void QMutex::unlock()

C++ documentation:

Unlocks the mutex. Attempting to unlock a mutex in a different thread to the one that locked it results in an error. Unlocking a mutex that is not locked results in undefined behavior.

See also lock().