Module qt_core::numeric::overloading

Traits

FloatDistanceArgs
This trait represents a set of arguments accepted by float_distance method.

IsFiniteArgs
This trait represents a set of arguments accepted by is_finite method.

IsInfArgs
This trait represents a set of arguments accepted by is_inf method.

IsNaNArgs
This trait represents a set of arguments accepted by is_na_n method.