Module qt_core::object::overloading

Traits

ObjectConnectArgs
This trait represents a set of arguments accepted by Object::connect method.

ObjectConnectStaticArgs
This trait represents a set of arguments accepted by Object::connect_static method.

ObjectDestroyedArgs
This trait represents a set of arguments accepted by Object::destroyed method.

ObjectDisconnectArgs
This trait represents a set of arguments accepted by Object::disconnect method.

ObjectDisconnectStaticArgs
This trait represents a set of arguments accepted by Object::disconnect_static method.

ObjectDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Object::dynamic_property_names method.

ObjectNewArgs
This trait represents a set of arguments accepted by Object::new method.

ObjectObjectNameArgs
This trait represents a set of arguments accepted by Object::object_name method.

ObjectPropertyArgs
This trait represents a set of arguments accepted by Object::property method.

ObjectStartTimerArgs
This trait represents a set of arguments accepted by Object::start_timer method.

ObjectTrArgs
This trait represents a set of arguments accepted by Object::tr method.

ObjectTrUtf8Args
This trait represents a set of arguments accepted by Object::tr_utf8 method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

QtQFindChildrenHelperArgs
This trait represents a set of arguments accepted by qt_q_find_children_helper method.