Module qt_core::object_cleanup_handler::overloading

Traits

ObjectCleanupHandlerConnectArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::connect method.

ObjectCleanupHandlerConnectStaticArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::connect_static method.

ObjectCleanupHandlerDestroyedArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::destroyed method.

ObjectCleanupHandlerDisconnectArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::disconnect method.

ObjectCleanupHandlerDisconnectStaticArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::disconnect_static method.

ObjectCleanupHandlerDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::dynamic_property_names method.

ObjectCleanupHandlerNewArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::new method.

ObjectCleanupHandlerObjectNameArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::object_name method.

ObjectCleanupHandlerPropertyArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::property method.

ObjectCleanupHandlerStartTimerArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::start_timer method.

ObjectCleanupHandlerTrArgs
This trait represents a set of arguments accepted by ObjectCleanupHandler::tr method.

ObjectCleanupHandlerTrUtf8Args
This trait represents a set of arguments accepted by ObjectCleanupHandler::tr_utf8 method.