Struct qt_core::persistent_model_index::PersistentModelIndex

pub struct PersistentModelIndex {
    // some fields omitted
}

C++ type: QPersistentModelIndex

C++ documentation:

The QPersistentModelIndex class is used to locate data in a data model.

A QPersistentModelIndex is a model index that can be stored by an application, and later used to access information in a model. Unlike the QModelIndex class, it is safe to store a QPersistentModelIndex since the model will ensure that references to items will continue to be valid as long as they can be accessed by the model.

It is good practice to check that persistent model indexes are valid before using them.

Methods

impl PersistentModelIndex

fn as_q_model_index_ref<'l0>(&'l0 self) -> &'l0 ModelIndex

C++ method: const QModelIndex& QPersistentModelIndex::operator const QModelIndex &() const

C++ documentation:

Cast operator that returns a const QModelIndex&.

fn child<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PersistentModelIndexChildArgs<'largs>

C++ method: QPersistentModelIndex::child

This is an overloaded function. Available variants:

Rust arguments:
1) fn child(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn child(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::model_index::ModelIndex
C++ method: QModelIndex QPersistentModelIndex::child(int row, int column) const

C++ documentation:

Returns the child of the model index that is stored in the given row and column.

See also parent() and sibling().

fn column(&self) -> c_int

C++ method: int QPersistentModelIndex::column() const

C++ documentation:

Returns the column this persistent model index refers to.

fn data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PersistentModelIndexDataArgs<'largs>

C++ method: QPersistentModelIndex::data

This is an overloaded function. Available variants:

Rust arguments:
1) fn data(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn data(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
3) fn data(&self, cpp_utils::AsStruct) -> ::variant::Variant
4) fn data(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QPersistentModelIndex::data(int role = ?) const

C++ documentation:

Returns the data for the given role for the item referred to by the index.

See also Qt::ItemDataRole and QAbstractItemModel::setData().

fn flags(&self) -> Flags<ItemFlag>

C++ method: QFlags<Qt::ItemFlag> QPersistentModelIndex::flags() const

C++ documentation:

Returns the flags for the item referred to by the index.

This function was introduced in Qt 4.2.

fn internal_id(&self) -> usize

C++ method: quintptr QPersistentModelIndex::internalId() const

fn internal_pointer(&self) -> *mut c_void

C++ method: void* QPersistentModelIndex::internalPointer() const

fn is_valid(&self) -> bool

C++ method: bool QPersistentModelIndex::isValid() const

C++ documentation:

Returns true if this persistent model index is valid; otherwise returns false.

A valid index belongs to a model, and has non-negative row and column numbers.

See also model(), row(), and column().

fn model(&self) -> *const AbstractItemModel

C++ method: const QAbstractItemModel* QPersistentModelIndex::model() const

C++ documentation:

Returns the model that the index belongs to.

fn new<Args>(args: Args) -> Args::ReturnType where Args: PersistentModelIndexNewArgs

C++ method: QPersistentModelIndex::QPersistentModelIndex

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::persistent_model_index::PersistentModelIndex
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
C++ method: [constructor] void QPersistentModelIndex::QPersistentModelIndex()

Variant 2

Rust arguments:
1) fn new((&::model_index::ModelIndex, cpp_utils::AsStruct)) -> ::persistent_model_index::PersistentModelIndex
2) fn new((&::model_index::ModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
C++ method: [constructor] void QPersistentModelIndex::QPersistentModelIndex(const QModelIndex& index)

Variant 3

Rust arguments:
1) fn new((&::persistent_model_index::PersistentModelIndex, cpp_utils::AsStruct)) -> ::persistent_model_index::PersistentModelIndex
2) fn new((&::persistent_model_index::PersistentModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
C++ method: [constructor] void QPersistentModelIndex::QPersistentModelIndex(const QPersistentModelIndex& other)

fn op_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut PersistentModelIndex where Args: PersistentModelIndexOpAssignArgs<'largs>

C++ method: QPersistentModelIndex::operator=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_assign(&mut self, &'l1 ::model_index::ModelIndex) -> &'l0 mut ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex& QPersistentModelIndex::operator=(const QModelIndex& other)

C++ documentation:

Sets the persistent model index to refer to the same item in a model as the other model index.

Variant 2

Rust arguments: fn op_assign(&mut self, &'l1 ::persistent_model_index::PersistentModelIndex) -> &'l0 mut ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex& QPersistentModelIndex::operator=(const QPersistentModelIndex& other)

C++ documentation:

Sets the persistent model index to refer to the same item in a model as the other persistent model index.

fn op_eq<'largs, Args>(&'largs self, args: Args) -> bool where Args: PersistentModelIndexOpEqArgs<'largs>

C++ method: QPersistentModelIndex::operator==

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_eq(&self, &::model_index::ModelIndex) -> bool
C++ method: bool QPersistentModelIndex::operator==(const QModelIndex& other) const

C++ documentation:

Returns true if this persistent model index refers to the same location as the other model index; otherwise returns false.

All values in the persistent model index are used when comparing with another model index.

Variant 2

Rust arguments: fn op_eq(&self, &::persistent_model_index::PersistentModelIndex) -> bool
C++ method: bool QPersistentModelIndex::operator==(const QPersistentModelIndex& other) const

C++ documentation:

Returns true if this persistent model index is equal to the other persistent model index; otherwise returns false.

All values in the persistent model index are used when comparing with another persistent model index.

fn op_lt(&self, other: &PersistentModelIndex) -> bool

C++ method: bool QPersistentModelIndex::operator<(const QPersistentModelIndex& other) const

C++ documentation:

Returns true if this persistent model index is smaller than the other persistent model index; otherwise returns false.

All values in the persistent model index are used when comparing with another persistent model index.

This function was introduced in Qt 4.1.

fn op_neq<'largs, Args>(&'largs self, args: Args) -> bool where Args: PersistentModelIndexOpNeqArgs<'largs>

C++ method: QPersistentModelIndex::operator!=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_neq(&self, &::model_index::ModelIndex) -> bool
C++ method: bool QPersistentModelIndex::operator!=(const QModelIndex& other) const

C++ documentation:

Returns true if this persistent model index does not refer to the same location as the other model index; otherwise returns false.

Variant 2

Rust arguments: fn op_neq(&self, &::persistent_model_index::PersistentModelIndex) -> bool
C++ method: bool QPersistentModelIndex::operator!=(const QPersistentModelIndex& other) const

C++ documentation:

Returns true if this persistent model index is not equal to the other persistent model index; otherwise returns false.

This function was introduced in Qt 4.2.

fn parent<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PersistentModelIndexParentArgs<'largs>

C++ method: QPersistentModelIndex::parent

This is an overloaded function. Available variants:

Rust arguments:
1) fn parent(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn parent(&self, cpp_utils::AsStruct) -> ::model_index::ModelIndex
C++ method: QModelIndex QPersistentModelIndex::parent() const

C++ documentation:

Returns the parent QModelIndex for this persistent index, or an invalid QModelIndex if it has no parent.

See also child(), sibling(), and model().

fn row(&self) -> c_int

C++ method: int QPersistentModelIndex::row() const

C++ documentation:

Returns the row this persistent model index refers to.

fn sibling<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PersistentModelIndexSiblingArgs<'largs>

C++ method: QPersistentModelIndex::sibling

This is an overloaded function. Available variants:

Rust arguments:
1) fn sibling(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn sibling(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::model_index::ModelIndex
C++ method: QModelIndex QPersistentModelIndex::sibling(int row, int column) const

C++ documentation:

Returns the sibling at row and column or an invalid QModelIndex if there is no sibling at this position.

See also parent() and child().

fn swap(&mut self, other: &mut PersistentModelIndex)

C++ method: void QPersistentModelIndex::swap(QPersistentModelIndex& other)

C++ documentation:

Swaps this persistent modelindex with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.