Module qt_core::plugin_loader::overloading

Traits

PluginLoaderConnectArgs
This trait represents a set of arguments accepted by PluginLoader::connect method.

PluginLoaderConnectStaticArgs
This trait represents a set of arguments accepted by PluginLoader::connect_static method.

PluginLoaderDestroyedArgs
This trait represents a set of arguments accepted by PluginLoader::destroyed method.

PluginLoaderDisconnectArgs
This trait represents a set of arguments accepted by PluginLoader::disconnect method.

PluginLoaderDisconnectStaticArgs
This trait represents a set of arguments accepted by PluginLoader::disconnect_static method.

PluginLoaderDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by PluginLoader::dynamic_property_names method.

PluginLoaderErrorStringArgs
This trait represents a set of arguments accepted by PluginLoader::error_string method.

PluginLoaderFileNameArgs
This trait represents a set of arguments accepted by PluginLoader::file_name method.

PluginLoaderMetaDataArgs
This trait represents a set of arguments accepted by PluginLoader::meta_data method.

PluginLoaderNewArgs
This trait represents a set of arguments accepted by PluginLoader::new method.

PluginLoaderObjectNameArgs
This trait represents a set of arguments accepted by PluginLoader::object_name method.

PluginLoaderPropertyArgs
This trait represents a set of arguments accepted by PluginLoader::property method.

PluginLoaderStartTimerArgs
This trait represents a set of arguments accepted by PluginLoader::start_timer method.

PluginLoaderStaticInstancesArgs
This trait represents a set of arguments accepted by PluginLoader::static_instances method.

PluginLoaderStaticPluginsArgs
This trait represents a set of arguments accepted by PluginLoader::static_plugins method.

PluginLoaderTrArgs
This trait represents a set of arguments accepted by PluginLoader::tr method.

PluginLoaderTrUtf8Args
This trait represents a set of arguments accepted by PluginLoader::tr_utf8 method.