Struct qt_core::point::Point

pub struct Point {
    // some fields omitted
}

C++ type: QPoint

C++ documentation:

The QPoint class defines a point in the plane using integer precision.

A point is specified by a x coordinate and an y coordinate which can be accessed using the x() and y() functions. The isNull() function returns true if both x and y are set to 0. The coordinates can be set (or altered) using the setX() and setY() functions, or alternatively the rx() and ry() functions which return references to the coordinates (allowing direct manipulation).

Given a point p, the following statements are all equivalent:

QPoint p;

p.setX(p.x() + 1); p += QPoint(1, 0); p.rx()++;
A QPoint object can also be used as a vector: Addition and subtraction are defined as for vectors (each component is added separately). A QPoint object can also be divided or multiplied by an int or a qreal.

In addition, the QPoint class provides the manhattanLength() function which gives an inexpensive approximation of the length of the QPoint object interpreted as a vector. Finally, QPoint objects can be streamed as well as compared.

Methods

impl Point

fn dot_product(p1: &Point, p2: &Point) -> c_int

C++ method: static int QPoint::dotProduct(const QPoint& p1, const QPoint& p2)

C++ documentation:

QPoint p( 3, 7);
QPoint q(-1, 4);
int lengthSquared = QPoint::dotProduct(p, q);   // lengthSquared becomes 25

Returns the dot product of p1 and p2.

This function was introduced in Qt 5.1.

fn is_null(&self) -> bool

C++ method: bool QPoint::isNull() const

C++ documentation:

Returns true if both the x and y coordinates are set to 0, otherwise returns false.

fn manhattan_length(&self) -> c_int

C++ method: int QPoint::manhattanLength() const

C++ documentation:

Returns the sum of the absolute values of x() and y(), traditionally known as the "Manhattan length" of the vector from the origin to the point. For example:

QPoint oldPosition;

MyWidget::mouseMoveEvent(QMouseEvent *event) { QPoint point = event->pos() - oldPosition; if (point.manhattanLength() > 3) // the mouse has moved more than 3 pixels since the oldPosition }
This is a useful, and quick to calculate, approximation to the true length:

double trueLength = std::sqrt(std::pow(x(), 2) + std::pow(y(), 2));

The tradition of "Manhattan length" arises because such distances apply to travelers who can only travel on a rectangular grid, like the streets of Manhattan.

fn new<Args>(args: Args) -> Args::ReturnType where Args: PointNewArgs

C++ method: QPoint::QPoint

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::point::Point
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::point::Point>
C++ method: [constructor] void QPoint::QPoint()

Variant 2

Rust arguments:
1) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::point::Point
2) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::point::Point>
C++ method: [constructor] void QPoint::QPoint(int xpos, int ypos)

fn op_add_assign<'l0, 'l1>(&'l0 mut self, p: &'l1 Point) -> &'l0 mut Point

C++ method: QPoint& QPoint::operator+=(const QPoint& p)

C++ documentation:

Adds the given point to this point and returns a reference to this point. For example:

QPoint p( 3, 7);
QPoint q(-1, 4);
p += q;    // p becomes (2, 11)

See also operator-=().

fn op_div_assign<'l0>(&'l0 mut self, divisor: c_double) -> &'l0 mut Point

C++ method: QPoint& QPoint::operator/=(double divisor)

C++ documentation:

This is an overloaded function.

Divides both x and y by the given divisor, and returns a reference to this point. For example:

QPoint p(-3, 10);
p /= 2.5;           // p becomes (-1, 4)

Note that the result is rounded to the nearest integer as points are held as integers. Use QPointF for floating point accuracy.

See also operator*=().

fn op_mul_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut Point where Args: PointOpMulAssignArgs<'largs>

C++ method: QPoint::operator*=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_mul_assign(&mut self, libc::c_double) -> &'l0 mut ::point::Point
C++ method: QPoint& QPoint::operator*=(double factor)

C++ documentation:

Multiplies this point's coordinates by the given factor, and returns a reference to this point. For example:

QPoint p(-1, 4);
p *= 2.5;    // p becomes (-3, 10)

Note that the result is rounded to the nearest integer as points are held as integers. Use QPointF for floating point accuracy.

See also operator/=().

Variant 2

Rust arguments: fn op_mul_assign(&mut self, libc::c_float) -> &'l0 mut ::point::Point
C++ method: QPoint& QPoint::operator*=(float factor)

C++ documentation:

Multiplies this point's coordinates by the given factor, and returns a reference to this point.

Note that the result is rounded to the nearest integer as points are held as integers. Use QPointF for floating point accuracy.

See also operator/=().

Variant 3

Rust arguments: fn op_mul_assign(&mut self, libc::c_int) -> &'l0 mut ::point::Point
C++ method: QPoint& QPoint::operator*=(int factor)

C++ documentation:

Multiplies this point's coordinates by the given factor, and returns a reference to this point.

See also operator/=().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, p: &'l1 Point) -> &'l0 mut Point

C++ method: QPoint& QPoint::operator-=(const QPoint& p)

C++ documentation:

Subtracts the given point from this point and returns a reference to this point. For example:

QPoint p( 3, 7);
QPoint q(-1, 4);
p -= q;    // p becomes (4, 3)

See also operator+=().

fn rx<'l0>(&'l0 mut self) -> &'l0 mut c_int

C++ method: int& QPoint::rx()

C++ documentation:

Returns a reference to the x coordinate of this point.

Using a reference makes it possible to directly manipulate x. For example:

QPoint p(1, 2);
p.rx()--;   // p becomes (0, 2)

See also x() and setX().

fn ry<'l0>(&'l0 mut self) -> &'l0 mut c_int

C++ method: int& QPoint::ry()

C++ documentation:

Returns a reference to the y coordinate of this point.

Using a reference makes it possible to directly manipulate y. For example:

QPoint p(1, 2);
p.ry()++;   // p becomes (1, 3)

See also y() and setY().

fn set_x(&mut self, x: c_int)

C++ method: void QPoint::setX(int x)

C++ documentation:

Sets the x coordinate of this point to the given x coordinate.

See also x() and setY().

fn set_y(&mut self, y: c_int)

C++ method: void QPoint::setY(int y)

C++ documentation:

Sets the y coordinate of this point to the given y coordinate.

See also y() and setX().

fn x(&self) -> c_int

C++ method: int QPoint::x() const

C++ documentation:

Returns the x coordinate of this point.

See also setX() and rx().

fn y(&self) -> c_int

C++ method: int QPoint::y() const

C++ documentation:

Returns the y coordinate of this point.

See also setY() and ry().