Struct qt_core::point_f::PointF

pub struct PointF {
    // some fields omitted
}

C++ type: QPointF

C++ documentation:

The QPointF class defines a point in the plane using floating point precision.

A point is specified by a x coordinate and an y coordinate which can be accessed using the x() and y() functions. The coordinates of the point are specified using floating point numbers for accuracy. The isNull() function returns true if both x and y are set to 0.0. The coordinates can be set (or altered) using the setX() and setY() functions, or alternatively the rx() and ry() functions which return references to the coordinates (allowing direct manipulation).

Given a point p, the following statements are all equivalent:

QPointF p;

p.setX(p.x() + 1.0); p += QPointF(1.0, 0.0); p.rx()++;
A QPointF object can also be used as a vector: Addition and subtraction are defined as for vectors (each component is added separately). A QPointF object can also be divided or multiplied by an int or a qreal.

In addition, the QPointF class provides a constructor converting a QPoint object into a QPointF object, and a corresponding toPoint() function which returns a QPoint copy of this point. Finally, QPointF objects can be streamed as well as compared.

Methods

impl PointF

fn dot_product(p1: &PointF, p2: &PointF) -> c_double

C++ method: static double QPointF::dotProduct(const QPointF& p1, const QPointF& p2)

C++ documentation:

QPointF p( 3.1, 7.1);
QPointF q(-1.0, 4.1);
int lengthSquared = QPointF::dotProduct(p, q);   // lengthSquared becomes 26.01

Returns the dot product of p1 and p2.

This function was introduced in Qt 5.1.

fn is_null(&self) -> bool

C++ method: bool QPointF::isNull() const

C++ documentation:

Returns true if both the x and y coordinates are set to 0.0 (ignoring the sign); otherwise returns false.

fn manhattan_length(&self) -> c_double

C++ method: double QPointF::manhattanLength() const

C++ documentation:

Returns the sum of the absolute values of x() and y(), traditionally known as the "Manhattan length" of the vector from the origin to the point.

This function was introduced in Qt 4.6.

See also QPoint::manhattanLength().

fn new<Args>(args: Args) -> Args::ReturnType where Args: PointFNewArgs

C++ method: QPointF::QPointF

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::point_f::PointF
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::point_f::PointF>
C++ method: [constructor] void QPointF::QPointF()

Variant 2

Rust arguments:
1) fn new((&::point::Point, cpp_utils::AsStruct)) -> ::point_f::PointF
2) fn new((&::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<::point_f::PointF>
C++ method: [constructor] void QPointF::QPointF(const QPoint& p)

Variant 3

Rust arguments:
1) fn new((libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::point_f::PointF
2) fn new((libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::point_f::PointF>
C++ method: [constructor] void QPointF::QPointF(double xpos, double ypos)

fn op_add_assign<'l0, 'l1>(&'l0 mut self, p: &'l1 PointF) -> &'l0 mut PointF

C++ method: QPointF& QPointF::operator+=(const QPointF& p)

C++ documentation:

Adds the given point to this point and returns a reference to this point. For example:

QPointF p( 3.1, 7.1);
QPointF q(-1.0, 4.1);
p += q;    // p becomes (2.1, 11.2)

See also operator-=().

fn op_div_assign<'l0>(&'l0 mut self, c: c_double) -> &'l0 mut PointF

C++ method: QPointF& QPointF::operator/=(double c)

C++ documentation:

Divides both x and y by the given divisor, and returns a reference to this point. For example:

QPointF p(-2.75, 10.25);
p /= 2.5;           // p becomes (-1.1, 4.1)

See also operator*=().

fn op_mul_assign<'l0>(&'l0 mut self, c: c_double) -> &'l0 mut PointF

C++ method: QPointF& QPointF::operator*=(double c)

C++ documentation:

Multiplies this point's coordinates by the given factor, and returns a reference to this point. For example:

QPointF p(-1.1, 4.1);
p *= 2.5;    // p becomes (-2.75, 10.25)

See also operator/=().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, p: &'l1 PointF) -> &'l0 mut PointF

C++ method: QPointF& QPointF::operator-=(const QPointF& p)

C++ documentation:

Subtracts the given point from this point and returns a reference to this point. For example:

QPointF p( 3.1, 7.1);
QPointF q(-1.0, 4.1);
p -= q;    // p becomes (4.1, 3.0)

See also operator+=().

fn rx<'l0>(&'l0 mut self) -> &'l0 mut c_double

C++ method: double& QPointF::rx()

C++ documentation:

Returns a reference to the x coordinate of this point.

Using a reference makes it possible to directly manipulate x. For example:

 QPointF p(1.1, 2.5);
 p.rx()--;   // p becomes (0.1, 2.5)

See also x() and setX().

fn ry<'l0>(&'l0 mut self) -> &'l0 mut c_double

C++ method: double& QPointF::ry()

C++ documentation:

Returns a reference to the y coordinate of this point.

Using a reference makes it possible to directly manipulate y. For example:

QPointF p(1.1, 2.5);
p.ry()++;   // p becomes (1.1, 3.5)

See also y() and setY().

fn set_x(&mut self, x: c_double)

C++ method: void QPointF::setX(double x)

C++ documentation:

Sets the x coordinate of this point to the given x coordinate.

See also x() and setY().

fn set_y(&mut self, y: c_double)

C++ method: void QPointF::setY(double y)

C++ documentation:

Sets the y coordinate of this point to the given y coordinate.

See also y() and setX().

fn to_point<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: PointFToPointArgs<'largs>

C++ method: QPointF::toPoint

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::point::Point>
2) fn to_point(&self, cpp_utils::AsStruct) -> ::point::Point
C++ method: QPoint QPointF::toPoint() const

C++ documentation:

Rounds the coordinates of this point to the nearest integer, and returns a QPoint object with the rounded coordinates.

See also QPointF().

fn x(&self) -> c_double

C++ method: double QPointF::x() const

C++ documentation:

Returns the x coordinate of this point.

See also setX() and rx().

fn y(&self) -> c_double

C++ method: double QPointF::y() const

C++ documentation:

Returns the y coordinate of this point.

See also setY() and ry().