Enum qt_core::process::ProcessError

pub enum ProcessError {
    FailedToStart,
    Crashed,
    Timedout,
    ReadError,
    WriteError,
    UnknownError,
}

C++ type: QProcess::ProcessError

C++ documentation:

This enum describes the different types of errors that are reported by QProcess.

See also error().

Variants
FailedToStart

The process failed to start. Either the invoked program is missing, or you may have insufficient permissions to invoke the program. (C++ enum variant: FailedToStart = 0)

Crashed

The process crashed some time after starting successfully. (C++ enum variant: Crashed = 1)

Timedout

The last waitFor...() function timed out. The state of QProcess is unchanged, and you can try calling waitFor...() again. (C++ enum variant: Timedout = 2)

ReadError

An error occurred when attempting to read from the process. For example, the process may not be running. (C++ enum variant: ReadError = 3)

WriteError

An error occurred when attempting to write to the process. For example, the process may not be running, or it may have closed its input channel. (C++ enum variant: WriteError = 4)

UnknownError

An unknown error occurred. This is the default return value of error(). (C++ enum variant: UnknownError = 5)