Struct qt_core::process_environment::ProcessEnvironment

pub struct ProcessEnvironment {
    // some fields omitted
}

C++ type: QProcessEnvironment

C++ documentation:

The QProcessEnvironment class holds the environment variables that can be passed to a program.

A process's environment is composed of a set of key=value pairs known as environment variables. The QProcessEnvironment class wraps that concept and allows easy manipulation of those variables. It's meant to be used along with QProcess, to set the environment for child processes. It cannot be used to change the current process's environment.

The environment of the calling process can be obtained using QProcessEnvironment::systemEnvironment().

On Unix systems, the variable names are case-sensitive. Note that the Unix environment allows both variable names and contents to contain arbitrary binary data (except for the NUL character). QProcessEnvironment will preserve such variables, but does not support manipulating variables whose names or values are not encodable by the current locale settings (see QTextCodec::codecForLocale).

On Windows, the variable names are case-insensitive, but case-preserving. QProcessEnvironment behaves accordingly.

On Windows CE, the concept of environment does not exist. This class will keep the values set for compatibility with other platforms, but the values set will have no effect on the processes being created.

Methods

impl ProcessEnvironment

fn clear(&mut self)

C++ method: void QProcessEnvironment::clear()

C++ documentation:

Removes all key=value pairs from this QProcessEnvironment object, making it empty.

See also isEmpty() and systemEnvironment().

fn contains(&self, name: &String) -> bool

C++ method: bool QProcessEnvironment::contains(const QString& name) const

C++ documentation:

Returns true if the environment variable of name name is found in this QProcessEnvironment object.

See also insert() and value().

fn insert<'largs, Args>(&'largs mut self, args: Args) where Args: ProcessEnvironmentInsertArgs<'largs>

C++ method: QProcessEnvironment::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert(&mut self, &::process_environment::ProcessEnvironment) -> ()
C++ method: void QProcessEnvironment::insert(const QProcessEnvironment& e)

C++ documentation:

This is an overloaded function.

Inserts the contents of e in this QProcessEnvironment object. Variables in this object that also exist in e will be overwritten.

This function was introduced in Qt 4.8.

Variant 2

Rust arguments: fn insert(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QProcessEnvironment::insert(const QString& name, const QString& value)

C++ documentation:

Inserts the environment variable of name name and contents value into this QProcessEnvironment object. If that variable already existed, it is replaced by the new value.

On most systems, inserting a variable with no contents will have the same effect for applications as if the variable had not been set at all. However, to guarantee that there are no incompatibilities, to remove a variable, please use the remove() function.

See also contains(), remove(), and value().

fn is_empty(&self) -> bool

C++ method: bool QProcessEnvironment::isEmpty() const

C++ documentation:

Returns true if this QProcessEnvironment object is empty: that is there are no key=value pairs set.

See also clear(), systemEnvironment(), and insert().

fn keys<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ProcessEnvironmentKeysArgs<'largs>

C++ method: QProcessEnvironment::keys

This is an overloaded function. Available variants:

Rust arguments:
1) fn keys(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn keys(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QProcessEnvironment::keys() const

C++ documentation:

Returns a list containing all the variable names in this QProcessEnvironment object.

This function was introduced in Qt 4.8.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ProcessEnvironmentNewArgs

C++ method: QProcessEnvironment::QProcessEnvironment

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::process_environment::ProcessEnvironment
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::process_environment::ProcessEnvironment>
C++ method: [constructor] void QProcessEnvironment::QProcessEnvironment()

Variant 2

Rust arguments:
1) fn new((&::process_environment::ProcessEnvironment, cpp_utils::AsStruct)) -> ::process_environment::ProcessEnvironment
2) fn new((&::process_environment::ProcessEnvironment, cpp_utils::AsBox)) -> cpp_utils::CppBox<::process_environment::ProcessEnvironment>
C++ method: [constructor] void QProcessEnvironment::QProcessEnvironment(const QProcessEnvironment& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 ProcessEnvironment) -> &'l0 mut ProcessEnvironment

C++ method: QProcessEnvironment& QProcessEnvironment::operator=(const QProcessEnvironment& other)

C++ documentation:

Copies the contents of the other QProcessEnvironment object into this one.

fn op_eq(&self, other: &ProcessEnvironment) -> bool

C++ method: bool QProcessEnvironment::operator==(const QProcessEnvironment& other) const

C++ documentation:

Returns true if this and the other QProcessEnvironment objects are equal.

Two QProcessEnvironment objects are considered equal if they have the same set of key=value pairs. The comparison of keys is done case-sensitive on platforms where the environment is case-sensitive.

See also operator!=() and contains().

fn op_neq(&self, other: &ProcessEnvironment) -> bool

C++ method: bool QProcessEnvironment::operator!=(const QProcessEnvironment& other) const

C++ documentation:

Returns true if this and the other QProcessEnvironment objects are different.

See also operator==().

fn remove(&mut self, name: &String)

C++ method: void QProcessEnvironment::remove(const QString& name)

C++ documentation:

Removes the environment variable identified by name from this QProcessEnvironment object. If that variable did not exist before, nothing happens.

See also contains(), insert(), and value().

fn swap(&mut self, other: &mut ProcessEnvironment)

C++ method: void QProcessEnvironment::swap(QProcessEnvironment& other)

C++ documentation:

Swaps this process environment instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.0.

fn system_environment<Args>(args: Args) -> Args::ReturnType where Args: ProcessEnvironmentSystemEnvironmentArgs

C++ method: QProcessEnvironment::systemEnvironment

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_environment(cpp_utils::AsBox) -> cpp_utils::CppBox<::process_environment::ProcessEnvironment>
2) fn system_environment(cpp_utils::AsStruct) -> ::process_environment::ProcessEnvironment
C++ method: static QProcessEnvironment QProcessEnvironment::systemEnvironment()

C++ documentation:

The systemEnvironment function returns the environment of the calling process.

It is returned as a QProcessEnvironment. This function does not cache the system environment. Therefore, it's possible to obtain an updated version of the environment if low-level C library functions like setenv ot putenv have been called.

However, note that repeated calls to this function will recreate the QProcessEnvironment object, which is a non-trivial operation.

This function was introduced in Qt 4.6.

See also QProcess::systemEnvironment().

fn to_string_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ProcessEnvironmentToStringListArgs<'largs>

C++ method: QProcessEnvironment::toStringList

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn to_string_list(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QProcessEnvironment::toStringList() const

C++ documentation:

Converts this QProcessEnvironment object into a list of strings, one for each environment variable that is set. The environment variable's name and its value are separated by an equal character ('=').

The QStringList contents returned by this function are suitable for presentation. Use with the QProcess::setEnvironment function is not recommended due to potential encoding problems under Unix, and worse performance.

See also systemEnvironment(), QProcess::systemEnvironment(), QProcess::environment(), and QProcess::setEnvironment().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ProcessEnvironmentValueArgs<'largs>

C++ method: QProcessEnvironment::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn value(&self, (&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn value(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
4) fn value(&self, (&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QProcessEnvironment::value(const QString& name, const QString& defaultValue = ?) const

C++ documentation:

Searches this QProcessEnvironment object for a variable identified by name and returns its value. If the variable is not found in this object, then defaultValue is returned instead.

See also contains(), insert(), and remove().