Module qt_core::property_animation::overloading

Traits

PropertyAnimationConnectArgs
This trait represents a set of arguments accepted by PropertyAnimation::connect method.

PropertyAnimationConnectStaticArgs
This trait represents a set of arguments accepted by PropertyAnimation::connect_static method.

PropertyAnimationCurrentValueArgs
This trait represents a set of arguments accepted by PropertyAnimation::current_value method.

PropertyAnimationDestroyedArgs
This trait represents a set of arguments accepted by PropertyAnimation::destroyed method.

PropertyAnimationDisconnectArgs
This trait represents a set of arguments accepted by PropertyAnimation::disconnect method.

PropertyAnimationDisconnectStaticArgs
This trait represents a set of arguments accepted by PropertyAnimation::disconnect_static method.

PropertyAnimationDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by PropertyAnimation::dynamic_property_names method.

PropertyAnimationEasingCurveArgs
This trait represents a set of arguments accepted by PropertyAnimation::easing_curve method.

PropertyAnimationEndValueArgs
This trait represents a set of arguments accepted by PropertyAnimation::end_value method.

PropertyAnimationKeyValueAtArgs
This trait represents a set of arguments accepted by PropertyAnimation::key_value_at method.

PropertyAnimationKeyValuesArgs
This trait represents a set of arguments accepted by PropertyAnimation::key_values method.

PropertyAnimationNewArgs
This trait represents a set of arguments accepted by PropertyAnimation::new method.

PropertyAnimationObjectNameArgs
This trait represents a set of arguments accepted by PropertyAnimation::object_name method.

PropertyAnimationPropertyArgs
This trait represents a set of arguments accepted by PropertyAnimation::property method.

PropertyAnimationPropertyNameArgs
This trait represents a set of arguments accepted by PropertyAnimation::property_name method.

PropertyAnimationStartArgs
This trait represents a set of arguments accepted by PropertyAnimation::start method.

PropertyAnimationStartTimerArgs
This trait represents a set of arguments accepted by PropertyAnimation::start_timer method.

PropertyAnimationStartValueArgs
This trait represents a set of arguments accepted by PropertyAnimation::start_value method.

PropertyAnimationTrArgs
This trait represents a set of arguments accepted by PropertyAnimation::tr method.

PropertyAnimationTrUtf8Args
This trait represents a set of arguments accepted by PropertyAnimation::tr_utf8 method.