Enum qt_core::qiod_evice::OpenModeFlag

pub enum OpenModeFlag {
    NotOpen,
    ReadOnly,
    WriteOnly,
    ReadWrite,
    Append,
    Truncate,
    Text,
    Unbuffered,
}

C++ type: QIODevice::OpenModeFlag

C++ documentation:

This enum is used with open() to describe the mode in which a device is opened. It is also returned by openMode().

Certain flags, such as Unbuffered and Truncate, are meaningless when used with some subclasses. Some of these restrictions are implied by the type of device that is represented by a subclass. In other cases, the restriction may be due to the implementation, or may be imposed by the underlying platform; for example, QTcpSocket does not support Unbuffered mode, and limitations in the native API prevent QFile from supporting Unbuffered on Windows.

The OpenMode type is a typedef for QFlags<OpenModeFlag>. It stores an OR combination of OpenModeFlag values.

Variants
NotOpen

The device is not open. (C++ enum variant: NotOpen = 0)

ReadOnly

The device is open for reading. (C++ enum variant: ReadOnly = 1)

WriteOnly

The device is open for writing. Note that this mode implies Truncate. (C++ enum variant: WriteOnly = 2)

ReadWrite

The device is open for reading and writing. (C++ enum variant: ReadWrite = 3)

Append

The device is opened in append mode so that all data is written to the end of the file. (C++ enum variant: Append = 4)

Truncate

If possible, the device is truncated before it is opened. All earlier contents of the device are lost. (C++ enum variant: Truncate = 8)

Text

When reading, the end-of-line terminators are translated to '\n'. When writing, the end-of-line terminators are translated to the local encoding, for example '\r\n' for Win32. (C++ enum variant: Text = 16)

Unbuffered

Any buffer in the device is bypassed. (C++ enum variant: Unbuffered = 32)