Module qt_core::qiod_evice::overloading

Traits

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

QiodEviceConnectArgs
This trait represents a set of arguments accepted by QiodEvice::connect method.

QiodEviceConnectStaticArgs
This trait represents a set of arguments accepted by QiodEvice::connect_static method.

QiodEviceDestroyedArgs
This trait represents a set of arguments accepted by QiodEvice::destroyed method.

QiodEviceDisconnectArgs
This trait represents a set of arguments accepted by QiodEvice::disconnect method.

QiodEviceDisconnectStaticArgs
This trait represents a set of arguments accepted by QiodEvice::disconnect_static method.

QiodEviceDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by QiodEvice::dynamic_property_names method.

QiodEviceErrorStringArgs
This trait represents a set of arguments accepted by QiodEvice::error_string method.

QiodEviceObjectNameArgs
This trait represents a set of arguments accepted by QiodEvice::object_name method.

QiodEvicePeekArgs
This trait represents a set of arguments accepted by QiodEvice::peek method.

QiodEvicePropertyArgs
This trait represents a set of arguments accepted by QiodEvice::property method.

QiodEviceReadAllArgs
This trait represents a set of arguments accepted by QiodEvice::read_all method.

QiodEviceReadArgs
This trait represents a set of arguments accepted by QiodEvice::read method.

QiodEviceReadLineArgs
This trait represents a set of arguments accepted by QiodEvice::read_line method.

QiodEviceStartTimerArgs
This trait represents a set of arguments accepted by QiodEvice::start_timer method.

QiodEviceTrArgs
This trait represents a set of arguments accepted by QiodEvice::tr method.

QiodEviceTrUtf8Args
This trait represents a set of arguments accepted by QiodEvice::tr_utf8 method.

QiodEviceWriteArgs
This trait represents a set of arguments accepted by QiodEvice::write method.