Enum qt_core::qt::AlignmentFlag

pub enum AlignmentFlag {
    Left,
    Right,
    HCenter,
    Justify,
    Absolute,
    HorizontalMask,
    Top,
    Bottom,
    VCenter,
    Center,
    Baseline,
    VerticalMask,
}

C++ type: Qt::AlignmentFlag

C++ documentation:

This enum type is used to describe alignment. It contains horizontal and vertical flags that can be combined to produce the required effect.

The TextElideMode enum can also be used in many situations to fine-tune the appearance of aligned text.

The horizontal flags are:

The vertical flags are:

You can use only one of the horizontal flags at a time. There is one two-dimensional flag:

You can use at most one horizontal and one vertical flag at a time. Qt::AlignCenter counts as both horizontal and vertical.

Three enum values are useful in applications that can be run in right-to-left mode:

Masks:

Conflicting combinations of flags have undefined meanings.

The Alignment type is a typedef for QFlags<AlignmentFlag>. It stores an OR combination of AlignmentFlag values.

Variants
Left

This variant corresponds to multiple C++ enum variants with the same value:

AlignLeft = 1: Aligns with the left edge.

AlignLeading = 1: Synonym for Qt::AlignLeft.

Right

This variant corresponds to multiple C++ enum variants with the same value:

AlignRight = 2: Aligns with the right edge.

AlignTrailing = 2: Synonym for Qt::AlignRight.

HCenter

Centers horizontally in the available space. (C++ enum variant: AlignHCenter = 4)

Justify

Justifies the text in the available space. (C++ enum variant: AlignJustify = 8)

Absolute

If the widget's layout direction is Qt::RightToLeft (instead of Qt::LeftToRight, the default), Qt::AlignLeft refers to the right edge and Qt::AlignRight to the left edge. This is normally the desired behavior. If you want Qt::AlignLeft to always mean "left" and Qt::AlignRight to always mean "right", combine the flag with Qt::AlignAbsolute. (C++ enum variant: AlignAbsolute = 16)

HorizontalMask

C++ enum variant: AlignHorizontal_Mask = 31

Top

Aligns with the top. (C++ enum variant: AlignTop = 32)

Bottom

Aligns with the bottom. (C++ enum variant: AlignBottom = 64)

VCenter

Centers vertically in the available space. (C++ enum variant: AlignVCenter = 128)

Center

Centers in both dimensions. (C++ enum variant: AlignCenter = 132)

Baseline

Aligns with the baseline. (C++ enum variant: AlignBaseline = 256)

VerticalMask

C++ enum variant: AlignVertical_Mask = 480