Enum qt_core::qt::ApplicationAttribute

pub enum ApplicationAttribute {
    AAImmediateWidgetCreation,
    AAMSWindowsUseDirect3DByDefault,
    AADontShowIconsInMenus,
    AANativeWindows,
    AADontCreateNativeWidgetSiblings,
    AAMacPluginApplication,
    AADontUseNativeMenuBar,
    AAMacDontSwapCtrlAndMeta,
    AAUse96Dpi,
    AAX11InitThreads,
    AASynthesizeTouchForUnhandledMouseEvents,
    AASynthesizeMouseForUnhandledTouchEvents,
    AAUseHighDpiPixmaps,
    AAForceRasterWidgets,
    AAUseDesktopOpenGL,
    AAUseOpenGLES,
    AAUseSoftwareOpenGL,
    AAShareOpenGLContexts,
    AASetPalette,
    AAAttributeCount,
}

C++ type: Qt::ApplicationAttribute

C++ documentation:

This enum describes attributes that change the behavior of application-wide features. These are enabled and disabled using QCoreApplication::setAttribute(), and can be tested for with QCoreApplication::testAttribute().

After setting this attribute, application code that uses pixmap sizes in layout geometry calculations should typically divide by devicePixelRatio() to get device-independent layout geometry.

The following values are obsolete:

Variants
AAImmediateWidgetCreation

This attribute is no longer fully supported in Qt 5. It ensures that widgets are created as soon as they are constructed. By default, resources for widgets are allocated on demand to improve efficiency and minimize resource usage. Setting or clearing this attribute affects widgets constructed after the change. Setting it tells Qt to create toplevel windows immediately. Therefore, if it is important to minimize resource consumption, do not set this attribute. (C++ enum variant: AA_ImmediateWidgetCreation = 0)

AAMSWindowsUseDirect3DByDefault

This value is obsolete and has no effect. (C++ enum variant: AA_MSWindowsUseDirect3DByDefault = 1)

AADontShowIconsInMenus

Actions with the Icon property won't be shown in any menus unless specifically set by the QAction::iconVisibleInMenu property. Menus that are currently open or menus already created in the native OS X menubar may not pick up a change in this attribute. Changes in the QAction::iconVisibleInMenu property will always be picked up. (C++ enum variant: AA_DontShowIconsInMenus = 2)

AANativeWindows

Ensures that widgets have native windows. (C++ enum variant: AA_NativeWindows = 3)

AADontCreateNativeWidgetSiblings

Ensures that siblings of native widgets stay non-native unless specifically set by the Qt::WA_NativeWindow attribute. (C++ enum variant: AA_DontCreateNativeWidgetSiblings = 4)

AAMacPluginApplication

Stops the Qt mac application from doing specific initializations that do not necessarily make sense when using Qt to author a plugin. This includes avoiding loading our nib for the main menu and not taking possession of the native menu bar. When setting this attribute to true will also set the AA_DontUseNativeMenuBar attribute to true. It also disables native event filters. (C++ enum variant: AA_MacPluginApplication = 5)

AADontUseNativeMenuBar

All menubars created while this attribute is set to true won't be used as a native menubar (e.g, the menubar at the top of the main screen on OS X or at the bottom in Windows CE). (C++ enum variant: AA_DontUseNativeMenuBar = 6)

AAMacDontSwapCtrlAndMeta

On OS X by default, Qt swaps the Control and Meta (Command) keys (i.e., whenever Control is pressed, Qt sends Meta, and whenever Meta is pressed Control is sent). When this attribute is true, Qt will not do the flip. QKeySequence::StandardKey will also flip accordingly (i.e., QKeySequence::Copy will be Command+C on the keyboard regardless of the value set, though what is output for QKeySequence::toString(QKeySequence::PortableText) will be different). (C++ enum variant: AA_MacDontSwapCtrlAndMeta = 7)

AAUse96Dpi

Assume the screen has a resolution of 96 DPI rather than using the OS-provided resolution. This will cause font rendering to be consistent in pixels-per-point across devices rather than defining 1 point as 1/72 inch. (C++ enum variant: AA_Use96Dpi = 8)

AAX11InitThreads

Calls XInitThreads() as part of the QApplication construction in order to make Xlib calls thread-safe. This attribute must be set before QApplication is constructed. (C++ enum variant: AA_X11InitThreads = 10)

AASynthesizeTouchForUnhandledMouseEvents

All mouse events that are not accepted by the application will be translated to touch events instead. (C++ enum variant: AA_SynthesizeTouchForUnhandledMouseEvents = 11)

AASynthesizeMouseForUnhandledTouchEvents

All touch events that are not accepted by the application will be translated to left button mouse events instead. This attribute is enabled by default. (C++ enum variant: AA_SynthesizeMouseForUnhandledTouchEvents = 12)

AAUseHighDpiPixmaps

Make QIcon::pixmap() generate high-dpi pixmaps that can be larger than the requested size. Such pixmaps will have devicePixelRatio() set to a value higher than 1. (C++ enum variant: AA_UseHighDpiPixmaps = 13)

AAForceRasterWidgets

Make top-level widgets use pure raster surfaces, and do not support non-native GL-based child widgets. (C++ enum variant: AA_ForceRasterWidgets = 14)

AAUseDesktopOpenGL

Forces the usage of desktop OpenGL (for example, opengl32.dll or libGL.so) on platforms that use dynamic loading of the OpenGL implementation. This value has been added in Qt 5.3. (C++ enum variant: AA_UseDesktopOpenGL = 15)

AAUseOpenGLES

Forces the usage of OpenGL ES 2.0 or higher on platforms that use dynamic loading of the OpenGL implementation. This value has been added in Qt 5.3. (C++ enum variant: AA_UseOpenGLES = 16)

AAUseSoftwareOpenGL

Forces the usage of a software based OpenGL implementation on platforms that use dynamic loading of the OpenGL implementation. This will typically be a patched build of Mesa llvmpipe, providing OpenGL 2.1. The value may have no effect if no such OpenGL implementation is available. The default name of this library is opengl32sw.dll and can be overridden by setting the environment variable QT_OPENGL_DLL. See the platform-specific pages, for instance Qt for Windows, for more information. This value has been added in Qt 5.4. (C++ enum variant: AA_UseSoftwareOpenGL = 17)

AAShareOpenGLContexts

Enables resource sharing between the OpenGL contexts used by classes like QOpenGLWidget and QQuickWidget. This allows sharing OpenGL resources, like textures, between QOpenGLWidget instances that belong to different top-level windows. This value has been added in Qt 5.4. (C++ enum variant: AA_ShareOpenGLContexts = 18)

AASetPalette

Indicates whether a palette was explicitly set on the QApplication/QGuiApplication. This value has been added in Qt 5.5. (C++ enum variant: AA_SetPalette = 19)

AAAttributeCount

C++ enum variant: AA_AttributeCount = 20