Enum qt_core::qt::ApplicationState

pub enum ApplicationState {
    Suspended,
    Hidden,
    Inactive,
    Active,
}

C++ type: Qt::ApplicationState

C++ documentation:

This enum type is used to specify the current state of the application.

The states are

This enum was introduced or modified in Qt 5.1.

The ApplicationStates type is a typedef for QFlags<ApplicationState>. It stores an OR combination of ApplicationState values.

Variants
Suspended

The application is about to suspend. When entering this state, the application should save its state, cease all activities, and be prepared for code execution to stop. While suspended, the application can be killed at any time without further warnings (e.g. when low memory forces the OS to purge suspended applications). (C++ enum variant: ApplicationSuspended = 0)

Hidden

The application is hidden and runs in the background. This is the normal state for applications that need to do background processing, like playing music, while the user interacts with other applications. The application should free up all graphical resources when entering this state. (C++ enum variant: ApplicationHidden = 1)

Inactive

The application is visible, but not selected to be in front. On desktop platforms, this typically means that the user activated another application. On mobile platforms, it is more common to enter this state when the OS is interrupting the user with e.g. incoming calls or SMS-messages. While in this state, consider reducing CPU-intensive tasks. (C++ enum variant: ApplicationInactive = 2)

Active

The application is visible and selected to be in front. (C++ enum variant: ApplicationActive = 4)