Enum qt_core::qt::DateFormat

pub enum DateFormat {
    TextDate,
    IsodAte,
    SystemLocaleDate,
    LocaleDate,
    SystemLocaleShortDate,
    SystemLocaleLongDate,
    DefaultLocaleShortDate,
    DefaultLocaleLongDate,
    Rfc2822Date,
}

C++ type: Qt::DateFormat

C++ documentation:

Note: For ISODate formats, each Y, M and D represents a single digit of the year, month and day used to specify the date. Each H, M and S represents a single digit of the hour, minute and second used to specify the time. The presence of a literal T character is used to separate the date and time when both are specified.

Variants
TextDate

The default Qt format, which includes the day and month name, the day number in the month, and the year in full. The day and month names will be short, localized names. This is basically equivalent to using the date format string, "ddd MMM d yyyy". See QDate::toString() for more information. (C++ enum variant: TextDate = 0)

IsodAte

ISO 8601 extended format: either YYYY-MM-DD for dates or YYYY-MM-DDTHH:mm:ss, YYYY-MM-DDTHH:mm:ssTZD (e.g., 1997-07-16T19:20:30+01:00) for combined dates and times. (C++ enum variant: ISODate = 1)

SystemLocaleDate

This variant corresponds to multiple C++ enum variants with the same value:

SystemLocaleDate = 2: This enum value is deprecated. Use Qt::SystemLocaleShortDate instead (or Qt::SystemLocaleLongDate if you want long dates).

LocalDate = 2: This enum value is deprecated. Use Qt::SystemLocaleShortDate instead (or Qt::SystemLocaleLongDate if you want long dates).

LocaleDate

This enum value is deprecated. Use Qt::DefaultLocaleShortDate instead (or Qt::DefaultLocaleLongDate if you want long dates). (C++ enum variant: LocaleDate = 3)

SystemLocaleShortDate

The short format used by the operating system. (C++ enum variant: SystemLocaleShortDate = 4)

SystemLocaleLongDate

The long format used by the operating system. (C++ enum variant: SystemLocaleLongDate = 5)

DefaultLocaleShortDate

The short format specified by the application's locale. (C++ enum variant: DefaultLocaleShortDate = 6)

DefaultLocaleLongDate

The long format used by the application's locale. (C++ enum variant: DefaultLocaleLongDate = 7)

Rfc2822Date

RFC 2822, RFC 850 and RFC 1036 format: either [ddd,] dd MMM yyyy hh:mm[:ss] +/-TZ or ddd MMM dd yyyy hh:mm[:ss] +/-TZ for combined dates and times. (C++ enum variant: RFC2822Date = 8)