Enum qt_core::qt::EventPriority

pub enum EventPriority {
    Low,
    Normal,
    High,
}

C++ type: Qt::EventPriority

C++ documentation:

This enum can be used to specify event priorities.

Note that these values are provided purely for convenience, since event priorities can be any value between INT_MAX and INT_MIN, inclusive. For example, you can define custom priorities as being relative to each other:

enum CustomEventPriority
{
    // An important event
    ImportantEventPriority = Qt::HighEventPriority,

    // A more important event
    MoreImportantEventPriority = ImportantEventPriority + 1,

    // A critical event
    CriticalEventPriority = 100 * MoreImportantEventPriority,

    // Not that important
    StatusEventPriority = Qt::LowEventPriority,

    // These are less important than Status events
    IdleProcessingDoneEventPriority = StatusEventPriority - 1
};

See also QCoreApplication::postEvent().

Variants
Low

Events with this priority are sent after events with HighEventPriority or NormalEventPriority. (C++ enum variant: LowEventPriority = -1)

Normal

Events with this priority are sent after events with HighEventPriority, but before events with LowEventPriority. (C++ enum variant: NormalEventPriority = 0)

High

Events with this priority are sent before events with NormalEventPriority or LowEventPriority. (C++ enum variant: HighEventPriority = 1)