Enum qt_core::qt::ImageConversionFlag

pub enum ImageConversionFlag {
    AutoColor,
    MonoOnly,
    ColorModeMask,
    OrderedAlphaDither,
    DiffuseAlphaDither,
    AlphaDitherMask,
    OrderedDither,
    ThresholdDither,
    DitherMask,
    PreferDither,
    AvoidDither,
    DitherModeMask,
    NoOpaqueDetection,
    NoFormatConversion,
}

C++ type: Qt::ImageConversionFlag

C++ documentation:

The options marked "(default)" are set if no other values from the list are included (since the defaults are zero):

Color/Mono preference (ignored for QBitmap):

Dithering mode preference for RGB channels:

Dithering mode preference for alpha channel:

Color matching versus dithering preference:

The ImageConversionFlags type is a typedef for QFlags<ImageConversionFlag>. It stores an OR combination of ImageConversionFlag values.

Variants
AutoColor

This variant corresponds to multiple C++ enum variants with the same value:

AutoColor = 0: (default) - If the image has depth 1 and contains only black and white pixels, the pixmap becomes monochrome.

ThresholdAlphaDither = 0: (default) - No dithering.

DiffuseDither = 0: (default) - A high-quality dither.

AutoDither = 0

MonoOnly

The pixmap becomes monochrome. If necessary, it is dithered using the chosen dithering algorithm. (C++ enum variant: MonoOnly = 2)

ColorModeMask

This variant corresponds to multiple C++ enum variants with the same value:

ColorMode_Mask = 3

ColorOnly = 3: The pixmap is dithered/converted to the native display depth.

OrderedAlphaDither

A faster, more ordered dither. (C++ enum variant: OrderedAlphaDither = 4)

DiffuseAlphaDither

A high-quality dither. (C++ enum variant: DiffuseAlphaDither = 8)

AlphaDitherMask

This variant corresponds to multiple C++ enum variants with the same value:

AlphaDither_Mask = 12

NoAlpha = 12

OrderedDither

A faster, more ordered dither. (C++ enum variant: OrderedDither = 16)

ThresholdDither

No dithering; closest color is used. (C++ enum variant: ThresholdDither = 32)

DitherMask

C++ enum variant: Dither_Mask = 48

PreferDither

(default when converting to a pixmap) - Always dither 32-bit images when the image is converted to 8 bits. (C++ enum variant: PreferDither = 64)

AvoidDither

(default when converting for the purpose of saving to file) - Dither 32-bit images only if the image has more than 256 colors and it is being converted to 8 bits. (C++ enum variant: AvoidDither = 128)

DitherModeMask

C++ enum variant: DitherMode_Mask = 192

NoOpaqueDetection

Do not check whether the image contains non-opaque pixels. Use this if you know that the image is semi-transparent and you want to avoid the overhead of checking the pixels in the image until a non-opaque pixel is found, or if you want the pixmap to retain an alpha channel for some other reason. If the image has no alpha channel this flag has no effect. (C++ enum variant: NoOpaqueDetection = 256)

NoFormatConversion

Don't do any format conversions on the image. Can be useful when converting a QImage to a QPixmap for a one-time rendering operation for example. Note that a QPixmap not in the preferred format will be much slower as a paint device. (C++ enum variant: NoFormatConversion = 512)