Enum qt_core::qt::InputMethodHint

pub enum InputMethodHint {
    ExclusiveInputMask,
    None,
    HiddenText,
    SensitiveData,
    NoAutoUppercase,
    PreferNumbers,
    PreferUppercase,
    PreferLowercase,
    NoPredictiveText,
    Date,
    Time,
    PreferLatin,
    MultiLine,
    DigitsOnly,
    FormattedNumbersOnly,
    UppercaseOnly,
    LowercaseOnly,
    DialableCharactersOnly,
    EmailCharactersOnly,
    UrlCharactersOnly,
    LatinOnly,
}

C++ type: Qt::InputMethodHint

C++ documentation:

Flags that alter the behavior:

Flags that restrict input (exclusive flags):

Masks:

Note: If several exclusive flags are OR-ed together, the resulting character set will consist of the union of the specified sets. For instance specifying ImhNumbersOnly and ImhUppercaseOnly would yield a set consisting of numbers and uppercase letters.

The InputMethodHints type is a typedef for QFlags<InputMethodHint>. It stores an OR combination of InputMethodHint values.

See also QGraphicsItem::inputMethodHints().

Variants
ExclusiveInputMask

This mask yields nonzero if any of the exclusive flags are used. (C++ enum variant: ImhExclusiveInputMask = -65536)

None

No hints. (C++ enum variant: ImhNone = 0)

HiddenText

The input method should not show the characters while typing. This is automatically set when setting QLineEdit::echoMode to Password. Note that setting ImhHiddenText does not change the echo mode. (C++ enum variant: ImhHiddenText = 1)

SensitiveData

Typed text should not be stored by the active input method in any persistent storage like predictive user dictionary. (C++ enum variant: ImhSensitiveData = 2)

NoAutoUppercase

The input method should not try to automatically switch to upper case when a sentence ends. (C++ enum variant: ImhNoAutoUppercase = 4)

PreferNumbers

Numbers are preferred (but not required). (C++ enum variant: ImhPreferNumbers = 8)

PreferUppercase

Upper case letters are preferred (but not required). (C++ enum variant: ImhPreferUppercase = 16)

PreferLowercase

Lower case letters are preferred (but not required). (C++ enum variant: ImhPreferLowercase = 32)

NoPredictiveText

Do not use predictive text (i.e. dictionary lookup) while typing. (C++ enum variant: ImhNoPredictiveText = 64)

Date

The text editor functions as a date field. (C++ enum variant: ImhDate = 128)

Time

The text editor functions as a time field. (C++ enum variant: ImhTime = 256)

PreferLatin

Latin characters are preferred (but not required). (C++ enum variant: ImhPreferLatin = 512)

MultiLine

Multiple lines can be entered into the text field. (C++ enum variant: ImhMultiLine = 1024)

DigitsOnly

Only digits are allowed. (C++ enum variant: ImhDigitsOnly = 65536)

FormattedNumbersOnly

Only number input is allowed. This includes decimal point and minus sign. (C++ enum variant: ImhFormattedNumbersOnly = 131072)

UppercaseOnly

Only upper case letter input is allowed. (C++ enum variant: ImhUppercaseOnly = 262144)

LowercaseOnly

Only lower case letter input is allowed. (C++ enum variant: ImhLowercaseOnly = 524288)

DialableCharactersOnly

Only characters suitable for phone dialing are allowed. (C++ enum variant: ImhDialableCharactersOnly = 1048576)

EmailCharactersOnly

Only characters suitable for email addresses are allowed. (C++ enum variant: ImhEmailCharactersOnly = 2097152)

UrlCharactersOnly

Only characters suitable for URLs are allowed. (C++ enum variant: ImhUrlCharactersOnly = 4194304)

LatinOnly

Only latin based input is allowed. (C++ enum variant: ImhLatinOnly = 8388608)