Enum qt_core::qt::InputMethodQuery

pub enum InputMethodQuery {
    PlatformData,
    QueryAll,
    Enabled,
    CursorRectangle,
    Font,
    CursorPosition,
    SurroundingText,
    CurrentSelection,
    MaximumTextLength,
    AnchorPosition,
    QueryInput,
    Hints,
    PreferredLanguage,
    AbsolutePosition,
    TextBeforeCursor,
    TextAfterCursor,
}

C++ type: Qt::InputMethodQuery

C++ documentation:

Masks:

The InputMethodQueries type is a typedef for QFlags<InputMethodQuery>. It stores an OR combination of InputMethodQuery values.

See also QInputMethodQueryEvent.

Variants
PlatformData

Platform specific data for input method. (C++ enum variant: ImPlatformData = -2147483648)

QueryAll

Query for all input method properties. (C++ enum variant: ImQueryAll = -1)

Enabled

The widget accepts input method input. (C++ enum variant: ImEnabled = 1)

CursorRectangle

This variant corresponds to multiple C++ enum variants with the same value:

ImCursorRectangle = 2: The rectangle covering the area of the input cursor in widget coordinates.

ImMicroFocus = 2: This query is obsolete. Use ImCursorRectangle instead.

Font

The currently used font for text input. (C++ enum variant: ImFont = 4)

CursorPosition

The logical position of the cursor within the text surrounding the input area (see ImSurroundingText). (C++ enum variant: ImCursorPosition = 8)

SurroundingText

The plain text around the input area, for example the current paragraph. (C++ enum variant: ImSurroundingText = 16)

CurrentSelection

The currently selected text. (C++ enum variant: ImCurrentSelection = 32)

MaximumTextLength

The maximum number of characters that the widget can hold. If there is no limit, QVariant::QVariant() is returned. (C++ enum variant: ImMaximumTextLength = 64)

AnchorPosition

The position of the selection anchor. This may be less or greater than ImCursorPosition, depending on which side of selection the cursor is. If there is no selection, it returns the same as ImCursorPosition. (C++ enum variant: ImAnchorPosition = 128)

QueryInput

Commonly changed properties on input. (C++ enum variant: ImQueryInput = 186)

Hints

The hints for input method on expected input. (See Qt::InputMethodHints) (C++ enum variant: ImHints = 256)

PreferredLanguage

The preferred input language. (C++ enum variant: ImPreferredLanguage = 512)

AbsolutePosition

The logical position of the cursor within the entire document. (C++ enum variant: ImAbsolutePosition = 1024)

TextBeforeCursor

The plain text before the cursor. The widget can decide how much text to return, but must not return an empty string unless the cursor is at the start of the document. (C++ enum variant: ImTextBeforeCursor = 2048)

TextAfterCursor

The plain text after the cursor. The widget can decide how much text to return, but must not return an empty string unless the cursor is at the end of the document. (C++ enum variant: ImTextAfterCursor = 4096)