Enum qt_core::qt::ItemDataRole

pub enum ItemDataRole {
    Display,
    Decoration,
    Edit,
    ToolTip,
    StatusTip,
    WhatsThis,
    Font,
    TextAlignment,
    BackgroundColor,
    TextColor,
    CheckState,
    AccessibleText,
    AccessibleDescription,
    SizeHint,
    InitialSortOrder,
    DisplayProperty,
    DecorationProperty,
    ToolTipProperty,
    StatusTipProperty,
    WhatsThisProperty,
    User,
}

C++ type: Qt::ItemDataRole

C++ documentation:

Each item in the model has a set of data elements associated with it, each with its own role. The roles are used by the view to indicate to the model which type of data it needs. Custom models should return data in these types.

The general purpose roles (and the associated types) are:

Roles describing appearance and meta data (with associated types):

Accessibility roles (with associated types):

User roles:

For user roles, it is up to the developer to decide which types to use and ensure that components use the correct types when accessing and setting data.

Variants
Display

The key data to be rendered in the form of text. (QString) (C++ enum variant: DisplayRole = 0)

Decoration

The data to be rendered as a decoration in the form of an icon. (QColor, QIcon or QPixmap) (C++ enum variant: DecorationRole = 1)

Edit

The data in a form suitable for editing in an editor. (QString) (C++ enum variant: EditRole = 2)

ToolTip

The data displayed in the item's tooltip. (QString) (C++ enum variant: ToolTipRole = 3)

StatusTip

The data displayed in the status bar. (QString) (C++ enum variant: StatusTipRole = 4)

WhatsThis

The data displayed for the item in "What's This?" mode. (QString) (C++ enum variant: WhatsThisRole = 5)

Font

The font used for items rendered with the default delegate. (QFont) (C++ enum variant: FontRole = 6)

TextAlignment

The alignment of the text for items rendered with the default delegate. (Qt::AlignmentFlag) (C++ enum variant: TextAlignmentRole = 7)

BackgroundColor

This variant corresponds to multiple C++ enum variants with the same value:

BackgroundColorRole = 8: This role is obsolete. Use BackgroundRole instead.

BackgroundRole = 8: The background brush used for items rendered with the default delegate. (QBrush)

TextColor

This variant corresponds to multiple C++ enum variants with the same value:

TextColorRole = 9: This role is obsolete. Use ForegroundRole instead.

ForegroundRole = 9: The foreground brush (text color, typically) used for items rendered with the default delegate. (QBrush)

CheckState

This role is used to obtain the checked state of an item. (Qt::CheckState) (C++ enum variant: CheckStateRole = 10)

AccessibleText

The text to be used by accessibility extensions and plugins, such as screen readers. (QString) (C++ enum variant: AccessibleTextRole = 11)

AccessibleDescription

A description of the item for accessibility purposes. (QString) (C++ enum variant: AccessibleDescriptionRole = 12)

SizeHint

The size hint for the item that will be supplied to views. (QSize) (C++ enum variant: SizeHintRole = 13)

InitialSortOrder

This role is used to obtain the initial sort order of a header view section. (Qt::SortOrder). This role was introduced in Qt 4.8. (C++ enum variant: InitialSortOrderRole = 14)

DisplayProperty

C++ enum variant: DisplayPropertyRole = 27

DecorationProperty

C++ enum variant: DecorationPropertyRole = 28

ToolTipProperty

C++ enum variant: ToolTipPropertyRole = 29

StatusTipProperty

C++ enum variant: StatusTipPropertyRole = 30

WhatsThisProperty

C++ enum variant: WhatsThisPropertyRole = 31

User

The first role that can be used for application-specific purposes. (C++ enum variant: UserRole = 256)