Enum qt_core::qt::ItemFlag

pub enum ItemFlag {
    NoItemFlags,
    ItemIsSelectable,
    ItemIsEditable,
    ItemIsDragEnabled,
    ItemIsDropEnabled,
    ItemIsUserCheckable,
    ItemIsEnabled,
    ItemIsTristate,
    ItemNeverHasChildren,
    ItemIsUserTristate,
}

C++ type: Qt::ItemFlag

C++ documentation:

This enum describes the properties of an item:

Note that checkable items need to be given both a suitable set of flags and an initial state, indicating whether the item is checked or not. This is handled automatically for model/view components, but needs to be explicitly set for instances of QListWidgetItem, QTableWidgetItem, and QTreeWidgetItem.

Note that it is undefined behavior to reimplement QAbstractItemModel::hasChildren to return true for an index if that index has the Qt::ItemNeverHasChildren flag set.

The ItemFlags type is a typedef for QFlags<ItemFlag>. It stores an OR combination of ItemFlag values.

See also QAbstractItemModel.

Variants
NoItemFlags

It does not have any properties set. (C++ enum variant: NoItemFlags = 0)

ItemIsSelectable

It can be selected. (C++ enum variant: ItemIsSelectable = 1)

ItemIsEditable

It can be edited. (C++ enum variant: ItemIsEditable = 2)

ItemIsDragEnabled

It can be dragged. (C++ enum variant: ItemIsDragEnabled = 4)

ItemIsDropEnabled

It can be used as a drop target. (C++ enum variant: ItemIsDropEnabled = 8)

ItemIsUserCheckable

It can be checked or unchecked by the user. (C++ enum variant: ItemIsUserCheckable = 16)

ItemIsEnabled

The user can interact with the item. (C++ enum variant: ItemIsEnabled = 32)

ItemIsTristate

The item can show three separate states. This enables automatic management of the state of parent items in QTreeWidget (checked if all children are checked, unchecked if all children are unchecked, or partially checked if only some children are checked). (C++ enum variant: ItemIsTristate = 64)

ItemNeverHasChildren

The item never has child items. (C++ enum variant: ItemNeverHasChildren = 128)

ItemIsUserTristate

The user can cycle through three separate states. This value has been added in Qt 5.5. (C++ enum variant: ItemIsUserTristate = 256)