Enum qt_core::qt::Key

pub enum Key {
    KeySpace,
    KeyExclam,
    KeyQuoteDbl,
    KeyNumberSign,
    KeyDollar,
    KeyPercent,
    KeyAmpersand,
    KeyApostrophe,
    KeyParenLeft,
    KeyParenRight,
    KeyAsterisk,
    KeyPlus,
    KeyComma,
    KeyMinus,
    KeyPeriod,
    KeySlash,
    Key0,
    Key1,
    Key2,
    Key3,
    Key4,
    Key5,
    Key6,
    Key7,
    Key8,
    Key9,
    KeyColon,
    KeySemicolon,
    KeyLess,
    KeyEqual,
    KeyGreater,
    KeyQuestion,
    KeyAt,
    KeyA,
    KeyB,
    KeyC,
    KeyD,
    KeyE,
    KeyF,
    KeyG,
    KeyH,
    KeyI,
    KeyJ,
    KeyK,
    KeyL,
    KeyM,
    KeyN,
    KeyO,
    KeyP,
    KeyQ,
    KeyR,
    KeyS,
    KeyT,
    KeyU,
    KeyV,
    KeyW,
    KeyX,
    KeyY,
    KeyZ,
    KeyBracketLeft,
    KeyBackslash,
    KeyBracketRight,
    KeyAsciiCircum,
    KeyUnderscore,
    KeyQuoteLeft,
    KeyBraceLeft,
    KeyBar,
    KeyBraceRight,
    KeyAsciiTilde,
    KeyNobreakspace,
    KeyExclamdown,
    KeyCent,
    KeySterling,
    KeyCurrency,
    KeyYen,
    KeyBrokenbar,
    KeySection,
    KeyDiaeresis,
    KeyCopyright,
    KeyOrdfeminine,
    KeyGuillemotleft,
    KeyNotsign,
    KeyHyphen,
    KeyRegistered,
    KeyMacron,
    KeyDegree,
    KeyPlusminus,
    KeyTwosuperior,
    KeyThreesuperior,
    KeyAcute,
    KeyMu,
    KeyParagraph,
    KeyPeriodcentered,
    KeyCedilla,
    KeyOnesuperior,
    KeyMasculine,
    KeyGuillemotright,
    KeyOnequarter,
    KeyOnehalf,
    KeyThreequarters,
    KeyQuestiondown,
    KeyAgrave,
    KeyAacute,
    KeyAcircumflex,
    KeyAtilde,
    KeyAdiaeresis,
    KeyAring,
    KeyAE,
    KeyCcedilla,
    KeyEgrave,
    KeyEacute,
    KeyEcircumflex,
    KeyEdiaeresis,
    KeyIgrave,
    KeyIacute,
    KeyIcircumflex,
    KeyIdiaeresis,
    KeyETH,
    KeyNtilde,
    KeyOgrave,
    KeyOacute,
    KeyOcircumflex,
    KeyOtilde,
    KeyOdiaeresis,
    KeyMultiply,
    KeyOoblique,
    KeyUgrave,
    KeyUacute,
    KeyUcircumflex,
    KeyUdiaeresis,
    KeyYacute,
    KeyTHORN,
    KeySsharp,
    KeyDivision,
    KeyYdiaeresis,
    KeyEscape,
    KeyTab,
    KeyBacktab,
    KeyBackspace,
    KeyReturn,
    KeyEnter,
    KeyInsert,
    KeyDelete,
    KeyPause,
    KeyPrint,
    KeySysReq,
    KeyClear,
    KeyHome,
    KeyEnd,
    KeyLeft,
    KeyUp,
    KeyRight,
    KeyDown,
    KeyPageUp,
    KeyPageDown,
    KeyShift,
    KeyControl,
    KeyMeta,
    KeyAlt,
    KeyCapsLock,
    KeyNumLock,
    KeyScrollLock,
    KeyF1,
    KeyF2,
    KeyF3,
    KeyF4,
    KeyF5,
    KeyF6,
    KeyF7,
    KeyF8,
    KeyF9,
    KeyF10,
    KeyF11,
    KeyF12,
    KeyF13,
    KeyF14,
    KeyF15,
    KeyF16,
    KeyF17,
    KeyF18,
    KeyF19,
    KeyF20,
    KeyF21,
    KeyF22,
    KeyF23,
    KeyF24,
    KeyF25,
    KeyF26,
    KeyF27,
    KeyF28,
    KeyF29,
    KeyF30,
    KeyF31,
    KeyF32,
    KeyF33,
    KeyF34,
    KeyF35,
    KeySuperL,
    KeySuperR,
    KeyMenu,
    KeyHyperL,
    KeyHyperR,
    KeyHelp,
    KeyDirectionL,
    KeyDirectionR,
    KeyBack,
    KeyForward,
    KeyStop,
    KeyRefresh,
    KeyVolumeDown,
    KeyVolumeMute,
    KeyVolumeUp,
    KeyBassBoost,
    KeyBassUp,
    KeyBassDown,
    KeyTrebleUp,
    KeyTrebleDown,
    KeyMediaPlay,
    KeyMediaStop,
    KeyMediaPrevious,
    KeyMediaNext,
    KeyMediaRecord,
    KeyMediaPause,
    KeyMediaTogglePlayPause,
    KeyHomePage,
    KeyFavorites,
    KeySearch,
    KeyStandby,
    KeyOpenUrl,
    KeyLaunchMail,
    KeyLaunchMedia,
    KeyLaunch0,
    KeyLaunch1,
    KeyLaunch2,
    KeyLaunch3,
    KeyLaunch4,
    KeyLaunch5,
    KeyLaunch6,
    KeyLaunch7,
    KeyLaunch8,
    KeyLaunch9,
    KeyLaunchA,
    KeyLaunchB,
    KeyLaunchC,
    KeyLaunchD,
    KeyLaunchE,
    KeyLaunchF,
    KeyMonBrightnessUp,
    KeyMonBrightnessDown,
    KeyKeyboardLightOnOff,
    KeyKeyboardBrightnessUp,
    KeyKeyboardBrightnessDown,
    KeyPowerOff,
    KeyWakeUp,
    KeyEject,
    KeyScreenSaver,
    KeyWWW,
    KeyMemo,
    KeyLightBulb,
    KeyShop,
    KeyHistory,
    KeyAddFavorite,
    KeyHotLinks,
    KeyBrightnessAdjust,
    KeyFinance,
    KeyCommunity,
    KeyAudioRewind,
    KeyBackForward,
    KeyApplicationLeft,
    KeyApplicationRight,
    KeyBook,
    KeyCD,
    KeyCalculator,
    KeyToDoList,
    KeyClearGrab,
    KeyClose,
    KeyCopy,
    KeyCut,
    KeyDisplay,
    KeyDOS,
    KeyDocuments,
    KeyExcel,
    KeyExplorer,
    KeyGame,
    KeyGo,
    KeyITouch,
    KeyLogOff,
    KeyMarket,
    KeyMeeting,
    KeyMenuKB,
    KeyMenuPB,
    KeyMySites,
    KeyNews,
    KeyOfficeHome,
    KeyOption,
    KeyPaste,
    KeyPhone,
    KeyCalendar,
    KeyReply,
    KeyReload,
    KeyRotateWindows,
    KeyRotationPB,
    KeyRotationKB,
    KeySave,
    KeySend,
    KeySpell,
    KeySplitScreen,
    KeySupport,
    KeyTaskPane,
    KeyTerminal,
    KeyTools,
    KeyTravel,
    KeyVideo,
    KeyWord,
    KeyXfer,
    KeyZoomIn,
    KeyZoomOut,
    KeyAway,
    KeyMessenger,
    KeyWebCam,
    KeyMailForward,
    KeyPictures,
    KeyMusic,
    KeyBattery,
    KeyBluetooth,
    KeyWLAN,
    KeyUWB,
    KeyAudioForward,
    KeyAudioRepeat,
    KeyAudioRandomPlay,
    KeySubtitle,
    KeyAudioCycleTrack,
    KeyTime,
    KeyHibernate,
    KeyView,
    KeyTopMenu,
    KeyPowerDown,
    KeySuspend,
    KeyContrastAdjust,
    KeyLaunchG,
    KeyLaunchH,
    KeyTouchpadToggle,
    KeyTouchpadOn,
    KeyTouchpadOff,
    KeyMicMute,
    KeyRed,
    KeyGreen,
    KeyYellow,
    KeyBlue,
    KeyChannelUp,
    KeyChannelDown,
    KeyGuide,
    KeyInfo,
    KeySettings,
    KeyMicVolumeUp,
    KeyMicVolumeDown,
    KeyNew,
    KeyOpen,
    KeyFind,
    KeyUndo,
    KeyRedo,
    KeyAltGr,
    KeyMultiKey,
    KeyKanji,
    KeyMuhenkan,
    KeyHenkan,
    KeyRomaji,
    KeyHiragana,
    KeyKatakana,
    KeyHiraganaKatakana,
    KeyZenkaku,
    KeyHankaku,
    KeyZenkakuHankaku,
    KeyTouroku,
    KeyMassyo,
    KeyKanaLock,
    KeyKanaShift,
    KeyEisuShift,
    KeyEisuToggle,
    KeyHangul,
    KeyHangulStart,
    KeyHangulEnd,
    KeyHangulHanja,
    KeyHangulJamo,
    KeyHangulRomaja,
    KeyCodeinput,
    KeyHangulJeonja,
    KeyHangulBanja,
    KeyHangulPreHanja,
    KeyHangulPostHanja,
    KeySingleCandidate,
    KeyMultipleCandidate,
    KeyPreviousCandidate,
    KeyHangulSpecial,
    KeyModeSwitch,
    KeyDeadGrave,
    KeyDeadAcute,
    KeyDeadCircumflex,
    KeyDeadTilde,
    KeyDeadMacron,
    KeyDeadBreve,
    KeyDeadAbovedot,
    KeyDeadDiaeresis,
    KeyDeadAbovering,
    KeyDeadDoubleacute,
    KeyDeadCaron,
    KeyDeadCedilla,
    KeyDeadOgonek,
    KeyDeadIota,
    KeyDeadVoicedSound,
    KeyDeadSemivoicedSound,
    KeyDeadBelowdot,
    KeyDeadHook,
    KeyDeadHorn,
    KeyMediaLast,
    KeySelect,
    KeyYes,
    KeyNo,
    KeyCancel,
    KeyPrinter,
    KeyExecute,
    KeySleep,
    KeyPlay,
    KeyZoom,
    KeyExit,
    KeyContext1,
    KeyContext2,
    KeyContext3,
    KeyContext4,
    KeyCall,
    KeyHangup,
    KeyFlip,
    KeyToggleCallHangup,
    KeyVoiceDial,
    KeyLastNumberRedial,
    KeyCamera,
    KeyCameraFocus,
    KeyUnknown,
}

C++ type: Qt::Key

C++ documentation:

The key names used by Qt.

See also QKeyEvent::key().

Variants
KeySpace

This variant corresponds to multiple C++ enum variants with the same value:

Key_Space = 32:

Key_Any = 32:

KeyExclam

(C++ enum variant: Key_Exclam = 33)

KeyQuoteDbl

(C++ enum variant: Key_QuoteDbl = 34)

KeyNumberSign

(C++ enum variant: Key_NumberSign = 35)

KeyDollar

(C++ enum variant: Key_Dollar = 36)

KeyPercent

(C++ enum variant: Key_Percent = 37)

KeyAmpersand

(C++ enum variant: Key_Ampersand = 38)

KeyApostrophe

(C++ enum variant: Key_Apostrophe = 39)

KeyParenLeft

(C++ enum variant: Key_ParenLeft = 40)

KeyParenRight

(C++ enum variant: Key_ParenRight = 41)

KeyAsterisk

(C++ enum variant: Key_Asterisk = 42)

KeyPlus

(C++ enum variant: Key_Plus = 43)

KeyComma

(C++ enum variant: Key_Comma = 44)

KeyMinus

(C++ enum variant: Key_Minus = 45)

KeyPeriod

(C++ enum variant: Key_Period = 46)

KeySlash

(C++ enum variant: Key_Slash = 47)

Key0

(C++ enum variant: Key_0 = 48)

Key1

(C++ enum variant: Key_1 = 49)

Key2

(C++ enum variant: Key_2 = 50)

Key3

(C++ enum variant: Key_3 = 51)

Key4

(C++ enum variant: Key_4 = 52)

Key5

(C++ enum variant: Key_5 = 53)

Key6

(C++ enum variant: Key_6 = 54)

Key7

(C++ enum variant: Key_7 = 55)

Key8

(C++ enum variant: Key_8 = 56)

Key9

(C++ enum variant: Key_9 = 57)

KeyColon

(C++ enum variant: Key_Colon = 58)

KeySemicolon

(C++ enum variant: Key_Semicolon = 59)

KeyLess

(C++ enum variant: Key_Less = 60)

KeyEqual

(C++ enum variant: Key_Equal = 61)

KeyGreater

(C++ enum variant: Key_Greater = 62)

KeyQuestion

(C++ enum variant: Key_Question = 63)

KeyAt

(C++ enum variant: Key_At = 64)

KeyA

(C++ enum variant: Key_A = 65)

KeyB

(C++ enum variant: Key_B = 66)

KeyC

(C++ enum variant: Key_C = 67)

KeyD

(C++ enum variant: Key_D = 68)

KeyE

(C++ enum variant: Key_E = 69)

KeyF

(C++ enum variant: Key_F = 70)

KeyG

(C++ enum variant: Key_G = 71)

KeyH

(C++ enum variant: Key_H = 72)

KeyI

(C++ enum variant: Key_I = 73)

KeyJ

(C++ enum variant: Key_J = 74)

KeyK

(C++ enum variant: Key_K = 75)

KeyL

(C++ enum variant: Key_L = 76)

KeyM

(C++ enum variant: Key_M = 77)

KeyN

(C++ enum variant: Key_N = 78)

KeyO

(C++ enum variant: Key_O = 79)

KeyP

(C++ enum variant: Key_P = 80)

KeyQ

(C++ enum variant: Key_Q = 81)

KeyR

(C++ enum variant: Key_R = 82)

KeyS

(C++ enum variant: Key_S = 83)

KeyT

(C++ enum variant: Key_T = 84)

KeyU

(C++ enum variant: Key_U = 85)

KeyV

(C++ enum variant: Key_V = 86)

KeyW

(C++ enum variant: Key_W = 87)

KeyX

(C++ enum variant: Key_X = 88)

KeyY

(C++ enum variant: Key_Y = 89)

KeyZ

(C++ enum variant: Key_Z = 90)

KeyBracketLeft

(C++ enum variant: Key_BracketLeft = 91)

KeyBackslash

(C++ enum variant: Key_Backslash = 92)

KeyBracketRight

(C++ enum variant: Key_BracketRight = 93)

KeyAsciiCircum

(C++ enum variant: Key_AsciiCircum = 94)

KeyUnderscore

(C++ enum variant: Key_Underscore = 95)

KeyQuoteLeft

(C++ enum variant: Key_QuoteLeft = 96)

KeyBraceLeft

(C++ enum variant: Key_BraceLeft = 123)

KeyBar

(C++ enum variant: Key_Bar = 124)

KeyBraceRight

(C++ enum variant: Key_BraceRight = 125)

KeyAsciiTilde

(C++ enum variant: Key_AsciiTilde = 126)

KeyNobreakspace

(C++ enum variant: Key_nobreakspace = 160)

KeyExclamdown

(C++ enum variant: Key_exclamdown = 161)

KeyCent

(C++ enum variant: Key_cent = 162)

KeySterling

(C++ enum variant: Key_sterling = 163)

KeyCurrency

(C++ enum variant: Key_currency = 164)

KeyYen

(C++ enum variant: Key_yen = 165)

KeyBrokenbar

(C++ enum variant: Key_brokenbar = 166)

KeySection

(C++ enum variant: Key_section = 167)

KeyDiaeresis

(C++ enum variant: Key_diaeresis = 168)

KeyCopyright

(C++ enum variant: Key_copyright = 169)

KeyOrdfeminine

(C++ enum variant: Key_ordfeminine = 170)

KeyGuillemotleft

(C++ enum variant: Key_guillemotleft = 171)

KeyNotsign

(C++ enum variant: Key_notsign = 172)

KeyHyphen

(C++ enum variant: Key_hyphen = 173)

KeyRegistered

(C++ enum variant: Key_registered = 174)

KeyMacron

(C++ enum variant: Key_macron = 175)

KeyDegree

(C++ enum variant: Key_degree = 176)

KeyPlusminus

(C++ enum variant: Key_plusminus = 177)

KeyTwosuperior

(C++ enum variant: Key_twosuperior = 178)

KeyThreesuperior

(C++ enum variant: Key_threesuperior = 179)

KeyAcute

(C++ enum variant: Key_acute = 180)

KeyMu

(C++ enum variant: Key_mu = 181)

KeyParagraph

(C++ enum variant: Key_paragraph = 182)

KeyPeriodcentered

(C++ enum variant: Key_periodcentered = 183)

KeyCedilla

(C++ enum variant: Key_cedilla = 184)

KeyOnesuperior

(C++ enum variant: Key_onesuperior = 185)

KeyMasculine

(C++ enum variant: Key_masculine = 186)

KeyGuillemotright

(C++ enum variant: Key_guillemotright = 187)

KeyOnequarter

(C++ enum variant: Key_onequarter = 188)

KeyOnehalf

(C++ enum variant: Key_onehalf = 189)

KeyThreequarters

(C++ enum variant: Key_threequarters = 190)

KeyQuestiondown

(C++ enum variant: Key_questiondown = 191)

KeyAgrave

(C++ enum variant: Key_Agrave = 192)

KeyAacute

(C++ enum variant: Key_Aacute = 193)

KeyAcircumflex

(C++ enum variant: Key_Acircumflex = 194)

KeyAtilde

(C++ enum variant: Key_Atilde = 195)

KeyAdiaeresis

(C++ enum variant: Key_Adiaeresis = 196)

KeyAring

(C++ enum variant: Key_Aring = 197)

KeyAE

(C++ enum variant: Key_AE = 198)

KeyCcedilla

(C++ enum variant: Key_Ccedilla = 199)

KeyEgrave

(C++ enum variant: Key_Egrave = 200)

KeyEacute

(C++ enum variant: Key_Eacute = 201)

KeyEcircumflex

(C++ enum variant: Key_Ecircumflex = 202)

KeyEdiaeresis

(C++ enum variant: Key_Ediaeresis = 203)

KeyIgrave

(C++ enum variant: Key_Igrave = 204)

KeyIacute

(C++ enum variant: Key_Iacute = 205)

KeyIcircumflex

(C++ enum variant: Key_Icircumflex = 206)

KeyIdiaeresis

(C++ enum variant: Key_Idiaeresis = 207)

KeyETH

(C++ enum variant: Key_ETH = 208)

KeyNtilde

(C++ enum variant: Key_Ntilde = 209)

KeyOgrave

(C++ enum variant: Key_Ograve = 210)

KeyOacute

(C++ enum variant: Key_Oacute = 211)

KeyOcircumflex

(C++ enum variant: Key_Ocircumflex = 212)

KeyOtilde

(C++ enum variant: Key_Otilde = 213)

KeyOdiaeresis

(C++ enum variant: Key_Odiaeresis = 214)

KeyMultiply

(C++ enum variant: Key_multiply = 215)

KeyOoblique

(C++ enum variant: Key_Ooblique = 216)

KeyUgrave

(C++ enum variant: Key_Ugrave = 217)

KeyUacute

(C++ enum variant: Key_Uacute = 218)

KeyUcircumflex

(C++ enum variant: Key_Ucircumflex = 219)

KeyUdiaeresis

(C++ enum variant: Key_Udiaeresis = 220)

KeyYacute

(C++ enum variant: Key_Yacute = 221)

KeyTHORN

(C++ enum variant: Key_THORN = 222)

KeySsharp

(C++ enum variant: Key_ssharp = 223)

KeyDivision

(C++ enum variant: Key_division = 247)

KeyYdiaeresis

(C++ enum variant: Key_ydiaeresis = 255)

KeyEscape

(C++ enum variant: Key_Escape = 16777216)

KeyTab

(C++ enum variant: Key_Tab = 16777217)

KeyBacktab

(C++ enum variant: Key_Backtab = 16777218)

KeyBackspace

(C++ enum variant: Key_Backspace = 16777219)

KeyReturn

(C++ enum variant: Key_Return = 16777220)

KeyEnter

Typically located on the keypad. (C++ enum variant: Key_Enter = 16777221)

KeyInsert

(C++ enum variant: Key_Insert = 16777222)

KeyDelete

(C++ enum variant: Key_Delete = 16777223)

KeyPause

The Pause/Break key (Note: Not related to pausing media) (C++ enum variant: Key_Pause = 16777224)

KeyPrint

(C++ enum variant: Key_Print = 16777225)

KeySysReq

(C++ enum variant: Key_SysReq = 16777226)

KeyClear

(C++ enum variant: Key_Clear = 16777227)

KeyHome

(C++ enum variant: Key_Home = 16777232)

KeyEnd

(C++ enum variant: Key_End = 16777233)

KeyLeft

(C++ enum variant: Key_Left = 16777234)

KeyUp

(C++ enum variant: Key_Up = 16777235)

KeyRight

(C++ enum variant: Key_Right = 16777236)

KeyDown

(C++ enum variant: Key_Down = 16777237)

KeyPageUp

(C++ enum variant: Key_PageUp = 16777238)

KeyPageDown

(C++ enum variant: Key_PageDown = 16777239)

KeyShift

(C++ enum variant: Key_Shift = 16777248)

KeyControl

On OS X, this corresponds to the Command keys. (C++ enum variant: Key_Control = 16777249)

KeyMeta

On OS X, this corresponds to the Control keys. On Windows keyboards, this key is mapped to the Windows key. (C++ enum variant: Key_Meta = 16777250)

KeyAlt

(C++ enum variant: Key_Alt = 16777251)

KeyCapsLock

(C++ enum variant: Key_CapsLock = 16777252)

KeyNumLock

(C++ enum variant: Key_NumLock = 16777253)

KeyScrollLock

(C++ enum variant: Key_ScrollLock = 16777254)

KeyF1

(C++ enum variant: Key_F1 = 16777264)

KeyF2

(C++ enum variant: Key_F2 = 16777265)

KeyF3

(C++ enum variant: Key_F3 = 16777266)

KeyF4

(C++ enum variant: Key_F4 = 16777267)

KeyF5

(C++ enum variant: Key_F5 = 16777268)

KeyF6

(C++ enum variant: Key_F6 = 16777269)

KeyF7

(C++ enum variant: Key_F7 = 16777270)

KeyF8

(C++ enum variant: Key_F8 = 16777271)

KeyF9

(C++ enum variant: Key_F9 = 16777272)

KeyF10

(C++ enum variant: Key_F10 = 16777273)

KeyF11

(C++ enum variant: Key_F11 = 16777274)

KeyF12

(C++ enum variant: Key_F12 = 16777275)

KeyF13

(C++ enum variant: Key_F13 = 16777276)

KeyF14

(C++ enum variant: Key_F14 = 16777277)

KeyF15

(C++ enum variant: Key_F15 = 16777278)

KeyF16

(C++ enum variant: Key_F16 = 16777279)

KeyF17

(C++ enum variant: Key_F17 = 16777280)

KeyF18

(C++ enum variant: Key_F18 = 16777281)

KeyF19

(C++ enum variant: Key_F19 = 16777282)

KeyF20

(C++ enum variant: Key_F20 = 16777283)

KeyF21

(C++ enum variant: Key_F21 = 16777284)

KeyF22

(C++ enum variant: Key_F22 = 16777285)

KeyF23

(C++ enum variant: Key_F23 = 16777286)

KeyF24

(C++ enum variant: Key_F24 = 16777287)

KeyF25

(C++ enum variant: Key_F25 = 16777288)

KeyF26

(C++ enum variant: Key_F26 = 16777289)

KeyF27

(C++ enum variant: Key_F27 = 16777290)

KeyF28

(C++ enum variant: Key_F28 = 16777291)

KeyF29

(C++ enum variant: Key_F29 = 16777292)

KeyF30

(C++ enum variant: Key_F30 = 16777293)

KeyF31

(C++ enum variant: Key_F31 = 16777294)

KeyF32

(C++ enum variant: Key_F32 = 16777295)

KeyF33

(C++ enum variant: Key_F33 = 16777296)

KeyF34

(C++ enum variant: Key_F34 = 16777297)

KeyF35

(C++ enum variant: Key_F35 = 16777298)

KeySuperL

(C++ enum variant: Key_Super_L = 16777299)

KeySuperR

(C++ enum variant: Key_Super_R = 16777300)

KeyMenu

(C++ enum variant: Key_Menu = 16777301)

KeyHyperL

(C++ enum variant: Key_Hyper_L = 16777302)

KeyHyperR

(C++ enum variant: Key_Hyper_R = 16777303)

KeyHelp

(C++ enum variant: Key_Help = 16777304)

KeyDirectionL

(C++ enum variant: Key_Direction_L = 16777305)

KeyDirectionR

(C++ enum variant: Key_Direction_R = 16777312)

KeyBack

(C++ enum variant: Key_Back = 16777313)

KeyForward

(C++ enum variant: Key_Forward = 16777314)

KeyStop

(C++ enum variant: Key_Stop = 16777315)

KeyRefresh

(C++ enum variant: Key_Refresh = 16777316)

KeyVolumeDown

(C++ enum variant: Key_VolumeDown = 16777328)

KeyVolumeMute

(C++ enum variant: Key_VolumeMute = 16777329)

KeyVolumeUp

(C++ enum variant: Key_VolumeUp = 16777330)

KeyBassBoost

(C++ enum variant: Key_BassBoost = 16777331)

KeyBassUp

(C++ enum variant: Key_BassUp = 16777332)

KeyBassDown

(C++ enum variant: Key_BassDown = 16777333)

KeyTrebleUp

(C++ enum variant: Key_TrebleUp = 16777334)

KeyTrebleDown

(C++ enum variant: Key_TrebleDown = 16777335)

KeyMediaPlay

A key setting the state of the media player to play (C++ enum variant: Key_MediaPlay = 16777344)

KeyMediaStop

A key setting the state of the media player to stop (C++ enum variant: Key_MediaStop = 16777345)

KeyMediaPrevious

(C++ enum variant: Key_MediaPrevious = 16777346)

KeyMediaNext

(C++ enum variant: Key_MediaNext = 16777347)

KeyMediaRecord

(C++ enum variant: Key_MediaRecord = 16777348)

KeyMediaPause

A key setting the state of the media player to pause (Note: not the pause/break key) (C++ enum variant: Key_MediaPause = 16777349)

KeyMediaTogglePlayPause

A key to toggle the play/pause state in the media player (rather than setting an absolute state) (C++ enum variant: Key_MediaTogglePlayPause = 16777350)

KeyHomePage

(C++ enum variant: Key_HomePage = 16777360)

KeyFavorites

(C++ enum variant: Key_Favorites = 16777361)

KeySearch

(C++ enum variant: Key_Search = 16777362)

KeyStandby

(C++ enum variant: Key_Standby = 16777363)

KeyOpenUrl

(C++ enum variant: Key_OpenUrl = 16777364)

KeyLaunchMail

(C++ enum variant: Key_LaunchMail = 16777376)

KeyLaunchMedia

(C++ enum variant: Key_LaunchMedia = 16777377)

KeyLaunch0

On X11 this key is mapped to "My Computer" (XF86XK_MyComputer) key for legacy reasons. (C++ enum variant: Key_Launch0 = 16777378)

KeyLaunch1

On X11 this key is mapped to "Calculator" (XF86XK_Calculator) key for legacy reasons. (C++ enum variant: Key_Launch1 = 16777379)

KeyLaunch2

On X11 this key is mapped to XF86XK_Launch0 key for legacy reasons. (C++ enum variant: Key_Launch2 = 16777380)

KeyLaunch3

On X11 this key is mapped to XF86XK_Launch1 key for legacy reasons. (C++ enum variant: Key_Launch3 = 16777381)

KeyLaunch4

On X11 this key is mapped to XF86XK_Launch2 key for legacy reasons. (C++ enum variant: Key_Launch4 = 16777382)

KeyLaunch5

On X11 this key is mapped to XF86XK_Launch3 key for legacy reasons. (C++ enum variant: Key_Launch5 = 16777383)

KeyLaunch6

On X11 this key is mapped to XF86XK_Launch4 key for legacy reasons. (C++ enum variant: Key_Launch6 = 16777384)

KeyLaunch7

On X11 this key is mapped to XF86XK_Launch5 key for legacy reasons. (C++ enum variant: Key_Launch7 = 16777385)

KeyLaunch8

On X11 this key is mapped to XF86XK_Launch6 key for legacy reasons. (C++ enum variant: Key_Launch8 = 16777386)

KeyLaunch9

On X11 this key is mapped to XF86XK_Launch7 key for legacy reasons. (C++ enum variant: Key_Launch9 = 16777387)

KeyLaunchA

On X11 this key is mapped to XF86XK_Launch8 key for legacy reasons. (C++ enum variant: Key_LaunchA = 16777388)

KeyLaunchB

On X11 this key is mapped to XF86XK_Launch9 key for legacy reasons. (C++ enum variant: Key_LaunchB = 16777389)

KeyLaunchC

On X11 this key is mapped to XF86XK_LaunchA key for legacy reasons. (C++ enum variant: Key_LaunchC = 16777390)

KeyLaunchD

On X11 this key is mapped to XF86XK_LaunchB key for legacy reasons. (C++ enum variant: Key_LaunchD = 16777391)

KeyLaunchE

On X11 this key is mapped to XF86XK_LaunchC key for legacy reasons. (C++ enum variant: Key_LaunchE = 16777392)

KeyLaunchF

On X11 this key is mapped to XF86XK_LaunchD key for legacy reasons. (C++ enum variant: Key_LaunchF = 16777393)

KeyMonBrightnessUp

(C++ enum variant: Key_MonBrightnessUp = 16777394)

KeyMonBrightnessDown

(C++ enum variant: Key_MonBrightnessDown = 16777395)

KeyKeyboardLightOnOff

(C++ enum variant: Key_KeyboardLightOnOff = 16777396)

KeyKeyboardBrightnessUp

(C++ enum variant: Key_KeyboardBrightnessUp = 16777397)

KeyKeyboardBrightnessDown

(C++ enum variant: Key_KeyboardBrightnessDown = 16777398)

KeyPowerOff

(C++ enum variant: Key_PowerOff = 16777399)

KeyWakeUp

(C++ enum variant: Key_WakeUp = 16777400)

KeyEject

(C++ enum variant: Key_Eject = 16777401)

KeyScreenSaver

(C++ enum variant: Key_ScreenSaver = 16777402)

KeyWWW

(C++ enum variant: Key_WWW = 16777403)

KeyMemo

(C++ enum variant: Key_Memo = 16777404)

KeyLightBulb

(C++ enum variant: Key_LightBulb = 16777405)

KeyShop

(C++ enum variant: Key_Shop = 16777406)

KeyHistory

(C++ enum variant: Key_History = 16777407)

KeyAddFavorite

(C++ enum variant: Key_AddFavorite = 16777408)

KeyHotLinks

(C++ enum variant: Key_HotLinks = 16777409)

KeyBrightnessAdjust

(C++ enum variant: Key_BrightnessAdjust = 16777410)

KeyFinance

(C++ enum variant: Key_Finance = 16777411)

KeyCommunity

(C++ enum variant: Key_Community = 16777412)

KeyAudioRewind

(C++ enum variant: Key_AudioRewind = 16777413)

KeyBackForward

(C++ enum variant: Key_BackForward = 16777414)

KeyApplicationLeft

(C++ enum variant: Key_ApplicationLeft = 16777415)

KeyApplicationRight

(C++ enum variant: Key_ApplicationRight = 16777416)

KeyBook

(C++ enum variant: Key_Book = 16777417)

KeyCD

(C++ enum variant: Key_CD = 16777418)

KeyCalculator

On X11 this key is not mapped for legacy reasons. Use Qt::Key_Launch1 instead. (C++ enum variant: Key_Calculator = 16777419)

KeyToDoList

(C++ enum variant: Key_ToDoList = 16777420)

KeyClearGrab

(C++ enum variant: Key_ClearGrab = 16777421)

KeyClose

(C++ enum variant: Key_Close = 16777422)

KeyCopy

(C++ enum variant: Key_Copy = 16777423)

KeyCut

(C++ enum variant: Key_Cut = 16777424)

KeyDisplay

(C++ enum variant: Key_Display = 16777425)

KeyDOS

(C++ enum variant: Key_DOS = 16777426)

KeyDocuments

(C++ enum variant: Key_Documents = 16777427)

KeyExcel

(C++ enum variant: Key_Excel = 16777428)

KeyExplorer

(C++ enum variant: Key_Explorer = 16777429)

KeyGame

(C++ enum variant: Key_Game = 16777430)

KeyGo

(C++ enum variant: Key_Go = 16777431)

KeyITouch

(C++ enum variant: Key_iTouch = 16777432)

KeyLogOff

(C++ enum variant: Key_LogOff = 16777433)

KeyMarket

(C++ enum variant: Key_Market = 16777434)

KeyMeeting

(C++ enum variant: Key_Meeting = 16777435)

KeyMenuKB

(C++ enum variant: Key_MenuKB = 16777436)

KeyMenuPB

(C++ enum variant: Key_MenuPB = 16777437)

KeyMySites

(C++ enum variant: Key_MySites = 16777438)

KeyNews

(C++ enum variant: Key_News = 16777439)

KeyOfficeHome

(C++ enum variant: Key_OfficeHome = 16777440)

KeyOption

(C++ enum variant: Key_Option = 16777441)

KeyPaste

(C++ enum variant: Key_Paste = 16777442)

KeyPhone

(C++ enum variant: Key_Phone = 16777443)

KeyCalendar

(C++ enum variant: Key_Calendar = 16777444)

KeyReply

(C++ enum variant: Key_Reply = 16777445)

KeyReload

(C++ enum variant: Key_Reload = 16777446)

KeyRotateWindows

(C++ enum variant: Key_RotateWindows = 16777447)

KeyRotationPB

(C++ enum variant: Key_RotationPB = 16777448)

KeyRotationKB

(C++ enum variant: Key_RotationKB = 16777449)

KeySave

(C++ enum variant: Key_Save = 16777450)

KeySend

(C++ enum variant: Key_Send = 16777451)

KeySpell

(C++ enum variant: Key_Spell = 16777452)

KeySplitScreen

(C++ enum variant: Key_SplitScreen = 16777453)

KeySupport

(C++ enum variant: Key_Support = 16777454)

KeyTaskPane

(C++ enum variant: Key_TaskPane = 16777455)

KeyTerminal

(C++ enum variant: Key_Terminal = 16777456)

KeyTools

(C++ enum variant: Key_Tools = 16777457)

KeyTravel

(C++ enum variant: Key_Travel = 16777458)

KeyVideo

(C++ enum variant: Key_Video = 16777459)

KeyWord

(C++ enum variant: Key_Word = 16777460)

KeyXfer

(C++ enum variant: Key_Xfer = 16777461)

KeyZoomIn

(C++ enum variant: Key_ZoomIn = 16777462)

KeyZoomOut

(C++ enum variant: Key_ZoomOut = 16777463)

KeyAway

(C++ enum variant: Key_Away = 16777464)

KeyMessenger

(C++ enum variant: Key_Messenger = 16777465)

KeyWebCam

(C++ enum variant: Key_WebCam = 16777466)

KeyMailForward

(C++ enum variant: Key_MailForward = 16777467)

KeyPictures

(C++ enum variant: Key_Pictures = 16777468)

KeyMusic

(C++ enum variant: Key_Music = 16777469)

KeyBattery

(C++ enum variant: Key_Battery = 16777470)

KeyBluetooth

(C++ enum variant: Key_Bluetooth = 16777471)

KeyWLAN

(C++ enum variant: Key_WLAN = 16777472)

KeyUWB

(C++ enum variant: Key_UWB = 16777473)

KeyAudioForward

(C++ enum variant: Key_AudioForward = 16777474)

KeyAudioRepeat

(C++ enum variant: Key_AudioRepeat = 16777475)

KeyAudioRandomPlay

(C++ enum variant: Key_AudioRandomPlay = 16777476)

KeySubtitle

(C++ enum variant: Key_Subtitle = 16777477)

KeyAudioCycleTrack

(C++ enum variant: Key_AudioCycleTrack = 16777478)

KeyTime

(C++ enum variant: Key_Time = 16777479)

KeyHibernate

(C++ enum variant: Key_Hibernate = 16777480)

KeyView

(C++ enum variant: Key_View = 16777481)

KeyTopMenu

(C++ enum variant: Key_TopMenu = 16777482)

KeyPowerDown

(C++ enum variant: Key_PowerDown = 16777483)

KeySuspend

(C++ enum variant: Key_Suspend = 16777484)

KeyContrastAdjust

(C++ enum variant: Key_ContrastAdjust = 16777485)

KeyLaunchG

On X11 this key is mapped to XF86XK_LaunchE key for legacy reasons. (C++ enum variant: Key_LaunchG = 16777486)

KeyLaunchH

On X11 this key is mapped to XF86XK_LaunchF key for legacy reasons. (C++ enum variant: Key_LaunchH = 16777487)

KeyTouchpadToggle

(C++ enum variant: Key_TouchpadToggle = 16777488)

KeyTouchpadOn

(C++ enum variant: Key_TouchpadOn = 16777489)

KeyTouchpadOff

(C++ enum variant: Key_TouchpadOff = 16777490)

KeyMicMute

(C++ enum variant: Key_MicMute = 16777491)

KeyRed

(C++ enum variant: Key_Red = 16777492)

KeyGreen

(C++ enum variant: Key_Green = 16777493)

KeyYellow

(C++ enum variant: Key_Yellow = 16777494)

KeyBlue

(C++ enum variant: Key_Blue = 16777495)

KeyChannelUp

(C++ enum variant: Key_ChannelUp = 16777496)

KeyChannelDown

(C++ enum variant: Key_ChannelDown = 16777497)

KeyGuide

C++ enum variant: Key_Guide = 16777498

KeyInfo

C++ enum variant: Key_Info = 16777499

KeySettings

C++ enum variant: Key_Settings = 16777500

KeyMicVolumeUp

C++ enum variant: Key_MicVolumeUp = 16777501

KeyMicVolumeDown

C++ enum variant: Key_MicVolumeDown = 16777502

KeyNew

C++ enum variant: Key_New = 16777504

KeyOpen

C++ enum variant: Key_Open = 16777505

KeyFind

C++ enum variant: Key_Find = 16777506

KeyUndo

C++ enum variant: Key_Undo = 16777507

KeyRedo

C++ enum variant: Key_Redo = 16777508

KeyAltGr

On Windows, when the KeyDown event for this key is sent, the Ctrl+Alt modifiers are also set. (C++ enum variant: Key_AltGr = 16781571)

KeyMultiKey

(C++ enum variant: Key_Multi_key = 16781600)

KeyKanji

(C++ enum variant: Key_Kanji = 16781601)

KeyMuhenkan

(C++ enum variant: Key_Muhenkan = 16781602)

KeyHenkan

(C++ enum variant: Key_Henkan = 16781603)

KeyRomaji

(C++ enum variant: Key_Romaji = 16781604)

KeyHiragana

(C++ enum variant: Key_Hiragana = 16781605)

KeyKatakana

(C++ enum variant: Key_Katakana = 16781606)

KeyHiraganaKatakana

(C++ enum variant: Key_Hiragana_Katakana = 16781607)

KeyZenkaku

(C++ enum variant: Key_Zenkaku = 16781608)

KeyHankaku

(C++ enum variant: Key_Hankaku = 16781609)

KeyZenkakuHankaku

(C++ enum variant: Key_Zenkaku_Hankaku = 16781610)

KeyTouroku

(C++ enum variant: Key_Touroku = 16781611)

KeyMassyo

(C++ enum variant: Key_Massyo = 16781612)

KeyKanaLock

(C++ enum variant: Key_Kana_Lock = 16781613)

KeyKanaShift

(C++ enum variant: Key_Kana_Shift = 16781614)

KeyEisuShift

(C++ enum variant: Key_Eisu_Shift = 16781615)

KeyEisuToggle

(C++ enum variant: Key_Eisu_toggle = 16781616)

KeyHangul

(C++ enum variant: Key_Hangul = 16781617)

KeyHangulStart

(C++ enum variant: Key_Hangul_Start = 16781618)

KeyHangulEnd

(C++ enum variant: Key_Hangul_End = 16781619)

KeyHangulHanja

(C++ enum variant: Key_Hangul_Hanja = 16781620)

KeyHangulJamo

(C++ enum variant: Key_Hangul_Jamo = 16781621)

KeyHangulRomaja

(C++ enum variant: Key_Hangul_Romaja = 16781622)

KeyCodeinput

(C++ enum variant: Key_Codeinput = 16781623)

KeyHangulJeonja

(C++ enum variant: Key_Hangul_Jeonja = 16781624)

KeyHangulBanja

(C++ enum variant: Key_Hangul_Banja = 16781625)

KeyHangulPreHanja

(C++ enum variant: Key_Hangul_PreHanja = 16781626)

KeyHangulPostHanja

(C++ enum variant: Key_Hangul_PostHanja = 16781627)

KeySingleCandidate

(C++ enum variant: Key_SingleCandidate = 16781628)

KeyMultipleCandidate

(C++ enum variant: Key_MultipleCandidate = 16781629)

KeyPreviousCandidate

(C++ enum variant: Key_PreviousCandidate = 16781630)

KeyHangulSpecial

(C++ enum variant: Key_Hangul_Special = 16781631)

KeyModeSwitch

(C++ enum variant: Key_Mode_switch = 16781694)

KeyDeadGrave

(C++ enum variant: Key_Dead_Grave = 16781904)

KeyDeadAcute

(C++ enum variant: Key_Dead_Acute = 16781905)

KeyDeadCircumflex

(C++ enum variant: Key_Dead_Circumflex = 16781906)

KeyDeadTilde

(C++ enum variant: Key_Dead_Tilde = 16781907)

KeyDeadMacron

(C++ enum variant: Key_Dead_Macron = 16781908)

KeyDeadBreve

(C++ enum variant: Key_Dead_Breve = 16781909)

KeyDeadAbovedot

(C++ enum variant: Key_Dead_Abovedot = 16781910)

KeyDeadDiaeresis

(C++ enum variant: Key_Dead_Diaeresis = 16781911)

KeyDeadAbovering

(C++ enum variant: Key_Dead_Abovering = 16781912)

KeyDeadDoubleacute

(C++ enum variant: Key_Dead_Doubleacute = 16781913)

KeyDeadCaron

(C++ enum variant: Key_Dead_Caron = 16781914)

KeyDeadCedilla

(C++ enum variant: Key_Dead_Cedilla = 16781915)

KeyDeadOgonek

(C++ enum variant: Key_Dead_Ogonek = 16781916)

KeyDeadIota

(C++ enum variant: Key_Dead_Iota = 16781917)

KeyDeadVoicedSound

(C++ enum variant: Key_Dead_Voiced_Sound = 16781918)

KeyDeadSemivoicedSound

(C++ enum variant: Key_Dead_Semivoiced_Sound = 16781919)

KeyDeadBelowdot

(C++ enum variant: Key_Dead_Belowdot = 16781920)

KeyDeadHook

(C++ enum variant: Key_Dead_Hook = 16781921)

KeyDeadHorn

(C++ enum variant: Key_Dead_Horn = 16781922)

KeyMediaLast

(C++ enum variant: Key_MediaLast = 16842751)

KeySelect

(C++ enum variant: Key_Select = 16842752)

KeyYes

(C++ enum variant: Key_Yes = 16842753)

KeyNo

(C++ enum variant: Key_No = 16842754)

KeyCancel

(C++ enum variant: Key_Cancel = 16908289)

KeyPrinter

(C++ enum variant: Key_Printer = 16908290)

KeyExecute

(C++ enum variant: Key_Execute = 16908291)

KeySleep

(C++ enum variant: Key_Sleep = 16908292)

KeyPlay

(C++ enum variant: Key_Play = 16908293)

KeyZoom

(C++ enum variant: Key_Zoom = 16908294)

KeyExit

C++ enum variant: Key_Exit = 16908298

KeyContext1

(C++ enum variant: Key_Context1 = 17825792)

KeyContext2

(C++ enum variant: Key_Context2 = 17825793)

KeyContext3

(C++ enum variant: Key_Context3 = 17825794)

KeyContext4

(C++ enum variant: Key_Context4 = 17825795)

KeyCall

A key to answer or initiate a call (see Qt::Key_ToggleCallHangup for a key to toggle current call state) (C++ enum variant: Key_Call = 17825796)

KeyHangup

A key to end an ongoing call (see Qt::Key_ToggleCallHangup for a key to toggle current call state) (C++ enum variant: Key_Hangup = 17825797)

KeyFlip

(C++ enum variant: Key_Flip = 17825798)

KeyToggleCallHangup

A key to toggle the current call state (ie. either answer, or hangup) depending on current call state (C++ enum variant: Key_ToggleCallHangup = 17825799)

KeyVoiceDial

(C++ enum variant: Key_VoiceDial = 17825800)

KeyLastNumberRedial

(C++ enum variant: Key_LastNumberRedial = 17825801)

KeyCamera

A key to activate the camera shutter (C++ enum variant: Key_Camera = 17825824)

KeyCameraFocus

A key to focus the camera (C++ enum variant: Key_CameraFocus = 17825825)

KeyUnknown

(C++ enum variant: Key_unknown = 33554431)