Enum qt_core::qt::MouseButton

pub enum MouseButton {
    MouseButtonMask,
    NoButton,
    LeftButton,
    RightButton,
    MidButton,
    BackButton,
    ForwardButton,
    TaskButton,
    ExtraButton4,
    ExtraButton5,
    ExtraButton6,
    ExtraButton7,
    ExtraButton8,
    ExtraButton9,
    ExtraButton10,
    ExtraButton11,
    ExtraButton12,
    ExtraButton13,
    ExtraButton14,
    ExtraButton15,
    ExtraButton16,
    ExtraButton17,
    ExtraButton18,
    ExtraButton19,
    ExtraButton20,
    ExtraButton21,
    ExtraButton22,
    ExtraButton23,
    ExtraButton24,
    AllButtons,
}

C++ type: Qt::MouseButton

C++ documentation:

This enum type describes the different mouse buttons.

Note: Some models of multi-button mice are pre-configured with high-numbered Buttons emulating keyboard sequences, for use in specific games. In order for these Buttons to be seen as actual 'Mouse Buttons', the device must be re-configured (using the vendor's configuration tool).

The MouseButtons type is a typedef for QFlags<MouseButton>. It stores an OR combination of MouseButton values.

See also KeyboardModifier and Modifier.

Variants
MouseButtonMask

C++ enum variant: MouseButtonMask = -1

NoButton

The button state does not refer to any button (see QMouseEvent::button()). (C++ enum variant: NoButton = 0)

LeftButton

The left button is pressed, or an event refers to the left button. (The left button may be the right button on left-handed mice.) (C++ enum variant: LeftButton = 1)

RightButton

The right button. (C++ enum variant: RightButton = 2)

MidButton

This variant corresponds to multiple C++ enum variants with the same value:

MidButton = 4: The middle button.

MiddleButton = 4: The middle button.

BackButton

This variant corresponds to multiple C++ enum variants with the same value:

BackButton = 8: The 'Back' button. (Typically present on the 'thumb' side of a mouse with extra buttons. This is NOT the tilt wheel.)

XButton1 = 8: The 'Back' Button.

ExtraButton1 = 8: The 'Back' Button.

ForwardButton

This variant corresponds to multiple C++ enum variants with the same value:

ForwardButton = 16: The 'Forward' Button. (Typically present beside the 'Back' button, and also pressed by the thumb.)

XButton2 = 16: The 'Forward Button.

ExtraButton2 = 16: The 'Forward' Button.

TaskButton

This variant corresponds to multiple C++ enum variants with the same value:

TaskButton = 32: The 'Task' Button.

ExtraButton3 = 32: The 'Task' Button.

ExtraButton4

The 7th non-wheel Mouse Button. (C++ enum variant: ExtraButton4 = 64)

ExtraButton5

The 8th non-wheel Mouse Button. (C++ enum variant: ExtraButton5 = 128)

ExtraButton6

The 9th non-wheel Mouse Button. (C++ enum variant: ExtraButton6 = 256)

ExtraButton7

The 10th non-wheel Mouse Button. (C++ enum variant: ExtraButton7 = 512)

ExtraButton8

The 11th non-wheel Mouse Button. (C++ enum variant: ExtraButton8 = 1024)

ExtraButton9

The 12th non-wheel Mouse Button. (C++ enum variant: ExtraButton9 = 2048)

ExtraButton10

The 13th non-wheel Mouse Button. (C++ enum variant: ExtraButton10 = 4096)

ExtraButton11

The 14th non-wheel Mouse Button. (C++ enum variant: ExtraButton11 = 8192)

ExtraButton12

The 15th non-wheel Mouse Button. (C++ enum variant: ExtraButton12 = 16384)

ExtraButton13

The 16th non-wheel Mouse Button. (C++ enum variant: ExtraButton13 = 32768)

ExtraButton14

The 17th non-wheel Mouse Button. (C++ enum variant: ExtraButton14 = 65536)

ExtraButton15

The 18th non-wheel Mouse Button. (C++ enum variant: ExtraButton15 = 131072)

ExtraButton16

The 19th non-wheel Mouse Button. (C++ enum variant: ExtraButton16 = 262144)

ExtraButton17

The 20th non-wheel Mouse Button. (C++ enum variant: ExtraButton17 = 524288)

ExtraButton18

The 21st non-wheel Mouse Button. (C++ enum variant: ExtraButton18 = 1048576)

ExtraButton19

The 22nd non-wheel Mouse Button. (C++ enum variant: ExtraButton19 = 2097152)

ExtraButton20

The 23rd non-wheel Mouse Button. (C++ enum variant: ExtraButton20 = 4194304)

ExtraButton21

The 24th non-wheel Mouse Button. (C++ enum variant: ExtraButton21 = 8388608)

ExtraButton22

The 25th non-wheel Mouse Button. (C++ enum variant: ExtraButton22 = 16777216)

ExtraButton23

The 26th non-wheel Mouse Button. (C++ enum variant: ExtraButton23 = 33554432)

ExtraButton24

This variant corresponds to multiple C++ enum variants with the same value:

ExtraButton24 = 67108864: The 27th non-wheel Mouse Button.

MaxMouseButton = 67108864

AllButtons

This value corresponds to a mask of all possible mouse buttons. Use to set the 'acceptedButtons' property of a MouseArea to accept ALL mouse buttons. (C++ enum variant: AllButtons = 134217727)