Enum qt_core::qt::ScreenOrientation

pub enum ScreenOrientation {
    Primary,
    Portrait,
    Landscape,
    InvertedPortrait,
    InvertedLandscape,
}

C++ type: Qt::ScreenOrientation

C++ documentation:

This enum type specifies the various orientations a screen might have.

The ScreenOrientations type is a typedef for QFlags<ScreenOrientation>. It stores an OR combination of ScreenOrientation values.

Variants
Primary

The display's primary orientation. (C++ enum variant: PrimaryOrientation = 0)

Portrait

Portrait orientation, display height is greater than display width, rotated 90 degree clockwise relative to landscape. (C++ enum variant: PortraitOrientation = 1)

Landscape

Landscape orientation, display width is greater than display height. (C++ enum variant: LandscapeOrientation = 2)

InvertedPortrait

Inverted portrait orientation, rotated 180 degrees relative to portrait. (C++ enum variant: InvertedPortraitOrientation = 4)

InvertedLandscape

Inverted landscape orientation, rotated 180 degrees relative to landscape. (C++ enum variant: InvertedLandscapeOrientation = 8)