Enum qt_core::qt::TabFocusBehavior

pub enum TabFocusBehavior {
    NoTabFocus,
    TabFocusTextControls,
    TabFocusListControls,
    TabFocusAllControls,
}

C++ type: Qt::TabFocusBehavior

C++ documentation:

This enum type provides different focus behaviors for tab navigation.

This enum was introduced or modified in Qt 5.5.

Variants
NoTabFocus

iterate nothing. (C++ enum variant: NoTabFocus = 0)

TabFocusTextControls

iterate text controls and widgets. (C++ enum variant: TabFocusTextControls = 1)

TabFocusListControls

iterate list controls and widgets. (C++ enum variant: TabFocusListControls = 2)

TabFocusAllControls

iterate all controls and widgets. (C++ enum variant: TabFocusAllControls = 255)