Enum qt_core::qt::UieFfect

pub enum UieFfect {
    UIGeneral,
    UIAnimateMenu,
    UIFadeMenu,
    UIAnimateCombo,
    UIAnimateTooltip,
    UIFadeTooltip,
    UIAnimateToolBox,
}

C++ type: Qt::UIEffect

C++ documentation:

This enum describes the available UI effects.

By default, Qt will try to use the platform specific desktop settings for each effect. Use the QApplication::setDesktopSettingsAware() function (passing false as argument) to prevent this, and the QApplication::setEffectEnabled() to enable or disable a particular effect.

Note that all effects are disabled on screens running at less than 16-bit color depth.

See also QApplication::setEffectEnabled() and QApplication::setDesktopSettingsAware().

Variants
UIGeneral

C++ enum variant: UI_General = 0

UIAnimateMenu

Show animated menus. (C++ enum variant: UI_AnimateMenu = 1)

UIFadeMenu

Show faded menus. (C++ enum variant: UI_FadeMenu = 2)

UIAnimateCombo

Show animated comboboxes. (C++ enum variant: UI_AnimateCombo = 3)

UIAnimateTooltip

Show tooltip animations. (C++ enum variant: UI_AnimateTooltip = 4)

UIFadeTooltip

Show tooltip fading effects. (C++ enum variant: UI_FadeTooltip = 5)

UIAnimateToolBox

Reserved (C++ enum variant: UI_AnimateToolBox = 6)