Enum qt_core::qt::WidgetAttribute

pub enum WidgetAttribute {
    WADisabled,
    WAUnderMouse,
    WAMouseTracking,
    WAContentsPropagated,
    WAOpaquePaintEvent,
    WAStaticContents,
    WALaidOut,
    WAPaintOnScreen,
    WANoSystemBackground,
    WAUpdatesDisabled,
    WAMapped,
    WAMacNoClickThrough,
    WAInputMethodEnabled,
    WAWStateVisible,
    WAWStateHidden,
    WAForceDisabled,
    WAKeyCompression,
    WAPendingMoveEvent,
    WAPendingResizeEvent,
    WASetPalette,
    WASetFont,
    WASetCursor,
    WANoChildEventsFromChildren,
    WAWindowModified,
    WAResized,
    WAMoved,
    WAPendingUpdate,
    WAInvalidSize,
    WAMacBrushedMetal,
    WACustomWhatsThis,
    WALayoutOnEntireRect,
    WAOutsideWSRange,
    WAGrabbedShortcut,
    WATransparentForMouseEvents,
    WAPaintUnclipped,
    WASetWindowIcon,
    WANoMouseReplay,
    WADeleteOnClose,
    WARightToLeft,
    WASetLayoutDirection,
    WANoChildEventsForParent,
    WAForceUpdatesDisabled,
    WAWStateCreated,
    WAWStateCompressKeys,
    WAWStateInPaintEvent,
    WAWStateReparented,
    WAWStateConfigPending,
    WAWStatePolished,
    WAWStateDND,
    WAWStateOwnSizePolicy,
    WAWStateExplicitShowHide,
    WAShowModal,
    WAMouseNoMask,
    WAGroupLeader,
    WANoMousePropagation,
    WAHover,
    WAInputMethodTransparent,
    WAQuitOnClose,
    WAKeyboardFocusChange,
    WAAcceptDrops,
    WADropSiteRegistered,
    WAWindowPropagation,
    WANoX11EventCompression,
    WATintedBackground,
    WAX11OpenGLOverlay,
    WAAlwaysShowToolTips,
    WAMacOpaqueSizeGrip,
    WASetStyle,
    WASetLocale,
    WAMacShowFocusRect,
    WAMacNormalSize,
    WAMacSmallSize,
    WAMacMiniSize,
    WALayoutUsesWidgetRect,
    WAStyledBackground,
    WAMSWindowsUseDirect3D,
    WACanHostQMdiSubWindowTitleBar,
    WAMacAlwaysShowToolWindow,
    WAStyleSheet,
    WAShowWithoutActivating,
    WAX11BypassTransientForHint,
    WANativeWindow,
    WADontCreateNativeAncestors,
    WAMacVariableSize,
    WADontShowOnScreen,
    WAX11NetWmWindowTypeDesktop,
    WAX11NetWmWindowTypeDock,
    WAX11NetWmWindowTypeToolBar,
    WAX11NetWmWindowTypeMenu,
    WAX11NetWmWindowTypeUtility,
    WAX11NetWmWindowTypeSplash,
    WAX11NetWmWindowTypeDialog,
    WAX11NetWmWindowTypeDropDownMenu,
    WAX11NetWmWindowTypePopupMenu,
    WAX11NetWmWindowTypeToolTip,
    WAX11NetWmWindowTypeNotification,
    WAX11NetWmWindowTypeCombo,
    WAX11NetWmWindowTypeDND,
    WAMacFrameworkScaled,
    WASetWindowModality,
    WAWStateWindowOpacitySet,
    WATranslucentBackground,
    WAAcceptTouchEvents,
    WAWStateAcceptedTouchBeginEvent,
    WATouchPadAcceptSingleTouchEvents,
    WAX11DoNotAcceptFocus,
    WAMacNoShadow,
    WAAlwaysStackOnTop,
    WAAttributeCount,
}

C++ type: Qt::WidgetAttribute

C++ documentation:

This enum type is used to specify various widget attributes. Attributes are set and cleared with QWidget::setAttribute(), and queried with QWidget::testAttribute(), although some have special convenience functions which are mentioned below.

Variants
WADisabled

Indicates that the widget is disabled, i.e. it does not receive any mouse or keyboard events. There is also a getter functions QWidget::isEnabled(). This is set/cleared by the Qt kernel. (C++ enum variant: WA_Disabled = 0)

WAUnderMouse

Indicates that the widget is under the mouse cursor. The value is not updated correctly during drag and drop operations. There is also a getter function, QWidget::underMouse(). This flag is set or cleared by the Qt kernel. (C++ enum variant: WA_UnderMouse = 1)

WAMouseTracking

Indicates that the widget has mouse tracking enabled. See QWidget::mouseTracking. (C++ enum variant: WA_MouseTracking = 2)

WAContentsPropagated

This flag is superfluous and obsolete; it no longer has any effect. Since Qt 4.1, all widgets that do not set WA_PaintOnScreen propagate their contents. (C++ enum variant: WA_ContentsPropagated = 3)

WAOpaquePaintEvent

This variant corresponds to multiple C++ enum variants with the same value:

WA_OpaquePaintEvent = 4: Indicates that the widget paints all its pixels when it receives a paint event. Thus, it is not required for operations like updating, resizing, scrolling and focus changes to erase the widget before generating paint events. The use of WA_OpaquePaintEvent provides a small optimization by helping to reduce flicker on systems that do not support double buffering and avoiding computational cycles necessary to erase the background prior to painting. Note: Unlike WA_NoSystemBackground, WA_OpaquePaintEvent makes an effort to avoid transparent window backgrounds. This flag is set or cleared by the widget's author.

WA_NoBackground = 4: This value is obsolete. Use WA_OpaquePaintEvent instead.

WAStaticContents

Indicates that the widget contents are north-west aligned and static. On resize, such a widget will receive paint events only for parts of itself that are newly visible. This flag is set or cleared by the widget's author. (C++ enum variant: WA_StaticContents = 5)

WALaidOut

C++ enum variant: WA_LaidOut = 7

WAPaintOnScreen

Indicates that the widget wants to draw directly onto the screen. Widgets with this attribute set do not participate in composition management, i.e. they cannot be semi-transparent or shine through semi-transparent overlapping widgets. Note: This flag is only supported on X11 and it disables double buffering. On Qt for Embedded Linux, the flag only works when set on a top-level widget and it relies on support from the active screen driver. This flag is set or cleared by the widget's author. To render outside of Qt's paint system, e.g., if you require native painting primitives, you need to reimplement QWidget::paintEngine() to return 0 and set this flag. (C++ enum variant: WA_PaintOnScreen = 8)

WANoSystemBackground

Indicates that the widget has no background, i.e. when the widget receives paint events, the background is not automatically repainted. Note: Unlike WA_OpaquePaintEvent, newly exposed areas are never filled with the background (e.g., after showing a window for the first time the user can see "through" it until the application processes the paint events). This flag is set or cleared by the widget's author. (C++ enum variant: WA_NoSystemBackground = 9)

WAUpdatesDisabled

Indicates that updates are blocked (including the system background). This flag is set or cleared by the Qt kernel. Warning: This flag must never be set or cleared by the widget's author. (C++ enum variant: WA_UpdatesDisabled = 10)

WAMapped

Indicates that the widget is mapped on screen. This is set/cleared by the Qt kernel. (C++ enum variant: WA_Mapped = 11)

WAMacNoClickThrough

When a widget that has this attribute set is clicked, and its window is inactive, the click will make the window active but won't be seen by the widget. Typical use of this attribute is on widgets with "destructive" actions, such as a "Delete" button. WA_MacNoClickThrough also applies to all child widgets of the widget that has it set. (C++ enum variant: WA_MacNoClickThrough = 12)

WAInputMethodEnabled

Enables input methods for Asian languages. Must be set when creating custom text editing widgets. On Windows CE this flag can be used in addition to QApplication::autoSipEnabled to automatically display the SIP when entering a widget. (C++ enum variant: WA_InputMethodEnabled = 14)

WAWStateVisible

C++ enum variant: WA_WState_Visible = 15

WAWStateHidden

C++ enum variant: WA_WState_Hidden = 16

WAForceDisabled

Indicates that the widget is explicitly disabled, i.e. it will remain disabled even when all its ancestors are set to the enabled state. This implies WA_Disabled. This is set/cleared by QWidget::setEnabled() and QWidget::setDisabled(). (C++ enum variant: WA_ForceDisabled = 32)

WAKeyCompression

Enables key event compression if set, and disables it if not set. By default key compression is off, so widgets receive one key press event for each key press (or more, since autorepeat is usually on). If you turn it on and your program doesn't keep up with key input, Qt may try to compress key events so that more than one character can be processed in each event. For example, a word processor widget might receive 2, 3 or more characters in each QKeyEvent::text(), if the layout recalculation takes too long for the CPU. If a widget supports multiple character unicode input, it is always safe to turn the compression on. Qt performs key event compression only for printable characters. Qt::Modifier keys, cursor movement keys, function keys and miscellaneous action keys (e.g. Escape, Enter, Backspace, PrintScreen) will stop key event compression, even if there are more compressible key events available. Platforms other than Mac and X11 do not support this compression, in which case turning it on will have no effect. This is set/cleared by the widget's author. (C++ enum variant: WA_KeyCompression = 33)

WAPendingMoveEvent

Indicates that a move event is pending, e.g., when a hidden widget was moved. This flag is set or cleared by the Qt kernel. (C++ enum variant: WA_PendingMoveEvent = 34)

WAPendingResizeEvent

Indicates that a resize event is pending, e.g., when a hidden widget was resized. This flag is set or cleared by the Qt kernel. (C++ enum variant: WA_PendingResizeEvent = 35)

WASetPalette

Indicates that the widget has a palette of its own. This flag is set or cleared by QWidget::setPalette(). (C++ enum variant: WA_SetPalette = 36)

WASetFont

Indicates that the widget has a font of its own. This flag is set or cleared by QWidget::setFont(). (C++ enum variant: WA_SetFont = 37)

WASetCursor

Indicates that the widget has a cursor of its own. This flag is set or cleared by QWidget::setCursor() and QWidget::unsetCursor(). (C++ enum variant: WA_SetCursor = 38)

WANoChildEventsFromChildren

Indicates that the widget does not want to receive ChildAdded or ChildRemoved events sent from its children. This is set by a widget's author. (C++ enum variant: WA_NoChildEventsFromChildren = 39)

WAWindowModified

Indicates that the window is marked as modified. On some platforms this flag will do nothing, on others (including OS X and Windows) the window will take a modified appearance. This flag is set or cleared by QWidget::setWindowModified(). (C++ enum variant: WA_WindowModified = 41)

WAResized

Indicates that the widget has an explicit size. This flag is set or cleared by QWidget::resize() and QWidget::setGeometry(). (C++ enum variant: WA_Resized = 42)

WAMoved

Indicates that the widget has an explicit position. This is set/cleared by QWidget::move() and by QWidget::setGeometry(). (C++ enum variant: WA_Moved = 43)

WAPendingUpdate

C++ enum variant: WA_PendingUpdate = 44

WAInvalidSize

C++ enum variant: WA_InvalidSize = 45

WAMacBrushedMetal

This variant corresponds to multiple C++ enum variants with the same value:

WA_MacBrushedMetal = 46: Indicates the widget should be drawn in the brushed metal style as supported by the windowing system. This attribute is only applicable to OS X.

WA_MacMetalStyle = 46

WACustomWhatsThis

Indicates that the widget wants to continue operating normally in "What's This?" mode. This is set by the widget's author. (C++ enum variant: WA_CustomWhatsThis = 47)

WALayoutOnEntireRect

Indicates that the widget wants QLayout to operate on the entire QWidget::rect(), not only on QWidget::contentsRect(). This is set by the widget's author. (C++ enum variant: WA_LayoutOnEntireRect = 48)

WAOutsideWSRange

Indicates that the widget is outside the valid range of the window system's coordinate system. A widget outside the valid range cannot be mapped on screen. This is set/cleared by the Qt kernel. (C++ enum variant: WA_OutsideWSRange = 49)

WAGrabbedShortcut

C++ enum variant: WA_GrabbedShortcut = 50

WATransparentForMouseEvents

When enabled, this attribute disables the delivery of mouse events to the widget and its children. Mouse events are delivered to other widgets as if the widget and its children were not present in the widget hierarchy; mouse clicks and other events effectively "pass through" them. This attribute is disabled by default. (C++ enum variant: WA_TransparentForMouseEvents = 51)

WAPaintUnclipped

Makes all painters operating on this widget unclipped. Children of this widget or other widgets in front of it do not clip the area the painter can paint on. This flag is only supported for widgets with the WA_PaintOnScreen flag set. The preferred way to do this in a cross platform way is to create a transparent widget that lies in front of the other widgets. (C++ enum variant: WA_PaintUnclipped = 52)

WASetWindowIcon

C++ enum variant: WA_SetWindowIcon = 53

WANoMouseReplay

Used for pop-up widgets. Indicates that the most recent mouse press event should not be replayed when the pop-up widget closes. The flag is set by the widget's author and cleared by the Qt kernel every time the widget receives a new mouse event. (C++ enum variant: WA_NoMouseReplay = 54)

WADeleteOnClose

Makes Qt delete this widget when the widget has accepted the close event (see QWidget::closeEvent()). (C++ enum variant: WA_DeleteOnClose = 55)

WARightToLeft

Indicates that the layout direction for the widget is right to left. (C++ enum variant: WA_RightToLeft = 56)

WASetLayoutDirection

C++ enum variant: WA_SetLayoutDirection = 57

WANoChildEventsForParent

Indicates that the widget does not want ChildAdded or ChildRemoved events sent to its parent. This is rarely necessary but can help to avoid automatic insertion widgets like splitters and layouts. This is set by a widget's author. (C++ enum variant: WA_NoChildEventsForParent = 58)

WAForceUpdatesDisabled

Indicates that updates are explicitly disabled for the widget; i.e. it will remain disabled even when all its ancestors are set to the updates-enabled state. This implies WA_UpdatesDisabled. This is set/cleared by QWidget::setUpdatesEnabled(). (C++ enum variant: WA_ForceUpdatesDisabled = 59)

WAWStateCreated

C++ enum variant: WA_WState_Created = 60

WAWStateCompressKeys

C++ enum variant: WA_WState_CompressKeys = 61

WAWStateInPaintEvent

C++ enum variant: WA_WState_InPaintEvent = 62

WAWStateReparented

C++ enum variant: WA_WState_Reparented = 63

WAWStateConfigPending

C++ enum variant: WA_WState_ConfigPending = 64

WAWStatePolished

C++ enum variant: WA_WState_Polished = 66

WAWStateDND

C++ enum variant: WA_WState_DND = 67

WAWStateOwnSizePolicy

C++ enum variant: WA_WState_OwnSizePolicy = 68

WAWStateExplicitShowHide

C++ enum variant: WA_WState_ExplicitShowHide = 69

WAShowModal

This attribute has been deprecated. Use QWidget::windowModality instead. (C++ enum variant: WA_ShowModal = 70)

WAMouseNoMask

Makes the widget receive mouse events for the entire widget regardless of the currently set mask, overriding QWidget::setMask(). This is not applicable for top-level windows. (C++ enum variant: WA_MouseNoMask = 71)

WAGroupLeader

This attribute has been deprecated. Use QWidget::windowModality instead. (C++ enum variant: WA_GroupLeader = 72)

WANoMousePropagation

Prohibits mouse events from being propagated to the widget's parent. This attribute is disabled by default. (C++ enum variant: WA_NoMousePropagation = 73)

WAHover

Forces Qt to generate paint events when the mouse enters or leaves the widget. This feature is typically used when implementing custom styles; see the Styles example for details. (C++ enum variant: WA_Hover = 74)

WAInputMethodTransparent

C++ enum variant: WA_InputMethodTransparent = 75

WAQuitOnClose

Makes Qt quit the application when the last widget with the attribute set has accepted closeEvent(). This behavior can be modified with the QApplication::quitOnLastWindowClosed property. By default this attribute is set for all widgets of type Qt::Window. (C++ enum variant: WA_QuitOnClose = 76)

WAKeyboardFocusChange

Set on a toplevel window when the users changes focus with the keyboard (tab, backtab, or shortcut). (C++ enum variant: WA_KeyboardFocusChange = 77)

WAAcceptDrops

Allows data from drag and drop operations to be dropped onto the widget (see QWidget::setAcceptDrops()). (C++ enum variant: WA_AcceptDrops = 78)

WADropSiteRegistered

This variant corresponds to multiple C++ enum variants with the same value:

WA_DropSiteRegistered = 79

WA_ForceAcceptDrops = 79

WAWindowPropagation

Makes a toplevel window inherit font and palette from its parent. (C++ enum variant: WA_WindowPropagation = 80)

WANoX11EventCompression

C++ enum variant: WA_NoX11EventCompression = 81

WATintedBackground

C++ enum variant: WA_TintedBackground = 82

WAX11OpenGLOverlay

C++ enum variant: WA_X11OpenGLOverlay = 83

WAAlwaysShowToolTips

Enables tooltips for inactive windows. (C++ enum variant: WA_AlwaysShowToolTips = 84)

WAMacOpaqueSizeGrip

Indicates that the native Carbon size grip should be opaque instead of transparent (the default). This attribute is only applicable to OS X and is set by the widget's author. (C++ enum variant: WA_MacOpaqueSizeGrip = 85)

WASetStyle

Indicates that the widget has a style of its own. This flag is set or cleared by QWidget::setStyle(). (C++ enum variant: WA_SetStyle = 86)

WASetLocale

Indicates the locale should be taken into consideration in the widget. (C++ enum variant: WA_SetLocale = 87)

WAMacShowFocusRect

Indicates that this widget should get a QFocusFrame around it. Some widgets draw their own focus halo regardless of this attribute. Not that the QWidget::focusPolicy also plays the main role in whether something is given focus or not, this only controls whether or not this gets the focus frame. This attribute is only applicable to OS X. (C++ enum variant: WA_MacShowFocusRect = 88)

WAMacNormalSize

Indicates the widget should have the normal size for widgets in OS X. This attribute is only applicable to OS X. (C++ enum variant: WA_MacNormalSize = 89)

WAMacSmallSize

Indicates the widget should have the small size for widgets in OS X. This attribute is only applicable to OS X. (C++ enum variant: WA_MacSmallSize = 90)

WAMacMiniSize

Indicates the widget should have the mini size for widgets in OS X. This attribute is only applicable to OS X. (C++ enum variant: WA_MacMiniSize = 91)

WALayoutUsesWidgetRect

Ignore the layout item rect from the style when laying out this widget with QLayout. (C++ enum variant: WA_LayoutUsesWidgetRect = 92)

WAStyledBackground

Indicates the widget should be drawn using a styled background. (C++ enum variant: WA_StyledBackground = 93)

WAMSWindowsUseDirect3D

This value is obsolete and has no effect. (C++ enum variant: WA_MSWindowsUseDirect3D = 94)

WACanHostQMdiSubWindowTitleBar

C++ enum variant: WA_CanHostQMdiSubWindowTitleBar = 95

WAMacAlwaysShowToolWindow

On OS X, show the tool window even when the application is not active. By default, all tool windows are hidden when the application is inactive. (C++ enum variant: WA_MacAlwaysShowToolWindow = 96)

WAStyleSheet

Indicates that the widget is styled using a style sheet. (C++ enum variant: WA_StyleSheet = 97)

WAShowWithoutActivating

Show the widget without making it active. (C++ enum variant: WA_ShowWithoutActivating = 98)

WAX11BypassTransientForHint

C++ enum variant: WA_X11BypassTransientForHint = 99

WANativeWindow

Indicates that a native window is created for the widget. Enabling this flag will also force a native window for the widget's ancestors unless Qt::WA_DontCreateNativeAncestors is set. (C++ enum variant: WA_NativeWindow = 100)

WADontCreateNativeAncestors

Indicates that the widget's ancestors are kept non-native even though the widget itself is native. (C++ enum variant: WA_DontCreateNativeAncestors = 101)

WAMacVariableSize

Indicates the widget can choose between alternative sizes for widgets to avoid clipping. This attribute is only applicable to OS X. (C++ enum variant: WA_MacVariableSize = 102)

WADontShowOnScreen

Indicates that the widget is hidden or is not a part of the viewable Desktop. (C++ enum variant: WA_DontShowOnScreen = 103)

WAX11NetWmWindowTypeDesktop

Adds _NET_WM_WINDOW_TYPE_DESKTOP to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. (C++ enum variant: WA_X11NetWmWindowTypeDesktop = 104)

WAX11NetWmWindowTypeDock

Adds _NET_WM_WINDOW_TYPE_DOCK to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. (C++ enum variant: WA_X11NetWmWindowTypeDock = 105)

WAX11NetWmWindowTypeToolBar

Adds _NET_WM_WINDOW_TYPE_TOOLBAR to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for QToolBar. (C++ enum variant: WA_X11NetWmWindowTypeToolBar = 106)

WAX11NetWmWindowTypeMenu

Adds _NET_WM_WINDOW_TYPE_MENU to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for QMenu when torn-off. (C++ enum variant: WA_X11NetWmWindowTypeMenu = 107)

WAX11NetWmWindowTypeUtility

Adds _NET_WM_WINDOW_TYPE_UTILITY to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for the Qt::Tool window type. (C++ enum variant: WA_X11NetWmWindowTypeUtility = 108)

WAX11NetWmWindowTypeSplash

Adds _NET_WM_WINDOW_TYPE_SPLASH to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for the Qt::SplashScreen window type. (C++ enum variant: WA_X11NetWmWindowTypeSplash = 109)

WAX11NetWmWindowTypeDialog

Adds _NET_WM_WINDOW_TYPE_DIALOG to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for the Qt::Dialog and Qt::Sheet window types. (C++ enum variant: WA_X11NetWmWindowTypeDialog = 110)

WAX11NetWmWindowTypeDropDownMenu

Adds _NET_WM_WINDOW_TYPE_DROPDOWN_MENU to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for QMenu objects added to a QMenuBar. (C++ enum variant: WA_X11NetWmWindowTypeDropDownMenu = 111)

WAX11NetWmWindowTypePopupMenu

Adds _NET_WM_WINDOW_TYPE_POPUP_MENU to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for QMenu. (C++ enum variant: WA_X11NetWmWindowTypePopupMenu = 112)

WAX11NetWmWindowTypeToolTip

Adds _NET_WM_WINDOW_TYPE_TOOLTIP to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for the Qt::ToolTip window type. (C++ enum variant: WA_X11NetWmWindowTypeToolTip = 113)

WAX11NetWmWindowTypeNotification

Adds _NET_WM_WINDOW_TYPE_NOTIFICATION to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. (C++ enum variant: WA_X11NetWmWindowTypeNotification = 114)

WAX11NetWmWindowTypeCombo

Adds _NET_WM_WINDOW_TYPE_COMBO to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute for the QComboBox pop-up. (C++ enum variant: WA_X11NetWmWindowTypeCombo = 115)

WAX11NetWmWindowTypeDND

Adds _NET_WM_WINDOW_TYPE_DND to the window's _NET_WM_WINDOW_TYPE X11 window property. See http://standards.freedesktop.org/wm-spec/ for more details. This attribute has no effect on non-X11 platforms. Note: Qt automatically sets this attribute on the feedback widget used during a drag. (C++ enum variant: WA_X11NetWmWindowTypeDND = 116)

WAMacFrameworkScaled

Enables resolution independence aware mode on Mac when using Carbon. This attribute has no effect on Cocoa. The attribute is off by default and can be enabled on a per-window basis. (C++ enum variant: WA_MacFrameworkScaled = 117)

WASetWindowModality

C++ enum variant: WA_SetWindowModality = 118

WAWStateWindowOpacitySet

C++ enum variant: WA_WState_WindowOpacitySet = 119

WATranslucentBackground

Indicates that the widget should have a translucent background, i.e., any non-opaque regions of the widgets will be translucent because the widget will have an alpha channel. Setting this flag causes WA_NoSystemBackground to be set. On Windows the widget also needs the Qt::FramelessWindowHint window flag to be set. This flag is set or cleared by the widget's author. (C++ enum variant: WA_TranslucentBackground = 120)

WAAcceptTouchEvents

Allows touch events (see QTouchEvent) to be sent to the widget. Must be set on all widgets that can handle touch events. Without this attribute set, events from a touch device will be sent as mouse events. (C++ enum variant: WA_AcceptTouchEvents = 121)

WAWStateAcceptedTouchBeginEvent

C++ enum variant: WA_WState_AcceptedTouchBeginEvent = 122

WATouchPadAcceptSingleTouchEvents

Allows touchpad single touch events to be sent to the widget. (C++ enum variant: WA_TouchPadAcceptSingleTouchEvents = 123)

WAX11DoNotAcceptFocus

Asks the window manager to not give focus to this top level window. This attribute has no effect on non-X11 platforms. (C++ enum variant: WA_X11DoNotAcceptFocus = 126)

WAMacNoShadow

C++ enum variant: WA_MacNoShadow = 127

WAAlwaysStackOnTop

Since Qt 5.4, this value forces QOpenGLWidget and QQuickWidget to be drawn last, on top of other widgets. Ignored for other type of widgets. Setting this attribute breaks the stacking order, but allows having a semi-transparent OpenGL widget with other widgets visible underneath. It is strongly recommended to call update() on the widget's top-level window after enabling or disabling this attribute. (C++ enum variant: WA_AlwaysStackOnTop = 128)

WAAttributeCount

C++ enum variant: WA_AttributeCount = 129