Enum qt_core::qt::WindowType

pub enum WindowType {
    WindowFullscreenButtonHint,
    Widget,
    Window,
    Dialog,
    Sheet,
    Drawer,
    Popup,
    Tool,
    ToolTip,
    SplashScreen,
    Desktop,
    SubWindow,
    ForeignWindow,
    CoverWindow,
    WindowTypeMask,
    MswIndowsFixedSizeDialogHint,
    MswIndowsOwnDC,
    BypassWindowManagerHint,
    FramelessWindowHint,
    WindowTitleHint,
    WindowSystemMenuHint,
    WindowMinimizeButtonHint,
    WindowMaximizeButtonHint,
    WindowMinMaxButtonsHint,
    WindowContextHelpButtonHint,
    WindowShadeButtonHint,
    WindowStaysOnTopHint,
    WindowTransparentForInput,
    WindowOverridesSystemGestures,
    WindowDoesNotAcceptFocus,
    MaximizeUsingFullscreenGeometryHint,
    CustomizeWindowHint,
    WindowStaysOnBottomHint,
    WindowCloseButtonHint,
    MacWindowToolBarButtonHint,
    BypassGraphicsProxyWidget,
    NoDropShadowWindowHint,
}

C++ type: Qt::WindowType

C++ documentation:

This enum type is used to specify various window-system properties for the widget. They are fairly unusual but necessary in a few cases. Some of these flags depend on whether the underlying window manager supports them.

The main types are

There are also a number of flags which you can use to customize the appearance of top-level windows. These have no effect on other windows:

The CustomizeWindowHint flag is used to enable customization of the window controls. This flag must be set to allow the WindowTitleHint, WindowSystemMenuHint, WindowMinimizeButtonHint, WindowMaximizeButtonHint and WindowCloseButtonHint flags to be changed.

The WindowFlags type is a typedef for QFlags<WindowType>. It stores an OR combination of WindowType values.

See also QWidget::windowFlags and Window Flags Example.

Variants
WindowFullscreenButtonHint

On OS X adds a fullscreen button. (C++ enum variant: WindowFullscreenButtonHint = -2147483648)

Widget

This is the default type for QWidget. Widgets of this type are child widgets if they have a parent, and independent windows if they have no parent. See also Qt::Window and Qt::SubWindow. (C++ enum variant: Widget = 0)

Window

Indicates that the widget is a window, usually with a window system frame and a title bar, irrespective of whether the widget has a parent or not. Note that it is not possible to unset this flag if the widget does not have a parent. (C++ enum variant: Window = 1)

Dialog

Indicates that the widget is a window that should be decorated as a dialog (i.e., typically no maximize or minimize buttons in the title bar). This is the default type for QDialog. If you want to use it as a modal dialog, it should be launched from another window, or have a parent and used with the QWidget::windowModality property. If you make it modal, the dialog will prevent other top-level windows in the application from getting any input. We refer to a top-level window that has a parent as a secondary window. (C++ enum variant: Dialog = 3)

Sheet

Indicates that the window is a Macintosh sheet. Since using a sheet implies window modality, the recommended way is to use QWidget::setWindowModality(), or QDialog::open(), instead. (C++ enum variant: Sheet = 5)

Drawer

Indicates that the widget is a Macintosh drawer. (C++ enum variant: Drawer = 7)

Popup

Indicates that the widget is a pop-up top-level window, i.e. that it is modal, but has a window system frame appropriate for pop-up menus. (C++ enum variant: Popup = 9)

Tool

Indicates that the widget is a tool window. A tool window is often a small window with a smaller than usual title bar and decoration, typically used for collections of tool buttons. If there is a parent, the tool window will always be kept on top of it. If there isn't a parent, you may consider using Qt::WindowStaysOnTopHint as well. If the window system supports it, a tool window can be decorated with a somewhat lighter frame. It can also be combined with Qt::FramelessWindowHint. On OS X, tool windows correspond to the Floating class of windows. This means that the window lives on a level above normal windows; it impossible to put a normal window on top of it. By default, tool windows will disappear when the application is inactive. This can be controlled by the Qt::WA_MacAlwaysShowToolWindow attribute. (C++ enum variant: Tool = 11)

ToolTip

Indicates that the widget is a tooltip. This is used internally to implement tooltips. (C++ enum variant: ToolTip = 13)

SplashScreen

Indicates that the window is a splash screen. This is the default type for QSplashScreen. (C++ enum variant: SplashScreen = 15)

Desktop

Indicates that this widget is the desktop. This is the type for QDesktopWidget. (C++ enum variant: Desktop = 17)

SubWindow

Indicates that this widget is a sub-window, such as a QMdiSubWindow widget. (C++ enum variant: SubWindow = 18)

ForeignWindow

Indicates that this window object is a handle representing a native platform window created by another process or by manually using native code. (C++ enum variant: ForeignWindow = 33)

CoverWindow

Indicates that the window represents a cover window, which is shown when the application is minimized on the BlackBerry platform for instance. (C++ enum variant: CoverWindow = 65)

WindowTypeMask

A mask for extracting the window type part of the window flags. (C++ enum variant: WindowType_Mask = 255)

MswIndowsFixedSizeDialogHint

Gives the window a thin dialog border on Windows. This style is traditionally used for fixed-size dialogs. (C++ enum variant: MSWindowsFixedSizeDialogHint = 256)

MswIndowsOwnDC

Gives the window its own display context on Windows. (C++ enum variant: MSWindowsOwnDC = 512)

BypassWindowManagerHint

This variant corresponds to multiple C++ enum variants with the same value:

BypassWindowManagerHint = 1024: This flag can be used to indicate to the platform plugin that "all" window manager protocols should be disabled. This flag will behave different depending on what operating system the application is running on and what window manager is running. The flag can be used to get a native window with no configuration set.

X11BypassWindowManagerHint = 1024: Bypass the window manager completely. This results in a borderless window that is not managed at all (i.e., no keyboard input unless you call QWidget::activateWindow() manually).

FramelessWindowHint

Produces a borderless window. The user cannot move or resize a borderless window via the window system. On X11, the result of the flag is dependent on the window manager and its ability to understand Motif and/or NETWM hints. Most existing modern window managers can handle this. (C++ enum variant: FramelessWindowHint = 2048)

WindowTitleHint

Gives the window a title bar. (C++ enum variant: WindowTitleHint = 4096)

WindowSystemMenuHint

Adds a window system menu, and possibly a close button (for example on Mac). If you need to hide or show a close button, it is more portable to use WindowCloseButtonHint. (C++ enum variant: WindowSystemMenuHint = 8192)

WindowMinimizeButtonHint

Adds a minimize button. On some platforms this implies Qt::WindowSystemMenuHint for it to work. (C++ enum variant: WindowMinimizeButtonHint = 16384)

WindowMaximizeButtonHint

Adds a maximize button. On some platforms this implies Qt::WindowSystemMenuHint for it to work. (C++ enum variant: WindowMaximizeButtonHint = 32768)

WindowMinMaxButtonsHint

Adds a minimize and a maximize button. On some platforms this implies Qt::WindowSystemMenuHint for it to work. (C++ enum variant: WindowMinMaxButtonsHint = 49152)

WindowContextHelpButtonHint

Adds a context help button to dialogs. On some platforms this implies Qt::WindowSystemMenuHint for it to work. (C++ enum variant: WindowContextHelpButtonHint = 65536)

WindowShadeButtonHint

Adds a shade button in place of the minimize button if the underlying window manager supports it. (C++ enum variant: WindowShadeButtonHint = 131072)

WindowStaysOnTopHint

Informs the window system that the window should stay on top of all other windows. Note that on some window managers on X11 you also have to pass Qt::X11BypassWindowManagerHint for this flag to work correctly. (C++ enum variant: WindowStaysOnTopHint = 262144)

WindowTransparentForInput

This variant corresponds to multiple C++ enum variants with the same value:

WindowTransparentForInput = 524288: Informs the window system that this window is used only for output (displaying something) and does not take input. Therefore input events should pass through as if it wasn't there.

WindowOkButtonHint = 524288: Adds an OK button to the window decoration of a dialog. Only supported for Windows CE.

WindowOverridesSystemGestures

This variant corresponds to multiple C++ enum variants with the same value:

WindowOverridesSystemGestures = 1048576: Informs the window system that this window implements its own set of gestures and that system level gestures, like for instance three-finger desktop switching, should be disabled.

WindowCancelButtonHint = 1048576: Adds a Cancel button to the window decoration of a dialog. Only supported for Windows CE.

WindowDoesNotAcceptFocus

Informs the window system that this window should not receive the input focus. (C++ enum variant: WindowDoesNotAcceptFocus = 2097152)

MaximizeUsingFullscreenGeometryHint

Informs the window system that when maximizing the window it should use as much of the available screen geometry as possible, including areas that may be covered by system UI such as status bars or application launchers. This may result in the window being placed under these system UIs, but does not guarantee it, depending on whether or not the platform supports it. When the flag is enabled the user is responsible for taking QScreen::availableGeometry() into account, so that any UI elements in the application that require user interaction are not covered by system UI. (C++ enum variant: MaximizeUsingFullscreenGeometryHint = 4194304)

CustomizeWindowHint

Turns off the default window title hints. (C++ enum variant: CustomizeWindowHint = 33554432)

WindowStaysOnBottomHint

Informs the window system that the window should stay on bottom of all other windows. Note that on X11 this hint will work only in window managers that support _NET_WM_STATE_BELOW atom. If a window always on the bottom has a parent, the parent will also be left on the bottom. This window hint is currently not implemented for OS X. (C++ enum variant: WindowStaysOnBottomHint = 67108864)

WindowCloseButtonHint

Adds a close button. On some platforms this implies Qt::WindowSystemMenuHint for it to work. (C++ enum variant: WindowCloseButtonHint = 134217728)

MacWindowToolBarButtonHint

On OS X adds a tool bar button (i.e., the oblong button that is on the top right of windows that have toolbars). (C++ enum variant: MacWindowToolBarButtonHint = 268435456)

BypassGraphicsProxyWidget

Prevents the window and its children from automatically embedding themselves into a QGraphicsProxyWidget if the parent widget is already embedded. You can set this flag if you want your widget to always be a toplevel widget on the desktop, regardless of whether the parent widget is embedded in a scene or not. (C++ enum variant: BypassGraphicsProxyWidget = 536870912)

NoDropShadowWindowHint

Disables window drop shadow on supporting platforms. (C++ enum variant: NoDropShadowWindowHint = 1073741824)