Struct qt_core::read_locker::ReadLocker

pub struct ReadLocker {
    // some fields omitted
}

C++ type: QReadLocker

C++ documentation:

The QReadLocker class is a convenience class that simplifies locking and unlocking read-write locks for read access.

The purpose of QReadLocker (and QWriteLocker) is to simplify QReadWriteLock locking and unlocking. Locking and unlocking statements or in exception handling code is error-prone and difficult to debug. QReadLocker can be used in such situations to ensure that the state of the lock is always well-defined.

Here's an example that uses QReadLocker to lock and unlock a read-write lock for reading:

QReadWriteLock lock;

QByteArray readData() { QReadLocker locker(&lock); ... return data; }
It is equivalent to the following code:

QReadWriteLock lock;

QByteArray readData() { lock.lockForRead(); ... lock.unlock(); return data; }
The QMutexLocker documentation shows examples where the use of a locker object greatly simplifies programming.

Methods

impl ReadLocker

fn new<Args>(args: Args) -> Args::ReturnType where Args: ReadLockerNewArgs

C++ method: QReadLocker::QReadLocker

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::read_write_lock::ReadWriteLock, cpp_utils::AsStruct)) -> ::read_locker::ReadLocker
2) fn new((*mut ::read_write_lock::ReadWriteLock, cpp_utils::AsBox)) -> cpp_utils::CppBox<::read_locker::ReadLocker>
C++ method: [constructor] void QReadLocker::QReadLocker(QReadWriteLock* readWriteLock)

fn read_write_lock(&self) -> *mut ReadWriteLock

C++ method: QReadWriteLock* QReadLocker::readWriteLock() const

C++ documentation:

Returns a pointer to the read-write lock that was passed to the constructor.

fn relock(&mut self)

C++ method: void QReadLocker::relock()

C++ documentation:

Relocks an unlocked lock.

See also unlock().

fn unlock(&mut self)

C++ method: void QReadLocker::unlock()

C++ documentation:

Unlocks the lock associated with this locker.

See also QReadWriteLock::unlock().