Struct qt_core::read_write_lock::ReadWriteLock

pub struct ReadWriteLock {
    // some fields omitted
}

C++ type: QReadWriteLock

C++ documentation:

The QReadWriteLock class provides read-write locking.

A read-write lock is a synchronization tool for protecting resources that can be accessed for reading and writing. This type of lock is useful if you want to allow multiple threads to have simultaneous read-only access, but as soon as one thread wants to write to the resource, all other threads must be blocked until the writing is complete.

In many cases, QReadWriteLock is a direct competitor to QMutex. QReadWriteLock is a good choice if there are many concurrent reads and writing occurs infrequently.

Example:

QReadWriteLock lock;

void ReaderThread::run() { ... lock.lockForRead(); read_file(); lock.unlock(); ... }

void WriterThread::run() { ... lock.lockForWrite(); write_file(); lock.unlock(); ... }
To ensure that writers aren't blocked forever by readers, readers attempting to obtain a lock will not succeed if there is a blocked writer waiting for access, even if the lock is currently only accessed by other readers. Also, if the lock is accessed by a writer and another writer comes in, that writer will have priority over any readers that might also be waiting.

Like QMutex, a QReadWriteLock can be recursively locked by the same thread when constructed with QReadWriteLock::Recursive as QReadWriteLock::RecursionMode. In such cases, unlock() must be called the same number of times lockForWrite() or lockForRead() was called. Note that the lock type cannot be changed when trying to lock recursively, i.e. it is not possible to lock for reading in a thread that already has locked for writing (and vice versa).

Methods

impl ReadWriteLock

fn lock_for_read(&mut self)

C++ method: void QReadWriteLock::lockForRead()

C++ documentation:

Locks the lock for reading. This function will block the current thread if another thread has locked for writing.

It is not possible to lock for read if the thread already has locked for write.

See also unlock(), lockForWrite(), and tryLockForRead().

fn lock_for_write(&mut self)

C++ method: void QReadWriteLock::lockForWrite()

C++ documentation:

Locks the lock for writing. This function will block the current thread if another thread (including the current) has locked for reading or writing (unless the lock has been created using the QReadWriteLock::Recursive mode).

It is not possible to lock for write if the thread already has locked for read.

See also unlock(), lockForRead(), and tryLockForWrite().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ReadWriteLockNewArgs

C++ method: QReadWriteLock::QReadWriteLock

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::read_write_lock::ReadWriteLock
2) fn new((::read_write_lock::RecursionMode, cpp_utils::AsStruct)) -> ::read_write_lock::ReadWriteLock
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::read_write_lock::ReadWriteLock>
4) fn new((::read_write_lock::RecursionMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::read_write_lock::ReadWriteLock>
C++ method: [constructor] void QReadWriteLock::QReadWriteLock(QReadWriteLock::RecursionMode recursionMode = ?)

fn try_lock_for_read<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: ReadWriteLockTryLockForReadArgs<'largs>

C++ method: QReadWriteLock::tryLockForRead

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn try_lock_for_read(&mut self, ()) -> bool
C++ method: bool QReadWriteLock::tryLockForRead()

C++ documentation:

Attempts to lock for reading. If the lock was obtained, this function returns true, otherwise it returns false instead of waiting for the lock to become available, i.e. it does not block.

The lock attempt will fail if another thread has locked for writing.

If the lock was obtained, the lock must be unlocked with unlock() before another thread can successfully lock it for writing.

It is not possible to lock for read if the thread already has locked for write.

See also unlock() and lockForRead().

Variant 2

Rust arguments: fn try_lock_for_read(&mut self, libc::c_int) -> bool
C++ method: bool QReadWriteLock::tryLockForRead(int timeout)

C++ documentation:

This is an overloaded function.

Attempts to lock for reading. This function returns true if the lock was obtained; otherwise it returns false. If another thread has locked for writing, this function will wait for at most timeout milliseconds for the lock to become available.

Note: Passing a negative number as the timeout is equivalent to calling lockForRead(), i.e. this function will wait forever until lock can be locked for reading when timeout is negative.

If the lock was obtained, the lock must be unlocked with unlock() before another thread can successfully lock it for writing.

It is not possible to lock for read if the thread already has locked for write.

See also unlock() and lockForRead().

fn try_lock_for_write<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: ReadWriteLockTryLockForWriteArgs<'largs>

C++ method: QReadWriteLock::tryLockForWrite

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn try_lock_for_write(&mut self, ()) -> bool
C++ method: bool QReadWriteLock::tryLockForWrite()

C++ documentation:

Attempts to lock for writing. If the lock was obtained, this function returns true; otherwise, it returns false immediately.

The lock attempt will fail if another thread has locked for reading or writing.

If the lock was obtained, the lock must be unlocked with unlock() before another thread can successfully lock it.

It is not possible to lock for write if the thread already has locked for read.

See also unlock() and lockForWrite().

Variant 2

Rust arguments: fn try_lock_for_write(&mut self, libc::c_int) -> bool
C++ method: bool QReadWriteLock::tryLockForWrite(int timeout)

C++ documentation:

This is an overloaded function.

Attempts to lock for writing. This function returns true if the lock was obtained; otherwise it returns false. If another thread has locked for reading or writing, this function will wait for at most timeout milliseconds for the lock to become available.

Note: Passing a negative number as the timeout is equivalent to calling lockForWrite(), i.e. this function will wait forever until lock can be locked for writing when timeout is negative.

If the lock was obtained, the lock must be unlocked with unlock() before another thread can successfully lock it.

It is not possible to lock for write if the thread already has locked for read.

See also unlock() and lockForWrite().

fn unlock(&mut self)

C++ method: void QReadWriteLock::unlock()

C++ documentation:

Unlocks the lock.

Attempting to unlock a lock that is not locked is an error, and will result in program termination.

See also lockForRead(), lockForWrite(), tryLockForRead(), and tryLockForWrite().