Module qt_core::rect::overloading

Traits

OpAddArgs
This trait represents a set of arguments accepted by op_add method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

OpShrArgs
This trait represents a set of arguments accepted by op_shr method.

OpSubArgs
This trait represents a set of arguments accepted by op_sub method.

RectAdjustedArgs
This trait represents a set of arguments accepted by Rect::adjusted method.

RectBottomLeftArgs
This trait represents a set of arguments accepted by Rect::bottom_left method.

RectBottomRightArgs
This trait represents a set of arguments accepted by Rect::bottom_right method.

RectCenterArgs
This trait represents a set of arguments accepted by Rect::center method.

RectContainsArgs
This trait represents a set of arguments accepted by Rect::contains method.

RectIntersectedArgs
This trait represents a set of arguments accepted by Rect::intersected method.

RectMarginsAddedArgs
This trait represents a set of arguments accepted by Rect::margins_added method.

RectMarginsRemovedArgs
This trait represents a set of arguments accepted by Rect::margins_removed method.

RectMoveToArgs
This trait represents a set of arguments accepted by Rect::move_to method.

RectNewArgs
This trait represents a set of arguments accepted by Rect::new method.

RectNormalizedArgs
This trait represents a set of arguments accepted by Rect::normalized method.

RectOpBitAndArgs
This trait represents a set of arguments accepted by Rect::op_bit_and method.

RectOpBitOrArgs
This trait represents a set of arguments accepted by Rect::op_bit_or method.

RectSizeArgs
This trait represents a set of arguments accepted by Rect::size method.

RectTopLeftArgs
This trait represents a set of arguments accepted by Rect::top_left method.

RectTopRightArgs
This trait represents a set of arguments accepted by Rect::top_right method.

RectTranslateArgs
This trait represents a set of arguments accepted by Rect::translate method.

RectTranslatedArgs
This trait represents a set of arguments accepted by Rect::translated method.

RectUnitedArgs
This trait represents a set of arguments accepted by Rect::united method.