Module qt_core::reg_exp::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

RegExpCapArgs
This trait represents a set of arguments accepted by RegExp::cap method.

RegExpCapMutArgs
This trait represents a set of arguments accepted by RegExp::cap_mut method.

RegExpCapturedTextsArgs
This trait represents a set of arguments accepted by RegExp::captured_texts method.

RegExpCapturedTextsMutArgs
This trait represents a set of arguments accepted by RegExp::captured_texts_mut method.

RegExpErrorStringArgs
This trait represents a set of arguments accepted by RegExp::error_string method.

RegExpErrorStringMutArgs
This trait represents a set of arguments accepted by RegExp::error_string_mut method.

RegExpEscapeArgs
This trait represents a set of arguments accepted by RegExp::escape method.

RegExpIndexInArgs
This trait represents a set of arguments accepted by RegExp::index_in method.

RegExpLastIndexInArgs
This trait represents a set of arguments accepted by RegExp::last_index_in method.

RegExpNewArgs
This trait represents a set of arguments accepted by RegExp::new method.

RegExpPatternArgs
This trait represents a set of arguments accepted by RegExp::pattern method.

RegExpPosArgs
This trait represents a set of arguments accepted by RegExp::pos method.

RegExpPosMutArgs
This trait represents a set of arguments accepted by RegExp::pos_mut method.