Struct qt_core::reg_exp::RegExp

pub struct RegExp {
    // some fields omitted
}

C++ type: QRegExp

C++ documentation:

The QRegExp class provides pattern matching using regular expressions.

A regular expression, or "regexp", is a pattern for matching substrings in a text. This is useful in many contexts, e.g.,

Validation	A regexp can test whether a substring meets some criteria, e.g. is an integer or contains no whitespace.

Searching	A regexp provides more powerful pattern matching than simple substring matching, e.g., match one of the words mail, letter or correspondence, but none of the words email, mailman, mailer, letterbox, etc.

Search and Replace	A regexp can replace all occurrences of a substring with a different substring, e.g., replace all occurrences of & with &amp; except where the & is already followed by an amp;.

String Splitting	A regexp can be used to identify where a string should be split apart, e.g. splitting tab-delimited strings.

A brief introduction to regexps is presented, a description of Qt's regexp language, some examples, and the function documentation itself. QRegExp is modeled on Perl's regexp language. It fully supports Unicode. QRegExp can also be used in a simpler, wildcard mode that is similar to the functionality found in command shells. The syntax rules used by QRegExp can be changed with setPatternSyntax(). In particular, the pattern syntax can be set to QRegExp::FixedString, which means the pattern to be matched is interpreted as a plain string, i.e., special characters (e.g., backslash) are not escaped.

A good text on regexps is Mastering Regular Expressions (Third Edition) by Jeffrey E. F. Friedl, ISBN 0-596-52812-4.

Methods

impl RegExp

fn cap<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegExpCapArgs<'largs>

C++ method: QRegExp::cap

This is an overloaded function. Available variants:

Rust arguments:
1) fn cap(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn cap(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn cap(&self, cpp_utils::AsStruct) -> ::string::String
4) fn cap(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QRegExp::cap(int nth = ?) const

C++ documentation:

Returns the text captured by the nth subexpression. The entire match has index 0 and the parenthesized subexpressions have indexes starting from 1 (excluding non-capturing parentheses).

QRegExp rxlen("(\d+)(?:\s*)(cm|inch)");
int pos = rxlen.indexIn("Length: 189cm");
if (pos > -1) {
    QString value = rxlen.cap(1); // "189"
    QString unit = rxlen.cap(2);  // "cm"
    // ...
}

The order of elements matched by cap() is as follows. The first element, cap(0), is the entire matching string. Each subsequent element corresponds to the next capturing open left parentheses. Thus cap(1) is the text of the first capturing parentheses, cap(2) is the text of the second, and so on.

See also capturedTexts() and pos().

fn cap_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: RegExpCapMutArgs<'largs>

C++ method: QRegExp::cap

This is an overloaded function. Available variants:

Rust arguments:
1) fn cap_mut(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn cap_mut(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn cap_mut(&mut self, cpp_utils::AsStruct) -> ::string::String
4) fn cap_mut(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QRegExp::cap(int nth = ?)

C++ documentation:

Returns the text captured by the nth subexpression. The entire match has index 0 and the parenthesized subexpressions have indexes starting from 1 (excluding non-capturing parentheses).

QRegExp rxlen("(\d+)(?:\s*)(cm|inch)");
int pos = rxlen.indexIn("Length: 189cm");
if (pos > -1) {
    QString value = rxlen.cap(1); // "189"
    QString unit = rxlen.cap(2);  // "cm"
    // ...
}

The order of elements matched by cap() is as follows. The first element, cap(0), is the entire matching string. Each subsequent element corresponds to the next capturing open left parentheses. Thus cap(1) is the text of the first capturing parentheses, cap(2) is the text of the second, and so on.

See also capturedTexts() and pos().

fn capture_count(&self) -> c_int

C++ method: int QRegExp::captureCount() const

C++ documentation:

Returns the number of captures contained in the regular expression.

This function was introduced in Qt 4.6.

fn captured_texts<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegExpCapturedTextsArgs<'largs>

C++ method: QRegExp::capturedTexts

This is an overloaded function. Available variants:

Rust arguments:
1) fn captured_texts(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn captured_texts(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QRegExp::capturedTexts() const

C++ documentation:

Returns a list of the captured text strings.

The first string in the list is the entire matched string. Each subsequent list element contains a string that matched a (capturing) subexpression of the regexp.

For example:

QRegExp rx("(\d+)(\s*)(cm|inch(es)?)");
int pos = rx.indexIn("Length: 36 inches");
QStringList list = rx.capturedTexts();
// list is now ("36 inches", "36", " ", "inches", "es")

The above example also captures elements that may be present but which we have no interest in. This problem can be solved by using non-capturing parentheses:

QRegExp rx("(\d+)(?:\s*)(cm|inch(?:es)?)");
int pos = rx.indexIn("Length: 36 inches");
QStringList list = rx.capturedTexts();
// list is now ("36 inches", "36", "inches")

Note that if you want to iterate over the list, you should iterate over a copy, e.g.

QStringList list = rx.capturedTexts();
QStringList::iterator it = list.begin();
while (it != list.end()) {
    myProcessing(*it);
    ++it;
}

Some regexps can match an indeterminate number of times. For example if the input string is "Offsets: 12 14 99 231 7" and the regexp, rx, is (\d+)+, we would hope to get a list of all the numbers matched. However, after calling rx.indexIn(str), capturedTexts() will return the list ("12", "12"), i.e. the entire match was "12" and the first subexpression matched was "12". The correct approach is to use cap() in a loop.

The order of elements in the string list is as follows. The first element is the entire matching string. Each subsequent element corresponds to the next capturing open left parentheses. Thus capturedTexts()[1] is the text of the first capturing parentheses, capturedTexts()[2] is the text of the second and so on (corresponding to $1, $2, etc., in some other regexp languages).

See also cap() and pos().

fn captured_texts_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: RegExpCapturedTextsMutArgs<'largs>

C++ method: QRegExp::capturedTexts

This is an overloaded function. Available variants:

Rust arguments:
1) fn captured_texts_mut(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn captured_texts_mut(&mut self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QRegExp::capturedTexts()

C++ documentation:

Returns a list of the captured text strings.

The first string in the list is the entire matched string. Each subsequent list element contains a string that matched a (capturing) subexpression of the regexp.

For example:

QRegExp rx("(\d+)(\s*)(cm|inch(es)?)");
int pos = rx.indexIn("Length: 36 inches");
QStringList list = rx.capturedTexts();
// list is now ("36 inches", "36", " ", "inches", "es")

The above example also captures elements that may be present but which we have no interest in. This problem can be solved by using non-capturing parentheses:

QRegExp rx("(\d+)(?:\s*)(cm|inch(?:es)?)");
int pos = rx.indexIn("Length: 36 inches");
QStringList list = rx.capturedTexts();
// list is now ("36 inches", "36", "inches")

Note that if you want to iterate over the list, you should iterate over a copy, e.g.

QStringList list = rx.capturedTexts();
QStringList::iterator it = list.begin();
while (it != list.end()) {
    myProcessing(*it);
    ++it;
}

Some regexps can match an indeterminate number of times. For example if the input string is "Offsets: 12 14 99 231 7" and the regexp, rx, is (\d+)+, we would hope to get a list of all the numbers matched. However, after calling rx.indexIn(str), capturedTexts() will return the list ("12", "12"), i.e. the entire match was "12" and the first subexpression matched was "12". The correct approach is to use cap() in a loop.

The order of elements in the string list is as follows. The first element is the entire matching string. Each subsequent element corresponds to the next capturing open left parentheses. Thus capturedTexts()[1] is the text of the first capturing parentheses, capturedTexts()[2] is the text of the second and so on (corresponding to $1, $2, etc., in some other regexp languages).

See also cap() and pos().

fn case_sensitivity(&self) -> CaseSensitivity

C++ method: Qt::CaseSensitivity QRegExp::caseSensitivity() const

C++ documentation:

Returns Qt::CaseSensitive if the regexp is matched case sensitively; otherwise returns Qt::CaseInsensitive.

See also setCaseSensitivity(), patternSyntax(), pattern(), and isMinimal().

fn error_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegExpErrorStringArgs<'largs>

C++ method: QRegExp::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn error_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QRegExp::errorString() const

C++ documentation:

Returns a text string that explains why a regexp pattern is invalid the case being; otherwise returns "no error occurred".

See also isValid().

fn error_string_mut<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: RegExpErrorStringMutArgs<'largs>

C++ method: QRegExp::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string_mut(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn error_string_mut(&mut self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QRegExp::errorString()

C++ documentation:

Returns a text string that explains why a regexp pattern is invalid the case being; otherwise returns "no error occurred".

See also isValid().

fn escape<Args>(args: Args) -> Args::ReturnType where Args: RegExpEscapeArgs

C++ method: QRegExp::escape

This is an overloaded function. Available variants:

Rust arguments:
1) fn escape((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn escape((&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QRegExp::escape(const QString& str)

C++ documentation:

Returns the string str with every regexp special character escaped with a backslash. The special characters are $, (,), *, +, ., ?, [, ,], , {, | and }.

Example:

s1 = QRegExp::escape("bingo");   // s1 == "bingo"
s2 = QRegExp::escape("f(x)");    // s2 == "f\(x\)"

This function is useful to construct regexp patterns dynamically:

QRegExp rx("(" + QRegExp::escape(name) +
           "|" + QRegExp::escape(alias) + ")");

See also setPatternSyntax().

fn exact_match(&self, str: &String) -> bool

C++ method: bool QRegExp::exactMatch(const QString& str) const

C++ documentation:

Returns true if str is matched exactly by this regular expression; otherwise returns false. You can determine how much of the string was matched by calling matchedLength().

For a given regexp string R, exactMatch("R") is the equivalent of indexIn("R$") since exactMatch() effectively encloses the regexp in the start of string and end of string anchors, except that it sets matchedLength() differently.

For example, if the regular expression is blue, then exactMatch() returns true only for input blue. For inputs bluebell, blutak and lightblue, exactMatch() returns false and matchedLength() will return 4, 3 and 0 respectively.

Although const, this function sets matchedLength(), capturedTexts(), and pos().

See also indexIn() and lastIndexIn().

fn index_in<'largs, Args>(&'largs self, args: Args) -> c_int where Args: RegExpIndexInArgs<'largs>

C++ method: QRegExp::indexIn

This is an overloaded function. Available variants:

Rust arguments:
1) fn index_in(&self, &::string::String) -> libc::c_int
2) fn index_in(&self, (&::string::String, libc::c_int)) -> libc::c_int
3) fn index_in(&self, (&::string::String, libc::c_int, ::reg_exp::CaretMode)) -> libc::c_int
C++ method: int QRegExp::indexIn(const QString& str, int offset = ?, QRegExp::CaretMode caretMode = ?) const

C++ documentation:

Attempts to find a match in str from position offset (0 by default). If offset is -1, the search starts at the last character; if -2, at the next to last character; etc.

Returns the position of the first match, or -1 if there was no match.

The caretMode parameter can be used to instruct whether should match at index 0 or at offset.

You might prefer to use QString::indexOf(), QString::contains(), or even QStringList::filter(). To replace matches use QString::replace().

Example:

QString str = "offsets: 1.23 .50 71.00 6.00";
QRegExp rx("\d*\.\d+");    // primitive floating point matching
int count = 0;
int pos = 0;
while ((pos = rx.indexIn(str, pos)) != -1) {
    ++count;
    pos += rx.matchedLength();
}
// pos will be 9, 14, 18 and finally 24; count will end up as 4

Although const, this function sets matchedLength(), capturedTexts() and pos().

If the QRegExp is a wildcard expression (see setPatternSyntax()) and want to test a string against the whole wildcard expression, use exactMatch() instead of this function.

See also lastIndexIn() and exactMatch().

fn is_empty(&self) -> bool

C++ method: bool QRegExp::isEmpty() const

C++ documentation:

Returns true if the pattern string is empty; otherwise returns false.

If you call exactMatch() with an empty pattern on an empty string it will return true; otherwise it returns false since it operates over the whole string. If you call indexIn() with an empty pattern on any string it will return the start offset (0 by default) because the empty pattern matches the 'emptiness' at the start of the string. In this case the length of the match returned by matchedLength() will be 0.

See QString::isEmpty().

fn is_minimal(&self) -> bool

C++ method: bool QRegExp::isMinimal() const

C++ documentation:

Returns true if minimal (non-greedy) matching is enabled; otherwise returns false.

See also caseSensitivity() and setMinimal().

fn is_valid(&self) -> bool

C++ method: bool QRegExp::isValid() const

C++ documentation:

Returns true if the regular expression is valid; otherwise returns false. An invalid regular expression never matches.

The pattern [a-z is an example of an invalid pattern, since it lacks a closing square bracket.

Note that the validity of a regexp may also depend on the setting of the wildcard flag, for example *.html is a valid wildcard regexp but an invalid full regexp.

See also errorString().

fn last_index_in<'largs, Args>(&'largs self, args: Args) -> c_int where Args: RegExpLastIndexInArgs<'largs>

C++ method: QRegExp::lastIndexIn

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_index_in(&self, &::string::String) -> libc::c_int
2) fn last_index_in(&self, (&::string::String, libc::c_int)) -> libc::c_int
3) fn last_index_in(&self, (&::string::String, libc::c_int, ::reg_exp::CaretMode)) -> libc::c_int
C++ method: int QRegExp::lastIndexIn(const QString& str, int offset = ?, QRegExp::CaretMode caretMode = ?) const

C++ documentation:

Attempts to find a match backwards in str from position offset. If offset is -1 (the default), the search starts at the last character; if -2, at the next to last character; etc.

Returns the position of the first match, or -1 if there was no match.

The caretMode parameter can be used to instruct whether should match at index 0 or at offset.

Although const, this function sets matchedLength(), capturedTexts() and pos().

Warning: Searching backwards is much slower than searching forwards.

See also indexIn() and exactMatch().

fn matched_length(&self) -> c_int

C++ method: int QRegExp::matchedLength() const

C++ documentation:

Returns the length of the last matched string, or -1 if there was no match.

See also exactMatch(), indexIn(), and lastIndexIn().

fn new<Args>(args: Args) -> Args::ReturnType where Args: RegExpNewArgs

C++ method: QRegExp::QRegExp

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::reg_exp::RegExp
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::reg_exp::RegExp>
C++ method: [constructor] void QRegExp::QRegExp()

Variant 2

Rust arguments:
1) fn new((&::reg_exp::RegExp, cpp_utils::AsStruct)) -> ::reg_exp::RegExp
2) fn new((&::reg_exp::RegExp, cpp_utils::AsBox)) -> cpp_utils::CppBox<::reg_exp::RegExp>
C++ method: [constructor] void QRegExp::QRegExp(const QRegExp& rx)

Variant 3

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::reg_exp::RegExp
2) fn new((&::string::String, ::qt::CaseSensitivity, cpp_utils::AsStruct)) -> ::reg_exp::RegExp
3) fn new((&::string::String, ::qt::CaseSensitivity, ::reg_exp::PatternSyntax, cpp_utils::AsStruct)) -> ::reg_exp::RegExp
4) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::reg_exp::RegExp>
5) fn new((&::string::String, ::qt::CaseSensitivity, cpp_utils::AsBox)) -> cpp_utils::CppBox<::reg_exp::RegExp>
6) fn new((&::string::String, ::qt::CaseSensitivity, ::reg_exp::PatternSyntax, cpp_utils::AsBox)) -> cpp_utils::CppBox<::reg_exp::RegExp>
C++ method: [constructor] void QRegExp::QRegExp(const QString& pattern, Qt::CaseSensitivity cs = ?, QRegExp::PatternSyntax syntax = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, rx: &'l1 RegExp) -> &'l0 mut RegExp

C++ method: QRegExp& QRegExp::operator=(const QRegExp& rx)

C++ documentation:

Copies the regular expression rx and returns a reference to the copy. The case sensitivity, wildcard, and minimal matching options are also copied.

fn op_eq(&self, rx: &RegExp) -> bool

C++ method: bool QRegExp::operator==(const QRegExp& rx) const

C++ documentation:

Returns true if this regular expression is equal to rx; otherwise returns false.

Two QRegExp objects are equal if they have the same pattern strings and the same settings for case sensitivity, wildcard and minimal matching.

fn op_neq(&self, rx: &RegExp) -> bool

C++ method: bool QRegExp::operator!=(const QRegExp& rx) const

C++ documentation:

Returns true if this regular expression is not equal to rx; otherwise returns false.

See also operator==().

fn pattern<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegExpPatternArgs<'largs>

C++ method: QRegExp::pattern

This is an overloaded function. Available variants:

Rust arguments:
1) fn pattern(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn pattern(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QRegExp::pattern() const

C++ documentation:

Returns the pattern string of the regular expression. The pattern has either regular expression syntax or wildcard syntax, depending on patternSyntax().

See also setPattern(), patternSyntax(), and caseSensitivity().

fn pattern_syntax(&self) -> PatternSyntax

C++ method: QRegExp::PatternSyntax QRegExp::patternSyntax() const

C++ documentation:

Returns the syntax used by the regular expression. The default is QRegExp::RegExp.

See also setPatternSyntax(), pattern(), and caseSensitivity().

fn pos<'largs, Args>(&'largs self, args: Args) -> c_int where Args: RegExpPosArgs<'largs>

C++ method: QRegExp::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos(&self, ()) -> libc::c_int
2) fn pos(&self, libc::c_int) -> libc::c_int
C++ method: int QRegExp::pos(int nth = ?) const

C++ documentation:

Returns the position of the nth captured text in the searched string. If nth is 0 (the default), pos() returns the position of the whole match.

Example:

QRegExp rx("/([a-z]+)/([a-z]+)");
rx.indexIn("Output /dev/null");   // returns 7 (position of /dev/null)
rx.pos(0);                        // returns 7 (position of /dev/null)
rx.pos(1);                        // returns 8 (position of dev)
rx.pos(2);                        // returns 12 (position of null)

For zero-length matches, pos() always returns -1. (For example, if cap(4) would return an empty string, pos(4) returns -1.) This is a feature of the implementation.

See also cap() and capturedTexts().

fn pos_mut<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: RegExpPosMutArgs<'largs>

C++ method: QRegExp::pos

This is an overloaded function. Available variants:

Rust arguments:
1) fn pos_mut(&mut self, ()) -> libc::c_int
2) fn pos_mut(&mut self, libc::c_int) -> libc::c_int
C++ method: int QRegExp::pos(int nth = ?)

C++ documentation:

Returns the position of the nth captured text in the searched string. If nth is 0 (the default), pos() returns the position of the whole match.

Example:

QRegExp rx("/([a-z]+)/([a-z]+)");
rx.indexIn("Output /dev/null");   // returns 7 (position of /dev/null)
rx.pos(0);                        // returns 7 (position of /dev/null)
rx.pos(1);                        // returns 8 (position of dev)
rx.pos(2);                        // returns 12 (position of null)

For zero-length matches, pos() always returns -1. (For example, if cap(4) would return an empty string, pos(4) returns -1.) This is a feature of the implementation.

See also cap() and capturedTexts().

fn set_case_sensitivity(&mut self, cs: CaseSensitivity)

C++ method: void QRegExp::setCaseSensitivity(Qt::CaseSensitivity cs)

C++ documentation:

Sets case sensitive matching to cs.

If cs is Qt::CaseSensitive, .txt$ matches readme.txt but not README.TXT.

See also caseSensitivity(), setPatternSyntax(), setPattern(), and setMinimal().

fn set_minimal(&mut self, minimal: bool)

C++ method: void QRegExp::setMinimal(bool minimal)

C++ documentation:

Enables or disables minimal matching. If minimal is false, matching is greedy (maximal) which is the default.

For example, suppose we have the input string "We must be <b>bold</b>, very <b>bold</b>!" and the pattern <b>.*</b>. With the default greedy (maximal) matching, the match is "We must be <b>bold</b>, very <b>bold</b>!". But with minimal (non-greedy) matching, the first match is: "We must be <b>bold</b>, very <b>bold</b>!" and the second match is "We must be <b>bold</b>, very <b>bold</b>!". In practice we might use the pattern <b>[<]*</b> instead, although this will still fail for nested tags.

See also isMinimal() and setCaseSensitivity().

fn set_pattern(&mut self, pattern: &String)

C++ method: void QRegExp::setPattern(const QString& pattern)

C++ documentation:

Sets the pattern string to pattern. The case sensitivity, wildcard, and minimal matching options are not changed.

See also pattern(), setPatternSyntax(), and setCaseSensitivity().

fn set_pattern_syntax(&mut self, syntax: PatternSyntax)

C++ method: void QRegExp::setPatternSyntax(QRegExp::PatternSyntax syntax)

C++ documentation:

Sets the syntax mode for the regular expression. The default is QRegExp::RegExp.

Setting syntax to QRegExp::Wildcard enables simple shell-like QRegExp wildcard matching. For example, r*.txt matches the string readme.txt in wildcard mode, but does not match readme.

Setting syntax to QRegExp::FixedString means that the pattern is interpreted as a plain string. Special characters (e.g., backslash) don't need to be escaped then.

See also patternSyntax(), setPattern(), setCaseSensitivity(), and escape().

fn swap(&mut self, other: &mut RegExp)

C++ method: void QRegExp::swap(QRegExp& other)

C++ documentation:

Swaps regular expression other with this regular expression. This operation is very fast and never fails.

This function was introduced in Qt 4.8.