Enum qt_core::regular_expression::PatternOption

pub enum PatternOption {
    NoPattern,
    CaseInsensitive,
    DotMatchesEverything,
    Multiline,
    ExtendedPatternSyntax,
    InvertedGreediness,
    DontCapture,
    UseUnicodeProperties,
    OptimizeOnFirstUsage,
    DontAutomaticallyOptimize,
}

C++ type: QRegularExpression::PatternOption

C++ documentation:

The PatternOption enum defines modifiers to the way the pattern string should be interpreted, and therefore the way the pattern matches against a subject string.

The PatternOptions type is a typedef for QFlags<PatternOption>. It stores an OR combination of PatternOption values.

Variants
NoPattern

No pattern options are set. (C++ enum variant: NoPatternOption = 0)

CaseInsensitive

The pattern should match against the subject string in a case insensitive way. This option corresponds to the /i modifier in Perl regular expressions. (C++ enum variant: CaseInsensitiveOption = 1)

DotMatchesEverything

The dot metacharacter (.) in the pattern string is allowed to match any character in the subject string, including newlines (normally, the dot does not match newlines). This option corresponds to the /s modifier in Perl regular expressions. (C++ enum variant: DotMatchesEverythingOption = 2)

Multiline

The caret () and the dollar ($) metacharacters in the pattern string are allowed to match, respectively, immediately after and immediately before any newline in the subject string, as well as at the very beginning and at the very end of the subject string. This option corresponds to the /m modifier in Perl regular expressions. (C++ enum variant: MultilineOption = 4)

ExtendedPatternSyntax

Any whitespace in the pattern string which is not escaped and outside a character class is ignored. Moreover, an unescaped sharp (#) outside a character class causes all the following characters, until the first newline (included), to be ignored. This can be used to increase the readability of a pattern string as well as put comments inside regular expressions; this is particulary useful if the pattern string is loaded from a file or written by the user, because in C++ code it is always possible to use the rules for string literals to put comments outside the pattern string. This option corresponds to the /x modifier in Perl regular expressions. (C++ enum variant: ExtendedPatternSyntaxOption = 8)

InvertedGreediness

The greediness of the quantifiers is inverted: , +, ?, {m,n}, etc. become lazy, while their lazy versions (?, +?, ??, {m,n}?, etc.) become greedy. There is no equivalent for this option in Perl regular expressions. (C++ enum variant: InvertedGreedinessOption = 16)

DontCapture

The non-named capturing groups do not capture substrings; named capturing groups still work as intended, as well as the implicit capturing group number 0 corresponding to the entire match. There is no equivalent for this option in Perl regular expressions. (C++ enum variant: DontCaptureOption = 32)

UseUnicodeProperties

The meaning of the \w, \d, etc., character classes, as well as the meaning of their counterparts (\W, \D, etc.), is changed from matching ASCII characters only to matching any character with the corresponding Unicode property. For instance, \d is changed to match any character with the Unicode Nd (decimal digit) property; \w to match any character with either the Unicode L (letter) or N (digit) property, plus underscore, and so on. This option corresponds to the /u modifier in Perl regular expressions. (C++ enum variant: UseUnicodePropertiesOption = 64)

OptimizeOnFirstUsage

The regular expression will be optimized (and possibly JIT-compiled) on its first usage, instead of after a certain (undefined) number of usages. See also optimize(). This enum value has been introduced in Qt 5.4. (C++ enum variant: OptimizeOnFirstUsageOption = 128)

DontAutomaticallyOptimize

Regular expressions are automatically optimized after a certain number of usages; setting this option prevents such optimizations, therefore avoiding possible unpredictable spikes in CPU and memory usage. If both this option and the OptimizeOnFirstUsageOption option are set, then this option takes precedence. Note: this option will still let the regular expression to be optimized by manually calling optimize(). This enum value has been introduced in Qt 5.4. (C++ enum variant: DontAutomaticallyOptimizeOption = 256)