Struct qt_core::regular_expression::RegularExpression

pub struct RegularExpression {
    // some fields omitted
}

C++ type: QRegularExpression

C++ documentation:

The QRegularExpression class provides pattern matching using regular expressions.

Regular expressions, or regexps, are a very powerful tool to handle strings and texts. This is useful in many contexts, e.g.,

Validation	A regexp can test whether a substring meets some criteria, e.g. is an integer or contains no whitespace.

Searching	A regexp provides more powerful pattern matching than simple substring matching, e.g., match one of the words mail, letter or correspondence, but none of the words email, mailman, mailer, letterbox, etc.

Search and Replace	A regexp can replace all occurrences of a substring with a different substring, e.g., replace all occurrences of & with &amp; except where the & is already followed by an amp;.

String Splitting	A regexp can be used to identify where a string should be split apart, e.g. splitting tab-delimited strings.

This document is by no means a complete reference to pattern matching using regular expressions, and the following parts will require the reader to have some basic knowledge about Perl-like regular expressions and their pattern syntax.

Good references about regular expressions include:

Mastering Regular Expressions (Third Edition) by Jeffrey E. F. Friedl, ISBN 0-596-52812-4;

the pcrepattern(3) man page, describing the pattern syntax supported by PCRE (the reference implementation of Perl-compatible regular expressions);

the Perl's regular expression documentation and the Perl's regular expression tutorial.

Methods

impl RegularExpression

fn capture_count(&self) -> c_int

C++ method: int QRegularExpression::captureCount() const

C++ documentation:

Returns the number of capturing groups inside the pattern string, or -1 if the regular expression is not valid.

Note: The implicit capturing group 0 is not included in the returned number.

See also isValid().

fn error_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionErrorStringArgs<'largs>

C++ method: QRegularExpression::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn error_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QRegularExpression::errorString() const

C++ documentation:

Returns a textual description of the error found when checking the validity of the regular expression, or "no error" if no error was found.

See also isValid() and patternErrorOffset().

fn escape<Args>(args: Args) -> Args::ReturnType where Args: RegularExpressionEscapeArgs

C++ method: QRegularExpression::escape

This is an overloaded function. Available variants:

Rust arguments:
1) fn escape((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn escape((&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QRegularExpression::escape(const QString& str)

C++ documentation:

Escapes all characters of str so that they no longer have any special meaning when used as a regular expression pattern string, and returns the escaped string. For instance:

QString escaped = QRegularExpression::escape("a(x) = f(x) + g(x)");
// escaped == "a\(x\)\ \=\ f\(x\)\ \+\ g\(x\)"

This is very convenient in order to build patterns from arbitrary strings:

QString pattern = "(" + QRegularExpression::escape(name) +
                  "|" + QRegularExpression::escape(nickname) + ")";
QRegularExpression re(pattern);

Note: This function implements Perl's quotemeta algorithm and escapes with a backslash all characters in str, except for the characters in the [A-Z], [a-z] and [0-9] ranges, as well as the underscore (_) character. The only difference with Perl is that a literal NUL inside str is escaped with the sequence "\0" (backslash + '0'), instead of "\\0" (backslash + NUL).

fn global_match<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionGlobalMatchArgs<'largs>

C++ method: QRegularExpression::globalMatch

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn global_match(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
2) fn global_match(&self, (&::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
3) fn global_match(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
4) fn global_match(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
5) fn global_match(&self, (&::string::String, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
6) fn global_match(&self, (&::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
7) fn global_match(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
8) fn global_match(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
C++ method: QRegularExpressionMatchIterator QRegularExpression::globalMatch(const QString& subject, int offset = ?, QRegularExpression::MatchType matchType = ?, QFlags<QRegularExpression::MatchOption> matchOptions = ?) const

C++ documentation:

Attempts to perform a global match of the regular expression against the given subject string, starting at the position offset inside the subject, using a match of type matchType and honoring the given matchOptions.

The returned QRegularExpressionMatchIterator is positioned before the first match result (if any).

See also QRegularExpressionMatchIterator and global matching.

Variant 2

Rust arguments:
1) fn global_match(&self, (&::string_ref::StringRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
2) fn global_match(&self, (&::string_ref::StringRef, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
3) fn global_match(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
4) fn global_match(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
5) fn global_match(&self, (&::string_ref::StringRef, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
6) fn global_match(&self, (&::string_ref::StringRef, libc::c_int, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
7) fn global_match(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
8) fn global_match(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
C++ method: QRegularExpressionMatchIterator QRegularExpression::globalMatch(const QStringRef& subjectRef, int offset = ?, QRegularExpression::MatchType matchType = ?, QFlags<QRegularExpression::MatchOption> matchOptions = ?) const

C++ documentation:

This is an overloaded function.

Attempts to perform a global match of the regular expression against the given subjectRef string reference, starting at the position offset inside the subject, using a match of type matchType and honoring the given matchOptions.

The returned QRegularExpressionMatchIterator is positioned before the first match result (if any).

This function was introduced in Qt 5.5.

See also QRegularExpressionMatchIterator and global matching.

fn is_valid(&self) -> bool

C++ method: bool QRegularExpression::isValid() const

C++ documentation:

Returns true if the regular expression is a valid regular expression (that is, it contains no syntax errors, etc.), or false otherwise. Use errorString() to obtain a textual description of the error.

See also errorString() and patternErrorOffset().

fn match_<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchArgs<'largs>

C++ method: QRegularExpression::match

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn match_(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
2) fn match_(&self, (&::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
3) fn match_(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
4) fn match_(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
5) fn match_(&self, (&::string::String, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
6) fn match_(&self, (&::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
7) fn match_(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
8) fn match_(&self, (&::string::String, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
C++ method: QRegularExpressionMatch QRegularExpression::match(const QString& subject, int offset = ?, QRegularExpression::MatchType matchType = ?, QFlags<QRegularExpression::MatchOption> matchOptions = ?) const

C++ documentation:

Attempts to match the regular expression against the given subject string, starting at the position offset inside the subject, using a match of type matchType and honoring the given matchOptions.

The returned QRegularExpressionMatch object contains the results of the match.

See also QRegularExpressionMatch and normal matching.

Variant 2

Rust arguments:
1) fn match_(&self, (&::string_ref::StringRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
2) fn match_(&self, (&::string_ref::StringRef, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
3) fn match_(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
4) fn match_(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
5) fn match_(&self, (&::string_ref::StringRef, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
6) fn match_(&self, (&::string_ref::StringRef, libc::c_int, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
7) fn match_(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
8) fn match_(&self, (&::string_ref::StringRef, libc::c_int, ::regular_expression::MatchType, ::flags::Flags<::regular_expression::MatchOption>, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
C++ method: QRegularExpressionMatch QRegularExpression::match(const QStringRef& subjectRef, int offset = ?, QRegularExpression::MatchType matchType = ?, QFlags<QRegularExpression::MatchOption> matchOptions = ?) const

C++ documentation:

This is an overloaded function.

Attempts to match the regular expression against the given subjectRef string reference, starting at the position offset inside the subject, using a match of type matchType and honoring the given matchOptions.

The returned QRegularExpressionMatch object contains the results of the match.

This function was introduced in Qt 5.5.

See also QRegularExpressionMatch and normal matching.

fn named_capture_groups<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionNamedCaptureGroupsArgs<'largs>

C++ method: QRegularExpression::namedCaptureGroups

This is an overloaded function. Available variants:

Rust arguments:
1) fn named_capture_groups(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn named_capture_groups(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QRegularExpression::namedCaptureGroups() const

C++ documentation:

Returns a list of captureCount() + 1 elements, containing the names of the named capturing groups in the pattern string. The list is sorted such that the element of the list at position i is the name of the i-th capturing group, if it has a name, or an empty string if that capturing group is unnamed.

For instance, given the regular expression

(?<day>\d\d)-(?<month>\d\d)-(?<year>\d\d\d\d) (\w+) (?<name>\w+)

namedCaptureGroups() will return the following list:

("", "day", "month", "year", "", "name")

which corresponds to the fact that the capturing group #0 (corresponding to the whole match) has no name, the capturing group #1 has name "day", the capturing group #2 has name "month", etc.

If the regular expression is not valid, returns an empty list.

This function was introduced in Qt 5.1.

See also isValid(), QRegularExpressionMatch::captured(), and QString::isEmpty().

fn new<Args>(args: Args) -> Args::ReturnType where Args: RegularExpressionNewArgs

C++ method: QRegularExpression::QRegularExpression

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::regular_expression::RegularExpression
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression::RegularExpression>
C++ method: [constructor] void QRegularExpression::QRegularExpression()

Variant 2

Rust arguments:
1) fn new((&::regular_expression::RegularExpression, cpp_utils::AsStruct)) -> ::regular_expression::RegularExpression
2) fn new((&::regular_expression::RegularExpression, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression::RegularExpression>
C++ method: [constructor] void QRegularExpression::QRegularExpression(const QRegularExpression& re)

Variant 3

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::regular_expression::RegularExpression
2) fn new((&::string::String, ::flags::Flags<::regular_expression::PatternOption>, cpp_utils::AsStruct)) -> ::regular_expression::RegularExpression
3) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression::RegularExpression>
4) fn new((&::string::String, ::flags::Flags<::regular_expression::PatternOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression::RegularExpression>
C++ method: [constructor] void QRegularExpression::QRegularExpression(const QString& pattern, QFlags<QRegularExpression::PatternOption> options = ?)

fn op_assign<'l0, 'l1>(&'l0 mut self, re: &'l1 RegularExpression) -> &'l0 mut RegularExpression

C++ method: QRegularExpression& QRegularExpression::operator=(const QRegularExpression& re)

C++ documentation:

Assigns the regular expression re to this object, and returns a reference to the copy. Both the pattern and the pattern options are copied.

fn op_eq(&self, re: &RegularExpression) -> bool

C++ method: bool QRegularExpression::operator==(const QRegularExpression& re) const

C++ documentation:

Returns true if the regular expression is equal to re, or false otherwise. Two QRegularExpression objects are equal if they have the same pattern string and the same pattern options.

See also operator!=().

fn op_neq(&self, re: &RegularExpression) -> bool

C++ method: bool QRegularExpression::operator!=(const QRegularExpression& re) const

C++ documentation:

Returns true if the regular expression is different from re, or false otherwise.

See also operator==().

fn optimize(&self)

C++ method: void QRegularExpression::optimize() const

C++ documentation:

Forces an immediate optimization of the pattern, including JIT-compiling it (if the JIT compiler is enabled).

Patterns are normally optimized only after a certain number of usages. If you can predict that this QRegularExpression object is going to be used for several matches, it may be convenient to optimize it in advance by calling this function.

This function was introduced in Qt 5.4.

See also QRegularExpression::OptimizeOnFirstUsageOption.

fn pattern<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionPatternArgs<'largs>

C++ method: QRegularExpression::pattern

This is an overloaded function. Available variants:

Rust arguments:
1) fn pattern(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn pattern(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QRegularExpression::pattern() const

C++ documentation:

Returns the pattern string of the regular expression.

See also setPattern() and patternOptions().

fn pattern_error_offset(&self) -> c_int

C++ method: int QRegularExpression::patternErrorOffset() const

C++ documentation:

Returns the offset, inside the pattern string, at which an error was found when checking the validity of the regular expression. If no error was found, then -1 is returned.

See also pattern(), isValid(), and errorString().

fn pattern_options(&self) -> Flags<PatternOption>

C++ method: QFlags<QRegularExpression::PatternOption> QRegularExpression::patternOptions() const

C++ documentation:

Returns the pattern options for the regular expression.

See also setPatternOptions() and pattern().

fn set_pattern(&mut self, pattern: &String)

C++ method: void QRegularExpression::setPattern(const QString& pattern)

C++ documentation:

Sets the pattern string of the regular expression to pattern. The pattern options are left unchanged.

See also pattern() and setPatternOptions().

fn set_pattern_options(&mut self, options: Flags<PatternOption>)

C++ method: void QRegularExpression::setPatternOptions(QFlags<QRegularExpression::PatternOption> options)

C++ documentation:

Sets the given options as the pattern options of the regular expression. The pattern string is left unchanged.

See also patternOptions() and setPattern().

fn swap(&mut self, other: &mut RegularExpression)

C++ method: void QRegularExpression::swap(QRegularExpression& other)

C++ documentation:

Swaps the regular expression other with this regular expression. This operation is very fast and never fails.