Module qt_core::regular_expression_match::overloading

Traits

RegularExpressionMatchCapturedArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::captured method.

RegularExpressionMatchCapturedEndArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::captured_end method.

RegularExpressionMatchCapturedLengthArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::captured_length method.

RegularExpressionMatchCapturedRefArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::captured_ref method.

RegularExpressionMatchCapturedStartArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::captured_start method.

RegularExpressionMatchCapturedTextsArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::captured_texts method.

RegularExpressionMatchNewArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::new method.

RegularExpressionMatchRegularExpressionArgs
This trait represents a set of arguments accepted by RegularExpressionMatch::regular_expression method.