Struct qt_core::regular_expression_match::RegularExpressionMatch

pub struct RegularExpressionMatch {
    // some fields omitted
}

C++ type: QRegularExpressionMatch

C++ documentation:

The QRegularExpressionMatch class provides the results of a matching a QRegularExpression against a string.

A QRegularExpressionMatch object can be obtained by calling the QRegularExpression::match() function, or as a single result of a global match from a QRegularExpressionMatchIterator.

The success or the failure of a match attempt can be inspected by calling the hasMatch() function. QRegularExpressionMatch also reports a successful partial match through the hasPartialMatch() function.

In addition, QRegularExpressionMatch returns the substrings captured by the capturing groups in the pattern string. The implicit capturing group with index 0 captures the result of the whole match. The captured() function returns each substring captured, either by the capturing group's index or by its name:

QRegularExpression re("(\d\d) (?<name>\w+)");
QRegularExpressionMatch match = re.match("23 Jordan");
if (match.hasMatch()) {
    QString number = match.captured(1); // first == "23"
    QString name = match.captured("name"); // name == "Jordan"
}

For each captured substring it is possible to query its starting and ending offsets in the subject string by calling the capturedStart() and the capturedEnd() function, respectively. The length of each captured substring is available using the capturedLength() function.

The convenience function capturedTexts() will return all the captured substrings at once (including the substring matched by the entire pattern) in the order they have been captured by captring groups; that is, captured(i) == capturedTexts().at(i).

You can retrieve the QRegularExpression object the subject string was matched against by calling the regularExpression() function; the match type and the match options are available as well by calling the matchType() and the matchOptions() respectively.

Please refer to the QRegularExpression documentation for more information about the Qt regular expression classes.

Methods

impl RegularExpressionMatch

fn captured<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchCapturedArgs<'largs>

C++ method: QRegularExpressionMatch::captured

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn captured(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn captured(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QRegularExpressionMatch::captured(const QString& name) const

C++ documentation:

Returns the substring captured by the capturing group named name. If the capturing group named name did not capture a string or doesn't exist, returns a null QString.

See also capturedRef(), capturedStart(), capturedEnd(), capturedLength(), and QString::isNull().

Variant 2

Rust arguments:
1) fn captured(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn captured(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn captured(&self, cpp_utils::AsStruct) -> ::string::String
4) fn captured(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QRegularExpressionMatch::captured(int nth = ?) const

C++ documentation:

Returns the substring captured by the nth capturing group. If the nth capturing group did not capture a string or doesn't exist, returns a null QString.

See also capturedRef(), lastCapturedIndex(), capturedStart(), capturedEnd(), capturedLength(), and QString::isNull().

fn captured_end<'largs, Args>(&'largs self, args: Args) -> c_int where Args: RegularExpressionMatchCapturedEndArgs<'largs>

C++ method: QRegularExpressionMatch::capturedEnd

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn captured_end(&self, &::string::String) -> libc::c_int
C++ method: int QRegularExpressionMatch::capturedEnd(const QString& name) const

C++ documentation:

Returns the offset inside the subject string immediately after the ending position of the substring captured by the capturing group named name. If the capturing group named name did not capture a string or doesn't exist, returns -1.

See also capturedStart(), capturedLength(), and captured().

Variant 2

Rust arguments:
1) fn captured_end(&self, ()) -> libc::c_int
2) fn captured_end(&self, libc::c_int) -> libc::c_int
C++ method: int QRegularExpressionMatch::capturedEnd(int nth = ?) const

C++ documentation:

Returns the offset inside the subject string immediately after the ending position of the substring captured by the nth capturing group. If the nth capturing group did not capture a string or doesn't exist, returns -1.

See also capturedStart(), capturedLength(), and captured().

fn captured_length<'largs, Args>(&'largs self, args: Args) -> c_int where Args: RegularExpressionMatchCapturedLengthArgs<'largs>

C++ method: QRegularExpressionMatch::capturedLength

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn captured_length(&self, &::string::String) -> libc::c_int
C++ method: int QRegularExpressionMatch::capturedLength(const QString& name) const

C++ documentation:

Returns the offset inside the subject string corresponding to the starting position of the substring captured by the capturing group named name.

Note: This function returns 0 if the capturing group named name did not capture a string or doesn't exist.

See also capturedStart(), capturedEnd(), and captured().

Variant 2

Rust arguments:
1) fn captured_length(&self, ()) -> libc::c_int
2) fn captured_length(&self, libc::c_int) -> libc::c_int
C++ method: int QRegularExpressionMatch::capturedLength(int nth = ?) const

C++ documentation:

Returns the length of the substring captured by the nth capturing group.

Note: This function returns 0 if the nth capturing group did not capture a string or doesn't exist.

See also capturedStart(), capturedEnd(), and captured().

fn captured_ref<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchCapturedRefArgs<'largs>

C++ method: QRegularExpressionMatch::capturedRef

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn captured_ref(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn captured_ref(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string_ref::StringRef
C++ method: QStringRef QRegularExpressionMatch::capturedRef(const QString& name) const

C++ documentation:

Returns a reference to the string captured by the capturing group named name. If the capturing group named name did not capture a string or doesn't exist, returns a null QStringRef.

See also captured(), capturedStart(), capturedEnd(), capturedLength(), and QStringRef::isNull().

Variant 2

Rust arguments:
1) fn captured_ref(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn captured_ref(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
3) fn captured_ref(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
4) fn captured_ref(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string_ref::StringRef
C++ method: QStringRef QRegularExpressionMatch::capturedRef(int nth = ?) const

C++ documentation:

Returns a reference to the substring captured by the nth capturing group. If the nth capturing group did not capture a string or doesn't exist, returns a null QStringRef.

See also captured(), lastCapturedIndex(), capturedStart(), capturedEnd(), capturedLength(), and QStringRef::isNull().

fn captured_start<'largs, Args>(&'largs self, args: Args) -> c_int where Args: RegularExpressionMatchCapturedStartArgs<'largs>

C++ method: QRegularExpressionMatch::capturedStart

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn captured_start(&self, &::string::String) -> libc::c_int
C++ method: int QRegularExpressionMatch::capturedStart(const QString& name) const

C++ documentation:

Returns the offset inside the subject string corresponding to the starting position of the substring captured by the capturing group named name. If the capturing group named name did not capture a string or doesn't exist, returns -1.

See also capturedEnd(), capturedLength(), and captured().

Variant 2

Rust arguments:
1) fn captured_start(&self, ()) -> libc::c_int
2) fn captured_start(&self, libc::c_int) -> libc::c_int
C++ method: int QRegularExpressionMatch::capturedStart(int nth = ?) const

C++ documentation:

Returns the offset inside the subject string corresponding to the starting position of the substring captured by the nth capturing group. If the nth capturing group did not capture a string or doesn't exist, returns -1.

See also capturedEnd(), capturedLength(), and captured().

fn captured_texts<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchCapturedTextsArgs<'largs>

C++ method: QRegularExpressionMatch::capturedTexts

This is an overloaded function. Available variants:

Rust arguments:
1) fn captured_texts(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn captured_texts(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QRegularExpressionMatch::capturedTexts() const

C++ documentation:

Returns a list of all strings captured by capturing groups, in the order the groups themselves appear in the pattern string.

fn has_match(&self) -> bool

C++ method: bool QRegularExpressionMatch::hasMatch() const

C++ documentation:

Returns true if the regular expression matched against the subject string, or false otherwise.

See also QRegularExpression::match() and hasPartialMatch().

fn has_partial_match(&self) -> bool

C++ method: bool QRegularExpressionMatch::hasPartialMatch() const

C++ documentation:

Returns true if the regular expression partially matched against the subject string, or false otherwise.

Note: Only a match that explicitly used the one of the partial match types can yield a partial match. Still, if such a match succeeds totally, this function will return false, while hasMatch() will return true.

See also QRegularExpression::match(), QRegularExpression::MatchType, and hasMatch().

fn is_valid(&self) -> bool

C++ method: bool QRegularExpressionMatch::isValid() const

C++ documentation:

Returns true if the match object was obtained as a result from the QRegularExpression::match() function invoked on a valid QRegularExpression object; returns false if the QRegularExpression was invalid.

See also QRegularExpression::match() and QRegularExpression::isValid().

fn last_captured_index(&self) -> c_int

C++ method: int QRegularExpressionMatch::lastCapturedIndex() const

C++ documentation:

Returns the index of the last capturing group that captured something, including the implicit capturing group 0. This can be used to extract all the substrings that were captured:

QRegularExpressionMatch match = re.match(string);
for (int i = 0; i <= match.lastCapturedIndex(); ++i) {
    QString captured = match.captured(i);
    // ...
}

Note that some of the capturing groups with an index less than lastCapturedIndex() could have not matched, and therefore captured nothing.

If the regular expression did not match, this function returns -1.

See also captured(), capturedStart(), capturedEnd(), and capturedLength().

fn match_options(&self) -> Flags<MatchOption>

C++ method: QFlags<QRegularExpression::MatchOption> QRegularExpressionMatch::matchOptions() const

C++ documentation:

Returns the match options that were used to get this QRegularExpressionMatch object, that is, the match options that were passed to QRegularExpression::match() or QRegularExpression::globalMatch().

See also QRegularExpression::match(), regularExpression(), and matchType().

fn match_type(&self) -> MatchType

C++ method: QRegularExpression::MatchType QRegularExpressionMatch::matchType() const

C++ documentation:

Returns the match type that was used to get this QRegularExpressionMatch object, that is, the match type that was passed to QRegularExpression::match() or QRegularExpression::globalMatch().

See also QRegularExpression::match(), regularExpression(), and matchOptions().

fn new<Args>(args: Args) -> Args::ReturnType where Args: RegularExpressionMatchNewArgs

C++ method: QRegularExpressionMatch::QRegularExpressionMatch

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::regular_expression_match::RegularExpressionMatch
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
C++ method: [constructor] void QRegularExpressionMatch::QRegularExpressionMatch()

Variant 2

Rust arguments:
1) fn new((&::regular_expression_match::RegularExpressionMatch, cpp_utils::AsStruct)) -> ::regular_expression_match::RegularExpressionMatch
2) fn new((&::regular_expression_match::RegularExpressionMatch, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
C++ method: [constructor] void QRegularExpressionMatch::QRegularExpressionMatch(const QRegularExpressionMatch& match)

fn op_assign<'l0, 'l1>(&'l0 mut self, match_: &'l1 RegularExpressionMatch) -> &'l0 mut RegularExpressionMatch

C++ method: QRegularExpressionMatch& QRegularExpressionMatch::operator=(const QRegularExpressionMatch& match)

C++ documentation:

Assigns the match result match to this object, and returns a reference to the copy.

fn regular_expression<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchRegularExpressionArgs<'largs>

C++ method: QRegularExpressionMatch::regularExpression

This is an overloaded function. Available variants:

Rust arguments:
1) fn regular_expression(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression::RegularExpression>
2) fn regular_expression(&self, cpp_utils::AsStruct) -> ::regular_expression::RegularExpression
C++ method: QRegularExpression QRegularExpressionMatch::regularExpression() const

C++ documentation:

Returns the QRegularExpression object whose match() function returned this object.

See also QRegularExpression::match(), matchType(), and matchOptions().

fn swap(&mut self, other: &mut RegularExpressionMatch)

C++ method: void QRegularExpressionMatch::swap(QRegularExpressionMatch& other)

C++ documentation:

Swaps the match result other with this match result. This operation is very fast and never fails.