Struct qt_core::regular_expression_match_iterator::RegularExpressionMatchIterator

pub struct RegularExpressionMatchIterator {
    // some fields omitted
}

C++ type: QRegularExpressionMatchIterator

C++ documentation:

The QRegularExpressionMatchIterator class provides an iterator on the results of a global match of a QRegularExpression object against a string.

A QRegularExpressionMatchIterator object is a forward only Java-like iterator; it can be obtained by calling the QRegularExpression::globalMatch() function. A new QRegularExpressionMatchIterator will be positioned before the first result. You can then call the hasNext() function to check if there are more results available; if so, the next() function will return the next result and advance the iterator.

Each result is a QRegularExpressionMatch object holding all the information for that result (including captured substrings).

For instance:

// extracts the words
QRegularExpression re("(\w+)");
QString subject("the quick fox");
QRegularExpressionMatchIterator i = re.globalMatch(subject);
while (i.hasNext()) {
    QRegularExpressionMatch match = i.next();
    // ...
}

Moreover, QRegularExpressionMatchIterator offers a peekNext() function to get the next result without advancing the iterator.

You can retrieve the QRegularExpression object the subject string was matched against by calling the regularExpression() function; the match type and the match options are available as well by calling the matchType() and the matchOptions() respectively.

Please refer to the QRegularExpression documentation for more information about the Qt regular expression classes.

Methods

impl RegularExpressionMatchIterator

fn has_next(&self) -> bool

C++ method: bool QRegularExpressionMatchIterator::hasNext() const

C++ documentation:

Returns true if there is at least one match result ahead of the iterator; otherwise it returns false.

See also next().

fn is_valid(&self) -> bool

C++ method: bool QRegularExpressionMatchIterator::isValid() const

C++ documentation:

Returns true if the iterator object was obtained as a result from the QRegularExpression::globalMatch() function invoked on a valid QRegularExpression object; returns false if the QRegularExpression was invalid.

See also QRegularExpression::globalMatch() and QRegularExpression::isValid().

fn match_options(&self) -> Flags<MatchOption>

C++ method: QFlags<QRegularExpression::MatchOption> QRegularExpressionMatchIterator::matchOptions() const

C++ documentation:

Returns the match options that were used to get this QRegularExpressionMatchIterator object, that is, the match options that were passed to QRegularExpression::globalMatch().

See also QRegularExpression::globalMatch(), regularExpression(), and matchType().

fn match_type(&self) -> MatchType

C++ method: QRegularExpression::MatchType QRegularExpressionMatchIterator::matchType() const

C++ documentation:

Returns the match type that was used to get this QRegularExpressionMatchIterator object, that is, the match type that was passed to QRegularExpression::globalMatch().

See also QRegularExpression::globalMatch(), regularExpression(), and matchOptions().

fn new<Args>(args: Args) -> Args::ReturnType where Args: RegularExpressionMatchIteratorNewArgs

C++ method: QRegularExpressionMatchIterator::QRegularExpressionMatchIterator

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
C++ method: [constructor] void QRegularExpressionMatchIterator::QRegularExpressionMatchIterator()

Variant 2

Rust arguments:
1) fn new((&::regular_expression_match_iterator::RegularExpressionMatchIterator, cpp_utils::AsStruct)) -> ::regular_expression_match_iterator::RegularExpressionMatchIterator
2) fn new((&::regular_expression_match_iterator::RegularExpressionMatchIterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::regular_expression_match_iterator::RegularExpressionMatchIterator>
C++ method: [constructor] void QRegularExpressionMatchIterator::QRegularExpressionMatchIterator(const QRegularExpressionMatchIterator& iterator)

fn next<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchIteratorNextArgs<'largs>

C++ method: QRegularExpressionMatchIterator::next

This is an overloaded function. Available variants:

Rust arguments:
1) fn next(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
2) fn next(&mut self, cpp_utils::AsStruct) -> ::regular_expression_match::RegularExpressionMatch
C++ method: QRegularExpressionMatch QRegularExpressionMatchIterator::next()

C++ documentation:

Returns the next match result and advances the iterator by one position.

Note: Calling this function when the iterator is at the end of the result set leads to undefined results.

fn op_assign<'l0, 'l1>(&'l0 mut self, iterator: &'l1 RegularExpressionMatchIterator) -> &'l0 mut RegularExpressionMatchIterator

C++ method: QRegularExpressionMatchIterator& QRegularExpressionMatchIterator::operator=(const QRegularExpressionMatchIterator& iterator)

C++ documentation:

Assigns the iterator iterator to this object, and returns a reference to the copy.

fn peek_next<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchIteratorPeekNextArgs<'largs>

C++ method: QRegularExpressionMatchIterator::peekNext

This is an overloaded function. Available variants:

Rust arguments:
1) fn peek_next(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression_match::RegularExpressionMatch>
2) fn peek_next(&self, cpp_utils::AsStruct) -> ::regular_expression_match::RegularExpressionMatch
C++ method: QRegularExpressionMatch QRegularExpressionMatchIterator::peekNext() const

C++ documentation:

Returns the next match result without moving the iterator.

Note: Calling this function when the iterator is at the end of the result set leads to undefined results.

fn regular_expression<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: RegularExpressionMatchIteratorRegularExpressionArgs<'largs>

C++ method: QRegularExpressionMatchIterator::regularExpression

This is an overloaded function. Available variants:

Rust arguments:
1) fn regular_expression(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression::RegularExpression>
2) fn regular_expression(&self, cpp_utils::AsStruct) -> ::regular_expression::RegularExpression
C++ method: QRegularExpression QRegularExpressionMatchIterator::regularExpression() const

C++ documentation:

Returns the QRegularExpression object whose globalMatch() function returned this object.

See also QRegularExpression::globalMatch(), matchType(), and matchOptions().

fn swap(&mut self, other: &mut RegularExpressionMatchIterator)

C++ method: void QRegularExpressionMatchIterator::swap(QRegularExpressionMatchIterator& other)

C++ documentation:

Swaps the iterator other with this iterator object. This operation is very fast and never fails.