Struct qt_core::resource::Resource

pub struct Resource {
    // some fields omitted
}

C++ type: QResource

C++ documentation:

The QResource class provides an interface for reading directly from resources.

QResource is an object that represents a set of data (and possibly children) relating to a single resource entity. QResource gives direct access to the bytes in their raw format. In this way direct access allows reading data without buffer copying or indirection. Indirection is often useful when interacting with the resource entity as if it is a file, this can be achieved with QFile. The data and children behind a QResource are normally compiled into an application/library, but it is also possible to load a resource at runtime. When loaded at run time the resource file will be loaded as one big set of data and then given out in pieces via references into the resource tree.

A QResource can either be loaded with an absolute path, either treated as a file system rooted with a / character, or in resource notation rooted with a : character. A relative resource can also be opened which will be found in the list of paths returned by QDir::searchPaths().

A QResource that is representing a file will have data backing it, this data can possibly be compressed, in which case qUncompress() must be used to access the real data; this happens implicitly when accessed through a QFile. A QResource that is representing a directory will have only children and no data.

Methods

impl Resource

fn absolute_file_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ResourceAbsoluteFilePathArgs<'largs>

C++ method: QResource::absoluteFilePath

This is an overloaded function. Available variants:

Rust arguments:
1) fn absolute_file_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn absolute_file_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QResource::absoluteFilePath() const

C++ documentation:

Returns the real path that this QResource represents, if the resource was found via the QDir::searchPaths() it will be indicated in the path.

See also fileName().

fn add_search_path(path: &String)

C++ method: static void QResource::addSearchPath(const QString& path)

C++ documentation:

Use QDir::addSearchPath() with a prefix instead.

Adds path to the search paths searched in to find resources that are not specified with an absolute path. The path must be an absolute path (start with /).

The default search path is to search only in the root (:/). The last path added will be consulted first upon next QResource creation.

fn data(&self) -> *const c_uchar

C++ method: const unsigned char* QResource::data() const

C++ documentation:

Returns direct access to a read only segment of data that this resource represents. If the resource is compressed the data returns is compressed and qUncompress() must be used to access the data. If the resource is a directory 0 is returned.

See also size(), isCompressed(), and isFile().

fn file_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ResourceFileNameArgs<'largs>

C++ method: QResource::fileName

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn file_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QResource::fileName() const

C++ documentation:

Returns the full path to the file that this QResource represents as it was passed.

See also setFileName() and absoluteFilePath().

fn is_compressed(&self) -> bool

C++ method: bool QResource::isCompressed() const

C++ documentation:

Returns true if the resource represents a file and the data backing it is in a compressed format, false otherwise.

See also data() and isFile().

fn is_valid(&self) -> bool

C++ method: bool QResource::isValid() const

C++ documentation:

Returns true if the resource really exists in the resource hierarchy, false otherwise.

fn locale<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ResourceLocaleArgs<'largs>

C++ method: QResource::locale

This is an overloaded function. Available variants:

Rust arguments:
1) fn locale(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::locale::Locale>
2) fn locale(&self, cpp_utils::AsStruct) -> ::locale::Locale
C++ method: QLocale QResource::locale() const

C++ documentation:

Returns the locale used to locate the data for the QResource.

See also setLocale().

fn new<Args>(args: Args) -> Args::ReturnType where Args: ResourceNewArgs

C++ method: QResource::QResource

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::resource::Resource
2) fn new((&::string::String, &::locale::Locale, cpp_utils::AsStruct)) -> ::resource::Resource
3) fn new(cpp_utils::AsStruct) -> ::resource::Resource
4) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::resource::Resource>
5) fn new((&::string::String, &::locale::Locale, cpp_utils::AsBox)) -> cpp_utils::CppBox<::resource::Resource>
6) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::resource::Resource>
C++ method: [constructor] void QResource::QResource(const QString& file = ?, const QLocale& locale = ?)

fn register_resource<Args>(args: Args) -> bool where Args: ResourceRegisterResourceArgs

C++ method: QResource::registerResource

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn register_resource(&::string::String) -> bool
2) fn register_resource((&::string::String, &::string::String)) -> bool
C++ method: static bool QResource::registerResource(const QString& rccFilename, const QString& resourceRoot = ?)

C++ documentation:

Registers the resource with the given rccFileName at the location in the resource tree specified by mapRoot, and returns true if the file is successfully opened; otherwise returns false.

See also unregisterResource().

Variant 2

Rust arguments:
1) fn register_resource(*const libc::c_uchar) -> bool
2) fn register_resource((*const libc::c_uchar, &::string::String)) -> bool
C++ method: static bool QResource::registerResource(const unsigned char* rccData, const QString& resourceRoot = ?)

C++ documentation:

Registers the resource with the given rccData at the location in the resource tree specified by mapRoot, and returns true if the file is successfully opened; otherwise returns false.

Warning: The data must remain valid throughout the life of any QFile that may reference the resource data.

This function was introduced in Qt 4.3.

See also unregisterResource().

fn search_paths<Args>(args: Args) -> Args::ReturnType where Args: ResourceSearchPathsArgs

C++ method: QResource::searchPaths

This is an overloaded function. Available variants:

Rust arguments:
1) fn search_paths(cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn search_paths(cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: static QStringList QResource::searchPaths()

C++ documentation:

Use QDir::searchPaths() instead.

Returns the current search path list. This list is consulted when creating a relative resource.

See also QDir::addSearchPath() and QDir::setSearchPaths().

fn set_file_name(&mut self, file: &String)

C++ method: void QResource::setFileName(const QString& file)

C++ documentation:

Sets a QResource to point to file. file can either be absolute, in which case it is opened directly, if relative then the file will be tried to be found in QDir::searchPaths().

See also fileName() and absoluteFilePath().

fn set_locale(&mut self, locale: &Locale)

C++ method: void QResource::setLocale(const QLocale& locale)

C++ documentation:

Sets a QResource to only load the localization of resource to for locale. If a resource for the specific locale is not found then the C locale is used.

See also locale() and setFileName().

fn size(&self) -> i64

C++ method: qint64 QResource::size() const

C++ documentation:

Returns the size of the data backing the resource.

See also data() and isFile().

fn unregister_resource<Args>(args: Args) -> bool where Args: ResourceUnregisterResourceArgs

C++ method: QResource::unregisterResource

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn unregister_resource(&::string::String) -> bool
2) fn unregister_resource((&::string::String, &::string::String)) -> bool
C++ method: static bool QResource::unregisterResource(const QString& rccFilename, const QString& resourceRoot = ?)

C++ documentation:

Unregisters the resource with the given rccFileName at the location in the resource tree specified by mapRoot, and returns true if the resource is successfully unloaded and no references exist for the resource; otherwise returns false.

See also registerResource().

Variant 2

Rust arguments:
1) fn unregister_resource(*const libc::c_uchar) -> bool
2) fn unregister_resource((*const libc::c_uchar, &::string::String)) -> bool
C++ method: static bool QResource::unregisterResource(const unsigned char* rccData, const QString& resourceRoot = ?)

C++ documentation:

Unregisters the resource with the given rccData at the location in the resource tree specified by mapRoot, and returns true if the resource is successfully unloaded and no references exist into the resource; otherwise returns false.

This function was introduced in Qt 4.3.

See also registerResource().