Struct qt_core::runnable::Runnable

pub struct Runnable {
    // some fields omitted
}

C++ type: QRunnable

C++ documentation:

The QRunnable class is the base class for all runnable objects.

The QRunnable class is an interface for representing a task or piece of code that needs to be executed, represented by your reimplementation of the run() function.

You can use QThreadPool to execute your code in a separate thread. QThreadPool deletes the QRunnable automatically if autoDelete() returns true (the default). Use setAutoDelete() to change the auto-deletion flag.

QThreadPool supports executing the same QRunnable more than once by calling QThreadPool::tryStart(this) from within the run() function. If autoDelete is enabled the QRunnable will be deleted when the last thread exits the run function. Calling QThreadPool::start() multiple times with the same QRunnable when autoDelete is enabled creates a race condition and is not recommended.

Methods

impl Runnable

fn auto_delete(&self) -> bool

C++ method: bool QRunnable::autoDelete() const

C++ documentation:

Returns true is auto-deletion is enabled; false otherwise.

If auto-deletion is enabled, QThreadPool will automatically delete this runnable after calling run(); otherwise, ownership remains with the application programmer.

See also setAutoDelete() and QThreadPool.

fn run(&mut self)

C++ method: pure virtual void QRunnable::run()

C++ documentation:

Implement this pure virtual function in your subclass.

fn set_auto_delete(&mut self, auto_delete: bool)

C++ method: void QRunnable::setAutoDelete(bool _autoDelete)

C++ documentation:

Enables auto-deletion if autoDelete is true; otherwise auto-deletion is disabled.

If auto-deletion is enabled, QThreadPool will automatically delete this runnable after calling run(); otherwise, ownership remains with the application programmer.

Note that this flag must be set before calling QThreadPool::start(). Calling this function after QThreadPool::start() results in undefined behavior.

See also autoDelete() and QThreadPool.