Module qt_core::save_file::overloading

Traits

SaveFileConnectArgs
This trait represents a set of arguments accepted by SaveFile::connect method.

SaveFileConnectStaticArgs
This trait represents a set of arguments accepted by SaveFile::connect_static method.

SaveFileDestroyedArgs
This trait represents a set of arguments accepted by SaveFile::destroyed method.

SaveFileDisconnectArgs
This trait represents a set of arguments accepted by SaveFile::disconnect method.

SaveFileDisconnectStaticArgs
This trait represents a set of arguments accepted by SaveFile::disconnect_static method.

SaveFileDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SaveFile::dynamic_property_names method.

SaveFileErrorStringArgs
This trait represents a set of arguments accepted by SaveFile::error_string method.

SaveFileFileNameArgs
This trait represents a set of arguments accepted by SaveFile::file_name method.

SaveFileMapArgs
This trait represents a set of arguments accepted by SaveFile::map method.

SaveFileNewArgs
This trait represents a set of arguments accepted by SaveFile::new method.

SaveFileObjectNameArgs
This trait represents a set of arguments accepted by SaveFile::object_name method.

SaveFilePeekArgs
This trait represents a set of arguments accepted by SaveFile::peek method.

SaveFilePropertyArgs
This trait represents a set of arguments accepted by SaveFile::property method.

SaveFileReadAllArgs
This trait represents a set of arguments accepted by SaveFile::read_all method.

SaveFileReadArgs
This trait represents a set of arguments accepted by SaveFile::read method.

SaveFileReadLineArgs
This trait represents a set of arguments accepted by SaveFile::read_line method.

SaveFileStartTimerArgs
This trait represents a set of arguments accepted by SaveFile::start_timer method.

SaveFileTrArgs
This trait represents a set of arguments accepted by SaveFile::tr method.

SaveFileTrUtf8Args
This trait represents a set of arguments accepted by SaveFile::tr_utf8 method.

SaveFileWriteArgs
This trait represents a set of arguments accepted by SaveFile::write method.