Struct qt_core::scoped_pointer::ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef

pub struct ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef {
    // some fields omitted
}

C++ type: QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>

Methods

impl ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef

fn data(&self) -> *mut ObjectData

C++ method: QObjectData* QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::data() const

C++ documentation:

Returns the value of the pointer referenced by this object. QScopedPointer still owns the object pointed to.

fn is_null(&self) -> bool

C++ method: bool QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::isNull() const

C++ documentation:

Returns true if this object is holding a pointer that is null.

fn new<Args>(args: Args) -> Args::ReturnType where Args: ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRefNewArgs

C++ method: QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::QScopedPointer

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::scoped_pointer::ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef
2) fn new((*mut ::object_data::ObjectData, cpp_utils::AsStruct)) -> ::scoped_pointer::ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::scoped_pointer::ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef>
4) fn new((*mut ::object_data::ObjectData, cpp_utils::AsBox)) -> cpp_utils::CppBox<::scoped_pointer::ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef>
C++ method: [constructor] void QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::QScopedPointer(QObjectData* p = ?)

fn op_indirection<'l0>(&'l0 self) -> &'l0 mut ObjectData

C++ method: QObjectData& QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::operator*() const

C++ documentation:

Provides access to the scoped pointer's object.

If the contained pointer is null, behavior is undefined.

See also isNull().

fn op_not(&self) -> bool

C++ method: bool QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::operator!() const

C++ documentation:

Returns true if the pointer referenced by this object is null, otherwise returns false.

See also isNull().

fn op_struct_deref(&self) -> *mut ObjectData

C++ method: QObjectData* QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::operator->() const

C++ documentation:

Provides access to the scoped pointer's object.

If the contained pointer is null, behavior is undefined.

See also isNull().

fn reset<'largs, Args>(&'largs mut self, args: Args) where Args: ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRefResetArgs<'largs>

C++ method: QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::reset

This is an overloaded function. Available variants:

Rust arguments:
1) fn reset(&mut self, ()) -> ()
2) fn reset(&mut self, *mut ::object_data::ObjectData) -> ()
C++ method: void QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::reset(QObjectData* other = ?)

C++ documentation:

Deletes the existing object it is pointing to if any, and sets its pointer to other. QScopedPointer now owns other and will delete it in its destructor.

fn swap(&mut self, other: &mut ScopedPointerObjectDataRefScopedPointerDeleterObjectDataRefRef)

C++ method: void QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::swap(QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>& other)

C++ documentation:

Swap this pointer with other.

fn take(&mut self) -> *mut ObjectData

C++ method: QObjectData* QScopedPointer<QObjectData, QScopedPointerDeleter<QObjectData>>::take()

C++ documentation:

Returns the value of the pointer referenced by this object. The pointer of this QScopedPointer object will be reset to null.

Callers of this function take ownership of the pointer.