Struct qt_core::semaphore::Semaphore

pub struct Semaphore {
    // some fields omitted
}

C++ type: QSemaphore

C++ documentation:

The QSemaphore class provides a general counting semaphore.

A semaphore is a generalization of a mutex. While a mutex can only be locked once, it's possible to acquire a semaphore multiple times. Semaphores are typically used to protect a certain number of identical resources.

Semaphores support two fundamental operations, acquire() and release():

acquire(n) tries to acquire n resources. If there aren't that many resources available, the call will block until this is the case.

release(n) releases n resources.

There's also a tryAcquire() function that returns immediately if it cannot acquire the resources, and an available() function that returns the number of available resources at any time.

Example:

QSemaphore sem(5);      // sem.available() == 5

sem.acquire(3); // sem.available() == 2 sem.acquire(2); // sem.available() == 0 sem.release(5); // sem.available() == 5 sem.release(5); // sem.available() == 10

sem.tryAcquire(1); // sem.available() == 9, returns true sem.tryAcquire(250); // sem.available() == 9, returns false
A typical application of semaphores is for controlling access to a circular buffer shared by a producer thread and a consumer thread. The Semaphores Example shows how to use QSemaphore to solve that problem.

A non-computing example of a semaphore would be dining at a restaurant. A semaphore is initialized with the number of chairs in the restaurant. As people arrive, they want a seat. As seats are filled, available() is decremented. As people leave, the available() is incremented, allowing more people to enter. If a party of 10 people want to be seated, but there are only 9 seats, those 10 people will wait, but a party of 4 people would be seated (taking the available seats to 5, making the party of 10 people wait longer).

Methods

impl Semaphore

fn acquire<'largs, Args>(&'largs mut self, args: Args) where Args: SemaphoreAcquireArgs<'largs>

C++ method: QSemaphore::acquire

This is an overloaded function. Available variants:

Rust arguments:
1) fn acquire(&mut self, libc::c_int) -> ()
2) fn acquire(&mut self, ()) -> ()
C++ method: void QSemaphore::acquire(int n = ?)

C++ documentation:

Tries to acquire n resources guarded by the semaphore. If n > available(), this call will block until enough resources are available.

See also release(), available(), and tryAcquire().

fn available(&self) -> c_int

C++ method: int QSemaphore::available() const

C++ documentation:

Returns the number of resources currently available to the semaphore. This number can never be negative.

See also acquire() and release().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SemaphoreNewArgs

C++ method: QSemaphore::QSemaphore

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::semaphore::Semaphore
2) fn new(cpp_utils::AsStruct) -> ::semaphore::Semaphore
3) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::semaphore::Semaphore>
4) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::semaphore::Semaphore>
C++ method: [constructor] void QSemaphore::QSemaphore(int n = ?)

fn release<'largs, Args>(&'largs mut self, args: Args) where Args: SemaphoreReleaseArgs<'largs>

C++ method: QSemaphore::release

This is an overloaded function. Available variants:

Rust arguments:
1) fn release(&mut self, libc::c_int) -> ()
2) fn release(&mut self, ()) -> ()
C++ method: void QSemaphore::release(int n = ?)

C++ documentation:

Releases n resources guarded by the semaphore.

This function can be used to "create" resources as well. For example:

QSemaphore sem(5);      // a semaphore that guards 5 resources
sem.acquire(5);         // acquire all 5 resources
sem.release(5);         // release the 5 resources
sem.release(10);        // "create" 10 new resources

See also acquire() and available().

fn try_acquire<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: SemaphoreTryAcquireArgs<'largs>

C++ method: QSemaphore::tryAcquire

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn try_acquire(&mut self, libc::c_int) -> bool
2) fn try_acquire(&mut self, ()) -> bool
C++ method: bool QSemaphore::tryAcquire(int n = ?)

C++ documentation:

Tries to acquire n resources guarded by the semaphore and returns true on success. If available() < n, this call immediately returns false without acquiring any resources.

Example:

QSemaphore sem(5);      // sem.available() == 5
sem.tryAcquire(250);    // sem.available() == 5, returns false
sem.tryAcquire(3);      // sem.available() == 2, returns true

See also acquire().

Variant 2

Rust arguments: fn try_acquire(&mut self, (libc::c_int, libc::c_int)) -> bool
C++ method: bool QSemaphore::tryAcquire(int n, int timeout)

C++ documentation:

Tries to acquire n resources guarded by the semaphore and returns true on success. If available() < n, this call will wait for at most timeout milliseconds for resources to become available.

Note: Passing a negative number as the timeout is equivalent to calling acquire(), i.e. this function will wait forever for resources to become available if timeout is negative.

Example:

QSemaphore sem(5);            // sem.available() == 5
sem.tryAcquire(250, 1000);    // sem.available() == 5, waits 1000 milliseconds and returns false
sem.tryAcquire(3, 30000);     // sem.available() == 2, returns true without waiting

See also acquire().