Module qt_core::sequential_animation_group::overloading

Traits

SequentialAnimationGroupConnectArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::connect method.

SequentialAnimationGroupConnectStaticArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::connect_static method.

SequentialAnimationGroupDestroyedArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::destroyed method.

SequentialAnimationGroupDisconnectArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::disconnect method.

SequentialAnimationGroupDisconnectStaticArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::disconnect_static method.

SequentialAnimationGroupDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::dynamic_property_names method.

SequentialAnimationGroupNewArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::new method.

SequentialAnimationGroupObjectNameArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::object_name method.

SequentialAnimationGroupPropertyArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::property method.

SequentialAnimationGroupStartArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::start method.

SequentialAnimationGroupStartTimerArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::start_timer method.

SequentialAnimationGroupTrArgs
This trait represents a set of arguments accepted by SequentialAnimationGroup::tr method.

SequentialAnimationGroupTrUtf8Args
This trait represents a set of arguments accepted by SequentialAnimationGroup::tr_utf8 method.