Module qt_core::sequential_iterable::overloading

Traits

ConstIteratorNewArgs
This trait represents a set of arguments accepted by ConstIterator::new method.

ConstIteratorOpAddArgs
This trait represents a set of arguments accepted by ConstIterator::op_add method.

ConstIteratorOpDecPostfixArgs
This trait represents a set of arguments accepted by ConstIterator::op_dec_postfix method.

ConstIteratorOpIncPostfixArgs
This trait represents a set of arguments accepted by ConstIterator::op_inc_postfix method.

ConstIteratorOpIndirectionArgs
This trait represents a set of arguments accepted by ConstIterator::op_indirection method.

ConstIteratorOpSubArgs
This trait represents a set of arguments accepted by ConstIterator::op_sub method.

SequentialIterableAtArgs
This trait represents a set of arguments accepted by SequentialIterable::at method.

SequentialIterableBeginArgs
This trait represents a set of arguments accepted by SequentialIterable::begin method.

SequentialIterableEndArgs
This trait represents a set of arguments accepted by SequentialIterable::end method.