Struct qt_core::sequential_iterable::ConstIterator

pub struct ConstIterator {
    // some fields omitted
}

C++ type: QSequentialIterable::const_iterator

C++ documentation:

The QSequentialIterable::const_iterator allows iteration over a container in a QVariant.

A QSequentialIterable::const_iterator can only be created by a QSequentialIterable instance, and can be used in a way similar to other stl-style iterators.

QList<int> intList;
intList.push_back(7);
intList.push_back(11);
intList.push_back(42);

QVariant variant = QVariant::fromValue(intList); if (variant.canConvert<QVariantList>()) { QSequentialIterable iterable = variant.value<QSequentialIterable>(); // Can use foreach: foreach (const QVariant &v, iterable) { qDebug() << v; } // Can use C++11 range-for: for (const QVariant &v : iterable) { qDebug() << v; } // Can use iterators: QSequentialIterable::const_iterator it = iterable.begin(); const QSequentialIterable::const_iterator end = iterable.end(); for ( ; it != end; ++it) { qDebug() << *it; } }

Methods

impl ConstIterator

fn new<Args>(args: Args) -> Args::ReturnType where Args: ConstIteratorNewArgs

C++ method: QSequentialIterable::const_iterator::const_iterator

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::sequential_iterable::ConstIterator, cpp_utils::AsStruct)) -> ::sequential_iterable::ConstIterator
2) fn new((&::sequential_iterable::ConstIterator, cpp_utils::AsBox)) -> cpp_utils::CppBox<::sequential_iterable::ConstIterator>
C++ method: [constructor] void QSequentialIterable::const_iterator::const_iterator(const QSequentialIterable::const_iterator& other)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpAddArgs<'largs>

C++ method: QSequentialIterable::const_iterator::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::sequential_iterable::ConstIterator>
2) fn op_add(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::sequential_iterable::ConstIterator
C++ method: QSequentialIterable::const_iterator QSequentialIterable::const_iterator::operator+(int j) const

C++ documentation:

Returns an iterator to the item at j positions forward from this iterator.

See also operator-() and operator+=().

fn op_add_assign<'l0>(&'l0 mut self, j: c_int) -> &'l0 mut ConstIterator

C++ method: QSequentialIterable::const_iterator& QSequentialIterable::const_iterator::operator+=(int j)

C++ documentation:

Advances the iterator by j items.

See also operator-=() and operator+().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 ConstIterator) -> &'l0 mut ConstIterator

C++ method: QSequentialIterable::const_iterator& QSequentialIterable::const_iterator::operator=(const QSequentialIterable::const_iterator& other)

C++ documentation:

Assigns other to this.

fn op_dec<'l0>(&'l0 mut self) -> &'l0 mut ConstIterator

C++ method: QSequentialIterable::const_iterator& QSequentialIterable::const_iterator::operator--()

C++ documentation:

The prefix -- operator (--it) makes the preceding item current and returns an iterator to the new current item.

Calling this function on QAssociativeIterable::begin() leads to undefined results.

See also operator++().

fn op_dec_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpDecPostfixArgs<'largs>

C++ method: QSequentialIterable::const_iterator::operator--

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::sequential_iterable::ConstIterator>
2) fn op_dec_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::sequential_iterable::ConstIterator
C++ method: QSequentialIterable::const_iterator QSequentialIterable::const_iterator::operator--(int arg1)

C++ documentation:

This is an overloaded function.

The postfix -- operator (it--) makes the preceding item current and returns an iterator to the previously current item.

fn op_eq(&self, o: &ConstIterator) -> bool

C++ method: bool QSequentialIterable::const_iterator::operator==(const QSequentialIterable::const_iterator& o) const

C++ documentation:

Returns true if other points to the same item as this iterator; otherwise returns false.

See also operator!=().

fn op_inc<'l0>(&'l0 mut self) -> &'l0 mut ConstIterator

C++ method: QSequentialIterable::const_iterator& QSequentialIterable::const_iterator::operator++()

C++ documentation:

The prefix ++ operator (++it) advances the iterator to the next item in the container and returns an iterator to the new current item.

Calling this function on QAssociativeIterable::end() leads to undefined results.

See also operator--().

fn op_inc_postfix<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpIncPostfixArgs<'largs>

C++ method: QSequentialIterable::const_iterator::operator++

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::sequential_iterable::ConstIterator>
2) fn op_inc_postfix(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::sequential_iterable::ConstIterator
C++ method: QSequentialIterable::const_iterator QSequentialIterable::const_iterator::operator++(int arg1)

C++ documentation:

This is an overloaded function.

The postfix ++ operator (it++) advances the iterator to the next item in the container and returns an iterator to the previously current item.

fn op_indirection<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpIndirectionArgs<'largs>

C++ method: QSequentialIterable::const_iterator::operator*

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_indirection(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn op_indirection(&self, cpp_utils::AsStruct) -> ::variant::Variant
C++ method: const QVariant QSequentialIterable::const_iterator::operator*() const

C++ documentation:

Returns the current value, converted to a QVariant.

fn op_neq(&self, o: &ConstIterator) -> bool

C++ method: bool QSequentialIterable::const_iterator::operator!=(const QSequentialIterable::const_iterator& o) const

C++ documentation:

Returns true if other points to a different item than this iterator; otherwise returns false.

See also operator==().

fn op_sub<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: ConstIteratorOpSubArgs<'largs>

C++ method: QSequentialIterable::const_iterator::operator-

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_sub(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::sequential_iterable::ConstIterator>
2) fn op_sub(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::sequential_iterable::ConstIterator
C++ method: QSequentialIterable::const_iterator QSequentialIterable::const_iterator::operator-(int j) const

C++ documentation:

Returns an iterator to the item at j positions backward from this iterator.

See also operator+() and operator-=().

fn op_sub_assign<'l0>(&'l0 mut self, j: c_int) -> &'l0 mut ConstIterator

C++ method: QSequentialIterable::const_iterator& QSequentialIterable::const_iterator::operator-=(int j)

C++ documentation:

Makes the iterator go back by j items.

See also operator+=() and operator-().