Struct qt_core::sequential_iterable::SequentialIterable

pub struct SequentialIterable {
    // some fields omitted
}

C++ type: QSequentialIterable

C++ documentation:

The QSequentialIterable class is an iterable interface for a container in a QVariant.

This class allows several methods of accessing the elements of a container held within a QVariant. An instance of QSequentialIterable can be extracted from a QVariant if it can be converted to a QVariantList.

QList<int> intList;
intList.push_back(7);
intList.push_back(11);
intList.push_back(42);

QVariant variant = QVariant::fromValue(intList); if (variant.canConvert<QVariantList>()) { QSequentialIterable iterable = variant.value<QSequentialIterable>(); // Can use foreach: foreach (const QVariant &v, iterable) { qDebug() << v; } // Can use C++11 range-for: for (const QVariant &v : iterable) { qDebug() << v; } // Can use iterators: QSequentialIterable::const_iterator it = iterable.begin(); const QSequentialIterable::const_iterator end = iterable.end(); for ( ; it != end; ++it) { qDebug() << *it; } }
The container itself is not copied before iterating over it.

Methods

impl SequentialIterable

fn at<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SequentialIterableAtArgs<'largs>

C++ method: QSequentialIterable::at

This is an overloaded function. Available variants:

Rust arguments:
1) fn at(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn at(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QSequentialIterable::at(int idx) const

C++ documentation:

Returns the element at position idx in the container.

fn begin<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SequentialIterableBeginArgs<'largs>

C++ method: QSequentialIterable::begin

This is an overloaded function. Available variants:

Rust arguments:
1) fn begin(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::sequential_iterable::ConstIterator>
2) fn begin(&self, cpp_utils::AsStruct) -> ::sequential_iterable::ConstIterator
C++ method: QSequentialIterable::const_iterator QSequentialIterable::begin() const

C++ documentation:

Returns a QSequentialIterable::const_iterator for the beginning of the container. This can be used in stl-style iteration.

See also end().

fn can_reverse_iterate(&self) -> bool

C++ method: bool QSequentialIterable::canReverseIterate() const

C++ documentation:

Returns whether it is possible to iterate over the container in reverse. This corresponds to the std::bidirectional_iterator_tag iterator trait of the const_iterator of the container.

fn end<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SequentialIterableEndArgs<'largs>

C++ method: QSequentialIterable::end

This is an overloaded function. Available variants:

Rust arguments:
1) fn end(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::sequential_iterable::ConstIterator>
2) fn end(&self, cpp_utils::AsStruct) -> ::sequential_iterable::ConstIterator
C++ method: QSequentialIterable::const_iterator QSequentialIterable::end() const

C++ documentation:

Returns a QSequentialIterable::const_iterator for the end of the container. This can be used in stl-style iteration.

See also begin().

fn size(&self) -> c_int

C++ method: int QSequentialIterable::size() const

C++ documentation:

Returns the number of elements in the container.