Struct qt_core::set::SetAbstractStateMutPtr

pub struct SetAbstractStateMutPtr {
    // some fields omitted
}

C++ type: QSet<QAbstractState*>

Methods

impl SetAbstractStateMutPtr

fn capacity(&self) -> c_int

C++ method: int QSet<QAbstractState*>::capacity() const

C++ documentation:

Returns the number of buckets in the set's internal hash table.

The sole purpose of this function is to provide a means of fine tuning QSet's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the set, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QSet<QAbstractState*>::clear()

C++ documentation:

Removes all elements from the set.

See also remove().

fn contains<'largs, Args>(&'largs self, args: Args) -> bool where Args: SetAbstractStateMutPtrContainsArgs<'largs>

C++ method: QSet<QAbstractState*>::contains

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn contains(&self, &*mut ::abstract_state::AbstractState) -> bool
C++ method: bool QSet<QAbstractState*>::contains(QAbstractState* const & value) const

C++ documentation:

Returns true if the set contains item value; otherwise returns false.

See also insert(), remove(), and find().

Variant 2

Rust arguments: fn contains(&self, &::set::SetAbstractStateMutPtr) -> bool
C++ method: bool QSet<QAbstractState*>::contains(const QSet<QAbstractState*>& set) const

C++ documentation:

Returns true if the set contains all items from the other set; otherwise returns false.

This function was introduced in Qt 4.6.

See also insert(), remove(), and find().

fn count(&self) -> c_int

C++ method: int QSet<QAbstractState*>::count() const

C++ documentation:

Same as size().

fn detach(&mut self)

C++ method: void QSet<QAbstractState*>::detach()

fn empty(&self) -> bool

C++ method: bool QSet<QAbstractState*>::empty() const

C++ documentation:

Returns true if the set is empty. This function is provided for STL compatibility. It is equivalent to isEmpty().

fn from_list<Args>(args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrFromListArgs

C++ method: QSet<QAbstractState*>::fromList

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_list((&::list::ListAbstractStateMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetAbstractStateMutPtr>
2) fn from_list((&::list::ListAbstractStateMutPtr, cpp_utils::AsStruct)) -> ::set::SetAbstractStateMutPtr
C++ method: static QSet<QAbstractState*> QSet<QAbstractState*>::fromList(const QList<QAbstractState*>& list)

C++ documentation:

Returns a new QSet object containing the data contained in list. Since QSet doesn't allow duplicates, the resulting QSet might be smaller than the list, because QList can contain duplicates.

Example:

QStringList list;
list << "Julia" << "Mike" << "Mike" << "Julia" << "Julia";

QSet<QString> set = QSet<QString>::fromList(list); set.contains("Julia"); // returns true set.contains("Mike"); // returns true set.size(); // returns 2
See also toList() and QList::toSet().

fn intersect<'l0, 'l1>(&'l0 mut self, other: &'l1 SetAbstractStateMutPtr) -> &'l0 mut SetAbstractStateMutPtr

C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::intersect(const QSet<QAbstractState*>& other)

C++ documentation:

Removes all items from this set that are not contained in the other set. A reference to this set is returned.

See also operator&=(), unite(), and subtract().

fn is_detached(&self) -> bool

C++ method: bool QSet<QAbstractState*>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QSet<QAbstractState*>::isEmpty() const

C++ documentation:

Returns true if the set contains no elements; otherwise returns false.

See also size().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrNewArgs

C++ method: QSet<QAbstractState*>::QSet

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::set::SetAbstractStateMutPtr
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::set::SetAbstractStateMutPtr>
C++ method: [constructor] void QSet<QAbstractState*>::QSet()

Variant 2

Rust arguments:
1) fn new((&::set::SetAbstractStateMutPtr, cpp_utils::AsStruct)) -> ::set::SetAbstractStateMutPtr
2) fn new((&::set::SetAbstractStateMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetAbstractStateMutPtr>
C++ method: [constructor] void QSet<QAbstractState*>::QSet(const QSet<QAbstractState*>& other)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrOpAddArgs<'largs>

C++ method: QSet<QAbstractState*>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetAbstractStateMutPtr>
2) fn op_add(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsStruct)) -> ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*> QSet<QAbstractState*>::operator+(const QSet<QAbstractState*>& other) const

C++ documentation:

Returns a new QSet that is the union of this set and the other set.

See also unite(), operator|=(), operator&(), and operator-().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetAbstractStateMutPtr where Args: SetAbstractStateMutPtrOpAddAssignArgs<'largs>

C++ method: QSet<QAbstractState*>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 *mut ::abstract_state::AbstractState) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator+=(QAbstractState* const & value)

C++ documentation:

Inserts a new item value and returns a reference to the set. If value already exists in the set, the set is left unchanged.

See also insert().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::set::SetAbstractStateMutPtr) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator+=(const QSet<QAbstractState*>& other)

C++ documentation:

Same as unite(other).

See also operator|(), operator&=(), and operator-=().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 SetAbstractStateMutPtr) -> &'l0 mut SetAbstractStateMutPtr

C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator=(const QSet<QAbstractState*>& other)

C++ documentation:

Assigns the other set to this set and returns a reference to this set.

fn op_bit_and<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrOpBitAndArgs<'largs>

C++ method: QSet<QAbstractState*>::operator&

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_bit_and(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetAbstractStateMutPtr>
2) fn op_bit_and(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsStruct)) -> ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*> QSet<QAbstractState*>::operator&(const QSet<QAbstractState*>& other) const

C++ documentation:

Returns a new QSet that is the intersection of this set and the other set.

See also intersect(), operator&=(), operator|(), and operator-().

fn op_bit_and_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetAbstractStateMutPtr where Args: SetAbstractStateMutPtrOpBitAndAssignArgs<'largs>

C++ method: QSet<QAbstractState*>::operator&=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_bit_and_assign(&mut self, &'l1 *mut ::abstract_state::AbstractState) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator&=(QAbstractState* const & value)

C++ documentation:

This is an overloaded function.

Same as intersect(other), if we consider other to be a set that contains the singleton value.

Variant 2

Rust arguments: fn op_bit_and_assign(&mut self, &'l1 ::set::SetAbstractStateMutPtr) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator&=(const QSet<QAbstractState*>& other)

C++ documentation:

Same as intersect(other).

See also operator&(), operator|=(), and operator-=().

fn op_bit_or<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrOpBitOrArgs<'largs>

C++ method: QSet<QAbstractState*>::operator|

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_bit_or(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetAbstractStateMutPtr>
2) fn op_bit_or(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsStruct)) -> ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*> QSet<QAbstractState*>::operator|(const QSet<QAbstractState*>& other) const

C++ documentation:

Returns a new QSet that is the union of this set and the other set.

See also unite(), operator|=(), operator&(), and operator-().

fn op_bit_or_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetAbstractStateMutPtr where Args: SetAbstractStateMutPtrOpBitOrAssignArgs<'largs>

C++ method: QSet<QAbstractState*>::operator|=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_bit_or_assign(&mut self, &'l1 *mut ::abstract_state::AbstractState) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator|=(QAbstractState* const & value)

C++ documentation:

Inserts a new item value and returns a reference to the set. If value already exists in the set, the set is left unchanged.

See also insert().

Variant 2

Rust arguments: fn op_bit_or_assign(&mut self, &'l1 ::set::SetAbstractStateMutPtr) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator|=(const QSet<QAbstractState*>& other)

C++ documentation:

Same as unite(other).

See also operator|(), operator&=(), and operator-=().

fn op_eq(&self, other: &SetAbstractStateMutPtr) -> bool

C++ method: bool QSet<QAbstractState*>::operator==(const QSet<QAbstractState*>& other) const

C++ documentation:

Returns true if the other set is equal to this set; otherwise returns false.

Two sets are considered equal if they contain the same elements.

This function requires the value type to implement operator==().

See also operator!=().

fn op_neq(&self, other: &SetAbstractStateMutPtr) -> bool

C++ method: bool QSet<QAbstractState*>::operator!=(const QSet<QAbstractState*>& other) const

C++ documentation:

Returns true if the other set is not equal to this set; otherwise returns false.

Two sets are considered equal if they contain the same elements.

This function requires the value type to implement operator==().

See also operator==().

fn op_shl<'l0, 'l1>(&'l0 mut self, value: &'l1 *mut AbstractState) -> &'l0 mut SetAbstractStateMutPtr

C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator<<(QAbstractState* const & value)

C++ documentation:

Inserts a new item value and returns a reference to the set. If value already exists in the set, the set is left unchanged.

See also insert().

fn op_sub<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrOpSubArgs<'largs>

C++ method: QSet<QAbstractState*>::operator-

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_sub(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsBox)) -> cpp_utils::CppBox<::set::SetAbstractStateMutPtr>
2) fn op_sub(&self, (&::set::SetAbstractStateMutPtr, cpp_utils::AsStruct)) -> ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*> QSet<QAbstractState*>::operator-(const QSet<QAbstractState*>& other) const

C++ documentation:

Returns a new QSet that is the set difference of this set and the other set, i.e., this set - other set.

See also subtract(), operator-=(), operator|(), and operator&().

fn op_sub_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut SetAbstractStateMutPtr where Args: SetAbstractStateMutPtrOpSubAssignArgs<'largs>

C++ method: QSet<QAbstractState*>::operator-=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_sub_assign(&mut self, &'l1 *mut ::abstract_state::AbstractState) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator-=(QAbstractState* const & value)

C++ documentation:

Removes the occurrence of item value from the set, if it is found, and returns a reference to the set. If the value is not contained the set, nothing is removed.

See also remove().

Variant 2

Rust arguments: fn op_sub_assign(&mut self, &'l1 ::set::SetAbstractStateMutPtr) -> &'l0 mut ::set::SetAbstractStateMutPtr
C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::operator-=(const QSet<QAbstractState*>& other)

C++ documentation:

Same as subtract(other).

See also operator-(), operator|=(), and operator&=().

fn remove(&mut self, value: &*mut AbstractState) -> bool

C++ method: bool QSet<QAbstractState*>::remove(QAbstractState* const & value)

C++ documentation:

Removes any occurrence of item value from the set. Returns true if an item was actually removed; otherwise returns false.

See also contains() and insert().

fn reserve(&mut self, size: c_int)

C++ method: void QSet<QAbstractState*>::reserve(int size)

C++ documentation:

Ensures that the set's internal hash table consists of at least size buckets.

This function is useful for code that needs to build a huge set and wants to avoid repeated reallocation. For example:

QSet<QString> set;
set.reserve(20000);
for (int i = 0; i < 20000; ++i)
    set.insert(values[i]);

Ideally, size should be slightly more than the maximum number of elements expected in the set. size doesn't have to be prime, because QSet will use a prime number internally anyway. If size is an underestimate, the worst that will happen is that the QSet will be a bit slower.

In general, you will rarely ever need to call this function. QSet's internal hash table automatically shrinks or grows to provide good performance without wasting too much memory.

See also squeeze() and capacity().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QSet<QAbstractState*>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QSet<QAbstractState*>::size() const

C++ documentation:

Returns the number of items in the set.

See also isEmpty() and count().

fn squeeze(&mut self)

C++ method: void QSet<QAbstractState*>::squeeze()

C++ documentation:

Reduces the size of the set's internal hash table to save memory.

The sole purpose of this function is to provide a means of fine tuning QSet's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn subtract<'l0, 'l1>(&'l0 mut self, other: &'l1 SetAbstractStateMutPtr) -> &'l0 mut SetAbstractStateMutPtr

C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::subtract(const QSet<QAbstractState*>& other)

C++ documentation:

Removes all items from this set that are contained in the other set. Returns a reference to this set.

See also operator-=(), unite(), and intersect().

fn swap(&mut self, other: &mut SetAbstractStateMutPtr)

C++ method: void QSet<QAbstractState*>::swap(QSet<QAbstractState*>& other)

C++ documentation:

Swaps set other with this set. This operation is very fast and never fails.

fn to_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrToListArgs<'largs>

C++ method: QSet<QAbstractState*>::toList

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListAbstractStateMutPtr>
2) fn to_list(&self, cpp_utils::AsStruct) -> ::list::ListAbstractStateMutPtr
C++ method: QList<QAbstractState*> QSet<QAbstractState*>::toList() const

C++ documentation:

Returns a new QList containing the elements in the set. The order of the elements in the QList is undefined.

Example:

QSet<QString> set;
set << "red" << "green" << "blue" << ... << "black";

QList<QString> list = set.toList(); qSort(list);
See also fromList() and QList::fromSet().

fn unite<'l0, 'l1>(&'l0 mut self, other: &'l1 SetAbstractStateMutPtr) -> &'l0 mut SetAbstractStateMutPtr

C++ method: QSet<QAbstractState*>& QSet<QAbstractState*>::unite(const QSet<QAbstractState*>& other)

C++ documentation:

Each item in the other set that isn't already in this set is inserted into this set. A reference to this set is returned.

See also operator|=(), intersect(), and subtract().

fn values<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SetAbstractStateMutPtrValuesArgs<'largs>

C++ method: QSet<QAbstractState*>::values

This is an overloaded function. Available variants:

Rust arguments:
1) fn values(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListAbstractStateMutPtr>
2) fn values(&self, cpp_utils::AsStruct) -> ::list::ListAbstractStateMutPtr
C++ method: QList<QAbstractState*> QSet<QAbstractState*>::values() const

C++ documentation:

Returns a new QList containing the elements in the set. The order of the elements in the QList is undefined.

This is the same as toList().

See also fromList() and QList::fromSet().