Enum qt_core::settings::Format

pub enum Format {
    NativeFormat,
    IniFormat,
    InvalidFormat,
    CustomFormat1,
    CustomFormat2,
    CustomFormat3,
    CustomFormat4,
    CustomFormat5,
    CustomFormat6,
    CustomFormat7,
    CustomFormat8,
    CustomFormat9,
    CustomFormat10,
    CustomFormat11,
    CustomFormat12,
    CustomFormat13,
    CustomFormat14,
    CustomFormat15,
    CustomFormat16,
}

C++ type: QSettings::Format

C++ documentation:

This enum type specifies the storage format used by QSettings.

On Unix, NativeFormat and IniFormat mean the same thing, except that the file extension is different (.conf for NativeFormat, .ini for IniFormat).

The INI file format is a Windows file format that Qt supports on all platforms. In the absence of an INI standard, we try to follow what Microsoft does, with the following exceptions:

If you store types that QVariant can't convert to QString (e.g., QPoint, QRect, and QSize), Qt uses an @-based syntax to encode the type. For example:
pos = @Point(100 100)

To minimize compatibility issues, any @ that doesn't appear at the first position in the value or that isn't followed by a Qt type (Point, Rect, Size, etc.) is treated as a normal character.

Although backslash is a special character in INI files, most Windows applications don't escape backslashes (\) in file paths:
windir = C:\Windows

QSettings always treats backslash as a special character and provides no API for reading or writing such entries.

The INI file format has severe restrictions on the syntax of a key. Qt works around this by using % as an escape character in keys. In addition, if you save a top-level setting (a key with no slashes in it, e.g., "someKey"), it will appear in the INI file's "General" section. To avoid overwriting other keys, if you save something using a key such as "General/someKey", the key will be located in the "%General" section, not in the "General" section.

Following the philosophy that we should be liberal in what we accept and conservative in what we generate, QSettings will accept Latin-1 encoded INI files, but generate pure ASCII files, where non-ASCII values are encoded using standard INI escape sequences. To make the INI files more readable (but potentially less compatible), call setIniCodec().

See also registerFormat() and setPath().

Variants
NativeFormat

Store the settings using the most appropriate storage format for the platform. On Windows, this means the system registry; on OS X and iOS, this means the CFPreferences API; on Unix, this means textual configuration files in INI format. (C++ enum variant: NativeFormat = 0)

IniFormat

Store the settings in INI files. (C++ enum variant: IniFormat = 1)

InvalidFormat

Special value returned by registerFormat(). (C++ enum variant: InvalidFormat = 16)

CustomFormat1

C++ enum variant: CustomFormat1 = 17

CustomFormat2

C++ enum variant: CustomFormat2 = 18

CustomFormat3

C++ enum variant: CustomFormat3 = 19

CustomFormat4

C++ enum variant: CustomFormat4 = 20

CustomFormat5

C++ enum variant: CustomFormat5 = 21

CustomFormat6

C++ enum variant: CustomFormat6 = 22

CustomFormat7

C++ enum variant: CustomFormat7 = 23

CustomFormat8

C++ enum variant: CustomFormat8 = 24

CustomFormat9

C++ enum variant: CustomFormat9 = 25

CustomFormat10

C++ enum variant: CustomFormat10 = 26

CustomFormat11

C++ enum variant: CustomFormat11 = 27

CustomFormat12

C++ enum variant: CustomFormat12 = 28

CustomFormat13

C++ enum variant: CustomFormat13 = 29

CustomFormat14

C++ enum variant: CustomFormat14 = 30

CustomFormat15

C++ enum variant: CustomFormat15 = 31

CustomFormat16

C++ enum variant: CustomFormat16 = 32