Module qt_core::settings::overloading

Traits

SettingsAllKeysArgs
This trait represents a set of arguments accepted by Settings::all_keys method.

SettingsApplicationNameArgs
This trait represents a set of arguments accepted by Settings::application_name method.

SettingsBeginWriteArrayArgs
This trait represents a set of arguments accepted by Settings::begin_write_array method.

SettingsChildGroupsArgs
This trait represents a set of arguments accepted by Settings::child_groups method.

SettingsChildKeysArgs
This trait represents a set of arguments accepted by Settings::child_keys method.

SettingsConnectArgs
This trait represents a set of arguments accepted by Settings::connect method.

SettingsConnectStaticArgs
This trait represents a set of arguments accepted by Settings::connect_static method.

SettingsDestroyedArgs
This trait represents a set of arguments accepted by Settings::destroyed method.

SettingsDisconnectArgs
This trait represents a set of arguments accepted by Settings::disconnect method.

SettingsDisconnectStaticArgs
This trait represents a set of arguments accepted by Settings::disconnect_static method.

SettingsDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Settings::dynamic_property_names method.

SettingsFileNameArgs
This trait represents a set of arguments accepted by Settings::file_name method.

SettingsGroupArgs
This trait represents a set of arguments accepted by Settings::group method.

SettingsNewArgs
This trait represents a set of arguments accepted by Settings::new method.

SettingsObjectNameArgs
This trait represents a set of arguments accepted by Settings::object_name method.

SettingsOrganizationNameArgs
This trait represents a set of arguments accepted by Settings::organization_name method.

SettingsPropertyArgs
This trait represents a set of arguments accepted by Settings::property method.

SettingsSetIniCodecArgs
This trait represents a set of arguments accepted by Settings::set_ini_codec method.

SettingsStartTimerArgs
This trait represents a set of arguments accepted by Settings::start_timer method.

SettingsTrArgs
This trait represents a set of arguments accepted by Settings::tr method.

SettingsTrUtf8Args
This trait represents a set of arguments accepted by Settings::tr_utf8 method.

SettingsValueArgs
This trait represents a set of arguments accepted by Settings::value method.