Module qt_core::shared_memory::overloading

Traits

SharedMemoryAttachArgs
This trait represents a set of arguments accepted by SharedMemory::attach method.

SharedMemoryConnectArgs
This trait represents a set of arguments accepted by SharedMemory::connect method.

SharedMemoryConnectStaticArgs
This trait represents a set of arguments accepted by SharedMemory::connect_static method.

SharedMemoryCreateArgs
This trait represents a set of arguments accepted by SharedMemory::create method.

SharedMemoryDestroyedArgs
This trait represents a set of arguments accepted by SharedMemory::destroyed method.

SharedMemoryDisconnectArgs
This trait represents a set of arguments accepted by SharedMemory::disconnect method.

SharedMemoryDisconnectStaticArgs
This trait represents a set of arguments accepted by SharedMemory::disconnect_static method.

SharedMemoryDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SharedMemory::dynamic_property_names method.

SharedMemoryErrorStringArgs
This trait represents a set of arguments accepted by SharedMemory::error_string method.

SharedMemoryKeyArgs
This trait represents a set of arguments accepted by SharedMemory::key method.

SharedMemoryNativeKeyArgs
This trait represents a set of arguments accepted by SharedMemory::native_key method.

SharedMemoryNewArgs
This trait represents a set of arguments accepted by SharedMemory::new method.

SharedMemoryObjectNameArgs
This trait represents a set of arguments accepted by SharedMemory::object_name method.

SharedMemoryPropertyArgs
This trait represents a set of arguments accepted by SharedMemory::property method.

SharedMemoryStartTimerArgs
This trait represents a set of arguments accepted by SharedMemory::start_timer method.

SharedMemoryTrArgs
This trait represents a set of arguments accepted by SharedMemory::tr method.

SharedMemoryTrUtf8Args
This trait represents a set of arguments accepted by SharedMemory::tr_utf8 method.