Struct qt_core::shared_pointer::SharedPointerObjectRef

pub struct SharedPointerObjectRef {
    // some fields omitted
}

C++ type: QSharedPointer<QObject>

Methods

impl SharedPointerObjectRef

fn clear(&mut self)

C++ method: void QSharedPointer<QObject>::clear()

C++ documentation:

Clears this QSharedPointer object, dropping the reference that it may have had to the pointer. If this was the last reference, then the pointer itself will be deleted.

fn data(&self) -> *mut Object

C++ method: QObject* QSharedPointer<QObject>::data() const

C++ documentation:

Returns the value of the pointer referenced by this object.

Note: do not delete the pointer returned by this function or pass it to another function that could delete it, including creating QSharedPointer or QWeakPointer objects.

fn is_null(&self) -> bool

C++ method: bool QSharedPointer<QObject>::isNull() const

C++ documentation:

Returns true if this object is holding a reference to a null pointer.

fn new<Args>(args: Args) -> Args::ReturnType where Args: SharedPointerObjectRefNewArgs

C++ method: QSharedPointer<QObject>::QSharedPointer

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::shared_pointer::SharedPointerObjectRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::shared_pointer::SharedPointerObjectRef>
C++ method: [constructor] void QSharedPointer<QObject>::QSharedPointer()

Variant 2

Rust arguments:
1) fn new((*mut ::object::Object, cpp_utils::AsStruct)) -> ::shared_pointer::SharedPointerObjectRef
2) fn new((*mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::shared_pointer::SharedPointerObjectRef>
C++ method: [constructor] void QSharedPointer<QObject>::QSharedPointer(QObject* ptr)

Variant 3

Rust arguments:
1) fn new((&::shared_pointer::SharedPointerObjectRef, cpp_utils::AsStruct)) -> ::shared_pointer::SharedPointerObjectRef
2) fn new((&::shared_pointer::SharedPointerObjectRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::shared_pointer::SharedPointerObjectRef>
C++ method: [constructor] void QSharedPointer<QObject>::QSharedPointer(const QSharedPointer<QObject>& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 SharedPointerObjectRef) -> &'l0 mut SharedPointerObjectRef

C++ method: QSharedPointer<QObject>& QSharedPointer<QObject>::operator=(const QSharedPointer<QObject>& other)

fn op_indirection<'l0>(&'l0 self) -> &'l0 mut Object

C++ method: QObject& QSharedPointer<QObject>::operator*() const

C++ documentation:

Provides access to the shared pointer's members.

See also isNull().

fn op_not(&self) -> bool

C++ method: bool QSharedPointer<QObject>::operator!() const

C++ documentation:

Returns true if this object is null. This function is suitable for use in if-constructs, like:

if (!sharedptr) { ... }

See also isNull().

fn op_struct_deref(&self) -> *mut Object

C++ method: QObject* QSharedPointer<QObject>::operator->() const

C++ documentation:

Provides access to the shared pointer's members.

See also isNull().

fn reset<'largs, Args>(&'largs mut self, args: Args) where Args: SharedPointerObjectRefResetArgs<'largs>

C++ method: QSharedPointer<QObject>::reset

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn reset(&mut self, ()) -> ()
C++ method: void QSharedPointer<QObject>::reset()

C++ documentation:

Same as clear(). For std::shared_ptr compatibility.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments: fn reset(&mut self, *mut ::object::Object) -> ()
C++ method: void QSharedPointer<QObject>::reset(QObject* t)

C++ documentation:

Resets this QSharedPointer object to point to t instead. Equivalent to:

QSharedPointer<T> other(t); this->swap(other);

This function was introduced in Qt 5.0.

fn swap(&mut self, other: &mut SharedPointerObjectRef)

C++ method: void QSharedPointer<QObject>::swap(QSharedPointer<QObject>& other)

Warning: no exact match found in C++ documentation.Below is the C++ documentation for void QSharedPointer::swap(QSharedPointer & other):

Swaps this shared pointer instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.3.

fn to_weak_ref<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SharedPointerObjectRefToWeakRefArgs<'largs>

C++ method: QSharedPointer<QObject>::toWeakRef

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_weak_ref(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::weak_pointer::WeakPointerObjectRef>
2) fn to_weak_ref(&self, cpp_utils::AsStruct) -> ::weak_pointer::WeakPointerObjectRef
C++ method: QWeakPointer<QObject> QSharedPointer<QObject>::toWeakRef() const

C++ documentation:

Returns a weak reference object that shares the pointer referenced by this object.

See also QWeakPointer::QWeakPointer().