Struct qt_core::signal_blocker::SignalBlocker

pub struct SignalBlocker {
    // some fields omitted
}

C++ type: QSignalBlocker

C++ documentation:

Exception-safe wrapper around QObject::blockSignals()

QSignalBlocker can be used whereever you would otherwise use a pair of calls to blockSignals(). It blocks signals in its constructor and in the destructor it resets the state to what it was before the constructor ran.

{
const QSignalBlocker blocker(someQObject);
// no signals here
}

is thus equivalent to

const bool wasBlocked = someQObject->blockSignals(true);
// no signals here
someQObject->blockSignals(wasBlocked);

except the code using QSignalBlocker is safe in the face of exceptions.

Methods

impl SignalBlocker

fn new<Args>(args: Args) -> Args::ReturnType where Args: SignalBlockerNewArgs

C++ method: QSignalBlocker::QSignalBlocker

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((&mut ::object::Object, cpp_utils::AsStruct)) -> ::signal_blocker::SignalBlocker
2) fn new((&mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::signal_blocker::SignalBlocker>
C++ method: [constructor] void QSignalBlocker::QSignalBlocker(QObject& o)

Variant 2

Rust arguments:
1) fn new((*mut ::object::Object, cpp_utils::AsStruct)) -> ::signal_blocker::SignalBlocker
2) fn new((*mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::signal_blocker::SignalBlocker>
C++ method: [constructor] void QSignalBlocker::QSignalBlocker(QObject* o)

fn reblock(&mut self)

C++ method: void QSignalBlocker::reblock()

C++ documentation:

Re-blocks signals after a previous unblock().

The numbers of reblock() and unblock() calls are not counted, so every reblock() undoes any number of unblock() calls.

fn unblock(&mut self)

C++ method: void QSignalBlocker::unblock()

C++ documentation:

Temporarily restores the QObject::signalsBlocked() state to what it was before this QSignaBlocker's constructor ran. To undo, use reblock().

The numbers of reblock() and unblock() calls are not counted, so every unblock() undoes any number of reblock() calls.