Module qt_core::signal_mapper::overloading

Traits

SignalMapperConnectArgs
This trait represents a set of arguments accepted by SignalMapper::connect method.

SignalMapperConnectStaticArgs
This trait represents a set of arguments accepted by SignalMapper::connect_static method.

SignalMapperDestroyedArgs
This trait represents a set of arguments accepted by SignalMapper::destroyed method.

SignalMapperDisconnectArgs
This trait represents a set of arguments accepted by SignalMapper::disconnect method.

SignalMapperDisconnectStaticArgs
This trait represents a set of arguments accepted by SignalMapper::disconnect_static method.

SignalMapperDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SignalMapper::dynamic_property_names method.

SignalMapperMapArgs
This trait represents a set of arguments accepted by SignalMapper::map method.

SignalMapperMappedArgs
This trait represents a set of arguments accepted by SignalMapper::mapped method.

SignalMapperMappingArgs
This trait represents a set of arguments accepted by SignalMapper::mapping method.

SignalMapperNewArgs
This trait represents a set of arguments accepted by SignalMapper::new method.

SignalMapperObjectNameArgs
This trait represents a set of arguments accepted by SignalMapper::object_name method.

SignalMapperPropertyArgs
This trait represents a set of arguments accepted by SignalMapper::property method.

SignalMapperSetMappingArgs
This trait represents a set of arguments accepted by SignalMapper::set_mapping method.

SignalMapperStartTimerArgs
This trait represents a set of arguments accepted by SignalMapper::start_timer method.

SignalMapperTrArgs
This trait represents a set of arguments accepted by SignalMapper::tr method.

SignalMapperTrUtf8Args
This trait represents a set of arguments accepted by SignalMapper::tr_utf8 method.