Module qt_core::signal_transition::overloading

Traits

SignalTransitionAnimationsArgs
This trait represents a set of arguments accepted by SignalTransition::animations method.

SignalTransitionConnectArgs
This trait represents a set of arguments accepted by SignalTransition::connect method.

SignalTransitionConnectStaticArgs
This trait represents a set of arguments accepted by SignalTransition::connect_static method.

SignalTransitionDestroyedArgs
This trait represents a set of arguments accepted by SignalTransition::destroyed method.

SignalTransitionDisconnectArgs
This trait represents a set of arguments accepted by SignalTransition::disconnect method.

SignalTransitionDisconnectStaticArgs
This trait represents a set of arguments accepted by SignalTransition::disconnect_static method.

SignalTransitionDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SignalTransition::dynamic_property_names method.

SignalTransitionNewArgs
This trait represents a set of arguments accepted by SignalTransition::new method.

SignalTransitionObjectNameArgs
This trait represents a set of arguments accepted by SignalTransition::object_name method.

SignalTransitionPropertyArgs
This trait represents a set of arguments accepted by SignalTransition::property method.

SignalTransitionSignalArgs
This trait represents a set of arguments accepted by SignalTransition::signal method.

SignalTransitionStartTimerArgs
This trait represents a set of arguments accepted by SignalTransition::start_timer method.

SignalTransitionTargetStatesArgs
This trait represents a set of arguments accepted by SignalTransition::target_states method.

SignalTransitionTrArgs
This trait represents a set of arguments accepted by SignalTransition::tr method.

SignalTransitionTrUtf8Args
This trait represents a set of arguments accepted by SignalTransition::tr_utf8 method.