Module qt_core::size::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

OpShrArgs
This trait represents a set of arguments accepted by op_shr method.

SizeBoundedToArgs
This trait represents a set of arguments accepted by Size::bounded_to method.

SizeExpandedToArgs
This trait represents a set of arguments accepted by Size::expanded_to method.

SizeNewArgs
This trait represents a set of arguments accepted by Size::new method.

SizeScaleArgs
This trait represents a set of arguments accepted by Size::scale method.

SizeScaledArgs
This trait represents a set of arguments accepted by Size::scaled method.

SizeTransposedArgs
This trait represents a set of arguments accepted by Size::transposed method.