Struct qt_core::size::Size

pub struct Size {
    // some fields omitted
}

C++ type: QSize

C++ documentation:

The QSize class defines the size of a two-dimensional object using integer point precision.

A size is specified by a width() and a height(). It can be set in the constructor and changed using the setWidth(), setHeight(), or scale() functions, or using arithmetic operators. A size can also be manipulated directly by retrieving references to the width and height using the rwidth() and rheight() functions. Finally, the width and height can be swapped using the transpose() function.

The isValid() function determines if a size is valid (a valid size has both width and height greater than or equal to zero). The isEmpty() function returns true if either of the width and height is less than, or equal to, zero, while the isNull() function returns true only if both the width and the height is zero.

Use the expandedTo() function to retrieve a size which holds the maximum height and width of this size and a given size. Similarly, the boundedTo() function returns a size which holds the minimum height and width of this size and a given size.

QSize objects can be streamed as well as compared.

Methods

impl Size

fn bounded_to<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeBoundedToArgs<'largs>

C++ method: QSize::boundedTo

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounded_to(&self, (&::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size::Size>
2) fn bounded_to(&self, (&::size::Size, cpp_utils::AsStruct)) -> ::size::Size
C++ method: QSize QSize::boundedTo(const QSize& arg1) const

C++ documentation:

Returns a size holding the minimum width and height of this size and the given otherSize.

See also expandedTo() and scale().

fn expanded_to<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeExpandedToArgs<'largs>

C++ method: QSize::expandedTo

This is an overloaded function. Available variants:

Rust arguments:
1) fn expanded_to(&self, (&::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size::Size>
2) fn expanded_to(&self, (&::size::Size, cpp_utils::AsStruct)) -> ::size::Size
C++ method: QSize QSize::expandedTo(const QSize& arg1) const

C++ documentation:

Returns a size holding the maximum width and height of this size and the given otherSize.

See also boundedTo() and scale().

fn height(&self) -> c_int

C++ method: int QSize::height() const

C++ documentation:

Returns the height.

See also width() and setHeight().

fn is_empty(&self) -> bool

C++ method: bool QSize::isEmpty() const

C++ documentation:

Returns true if either of the width and height is less than or equal to 0; otherwise returns false.

See also isNull() and isValid().

fn is_null(&self) -> bool

C++ method: bool QSize::isNull() const

C++ documentation:

Returns true if both the width and height is 0; otherwise returns false.

See also isValid() and isEmpty().

fn is_valid(&self) -> bool

C++ method: bool QSize::isValid() const

C++ documentation:

Returns true if both the width and height is equal to or greater than 0; otherwise returns false.

See also isNull() and isEmpty().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SizeNewArgs

C++ method: QSize::QSize

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::size::Size
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::size::Size>
C++ method: [constructor] void QSize::QSize()

Variant 2

Rust arguments:
1) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::size::Size
2) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size::Size>
C++ method: [constructor] void QSize::QSize(int w, int h)

fn op_add_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 Size) -> &'l0 mut Size

C++ method: QSize& QSize::operator+=(const QSize& arg1)

C++ documentation:

Adds the given size to this size, and returns a reference to this size. For example:

QSize s( 3, 7);
QSize r(-1, 4);
s += r;

// s becomes (2,11)

fn op_div_assign<'l0>(&'l0 mut self, c: c_double) -> &'l0 mut Size

C++ method: QSize& QSize::operator/=(double c)

C++ documentation:

This is an overloaded function.

Divides both the width and height by the given divisor, and returns a reference to the size.

Note that the result is rounded to the nearest integer.

See also QSize::scale().

fn op_mul_assign<'l0>(&'l0 mut self, c: c_double) -> &'l0 mut Size

C++ method: QSize& QSize::operator*=(double c)

C++ documentation:

This is an overloaded function.

Multiplies both the width and height by the given factor, and returns a reference to the size.

Note that the result is rounded to the nearest integer.

See also scale().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 Size) -> &'l0 mut Size

C++ method: QSize& QSize::operator-=(const QSize& arg1)

C++ documentation:

Subtracts the given size from this size, and returns a reference to this size. For example:

QSize s( 3, 7);
QSize r(-1, 4);
s -= r;

// s becomes (4,3)

fn rheight<'l0>(&'l0 mut self) -> &'l0 mut c_int

C++ method: int& QSize::rheight()

C++ documentation:

Returns a reference to the height.

Using a reference makes it possible to manipulate the height directly. For example:

QSize size(100, 10);
size.rheight() += 5;

// size becomes (100,15)
See also rwidth() and setHeight().

fn rwidth<'l0>(&'l0 mut self) -> &'l0 mut c_int

C++ method: int& QSize::rwidth()

C++ documentation:

Returns a reference to the width.

Using a reference makes it possible to manipulate the width directly. For example:

QSize size(100, 10);
size.rwidth() += 20;

// size becomes (120,10)
See also rheight() and setWidth().

fn scale<'largs, Args>(&'largs mut self, args: Args) where Args: SizeScaleArgs<'largs>

C++ method: QSize::scale

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn scale(&mut self, (&::size::Size, ::qt::AspectRatioMode)) -> ()
C++ method: void QSize::scale(const QSize& s, Qt::AspectRatioMode mode)

C++ documentation:

This is an overloaded function.

Scales the size to a rectangle with the given size, according to the specified mode.

Variant 2

Rust arguments: fn scale(&mut self, (libc::c_int, libc::c_int, ::qt::AspectRatioMode)) -> ()
C++ method: void QSize::scale(int w, int h, Qt::AspectRatioMode mode)

C++ documentation:

Scales the size to a rectangle with the given width and height, according to the specified mode:

If mode is Qt::IgnoreAspectRatio, the size is set to (width, height).

If mode is Qt::KeepAspectRatio, the current size is scaled to a rectangle as large as possible inside (width, height), preserving the aspect ratio.

If mode is Qt::KeepAspectRatioByExpanding, the current size is scaled to a rectangle as small as possible outside (width, height), preserving the aspect ratio.

Example:

QSize t1(10, 12);
t1.scale(60, 60, Qt::IgnoreAspectRatio);
// t1 is (60, 60)

QSize t2(10, 12); t2.scale(60, 60, Qt::KeepAspectRatio); // t2 is (50, 60)

QSize t3(10, 12); t3.scale(60, 60, Qt::KeepAspectRatioByExpanding); // t3 is (60, 72)
See also setWidth(), setHeight(), and scaled().

fn scaled<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeScaledArgs<'largs>

C++ method: QSize::scaled

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn scaled(&self, (&::size::Size, ::qt::AspectRatioMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size::Size>
2) fn scaled(&self, (&::size::Size, ::qt::AspectRatioMode, cpp_utils::AsStruct)) -> ::size::Size
C++ method: QSize QSize::scaled(const QSize& s, Qt::AspectRatioMode mode) const

C++ documentation:

This is an overloaded function.

Return a size scaled to a rectangle with the given size s, according to the specified mode.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments:
1) fn scaled(&self, (libc::c_int, libc::c_int, ::qt::AspectRatioMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size::Size>
2) fn scaled(&self, (libc::c_int, libc::c_int, ::qt::AspectRatioMode, cpp_utils::AsStruct)) -> ::size::Size
C++ method: QSize QSize::scaled(int w, int h, Qt::AspectRatioMode mode) const

C++ documentation:

Return a size scaled to a rectangle with the given width and height, according to the specified mode.

This function was introduced in Qt 5.0.

See also scale().

fn set_height(&mut self, h: c_int)

C++ method: void QSize::setHeight(int h)

C++ documentation:

Sets the height to the given height.

See also rheight(), height(), and setWidth().

fn set_width(&mut self, w: c_int)

C++ method: void QSize::setWidth(int w)

C++ documentation:

Sets the width to the given width.

See also rwidth(), width(), and setHeight().

fn transpose(&mut self)

C++ method: void QSize::transpose()

C++ documentation:

Swaps the width and height values.

See also setWidth(), setHeight(), and transposed().

fn transposed<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeTransposedArgs<'largs>

C++ method: QSize::transposed

This is an overloaded function. Available variants:

Rust arguments:
1) fn transposed(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size::Size>
2) fn transposed(&self, cpp_utils::AsStruct) -> ::size::Size
C++ method: QSize QSize::transposed() const

C++ documentation:

Returns a QSize with width and height swapped.

This function was introduced in Qt 5.0.

See also transpose().

fn width(&self) -> c_int

C++ method: int QSize::width() const

C++ documentation:

Returns the width.

See also height() and setWidth().