Module qt_core::size_f::overloading

Traits

SizeFBoundedToArgs
This trait represents a set of arguments accepted by SizeF::bounded_to method.

SizeFExpandedToArgs
This trait represents a set of arguments accepted by SizeF::expanded_to method.

SizeFNewArgs
This trait represents a set of arguments accepted by SizeF::new method.

SizeFScaleArgs
This trait represents a set of arguments accepted by SizeF::scale method.

SizeFScaledArgs
This trait represents a set of arguments accepted by SizeF::scaled method.

SizeFToSizeArgs
This trait represents a set of arguments accepted by SizeF::to_size method.

SizeFTransposedArgs
This trait represents a set of arguments accepted by SizeF::transposed method.