Struct qt_core::size_f::SizeF

pub struct SizeF {
    // some fields omitted
}

C++ type: QSizeF

C++ documentation:

The QSizeF class defines the size of a two-dimensional object using floating point precision.

A size is specified by a width() and a height(). It can be set in the constructor and changed using the setWidth(), setHeight(), or scale() functions, or using arithmetic operators. A size can also be manipulated directly by retrieving references to the width and height using the rwidth() and rheight() functions. Finally, the width and height can be swapped using the transpose() function.

The isValid() function determines if a size is valid. A valid size has both width and height greater than or equal to zero. The isEmpty() function returns true if either of the width and height is less than (or equal to) zero, while the isNull() function returns true only if both the width and the height is zero.

Use the expandedTo() function to retrieve a size which holds the maximum height and width of this size and a given size. Similarly, the boundedTo() function returns a size which holds the minimum height and width of this size and a given size.

The QSizeF class also provides the toSize() function returning a QSize copy of this size, constructed by rounding the width and height to the nearest integers.

QSizeF objects can be streamed as well as compared.

Methods

impl SizeF

fn bounded_to<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeFBoundedToArgs<'largs>

C++ method: QSizeF::boundedTo

This is an overloaded function. Available variants:

Rust arguments:
1) fn bounded_to(&self, (&::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_f::SizeF>
2) fn bounded_to(&self, (&::size_f::SizeF, cpp_utils::AsStruct)) -> ::size_f::SizeF
C++ method: QSizeF QSizeF::boundedTo(const QSizeF& arg1) const

C++ documentation:

Returns a size holding the minimum width and height of this size and the given otherSize.

See also expandedTo() and scale().

fn expanded_to<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeFExpandedToArgs<'largs>

C++ method: QSizeF::expandedTo

This is an overloaded function. Available variants:

Rust arguments:
1) fn expanded_to(&self, (&::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_f::SizeF>
2) fn expanded_to(&self, (&::size_f::SizeF, cpp_utils::AsStruct)) -> ::size_f::SizeF
C++ method: QSizeF QSizeF::expandedTo(const QSizeF& arg1) const

C++ documentation:

Returns a size holding the maximum width and height of this size and the given otherSize.

See also boundedTo() and scale().

fn height(&self) -> c_double

C++ method: double QSizeF::height() const

C++ documentation:

Returns the height.

See also width() and setHeight().

fn is_empty(&self) -> bool

C++ method: bool QSizeF::isEmpty() const

C++ documentation:

Returns true if either of the width and height is less than or equal to 0; otherwise returns false.

See also isNull() and isValid().

fn is_null(&self) -> bool

C++ method: bool QSizeF::isNull() const

C++ documentation:

Returns true if both the width and height are 0.0 (ignoring the sign); otherwise returns false.

See also isValid() and isEmpty().

fn is_valid(&self) -> bool

C++ method: bool QSizeF::isValid() const

C++ documentation:

Returns true if both the width and height is equal to or greater than 0; otherwise returns false.

See also isNull() and isEmpty().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SizeFNewArgs

C++ method: QSizeF::QSizeF

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::size_f::SizeF
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::size_f::SizeF>
C++ method: [constructor] void QSizeF::QSizeF()

Variant 2

Rust arguments:
1) fn new((&::size::Size, cpp_utils::AsStruct)) -> ::size_f::SizeF
2) fn new((&::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_f::SizeF>
C++ method: [constructor] void QSizeF::QSizeF(const QSize& sz)

Variant 3

Rust arguments:
1) fn new((libc::c_double, libc::c_double, cpp_utils::AsStruct)) -> ::size_f::SizeF
2) fn new((libc::c_double, libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_f::SizeF>
C++ method: [constructor] void QSizeF::QSizeF(double w, double h)

fn op_add_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 SizeF) -> &'l0 mut SizeF

C++ method: QSizeF& QSizeF::operator+=(const QSizeF& arg1)

C++ documentation:

Adds the given size to this size and returns a reference to this size. For example:

QSizeF s( 3, 7);
QSizeF r(-1, 4);
s += r;

// s becomes (2,11)

fn op_div_assign<'l0>(&'l0 mut self, c: c_double) -> &'l0 mut SizeF

C++ method: QSizeF& QSizeF::operator/=(double c)

C++ documentation:

This is an overloaded function.

Divides both the width and height by the given divisor and returns a reference to the size.

See also scale().

fn op_mul_assign<'l0>(&'l0 mut self, c: c_double) -> &'l0 mut SizeF

C++ method: QSizeF& QSizeF::operator*=(double c)

C++ documentation:

This is an overloaded function.

Multiplies both the width and height by the given factor and returns a reference to the size.

See also scale().

fn op_sub_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 SizeF) -> &'l0 mut SizeF

C++ method: QSizeF& QSizeF::operator-=(const QSizeF& arg1)

C++ documentation:

Subtracts the given size from this size and returns a reference to this size. For example:

QSizeF s( 3, 7);
QSizeF r(-1, 4);
s -= r;

// s becomes (4,3)

fn rheight<'l0>(&'l0 mut self) -> &'l0 mut c_double

C++ method: double& QSizeF::rheight()

C++ documentation:

Returns a reference to the height.

Using a reference makes it possible to manipulate the height directly. For example:

QSizeF size(100, 10.2);
size.rheight() += 5.5;

// size becomes (100,15.7)
See also rwidth() and setHeight().

fn rwidth<'l0>(&'l0 mut self) -> &'l0 mut c_double

C++ method: double& QSizeF::rwidth()

C++ documentation:

Returns a reference to the width.

Using a reference makes it possible to manipulate the width directly. For example:

QSizeF size(100.3, 10);
size.rwidth() += 20.5;

// size becomes (120.8,10)
See also rheight() and setWidth().

fn scale<'largs, Args>(&'largs mut self, args: Args) where Args: SizeFScaleArgs<'largs>

C++ method: QSizeF::scale

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn scale(&mut self, (&::size_f::SizeF, ::qt::AspectRatioMode)) -> ()
C++ method: void QSizeF::scale(const QSizeF& s, Qt::AspectRatioMode mode)

C++ documentation:

This is an overloaded function.

Scales the size to a rectangle with the given size, according to the specified mode.

Variant 2

Rust arguments: fn scale(&mut self, (libc::c_double, libc::c_double, ::qt::AspectRatioMode)) -> ()
C++ method: void QSizeF::scale(double w, double h, Qt::AspectRatioMode mode)

C++ documentation:

Scales the size to a rectangle with the given width and height, according to the specified mode.

If mode is Qt::IgnoreAspectRatio, the size is set to (width, height).

If mode is Qt::KeepAspectRatio, the current size is scaled to a rectangle as large as possible inside (width, height), preserving the aspect ratio.

If mode is Qt::KeepAspectRatioByExpanding, the current size is scaled to a rectangle as small as possible outside (width, height), preserving the aspect ratio.

Example:

QSizeF t1(10, 12);
t1.scale(60, 60, Qt::IgnoreAspectRatio);
// t1 is (60, 60)

QSizeF t2(10, 12); t2.scale(60, 60, Qt::KeepAspectRatio); // t2 is (50, 60)

QSizeF t3(10, 12); t3.scale(60, 60, Qt::KeepAspectRatioByExpanding); // t3 is (60, 72)
See also setWidth(), setHeight(), and scaled().

fn scaled<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeFScaledArgs<'largs>

C++ method: QSizeF::scaled

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn scaled(&self, (&::size_f::SizeF, ::qt::AspectRatioMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_f::SizeF>
2) fn scaled(&self, (&::size_f::SizeF, ::qt::AspectRatioMode, cpp_utils::AsStruct)) -> ::size_f::SizeF
C++ method: QSizeF QSizeF::scaled(const QSizeF& s, Qt::AspectRatioMode mode) const

C++ documentation:

This is an overloaded function.

Returns a size scaled to a rectangle with the given size s, according to the specified mode.

This function was introduced in Qt 5.0.

Variant 2

Rust arguments:
1) fn scaled(&self, (libc::c_double, libc::c_double, ::qt::AspectRatioMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::size_f::SizeF>
2) fn scaled(&self, (libc::c_double, libc::c_double, ::qt::AspectRatioMode, cpp_utils::AsStruct)) -> ::size_f::SizeF
C++ method: QSizeF QSizeF::scaled(double w, double h, Qt::AspectRatioMode mode) const

C++ documentation:

Returns a size scaled to a rectangle with the given width and height, according to the specified mode.

This function was introduced in Qt 5.0.

See also scale().

fn set_height(&mut self, h: c_double)

C++ method: void QSizeF::setHeight(double h)

C++ documentation:

Sets the height to the given height.

See also height(), rheight(), and setWidth().

fn set_width(&mut self, w: c_double)

C++ method: void QSizeF::setWidth(double w)

C++ documentation:

Sets the width to the given width.

See also width(), rwidth(), and setHeight().

fn to_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeFToSizeArgs<'largs>

C++ method: QSizeF::toSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size::Size>
2) fn to_size(&self, cpp_utils::AsStruct) -> ::size::Size
C++ method: QSize QSizeF::toSize() const

C++ documentation:

Returns an integer based copy of this size.

Note that the coordinates in the returned size will be rounded to the nearest integer.

See also QSizeF().

fn transpose(&mut self)

C++ method: void QSizeF::transpose()

C++ documentation:

Swaps the width and height values.

See also setWidth(), setHeight(), and transposed().

fn transposed<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SizeFTransposedArgs<'largs>

C++ method: QSizeF::transposed

This is an overloaded function. Available variants:

Rust arguments:
1) fn transposed(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size_f::SizeF>
2) fn transposed(&self, cpp_utils::AsStruct) -> ::size_f::SizeF
C++ method: QSizeF QSizeF::transposed() const

C++ documentation:

Returns the size with width and height values swapped.

This function was introduced in Qt 5.0.

See also transpose().

fn width(&self) -> c_double

C++ method: double QSizeF::width() const

C++ documentation:

Returns the width.

See also height() and setWidth().