Module qt_core::socket_notifier::overloading

Traits

SocketNotifierConnectArgs
This trait represents a set of arguments accepted by SocketNotifier::connect method.

SocketNotifierConnectStaticArgs
This trait represents a set of arguments accepted by SocketNotifier::connect_static method.

SocketNotifierDestroyedArgs
This trait represents a set of arguments accepted by SocketNotifier::destroyed method.

SocketNotifierDisconnectArgs
This trait represents a set of arguments accepted by SocketNotifier::disconnect method.

SocketNotifierDisconnectStaticArgs
This trait represents a set of arguments accepted by SocketNotifier::disconnect_static method.

SocketNotifierDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by SocketNotifier::dynamic_property_names method.

SocketNotifierNewArgs
This trait represents a set of arguments accepted by SocketNotifier::new method.

SocketNotifierObjectNameArgs
This trait represents a set of arguments accepted by SocketNotifier::object_name method.

SocketNotifierPropertyArgs
This trait represents a set of arguments accepted by SocketNotifier::property method.

SocketNotifierStartTimerArgs
This trait represents a set of arguments accepted by SocketNotifier::start_timer method.

SocketNotifierTrArgs
This trait represents a set of arguments accepted by SocketNotifier::tr method.

SocketNotifierTrUtf8Args
This trait represents a set of arguments accepted by SocketNotifier::tr_utf8 method.