Enum qt_core::standard_paths::StandardLocation

pub enum StandardLocation {
    Desktop,
    Documents,
    Fonts,
    Applications,
    Music,
    Movies,
    Pictures,
    Temp,
    Home,
    Data,
    Cache,
    GenericData,
    Runtime,
    Config,
    Download,
    GenericCache,
    GenericConfig,
    AppData,
    AppConfig,
}

C++ type: QStandardPaths::StandardLocation

C++ documentation:

This enum describes the different locations that can be queried using methods such as QStandardPaths::writableLocation, QStandardPaths::standardLocations, and QStandardPaths::displayName.

Some of the values in this enum represent a user configuration. Such enum values will return the same paths in different applications, so they could be used to share data with other applications. Other values are specific to this application. Each enum value in the table below describes whether it's application-specific or generic.

Application-specific directories should be assumed to be unreachable by other applications. Therefore, files placed there might not be readable by other applications, even if run by the same user. On the other hand, generic directories should be assumed to be accessible by all applications run by this user, but should still be assumed to be unreachable by applications by other users.

Data interchange with other users is out of the scope of QStandardPaths.

The following table gives examples of paths on different operating systems. The first path is the writable path (unless noted). Other, additional paths, if any, represent non-writable locations.

Path type	OS X	Windows

DesktopLocation	"~/Desktop"	"C:/Users/<USER>/Desktop"

DocumentsLocation	"~/Documents"	"C:/Users/<USER>/Documents"

FontsLocation	"/System/Library/Fonts" (not writable)	"C:/Windows/Fonts" (not writable)

ApplicationsLocation	"/Applications" (not writable)	"C:/Users/<USER>/AppData/Roaming/Microsoft/Windows/Start Menu/Programs"

MusicLocation	"~/Music"	"C:/Users/<USER>/Music"

MoviesLocation	"~/Movies"	"C:/Users/<USER>/Videos"

PicturesLocation	"~/Pictures"	"C:/Users/<USER>/Pictures"

TempLocation	randomly generated by the OS	"C:/Users/<USER>/AppData/Local/Temp"

HomeLocation	"~"	"C:/Users/<USER>"

DataLocation	"~/Library/Application Support/<APPNAME>", "/Library/Application Support/<APPNAME>". "<APPDIR>/../Resources"	"C:/Users/<USER>/AppData/Local/<APPNAME>", "C:/ProgramData/<APPNAME>", "<APPDIR>", "<APPDIR>/data"

CacheLocation	"~/Library/Caches/<APPNAME>", "/Library/Caches/<APPNAME>"	"C:/Users/<USER>/AppData/Local/<APPNAME>/cache"

GenericDataLocation	"~/Library/Application Support", "/Library/Application Support"	"C:/Users/<USER>/AppData/Local", "C:/ProgramData"

RuntimeLocation	"~/Library/Application Support"	"C:/Users/<USER>"

ConfigLocation	"~/Library/Preferences"	"C:/Users/<USER>/AppData/Local/<APPNAME>", "C:/ProgramData/<APPNAME>"

GenericConfigLocation	"~/Library/Preferences"	"C:/Users/<USER>/AppData/Local", "C:/ProgramData"

DownloadLocation	"~/Downloads"	"C:/Users/<USER>/Documents"

GenericCacheLocation	"~/Library/Caches", "/Library/Caches"	"C:/Users/<USER>/AppData/Local/cache"

AppDataLocation	"~/Library/Application Support/<APPNAME>", "/Library/Application Support/<APPNAME>". "<APPDIR>/../Resources"	"C:/Users/<USER>/AppData/Roaming/<APPNAME>", "C:/ProgramData/<APPNAME>", "<APPDIR>", "<APPDIR>/data"

AppLocalDataLocation	"~/Library/Application Support/<APPNAME>", "/Library/Application Support/<APPNAME>". "<APPDIR>/../Resources"	"C:/Users/<USER>/AppData/Local/<APPNAME>", "C:/ProgramData/<APPNAME>", "<APPDIR>", "<APPDIR>/data"

AppConfigLocation	"~/Library/Preferences/<APPNAME>"	"C:/Users/<USER>/AppData/Local/<APPNAME>", "C:/ProgramData/<APPNAME>"

Path type	Blackberry	Linux

DesktopLocation	"<APPROOT>/data"	"~/Desktop"

DocumentsLocation	"<APPROOT>/shared/documents"	"~/Documents"

FontsLocation	"/base/usr/fonts" (not writable)	"~/.fonts"

ApplicationsLocation	not supported (directory not readable)	"~/.local/share/applications", "/usr/local/share/applications", "/usr/share/applications"

MusicLocation	"<APPROOT>/shared/music"	"~/Music"

MoviesLocation	"<APPROOT>/shared/videos"	"~/Videos"

PicturesLocation	"<APPROOT>/shared/photos"	"~/Pictures"

TempLocation	"/var/tmp"	"/tmp"

HomeLocation	"<APPROOT>/data"	"~"

DataLocation	"<APPROOT>/data", "<APPROOT>/app/native/assets"	"~/.local/share/<APPNAME>", "/usr/local/share/<APPNAME>", "/usr/share/<APPNAME>"

CacheLocation	"<APPROOT>/data/Cache"	"~/.cache/<APPNAME>"

GenericDataLocation	"<APPROOT>/shared/misc"	"~/.local/share", "/usr/local/share", "/usr/share"

RuntimeLocation	"/var/tmp"	"/run/user/<USER>"

ConfigLocation	"<APPROOT>/data/Settings"	"~/.config", "/etc/xdg"

GenericConfigLocation	"<APPROOT>/data/Settings"	"~/.config", "/etc/xdg"

DownloadLocation	"<APPROOT>/shared/downloads"	"~/Downloads"

GenericCacheLocation	"<APPROOT>/data/Cache" (there is no shared cache)	"~/.cache"

AppDataLocation	"<APPROOT>/data", "<APPROOT>/app/native/assets"	"~/.local/share/<APPNAME>", "/usr/local/share/<APPNAME>", "/usr/share/<APPNAME>"

AppLocalDataLocation	"<APPROOT>/data", "<APPROOT>/app/native/assets"	"~/.local/share/<APPNAME>", "/usr/local/share/<APPNAME>", "/usr/share/<APPNAME>"

AppConfigLocation	"<APPROOT>/data/Settings"	"~/.config/<APPNAME>", "/etc/xdg/<APPNAME>"

Path type	Android	iOS

DesktopLocation	"<APPROOT>/files"	"<APPROOT>/<APPDIR>" (not writable)

DocumentsLocation	"<USER>/Documents", "<USER>/<APPNAME>/Documents"	"<APPROOT>/Documents"

FontsLocation	"/system/fonts" (not writable)	"<APPROOT>/Documents/.fonts"

ApplicationsLocation	not supported (directory not readable)	not supported

MusicLocation	"<USER>/Music", "<USER>/<APPNAME>/Music"	"<APPROOT>/Documents/Music"

MoviesLocation	"<USER>/Movies", "<USER>/<APPNAME>/Movies"	"<APPROOT>/Documents/Movies"

PicturesLocation	"<USER>/Pictures", "<USER>/<APPNAME>/Pictures"	"<APPROOT>/Documents/Pictures", "assets-library://"

TempLocation	"<APPROOT>/cache"	"<APPROOT>/tmp"

HomeLocation	"<APPROOT>/files"	"<APPROOT>/<APPDIR>" (not writable)

DataLocation	"<APPROOT>/files", "<USER>/<APPNAME>/files"	"<APPROOT>/Library/Application Support"

CacheLocation	"<APPROOT>/cache", "<USER>/<APPNAME>/cache"	"<APPROOT>/Library/Caches"

GenericDataLocation	"<USER>"	"<APPROOT>/Documents"

RuntimeLocation	"<APPROOT>/cache"	not supported

ConfigLocation	"<APPROOT>/files/settings"	"<APPROOT>/Documents"

GenericConfigLocation	"<APPROOT>/files/settings" (there is no shared settings)	"<APPROOT>/Documents"

DownloadLocation	"<USER>/Downloads", "<USER>/<APPNAME>/Downloads"	"<APPROOT>/Documents/Download"

GenericCacheLocation	"<APPROOT>/cache" (there is no shared cache)	"<APPROOT>/Library/Caches"

AppDataLocation	"<APPROOT>/files", "<USER>/<APPNAME>/files"	"<APPROOT>/Library/Application Support"

AppConfigLocation	"<APPROOT>/files/settings"	"<APPROOT>/Documents"

AppLocalDataLocation	"<APPROOT>/files", "<USER>/<APPNAME>/files"	"<APPROOT>/Library/Application Support"

In the table above, <APPNAME> is usually the organization name, the application name, or both, or a unique name generated at packaging. Similarly, <APPROOT> is the location where this application is installed (often a sandbox). <APPDIR> is the directory containing the application executable.

The paths above should not be relied upon, as they may change according to OS configuration, locale, or they may change in future Qt versions.

Note: On Android, applications with open files on the external storage (<USER> locations), will be killed if the external storage is unmounted.

Note: On iOS, if you do pass QStandardPaths::standardLocations(QStandardPaths::PicturesLocation).last() as argument to QFileDialog::setDirectory(), a native image picker dialog will be used for accessing the user's photo album. The filename returned can be loaded using QFile and related APIs. This feature was added in Qt 5.5.

See also writableLocation(), standardLocations(), displayName(), locate(), and locateAll().

Variants
Desktop

Returns the user's desktop directory. This is a generic value. On systems with no concept of a desktop. (C++ enum variant: DesktopLocation = 0)

Documents

Returns the directory containing user document files. This is a generic value. The returned path is never empty. (C++ enum variant: DocumentsLocation = 1)

Fonts

Returns the directory containing user's fonts. This is a generic value. Note that installing fonts may require additional, platform-specific operations. (C++ enum variant: FontsLocation = 2)

Applications

Returns the directory containing the user applications (either executables, application bundles, or shortcuts to them). This is a generic value. Note that installing applications may require additional, platform-specific operations. Files, folders or shortcuts in this directory are platform-specific. (C++ enum variant: ApplicationsLocation = 3)

Music

Returns the directory containing the user's music or other audio files. This is a generic value. If no directory specific for music files exists, a sensible fallback for storing user documents is returned. (C++ enum variant: MusicLocation = 4)

Movies

Returns the directory containing the user's movies and videos. This is a generic value. If no directory specific for movie files exists, a sensible fallback for storing user documents is returned. (C++ enum variant: MoviesLocation = 5)

Pictures

Returns the directory containing the user's pictures or photos. This is a generic value. If no directory specific for picture files exists, a sensible fallback for storing user documents is returned. (C++ enum variant: PicturesLocation = 6)

Temp

Returns a directory where temporary files can be stored. The returned value might be application-specific, shared among other applications for this user, or even system-wide. The returned path is never empty. (C++ enum variant: TempLocation = 7)

Home

Returns the user's home directory (the same as QDir::homePath()). On Unix systems, this is equal to the HOME environment variable. This value might be generic or application-specific, but the returned path is never empty. (C++ enum variant: HomeLocation = 8)

Data

This variant corresponds to multiple C++ enum variants with the same value:

DataLocation = 9: Returns the same value as AppLocalDataLocation. This enumeration value is deprecated. Using AppDataLocation is preferable since on Windows, the roaming path is recommended.

AppLocalDataLocation = 9: Returns the local settings path on the Windows operating system. On all other platforms, it returns the same value as AppDataLocation. This enum value was added in Qt 5.4.

Cache

Returns a directory location where user-specific non-essential (cached) data should be written. This is an application-specific directory. The returned path is never empty. (C++ enum variant: CacheLocation = 10)

GenericData

Returns a directory location where persistent data shared across applications can be stored. This is a generic value. The returned path is never empty. (C++ enum variant: GenericDataLocation = 11)

Runtime

Returns a directory location where runtime communication files should be written, like Unix local sockets. This is a generic value. The returned path may be empty on some systems. (C++ enum variant: RuntimeLocation = 12)

Config

Returns a directory location where user-specific configuration files should be written. This may be either a generic value or application-specific, and the returned path is never empty. (C++ enum variant: ConfigLocation = 13)

Download

Returns a directory for user's downloaded files. This is a generic value. If no directory specific for downloads exists, a sensible fallback for storing user documents is returned. (C++ enum variant: DownloadLocation = 14)

GenericCache

Returns a directory location where user-specific non-essential (cached) data, shared across applications, should be written. This is a generic value. Note that the returned path may be empty if the system has no concept of shared cache. (C++ enum variant: GenericCacheLocation = 15)

GenericConfig

Returns a directory location where user-specific configuration files shared between multiple applications should be written. This is a generic value and the returned path is never empty. (C++ enum variant: GenericConfigLocation = 16)

AppData

Returns a directory location where persistent application data can be stored. This is an application-specific directory. To obtain a path to store data to be shared with other applications, use QStandardPaths::GenericDataLocation. The returned path is never empty. On the Windows operating system, this returns the roaming path. This enum value was added in Qt 5.4. (C++ enum variant: AppDataLocation = 17)

AppConfig

Returns a directory location where user-specific configuration files should be written. This is an application-specific directory, and the returned path is never empty. This enum value was added in Qt 5.5. (C++ enum variant: AppConfigLocation = 18)