Struct qt_core::standard_paths::StandardPaths

pub struct StandardPaths {
    // some fields omitted
}

C++ type: QStandardPaths

C++ documentation:

The QStandardPaths class provides methods for accessing standard paths.

This class contains functions to query standard locations on the local filesystem, for common tasks such as user-specific directories or system-wide configuration directories.

Methods

impl StandardPaths

fn display_name<Args>(args: Args) -> Args::ReturnType where Args: StandardPathsDisplayNameArgs

C++ method: QStandardPaths::displayName

This is an overloaded function. Available variants:

Rust arguments:
1) fn display_name((::standard_paths::StandardLocation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn display_name((::standard_paths::StandardLocation, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QStandardPaths::displayName(QStandardPaths::StandardLocation type)

C++ documentation:

Returns a localized display name for the given location type or an empty QString if no relevant location can be found.

fn enable_test_mode(test_mode: bool)

C++ method: static void QStandardPaths::enableTestMode(bool testMode)

fn find_executable<Args>(args: Args) -> Args::ReturnType where Args: StandardPathsFindExecutableArgs

C++ method: QStandardPaths::findExecutable

This is an overloaded function. Available variants:

Rust arguments:
1) fn find_executable((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn find_executable((&::string::String, &::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn find_executable((&::string::String, cpp_utils::AsStruct)) -> ::string::String
4) fn find_executable((&::string::String, &::string_list::StringList, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QStandardPaths::findExecutable(const QString& executableName, const QStringList& paths = ?)

C++ documentation:

Finds the executable named executableName in the paths specified by paths, or the system paths if paths is empty.

On most operating systems the system path is determined by the PATH environment variable.

The directories where to search for the executable can be set in the paths argument. To search in both your own paths and the system paths, call findExecutable twice, once with paths set and once with paths empty.

Symlinks are not resolved, in order to preserve behavior for the case of executables whose behavior depends on the name they are invoked with.

Note: On Windows, the usual executable extensions (from the PATHEXT environment variable) are automatically appended, so that for instance findExecutable("foo") will find foo.exe or foo.bat if present.

Returns the absolute file path to the executable, or an empty string if not found.

fn is_test_mode_enabled() -> bool

C++ method: static bool QStandardPaths::isTestModeEnabled()

fn locate<Args>(args: Args) -> Args::ReturnType where Args: StandardPathsLocateArgs

C++ method: QStandardPaths::locate

This is an overloaded function. Available variants:

Rust arguments:
1) fn locate((::standard_paths::StandardLocation, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn locate((::standard_paths::StandardLocation, &::string::String, ::flags::Flags<::standard_paths::LocateOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn locate((::standard_paths::StandardLocation, &::string::String, cpp_utils::AsStruct)) -> ::string::String
4) fn locate((::standard_paths::StandardLocation, &::string::String, ::flags::Flags<::standard_paths::LocateOption>, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QStandardPaths::locate(QStandardPaths::StandardLocation type, const QString& fileName, QFlags<QStandardPaths::LocateOption> options = ?)

C++ documentation:

Tries to find a file or directory called fileName in the standard locations for type.

The full path to the first file or directory (depending on options) found is returned. If no such file or directory can be found, an empty string is returned.

fn locate_all<Args>(args: Args) -> Args::ReturnType where Args: StandardPathsLocateAllArgs

C++ method: QStandardPaths::locateAll

This is an overloaded function. Available variants:

Rust arguments:
1) fn locate_all((::standard_paths::StandardLocation, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn locate_all((::standard_paths::StandardLocation, &::string::String, ::flags::Flags<::standard_paths::LocateOption>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
3) fn locate_all((::standard_paths::StandardLocation, &::string::String, cpp_utils::AsStruct)) -> ::string_list::StringList
4) fn locate_all((::standard_paths::StandardLocation, &::string::String, ::flags::Flags<::standard_paths::LocateOption>, cpp_utils::AsStruct)) -> ::string_list::StringList
C++ method: static QStringList QStandardPaths::locateAll(QStandardPaths::StandardLocation type, const QString& fileName, QFlags<QStandardPaths::LocateOption> options = ?)

C++ documentation:

Tries to find all files or directories called fileName in the standard locations for type.

The options flag allows to specify whether to look for files or directories.

Returns the list of all the files that were found.

fn set_test_mode_enabled(test_mode: bool)

C++ method: static void QStandardPaths::setTestModeEnabled(bool testMode)

C++ documentation:

If testMode is true, this enables a special "test mode" in QStandardPaths, which changes writable locations to point to test directories, in order to prevent auto tests from reading from or writing to the current user's configuration.

This affects the locations into which test programs might write files: GenericDataLocation, DataLocation, ConfigLocation, GenericConfigLocation, AppConfigLocation, GenericCacheLocation, CacheLocation. Other locations are not affected.

On Unix, XDG_DATA_HOME is set to ~/.qttest/share, XDG_CONFIG_HOME is set to ~/.qttest/config, and XDG_CACHE_HOME is set to ~/.qttest/cache.

On OS X, data goes to ~/.qttest/Application Support, cache goes to ~/.qttest/Cache, and config goes to ~/.qttest/Preferences.

On Windows, everything goes to a "qttest" directory under Application Data.

fn standard_locations<Args>(args: Args) -> Args::ReturnType where Args: StandardPathsStandardLocationsArgs

C++ method: QStandardPaths::standardLocations

This is an overloaded function. Available variants:

Rust arguments:
1) fn standard_locations((::standard_paths::StandardLocation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn standard_locations((::standard_paths::StandardLocation, cpp_utils::AsStruct)) -> ::string_list::StringList
C++ method: static QStringList QStandardPaths::standardLocations(QStandardPaths::StandardLocation type)

C++ documentation:

Returns all the directories where files of type belong.

The list of directories is sorted from high to low priority, starting with writableLocation() if it can be determined. This list is empty if no locations for type are defined.

See also writableLocation().

fn writable_location<Args>(args: Args) -> Args::ReturnType where Args: StandardPathsWritableLocationArgs

C++ method: QStandardPaths::writableLocation

This is an overloaded function. Available variants:

Rust arguments:
1) fn writable_location((::standard_paths::StandardLocation, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn writable_location((::standard_paths::StandardLocation, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QStandardPaths::writableLocation(QStandardPaths::StandardLocation type)

C++ documentation:

Returns the directory where files of type should be written to, or an empty string if the location cannot be determined.

Note: The storage location returned can be a directory that does not exist; i.e., it may need to be created by the system or the user.