Module qt_core::state::overloading

Traits

StateAddTransitionArgs
This trait represents a set of arguments accepted by State::add_transition method.

StateConnectArgs
This trait represents a set of arguments accepted by State::connect method.

StateConnectStaticArgs
This trait represents a set of arguments accepted by State::connect_static method.

StateDestroyedArgs
This trait represents a set of arguments accepted by State::destroyed method.

StateDisconnectArgs
This trait represents a set of arguments accepted by State::disconnect method.

StateDisconnectStaticArgs
This trait represents a set of arguments accepted by State::disconnect_static method.

StateDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by State::dynamic_property_names method.

StateNewArgs
This trait represents a set of arguments accepted by State::new method.

StateObjectNameArgs
This trait represents a set of arguments accepted by State::object_name method.

StatePropertyArgs
This trait represents a set of arguments accepted by State::property method.

StateStartTimerArgs
This trait represents a set of arguments accepted by State::start_timer method.

StateTrArgs
This trait represents a set of arguments accepted by State::tr method.

StateTrUtf8Args
This trait represents a set of arguments accepted by State::tr_utf8 method.

StateTransitionsArgs
This trait represents a set of arguments accepted by State::transitions method.