Enum qt_core::state_machine::Error

pub enum Error {
    NoError,
    NoInitialStateError,
    NoDefaultStateInHistoryStateError,
    NoCommonAncestorForTransitionError,
}

C++ type: QStateMachine::Error

C++ documentation:

This enum type defines errors that can occur in the state machine at run time. When the state machine encounters an unrecoverable error at run time, it will set the error code returned by error(), the error message returned by errorString(), and enter an error state based on the context of the error.

See also setErrorState().

Variants
NoError

No error has occurred. (C++ enum variant: NoError = 0)

NoInitialStateError

The machine has entered a QState with children which does not have an initial state set. The context of this error is the state which is missing an initial state. (C++ enum variant: NoInitialStateError = 1)

NoDefaultStateInHistoryStateError

The machine has entered a QHistoryState which does not have a default state set. The context of this error is the QHistoryState which is missing a default state. (C++ enum variant: NoDefaultStateInHistoryStateError = 2)

NoCommonAncestorForTransitionError

The machine has selected a transition whose source and targets are not part of the same tree of states, and thus are not part of the same state machine. Commonly, this could mean that one of the states has not been given any parent or added to any machine. The context of this error is the source state of the transition. (C++ enum variant: NoCommonAncestorForTransitionError = 3)