Enum qt_core::state_machine::EventPriority

pub enum EventPriority {
    Normal,
    High,
}

C++ type: QStateMachine::EventPriority

C++ documentation:

This enum type specifies the priority of an event posted to the state machine using postEvent().

Events of high priority are processed before events of normal priority.

Variants
Normal

The event has normal priority. (C++ enum variant: NormalPriority = 0)

High

The event has high priority. (C++ enum variant: HighPriority = 1)