Struct qt_core::state_machine::SignalEvent

pub struct SignalEvent {
    // some fields omitted
}

C++ type: QStateMachine::SignalEvent

C++ documentation:

The SignalEvent class represents a Qt signal event.

A signal event is generated by a QStateMachine in response to a Qt signal. The QSignalTransition class provides a transition associated with a signal event. QStateMachine::SignalEvent is part of The State Machine Framework.

The sender() function returns the object that generated the signal. The signalIndex() function returns the index of the signal. The arguments() function returns the arguments of the signal.

Methods

impl SignalEvent

fn accept(&mut self)

C++ method: void QStateMachine::SignalEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn arguments<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SignalEventArgumentsArgs<'largs>

C++ method: QStateMachine::SignalEvent::arguments

This is an overloaded function. Available variants:

Rust arguments:
1) fn arguments(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListVariantRef>
2) fn arguments(&self, cpp_utils::AsStruct) -> ::list::ListVariantRef
C++ method: QList<QVariant> QStateMachine::SignalEvent::arguments() const

C++ documentation:

Returns the arguments of the signal.

fn ignore(&mut self)

C++ method: void QStateMachine::SignalEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QStateMachine::SignalEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn new<Args>(args: Args) -> Args::ReturnType where Args: SignalEventNewArgs

C++ method: QStateMachine::SignalEvent::SignalEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::object::Object, libc::c_int, &::list::ListVariantRef, cpp_utils::AsStruct)) -> ::state_machine::SignalEvent
2) fn new((*mut ::object::Object, libc::c_int, &::list::ListVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::state_machine::SignalEvent>
C++ method: [constructor] void QStateMachine::SignalEvent::SignalEvent(QObject* sender, int signalIndex, const QList<QVariant>& arguments)

fn register_event_type<Args>(args: Args) -> c_int where Args: SignalEventRegisterEventTypeArgs

C++ method: QStateMachine::SignalEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QStateMachine::SignalEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn sender(&self) -> *mut Object

C++ method: QObject* QStateMachine::SignalEvent::sender() const

C++ documentation:

Returns the object that emitted the signal.

See also QObject::sender().

fn set_accepted(&mut self, accepted: bool)

C++ method: void QStateMachine::SignalEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn signal_index(&self) -> c_int

C++ method: int QStateMachine::SignalEvent::signalIndex() const

C++ documentation:

Returns the index of the signal.

See also QMetaObject::indexOfSignal() and QMetaObject::method().

fn spontaneous(&self) -> bool

C++ method: bool QStateMachine::SignalEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QStateMachine::SignalEvent::type() const

C++ documentation:

Returns the event type.