Struct qt_core::state_machine::WrappedEvent

pub struct WrappedEvent {
    // some fields omitted
}

C++ type: QStateMachine::WrappedEvent

C++ documentation:

The WrappedEvent class inherits QEvent and holds a clone of an event associated with a QObject.

A wrapped event is generated by a QStateMachine in response to a Qt event. The QEventTransition class provides a transition associated with a such an event. QStateMachine::WrappedEvent is part of The State Machine Framework.

The object() function returns the object that generated the event. The event() function returns a clone of the original event.

Methods

impl WrappedEvent

fn accept(&mut self)

C++ method: void QStateMachine::WrappedEvent::accept()

C++ documentation:

Sets the accept flag of the event object, the equivalent of calling setAccepted(true).

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget.

See also ignore().

fn event(&self) -> *mut Event

C++ method: QEvent* QStateMachine::WrappedEvent::event() const

C++ documentation:

Returns a clone of the original event.

fn ignore(&mut self)

C++ method: void QStateMachine::WrappedEvent::ignore()

C++ documentation:

Clears the accept flag parameter of the event object, the equivalent of calling setAccepted(false).

Clearing the accept parameter indicates that the event receiver does not want the event. Unwanted events might be propagated to the parent widget.

See also accept().

fn is_accepted(&self) -> bool

C++ method: bool QStateMachine::WrappedEvent::isAccepted() const

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn new<Args>(args: Args) -> Args::ReturnType where Args: WrappedEventNewArgs

C++ method: QStateMachine::WrappedEvent::WrappedEvent

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((*mut ::object::Object, *mut ::event::Event, cpp_utils::AsStruct)) -> ::state_machine::WrappedEvent
2) fn new((*mut ::object::Object, *mut ::event::Event, cpp_utils::AsBox)) -> cpp_utils::CppBox<::state_machine::WrappedEvent>
C++ method: [constructor] void QStateMachine::WrappedEvent::WrappedEvent(QObject* object, QEvent* event)

fn object(&self) -> *mut Object

C++ method: QObject* QStateMachine::WrappedEvent::object() const

C++ documentation:

Returns the object that the event is associated with.

fn register_event_type<Args>(args: Args) -> c_int where Args: WrappedEventRegisterEventTypeArgs

C++ method: QStateMachine::WrappedEvent::registerEventType

This is an overloaded function. Available variants:

Rust arguments:
1) fn register_event_type(libc::c_int) -> libc::c_int
2) fn register_event_type(()) -> libc::c_int
C++ method: static int QStateMachine::WrappedEvent::registerEventType(int hint = ?)

C++ documentation:

Registers and returns a custom event type. The hint provided will be used if it is available, otherwise it will return a value between QEvent::User and QEvent::MaxUser that has not yet been registered. The hint is ignored if its value is not between QEvent::User and QEvent::MaxUser.

Returns -1 if all available values are already taken or the program is shutting down.

Note: This function is thread-safe.

This function was introduced in Qt 4.4.

fn set_accepted(&mut self, accepted: bool)

C++ method: void QStateMachine::WrappedEvent::setAccepted(bool accepted)

C++ documentation:

the accept flag of the event object

Setting the accept parameter indicates that the event receiver wants the event. Unwanted events might be propagated to the parent widget. By default, isAccepted() is set to true, but don't rely on this as subclasses may choose to clear it in their constructor.

For convenience, the accept flag can also be set with accept(), and cleared with ignore().

Access functions:

bool 	isAccepted() const

void 	setAccepted(bool accepted)

fn spontaneous(&self) -> bool

C++ method: bool QStateMachine::WrappedEvent::spontaneous() const

C++ documentation:

Returns true if the event originated outside the application (a system event); otherwise returns false.

The return value of this function is not defined for paint events.

fn type_(&self) -> Type

C++ method: QEvent::Type QStateMachine::WrappedEvent::type() const

C++ documentation:

Returns the event type.