Struct qt_core::static_plugin::StaticPlugin

pub struct StaticPlugin {
    // some fields omitted
}

C++ type: QStaticPlugin

C++ documentation:

QStaticPlugin is a struct containing a reference to a static plugin instance together with its meta data.

Methods

impl StaticPlugin

fn instance(&self) -> extern fn() -> *mut Object

C++ method: QObject* (*FN_PTR)() QStaticPlugin::instance() const

fn meta_data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StaticPluginMetaDataArgs<'largs>

C++ method: QStaticPlugin::metaData

This is an overloaded function. Available variants:

Rust arguments:
1) fn meta_data(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::JsonObject>
2) fn meta_data(&self, cpp_utils::AsStruct) -> ::json_object::JsonObject
C++ method: QJsonObject QStaticPlugin::metaData() const

C++ documentation:

Returns a the meta data for the plugin as a QJsonObject.

See also rawMetaData().

fn raw_meta_data(&self) -> extern fn() -> *const c_char

C++ method: const char* (*FN_PTR)() QStaticPlugin::rawMetaData() const

fn set_instance(&mut self, value: extern fn() -> *mut Object)

C++ method: void QStaticPlugin::set_instance(QObject* (*FN_PTR)() value)

fn set_raw_meta_data(&mut self, value: extern fn() -> *const c_char)

C++ method: void QStaticPlugin::set_rawMetaData(const char* (*FN_PTR)() value)