Struct qt_core::storage_info::StorageInfo

pub struct StorageInfo {
    // some fields omitted
}

C++ type: QStorageInfo

C++ documentation:

Provides information about currently mounted storage and drives.

Allows retrieving information about the volume's space, its mount point, label, and filesystem name.

You can create an instance of QStorageInfo by passing the path to the volume's mount point as a constructor parameter, or you can set it using the setPath() method. The static mountedVolumes() method can be used to get the list of all mounted filesystems.

QStorageInfo always caches the retrieved information, but you can call refresh() to invalidate the cache.

The following example retrieves the most common information about the root volume of the system, and prints information about it.

    QStorageInfo storage = QStorageInfo::root();

qDebug() << storage.rootPath(); if (storage.isReadOnly()) qDebug() << "isReadOnly:" << storage.isReadOnly();

qDebug() << "name:" << storage.name(); qDebug() << "fileSystemType:" << storage.fileSystemType(); qDebug() << "size:" << storage.bytesTotal()/1000/1000 << "MB"; qDebug() << "availableSize:" << storage.bytesAvailable()/1000/1000 << "MB";

Methods

impl StorageInfo

fn bytes_available(&self) -> i64

C++ method: qint64 QStorageInfo::bytesAvailable() const

C++ documentation:

Returns the size (in bytes) available for the current user. It returns the total size available if the user is the root user or a system administrator.

This size can be less than or equal to the free size returned by bytesFree() function.

Returns -1 if QStorageInfo object is not valid.

See also bytesTotal() and bytesFree().

fn bytes_free(&self) -> i64

C++ method: qint64 QStorageInfo::bytesFree() const

C++ documentation:

Returns the number of free bytes in a volume. Note that if there are quotas on the filesystem, this value can be larger than the value returned by bytesAvailable().

Returns -1 if QStorageInfo object is not valid.

See also bytesTotal() and bytesAvailable().

fn bytes_total(&self) -> i64

C++ method: qint64 QStorageInfo::bytesTotal() const

C++ documentation:

Returns the total volume size in bytes.

Returns -1 if QStorageInfo object is not valid.

See also bytesFree() and bytesAvailable().

fn device<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StorageInfoDeviceArgs<'largs>

C++ method: QStorageInfo::device

This is an overloaded function. Available variants:

Rust arguments:
1) fn device(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn device(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QStorageInfo::device() const

C++ documentation:

Returns the device for this volume.

For example, on Unix filesystems (including OS X), this returns the devpath like /dev/sda0 for local storages. On Windows, it returns the UNC path starting with \\?\ for local storages (in other words, the volume GUID).

See also rootPath().

fn display_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StorageInfoDisplayNameArgs<'largs>

C++ method: QStorageInfo::displayName

This is an overloaded function. Available variants:

Rust arguments:
1) fn display_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn display_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QStorageInfo::displayName() const

C++ documentation:

Returns the volume's name, if available, or the root path if not.

fn file_system_type<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StorageInfoFileSystemTypeArgs<'largs>

C++ method: QStorageInfo::fileSystemType

This is an overloaded function. Available variants:

Rust arguments:
1) fn file_system_type(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn file_system_type(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QStorageInfo::fileSystemType() const

C++ documentation:

Returns the type name of the filesystem.

This is a platform-dependent function, and filesystem names can vary between different operating systems. For example, on Windows filesystems they can be named NTFS, and on Linux they can be named ntfs-3g or fuseblk.

See also name().

fn is_read_only(&self) -> bool

C++ method: bool QStorageInfo::isReadOnly() const

C++ documentation:

Returns true if the current filesystem is protected from writing; false otherwise.

fn is_ready(&self) -> bool

C++ method: bool QStorageInfo::isReady() const

C++ documentation:

Returns true if the current filesystem is ready to work; false otherwise. For example, false is returned if the CD volume is not inserted.

Note that fileSystemType(), name(), bytesTotal(), bytesFree(), and bytesAvailable() will return invalid data until the volume is ready.

See also isValid().

fn is_root(&self) -> bool

C++ method: bool QStorageInfo::isRoot() const

C++ documentation:

Returns true if this QStorageInfo represents the system root volume; false otherwise.

On Unix filesystems, the root volume is a volume mounted on /. On Windows, the root volume is the volume where the OS is installed.

See also root().

fn is_valid(&self) -> bool

C++ method: bool QStorageInfo::isValid() const

C++ documentation:

Returns true if the QStorageInfo specified by rootPath exists and is mounted correctly.

See also isReady().

fn mounted_volumes<Args>(args: Args) -> Args::ReturnType where Args: StorageInfoMountedVolumesArgs

C++ method: QStorageInfo::mountedVolumes

This is an overloaded function. Available variants:

Rust arguments:
1) fn mounted_volumes(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListStorageInfoRef>
2) fn mounted_volumes(cpp_utils::AsStruct) -> ::list::ListStorageInfoRef
C++ method: static QList<QStorageInfo> QStorageInfo::mountedVolumes()

C++ documentation:

Returns the list of QStorageInfo objects that corresponds to the list of currently mounted filesystems.

On Windows, this returns the drives visible in the My Computer folder. On Unix operating systems, it returns the list of all mounted filesystems (except for pseudo filesystems).

Returns all currently mounted filesystems by default.

The example shows how to retrieve all available filesystems, skipping read-only ones.

    foreach (const QStorageInfo &storage, QStorageInfo::mountedVolumes()) {
        if (storage.isValid() && storage.isReady()) {
            if (!storage.isReadOnly()) {
                // ...
            }
        }
    }

See also root().

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StorageInfoNameArgs<'largs>

C++ method: QStorageInfo::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QStorageInfo::name() const

C++ documentation:

Returns the human-readable name of a filesystem, usually called label.

Not all filesystems support this feature. In this case, the value returned by this method could be empty. An empty string is returned if the file system does not support labels, or if no label is set.

On Linux, retrieving the volume's label requires udev to be present in the system.

See also fileSystemType().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StorageInfoNewArgs

C++ method: QStorageInfo::QStorageInfo

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::storage_info::StorageInfo
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::storage_info::StorageInfo>
C++ method: [constructor] void QStorageInfo::QStorageInfo()

Variant 2

Rust arguments:
1) fn new((&::dir::Dir, cpp_utils::AsStruct)) -> ::storage_info::StorageInfo
2) fn new((&::dir::Dir, cpp_utils::AsBox)) -> cpp_utils::CppBox<::storage_info::StorageInfo>
C++ method: [constructor] void QStorageInfo::QStorageInfo(const QDir& dir)

Variant 3

Rust arguments:
1) fn new((&::storage_info::StorageInfo, cpp_utils::AsStruct)) -> ::storage_info::StorageInfo
2) fn new((&::storage_info::StorageInfo, cpp_utils::AsBox)) -> cpp_utils::CppBox<::storage_info::StorageInfo>
C++ method: [constructor] void QStorageInfo::QStorageInfo(const QStorageInfo& other)

Variant 4

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::storage_info::StorageInfo
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::storage_info::StorageInfo>
C++ method: [constructor] void QStorageInfo::QStorageInfo(const QString& path)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 StorageInfo) -> &'l0 mut StorageInfo

C++ method: QStorageInfo& QStorageInfo::operator=(const QStorageInfo& other)

C++ documentation:

Makes a copy of the QStorageInfo object other and assigns it to this QStorageInfo object.

fn refresh(&mut self)

C++ method: void QStorageInfo::refresh()

C++ documentation:

Resets QStorageInfo's internal cache.

QStorageInfo caches information about storage to speed up performance. QStorageInfo retrieves information during object construction and/or when calling the setPath() method. You have to manually reset the cache by calling this function to update storage information.

fn root<Args>(args: Args) -> Args::ReturnType where Args: StorageInfoRootArgs

C++ method: QStorageInfo::root

This is an overloaded function. Available variants:

Rust arguments:
1) fn root(cpp_utils::AsBox) -> cpp_utils::CppBox<::storage_info::StorageInfo>
2) fn root(cpp_utils::AsStruct) -> ::storage_info::StorageInfo
C++ method: static QStorageInfo QStorageInfo::root()

C++ documentation:

Returns a QStorageInfo object that represents the system root volume.

On Unix systems this call returns the root ('/') volume; in Windows the volume where the operating system is installed.

See also isRoot().

fn root_path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StorageInfoRootPathArgs<'largs>

C++ method: QStorageInfo::rootPath

This is an overloaded function. Available variants:

Rust arguments:
1) fn root_path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn root_path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QStorageInfo::rootPath() const

C++ documentation:

Returns the mount point of the filesystem this QStorageInfo object represents.

On Windows, it returns the volume letter in case the volume is not mounted to a directory.

Note that the value returned by rootPath() is the real mount point of a volume, and may not be equal to the value passed to the constructor or setPath() method. For example, if you have only the root volume in the system, and pass '/directory' to setPath(), then this method will return '/'.

See also setPath() and device().

fn set_path(&mut self, path: &String)

C++ method: void QStorageInfo::setPath(const QString& path)

C++ documentation:

Sets this QStorageInfo object to the filesystem mounted where path is located.

path can either be a root path of the filesystem, a directory, or a file within that filesystem.

See also rootPath().

fn swap(&mut self, other: &mut StorageInfo)

C++ method: void QStorageInfo::swap(QStorageInfo& other)

C++ documentation:

Swaps this volume info with other. This function is very fast and never fails.