Enum qt_core::string::SectionFlag

pub enum SectionFlag {
    Default,
    SkipEmpty,
    IncludeLeadingSep,
    IncludeTrailingSep,
    CaseInsensitiveSeps,
}

C++ type: QString::SectionFlag

C++ documentation:

This enum specifies flags that can be used to affect various aspects of the section() function's behavior with respect to separators and empty fields.

The SectionFlags type is a typedef for QFlags<SectionFlag>. It stores an OR combination of SectionFlag values.

See also section().

Variants
Default

Empty fields are counted, leading and trailing separators are not included, and the separator is compared case sensitively. (C++ enum variant: SectionDefault = 0)

SkipEmpty

Treat empty fields as if they don't exist, i.e. they are not considered as far as start and end are concerned. (C++ enum variant: SectionSkipEmpty = 1)

IncludeLeadingSep

Include the leading separator (if any) in the result string. (C++ enum variant: SectionIncludeLeadingSep = 2)

IncludeTrailingSep

Include the trailing separator (if any) in the result string. (C++ enum variant: SectionIncludeTrailingSep = 4)

CaseInsensitiveSeps

Compare the separator case-insensitively. (C++ enum variant: SectionCaseInsensitiveSeps = 8)