Module qt_core::string::overloading

Traits

OpAddArgs
This trait represents a set of arguments accepted by op_add method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

OpEqArgs
This trait represents a set of arguments accepted by op_eq method.

OpGeArgs
This trait represents a set of arguments accepted by op_ge method.

OpGtArgs
This trait represents a set of arguments accepted by op_gt method.

OpLeArgs
This trait represents a set of arguments accepted by op_le method.

OpLtArgs
This trait represents a set of arguments accepted by op_lt method.

OpNeqArgs
This trait represents a set of arguments accepted by op_neq method.

StringAppendArgs
This trait represents a set of arguments accepted by String::append method.

StringArgArgs
This trait represents a set of arguments accepted by String::arg method.

StringAtArgs
This trait represents a set of arguments accepted by String::at method.

StringCompareArgs
This trait represents a set of arguments accepted by String::compare method.

StringCompareStaticArgs
This trait represents a set of arguments accepted by String::compare_static method.

StringContainsArgs
This trait represents a set of arguments accepted by String::contains method.

StringCountArgs
This trait represents a set of arguments accepted by String::count method.

StringEndsWithArgs
This trait represents a set of arguments accepted by String::ends_with method.

StringFillArgs
This trait represents a set of arguments accepted by String::fill method.

StringFromLatin1Args
This trait represents a set of arguments accepted by String::from_latin1 method.

StringFromLocal8BitArgs
This trait represents a set of arguments accepted by String::from_local8_bit method.

StringFromRawDataArgs
This trait represents a set of arguments accepted by String::from_raw_data method.

StringFromUcs4Args
This trait represents a set of arguments accepted by String::from_ucs4 method.

StringFromUtf16Args
This trait represents a set of arguments accepted by String::from_utf16 method.

StringFromUtf8Args
This trait represents a set of arguments accepted by String::from_utf8 method.

StringFromWCharArrayArgs
This trait represents a set of arguments accepted by String::from_w_char_array method.

StringIndexOfArgs
This trait represents a set of arguments accepted by String::index_of method.

StringInsertArgs
This trait represents a set of arguments accepted by String::insert method.

StringLastIndexOfArgs
This trait represents a set of arguments accepted by String::last_index_of method.

StringLeftArgs
This trait represents a set of arguments accepted by String::left method.

StringLeftJustifiedArgs
This trait represents a set of arguments accepted by String::left_justified method.

StringLeftRefArgs
This trait represents a set of arguments accepted by String::left_ref method.

StringLocaleAwareCompareArgs
This trait represents a set of arguments accepted by String::locale_aware_compare method.

StringLocaleAwareCompareStaticArgs
This trait represents a set of arguments accepted by String::locale_aware_compare_static method.

StringMidArgs
This trait represents a set of arguments accepted by String::mid method.

StringMidRefArgs
This trait represents a set of arguments accepted by String::mid_ref method.

StringNewArgs
This trait represents a set of arguments accepted by String::new method.

StringNormalizedArgs
This trait represents a set of arguments accepted by String::normalized method.

StringNumberArgs
This trait represents a set of arguments accepted by String::number method.

StringOpAddAssignArgs
This trait represents a set of arguments accepted by String::op_add_assign method.

StringOpAssignArgs
This trait represents a set of arguments accepted by String::op_assign method.

StringOpEqArgs
This trait represents a set of arguments accepted by String::op_eq method.

StringOpGeArgs
This trait represents a set of arguments accepted by String::op_ge method.

StringOpGtArgs
This trait represents a set of arguments accepted by String::op_gt method.

StringOpIndexArgs
This trait represents a set of arguments accepted by String::op_index method.

StringOpIndexMutArgs
This trait represents a set of arguments accepted by String::op_index_mut method.

StringOpLeArgs
This trait represents a set of arguments accepted by String::op_le method.

StringOpLtArgs
This trait represents a set of arguments accepted by String::op_lt method.

StringOpNeqArgs
This trait represents a set of arguments accepted by String::op_neq method.

StringPrependArgs
This trait represents a set of arguments accepted by String::prepend method.

StringPushBackArgs
This trait represents a set of arguments accepted by String::push_back method.

StringPushFrontArgs
This trait represents a set of arguments accepted by String::push_front method.

StringRemoveArgs
This trait represents a set of arguments accepted by String::remove method.

StringRepeatedArgs
This trait represents a set of arguments accepted by String::repeated method.

StringReplaceArgs
This trait represents a set of arguments accepted by String::replace method.

StringRightArgs
This trait represents a set of arguments accepted by String::right method.

StringRightJustifiedArgs
This trait represents a set of arguments accepted by String::right_justified method.

StringRightRefArgs
This trait represents a set of arguments accepted by String::right_ref method.

StringSectionArgs
This trait represents a set of arguments accepted by String::section method.

StringSetNumArgs
This trait represents a set of arguments accepted by String::set_num method.

StringSimplifiedArgs
This trait represents a set of arguments accepted by String::simplified method.

StringSimplifiedMutArgs
This trait represents a set of arguments accepted by String::simplified_mut method.

StringSplitArgs
This trait represents a set of arguments accepted by String::split method.

StringSplitRefArgs
This trait represents a set of arguments accepted by String::split_ref method.

StringStartsWithArgs
This trait represents a set of arguments accepted by String::starts_with method.

StringToCaseFoldedArgs
This trait represents a set of arguments accepted by String::to_case_folded method.

StringToCaseFoldedMutArgs
This trait represents a set of arguments accepted by String::to_case_folded_mut method.

StringToDoubleArgs
This trait represents a set of arguments accepted by String::to_double method.

StringToFloatArgs
This trait represents a set of arguments accepted by String::to_float method.

StringToHtmlEscapedArgs
This trait represents a set of arguments accepted by String::to_html_escaped method.

StringToIntArgs
This trait represents a set of arguments accepted by String::to_int method.

StringToLatin1Args
This trait represents a set of arguments accepted by String::to_latin1 method.

StringToLocal8BitArgs
This trait represents a set of arguments accepted by String::to_local8_bit method.

StringToLongArgs
This trait represents a set of arguments accepted by String::to_long method.

StringToLongLongArgs
This trait represents a set of arguments accepted by String::to_long_long method.

StringToLowerArgs
This trait represents a set of arguments accepted by String::to_lower method.

StringToLowerMutArgs
This trait represents a set of arguments accepted by String::to_lower_mut method.

StringToShortArgs
This trait represents a set of arguments accepted by String::to_short method.

StringToUIntArgs
This trait represents a set of arguments accepted by String::to_u_int method.

StringToULongArgs
This trait represents a set of arguments accepted by String::to_u_long method.

StringToULongLongArgs
This trait represents a set of arguments accepted by String::to_u_long_long method.

StringToUShortArgs
This trait represents a set of arguments accepted by String::to_u_short method.

StringToUcs4Args
This trait represents a set of arguments accepted by String::to_ucs4 method.

StringToUpperArgs
This trait represents a set of arguments accepted by String::to_upper method.

StringToUpperMutArgs
This trait represents a set of arguments accepted by String::to_upper_mut method.

StringToUtf8Args
This trait represents a set of arguments accepted by String::to_utf8 method.

StringTrimmedArgs
This trait represents a set of arguments accepted by String::trimmed method.

StringTrimmedMutArgs
This trait represents a set of arguments accepted by String::trimmed_mut method.