Struct qt_core::string_list::StringList

pub struct StringList {
    // some fields omitted
}

C++ type: QStringList

C++ documentation:

The QStringList class provides a list of strings.

QStringList inherits from QList<QString>. Like QList, QStringList is implicitly shared. It provides fast index-based access as well as fast insertions and removals. Passing string lists as value parameters is both fast and safe.

All of QList's functionality also applies to QStringList. For example, you can use isEmpty() to test whether the list is empty, and you can call functions like append(), prepend(), insert(), replace(), removeAll(), removeAt(), removeFirst(), removeLast(), and removeOne() to modify a QStringList. In addition, QStringList provides a few convenience functions that make handling lists of strings easier:

Methods

impl StringList

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: StringListAppendArgs<'largs>

C++ method: QStringList::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &::list::ListStringRef) -> ()
C++ method: void QStringList::append(const QList<QString>& t)

C++ documentation:

This is an overloaded function.

Appends the items of the value list to this list.

This function was introduced in Qt 4.5.

See also operator<<() and operator+=().

Variant 2

Rust arguments: fn append(&mut self, &::string::String) -> ()
C++ method: void QStringList::append(const QString& t)

C++ documentation:

Inserts value at the end of the list.

Example:

QList<QString> list;
list.append("one");
list.append("two");
list.append("three");
// list: ["one", "two", "three"]

This is the same as list.insert(size(), value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also operator<<(), prepend(), and insert().

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 String

C++ method: const QString& QStringList::at(int i) const

C++ documentation:

Returns the item at index position i in the list. i must be a valid index position in the list (i.e., 0 <= i < size()).

This function is very fast (constant time).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStringList::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStringList::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn clear(&mut self)

C++ method: void QStringList::clear()

C++ documentation:

Removes all items from the list.

See also removeAll().

fn contains<'largs, Args>(&'largs self, args: Args) -> bool where Args: StringListContainsArgs<'largs>

C++ method: QStringList::contains

This is an overloaded function. Available variants:

Rust arguments:
1) fn contains(&self, &::string::String) -> bool
2) fn contains(&self, (&::string::String, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringList::contains(const QString& str, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns true if the list contains the string str; otherwise returns false. The search is case insensitive if cs is Qt::CaseInsensitive; the search is case sensitive by default.

See also indexOf(), lastIndexOf(), and QString::contains().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringListCountArgs<'largs>

C++ method: QStringList::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QStringList::count() const

C++ documentation:

Returns the number of items in the list. This is effectively the same as size().

Variant 2

Rust arguments: fn count(&self, &::string::String) -> libc::c_int
C++ method: int QStringList::count(const QString& t) const

C++ documentation:

Returns the number of occurrences of value in the list.

This function requires the value type to have an implementation of operator==().

See also contains() and indexOf().

fn detach(&mut self)

C++ method: void QStringList::detach()

fn detach_shared(&mut self)

C++ method: void QStringList::detachShared()

fn empty(&self) -> bool

C++ method: bool QStringList::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty() and returns true if the list is empty.

fn ends_with(&self, t: &String) -> bool

C++ method: bool QStringList::endsWith(const QString& t) const

C++ documentation:

Returns true if this list is not empty and its last item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and contains().

fn first<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStringList::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStringList::first()

C++ documentation:

Returns a reference to the first item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStringList::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStringList::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringListIndexOfArgs<'largs>

C++ method: QStringList::indexOf

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn index_of(&self, &mut ::reg_exp::RegExp) -> libc::c_int
2) fn index_of(&self, (&mut ::reg_exp::RegExp, libc::c_int)) -> libc::c_int
C++ method: int QStringList::indexOf(QRegExp& rx, int from = ?) const

C++ documentation:

This function overloads indexOf().

Returns the index position of the first exact match of rx in the list, searching forward from index position from. Returns -1 if no item matched.

By default, this function is case sensitive.

If an item matched, the rx regular expression will contain the matched objects (see QRegExp::matchedLength, QRegExp::cap).

This function was introduced in Qt 4.5.

See also lastIndexOf(), contains(), and QRegExp::exactMatch().

Variant 2

Rust arguments:
1) fn index_of(&self, &::reg_exp::RegExp) -> libc::c_int
2) fn index_of(&self, (&::reg_exp::RegExp, libc::c_int)) -> libc::c_int
C++ method: int QStringList::indexOf(const QRegExp& rx, int from = ?) const

C++ documentation:

Returns the index position of the first exact match of rx in the list, searching forward from index position from. Returns -1 if no item matched.

By default, this function is case sensitive.

See also lastIndexOf(), contains(), and QRegExp::exactMatch().

Variant 3

Rust arguments:
1) fn index_of(&self, &::regular_expression::RegularExpression) -> libc::c_int
2) fn index_of(&self, (&::regular_expression::RegularExpression, libc::c_int)) -> libc::c_int
C++ method: int QStringList::indexOf(const QRegularExpression& re, int from = ?) const

C++ documentation:

This is an overloaded function.

Returns the index position of the first match of re in the list, searching forward from index position from. Returns -1 if no item matched.

This function was introduced in Qt 5.0.

See also lastIndexOf().

Variant 4

Rust arguments:
1) fn index_of(&self, &::string::String) -> libc::c_int
2) fn index_of(&self, (&::string::String, libc::c_int)) -> libc::c_int
C++ method: int QStringList::indexOf(const QString& t, int from = ?) const

C++ documentation:

Returns the index position of the first occurrence of value in the list, searching forward from index position from. Returns -1 if no item matched.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "B" << "A";
list.indexOf("B");          // returns 1
list.indexOf("B", 1);       // returns 1
list.indexOf("B", 2);       // returns 3
list.indexOf("X");          // returns -1

This function requires the value type to have an implementation of operator==().

Note that QList uses 0-based indexes, just like C++ arrays. Negative indexes are not supported with the exception of the value mentioned above.

See also lastIndexOf() and contains().

fn insert(&mut self, i: c_int, t: &String)

C++ method: void QStringList::insert(int i, const QString& t)

C++ documentation:

Inserts value at index position i in the list. If i is 0, the value is prepended to the list. If i is size(), the value is appended to the list.

Example:

QList<QString> list;
list << "alpha" << "beta" << "delta";
list.insert(2, "gamma");
// list: ["alpha", "beta", "gamma", "delta"]

See also append(), prepend(), replace(), and removeAt().

fn is_detached(&self) -> bool

C++ method: bool QStringList::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QStringList::isEmpty() const

C++ documentation:

Returns true if the list contains no items; otherwise returns false.

See also size().

fn is_shared_with(&self, other: &ListStringRef) -> bool

C++ method: bool QStringList::isSharedWith(const QList<QString>& other) const

fn last<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QStringList::last() const

C++ documentation:

This is an overloaded function.

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringListLastIndexOfArgs<'largs>

C++ method: QStringList::lastIndexOf

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn last_index_of(&self, &mut ::reg_exp::RegExp) -> libc::c_int
2) fn last_index_of(&self, (&mut ::reg_exp::RegExp, libc::c_int)) -> libc::c_int
C++ method: int QStringList::lastIndexOf(QRegExp& rx, int from = ?) const

C++ documentation:

This function overloads lastIndexOf().

Returns the index position of the last exact match of rx in the list, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

By default, this function is case sensitive.

If an item matched, the rx regular expression will contain the matched objects (see QRegExp::matchedLength, QRegExp::cap).

This function was introduced in Qt 4.5.

See also indexOf(), contains(), and QRegExp::exactMatch().

Variant 2

Rust arguments:
1) fn last_index_of(&self, &::reg_exp::RegExp) -> libc::c_int
2) fn last_index_of(&self, (&::reg_exp::RegExp, libc::c_int)) -> libc::c_int
C++ method: int QStringList::lastIndexOf(const QRegExp& rx, int from = ?) const

C++ documentation:

Returns the index position of the last exact match of rx in the list, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

By default, this function is case sensitive.

See also indexOf(), contains(), and QRegExp::exactMatch().

Variant 3

Rust arguments:
1) fn last_index_of(&self, &::regular_expression::RegularExpression) -> libc::c_int
2) fn last_index_of(&self, (&::regular_expression::RegularExpression, libc::c_int)) -> libc::c_int
C++ method: int QStringList::lastIndexOf(const QRegularExpression& re, int from = ?) const

C++ documentation:

This is an overloaded function.

Returns the index position of the last exact match of re in the list, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

This function was introduced in Qt 5.0.

See also indexOf().

Variant 4

Rust arguments:
1) fn last_index_of(&self, &::string::String) -> libc::c_int
2) fn last_index_of(&self, (&::string::String, libc::c_int)) -> libc::c_int
C++ method: int QStringList::lastIndexOf(const QString& t, int from = ?) const

C++ documentation:

Returns the index position of the last occurrence of value in the list, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "B" << "A";
list.lastIndexOf("B");      // returns 3
list.lastIndexOf("B", 3);   // returns 3
list.lastIndexOf("B", 2);   // returns 1
list.lastIndexOf("X");      // returns -1

This function requires the value type to have an implementation of operator==().

Note that QList uses 0-based indexes, just like C++ arrays. Negative indexes are not supported with the exception of the value mentioned above.

See also indexOf().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QStringList::last()

C++ documentation:

Returns a reference to the last item in the list. The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QStringList::length() const

C++ documentation:

This function is identical to count().

This function was introduced in Qt 4.5.

See also count().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringListMidArgs<'largs>

C++ method: QStringList::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListStringRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListStringRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::list::ListStringRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::list::ListStringRef
C++ method: QList<QString> QStringList::mid(int pos, int length = ?) const

C++ documentation:

Returns a sub-list which includes elements from this list, starting at position pos. If length is -1 (the default), all elements from pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn move_(&mut self, from: c_int, to: c_int)

C++ method: void QStringList::move(int from, int to)

C++ documentation:

Moves the item at index position from to index position to.

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.move(1, 4);
// list: ["A", "C", "D", "E", "B", "F"]

This is the same as insert(to, takeAt(from)).This function assumes that both from and to are at least 0 but less than size(). To avoid failure, test that both from and to are at least 0 and less than size().

See also swap(), insert(), and takeAt().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StringListNewArgs

C++ method: QStringList::QStringList

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::string_list::StringList
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
C++ method: [constructor] void QStringList::QStringList()

Variant 2

Rust arguments:
1) fn new((&::list::ListStringRef, cpp_utils::AsStruct)) -> ::string_list::StringList
2) fn new((&::list::ListStringRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
C++ method: [constructor] void QStringList::QStringList(const QList<QString>& l)

Variant 3

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::string_list::StringList
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
C++ method: [constructor] void QStringList::QStringList(const QString& i)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringListOpAddArgs<'largs>

C++ method: QStringList::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_list::StringList>
2) fn op_add(&self, (&::string_list::StringList, cpp_utils::AsStruct)) -> ::string_list::StringList
C++ method: QStringList QStringList::operator+(const QStringList& other) const

C++ documentation:

Returns a string list that is the concatenation of this string list with the other string list.

See also append().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut ListStringRef where Args: StringListOpAddAssignArgs<'largs>

C++ method: QStringList::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::list::ListStringRef) -> &'l0 mut ::list::ListStringRef
C++ method: QList<QString>& QStringList::operator+=(const QList<QString>& l)

C++ documentation:

Appends the items of the other list to this list and returns a reference to this list.

See also operator+() and append().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::string::String) -> &'l0 mut ::list::ListStringRef
C++ method: QList<QString>& QStringList::operator+=(const QString& t)

C++ documentation:

This is an overloaded function.

Appends value to the list.

See also append() and operator<<().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 ListStringRef) -> &'l0 mut StringList

C++ method: QStringList& QStringList::operator=(const QList<QString>& other)

C++ documentation:

Copy assignment operator from QList<QString>. Assigns the other list of strings to this string list.

After the operation, other and *this will be equal.

This function was introduced in Qt 5.4.

fn op_eq(&self, l: &ListStringRef) -> bool

C++ method: bool QStringList::operator==(const QList<QString>& l) const

C++ documentation:

Returns true if other is equal to this list; otherwise returns false.

Two lists are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator!=().

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 String

C++ method: const QString& QStringList::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(). This function runs in constant time.

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut String

C++ method: QString& QStringList::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference. i must be a valid index position in the list (i.e., 0 <= i < size()).

If this function is called on a list that is currently being shared, it will trigger a copy of all elements. Otherwise, this function runs in constant time. If you do not want to modify the list you should use QList::at().

See also at() and value().

fn op_neq(&self, l: &ListStringRef) -> bool

C++ method: bool QStringList::operator!=(const QList<QString>& l) const

C++ documentation:

Returns true if other is not equal to this list; otherwise returns false.

Two lists are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator==().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut StringList where Args: StringListOpShlArgs<'largs>

C++ method: QStringList::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::list::ListStringRef) -> &'l0 mut ::string_list::StringList
C++ method: QStringList& QStringList::operator<<(const QList<QString>& l)

C++ documentation:

This is an overloaded function.

Appends the other string list to the string list and returns a reference to the latter string list.

This function was introduced in Qt 5.4.

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::string::String) -> &'l0 mut ::string_list::StringList
C++ method: QStringList& QStringList::operator<<(const QString& str)

C++ documentation:

Appends the given string, str, to this string list and returns a reference to the string list.

See also append().

Variant 3

Rust arguments: fn op_shl(&mut self, &'l1 ::string_list::StringList) -> &'l0 mut ::string_list::StringList
C++ method: QStringList& QStringList::operator<<(const QStringList& l)

C++ documentation:

This is an overloaded function.

Appends the other string list to the string list and returns a reference to the latter string list.

fn pop_back(&mut self)

C++ method: void QStringList::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn pop_front(&mut self)

C++ method: void QStringList::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst(). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

fn prepend(&mut self, t: &String)

C++ method: void QStringList::prepend(const QString& t)

C++ documentation:

Inserts value at the beginning of the list.

Example:

QList<QString> list;
list.prepend("one");
list.prepend("two");
list.prepend("three");
// list: ["three", "two", "one"]

This is the same as list.insert(0, value).

If this list is not shared, this operation is typically very fast (amortized constant time), because QList preallocates extra space on both sides of its internal buffer to allow for fast growth at both ends of the list.

See also append() and insert().

fn push_back(&mut self, t: &String)

C++ method: void QStringList::push_back(const QString& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &String)

C++ method: void QStringList::push_front(const QString& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove_all(&mut self, t: &String) -> c_int

C++ method: int QStringList::removeAll(const QString& t)

C++ documentation:

Removes all occurrences of value in the list and returns the number of entries removed.

Example:

QList<QString> list;
list << "sun" << "cloud" << "sun" << "rain";
list.removeAll("sun");
// list: ["cloud", "rain"]

This function requires the value type to have an implementation of operator==().

See also removeOne(), removeAt(), takeAt(), and replace().

fn remove_at(&mut self, i: c_int)

C++ method: void QStringList::removeAt(int i)

C++ documentation:

Removes the item at index position i. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also takeAt(), removeFirst(), removeLast(), and removeOne().

fn remove_first(&mut self)

C++ method: void QStringList::removeFirst()

C++ documentation:

Removes the first item in the list. Calling this function is equivalent to calling removeAt(0). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeFirst().

fn remove_last(&mut self)

C++ method: void QStringList::removeLast()

C++ documentation:

Removes the last item in the list. Calling this function is equivalent to calling removeAt(size() - 1). The list must not be empty. If the list can be empty, call isEmpty() before calling this function.

See also removeAt() and takeLast().

fn remove_one(&mut self, t: &String) -> bool

C++ method: bool QStringList::removeOne(const QString& t)

C++ documentation:

Removes the first occurrence of value in the list and returns true on success; otherwise returns false.

Example:

QList<QString> list;
list << "sun" << "cloud" << "sun" << "rain";
list.removeOne("sun");
// list: ["cloud", ,"sun", "rain"]

This function requires the value type to have an implementation of operator==().

This function was introduced in Qt 4.4.

See also removeAll(), removeAt(), takeAt(), and replace().

fn replace(&mut self, i: c_int, t: &String)

C++ method: void QStringList::replace(int i, const QString& t)

C++ documentation:

Replaces the item at index position i with value. i must be a valid index position in the list (i.e., 0 <= i < size()).

See also operator[]() and removeAt().

fn reserve(&mut self, size: c_int)

C++ method: void QStringList::reserve(int size)

C++ documentation:

Reserve space for alloc elements.

If alloc is smaller than the current size of the list, nothing will happen.

Use this function to avoid repetetive reallocation of QList's internal data if you can predict how many elements will be appended. Note that the reservation applies only to the internal pointer array.

This function was introduced in Qt 4.7.

fn set_sharable(&mut self, sharable: bool)

C++ method: void QStringList::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QStringList::size() const

C++ documentation:

Returns the number of items in the list.

See also isEmpty() and count().

fn starts_with(&self, t: &String) -> bool

C++ method: bool QStringList::startsWith(const QString& t) const

C++ documentation:

Returns true if this list is not empty and its first item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and contains().

fn swap<'largs, Args>(&'largs mut self, args: Args) where Args: StringListSwapArgs<'largs>

C++ method: QStringList::swap

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn swap(&mut self, &mut ::list::ListStringRef) -> ()
C++ method: void QStringList::swap(QList<QString>& other)

C++ documentation:

Swaps list other with this list. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

Variant 2

Rust arguments: fn swap(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QStringList::swap(int i, int j)

C++ documentation:

Exchange the item at index position i with the item at index position j. This function assumes that both i and j are at least 0 but less than size(). To avoid failure, test that both i and j are at least 0 and less than size().

Example:

QList<QString> list;
list << "A" << "B" << "C" << "D" << "E" << "F";
list.swap(1, 4);
// list: ["A", "E", "C", "D", "B", "F"]

See also move().

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: StringListTakeAtArgs<'largs>

C++ method: QStringList::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QStringList::takeAt(int i)

C++ documentation:

Removes the item at index position i and returns it. i must be a valid index position in the list (i.e., 0 <= i < size()).

If you don't use the return value, removeAt() is more efficient.

See also removeAt(), takeFirst(), and takeLast().

fn take_first<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: StringListTakeFirstArgs<'largs>

C++ method: QStringList::takeFirst

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_first(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn take_first(&mut self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QStringList::takeFirst()

C++ documentation:

Removes the first item in the list and returns it. This is the same as takeAt(0). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeFirst() is more efficient.

See also takeLast(), takeAt(), and removeFirst().

fn take_last<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: StringListTakeLastArgs<'largs>

C++ method: QStringList::takeLast

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_last(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn take_last(&mut self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QStringList::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This is the same as takeAt(size() - 1). This function assumes the list is not empty. To avoid failure, call isEmpty() before calling this function.

If this list is not shared, this operation takes constant time.

If you don't use the return value, removeLast() is more efficient.

See also takeFirst(), takeAt(), and removeLast().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringListValueArgs<'largs>

C++ method: QStringList::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn value(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn value(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QStringList::value(int i) const

C++ documentation:

Returns the value at index position i in the list.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that the index is going to be within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments:
1) fn value(&self, (libc::c_int, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn value(&self, (libc::c_int, &::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QStringList::value(int i, const QString& defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.