Struct qt_core::string_matcher::StringMatcher

pub struct StringMatcher {
    // some fields omitted
}

C++ type: QStringMatcher

C++ documentation:

The QStringMatcher class holds a sequence of characters that can be quickly matched in a Unicode string.

This class is useful when you have a sequence of QChars that you want to repeatedly match against some strings (perhaps in a loop), or when you want to search for the same sequence of characters multiple times in the same string. Using a matcher object and indexIn() is faster than matching a plain QString with QString::indexOf() if repeated matching takes place. This class offers no benefit if you are doing one-off string matches.

Create the QStringMatcher with the QString you want to search for. Then call indexIn() on the QString that you want to search.

Methods

impl StringMatcher

fn case_sensitivity(&self) -> CaseSensitivity

C++ method: Qt::CaseSensitivity QStringMatcher::caseSensitivity() const

C++ documentation:

Returns the case sensitivity setting for this string matcher.

See also setCaseSensitivity().

fn index_in<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringMatcherIndexInArgs<'largs>

C++ method: QStringMatcher::indexIn

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn index_in(&self, (*const ::char::Char, libc::c_int)) -> libc::c_int
2) fn index_in(&self, (*const ::char::Char, libc::c_int, libc::c_int)) -> libc::c_int
C++ method: int QStringMatcher::indexIn(const QChar* str, int length, int from = ?) const

C++ documentation:

Searches the string starting at str (of length length) from character position from (default 0, i.e. from the first character), for the string pattern() that was set in the constructor or in the most recent call to setPattern(). Returns the position where the pattern() matched in str, or -1 if no match was found.

This function was introduced in Qt 4.5.

See also setPattern() and setCaseSensitivity().

Variant 2

Rust arguments:
1) fn index_in(&self, &::string::String) -> libc::c_int
2) fn index_in(&self, (&::string::String, libc::c_int)) -> libc::c_int
C++ method: int QStringMatcher::indexIn(const QString& str, int from = ?) const

C++ documentation:

Searches the string str from character position from (default 0, i.e. from the first character), for the string pattern() that was set in the constructor or in the most recent call to setPattern(). Returns the position where the pattern() matched in str, or -1 if no match was found.

See also setPattern() and setCaseSensitivity().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StringMatcherNewArgs

C++ method: QStringMatcher::QStringMatcher

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::string_matcher::StringMatcher
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::string_matcher::StringMatcher>
C++ method: [constructor] void QStringMatcher::QStringMatcher()

Variant 2

Rust arguments:
1) fn new((*const ::char::Char, libc::c_int, cpp_utils::AsStruct)) -> ::string_matcher::StringMatcher
2) fn new((*const ::char::Char, libc::c_int, ::qt::CaseSensitivity, cpp_utils::AsStruct)) -> ::string_matcher::StringMatcher
3) fn new((*const ::char::Char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_matcher::StringMatcher>
4) fn new((*const ::char::Char, libc::c_int, ::qt::CaseSensitivity, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_matcher::StringMatcher>
C++ method: [constructor] void QStringMatcher::QStringMatcher(const QChar* uc, int len, Qt::CaseSensitivity cs = ?)

Variant 3

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::string_matcher::StringMatcher
2) fn new((&::string::String, ::qt::CaseSensitivity, cpp_utils::AsStruct)) -> ::string_matcher::StringMatcher
3) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_matcher::StringMatcher>
4) fn new((&::string::String, ::qt::CaseSensitivity, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_matcher::StringMatcher>
C++ method: [constructor] void QStringMatcher::QStringMatcher(const QString& pattern, Qt::CaseSensitivity cs = ?)

Variant 4

Rust arguments:
1) fn new((&::string_matcher::StringMatcher, cpp_utils::AsStruct)) -> ::string_matcher::StringMatcher
2) fn new((&::string_matcher::StringMatcher, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_matcher::StringMatcher>
C++ method: [constructor] void QStringMatcher::QStringMatcher(const QStringMatcher& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 StringMatcher) -> &'l0 mut StringMatcher

C++ method: QStringMatcher& QStringMatcher::operator=(const QStringMatcher& other)

C++ documentation:

Assigns the other string matcher to this string matcher.

fn pattern<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringMatcherPatternArgs<'largs>

C++ method: QStringMatcher::pattern

This is an overloaded function. Available variants:

Rust arguments:
1) fn pattern(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn pattern(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QStringMatcher::pattern() const

C++ documentation:

Returns the string pattern that this string matcher will search for.

See also setPattern().

fn set_case_sensitivity(&mut self, cs: CaseSensitivity)

C++ method: void QStringMatcher::setCaseSensitivity(Qt::CaseSensitivity cs)

C++ documentation:

Sets the case sensitivity setting of this string matcher to cs.

See also caseSensitivity(), setPattern(), and indexIn().

fn set_pattern(&mut self, pattern: &String)

C++ method: void QStringMatcher::setPattern(const QString& pattern)

C++ documentation:

Sets the string that this string matcher will search for to pattern.

See also pattern(), setCaseSensitivity(), and indexIn().