Struct qt_core::string_ref::StringRef

pub struct StringRef {
    // some fields omitted
}

C++ type: QStringRef

C++ documentation:

The QStringRef class provides a thin wrapper around QString substrings.

QStringRef provides a read-only subset of the QString API.

A string reference explicitly references a portion of a string() with a given size(), starting at a specific position(). Calling toString() returns a copy of the data as a real QString instance.

This class is designed to improve the performance of substring handling when manipulating substrings obtained from existing QString instances. QStringRef avoids the memory allocation and reference counting overhead of a standard QString by simply referencing a part of the original string. This can prove to be advantageous in low level code, such as that used in a parser, at the expense of potentially more complex code.

For most users, there are no semantic benefits to using QStringRef instead of QString since QStringRef requires attention to be paid to memory management issues, potentially making code more complex to write and maintain.

Warning: A QStringRef is only valid as long as the referenced string exists. If the original string is deleted, the string reference points to an invalid memory location.

We suggest that you only use this class in stable code where profiling has clearly identified that performance improvements can be made by replacing standard string operations with the optimized substring handling provided by this class.

Methods

impl StringRef

fn append_to<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefAppendToArgs<'largs>

C++ method: QStringRef::appendTo

This is an overloaded function. Available variants:

Rust arguments:
1) fn append_to(&self, (*mut ::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn append_to(&self, (*mut ::string::String, cpp_utils::AsStruct)) -> ::string_ref::StringRef
C++ method: QStringRef QStringRef::appendTo(QString* string) const

C++ documentation:

Appends the string reference to string, and returns a new reference to the combined string data.

fn at<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefAtArgs<'largs>

C++ method: QStringRef::at

This is an overloaded function. Available variants:

Rust arguments:
1) fn at(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::char::Char>
2) fn at(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::char::Char
C++ method: const QChar QStringRef::at(int i) const

C++ documentation:

Returns the character at the given index position in the string reference.

The position must be a valid index position in the string (i.e., 0 <= position < size()).

fn begin(&self) -> *const Char

C++ method: const QChar* QStringRef::begin() const

C++ documentation:

Same as unicode().

This function was introduced in Qt 5.4.

fn cbegin(&self) -> *const Char

C++ method: const QChar* QStringRef::cbegin() const

C++ documentation:

Same as unicode().

This function was introduced in Qt 5.4.

fn cend(&self) -> *const Char

C++ method: const QChar* QStringRef::cend() const

C++ documentation:

Returns a pointer to one character past the last one in this string. (It is the same as unicode() + size().)

This function was introduced in Qt 5.4.

fn clear(&mut self)

C++ method: void QStringRef::clear()

C++ documentation:

Clears the contents of the string reference by making it null and empty.

See also isEmpty() and isNull().

fn compare<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringRefCompareArgs<'largs>

C++ method: QStringRef::compare

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn compare(&self, &::latin1_string::Latin1String) -> libc::c_int
2) fn compare(&self, (&::latin1_string::Latin1String, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::compare(QLatin1String s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This is an overloaded function.

Compares this string with the other string and returns an integer less than, equal to, or greater than zero if this string is less than, equal to, or greater than the other string.

If cs is Qt::CaseSensitive, the comparison is case sensitive; otherwise the comparison is case insensitive.

Equivalent to compare(*this, other, cs).

This function was introduced in Qt 4.5.

See also QString::compare().

Variant 2

Rust arguments:
1) fn compare(&self, &::string::String) -> libc::c_int
2) fn compare(&self, (&::string::String, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::compare(const QString& s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This is an overloaded function.

Compares this string with the other string and returns an integer less than, equal to, or greater than zero if this string is less than, equal to, or greater than the other string.

If cs is Qt::CaseSensitive, the comparison is case sensitive; otherwise the comparison is case insensitive.

Equivalent to compare(*this, other, cs).

This function was introduced in Qt 4.5.

See also QString::compare().

Variant 3

Rust arguments:
1) fn compare(&self, &::string_ref::StringRef) -> libc::c_int
2) fn compare(&self, (&::string_ref::StringRef, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::compare(const QStringRef& s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This is an overloaded function.

Compares this string with the other string and returns an integer less than, equal to, or greater than zero if this string is less than, equal to, or greater than the other string.

If cs is Qt::CaseSensitive, the comparison is case sensitive; otherwise the comparison is case insensitive.

Equivalent to compare(*this, other, cs).

This function was introduced in Qt 4.5.

See also QString::compare().

fn compare_static<Args>(args: Args) -> c_int where Args: StringRefCompareStaticArgs

C++ method: QStringRef::compare

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn compare_static((&::string_ref::StringRef, &::latin1_string::Latin1String)) -> libc::c_int
2) fn compare_static((&::string_ref::StringRef, &::latin1_string::Latin1String, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: static int QStringRef::compare(const QStringRef& s1, QLatin1String s2, Qt::CaseSensitivity cs = ?)

C++ documentation:

This is an overloaded function.

Compares the string s1 with the string s2 and returns an integer less than, equal to, or greater than zero if s1 is less than, equal to, or greater than s2.

If cs is Qt::CaseSensitive, the comparison is case sensitive; otherwise the comparison is case insensitive.

This function was introduced in Qt 4.5.

Variant 2

Rust arguments:
1) fn compare_static((&::string_ref::StringRef, &::string::String)) -> libc::c_int
2) fn compare_static((&::string_ref::StringRef, &::string::String, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: static int QStringRef::compare(const QStringRef& s1, const QString& s2, Qt::CaseSensitivity arg3 = ?)

C++ documentation:

Compares the string s1 with the string s2 and returns an integer less than, equal to, or greater than zero if s1 is less than, equal to, or greater than s2.

If cs is Qt::CaseSensitive, the comparison is case sensitive; otherwise the comparison is case insensitive.

This function was introduced in Qt 4.5.

Variant 3

Rust arguments:
1) fn compare_static((&::string_ref::StringRef, &::string_ref::StringRef)) -> libc::c_int
2) fn compare_static((&::string_ref::StringRef, &::string_ref::StringRef, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: static int QStringRef::compare(const QStringRef& s1, const QStringRef& s2, Qt::CaseSensitivity arg3 = ?)

C++ documentation:

This is an overloaded function.

Compares the string s1 with the string s2 and returns an integer less than, equal to, or greater than zero if s1 is less than, equal to, or greater than s2.

If cs is Qt::CaseSensitive, the comparison is case sensitive; otherwise the comparison is case insensitive.

This function was introduced in Qt 4.5.

fn const_data(&self) -> *const Char

C++ method: const QChar* QStringRef::constData() const

C++ documentation:

Same as unicode().

fn contains<'largs, Args>(&'largs self, args: Args) -> bool where Args: StringRefContainsArgs<'largs>

C++ method: QStringRef::contains

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn contains(&self, &::char::Char) -> bool
2) fn contains(&self, (&::char::Char, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::contains(QChar ch, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads contains().

Returns true if this string contains an occurrence of the character ch; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

Variant 2

Rust arguments:
1) fn contains(&self, &::latin1_string::Latin1String) -> bool
2) fn contains(&self, (&::latin1_string::Latin1String, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::contains(QLatin1String str, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads contains().

Returns true if this string reference contains an occurrence of the string str; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4,8.

See also indexOf() and count().

Variant 3

Rust arguments:
1) fn contains(&self, &::string::String) -> bool
2) fn contains(&self, (&::string::String, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::contains(const QString& str, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns true if this string reference contains an occurrence of the string str; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also indexOf() and count().

Variant 4

Rust arguments:
1) fn contains(&self, &::string_ref::StringRef) -> bool
2) fn contains(&self, (&::string_ref::StringRef, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::contains(const QStringRef& str, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads contains().

Returns true if this string reference contains an occurrence of the string reference str; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also indexOf() and count().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringRefCountArgs<'largs>

C++ method: QStringRef::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QStringRef::count() const

C++ documentation:

Returns the number of characters referred to by the string reference. Equivalent to size() and length().

See also position() and string().

Variant 2

Rust arguments:
1) fn count(&self, &::char::Char) -> libc::c_int
2) fn count(&self, (&::char::Char, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::count(QChar c, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads count().

Returns the number of occurrences of the character ch in the string reference.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::count(), contains(), and indexOf().

Variant 3

Rust arguments:
1) fn count(&self, &::string::String) -> libc::c_int
2) fn count(&self, (&::string::String, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::count(const QString& s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns the number of (potentially overlapping) occurrences of the string str in this string reference.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::count(), contains(), and indexOf().

Variant 4

Rust arguments:
1) fn count(&self, &::string_ref::StringRef) -> libc::c_int
2) fn count(&self, (&::string_ref::StringRef, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::count(const QStringRef& s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads count().

Returns the number of (potentially overlapping) occurrences of the string reference str in this string reference.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::count(), contains(), and indexOf().

fn data(&self) -> *const Char

C++ method: const QChar* QStringRef::data() const

C++ documentation:

Same as unicode().

fn end(&self) -> *const Char

C++ method: const QChar* QStringRef::end() const

C++ documentation:

Returns a pointer to one character past the last one in this string. (It is the same as unicode() + size().)

This function was introduced in Qt 5.4.

fn ends_with<'largs, Args>(&'largs self, args: Args) -> bool where Args: StringRefEndsWithArgs<'largs>

C++ method: QStringRef::endsWith

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn ends_with(&self, &::char::Char) -> bool
2) fn ends_with(&self, (&::char::Char, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::endsWith(QChar c, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads endsWith().

Returns true if the string reference ends with ch; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::endsWith() and endsWith().

Variant 2

Rust arguments:
1) fn ends_with(&self, &::latin1_string::Latin1String) -> bool
2) fn ends_with(&self, (&::latin1_string::Latin1String, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::endsWith(QLatin1String s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads endsWith().

This function was introduced in Qt 4.8.

See also QString::endsWith() and endsWith().

Variant 3

Rust arguments:
1) fn ends_with(&self, &::string::String) -> bool
2) fn ends_with(&self, (&::string::String, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::endsWith(const QString& s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns true if the string reference ends with str; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::endsWith() and startsWith().

Variant 4

Rust arguments:
1) fn ends_with(&self, &::string_ref::StringRef) -> bool
2) fn ends_with(&self, (&::string_ref::StringRef, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::endsWith(const QStringRef& c, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads endsWith().

This function was introduced in Qt 4.8.

See also QString::endsWith() and endsWith().

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringRefIndexOfArgs<'largs>

C++ method: QStringRef::indexOf

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn index_of(&self, &::char::Char) -> libc::c_int
2) fn index_of(&self, (&::char::Char, libc::c_int)) -> libc::c_int
3) fn index_of(&self, (&::char::Char, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::indexOf(QChar ch, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads indexOf().

Returns the index position of the first occurrence of the character ch in the string reference, searching forward from index position from. Returns -1 if ch could not be found.

This function was introduced in Qt 4.8.

See also QString::indexOf(), lastIndexOf(), contains(), and count().

Variant 2

Rust arguments:
1) fn index_of(&self, &::latin1_string::Latin1String) -> libc::c_int
2) fn index_of(&self, (&::latin1_string::Latin1String, libc::c_int)) -> libc::c_int
3) fn index_of(&self, (&::latin1_string::Latin1String, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::indexOf(QLatin1String str, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns the index position of the first occurrence of the string str in this string reference, searching forward from index position from. Returns -1 if str is not found.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

If from is -1, the search starts at the last character; if it is -2, at the next to last character and so on.

This function was introduced in Qt 4.8.

See also QString::indexOf(), lastIndexOf(), contains(), and count().

Variant 3

Rust arguments:
1) fn index_of(&self, &::string::String) -> libc::c_int
2) fn index_of(&self, (&::string::String, libc::c_int)) -> libc::c_int
3) fn index_of(&self, (&::string::String, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::indexOf(const QString& str, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns the index position of the first occurrence of the string str in this string reference, searching forward from index position from. Returns -1 if str is not found.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

If from is -1, the search starts at the last character; if it is -2, at the next to last character and so on.

This function was introduced in Qt 4.8.

See also QString::indexOf(), lastIndexOf(), contains(), and count().

Variant 4

Rust arguments:
1) fn index_of(&self, &::string_ref::StringRef) -> libc::c_int
2) fn index_of(&self, (&::string_ref::StringRef, libc::c_int)) -> libc::c_int
3) fn index_of(&self, (&::string_ref::StringRef, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::indexOf(const QStringRef& str, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads indexOf().

Returns the index position of the first occurrence of the string reference str in this string reference, searching forward from index position from. Returns -1 if str is not found.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::indexOf(), lastIndexOf(), contains(), and count().

fn is_empty(&self) -> bool

C++ method: bool QStringRef::isEmpty() const

C++ documentation:

Returns true if the string reference has no characters; otherwise returns false.

A string reference is empty if its size is zero.

See also size().

fn is_null(&self) -> bool

C++ method: bool QStringRef::isNull() const

C++ documentation:

Returns true if string() returns a null pointer or a pointer to a null string; otherwise returns true.

See also size().

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringRefLastIndexOfArgs<'largs>

C++ method: QStringRef::lastIndexOf

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn last_index_of(&self, &::char::Char) -> libc::c_int
2) fn last_index_of(&self, (&::char::Char, libc::c_int)) -> libc::c_int
3) fn last_index_of(&self, (&::char::Char, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::lastIndexOf(QChar ch, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads lastIndexOf().

Returns the index position of the last occurrence of the character ch, searching backward from position from.

This function was introduced in Qt 4.8.

See also QString::lastIndexOf(), indexOf(), contains(), and count().

Variant 2

Rust arguments:
1) fn last_index_of(&self, &::latin1_string::Latin1String) -> libc::c_int
2) fn last_index_of(&self, (&::latin1_string::Latin1String, libc::c_int)) -> libc::c_int
3) fn last_index_of(&self, (&::latin1_string::Latin1String, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::lastIndexOf(QLatin1String str, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads lastIndexOf().

Returns the index position of the last occurrence of the string str in this string reference, searching backward from index position from. If from is -1 (default), the search starts at the last character; if from is -2, at the next to last character and so on. Returns -1 if str is not found.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::lastIndexOf(), indexOf(), contains(), and count().

Variant 3

Rust arguments:
1) fn last_index_of(&self, &::string::String) -> libc::c_int
2) fn last_index_of(&self, (&::string::String, libc::c_int)) -> libc::c_int
3) fn last_index_of(&self, (&::string::String, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::lastIndexOf(const QString& str, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns the index position of the last occurrence of the string str in this string reference, searching backward from index position from. If from is -1 (default), the search starts at the last character; if from is -2, at the next to last character and so on. Returns -1 if str is not found.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::lastIndexOf(), indexOf(), contains(), and count().

Variant 4

Rust arguments:
1) fn last_index_of(&self, &::string_ref::StringRef) -> libc::c_int
2) fn last_index_of(&self, (&::string_ref::StringRef, libc::c_int)) -> libc::c_int
3) fn last_index_of(&self, (&::string_ref::StringRef, libc::c_int, ::qt::CaseSensitivity)) -> libc::c_int
C++ method: int QStringRef::lastIndexOf(const QStringRef& str, int from = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads lastIndexOf().

Returns the index position of the last occurrence of the string reference str in this string reference, searching backward from index position from. If from is -1 (default), the search starts at the last character; if from is -2, at the next to last character and so on. Returns -1 if str is not found.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::lastIndexOf(), indexOf(), contains(), and count().

fn left<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefLeftArgs<'largs>

C++ method: QStringRef::left

This is an overloaded function. Available variants:

Rust arguments:
1) fn left(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn left(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string_ref::StringRef
C++ method: QStringRef QStringRef::left(int n) const

C++ documentation:

Returns a substring reference to the n leftmost characters of the string.

If n is greater than size() or less than zero, a reference to the entire string is returned.

This function was introduced in Qt 5.2.

See also right(), mid(), and startsWith().

fn length(&self) -> c_int

C++ method: int QStringRef::length() const

C++ documentation:

Returns the number of characters referred to by the string reference. Equivalent to size() and count().

See also position() and string().

fn locale_aware_compare<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringRefLocaleAwareCompareArgs<'largs>

C++ method: QStringRef::localeAwareCompare

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn locale_aware_compare(&self, &::string::String) -> libc::c_int
C++ method: int QStringRef::localeAwareCompare(const QString& s) const

C++ documentation:

This is an overloaded function.

Compares this string with the other string and returns an integer less than, equal to, or greater than zero if this string is less than, equal to, or greater than the other string.

The comparison is performed in a locale- and also platform-dependent manner. Use this function to present sorted lists of strings to the user.

This function was introduced in Qt 4.5.

Variant 2

Rust arguments: fn locale_aware_compare(&self, &::string_ref::StringRef) -> libc::c_int
C++ method: int QStringRef::localeAwareCompare(const QStringRef& s) const

C++ documentation:

This is an overloaded function.

Compares this string with the other string and returns an integer less than, equal to, or greater than zero if this string is less than, equal to, or greater than the other string.

The comparison is performed in a locale- and also platform-dependent manner. Use this function to present sorted lists of strings to the user.

This function was introduced in Qt 4.5.

fn locale_aware_compare_static<Args>(args: Args) -> c_int where Args: StringRefLocaleAwareCompareStaticArgs

C++ method: QStringRef::localeAwareCompare

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn locale_aware_compare_static((&::string_ref::StringRef, &::string::String)) -> libc::c_int
C++ method: static int QStringRef::localeAwareCompare(const QStringRef& s1, const QString& s2)

C++ documentation:

Compares s1 with s2 and returns an integer less than, equal to, or greater than zero if s1 is less than, equal to, or greater than s2.

The comparison is performed in a locale- and also platform-dependent manner. Use this function to present sorted lists of strings to the user.

On OS X and iOS, this function compares according the "Order for sorted lists" setting in the International prefereces panel.

This function was introduced in Qt 4.5.

See also compare() and QLocale.

Variant 2

Rust arguments: fn locale_aware_compare_static((&::string_ref::StringRef, &::string_ref::StringRef)) -> libc::c_int
C++ method: static int QStringRef::localeAwareCompare(const QStringRef& s1, const QStringRef& s2)

C++ documentation:

This is an overloaded function.

Compares s1 with s2 and returns an integer less than, equal to, or greater than zero if s1 is less than, equal to, or greater than s2.

The comparison is performed in a locale- and also platform-dependent manner. Use this function to present sorted lists of strings to the user.

This function was introduced in Qt 4.5.

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefMidArgs<'largs>

C++ method: QStringRef::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string_ref::StringRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::string_ref::StringRef
C++ method: QStringRef QStringRef::mid(int pos, int n = ?) const

C++ documentation:

Returns a substring reference to n characters of this string, starting at the specified position.

If the position exceeds the length of the string, a null reference is returned.

If there are less than n characters available in the string, starting at the given position, or if n is -1 (default), the function returns all characters from the specified position onwards.

This function was introduced in Qt 5.2.

See also left() and right().

fn new<Args>(args: Args) -> Args::ReturnType where Args: StringRefNewArgs

C++ method: QStringRef::QStringRef

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::string_ref::StringRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
C++ method: [constructor] void QStringRef::QStringRef()

Variant 2

Rust arguments:
1) fn new((*const ::string::String, cpp_utils::AsStruct)) -> ::string_ref::StringRef
2) fn new((*const ::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
C++ method: [constructor] void QStringRef::QStringRef(const QString* string)

Variant 3

Rust arguments:
1) fn new((*const ::string::String, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::string_ref::StringRef
2) fn new((*const ::string::String, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
C++ method: [constructor] void QStringRef::QStringRef(const QString* string, int position, int size)

Variant 4

Rust arguments:
1) fn new((&::string_ref::StringRef, cpp_utils::AsStruct)) -> ::string_ref::StringRef
2) fn new((&::string_ref::StringRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
C++ method: [constructor] void QStringRef::QStringRef(const QStringRef& other)

fn op_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut StringRef where Args: StringRefOpAssignArgs<'largs>

C++ method: QStringRef::operator=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_assign(&mut self, *const ::string::String) -> &'l0 mut ::string_ref::StringRef
C++ method: QStringRef& QStringRef::operator=(const QString* string)

C++ documentation:

Constructs a string reference to the given string and assigns it to this string reference, returning the result.

Variant 2

Rust arguments: fn op_assign(&mut self, &'l1 ::string_ref::StringRef) -> &'l0 mut ::string_ref::StringRef
C++ method: QStringRef& QStringRef::operator=(const QStringRef& other)

C++ documentation:

Assigns the other string reference to this string reference, and returns the result.

fn op_eq(&self, s: *const c_char) -> bool

C++ method: bool QStringRef::operator==(const char* s) const

C++ documentation:

This function overloads operator==().

The s byte array is converted to a QStringRef using the fromUtf8() function. This function stops conversion at the first NUL character found, or the end of the byte array.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

Returns true if this string is lexically equal to the parameter string s. Otherwise returns false.

fn op_ge(&self, s: *const c_char) -> bool

C++ method: bool QStringRef::operator>=(const char* s) const

C++ documentation:

This function overloads operator>=().

The s const char pointer is converted to a QStringRef using the fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

Returns true if this string is lexically greater than or equal to the parameter string s. Otherwise returns false.

fn op_gt(&self, s: *const c_char) -> bool

C++ method: bool QStringRef::operator>(const char* s) const

C++ documentation:

This function overloads operator>().

The s const char pointer is converted to a QStringRef using the fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

Returns true if this string is lexically greater than the parameter string s. Otherwise returns false.

fn op_le(&self, s: *const c_char) -> bool

C++ method: bool QStringRef::operator<=(const char* s) const

C++ documentation:

This function overloads operator<=().

The s const char pointer is converted to a QStringRef using the fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

Returns true if this string is lexically smaller than or equal to the parameter string s. Otherwise returns false.

fn op_lt(&self, s: *const c_char) -> bool

C++ method: bool QStringRef::operator<(const char* s) const

C++ documentation:

This function overloads operator<().

The s const char pointer is converted to a QStringRef using the fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

Returns true if this string is lexically smaller than the parameter string s. Otherwise returns false.

fn op_neq(&self, s: *const c_char) -> bool

C++ method: bool QStringRef::operator!=(const char* s) const

C++ documentation:

This function overloads operator!=().

The s const char pointer is converted to a QStringRef using the fromUtf8() function.

You can disable this operator by defining QT_NO_CAST_FROM_ASCII when you compile your applications. This can be useful if you want to ensure that all user-visible strings go through QObject::tr(), for example.

Returns true if this string is not lexically equal to the parameter string s. Otherwise returns false.

fn position(&self) -> c_int

C++ method: int QStringRef::position() const

C++ documentation:

Returns the starting position in the referenced string that is referred to by the string reference.

See also size() and string().

fn right<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefRightArgs<'largs>

C++ method: QStringRef::right

This is an overloaded function. Available variants:

Rust arguments:
1) fn right(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn right(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::string_ref::StringRef
C++ method: QStringRef QStringRef::right(int n) const

C++ documentation:

Returns a substring reference to the n rightmost characters of the string.

If n is greater than size() or less than zero, a reference to the entire string is returned.

This function was introduced in Qt 5.2.

See also left(), mid(), and endsWith().

fn size(&self) -> c_int

C++ method: int QStringRef::size() const

C++ documentation:

Returns the number of characters referred to by the string reference. Equivalent to length() and count().

See also position() and string().

fn split<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefSplitArgs<'largs>

C++ method: QStringRef::split

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn split(&self, (&::char::Char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorStringRefRef>
2) fn split(&self, (&::char::Char, ::string::SplitBehavior, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorStringRefRef>
3) fn split(&self, (&::char::Char, ::string::SplitBehavior, ::qt::CaseSensitivity, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorStringRefRef>
4) fn split(&self, (&::char::Char, cpp_utils::AsStruct)) -> ::vector::VectorStringRefRef
5) fn split(&self, (&::char::Char, ::string::SplitBehavior, cpp_utils::AsStruct)) -> ::vector::VectorStringRefRef
6) fn split(&self, (&::char::Char, ::string::SplitBehavior, ::qt::CaseSensitivity, cpp_utils::AsStruct)) -> ::vector::VectorStringRefRef
C++ method: QVector<QStringRef> QStringRef::split(QChar sep, QString::SplitBehavior behavior = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This is an overloaded function.

This function was introduced in Qt 5.4.

Variant 2

Rust arguments:
1) fn split(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorStringRefRef>
2) fn split(&self, (&::string::String, ::string::SplitBehavior, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorStringRefRef>
3) fn split(&self, (&::string::String, ::string::SplitBehavior, ::qt::CaseSensitivity, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorStringRefRef>
4) fn split(&self, (&::string::String, cpp_utils::AsStruct)) -> ::vector::VectorStringRefRef
5) fn split(&self, (&::string::String, ::string::SplitBehavior, cpp_utils::AsStruct)) -> ::vector::VectorStringRefRef
6) fn split(&self, (&::string::String, ::string::SplitBehavior, ::qt::CaseSensitivity, cpp_utils::AsStruct)) -> ::vector::VectorStringRefRef
C++ method: QVector<QStringRef> QStringRef::split(const QString& sep, QString::SplitBehavior behavior = ?, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Splits the string into substrings references wherever sep occurs, and returns the list of those strings. If sep does not match anywhere in the string, split() returns a single-element vector containing this string reference.

cs specifies whether sep should be matched case sensitively or case insensitively.

If behavior is QString::SkipEmptyParts, empty entries don't appear in the result. By default, empty entries are kept.

Note: All references are valid as long this string is alive. Destroying this string will cause all references be dangling pointers.

This function was introduced in Qt 5.4.

fn starts_with<'largs, Args>(&'largs self, args: Args) -> bool where Args: StringRefStartsWithArgs<'largs>

C++ method: QStringRef::startsWith

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn starts_with(&self, &::char::Char) -> bool
2) fn starts_with(&self, (&::char::Char, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::startsWith(QChar c, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads startsWith().

Returns true if the string reference starts with ch; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::startsWith() and endsWith().

Variant 2

Rust arguments:
1) fn starts_with(&self, &::latin1_string::Latin1String) -> bool
2) fn starts_with(&self, (&::latin1_string::Latin1String, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::startsWith(QLatin1String s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads startsWith().

This function was introduced in Qt 4.8.

See also QString::startsWith() and endsWith().

Variant 3

Rust arguments:
1) fn starts_with(&self, &::string::String) -> bool
2) fn starts_with(&self, (&::string::String, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::startsWith(const QString& s, Qt::CaseSensitivity cs = ?) const

C++ documentation:

Returns true if the string reference starts with str; otherwise returns false.

If cs is Qt::CaseSensitive (default), the search is case sensitive; otherwise the search is case insensitive.

This function was introduced in Qt 4.8.

See also QString::startsWith() and endsWith().

Variant 4

Rust arguments:
1) fn starts_with(&self, &::string_ref::StringRef) -> bool
2) fn starts_with(&self, (&::string_ref::StringRef, ::qt::CaseSensitivity)) -> bool
C++ method: bool QStringRef::startsWith(const QStringRef& c, Qt::CaseSensitivity cs = ?) const

C++ documentation:

This function overloads startsWith().

This function was introduced in Qt 4.8.

See also QString::startsWith() and endsWith().

fn string(&self) -> *const String

C++ method: const QString* QStringRef::string() const

C++ documentation:

Returns a pointer to the string referred to by the string reference, or 0 if it does not reference a string.

See also unicode().

fn to_double<'largs, Args>(&'largs self, args: Args) -> c_double where Args: StringRefToDoubleArgs<'largs>

C++ method: QStringRef::toDouble

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_double(&self, ()) -> libc::c_double
2) fn to_double(&self, *mut bool) -> libc::c_double
C++ method: double QStringRef::toDouble(bool* ok = ?) const

C++ documentation:

Returns the string converted to a double value.

Returns 0.0 if the conversion fails.

If a conversion error occurs, ok is set to false; otherwise ok is set to true.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toDouble()

For historic reasons, this function does not handle thousands group separators. If you need to convert such numbers, use QLocale::toDouble().

This function was introduced in Qt 5.1.

See also QString::toDouble().

fn to_float<'largs, Args>(&'largs self, args: Args) -> c_float where Args: StringRefToFloatArgs<'largs>

C++ method: QStringRef::toFloat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_float(&self, ()) -> libc::c_float
2) fn to_float(&self, *mut bool) -> libc::c_float
C++ method: float QStringRef::toFloat(bool* ok = ?) const

C++ documentation:

Returns the string converted to a float value.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true. Returns 0.0 if the conversion fails.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toFloat()

This function was introduced in Qt 5.1.

See also QString::toFloat().

fn to_int<'largs, Args>(&'largs self, args: Args) -> c_int where Args: StringRefToIntArgs<'largs>

C++ method: QStringRef::toInt

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_int(&self, ()) -> libc::c_int
2) fn to_int(&self, *mut bool) -> libc::c_int
3) fn to_int(&self, (*mut bool, libc::c_int)) -> libc::c_int
C++ method: int QStringRef::toInt(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to an int using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toInt()

This function was introduced in Qt 5.1.

See also QString::toInt().

fn to_latin1<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefToLatin1Args<'largs>

C++ method: QStringRef::toLatin1

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_latin1(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_latin1(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QStringRef::toLatin1() const

C++ documentation:

Returns a Latin-1 representation of the string as a QByteArray.

The returned byte array is undefined if the string contains non-Latin1 characters. Those characters may be suppressed or replaced with a question mark.

This function was introduced in Qt 4.8.

See also toUtf8(), toLocal8Bit(), and QTextCodec.

fn to_local8_bit<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefToLocal8BitArgs<'largs>

C++ method: QStringRef::toLocal8Bit

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_local8_bit(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_local8_bit(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QStringRef::toLocal8Bit() const

C++ documentation:

Returns the local 8-bit representation of the string as a QByteArray. The returned byte array is undefined if the string contains characters not supported by the local 8-bit encoding.

QTextCodec::codecForLocale() is used to perform the conversion from Unicode. If the locale encoding could not be determined, this function does the same as toLatin1().

If this string contains any characters that cannot be encoded in the locale, the returned byte array is undefined. Those characters may be suppressed or replaced by another.

This function was introduced in Qt 4.8.

See also toLatin1(), toUtf8(), and QTextCodec.

fn to_long<'largs, Args>(&'largs self, args: Args) -> c_long where Args: StringRefToLongArgs<'largs>

C++ method: QStringRef::toLong

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_long(&self, ()) -> libc::c_long
2) fn to_long(&self, *mut bool) -> libc::c_long
3) fn to_long(&self, (*mut bool, libc::c_int)) -> libc::c_long
C++ method: long QStringRef::toLong(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to a long using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toLong()

This function was introduced in Qt 5.1.

See also QString::toLong().

fn to_long_long<'largs, Args>(&'largs self, args: Args) -> i64 where Args: StringRefToLongLongArgs<'largs>

C++ method: QStringRef::toLongLong

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_long_long(&self, ()) -> i64
2) fn to_long_long(&self, *mut bool) -> i64
3) fn to_long_long(&self, (*mut bool, libc::c_int)) -> i64
C++ method: qlonglong QStringRef::toLongLong(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to a long long using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toLongLong()

This function was introduced in Qt 5.1.

See also QString::toLongLong().

fn to_short<'largs, Args>(&'largs self, args: Args) -> c_short where Args: StringRefToShortArgs<'largs>

C++ method: QStringRef::toShort

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_short(&self, ()) -> libc::c_short
2) fn to_short(&self, *mut bool) -> libc::c_short
3) fn to_short(&self, (*mut bool, libc::c_int)) -> libc::c_short
C++ method: short QStringRef::toShort(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to a short using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toShort()

This function was introduced in Qt 5.1.

See also QString::toShort().

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefToStringArgs<'largs>

C++ method: QStringRef::toString

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QStringRef::toString() const

C++ documentation:

Returns a copy of the string reference as a QString object.

If the string reference is not a complete reference of the string (meaning that position() is 0 and size() equals string()->size()), this function will allocate a new string to return.

See also string().

fn to_u_int<'largs, Args>(&'largs self, args: Args) -> c_uint where Args: StringRefToUIntArgs<'largs>

C++ method: QStringRef::toUInt

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_u_int(&self, ()) -> libc::c_uint
2) fn to_u_int(&self, *mut bool) -> libc::c_uint
3) fn to_u_int(&self, (*mut bool, libc::c_int)) -> libc::c_uint
C++ method: unsigned int QStringRef::toUInt(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to an unsigned int using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toUInt()

This function was introduced in Qt 5.1.

See also QString::toUInt().

fn to_u_long<'largs, Args>(&'largs self, args: Args) -> c_ulong where Args: StringRefToULongArgs<'largs>

C++ method: QStringRef::toULong

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_u_long(&self, ()) -> libc::c_ulong
2) fn to_u_long(&self, *mut bool) -> libc::c_ulong
3) fn to_u_long(&self, (*mut bool, libc::c_int)) -> libc::c_ulong
C++ method: unsigned long QStringRef::toULong(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to an unsigned long using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toULongLong()

This function was introduced in Qt 5.1.

See also QString::toULong().

fn to_u_long_long<'largs, Args>(&'largs self, args: Args) -> u64 where Args: StringRefToULongLongArgs<'largs>

C++ method: QStringRef::toULongLong

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_u_long_long(&self, ()) -> u64
2) fn to_u_long_long(&self, *mut bool) -> u64
3) fn to_u_long_long(&self, (*mut bool, libc::c_int)) -> u64
C++ method: qulonglong QStringRef::toULongLong(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to an unsigned long long using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toULongLong()

This function was introduced in Qt 5.1.

See also QString::toULongLong().

fn to_u_short<'largs, Args>(&'largs self, args: Args) -> c_ushort where Args: StringRefToUShortArgs<'largs>

C++ method: QStringRef::toUShort

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_u_short(&self, ()) -> libc::c_ushort
2) fn to_u_short(&self, *mut bool) -> libc::c_ushort
3) fn to_u_short(&self, (*mut bool, libc::c_int)) -> libc::c_ushort
C++ method: unsigned short QStringRef::toUShort(bool* ok = ?, int base = ?) const

C++ documentation:

Returns the string converted to an unsigned short using base base, which is 10 by default and must be between 2 and 36, or 0. Returns 0 if the conversion fails.

If a conversion error occurs, *ok is set to false; otherwise *ok is set to true.

If base is 0, the C language convention is used: If the string begins with "0x", base 16 is used; if the string begins with "0", base 8 is used; otherwise, base 10 is used.

The string conversion will always happen in the 'C' locale. For locale dependent conversion use QLocale::toUShort()

This function was introduced in Qt 5.1.

See also QString::toUShort().

fn to_ucs4<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefToUcs4Args<'largs>

C++ method: QStringRef::toUcs4

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_ucs4(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorCUint>
2) fn to_ucs4(&self, cpp_utils::AsStruct) -> ::vector::VectorCUint
C++ method: QVector<unsigned int> QStringRef::toUcs4() const

C++ documentation:

Returns a UCS-4/UTF-32 representation of the string as a QVector<uint>.

UCS-4 is a Unicode codec and therefore it is lossless. All characters from this string will be encoded in UCS-4. Any invalid sequence of code units in this string is replaced by the Unicode's replacement character (QChar::ReplacementCharacter, which corresponds to U+FFFD).

The returned vector is not NUL terminated.

This function was introduced in Qt 4.8.

See also toUtf8(), toLatin1(), toLocal8Bit(), and QTextCodec.

fn to_utf8<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefToUtf8Args<'largs>

C++ method: QStringRef::toUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_utf8(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_utf8(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QStringRef::toUtf8() const

C++ documentation:

Returns a UTF-8 representation of the string as a QByteArray.

UTF-8 is a Unicode codec and can represent all characters in a Unicode string like QString.

This function was introduced in Qt 4.8.

See also toLatin1(), toLocal8Bit(), and QTextCodec.

fn trimmed<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: StringRefTrimmedArgs<'largs>

C++ method: QStringRef::trimmed

This is an overloaded function. Available variants:

Rust arguments:
1) fn trimmed(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn trimmed(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QStringRef::trimmed() const

C++ documentation:

Returns a string that has whitespace removed from the start and the end.

Whitespace means any character for which QChar::isSpace() returns true. This includes the ASCII characters '\t', '\n', '\v', '\f', '\r', and ' '.

Unlike QString::simplified(), trimmed() leaves internal whitespace alone.

This function was introduced in Qt 5.1.

See also QString::trimmed().

fn unicode(&self) -> *const Char

C++ method: const QChar* QStringRef::unicode() const

C++ documentation:

Returns a Unicode representation of the string reference. Since the data stems directly from the referenced string, it is not null-terminated unless the string reference includes the string's null terminator.

See also string().