Enum qt_core::sys_info::MacVersion

pub enum MacVersion {
    MVUnknown,
    MV9,
    MV100,
    MV101,
    MV102,
    MV103,
    MV104,
    MV105,
    MV106,
    MV107,
    MV108,
    MV109,
    MV1010,
    MV1011,
    MVIOS,
    MVIOS43,
    MVIOS50,
    MVIOS51,
    MVIOS60,
    MVIOS61,
    MVIOS70,
    MVIOS71,
    MVIOS80,
    MVIOS81,
    MVIOS82,
    MVIOS83,
    MVIOS84,
    MVIOS90,
    MVNone,
}

C++ type: QSysInfo::MacVersion

C++ documentation:

This enum provides symbolic names for the various versions of the Darwin operating system, covering both OS X and iOS. The QSysInfo::MacintoshVersion variable gives the version of the system on which the application is run.

See also WinVersion.

Variants
MVUnknown

An unknown and currently unsupported platform (C++ enum variant: MV_Unknown = 0)

MV9

Mac OS 9 (unsupported) (C++ enum variant: MV_9 = 1)

MV100

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_0 = 2: Mac OS X 10.0 (unsupported)

MV_CHEETAH = 2: Apple codename for MV_10_0

MV101

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_1 = 3: Mac OS X 10.1 (unsupported)

MV_PUMA = 3: Apple codename for MV_10_1

MV102

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_2 = 4: Mac OS X 10.2 (unsupported)

MV_JAGUAR = 4: Apple codename for MV_10_2

MV103

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_3 = 5: Mac OS X 10.3 (unsupported)

MV_PANTHER = 5: Apple codename for MV_10_3

MV104

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_4 = 6: Mac OS X 10.4 (unsupported)

MV_TIGER = 6: Apple codename for MV_10_4

MV105

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_5 = 7: Mac OS X 10.5 (unsupported)

MV_LEOPARD = 7: Apple codename for MV_10_5

MV106

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_6 = 8: Mac OS X 10.6

MV_SNOWLEOPARD = 8: Apple codename for MV_10_6

MV107

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_7 = 9: OS X 10.7

MV_LION = 9: Apple codename for MV_10_7

MV108

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_8 = 10: OS X 10.8

MV_MOUNTAINLION = 10: Apple codename for MV_10_8

MV109

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_9 = 11: OS X 10.9

MV_MAVERICKS = 11: Apple codename for MV_10_9

MV1010

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_10 = 12: OS X 10.10

MV_YOSEMITE = 12: Apple codename for MV_10_10

MV1011

This variant corresponds to multiple C++ enum variants with the same value:

MV_10_11 = 13: OS X 10.11

MV_ELCAPITAN = 13: Apple codename for MV_10_11

MVIOS

iOS (any) (C++ enum variant: MV_IOS = 256)

MVIOS43

iOS 4.3 (C++ enum variant: MV_IOS_4_3 = 323)

MVIOS50

iOS 5.0 (C++ enum variant: MV_IOS_5_0 = 336)

MVIOS51

iOS 5.1 (C++ enum variant: MV_IOS_5_1 = 337)

MVIOS60

iOS 6.0 (C++ enum variant: MV_IOS_6_0 = 352)

MVIOS61

iOS 6.1 (C++ enum variant: MV_IOS_6_1 = 353)

MVIOS70

iOS 7.0 (C++ enum variant: MV_IOS_7_0 = 368)

MVIOS71

iOS 7.1 (C++ enum variant: MV_IOS_7_1 = 369)

MVIOS80

iOS 8.0 (C++ enum variant: MV_IOS_8_0 = 384)

MVIOS81

iOS 8.1 (C++ enum variant: MV_IOS_8_1 = 385)

MVIOS82

iOS 8.2 (C++ enum variant: MV_IOS_8_2 = 386)

MVIOS83

iOS 8.3 (C++ enum variant: MV_IOS_8_3 = 387)

MVIOS84

iOS 8.4 (C++ enum variant: MV_IOS_8_4 = 388)

MVIOS90

iOS 9.0 (C++ enum variant: MV_IOS_9_0 = 400)

MVNone

Not a Darwin operating system (C++ enum variant: MV_None = 65535)