Enum qt_core::sys_info::WinVersion

pub enum WinVersion {
    WVNone,
    WV32s,
    WV95,
    WV98,
    WVMe,
    WVDOSBased,
    WVNT,
    WV2000,
    WVXP,
    WV2003,
    WVVISTA,
    WVWINDOWS7,
    WVWINDOWS8,
    WVWINDOWS81,
    WVWINDOWS10,
    WVNTBased,
    WVCE,
    WVCENET,
    WVCE5,
    WVCE6,
    WVCEBased,
}

C++ type: QSysInfo::WinVersion

C++ documentation:

This enum provides symbolic names for the various versions of the Windows operating system. On Windows, the QSysInfo::WindowsVersion variable gives the version of the system on which the application is run.

MS-DOS-based versions:

NT-based versions (note that each operating system version is only represented once rather than each Windows edition):

Alternatively, you may use the following macros which correspond directly to the Windows operating system version number:

CE-based versions:

The following masks can be used for testing whether a Windows version is MS-DOS-based, NT-based, or CE-based:

See also MacVersion.

Variants
WVNone

Operating system other than Windows. (C++ enum variant: WV_None = 0)

WV32s

Windows 3.1 with Win 32s (C++ enum variant: WV_32s = 1)

WV95

Windows 95 (C++ enum variant: WV_95 = 2)

WV98

Windows 98 (C++ enum variant: WV_98 = 3)

WVMe

Windows Me (C++ enum variant: WV_Me = 4)

WVDOSBased

MS-DOS-based version of Windows (C++ enum variant: WV_DOS_based = 15)

WVNT

This variant corresponds to multiple C++ enum variants with the same value:

WV_NT = 16: Windows NT (operating system version 4.0)

WV_4_0 = 16: Operating system version 4.0, corresponds to Windows NT

WV2000

This variant corresponds to multiple C++ enum variants with the same value:

WV_2000 = 32: Windows 2000 (operating system version 5.0)

WV_5_0 = 32: Operating system version 5.0, corresponds to Windows 2000

WVXP

This variant corresponds to multiple C++ enum variants with the same value:

WV_XP = 48: Windows XP (operating system version 5.1)

WV_5_1 = 48: Operating system version 5.1, corresponds to Windows XP

WV2003

This variant corresponds to multiple C++ enum variants with the same value:

WV_2003 = 64: Windows Server 2003, Windows Server 2003 R2, Windows Home Server, Windows XP Professional x64 Edition (operating system version 5.2)

WV_5_2 = 64: Operating system version 5.2, corresponds to Windows Server 2003, Windows Server 2003 R2, Windows Home Server, and Windows XP Professional x64 Edition

WVVISTA

This variant corresponds to multiple C++ enum variants with the same value:

WV_VISTA = 128: Windows Vista, Windows Server 2008 (operating system version 6.0)

WV_6_0 = 128: Operating system version 6.0, corresponds to Windows Vista and Windows Server 2008

WVWINDOWS7

This variant corresponds to multiple C++ enum variants with the same value:

WV_WINDOWS7 = 144: Windows 7, Windows Server 2008 R2 (operating system version 6.1)

WV_6_1 = 144: Operating system version 6.1, corresponds to Windows 7 and Windows Server 2008 R2

WVWINDOWS8

This variant corresponds to multiple C++ enum variants with the same value:

WV_WINDOWS8 = 160: Windows 8 (operating system version 6.2)

WV_6_2 = 160: Operating system version 6.2, corresponds to Windows 8

WVWINDOWS81

This variant corresponds to multiple C++ enum variants with the same value:

WV_WINDOWS8_1 = 176: Windows 8.1 (operating system version 6.3), introduced in Qt 5.2

WV_6_3 = 176: Operating system version 6.3, corresponds to Windows 8.1, introduced in Qt 5.2

WVWINDOWS10

This variant corresponds to multiple C++ enum variants with the same value:

WV_WINDOWS10 = 192: Windows 10 (operating system version 10.0), introduced in Qt 5.5

WV_10_0 = 192: Operating system version 10.0, corresponds to Windows 10, introduced in Qt 5.5

WVNTBased

NT-based version of Windows (C++ enum variant: WV_NT_based = 240)

WVCE

Windows CE (C++ enum variant: WV_CE = 256)

WVCENET

Windows CE .NET (C++ enum variant: WV_CENET = 512)

WVCE5

Windows CE 5.x (C++ enum variant: WV_CE_5 = 768)

WVCE6

Windows CE 6.x (C++ enum variant: WV_CE_6 = 1024)

WVCEBased

CE-based version of Windows (C++ enum variant: WV_CE_based = 3840)