Struct qt_core::sys_info::SysInfo

pub struct SysInfo {
    // some fields omitted
}

C++ type: QSysInfo

C++ documentation:

The QSysInfo class provides information about the system.

WordSize specifies the size of a pointer for the platform on which the application is compiled.

ByteOrder specifies whether the platform is big-endian or little-endian.

WindowsVersion specifies the version of the Windows operating system on which the application is run.

MacintoshVersion specifies the version of the Macintosh operating system on which the application is run.

Some constants are defined only on certain platforms. You can use the preprocessor symbols Q_OS_WIN and Q_OS_OSX to test that the application is compiled under Windows or OS X.

Methods

impl SysInfo

fn build_abi<Args>(args: Args) -> Args::ReturnType where Args: SysInfoBuildAbiArgs

C++ method: QSysInfo::buildAbi

This is an overloaded function. Available variants:

Rust arguments:
1) fn build_abi(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn build_abi(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::buildAbi()

C++ documentation:

Returns the full architecture string that Qt was compiled for. This string is useful for identifying different, incompatible builds. For example, it can be used as an identifier to request an upgrade package from a server.

The values returned from this function are kept stable as follows: the mandatory components of the result will not change in future versions of Qt, but optional suffixes may be added.

The returned value is composed of three or more parts, separated by dashes ("-"). They are:

Component	Value

CPU Architecture	The same as QSysInfo::buildCpuArchitecture(), such as "arm", "i386", "mips" or "x86_64"

Endianness	"little_endian" or "big_endian"

Word size	Whether it's a 32- or 64-bit application. Possible values are: "llp64" (Windows 64-bit), "lp64" (Unix 64-bit), "ilp32" (32-bit)

(Optional) ABI	Zero or more components identifying different ABIs possible in this architecture. Currently, Qt has optional ABI components for ARM and MIPS processors: one component is the main ABI (such as "eabi", "o32", "n32", "o64"); another is whether the calling convention is using hardware floating point registers ("hardfloat" is present).
Additionally, if Qt was configured with -qreal float, the ABI option tag "qreal_float" will be present. If Qt was configured with another type as qreal, that type is present after "qreal_", with all characters other than letters and digits escaped by an underscore, followed by two hex digits. For example, -qreal long double becomes "qreal_long_20double".

This function was introduced in Qt 5.4.

See also QSysInfo::buildCpuArchitecture().

fn build_cpu_architecture<Args>(args: Args) -> Args::ReturnType where Args: SysInfoBuildCpuArchitectureArgs

C++ method: QSysInfo::buildCpuArchitecture

This is an overloaded function. Available variants:

Rust arguments:
1) fn build_cpu_architecture(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn build_cpu_architecture(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::buildCpuArchitecture()

C++ documentation:

Returns the architecture of the CPU that Qt was compiled for, in text format. Note that this may not match the actual CPU that the application is running on if there's an emulation layer or if the CPU supports multiple architectures (like x86-64 processors supporting i386 applications). To detect that, use currentCpuArchitecture().

Values returned by this function are stable and will not change over time, so applications can rely on the returned value as an identifier, except that new CPU types may be added over time.

Typical returned values are (note: list not exhaustive):

"arm"

"arm64"

"i386"

"ia64"

"mips"

"mips64"

"power"

"power64"

"sparc"

"sparcv9"

"x86_64"

This function was introduced in Qt 5.4.

See also QSysInfo::buildAbi() and QSysInfo::currentCpuArchitecture().

fn current_cpu_architecture<Args>(args: Args) -> Args::ReturnType where Args: SysInfoCurrentCpuArchitectureArgs

C++ method: QSysInfo::currentCpuArchitecture

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_cpu_architecture(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn current_cpu_architecture(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::currentCpuArchitecture()

C++ documentation:

Returns the architecture of the CPU that the application is running on, in text format. Note that this function depends on what the OS will report and may not detect the actual CPU architecture if the OS hides that information or is unable to provide it. For example, a 32-bit OS running on a 64-bit CPU is usually unable to determine the CPU is actually capable of running 64-bit programs.

Values returned by this function are mostly stable: an attempt will be made to ensure that they stay constant over time and match the values returned by QSysInfo::builldCpuArchitecture(). However, due to the nature of the operating system functions being used, there may be discrepancies.

Typical returned values are (note: list not exhaustive):

"arm"

"arm64"

"i386"

"ia64"

"mips"

"mips64"

"power"

"power64"

"sparc"

"sparcv9"

"x86_64"

This function was introduced in Qt 5.4.

See also QSysInfo::buildAbi() and QSysInfo::buildCpuArchitecture().

fn kernel_type<Args>(args: Args) -> Args::ReturnType where Args: SysInfoKernelTypeArgs

C++ method: QSysInfo::kernelType

This is an overloaded function. Available variants:

Rust arguments:
1) fn kernel_type(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn kernel_type(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::kernelType()

C++ documentation:

Returns the type of the operating system kernel Qt was compiled for. It's also the kernel the application is running on, unless the host operating system is running a form of compatibility or virtualization layer.

Values returned by this function are stable and will not change over time, so applications can rely on the returned value as an identifier, except that new OS kernel types may be added over time.

On Windows, this function returns the type of Windows kernel, like "wince" or "winnt". On Unix systems, it returns the same as the output of uname -s (lowercased).

Note that this function may return surprising values: it returns "linux" for all operating systems running Linux (including Android), "qnx" for all operating systems running QNX (including BlackBerry 10), "freebsd" for Debian/kFreeBSD, and "darwin" for OS X and iOS. For information on the type of product the application is running on, see productType().

This function was introduced in Qt 5.4.

See also QFileSelector, kernelVersion(), productType(), productVersion(), and prettyProductName().

fn kernel_version<Args>(args: Args) -> Args::ReturnType where Args: SysInfoKernelVersionArgs

C++ method: QSysInfo::kernelVersion

This is an overloaded function. Available variants:

Rust arguments:
1) fn kernel_version(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn kernel_version(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::kernelVersion()

C++ documentation:

Returns the release version of the operating system kernel. On Windows, it returns the version of the NT or CE kernel. On Unix systems, including Android, BlackBerry and OS X, it returns the same as the uname -r command would return.

If the version could not be determined, this function may return an empty string.

This function was introduced in Qt 5.4.

See also kernelType(), productType(), productVersion(), and prettyProductName().

fn mac_version() -> MacVersion

C++ method: static QSysInfo::MacVersion QSysInfo::macVersion()

C++ documentation:

Returns the version of Darwin (OS X or iOS) on which the application is run, or MV_None if the operating system is not a version of Darwin.

fn pretty_product_name<Args>(args: Args) -> Args::ReturnType where Args: SysInfoPrettyProductNameArgs

C++ method: QSysInfo::prettyProductName

This is an overloaded function. Available variants:

Rust arguments:
1) fn pretty_product_name(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn pretty_product_name(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::prettyProductName()

C++ documentation:

Returns a prettier form of productType() and productVersion(), containing other tokens like the operating system type, codenames and other information. The result of this function is suitable for displaying to the user, but not for long-term storage, as the string may change with updates to Qt.

If productType() is "unknown", this function will instead use the kernelType() and kernelVersion() functions.

This function was introduced in Qt 5.4.

See also kernelType(), kernelVersion(), productType(), and productVersion().

fn product_type<Args>(args: Args) -> Args::ReturnType where Args: SysInfoProductTypeArgs

C++ method: QSysInfo::productType

This is an overloaded function. Available variants:

Rust arguments:
1) fn product_type(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn product_type(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::productType()

C++ documentation:

Returns the product name of the operating system this application is running in. If the application is running on some sort of emulation or virtualization layer (such as WINE on a Unix system), this function will inspect the emulation / virtualization layer.

Values returned by this function are stable and will not change over time, so applications can rely on the returned value as an identifier, except that new OS types may be added over time.

Linux and Android note: this function returns "android" for Linux systems running Android userspace, notably when using the Bionic library. For all other Linux systems, regardless of C library being used, it tries to determine the distribution name and returns that. If determining the distribution name failed, it returns "unknown".

BlackBerry note: this function returns "blackberry" for QNX systems running the BlackBerry userspace, but "qnx" for all other QNX-based systems.

Darwin, OS X and iOS note: this function returns "osx" for OS X systems, "ios" for iOS systems and "darwin" in case the system could not be determined.

FreeBSD note: this function returns "debian" for Debian/kFreeBSD and "unknown" otherwise.

Windows note: this function returns "winphone" for builds for Windows Phone, "winrt" for WinRT builds, "wince" for Windows CE and Embedded Compact builds, and "windows" for normal desktop builds.

For other Unix-type systems, this function usually returns "unknown".

This function was introduced in Qt 5.4.

See also QFileSelector, kernelType(), kernelVersion(), productVersion(), and prettyProductName().

fn product_version<Args>(args: Args) -> Args::ReturnType where Args: SysInfoProductVersionArgs

C++ method: QSysInfo::productVersion

This is an overloaded function. Available variants:

Rust arguments:
1) fn product_version(cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn product_version(cpp_utils::AsStruct) -> ::string::String
C++ method: static QString QSysInfo::productVersion()

C++ documentation:

Returns the product version of the operating system in string form. If the version could not be determined, this function returns "unknown".

It will return the Android, BlackBerry, iOS, OS X, Windows full-product versions on those systems. In particular, on OS X, iOS and Windows, the returned string is similar to the macVersion() or windowsVersion() enums.

On Linux systems, it will try to determine the distribution version and will return that. This is also done on Debian/kFreeBSD, so this function will return Debian version in that case.

In all other Unix-type systems, this function always returns "unknown".

Note: The version string returned from this function is only guaranteed to be orderable on Android, BlackBerry, OS X and iOS. On Windows, some Windows versions are text ("XP" and "Vista", for example). On Linux, the version of the distribution may jump unexpectedly, please refer to the distribution's documentation for versioning practices.

This function was introduced in Qt 5.4.

See also kernelType(), kernelVersion(), productType(), and prettyProductName().

fn windows_version() -> WinVersion

C++ method: static QSysInfo::WinVersion QSysInfo::windowsVersion()

C++ documentation:

Returns the version of the Windows operating system on which the application is run, or WV_None if the operating system is not Windows.

This function was introduced in Qt 4.4.