Struct qt_core::system_semaphore::SystemSemaphore

pub struct SystemSemaphore {
    // some fields omitted
}

C++ type: QSystemSemaphore

C++ documentation:

The QSystemSemaphore class provides a general counting system semaphore.

A semaphore is a generalization of a mutex. While a mutex can be locked only once, a semaphore can be acquired multiple times. Typically, a semaphore is used to protect a certain number of identical resources.

Like its lighter counterpart QSemaphore, a QSystemSemaphore can be accessed from multiple threads. Unlike QSemaphore, a QSystemSemaphore can also be accessed from multiple processes. This means QSystemSemaphore is a much heavier class, so if your application doesn't need to access your semaphores across multiple processes, you will probably want to use QSemaphore.

Semaphores support two fundamental operations, acquire() and release():

acquire() tries to acquire one resource. If there isn't a resource available, the call blocks until a resource becomes available. Then the resource is acquired and the call returns.

release() releases one resource so it can be acquired by another process. The function can also be called with a parameter n > 1, which releases n resources.

A system semaphore is created with a string key that other processes can use to use the same semaphore.

Example: Create a system semaphore

QSystemSemaphore sem("market", 3, QSystemSemaphore::Create);
                             // resources available == 3
sem.acquire();               // resources available == 2
sem.acquire();               // resources available == 1
sem.acquire();               // resources available == 0
sem.release();               // resources available == 1
sem.release(2);              // resources available == 3

A typical application of system semaphores is for controlling access to a circular buffer shared by a producer process and a consumer processes.

Methods

impl SystemSemaphore

fn acquire(&mut self) -> bool

C++ method: bool QSystemSemaphore::acquire()

C++ documentation:

Acquires one of the resources guarded by this semaphore, if there is one available, and returns true. If all the resources guarded by this semaphore have already been acquired, the call blocks until one of them is released by another process or thread having a semaphore with the same key.

If false is returned, a system error has occurred. Call error() to get a value of QSystemSemaphore::SystemSemaphoreError that indicates which error occurred.

See also release().

fn error(&self) -> SystemSemaphoreError

C++ method: QSystemSemaphore::SystemSemaphoreError QSystemSemaphore::error() const

C++ documentation:

Returns a value indicating whether an error occurred, and, if so, which error it was.

See also errorString().

fn error_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SystemSemaphoreErrorStringArgs<'largs>

C++ method: QSystemSemaphore::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn error_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QSystemSemaphore::errorString() const

C++ documentation:

Returns a text description of the last error that occurred. If error() returns an error value, call this function to get a text string that describes the error.

See also error().

fn key<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: SystemSemaphoreKeyArgs<'largs>

C++ method: QSystemSemaphore::key

This is an overloaded function. Available variants:

Rust arguments:
1) fn key(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn key(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QSystemSemaphore::key() const

C++ documentation:

Returns the key assigned to this system semaphore. The key is the name by which the semaphore can be accessed from other processes.

See also setKey().

fn new<Args>(args: Args) -> Args::ReturnType where Args: SystemSemaphoreNewArgs

C++ method: QSystemSemaphore::QSystemSemaphore

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::system_semaphore::SystemSemaphore
2) fn new((&::string::String, libc::c_int, cpp_utils::AsStruct)) -> ::system_semaphore::SystemSemaphore
3) fn new((&::string::String, libc::c_int, ::system_semaphore::AccessMode, cpp_utils::AsStruct)) -> ::system_semaphore::SystemSemaphore
4) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::system_semaphore::SystemSemaphore>
5) fn new((&::string::String, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::system_semaphore::SystemSemaphore>
6) fn new((&::string::String, libc::c_int, ::system_semaphore::AccessMode, cpp_utils::AsBox)) -> cpp_utils::CppBox<::system_semaphore::SystemSemaphore>
C++ method: [constructor] void QSystemSemaphore::QSystemSemaphore(const QString& key, int initialValue = ?, QSystemSemaphore::AccessMode mode = ?)

fn release<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: SystemSemaphoreReleaseArgs<'largs>

C++ method: QSystemSemaphore::release

This is an overloaded function. Available variants:

Rust arguments:
1) fn release(&mut self, libc::c_int) -> bool
2) fn release(&mut self, ()) -> bool
C++ method: bool QSystemSemaphore::release(int n = ?)

C++ documentation:

Releases n resources guarded by the semaphore. Returns true unless there is a system error.

Example: Create a system semaphore having five resources; acquire them all and then release them all.

QSystemSemaphore sem("market", 5, QSystemSemaphore::Create);
for (int i = 0; i < 5; ++i)  // acquire all 5 resources
    sem.acquire();
sem.release(5);              // release the 5 resources

This function can also "create" resources. For example, immediately following the sequence of statements above, suppose we add the statement:

sem.release(10);          // "create" 10 new resources

Ten new resources are now guarded by the semaphore, in addition to the five that already existed. You would not normally use this function to create more resources.

See also acquire().

fn set_key<'largs, Args>(&'largs mut self, args: Args) where Args: SystemSemaphoreSetKeyArgs<'largs>

C++ method: QSystemSemaphore::setKey

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_key(&mut self, &::string::String) -> ()
2) fn set_key(&mut self, (&::string::String, libc::c_int)) -> ()
3) fn set_key(&mut self, (&::string::String, libc::c_int, ::system_semaphore::AccessMode)) -> ()
C++ method: void QSystemSemaphore::setKey(const QString& key, int initialValue = ?, QSystemSemaphore::AccessMode mode = ?)

C++ documentation:

This function works the same as the constructor. It reconstructs this QSystemSemaphore object. If the new key is different from the old key, calling this function is like calling the destructor of the semaphore with the old key, then calling the constructor to create a new semaphore with the new key. The initialValue and mode parameters are as defined for the constructor.

See also QSystemSemaphore() and key().