Struct qt_core::temporary_dir::TemporaryDir

pub struct TemporaryDir {
    // some fields omitted
}

C++ type: QTemporaryDir

C++ documentation:

The QTemporaryDir class creates a unique directory for temporary use.

QTemporaryDir is used to create unique temporary dirs safely. The dir itself is created by the constructor. The name of the temporary directory is guaranteed to be unique (i.e., you are guaranteed to not overwrite an existing dir), and the directory will subsequently be removed upon destruction of the QTemporaryDir object. The directory name is either auto-generated, or created based on a template, which is passed to QTemporaryDir's constructor.

Example:

    // Within a function/method...

QTemporaryDir dir; if (dir.isValid()) { // dir.path() returns the unique directory path }

// The QTemporaryDir destructor removes the temporary directory // as it goes out of scope.
It is very important to test that the temporary directory could be created, using isValid(). Do not use exists(), since a default-constructed QDir represents the current directory, which exists.

The path to the temporary dir can be found by calling path().

A temporary directory will have some static part of the name and some part that is calculated to be unique. The default path will be determined from QCoreApplication::applicationName() (otherwise qt_temp) and will be placed into the temporary path as returned by QDir::tempPath(). If you specify your own path, a relative path will not be placed in the temporary directory by default, but be relative to the current working directory. In all cases, a random string will be appended to the path in order to make it unique.

Methods

impl TemporaryDir

fn auto_remove(&self) -> bool

C++ method: bool QTemporaryDir::autoRemove() const

C++ documentation:

Returns true if the QTemporaryDir is in auto remove mode. Auto-remove mode will automatically delete the directory from disk upon destruction. This makes it very easy to create your QTemporaryDir object on the stack, fill it with files, do something with the files, and finally on function return it will automatically clean up after itself.

Auto-remove is on by default.

See also setAutoRemove() and remove().

fn is_valid(&self) -> bool

C++ method: bool QTemporaryDir::isValid() const

C++ documentation:

Returns true if the QTemporaryDir was created successfully.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TemporaryDirNewArgs

C++ method: QTemporaryDir::QTemporaryDir

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::temporary_dir::TemporaryDir
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::temporary_dir::TemporaryDir>
C++ method: [constructor] void QTemporaryDir::QTemporaryDir()

Variant 2

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::temporary_dir::TemporaryDir
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::temporary_dir::TemporaryDir>
C++ method: [constructor] void QTemporaryDir::QTemporaryDir(const QString& templateName)

fn path<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TemporaryDirPathArgs<'largs>

C++ method: QTemporaryDir::path

This is an overloaded function. Available variants:

Rust arguments:
1) fn path(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn path(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QTemporaryDir::path() const

C++ documentation:

Returns the path to the temporary directory. Empty if the QTemporaryDir could not be created.

fn remove(&mut self) -> bool

C++ method: bool QTemporaryDir::remove()

C++ documentation:

Removes the temporary directory, including all its contents.

Returns true if removing was successful.

fn set_auto_remove(&mut self, b: bool)

C++ method: void QTemporaryDir::setAutoRemove(bool b)

C++ documentation:

Sets the QTemporaryDir into auto-remove mode if b is true.

Auto-remove is on by default.

See also autoRemove() and remove().