Struct qt_core::text_boundary_finder::TextBoundaryFinder

pub struct TextBoundaryFinder {
    // some fields omitted
}

C++ type: QTextBoundaryFinder

C++ documentation:

The QTextBoundaryFinder class provides a way of finding Unicode text boundaries in a string.

QTextBoundaryFinder allows to find Unicode text boundaries in a string, accordingly to the Unicode text boundary specification (see Unicode Standard Annex #14 and Unicode Standard Annex #29).

QTextBoundaryFinder can operate on a QString in four possible modes depending on the value of BoundaryType.

Units of Unicode characters that make up what the user thinks of as a character or basic unit of the language are here called Grapheme clusters. The two unicode characters 'A' + diaeresis do for example form one grapheme cluster as the user thinks of them as one character, yet it is in this case represented by two unicode code points (see http://www.unicode.org/reports/tr29/#Grapheme_Cluster_Boundaries).

Word boundaries are there to locate the start and end of what a language considers to be a word (see http://www.unicode.org/reports/tr29/#Word_Boundaries).

Line break boundaries give possible places where a line break might happen and sentence boundaries will show the beginning and end of whole sentences (see http://www.unicode.org/reports/tr29/#Sentence_Boundaries and http://www.unicode.org/reports/tr14/).

The first position in a string is always a valid boundary and refers to the position before the first character. The last position at the length of the string is also valid and refers to the position after the last character.

Methods

impl TextBoundaryFinder

fn boundary_reasons(&self) -> Flags<BoundaryReason>

C++ method: QFlags<QTextBoundaryFinder::BoundaryReason> QTextBoundaryFinder::boundaryReasons() const

C++ documentation:

Returns the reasons for the boundary finder to have chosen the current position as a boundary.

fn is_at_boundary(&self) -> bool

C++ method: bool QTextBoundaryFinder::isAtBoundary() const

C++ documentation:

Returns true if the object's position() is currently at a valid text boundary.

fn is_valid(&self) -> bool

C++ method: bool QTextBoundaryFinder::isValid() const

C++ documentation:

Returns true if the text boundary finder is valid; otherwise returns false. A default QTextBoundaryFinder is invalid.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextBoundaryFinderNewArgs

C++ method: QTextBoundaryFinder::QTextBoundaryFinder

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_boundary_finder::TextBoundaryFinder
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_boundary_finder::TextBoundaryFinder>
C++ method: [constructor] void QTextBoundaryFinder::QTextBoundaryFinder()

Variant 2

Rust arguments:
1) fn new((::text_boundary_finder::BoundaryType, *const ::char::Char, libc::c_int, cpp_utils::AsStruct)) -> ::text_boundary_finder::TextBoundaryFinder
2) fn new((::text_boundary_finder::BoundaryType, *const ::char::Char, libc::c_int, *mut libc::c_uchar, cpp_utils::AsStruct)) -> ::text_boundary_finder::TextBoundaryFinder
3) fn new((::text_boundary_finder::BoundaryType, *const ::char::Char, libc::c_int, *mut libc::c_uchar, libc::c_int, cpp_utils::AsStruct)) -> ::text_boundary_finder::TextBoundaryFinder
4) fn new((::text_boundary_finder::BoundaryType, *const ::char::Char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_boundary_finder::TextBoundaryFinder>
5) fn new((::text_boundary_finder::BoundaryType, *const ::char::Char, libc::c_int, *mut libc::c_uchar, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_boundary_finder::TextBoundaryFinder>
6) fn new((::text_boundary_finder::BoundaryType, *const ::char::Char, libc::c_int, *mut libc::c_uchar, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_boundary_finder::TextBoundaryFinder>
C++ method: [constructor] void QTextBoundaryFinder::QTextBoundaryFinder(QTextBoundaryFinder::BoundaryType type, const QChar* chars, int length, unsigned char* buffer = ?, int bufferSize = ?)

Variant 3

Rust arguments:
1) fn new((::text_boundary_finder::BoundaryType, &::string::String, cpp_utils::AsStruct)) -> ::text_boundary_finder::TextBoundaryFinder
2) fn new((::text_boundary_finder::BoundaryType, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_boundary_finder::TextBoundaryFinder>
C++ method: [constructor] void QTextBoundaryFinder::QTextBoundaryFinder(QTextBoundaryFinder::BoundaryType type, const QString& string)

Variant 4

Rust arguments:
1) fn new((&::text_boundary_finder::TextBoundaryFinder, cpp_utils::AsStruct)) -> ::text_boundary_finder::TextBoundaryFinder
2) fn new((&::text_boundary_finder::TextBoundaryFinder, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_boundary_finder::TextBoundaryFinder>
C++ method: [constructor] void QTextBoundaryFinder::QTextBoundaryFinder(const QTextBoundaryFinder& other)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 TextBoundaryFinder) -> &'l0 mut TextBoundaryFinder

C++ method: QTextBoundaryFinder& QTextBoundaryFinder::operator=(const QTextBoundaryFinder& other)

C++ documentation:

Assigns the object, other, to another QTextBoundaryFinder object.

fn position(&self) -> c_int

C++ method: int QTextBoundaryFinder::position() const

C++ documentation:

Returns the current position of the QTextBoundaryFinder.

The range is from 0 (the beginning of the string) to the length of the string inclusive.

See also setPosition().

fn set_position(&mut self, position: c_int)

C++ method: void QTextBoundaryFinder::setPosition(int position)

C++ documentation:

Sets the current position of the QTextBoundaryFinder to position.

If position is out of bounds, it will be bound to only valid positions. In this case, valid positions are from 0 to the length of the string inclusive.

See also position().

fn string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextBoundaryFinderStringArgs<'largs>

C++ method: QTextBoundaryFinder::string

This is an overloaded function. Available variants:

Rust arguments:
1) fn string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QTextBoundaryFinder::string() const

C++ documentation:

Returns the string the QTextBoundaryFinder object operates on.

fn to_end(&mut self)

C++ method: void QTextBoundaryFinder::toEnd()

C++ documentation:

Moves the finder to the end of the string. This is equivalent to setPosition(string.length()).

See also setPosition() and position().

fn to_next_boundary(&mut self) -> c_int

C++ method: int QTextBoundaryFinder::toNextBoundary()

C++ documentation:

Moves the QTextBoundaryFinder to the next boundary position and returns that position.

Returns -1 if there is no next boundary.

fn to_previous_boundary(&mut self) -> c_int

C++ method: int QTextBoundaryFinder::toPreviousBoundary()

C++ documentation:

Moves the QTextBoundaryFinder to the previous boundary position and returns that position.

Returns -1 if there is no previous boundary.

fn to_start(&mut self)

C++ method: void QTextBoundaryFinder::toStart()

C++ documentation:

Moves the finder to the start of the string. This is equivalent to setPosition(0).

See also setPosition() and position().

fn type_(&self) -> BoundaryType

C++ method: QTextBoundaryFinder::BoundaryType QTextBoundaryFinder::type() const

C++ documentation:

Returns the type of the QTextBoundaryFinder.