Module qt_core::text_codec::overloading

Traits

ConverterStateNewArgs
This trait represents a set of arguments accepted by ConverterState::new method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

TextCodecAliasesArgs
This trait represents a set of arguments accepted by TextCodec::aliases method.

TextCodecAvailableCodecsArgs
This trait represents a set of arguments accepted by TextCodec::available_codecs method.

TextCodecAvailableMibsArgs
This trait represents a set of arguments accepted by TextCodec::available_mibs method.

TextCodecCanEncodeArgs
This trait represents a set of arguments accepted by TextCodec::can_encode method.

TextCodecCodecForHtmlArgs
This trait represents a set of arguments accepted by TextCodec::codec_for_html method.

TextCodecCodecForNameArgs
This trait represents a set of arguments accepted by TextCodec::codec_for_name method.

TextCodecCodecForUtfTextArgs
This trait represents a set of arguments accepted by TextCodec::codec_for_utf_text method.

TextCodecFromUnicodeArgs
This trait represents a set of arguments accepted by TextCodec::from_unicode method.

TextCodecMakeDecoderArgs
This trait represents a set of arguments accepted by TextCodec::make_decoder method.

TextCodecMakeEncoderArgs
This trait represents a set of arguments accepted by TextCodec::make_encoder method.

TextCodecNameArgs
This trait represents a set of arguments accepted by TextCodec::name method.

TextCodecToUnicodeArgs
This trait represents a set of arguments accepted by TextCodec::to_unicode method.