Struct qt_core::text_codec::TextCodec

pub struct TextCodec {
    // some fields omitted
}

C++ type: QTextCodec

C++ documentation:

The QTextCodec class provides conversions between text encodings.

Qt uses Unicode to store, draw and manipulate strings. In many situations you may wish to deal with data that uses a different encoding. For example, most Japanese documents are still stored in Shift-JIS or ISO 2022-JP, while Russian users often have their documents in KOI8-R or Windows-1251.

Qt provides a set of QTextCodec classes to help with converting non-Unicode formats to and from Unicode. You can also create your own codec classes.

The supported encodings are:

Big5

Big5-HKSCS

CP949

EUC-JP

EUC-KR

GB18030

HP-ROMAN8

IBM 850

IBM 866

IBM 874

ISO 2022-JP

ISO 8859-1 to 10

ISO 8859-13 to 16

Iscii-Bng, Dev, Gjr, Knd, Mlm, Ori, Pnj, Tlg, and Tml

KOI8-R

KOI8-U

Macintosh

Shift-JIS

TIS-620

TSCII

UTF-8

UTF-16

UTF-16BE

UTF-16LE

UTF-32

UTF-32BE

UTF-32LE

Windows-1250 to 1258

If Qt is compiled with ICU support enabled, most codecs supported by ICU will also be available to the application.

QTextCodecs can be used as follows to convert some locally encoded string to Unicode. Suppose you have some string encoded in Russian KOI8-R encoding, and want to convert it to Unicode. The simple way to do it is like this:

QByteArray encodedString = "...";
QTextCodec codec = QTextCodec::codecForName("KOI8-R");
QString string = codec->toUnicode(encodedString);

After this, string holds the text converted to Unicode. Converting a string from Unicode to the local encoding is just as easy:

QString string = "...";
QTextCodec codec = QTextCodec::codecForName("KOI8-R");
QByteArray encodedString = codec->fromUnicode(string);

To read or write files in various encodings, use QTextStream and its setCodec() function. See the Codecs example for an application of QTextCodec to file I/O.

Some care must be taken when trying to convert the data in chunks, for example, when receiving it over a network. In such cases it is possible that a multi-byte character will be split over two chunks. At best this might result in the loss of a character and at worst cause the entire conversion to fail.

The approach to use in these situations is to create a QTextDecoder object for the codec and use this QTextDecoder for the whole decoding process, as shown below:

QTextCodec codec = QTextCodec::codecForName("Shift-JIS");
QTextDecoder decoder = codec->makeDecoder();

QString string; while (new_data_available()) { QByteArray chunk = get_new_data(); string += decoder->toUnicode(chunk); } delete decoder;
The QTextDecoder object maintains state between chunks and therefore works correctly even if a multi-byte character is split between chunks.

Methods

impl TextCodec

fn aliases<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCodecAliasesArgs<'largs>

C++ method: QTextCodec::aliases

This is an overloaded function. Available variants:

Rust arguments:
1) fn aliases(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn aliases(&self, cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: virtual QList<QByteArray> QTextCodec::aliases() const

C++ documentation:

Subclasses can return a number of aliases for the codec in question.

Standard aliases for codecs can be found in the IANA character-sets encoding file.

fn available_codecs<Args>(args: Args) -> Args::ReturnType where Args: TextCodecAvailableCodecsArgs

C++ method: QTextCodec::availableCodecs

This is an overloaded function. Available variants:

Rust arguments:
1) fn available_codecs(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn available_codecs(cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: static QList<QByteArray> QTextCodec::availableCodecs()

C++ documentation:

Returns the list of all available codecs, by name. Call QTextCodec::codecForName() to obtain the QTextCodec for the name.

The list may contain many mentions of the same codec if the codec has aliases.

Note: This function is thread-safe.

See also availableMibs(), name(), and aliases().

fn available_mibs<Args>(args: Args) -> Args::ReturnType where Args: TextCodecAvailableMibsArgs

C++ method: QTextCodec::availableMibs

This is an overloaded function. Available variants:

Rust arguments:
1) fn available_mibs(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListCInt>
2) fn available_mibs(cpp_utils::AsStruct) -> ::list::ListCInt
C++ method: static QList<int> QTextCodec::availableMibs()

C++ documentation:

Returns the list of MIBs for all available codecs. Call QTextCodec::codecForMib() to obtain the QTextCodec for the MIB.

Note: This function is thread-safe.

See also availableCodecs() and mibEnum().

fn can_encode<'largs, Args>(&'largs self, args: Args) -> bool where Args: TextCodecCanEncodeArgs<'largs>

C++ method: QTextCodec::canEncode

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn can_encode(&self, &::char::Char) -> bool
C++ method: bool QTextCodec::canEncode(QChar arg1) const

C++ documentation:

Returns true if the Unicode character ch can be fully encoded with this codec; otherwise returns false.

Variant 2

Rust arguments: fn can_encode(&self, &::string::String) -> bool
C++ method: bool QTextCodec::canEncode(const QString& arg1) const

C++ documentation:

This is an overloaded function.

s contains the string being tested for encode-ability.

fn codec_for_html<Args>(args: Args) -> *mut TextCodec where Args: TextCodecCodecForHtmlArgs

C++ method: QTextCodec::codecForHtml

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn codec_for_html(&::byte_array::ByteArray) -> *mut ::text_codec::TextCodec
C++ method: static QTextCodec* QTextCodec::codecForHtml(const QByteArray& ba)

C++ documentation:

This is an overloaded function.

Tries to detect the encoding of the provided snippet of HTML in the given byte array, ba, by checking the BOM (Byte Order Mark) and the content-type meta header and returns a QTextCodec instance that is capable of decoding the html to unicode. If the codec cannot be detected, this overload returns a Latin-1 QTextCodec.

Variant 2

Rust arguments: fn codec_for_html((&::byte_array::ByteArray, *mut ::text_codec::TextCodec)) -> *mut ::text_codec::TextCodec
C++ method: static QTextCodec* QTextCodec::codecForHtml(const QByteArray& ba, QTextCodec* defaultCodec)

C++ documentation:

Tries to detect the encoding of the provided snippet of HTML in the given byte array, ba, by checking the BOM (Byte Order Mark) and the content-type meta header and returns a QTextCodec instance that is capable of decoding the html to unicode. If the codec cannot be detected from the content provided, defaultCodec is returned.

This function was introduced in Qt 4.4.

See also codecForUtfText().

fn codec_for_locale() -> *mut TextCodec

C++ method: static QTextCodec* QTextCodec::codecForLocale()

C++ documentation:

Returns a pointer to the codec most suitable for this locale.

On Windows, the codec will be based on a system locale. On Unix systems, the codec will might fall back to using the iconv library if no builtin codec for the locale can be found.

Note that in these cases the codec's name will be "System".

Note: This function is thread-safe.

See also setCodecForLocale().

fn codec_for_mib(mib: c_int) -> *mut TextCodec

C++ method: static QTextCodec* QTextCodec::codecForMib(int mib)

C++ documentation:

Returns the QTextCodec which matches the MIBenum mib.

Note: This function is thread-safe.

fn codec_for_name<Args>(args: Args) -> *mut TextCodec where Args: TextCodecCodecForNameArgs

C++ method: QTextCodec::codecForName

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn codec_for_name(&::byte_array::ByteArray) -> *mut ::text_codec::TextCodec
C++ method: static QTextCodec* QTextCodec::codecForName(const QByteArray& name)

C++ documentation:

Searches all installed QTextCodec objects and returns the one which best matches name; the match is case-insensitive. Returns 0 if no codec matching the name name could be found.

Note: This function is thread-safe.

Variant 2

Rust arguments: fn codec_for_name(*const libc::c_char) -> *mut ::text_codec::TextCodec
C++ method: static QTextCodec* QTextCodec::codecForName(const char* name)

C++ documentation:

Searches all installed QTextCodec objects and returns the one which best matches name; the match is case-insensitive. Returns 0 if no codec matching the name name could be found.

fn codec_for_utf_text<Args>(args: Args) -> *mut TextCodec where Args: TextCodecCodecForUtfTextArgs

C++ method: QTextCodec::codecForUtfText

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn codec_for_utf_text(&::byte_array::ByteArray) -> *mut ::text_codec::TextCodec
C++ method: static QTextCodec* QTextCodec::codecForUtfText(const QByteArray& ba)

C++ documentation:

This is an overloaded function.

Tries to detect the encoding of the provided snippet ba by using the BOM (Byte Order Mark) and returns a QTextCodec instance that is capable of decoding the text to unicode. If the codec cannot be detected, this overload returns a Latin-1 QTextCodec.

See also codecForHtml().

Variant 2

Rust arguments: fn codec_for_utf_text((&::byte_array::ByteArray, *mut ::text_codec::TextCodec)) -> *mut ::text_codec::TextCodec
C++ method: static QTextCodec* QTextCodec::codecForUtfText(const QByteArray& ba, QTextCodec* defaultCodec)

C++ documentation:

Tries to detect the encoding of the provided snippet ba by using the BOM (Byte Order Mark) and returns a QTextCodec instance that is capable of decoding the text to unicode. If the codec cannot be detected from the content provided, defaultCodec is returned.

This function was introduced in Qt 4.6.

See also codecForHtml().

fn from_unicode<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCodecFromUnicodeArgs<'largs>

C++ method: QTextCodec::fromUnicode

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_unicode(&self, (*const ::char::Char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn from_unicode(&self, (*const ::char::Char, libc::c_int, *mut ::text_codec::ConverterState, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
3) fn from_unicode(&self, (*const ::char::Char, libc::c_int, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
4) fn from_unicode(&self, (*const ::char::Char, libc::c_int, *mut ::text_codec::ConverterState, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray QTextCodec::fromUnicode(const QChar* in, int length, QTextCodec::ConverterState* state = ?) const

C++ documentation:

Converts the first number of characters from the input array from Unicode to the encoding of this codec, and returns the result in a QByteArray.

The state of the convertor used is updated.

Variant 2

Rust arguments:
1) fn from_unicode(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn from_unicode(&self, (&::string::String, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray QTextCodec::fromUnicode(const QString& uc) const

C++ documentation:

Converts str from Unicode to the encoding of this codec, and returns the result in a QByteArray.

fn make_decoder<'largs, Args>(&'largs self, args: Args) -> *mut TextDecoder where Args: TextCodecMakeDecoderArgs<'largs>

C++ method: QTextCodec::makeDecoder

This is an overloaded function. Available variants:

Rust arguments:
1) fn make_decoder(&self, ::flags::Flags<::text_codec::ConversionFlag>) -> *mut ::text_decoder::TextDecoder
2) fn make_decoder(&self, ()) -> *mut ::text_decoder::TextDecoder
C++ method: QTextDecoder* QTextCodec::makeDecoder(QFlags<QTextCodec::ConversionFlag> flags = ?) const

C++ documentation:

Creates a QTextDecoder with a specified flags to decode chunks of char * data to create chunks of Unicode data.

The caller is responsible for deleting the returned object.

This function was introduced in Qt 4.7.

fn make_encoder<'largs, Args>(&'largs self, args: Args) -> *mut TextEncoder where Args: TextCodecMakeEncoderArgs<'largs>

C++ method: QTextCodec::makeEncoder

This is an overloaded function. Available variants:

Rust arguments:
1) fn make_encoder(&self, ::flags::Flags<::text_codec::ConversionFlag>) -> *mut ::text_encoder::TextEncoder
2) fn make_encoder(&self, ()) -> *mut ::text_encoder::TextEncoder
C++ method: QTextEncoder* QTextCodec::makeEncoder(QFlags<QTextCodec::ConversionFlag> flags = ?) const

C++ documentation:

Creates a QTextEncoder with a specified flags to encode chunks of Unicode data as char * data.

The caller is responsible for deleting the returned object.

This function was introduced in Qt 4.7.

fn mib_enum(&self) -> c_int

C++ method: pure virtual int QTextCodec::mibEnum() const

C++ documentation:

Subclasses of QTextCodec must reimplement this function. It returns the MIBenum (see IANA character-sets encoding file for more information). It is important that each QTextCodec subclass returns the correct unique value for this function.

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCodecNameArgs<'largs>

C++ method: QTextCodec::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn name(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: pure virtual QByteArray QTextCodec::name() const

C++ documentation:

QTextCodec subclasses must reimplement this function. It returns the name of the encoding supported by the subclass.

If the codec is registered as a character set in the IANA character-sets encoding file this method should return the preferred mime name for the codec if defined, otherwise its name.

fn set_codec_for_locale(c: *mut TextCodec)

C++ method: static void QTextCodec::setCodecForLocale(QTextCodec* c)

C++ documentation:

Set the codec to c; this will be returned by codecForLocale(). If c is a null pointer, the codec is reset to the default.

This might be needed for some applications that want to use their own mechanism for setting the locale.

Warning: This function is not reentrant.

See also codecForLocale().

fn to_unicode<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextCodecToUnicodeArgs<'largs>

C++ method: QTextCodec::toUnicode

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_unicode(&self, (&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_unicode(&self, (&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTextCodec::toUnicode(const QByteArray& arg1) const

C++ documentation:

Converts a from the encoding of this codec to Unicode, and returns the result in a QString.

Variant 2

Rust arguments:
1) fn to_unicode(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_unicode(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTextCodec::toUnicode(const char* chars) const

C++ documentation:

This is an overloaded function.

chars contains the source characters.

Variant 3

Rust arguments:
1) fn to_unicode(&self, (*const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_unicode(&self, (*const libc::c_char, libc::c_int, *mut ::text_codec::ConverterState, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn to_unicode(&self, (*const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
4) fn to_unicode(&self, (*const libc::c_char, libc::c_int, *mut ::text_codec::ConverterState, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTextCodec::toUnicode(const char* in, int length, QTextCodec::ConverterState* state = ?) const

C++ documentation:

Converts the first size characters from the input from the encoding of this codec to Unicode, and returns the result in a QString.

The state of the convertor used is updated.