Struct qt_core::text_decoder::TextDecoder

pub struct TextDecoder {
    // some fields omitted
}

C++ type: QTextDecoder

C++ documentation:

The QTextDecoder class provides a state-based decoder.

A text decoder converts text from an encoded text format into Unicode using a specific codec.

The decoder converts text in this format into Unicode, remembering any state that is required between calls.

Methods

impl TextDecoder

fn has_failure(&self) -> bool

C++ method: bool QTextDecoder::hasFailure() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextDecoderNewArgs

C++ method: QTextDecoder::QTextDecoder

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*const ::text_codec::TextCodec, cpp_utils::AsStruct)) -> ::text_decoder::TextDecoder
2) fn new((*const ::text_codec::TextCodec, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_decoder::TextDecoder>
C++ method: [constructor] void QTextDecoder::QTextDecoder(const QTextCodec* codec)

Variant 2

Rust arguments:
1) fn new((*const ::text_codec::TextCodec, ::flags::Flags<::text_codec::ConversionFlag>, cpp_utils::AsStruct)) -> ::text_decoder::TextDecoder
2) fn new((*const ::text_codec::TextCodec, ::flags::Flags<::text_codec::ConversionFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_decoder::TextDecoder>
C++ method: [constructor] void QTextDecoder::QTextDecoder(const QTextCodec* codec, QFlags<QTextCodec::ConversionFlag> flags)

fn to_unicode<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TextDecoderToUnicodeArgs<'largs>

C++ method: QTextDecoder::toUnicode

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_unicode(&mut self, (&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_unicode(&mut self, (&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTextDecoder::toUnicode(const QByteArray& ba)

C++ documentation:

This is an overloaded function.

Converts the bytes in the byte array specified by ba to Unicode and returns the result.

Variant 2

Rust arguments:
1) fn to_unicode(&mut self, (*const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_unicode(&mut self, (*const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTextDecoder::toUnicode(const char* chars, int len)

C++ documentation:

Converts the first len bytes in chars to Unicode, returning the result.

If not all characters are used (e.g. if only part of a multi-byte encoding is at the end of the characters), the decoder remembers enough state to continue with the next call to this function.

Variant 3

Rust arguments: fn to_unicode(&mut self, (*mut ::string::String, *const libc::c_char, libc::c_int)) -> ()
C++ method: void QTextDecoder::toUnicode(QString* target, const char* chars, int len)

C++ documentation:

This is an overloaded function.

The converted string is returned in target.