Struct qt_core::text_encoder::TextEncoder

pub struct TextEncoder {
    // some fields omitted
}

C++ type: QTextEncoder

C++ documentation:

The QTextEncoder class provides a state-based encoder.

A text encoder converts text from Unicode into an encoded text format using a specific codec.

The encoder converts Unicode into another format, remembering any state that is required between calls.

Methods

impl TextEncoder

fn from_unicode<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TextEncoderFromUnicodeArgs<'largs>

C++ method: QTextEncoder::fromUnicode

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_unicode(&mut self, (*const ::char::Char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn from_unicode(&mut self, (*const ::char::Char, libc::c_int, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray QTextEncoder::fromUnicode(const QChar* uc, int len)

C++ documentation:

This is an overloaded function.

Converts len characters (not bytes) from uc, and returns the result in a QByteArray.

Variant 2

Rust arguments:
1) fn from_unicode(&mut self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn from_unicode(&mut self, (&::string::String, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: QByteArray QTextEncoder::fromUnicode(const QString& str)

C++ documentation:

Converts the Unicode string str into an encoded QByteArray.

fn has_failure(&self) -> bool

C++ method: bool QTextEncoder::hasFailure() const

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextEncoderNewArgs

C++ method: QTextEncoder::QTextEncoder

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new((*const ::text_codec::TextCodec, cpp_utils::AsStruct)) -> ::text_encoder::TextEncoder
2) fn new((*const ::text_codec::TextCodec, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_encoder::TextEncoder>
C++ method: [constructor] void QTextEncoder::QTextEncoder(const QTextCodec* codec)

Variant 2

Rust arguments:
1) fn new((*const ::text_codec::TextCodec, ::flags::Flags<::text_codec::ConversionFlag>, cpp_utils::AsStruct)) -> ::text_encoder::TextEncoder
2) fn new((*const ::text_codec::TextCodec, ::flags::Flags<::text_codec::ConversionFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_encoder::TextEncoder>
C++ method: [constructor] void QTextEncoder::QTextEncoder(const QTextCodec* codec, QFlags<QTextCodec::ConversionFlag> flags)