Struct qt_core::text_stream::TextStream

pub struct TextStream {
    // some fields omitted
}

C++ type: QTextStream

C++ documentation:

The QTextStream class provides a convenient interface for reading and writing text.

QTextStream can operate on a QIODevice, a QByteArray or a QString. Using QTextStream's streaming operators, you can conveniently read and write words, lines and numbers. For generating text, QTextStream supports formatting options for field padding and alignment, and formatting of numbers. Example:

QFile data("output.txt");
if (data.open(QFile::WriteOnly | QFile::Truncate)) {
    QTextStream out(&data);
    out << "Result: " << qSetFieldWidth(10) << left << 3.14 << 2.7;
    // writes "Result: 3.14      2.7       "
}

It's also common to use QTextStream to read console input and write console output. QTextStream is locale aware, and will automatically decode standard input using the correct codec. Example:

QTextStream stream(stdin);
QString line;
while (stream.readLineInto(&line)) {
    ...
}

Besides using QTextStream's constructors, you can also set the device or string QTextStream operates on by calling setDevice() or setString(). You can seek to a position by calling seek(), and atEnd() will return true when there is no data left to be read. If you call flush(), QTextStream will empty all data from its write buffer into the device and call flush() on the device.

Internally, QTextStream uses a Unicode based buffer, and QTextCodec is used by QTextStream to automatically support different character sets. By default, QTextCodec::codecForLocale() is used for reading and writing, but you can also set the codec by calling setCodec(). Automatic Unicode detection is also supported. When this feature is enabled (the default behavior), QTextStream will detect the UTF-16 or the UTF-32 BOM (Byte Order Mark) and switch to the appropriate UTF codec when reading. QTextStream does not write a BOM by default, but you can enable this by calling setGenerateByteOrderMark(true). When QTextStream operates on a QString directly, the codec is disabled.

There are three general ways to use QTextStream when reading text files:

Chunk by chunk, by calling readLine() or readAll().

Word by word. QTextStream supports streaming into QStrings, QByteArrays and char* buffers. Words are delimited by space, and leading white space is automatically skipped.

Character by character, by streaming into QChar or char types. This method is often used for convenient input handling when parsing files, independent of character encoding and end-of-line semantics. To skip white space, call skipWhiteSpace().

Since the text stream uses a buffer, you should not read from the stream using the implementation of a superclass. For instance, if you have a QFile and read from it directly using QFile::readLine() instead of using the stream, the text stream's internal position will be out of sync with the file's position.

By default, when reading numbers from a stream of text, QTextStream will automatically detect the number's base representation. For example, if the number starts with "0x", it is assumed to be in hexadecimal form. If it starts with the digits 1-9, it is assumed to be in decimal form, and so on. You can set the integer base, thereby disabling the automatic detection, by calling setIntegerBase(). Example:

QTextStream in("0x50 0x20");
int firstNumber, secondNumber;

in >> firstNumber; // firstNumber == 80 in >> dec >> secondNumber; // secondNumber == 0

char ch; in >> ch; // ch == 'x'
QTextStream supports many formatting options for generating text. You can set the field width and pad character by calling setFieldWidth() and setPadChar(). Use setFieldAlignment() to set the alignment within each field. For real numbers, call setRealNumberNotation() and setRealNumberPrecision() to set the notation (SmartNotation, ScientificNotation, FixedNotation) and precision in digits of the generated number. Some extra number formatting options are also available through setNumberFlags().

Like <iostream> in the standard C++ library, QTextStream also defines several global manipulator functions:

Manipulator	Description

bin	Same as setIntegerBase(2).

oct	Same as setIntegerBase(8).

dec	Same as setIntegerBase(10).

hex	Same as setIntegerBase(16).

showbase	Same as setNumberFlags(numberFlags() | ShowBase).

forcesign	Same as setNumberFlags(numberFlags() | ForceSign).

forcepoint	Same as setNumberFlags(numberFlags() | ForcePoint).

noshowbase	Same as setNumberFlags(numberFlags() & ~ShowBase).

noforcesign	Same as setNumberFlags(numberFlags() & ~ForceSign).

noforcepoint	Same as setNumberFlags(numberFlags() & ~ForcePoint).

uppercasebase	Same as setNumberFlags(numberFlags() | UppercaseBase).

uppercasedigits	Same as setNumberFlags(numberFlags() | UppercaseDigits).

lowercasebase	Same as setNumberFlags(numberFlags() & ~UppercaseBase).

lowercasedigits	Same as setNumberFlags(numberFlags() & ~UppercaseDigits).

fixed	Same as setRealNumberNotation(FixedNotation).

scientific	Same as setRealNumberNotation(ScientificNotation).

left	Same as setFieldAlignment(AlignLeft).

right	Same as setFieldAlignment(AlignRight).

center	Same as setFieldAlignment(AlignCenter).

endl	Same as operator<<('\n') and flush().

flush	Same as flush().

reset	Same as reset().

ws	Same as skipWhiteSpace().

bom	Same as setGenerateByteOrderMark(true).

In addition, Qt provides three global manipulators that take a parameter: qSetFieldWidth(), qSetPadChar(), and qSetRealNumberPrecision().

Methods

impl TextStream

fn at_end(&self) -> bool

C++ method: bool QTextStream::atEnd() const

C++ documentation:

Returns true if there is no more data to be read from the QTextStream; otherwise returns false. This is similar to, but not the same as calling QIODevice::atEnd(), as QTextStream also takes into account its internal Unicode buffer.

fn auto_detect_unicode(&self) -> bool

C++ method: bool QTextStream::autoDetectUnicode() const

C++ documentation:

Returns true if automatic Unicode detection is enabled, otherwise returns false. Automatic Unicode detection is enabled by default.

See also setAutoDetectUnicode() and setCodec().

fn codec(&self) -> *mut TextCodec

C++ method: QTextCodec* QTextStream::codec() const

C++ documentation:

Returns the codec that is current assigned to the stream.

See also setCodec(), setAutoDetectUnicode(), and locale().

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QTextStream::device() const

C++ documentation:

Returns the current device associated with the QTextStream, or 0 if no device has been assigned.

See also setDevice() and string().

fn field_alignment(&self) -> FieldAlignment

C++ method: QTextStream::FieldAlignment QTextStream::fieldAlignment() const

C++ documentation:

Returns the current field alignment.

See also setFieldAlignment() and fieldWidth().

fn field_width(&self) -> c_int

C++ method: int QTextStream::fieldWidth() const

C++ documentation:

Returns the current field width.

See also setFieldWidth().

fn flush(&mut self)

C++ method: void QTextStream::flush()

C++ documentation:

Flushes any buffered data waiting to be written to the device.

If QTextStream operates on a string, this function does nothing.

fn generate_byte_order_mark(&self) -> bool

C++ method: bool QTextStream::generateByteOrderMark() const

C++ documentation:

Returns true if QTextStream is set to generate the UTF BOM (Byte Order Mark) when using a UTF codec; otherwise returns false. UTF BOM generation is set to false by default.

See also setGenerateByteOrderMark().

fn integer_base(&self) -> c_int

C++ method: int QTextStream::integerBase() const

C++ documentation:

Returns the current base of integers. 0 means that the base is detected when reading, or 10 (decimal) when generating numbers.

See also setIntegerBase(), QString::number(), and numberFlags().

fn locale<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextStreamLocaleArgs<'largs>

C++ method: QTextStream::locale

This is an overloaded function. Available variants:

Rust arguments:
1) fn locale(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::locale::Locale>
2) fn locale(&self, cpp_utils::AsStruct) -> ::locale::Locale
C++ method: QLocale QTextStream::locale() const

C++ documentation:

Returns the locale for this stream. The default locale is C.

This function was introduced in Qt 4.5.

See also setLocale().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TextStreamNewArgs

C++ method: QTextStream::QTextStream

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::text_stream::TextStream
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::text_stream::TextStream>
C++ method: [constructor] void QTextStream::QTextStream()

Variant 2

Rust arguments:
1) fn new((*mut ::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::text_stream::TextStream
2) fn new((*mut ::byte_array::ByteArray, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsStruct)) -> ::text_stream::TextStream
3) fn new((*mut ::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_stream::TextStream>
4) fn new((*mut ::byte_array::ByteArray, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_stream::TextStream>
C++ method: [constructor] void QTextStream::QTextStream(QByteArray* array, QFlags<QIODevice::OpenModeFlag> openMode = ?)

Variant 3

Rust arguments:
1) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::text_stream::TextStream
2) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_stream::TextStream>
C++ method: [constructor] void QTextStream::QTextStream(QIODevice* device)

Variant 4

Rust arguments:
1) fn new((*mut ::string::String, cpp_utils::AsStruct)) -> ::text_stream::TextStream
2) fn new((*mut ::string::String, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsStruct)) -> ::text_stream::TextStream
3) fn new((*mut ::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_stream::TextStream>
4) fn new((*mut ::string::String, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_stream::TextStream>
C++ method: [constructor] void QTextStream::QTextStream(QString* string, QFlags<QIODevice::OpenModeFlag> openMode = ?)

Variant 5

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::text_stream::TextStream
2) fn new((&::byte_array::ByteArray, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsStruct)) -> ::text_stream::TextStream
3) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_stream::TextStream>
4) fn new((&::byte_array::ByteArray, ::flags::Flags<::qiod_evice::OpenModeFlag>, cpp_utils::AsBox)) -> cpp_utils::CppBox<::text_stream::TextStream>
C++ method: [constructor] void QTextStream::QTextStream(const QByteArray& array, QFlags<QIODevice::OpenModeFlag> openMode = ?)

fn number_flags(&self) -> Flags<NumberFlag>

C++ method: QFlags<QTextStream::NumberFlag> QTextStream::numberFlags() const

C++ documentation:

Returns the current number flags.

See also setNumberFlags(), integerBase(), and realNumberNotation().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut TextStream where Args: TextStreamOpShlArgs<'largs>

C++ method: QTextStream::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::char::Char) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(QChar ch)

C++ documentation:

Writes the character c to the stream, then returns a reference to the QTextStream.

See also setFieldWidth().

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::latin1_string::Latin1String) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(QLatin1String s)

C++ documentation:

This is an overloaded function.

Writes string to the stream, and returns a reference to the QTextStream. The contents of string are converted with the QString constructor that takes a QLatin1String as argument.

Variant 3

Rust arguments: fn op_shl(&mut self, libc::c_char) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(char ch)

C++ documentation:

This is an overloaded function.

Converts c from ASCII to a QChar, then writes it to the stream.

Variant 4

Rust arguments: fn op_shl(&mut self, &'l1 ::byte_array::ByteArray) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(const QByteArray& array)

C++ documentation:

This is an overloaded function.

Writes array to the stream. The contents of array are converted with QString::fromUtf8().

Variant 5

Rust arguments: fn op_shl(&mut self, &'l1 ::string::String) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(const QString& s)

C++ documentation:

Writes the string string to the stream, and returns a reference to the QTextStream. The string is first encoded using the assigned codec (the default codec is QTextCodec::codecForLocale()) before it is written to the stream.

See also setFieldWidth() and setCodec().

Variant 6

Rust arguments: fn op_shl(&mut self, *const libc::c_char) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(const char* c)

C++ documentation:

This is an overloaded function.

Writes the constant string pointed to by string to the stream. string is assumed to be in ISO-8859-1 encoding. This operator is convenient when working with constant string data. Example:

QTextStream out(stdout);
out << "Qt rocks!" << endl;

Warning: QTextStream assumes that string points to a string of text, terminated by a '\0' character. If there is no terminating '\0' character, your application may crash.

Variant 7

Rust arguments: fn op_shl(&mut self, *const libc::c_void) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(const void* ptr)

C++ documentation:

This is an overloaded function.

Writes ptr to the stream as a hexadecimal number with a base.

Variant 8

Rust arguments: fn op_shl(&mut self, libc::c_double) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(double f)

C++ documentation:

This is an overloaded function.

Writes the double f to the stream.

Variant 9

Rust arguments: fn op_shl(&mut self, libc::c_float) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(float f)

C++ documentation:

Writes the real number f to the stream, then returns a reference to the QTextStream. By default, QTextStream stores it using SmartNotation, with up to 6 digits of precision. You can change the textual representation QTextStream will use for real numbers by calling setRealNumberNotation(), setRealNumberPrecision() and setNumberFlags().

See also setFieldWidth(), setRealNumberNotation(), setRealNumberPrecision(), and setNumberFlags().

Variant 10

Rust arguments: fn op_shl(&mut self, libc::c_int) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(int i)

C++ documentation:

This is an overloaded function.

Writes the signed int i to the stream.

Variant 11

Rust arguments: fn op_shl(&mut self, libc::c_long) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(long i)

C++ documentation:

This is an overloaded function.

Writes the signed long i to the stream.

Variant 12

Rust arguments: fn op_shl(&mut self, u64) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(qulonglong i)

C++ documentation:

This is an overloaded function.

Writes the qulonglong i to the stream.

Variant 13

Rust arguments: fn op_shl(&mut self, libc::c_short) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(short i)

C++ documentation:

Writes the integer number i to the stream, then returns a reference to the QTextStream. By default, the number is stored in decimal form, but you can also set the base by calling setIntegerBase().

See also setFieldWidth() and setNumberFlags().

Variant 14

Rust arguments: fn op_shl(&mut self, libc::c_uint) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(unsigned int i)

C++ documentation:

This is an overloaded function.

Writes the unsigned int i to the stream.

Variant 15

Rust arguments: fn op_shl(&mut self, libc::c_ushort) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator<<(unsigned short i)

C++ documentation:

This is an overloaded function.

Writes the unsigned short i to the stream.

fn op_shr<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut TextStream where Args: TextStreamOpShrArgs<'largs>

C++ method: QTextStream::operator>>

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shr(&mut self, &'l1 mut ::byte_array::ByteArray) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(QByteArray& array)

C++ documentation:

This is an overloaded function.

Converts the word to ISO-8859-1, then stores it in array.

See also QString::toLatin1().

Variant 2

Rust arguments: fn op_shr(&mut self, &'l1 mut ::char::Char) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(QChar& ch)

C++ documentation:

Reads a character from the stream and stores it in c. Returns a reference to the QTextStream, so several operators can be nested. Example:

QTextStream in(file);
QChar ch1, ch2, ch3;
in >> ch1 >> ch2 >> ch3;

Whitespace is not skipped.

Variant 3

Rust arguments: fn op_shr(&mut self, &'l1 mut ::string::String) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(QString& s)

C++ documentation:

Reads a word from the stream and stores it in str, then returns a reference to the stream. Words are separated by whitespace (i.e., all characters for which QChar::isSpace() returns true).

Leading whitespace is skipped.

Variant 4

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_char) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(char& ch)

C++ documentation:

This is an overloaded function.

Reads a character from the stream and stores it in c. The character from the stream is converted to ISO-5589-1 before it is stored.

See also QChar::toLatin1().

Variant 5

Rust arguments: fn op_shr(&mut self, *mut libc::c_char) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(char* c)

C++ documentation:

This is an overloaded function.

Stores the word in c, terminated by a '\0' character. If no word is available, only the '\0' character is stored.

Warning: Although convenient, this operator is dangerous and must be used with care. QTextStream assumes that c points to a buffer with enough space to hold the word. If the buffer is too small, your application may crash.

If possible, use the QByteArray operator instead.

Variant 6

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_double) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(double& f)

C++ documentation:

This is an overloaded function.

Stores the real number in the double f.

Variant 7

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_float) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(float& f)

C++ documentation:

Reads a real number from the stream and stores it in f, then returns a reference to the QTextStream. The number is cast to the correct type. If no real number is detect on the stream, f is set to 0.0.

As a special exception, QTextStream allows the strings "nan" and "inf" to represent NAN and INF floats or doubles.

Leading whitespace is skipped.

Variant 8

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_int) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(int& i)

C++ documentation:

This is an overloaded function.

Stores the integer in the signed int i.

Variant 9

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_long) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(long& i)

C++ documentation:

This is an overloaded function.

Stores the integer in the signed long i.

Variant 10

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_short) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(short& i)

C++ documentation:

Reads an integer from the stream and stores it in i, then returns a reference to the QTextStream. The number is cast to the correct type before it is stored. If no number was detected on the stream, i is set to 0.

By default, QTextStream will attempt to detect the base of the number using the following rules:

Prefix	Base

"0b" or "0B"	2 (binary)

"0" followed by "0-7"	8 (octal)

"0" otherwise	10 (decimal)

"0x" or "0X"	16 (hexadecimal)

"1" to "9"	10 (decimal)

By calling setIntegerBase(), you can specify the integer base explicitly. This will disable the auto-detection, and speed up QTextStream slightly.

Leading whitespace is skipped.

Variant 11

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_uint) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(unsigned int& i)

C++ documentation:

This is an overloaded function.

Stores the integer in the unsigned int i.

Variant 12

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_ulong) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(unsigned long& i)

C++ documentation:

This is an overloaded function.

Stores the integer in the unsigned long i.

Variant 13

Rust arguments: fn op_shr(&mut self, &'l1 mut libc::c_ushort) -> &'l0 mut ::text_stream::TextStream
C++ method: QTextStream& QTextStream::operator>>(unsigned short& i)

C++ documentation:

This is an overloaded function.

Stores the integer in the unsigned short i.

fn pad_char<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TextStreamPadCharArgs<'largs>

C++ method: QTextStream::padChar

This is an overloaded function. Available variants:

Rust arguments:
1) fn pad_char(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn pad_char(&self, cpp_utils::AsStruct) -> ::char::Char
C++ method: QChar QTextStream::padChar() const

C++ documentation:

Returns the current pad character.

See also setPadChar() and setFieldWidth().

fn pos(&self) -> i64

C++ method: qint64 QTextStream::pos() const

C++ documentation:

Returns the device position corresponding to the current position of the stream, or -1 if an error occurs (e.g., if there is no device or string, or if there's a device error).

Because QTextStream is buffered, this function may have to seek the device to reconstruct a valid device position. This operation can be expensive, so you may want to avoid calling this function in a tight loop.

This function was introduced in Qt 4.2.

See also seek().

fn read<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TextStreamReadArgs<'largs>

C++ method: QTextStream::read

This is an overloaded function. Available variants:

Rust arguments:
1) fn read(&mut self, (i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn read(&mut self, (i64, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTextStream::read(qint64 maxlen)

C++ documentation:

Reads at most maxlen characters from the stream, and returns the data read as a QString.

This function was introduced in Qt 4.1.

See also readAll(), readLine(), and QIODevice::read().

fn read_all<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TextStreamReadAllArgs<'largs>

C++ method: QTextStream::readAll

This is an overloaded function. Available variants:

Rust arguments:
1) fn read_all(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn read_all(&mut self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QTextStream::readAll()

C++ documentation:

Reads the entire content of the stream, and returns it as a QString. Avoid this function when working on large files, as it will consume a significant amount of memory.

Calling readLine() is better if you do not know how much data is available.

See also readLine().

fn read_line<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: TextStreamReadLineArgs<'largs>

C++ method: QTextStream::readLine

This is an overloaded function. Available variants:

Rust arguments:
1) fn read_line(&mut self, (i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn read_line(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
3) fn read_line(&mut self, (i64, cpp_utils::AsStruct)) -> ::string::String
4) fn read_line(&mut self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QTextStream::readLine(qint64 maxlen = ?)

C++ documentation:

Reads one line of text from the stream, and returns it as a QString. The maximum allowed line length is set to maxlen. If the stream contains lines longer than this, then the lines will be split after maxlen characters and returned in parts.

If maxlen is 0, the lines can be of any length.

The returned line has no trailing end-of-line characters ("\n" or "\r\n"), so calling QString::trimmed() can be unnecessary.

If the stream has read to the end of the file, readLine() will return a null QString. For strings, or for devices that support it, you can explicitly test for the end of the stream using atEnd().

See also readAll() and QIODevice::readLine().

fn read_line_into<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: TextStreamReadLineIntoArgs<'largs>

C++ method: QTextStream::readLineInto

This is an overloaded function. Available variants:

Rust arguments:
1) fn read_line_into(&mut self, *mut ::string::String) -> bool
2) fn read_line_into(&mut self, (*mut ::string::String, i64)) -> bool
C++ method: bool QTextStream::readLineInto(QString* line, qint64 maxlen = ?)

C++ documentation:

Reads one line of text from the stream into line. If line is 0, the read line is not stored.

The maximum allowed line length is set to maxlen. If the stream contains lines longer than this, then the lines will be split after maxlen characters and returned in parts.

If maxlen is 0, the lines can be of any length.

The resulting line has no trailing end-of-line characters ("\n" or "\r\n"), so calling QString::trimmed() can be unnecessary.

If line has sufficient capacity for the data that is about to be read, this function may not need to allocate new memory. Because of this, it can be faster than readLine().

Returns false if the stream has read to the end of the file or an error has occurred; otherwise returns true. The contents in line before the call are discarded in any case.

This function was introduced in Qt 5.5.

See also readAll() and QIODevice::readLine().

fn real_number_notation(&self) -> RealNumberNotation

C++ method: QTextStream::RealNumberNotation QTextStream::realNumberNotation() const

C++ documentation:

Returns the current real number notation.

See also setRealNumberNotation(), realNumberPrecision(), numberFlags(), and integerBase().

fn real_number_precision(&self) -> c_int

C++ method: int QTextStream::realNumberPrecision() const

C++ documentation:

Returns the current real number precision, or the number of fraction digits QTextStream will write when generating real numbers.

See also setRealNumberPrecision(), setRealNumberNotation(), realNumberNotation(), numberFlags(), and integerBase().

fn reset(&mut self)

C++ method: void QTextStream::reset()

C++ documentation:

Resets QTextStream's formatting options, bringing it back to its original constructed state. The device, string and any buffered data is left untouched.

fn reset_status(&mut self)

C++ method: void QTextStream::resetStatus()

C++ documentation:

Resets the status of the text stream.

This function was introduced in Qt 4.1.

See also QTextStream::Status, status(), and setStatus().

fn seek(&mut self, pos: i64) -> bool

C++ method: bool QTextStream::seek(qint64 pos)

C++ documentation:

Seeks to the position pos in the device. Returns true on success; otherwise returns false.

fn set_auto_detect_unicode(&mut self, enabled: bool)

C++ method: void QTextStream::setAutoDetectUnicode(bool enabled)

C++ documentation:

If enabled is true, QTextStream will attempt to detect Unicode encoding by peeking into the stream data to see if it can find the UTF-16 or UTF-32 BOM (Byte Order Mark). If this mark is found, QTextStream will replace the current codec with the UTF codec.

This function can be used together with setCodec(). It is common to set the codec to UTF-8, and then enable UTF-16 detection.

See also autoDetectUnicode() and setCodec().

fn set_codec<'largs, Args>(&'largs mut self, args: Args) where Args: TextStreamSetCodecArgs<'largs>

C++ method: QTextStream::setCodec

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_codec(&mut self, *mut ::text_codec::TextCodec) -> ()
C++ method: void QTextStream::setCodec(QTextCodec* codec)

C++ documentation:

Sets the codec for this stream to codec. The codec is used for decoding any data that is read from the assigned device, and for encoding any data that is written. By default, QTextCodec::codecForLocale() is used, and automatic unicode detection is enabled.

If QTextStream operates on a string, this function does nothing.

Warning: If you call this function while the text stream is reading from an open sequential socket, the internal buffer may still contain text decoded using the old codec.

See also codec(), setAutoDetectUnicode(), and setLocale().

Variant 2

Rust arguments: fn set_codec(&mut self, *const libc::c_char) -> ()
C++ method: void QTextStream::setCodec(const char* codecName)

C++ documentation:

Sets the codec for this stream to the QTextCodec for the encoding specified by codecName. Common values for codecName include "ISO 8859-1", "UTF-8", and "UTF-16". If the encoding isn't recognized, nothing happens.

Example:

QTextStream out(&file);
out.setCodec("UTF-8");

See also QTextCodec::codecForName() and setLocale().

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QTextStream::setDevice(QIODevice* device)

C++ documentation:

Sets the current device to device. If a device has already been assigned, QTextStream will call flush() before the old device is replaced.

Note: This function resets locale to the default locale ('C') and codec to the default codec, QTextCodec::codecForLocale().

See also device() and setString().

fn set_field_alignment(&mut self, alignment: FieldAlignment)

C++ method: void QTextStream::setFieldAlignment(QTextStream::FieldAlignment alignment)

C++ documentation:

Sets the field alignment to mode. When used together with setFieldWidth(), this function allows you to generate formatted output with text aligned to the left, to the right or center aligned.

See also fieldAlignment() and setFieldWidth().

fn set_field_width(&mut self, width: c_int)

C++ method: void QTextStream::setFieldWidth(int width)

C++ documentation:

Sets the current field width to width. If width is 0 (the default), the field width is equal to the length of the generated text.

Note: The field width applies to every element appended to this stream after this function has been called (e.g., it also pads endl). This behavior is different from similar classes in the STL, where the field width only applies to the next element.

See also fieldWidth() and setPadChar().

fn set_generate_byte_order_mark(&mut self, generate: bool)

C++ method: void QTextStream::setGenerateByteOrderMark(bool generate)

C++ documentation:

If generate is true and a UTF codec is used, QTextStream will insert the BOM (Byte Order Mark) before any data has been written to the device. If generate is false, no BOM will be inserted. This function must be called before any data is written. Otherwise, it does nothing.

See also generateByteOrderMark() and bom().

fn set_integer_base(&mut self, base: c_int)

C++ method: void QTextStream::setIntegerBase(int base)

C++ documentation:

Sets the base of integers to base, both for reading and for generating numbers. base can be either 2 (binary), 8 (octal), 10 (decimal) or 16 (hexadecimal). If base is 0, QTextStream will attempt to detect the base by inspecting the data on the stream. When generating numbers, QTextStream assumes base is 10 unless the base has been set explicitly.

See also integerBase(), QString::number(), and setNumberFlags().

fn set_locale(&mut self, locale: &Locale)

C++ method: void QTextStream::setLocale(const QLocale& locale)

C++ documentation:

Sets the locale for this stream to locale. The specified locale is used for conversions between numbers and their string representations.

The default locale is C and it is a special case - the thousands group separator is not used for backward compatibility reasons.

This function was introduced in Qt 4.5.

See also locale().

fn set_number_flags(&mut self, flags: Flags<NumberFlag>)

C++ method: void QTextStream::setNumberFlags(QFlags<QTextStream::NumberFlag> flags)

C++ documentation:

Sets the current number flags to flags. flags is a set of flags from the NumberFlag enum, and describes options for formatting generated code (e.g., whether or not to always write the base or sign of a number).

See also numberFlags(), setIntegerBase(), and setRealNumberNotation().

fn set_pad_char(&mut self, ch: &Char)

C++ method: void QTextStream::setPadChar(QChar ch)

C++ documentation:

Sets the pad character to ch. The default value is the ASCII space character (' '), or QChar(0x20). This character is used to fill in the space in fields when generating text.

Example:

QString s;
QTextStream out(&s);
out.setFieldWidth(10);
out.setFieldAlignment(QTextStream::AlignCenter);
out.setPadChar('-');
out << "Qt" << "rocks!";

The string s contains:

----Qt------rocks!--

See also padChar() and setFieldWidth().

fn set_real_number_notation(&mut self, notation: RealNumberNotation)

C++ method: void QTextStream::setRealNumberNotation(QTextStream::RealNumberNotation notation)

C++ documentation:

Sets the real number notation to notation (SmartNotation, FixedNotation, ScientificNotation). When reading and generating numbers, QTextStream uses this value to detect the formatting of real numbers.

See also realNumberNotation(), setRealNumberPrecision(), setNumberFlags(), and setIntegerBase().

fn set_real_number_precision(&mut self, precision: c_int)

C++ method: void QTextStream::setRealNumberPrecision(int precision)

C++ documentation:

Sets the precision of real numbers to precision. This value describes the number of fraction digits QTextStream should write when generating real numbers.

The precision cannot be a negative value. The default value is 6.

See also realNumberPrecision() and setRealNumberNotation().

fn set_status(&mut self, status: Status)

C++ method: void QTextStream::setStatus(QTextStream::Status status)

C++ documentation:

Sets the status of the text stream to the status given.

Subsequent calls to setStatus() are ignored until resetStatus() is called.

This function was introduced in Qt 4.1.

See also Status, status(), and resetStatus().

fn set_string<'largs, Args>(&'largs mut self, args: Args) where Args: TextStreamSetStringArgs<'largs>

C++ method: QTextStream::setString

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_string(&mut self, *mut ::string::String) -> ()
2) fn set_string(&mut self, (*mut ::string::String, ::flags::Flags<::qiod_evice::OpenModeFlag>)) -> ()
C++ method: void QTextStream::setString(QString* string, QFlags<QIODevice::OpenModeFlag> openMode = ?)

C++ documentation:

Sets the current string to string, using the given openMode. If a device has already been assigned, QTextStream will call flush() before replacing it.

See also string() and setDevice().

fn skip_white_space(&mut self)

C++ method: void QTextStream::skipWhiteSpace()

C++ documentation:

Reads and discards whitespace from the stream until either a non-space character is detected, or until atEnd() returns true. This function is useful when reading a stream character by character.

Whitespace characters are all characters for which QChar::isSpace() returns true.

See also operator>>().

fn status(&self) -> Status

C++ method: QTextStream::Status QTextStream::status() const

C++ documentation:

Returns the status of the text stream.

See also QTextStream::Status, setStatus(), and resetStatus().

fn string(&self) -> *mut String

C++ method: QString* QTextStream::string() const

C++ documentation:

Returns the current string assigned to the QTextStream, or 0 if no string has been assigned.

See also setString() and device().