Module qt_core::thread::overloading

Traits

ThreadConnectArgs
This trait represents a set of arguments accepted by Thread::connect method.

ThreadConnectStaticArgs
This trait represents a set of arguments accepted by Thread::connect_static method.

ThreadDestroyedArgs
This trait represents a set of arguments accepted by Thread::destroyed method.

ThreadDisconnectArgs
This trait represents a set of arguments accepted by Thread::disconnect method.

ThreadDisconnectStaticArgs
This trait represents a set of arguments accepted by Thread::disconnect_static method.

ThreadDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Thread::dynamic_property_names method.

ThreadExitArgs
This trait represents a set of arguments accepted by Thread::exit method.

ThreadNewArgs
This trait represents a set of arguments accepted by Thread::new method.

ThreadObjectNameArgs
This trait represents a set of arguments accepted by Thread::object_name method.

ThreadPropertyArgs
This trait represents a set of arguments accepted by Thread::property method.

ThreadStartArgs
This trait represents a set of arguments accepted by Thread::start method.

ThreadStartTimerArgs
This trait represents a set of arguments accepted by Thread::start_timer method.

ThreadTrArgs
This trait represents a set of arguments accepted by Thread::tr method.

ThreadTrUtf8Args
This trait represents a set of arguments accepted by Thread::tr_utf8 method.

ThreadWaitArgs
This trait represents a set of arguments accepted by Thread::wait method.