Module qt_core::thread_pool::overloading

Traits

ThreadPoolConnectArgs
This trait represents a set of arguments accepted by ThreadPool::connect method.

ThreadPoolConnectStaticArgs
This trait represents a set of arguments accepted by ThreadPool::connect_static method.

ThreadPoolDestroyedArgs
This trait represents a set of arguments accepted by ThreadPool::destroyed method.

ThreadPoolDisconnectArgs
This trait represents a set of arguments accepted by ThreadPool::disconnect method.

ThreadPoolDisconnectStaticArgs
This trait represents a set of arguments accepted by ThreadPool::disconnect_static method.

ThreadPoolDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by ThreadPool::dynamic_property_names method.

ThreadPoolNewArgs
This trait represents a set of arguments accepted by ThreadPool::new method.

ThreadPoolObjectNameArgs
This trait represents a set of arguments accepted by ThreadPool::object_name method.

ThreadPoolPropertyArgs
This trait represents a set of arguments accepted by ThreadPool::property method.

ThreadPoolStartArgs
This trait represents a set of arguments accepted by ThreadPool::start method.

ThreadPoolStartTimerArgs
This trait represents a set of arguments accepted by ThreadPool::start_timer method.

ThreadPoolTrArgs
This trait represents a set of arguments accepted by ThreadPool::tr method.

ThreadPoolTrUtf8Args
This trait represents a set of arguments accepted by ThreadPool::tr_utf8 method.

ThreadPoolWaitForDoneArgs
This trait represents a set of arguments accepted by ThreadPool::wait_for_done method.