Module qt_core::time::overloading

Traits

TimeAddMSecsArgs
This trait represents a set of arguments accepted by Time::add_m_secs method.

TimeAddSecsArgs
This trait represents a set of arguments accepted by Time::add_secs method.

TimeCurrentTimeArgs
This trait represents a set of arguments accepted by Time::current_time method.

TimeFromMSecsSinceStartOfDayArgs
This trait represents a set of arguments accepted by Time::from_m_secs_since_start_of_day method.

TimeFromStringArgs
This trait represents a set of arguments accepted by Time::from_string method.

TimeIsValidStaticArgs
This trait represents a set of arguments accepted by Time::is_valid_static method.

TimeNewArgs
This trait represents a set of arguments accepted by Time::new method.

TimeSetHmsArgs
This trait represents a set of arguments accepted by Time::set_hms method.

TimeToStringArgs
This trait represents a set of arguments accepted by Time::to_string method.