Struct qt_core::time::Time

pub struct Time {
    // some fields omitted
}

C++ type: QTime

C++ documentation:

The QTime class provides clock time functions.

A QTime object contains a clock time, i.e. the number of hours, minutes, seconds, and milliseconds since midnight. It can read the current time from the system clock and measure a span of elapsed time. It provides functions for comparing times and for manipulating a time by adding a number of milliseconds.

QTime uses the 24-hour clock format; it has no concept of AM/PM. Unlike QDateTime, QTime knows nothing about time zones or daylight savings time (DST).

A QTime object is typically created either by giving the number of hours, minutes, seconds, and milliseconds explicitly, or by using the static function currentTime(), which creates a QTime object that contains the system's local time. Note that the accuracy depends on the accuracy of the underlying operating system; not all systems provide 1-millisecond accuracy.

The hour(), minute(), second(), and msec() functions provide access to the number of hours, minutes, seconds, and milliseconds of the time. The same information is provided in textual format by the toString() function.

QTime provides a full set of operators to compare two QTime objects. QTime A is considered smaller than QTime B if A is earlier than B.

The addSecs() and addMSecs() functions provide the time a given number of seconds or milliseconds later than a given time. Correspondingly, the number of seconds or milliseconds between two times can be found using secsTo() or msecsTo().

QTime can be used to measure a span of elapsed time using the start(), restart(), and elapsed() functions.

Methods

impl Time

fn add_m_secs<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeAddMSecsArgs<'largs>

C++ method: QTime::addMSecs

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_m_secs(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
2) fn add_m_secs(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::time::Time
C++ method: QTime QTime::addMSecs(int ms) const

C++ documentation:

Returns a QTime object containing a time ms milliseconds later than the time of this object (or earlier if ms is negative).

Note that the time will wrap if it passes midnight. See addSecs() for an example.

Returns a null time if this time is invalid.

See also addSecs(), msecsTo(), and QDateTime::addMSecs().

fn add_secs<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeAddSecsArgs<'largs>

C++ method: QTime::addSecs

This is an overloaded function. Available variants:

Rust arguments:
1) fn add_secs(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
2) fn add_secs(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::time::Time
C++ method: QTime QTime::addSecs(int secs) const

C++ documentation:

Returns a QTime object containing a time s seconds later than the time of this object (or earlier if s is negative).

Note that the time will wrap if it passes midnight.

Returns a null time if this time is invalid.

Example:

QTime n(14, 0, 0);                // n == 14:00:00
QTime t;
t = n.addSecs(70);                // t == 14:01:10
t = n.addSecs(-70);               // t == 13:58:50
t = n.addSecs(10  60  60 + 5);  // t == 00:00:05
t = n.addSecs(-15  60  60);     // t == 23:00:00

See also addMSecs(), secsTo(), and QDateTime::addSecs().

fn current_time<Args>(args: Args) -> Args::ReturnType where Args: TimeCurrentTimeArgs

C++ method: QTime::currentTime

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_time(cpp_utils::AsBox) -> cpp_utils::CppBox<::time::Time>
2) fn current_time(cpp_utils::AsStruct) -> ::time::Time
C++ method: static QTime QTime::currentTime()

C++ documentation:

Returns the current time as reported by the system clock.

Note that the accuracy depends on the accuracy of the underlying operating system; not all systems provide 1-millisecond accuracy.

fn elapsed(&self) -> c_int

C++ method: int QTime::elapsed() const

C++ documentation:

Returns the number of milliseconds that have elapsed since the last time start() or restart() was called.

Note that the counter wraps to zero 24 hours after the last call to start() or restart.

Note that the accuracy depends on the accuracy of the underlying operating system; not all systems provide 1-millisecond accuracy.

Warning: If the system's clock setting has been changed since the last time start() or restart() was called, the result is undefined. This can happen when daylight savings time is turned on or off.

See also start() and restart().

fn from_m_secs_since_start_of_day<Args>(args: Args) -> Args::ReturnType where Args: TimeFromMSecsSinceStartOfDayArgs

C++ method: QTime::fromMSecsSinceStartOfDay

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_m_secs_since_start_of_day((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
2) fn from_m_secs_since_start_of_day((libc::c_int, cpp_utils::AsStruct)) -> ::time::Time
C++ method: static QTime QTime::fromMSecsSinceStartOfDay(int msecs)

C++ documentation:

Returns a new QTime instance with the time set to the number of msecs since the start of the day, i.e. since 00:00:00.

If msecs falls outside the valid range an invalid QTime will be returned.

See also msecsSinceStartOfDay().

fn from_string<Args>(args: Args) -> Args::ReturnType where Args: TimeFromStringArgs

C++ method: QTime::fromString

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn from_string((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
2) fn from_string((&::string::String, ::qt::DateFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
3) fn from_string((&::string::String, cpp_utils::AsStruct)) -> ::time::Time
4) fn from_string((&::string::String, ::qt::DateFormat, cpp_utils::AsStruct)) -> ::time::Time
C++ method: static QTime QTime::fromString(const QString& s, Qt::DateFormat f = ?)

C++ documentation:

Returns the time represented in the string as a QTime using the format given, or an invalid time if this is not possible.

Note that fromString() uses a "C" locale encoded string to convert milliseconds to a float value. If the default locale is not "C", this may result in two conversion attempts (if the conversion fails for the default locale). This should be considered an implementation detail.

Variant 2

Rust arguments:
1) fn from_string((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
2) fn from_string((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::time::Time
C++ method: static QTime QTime::fromString(const QString& s, const QString& format)

C++ documentation:

Returns the QTime represented by the string, using the format given, or an invalid time if the string cannot be parsed.

These expressions may be used for the format:

Expression	Output

h	the hour without a leading zero (0 to 23 or 1 to 12 if AM/PM display)

hh	the hour with a leading zero (00 to 23 or 01 to 12 if AM/PM display)

m	the minute without a leading zero (0 to 59)

mm	the minute with a leading zero (00 to 59)

s	the second without a leading zero (0 to 59)

ss	the second with a leading zero (00 to 59)

z	the milliseconds without leading zeroes (0 to 999)

zzz	the milliseconds with leading zeroes (000 to 999)

AP	interpret as an AM/PM time. AP must be either "AM" or "PM".

ap	Interpret as an AM/PM time. ap must be either "am" or "pm".

All other input characters will be treated as text. Any sequence of characters that are enclosed in single quotes will also be treated as text and not be used as an expression.

QTime time = QTime::fromString("1mm12car00", "m'mm'hcarss");
// time is 12:01.00

If the format is not satisfied, an invalid QTime is returned. Expressions that do not expect leading zeroes to be given (h, m, s and z) are greedy. This means that they will use two digits even if this puts them outside the range of accepted values and leaves too few digits for other sections. For example, the following string could have meant 00:07:10, but the m will grab two digits, resulting in an invalid time:

QTime time = QTime::fromString("00:710", "hh:ms"); // invalid

Any field that is not represented in the format will be set to zero. For example:

QTime time = QTime::fromString("1.30", "m.s");
// time is 00:01:30.000

See also QDateTime::fromString(), QDate::fromString(), QDate::toString(), QDateTime::toString(), and QTime::toString().

fn hour(&self) -> c_int

C++ method: int QTime::hour() const

C++ documentation:

Returns the hour part (0 to 23) of the time.

Returns -1 if the time is invalid.

See also minute(), second(), and msec().

fn is_null(&self) -> bool

C++ method: bool QTime::isNull() const

C++ documentation:

Returns true if the time is null (i.e., the QTime object was constructed using the default constructor); otherwise returns false. A null time is also an invalid time.

See also isValid().

fn is_valid(&self) -> bool

C++ method: bool QTime::isValid() const

C++ documentation:

Returns true if the time is valid; otherwise returns false. For example, the time 23:30:55.746 is valid, but 24:12:30 is invalid.

See also isNull().

fn is_valid_static<Args>(args: Args) -> bool where Args: TimeIsValidStaticArgs

C++ method: QTime::isValid

This is an overloaded function. Available variants:

Rust arguments:
1) fn is_valid_static((libc::c_int, libc::c_int, libc::c_int)) -> bool
2) fn is_valid_static((libc::c_int, libc::c_int, libc::c_int, libc::c_int)) -> bool
C++ method: static bool QTime::isValid(int h, int m, int s, int ms = ?)

C++ documentation:

This is an overloaded function.

Returns true if the specified time is valid; otherwise returns false.

The time is valid if h is in the range 0 to 23, m and s are in the range 0 to 59, and ms is in the range 0 to 999.

Example:

QTime::isValid(21, 10, 30); // returns true
QTime::isValid(22, 5,  62); // returns false

fn minute(&self) -> c_int

C++ method: int QTime::minute() const

C++ documentation:

Returns the minute part (0 to 59) of the time.

Returns -1 if the time is invalid.

See also hour(), second(), and msec().

fn msec(&self) -> c_int

C++ method: int QTime::msec() const

C++ documentation:

Returns the millisecond part (0 to 999) of the time.

Returns -1 if the time is invalid.

See also hour(), minute(), and second().

fn msecs_since_start_of_day(&self) -> c_int

C++ method: int QTime::msecsSinceStartOfDay() const

C++ documentation:

Returns the number of msecs since the start of the day, i.e. since 00:00:00.

See also fromMSecsSinceStartOfDay().

fn msecs_to(&self, arg1: &Time) -> c_int

C++ method: int QTime::msecsTo(const QTime& arg1) const

C++ documentation:

Returns the number of milliseconds from this time to t. If t is earlier than this time, the number of milliseconds returned is negative.

Because QTime measures time within a day and there are 86400 seconds in a day, the result is always between -86400000 and 86400000 ms.

Returns 0 if either time is invalid.

See also secsTo(), addMSecs(), and QDateTime::msecsTo().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TimeNewArgs

C++ method: QTime::QTime

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::time::Time
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::time::Time>
C++ method: [constructor] void QTime::QTime()

Variant 2

Rust arguments:
1) fn new((libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::time::Time
2) fn new((libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::time::Time
3) fn new((libc::c_int, libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::time::Time
4) fn new((libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
5) fn new((libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
6) fn new((libc::c_int, libc::c_int, libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time::Time>
C++ method: [constructor] void QTime::QTime(int h, int m, int s = ?, int ms = ?)

fn op_eq(&self, other: &Time) -> bool

C++ method: bool QTime::operator==(const QTime& other) const

C++ documentation:

Returns true if this time is equal to t; otherwise returns false.

fn op_ge(&self, other: &Time) -> bool

C++ method: bool QTime::operator>=(const QTime& other) const

C++ documentation:

Returns true if this time is later than or equal to t; otherwise returns false.

fn op_gt(&self, other: &Time) -> bool

C++ method: bool QTime::operator>(const QTime& other) const

C++ documentation:

Returns true if this time is later than t; otherwise returns false.

fn op_le(&self, other: &Time) -> bool

C++ method: bool QTime::operator<=(const QTime& other) const

C++ documentation:

Returns true if this time is earlier than or equal to t; otherwise returns false.

fn op_lt(&self, other: &Time) -> bool

C++ method: bool QTime::operator<(const QTime& other) const

C++ documentation:

Returns true if this time is earlier than t; otherwise returns false.

fn op_neq(&self, other: &Time) -> bool

C++ method: bool QTime::operator!=(const QTime& other) const

C++ documentation:

Returns true if this time is different from t; otherwise returns false.

fn restart(&mut self) -> c_int

C++ method: int QTime::restart()

C++ documentation:

Sets this time to the current time and returns the number of milliseconds that have elapsed since the last time start() or restart() was called.

This function is guaranteed to be atomic and is thus very handy for repeated measurements. Call start() to start the first measurement, and restart() for each later measurement.

Note that the counter wraps to zero 24 hours after the last call to start() or restart().

Warning: If the system's clock setting has been changed since the last time start() or restart() was called, the result is undefined. This can happen when daylight savings time is turned on or off.

See also start(), elapsed(), and currentTime().

fn second(&self) -> c_int

C++ method: int QTime::second() const

C++ documentation:

Returns the second part (0 to 59) of the time.

Returns -1 if the time is invalid.

See also hour(), minute(), and msec().

fn secs_to(&self, arg1: &Time) -> c_int

C++ method: int QTime::secsTo(const QTime& arg1) const

C++ documentation:

Returns the number of seconds from this time to t. If t is earlier than this time, the number of seconds returned is negative.

Because QTime measures time within a day and there are 86400 seconds in a day, the result is always between -86400 and 86400.

secsTo() does not take into account any milliseconds.

Returns 0 if either time is invalid.

See also addSecs() and QDateTime::secsTo().

fn set_hms<'largs, Args>(&'largs mut self, args: Args) -> bool where Args: TimeSetHmsArgs<'largs>

C++ method: QTime::setHMS

This is an overloaded function. Available variants:

Rust arguments:
1) fn set_hms(&mut self, (libc::c_int, libc::c_int, libc::c_int)) -> bool
2) fn set_hms(&mut self, (libc::c_int, libc::c_int, libc::c_int, libc::c_int)) -> bool
C++ method: bool QTime::setHMS(int h, int m, int s, int ms = ?)

C++ documentation:

Sets the time to hour h, minute m, seconds s and milliseconds ms.

h must be in the range 0 to 23, m and s must be in the range 0 to 59, and ms must be in the range 0 to 999. Returns true if the set time is valid; otherwise returns false.

See also isValid().

fn start(&mut self)

C++ method: void QTime::start()

C++ documentation:

Sets this time to the current time. This is practical for timing:

QTime t;
t.start();
some_lengthy_task();
qDebug("Time elapsed: %d ms", t.elapsed());

See also restart(), elapsed(), and currentTime().

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeToStringArgs<'largs>

C++ method: QTime::toString

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn to_string(&self, (::qt::DateFormat, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
3) fn to_string(&self, (::qt::DateFormat, cpp_utils::AsStruct)) -> ::string::String
4) fn to_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QTime::toString(Qt::DateFormat f = ?) const

C++ documentation:

This is an overloaded function.

Returns the time as a string. The format parameter determines the format of the string.

If format is Qt::TextDate, the string format is HH:mm:ss; e.g. 1 second before midnight would be "23:59:59".

If format is Qt::ISODate, the string format corresponds to the ISO 8601 extended specification for representations of dates, which is also HH:mm:ss.

If the format is Qt::SystemLocaleShortDate or Qt::SystemLocaleLongDate, the string format depends on the locale settings of the system. Identical to calling QLocale::system().toString(time, QLocale::ShortFormat) or QLocale::system().toString(time, QLocale::LongFormat).

If the format is Qt::DefaultLocaleShortDate or Qt::DefaultLocaleLongDate, the string format depends on the default application locale. This is the locale set with QLocale::setDefault(), or the system locale if no default locale has been set. Identical to calling

QLocale().toString(time, QLocale::ShortFormat) or QLocale().toString(time, QLocale::LongFormat).

If the format is Qt::RFC2822Date, the string is formatted in an RFC 2822 compatible way. An example of this formatting is "23:59:20".

If the time is invalid, an empty string will be returned.

See also QDate::toString() and QDateTime::toString().

Variant 2

Rust arguments:
1) fn to_string(&self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTime::toString(const QString& format) const

C++ documentation:

Returns the time as a string. The format parameter determines the format of the result string.

These expressions may be used:

Expression	Output

h	the hour without a leading zero (0 to 23 or 1 to 12 if AM/PM display)

hh	the hour with a leading zero (00 to 23 or 01 to 12 if AM/PM display)

H	the hour without a leading zero (0 to 23, even with AM/PM display)

HH	the hour with a leading zero (00 to 23, even with AM/PM display)

m	the minute without a leading zero (0 to 59)

mm	the minute with a leading zero (00 to 59)

s	the second without a leading zero (0 to 59)

ss	the second with a leading zero (00 to 59)

z	the milliseconds without leading zeroes (0 to 999)

zzz	the milliseconds with leading zeroes (000 to 999)

AP or A	use AM/PM display. A/AP will be replaced by either "AM" or "PM".

ap or a	use am/pm display. a/ap will be replaced by either "am" or "pm".

t	the timezone (for example "CEST")

All other input characters will be ignored. Any sequence of characters that are enclosed in single quotes will be treated as text and not be used as an expression. Two consecutive single quotes ("''") are replaced by a singlequote in the output. Formats without separators (e.g. "HHmm") are currently not supported.

Example format strings (assuming that the QTime is 14:13:09.042)

Format	Result

hh:mm:ss.zzz	14:13:09.042

h:m:s ap	2:13:9 pm

H:m:s a	14:13:9 pm

If the time is invalid, an empty string will be returned. If format is empty, the default format "hh:mm:ss" is used.

See also QDate::toString(), QDateTime::toString(), and QLocale::toString().