Struct qt_core::time_line::TimeLine

pub struct TimeLine {
    // some fields omitted
}

C++ type: QTimeLine

C++ documentation:

The QTimeLine class provides a timeline for controlling animations.

It's most commonly used to animate a GUI control by calling a slot periodically. You can construct a timeline by passing its duration in milliseconds to QTimeLine's constructor. The timeline's duration describes for how long the animation will run. Then you set a suitable frame range by calling setFrameRange(). Finally connect the frameChanged() signal to a suitable slot in the widget you wish to animate (for example, setValue() in QProgressBar). When you proceed to calling start(), QTimeLine will enter Running state, and start emitting frameChanged() at regular intervals, causing your widget's connected property's value to grow from the lower end to the upper and of your frame range, at a steady rate. You can specify the update interval by calling setUpdateInterval(). When done, QTimeLine enters NotRunning state, and emits finished().

Example:

...
progressBar = new QProgressBar(this);
progressBar->setRange(0, 100);

// Construct a 1-second timeline with a frame range of 0 - 100 QTimeLine *timeLine = new QTimeLine(1000, this); timeLine->setFrameRange(0, 100); connect(timeLine, SIGNAL(frameChanged(int)), progressBar, SLOT(setValue(int)));

// Clicking the push button will start the progress bar animation pushButton = new QPushButton(tr("Start animation"), this); connect(pushButton, SIGNAL(clicked()), timeLine, SLOT(start())); ...
By default the timeline runs once, from the beginning and towards the end, upon which you must call start() again to restart from the beginning. To make the timeline loop, you can call setLoopCount(), passing the number of times the timeline should run before finishing. The direction can also be changed, causing the timeline to run backward, by calling setDirection(). You can also pause and unpause the timeline while it's running by calling setPaused(). For interactive control, the setCurrentTime() function is provided, which sets the time position of the time line directly. Although most useful in NotRunning state, (e.g., connected to a valueChanged() signal in a QSlider,) this function can be called at any time.

The frame interface is useful for standard widgets, but QTimeLine can be used to control any type of animation. The heart of QTimeLine lies in the valueForTime() function, which generates a value between 0 and 1 for a given time. This value is typically used to describe the steps of an animation, where 0 is the first step of an animation, and 1 is the last step. When running, QTimeLine generates values between 0 and 1 by calling valueForTime() and emitting valueChanged(). By default, valueForTime() applies an interpolation algorithm to generate these value. You can choose from a set of predefined timeline algorithms by calling setCurveShape().

Note that by default, QTimeLine uses the EaseInOut curve shape, which provides a value that grows slowly, then grows steadily, and finally grows slowly. For a custom timeline, you can reimplement valueForTime(), in which case QTimeLine's curveShape property is ignored.

Methods

impl TimeLine

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QTimeLine::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QTimeLine::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeLineConnectArgs<'largs>

C++ method: QTimeLine::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: QMetaObject::Connection QTimeLine::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: TimeLineConnectStaticArgs

C++ method: QTimeLine::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QTimeLine::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QTimeLine::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn current_frame(&self) -> c_int

C++ method: int QTimeLine::currentFrame() const

C++ documentation:

Returns the frame corresponding to the current time.

See also currentTime(), frameForTime(), and setFrameRange().

fn current_time(&self) -> c_int

C++ method: int QTimeLine::currentTime() const

C++ documentation:

This property holds the current time of the time line.

When QTimeLine is in Running state, this value is updated continuously as a function of the duration and direction of the timeline. Otherwise, it is value that was current when stop() was called last, or the value set by setCurrentTime().

By default, this property contains a value of 0.

Access functions:

int 	currentTime() const

void 	setCurrentTime(int msec)

fn current_value(&self) -> c_double

C++ method: double QTimeLine::currentValue() const

C++ documentation:

Returns the value corresponding to the current time.

See also valueForTime() and currentFrame().

fn curve_shape(&self) -> CurveShape

C++ method: QTimeLine::CurveShape QTimeLine::curveShape() const

C++ documentation:

This property holds the shape of the timeline curve.

The curve shape describes the relation between the time and value for the base implementation of valueForTime().

If you have reimplemented valueForTime(), this value is ignored.

By default, this property is set to EaseInOutCurve.

Access functions:

CurveShape 	curveShape() const

void 	setCurveShape(CurveShape shape)

See also valueForTime().

fn delete_later(&mut self)

C++ method: void QTimeLine::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: TimeLineDestroyedArgs<'largs>

C++ method: QTimeLine::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut ::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QTimeLine::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn direction(&self) -> Direction

C++ method: QTimeLine::Direction QTimeLine::direction() const

C++ documentation:

This property holds the direction of the timeline when QTimeLine is in Running state.

This direction indicates whether the time moves from 0 towards the timeline duration, or from the value of the duration and towards 0 after start() has been called.

By default, this property is set to Forward.

Access functions:

Direction 	direction() const

void 	setDirection(Direction direction)

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: TimeLineDisconnectArgs<'largs>

C++ method: QTimeLine::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const ::object::Object) -> bool
2) fn disconnect(&self, (*const ::object::Object, *const libc::c_char)) -> bool
C++ method: bool QTimeLine::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const ::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QTimeLine::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: TimeLineDisconnectStaticArgs

C++ method: QTimeLine::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&::meta_object::Connection) -> bool
C++ method: static bool QTimeLine::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod)) -> bool
C++ method: static bool QTimeLine::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QTimeLine::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QTimeLine::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QTimeLine::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn duration(&self) -> c_int

C++ method: int QTimeLine::duration() const

C++ documentation:

This property holds the total duration of the timeline in milliseconds.

By default, this value is 1000 (i.e., 1 second), but you can change this by either passing a duration to QTimeLine's constructor, or by calling setDuration(). The duration must be larger than 0.

Note: Changing the duration does not cause the current time to be reset to zero or the new duration. You also need to call setCurrentTime() with the desired value.

Access functions:

int 	duration() const

void 	setDuration(int duration)

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeLineDynamicPropertyNamesArgs<'largs>

C++ method: QTimeLine::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: QList<QByteArray> QTimeLine::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn easing_curve<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeLineEasingCurveArgs<'largs>

C++ method: QTimeLine::easingCurve

This is an overloaded function. Available variants:

Rust arguments:
1) fn easing_curve(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::easing_curve::EasingCurve>
2) fn easing_curve(&self, cpp_utils::AsStruct) -> ::easing_curve::EasingCurve
C++ method: QEasingCurve QTimeLine::easingCurve() const

C++ documentation:

Specifies the easing curve that the timeline will use. If both easing curve and curveShape are set, the last set property will override the previous one. (If valueForTime() is reimplemented it will override both)

This property was introduced in Qt 4.6.

Access functions:

QEasingCurve 	easingCurve() const

void 	setEasingCurve(const QEasingCurve & curve)

fn end_frame(&self) -> c_int

C++ method: int QTimeLine::endFrame() const

C++ documentation:

Returns the end frame, which is the frame corresponding to the end of the timeline (i.e., the frame for which the current value is 1).

See also setEndFrame() and setFrameRange().

fn event(&mut self, arg1: *mut Event) -> bool

C++ method: virtual bool QTimeLine::event(QEvent* arg1)

C++ documentation:

This virtual function receives events to an object and should return true if the event e was recognized and processed.

The event() function can be reimplemented to customize the behavior of an object.

See also installEventFilter(), timerEvent(), QCoreApplication::sendEvent(), and QCoreApplication::postEvent().

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QTimeLine::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn frame_for_time(&self, msec: c_int) -> c_int

C++ method: int QTimeLine::frameForTime(int msec) const

C++ documentation:

Returns the frame corresponding to the time msec. This value is calculated using a linear interpolation of the start and end frame, based on the value returned by valueForTime().

See also valueForTime() and setFrameRange().

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QTimeLine::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QTimeLine::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_widget_type(&self) -> bool

C++ method: bool QTimeLine::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QTimeLine::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn kill_timer(&mut self, id: c_int)

C++ method: void QTimeLine::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn loop_count(&self) -> c_int

C++ method: int QTimeLine::loopCount() const

C++ documentation:

This property holds the number of times the timeline should loop before it's finished.

A loop count of of 0 means that the timeline will loop forever.

By default, this property contains a value of 1.

Access functions:

int 	loopCount() const

void 	setLoopCount(int count)

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QTimeLine::metaObject() const

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QTimeLine::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn new<Args>(args: Args) -> Args::ReturnType where Args: TimeLineNewArgs

C++ method: QTimeLine::QTimeLine

This is an overloaded function. Available variants:

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::time_line::TimeLine
2) fn new((libc::c_int, *mut ::object::Object, cpp_utils::AsStruct)) -> ::time_line::TimeLine
3) fn new(cpp_utils::AsStruct) -> ::time_line::TimeLine
4) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_line::TimeLine>
5) fn new((libc::c_int, *mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_line::TimeLine>
6) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::time_line::TimeLine>
C++ method: [constructor] void QTimeLine::QTimeLine(int duration = ?, QObject* parent = ?)

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeLineObjectNameArgs<'largs>

C++ method: QTimeLine::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QTimeLine::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QTimeLine::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeLinePropertyArgs<'largs>

C++ method: QTimeLine::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QTimeLine::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QTimeLine::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QTimeLine::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QTimeLine::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QTimeLine::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn resume(&mut self)

C++ method: void QTimeLine::resume()

C++ documentation:

Resumes the timeline from the current time. QTimeLine will reenter Running state, and once it enters the event loop, it will update its current time, frame and value at regular intervals.

In contrast to start(), this function does not restart the timeline before it resumes.

See also start(), updateInterval(), frameChanged(), and valueChanged().

fn set_current_time(&mut self, msec: c_int)

C++ method: void QTimeLine::setCurrentTime(int msec)

C++ documentation:

This property holds the current time of the time line.

When QTimeLine is in Running state, this value is updated continuously as a function of the duration and direction of the timeline. Otherwise, it is value that was current when stop() was called last, or the value set by setCurrentTime().

By default, this property contains a value of 0.

Access functions:

int 	currentTime() const

void 	setCurrentTime(int msec)

fn set_curve_shape(&mut self, shape: CurveShape)

C++ method: void QTimeLine::setCurveShape(QTimeLine::CurveShape shape)

C++ documentation:

This property holds the shape of the timeline curve.

The curve shape describes the relation between the time and value for the base implementation of valueForTime().

If you have reimplemented valueForTime(), this value is ignored.

By default, this property is set to EaseInOutCurve.

Access functions:

CurveShape 	curveShape() const

void 	setCurveShape(CurveShape shape)

See also valueForTime().

fn set_direction(&mut self, direction: Direction)

C++ method: void QTimeLine::setDirection(QTimeLine::Direction direction)

C++ documentation:

This property holds the direction of the timeline when QTimeLine is in Running state.

This direction indicates whether the time moves from 0 towards the timeline duration, or from the value of the duration and towards 0 after start() has been called.

By default, this property is set to Forward.

Access functions:

Direction 	direction() const

void 	setDirection(Direction direction)

fn set_duration(&mut self, duration: c_int)

C++ method: void QTimeLine::setDuration(int duration)

C++ documentation:

This property holds the total duration of the timeline in milliseconds.

By default, this value is 1000 (i.e., 1 second), but you can change this by either passing a duration to QTimeLine's constructor, or by calling setDuration(). The duration must be larger than 0.

Note: Changing the duration does not cause the current time to be reset to zero or the new duration. You also need to call setCurrentTime() with the desired value.

Access functions:

int 	duration() const

void 	setDuration(int duration)

fn set_easing_curve(&mut self, curve: &EasingCurve)

C++ method: void QTimeLine::setEasingCurve(const QEasingCurve& curve)

C++ documentation:

Specifies the easing curve that the timeline will use. If both easing curve and curveShape are set, the last set property will override the previous one. (If valueForTime() is reimplemented it will override both)

This property was introduced in Qt 4.6.

Access functions:

QEasingCurve 	easingCurve() const

void 	setEasingCurve(const QEasingCurve & curve)

fn set_end_frame(&mut self, frame: c_int)

C++ method: void QTimeLine::setEndFrame(int frame)

C++ documentation:

Sets the end frame, which is the frame corresponding to the end of the timeline (i.e., the frame for which the current value is 1), to frame.

See also endFrame(), startFrame(), and setFrameRange().

fn set_frame_range(&mut self, start_frame: c_int, end_frame: c_int)

C++ method: void QTimeLine::setFrameRange(int startFrame, int endFrame)

C++ documentation:

Sets the timeline's frame counter to start at startFrame, and end and endFrame. For each time value, QTimeLine will find the corresponding frame when you call currentFrame() or frameForTime() by interpolating, using the return value of valueForTime().

When in Running state, QTimeLine also emits the frameChanged() signal when the frame changes.

See also startFrame(), endFrame(), start(), and currentFrame().

fn set_loop_count(&mut self, count: c_int)

C++ method: void QTimeLine::setLoopCount(int count)

C++ documentation:

This property holds the number of times the timeline should loop before it's finished.

A loop count of of 0 means that the timeline will loop forever.

By default, this property contains a value of 1.

Access functions:

int 	loopCount() const

void 	setLoopCount(int count)

fn set_object_name(&mut self, name: &String)

C++ method: void QTimeLine::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QTimeLine::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_paused(&mut self, paused: bool)

C++ method: void QTimeLine::setPaused(bool paused)

C++ documentation:

If paused is true, the timeline is paused, causing QTimeLine to enter Paused state. No updates will be signaled until either start() or setPaused(false) is called. If paused is false, the timeline is resumed and continues where it left.

See also state() and start().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QTimeLine::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_start_frame(&mut self, frame: c_int)

C++ method: void QTimeLine::setStartFrame(int frame)

C++ documentation:

Sets the start frame, which is the frame corresponding to the start of the timeline (i.e., the frame for which the current value is 0), to frame.

See also startFrame(), endFrame(), and setFrameRange().

fn set_update_interval(&mut self, interval: c_int)

C++ method: void QTimeLine::setUpdateInterval(int interval)

C++ documentation:

This property holds the time in milliseconds between each time QTimeLine updates its current time.

When updating the current time, QTimeLine will emit valueChanged() if the current value changed, and frameChanged() if the frame changed.

By default, the interval is 40 ms, which corresponds to a rate of 25 updates per second.

Access functions:

int 	updateInterval() const

void 	setUpdateInterval(int interval)

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QTimeLine::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QTimeLine::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn start(&mut self)

C++ method: void QTimeLine::start()

C++ documentation:

Starts the timeline. QTimeLine will enter Running state, and once it enters the event loop, it will update its current time, frame and value at regular intervals. The default interval is 40 ms (i.e., 25 times per second). You can change the update interval by calling setUpdateInterval().

The timeline will start from position 0, or the end if going backward. If you want to resume a stopped timeline without restarting, you can call resume() instead.

See also resume(), updateInterval(), frameChanged(), and valueChanged().

fn start_frame(&self) -> c_int

C++ method: int QTimeLine::startFrame() const

C++ documentation:

Returns the start frame, which is the frame corresponding to the start of the timeline (i.e., the frame for which the current value is 0).

See also setStartFrame() and setFrameRange().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: TimeLineStartTimerArgs<'largs>

C++ method: QTimeLine::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, ::qt::TimerType)) -> libc::c_int
C++ method: int QTimeLine::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn state(&self) -> State

C++ method: QTimeLine::State QTimeLine::state() const

C++ documentation:

Returns the state of the timeline.

See also start(), setPaused(), and stop().

fn stop(&mut self)

C++ method: void QTimeLine::stop()

C++ documentation:

Stops the timeline, causing QTimeLine to enter NotRunning state.

See also start().

fn thread(&self) -> *mut Thread

C++ method: QThread* QTimeLine::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn toggle_direction(&mut self)

C++ method: void QTimeLine::toggleDirection()

C++ documentation:

Toggles the direction of the timeline. If the direction was Forward, it becomes Backward, and vice verca.

See also setDirection().

fn tr<Args>(args: Args) -> Args::ReturnType where Args: TimeLineTrArgs

C++ method: QTimeLine::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QTimeLine::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: TimeLineTrUtf8Args

C++ method: QTimeLine::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QTimeLine::trUtf8(const char* s, const char* c, int n)

fn update_interval(&self) -> c_int

C++ method: int QTimeLine::updateInterval() const

C++ documentation:

This property holds the time in milliseconds between each time QTimeLine updates its current time.

When updating the current time, QTimeLine will emit valueChanged() if the current value changed, and frameChanged() if the frame changed.

By default, the interval is 40 ms, which corresponds to a rate of 25 updates per second.

Access functions:

int 	updateInterval() const

void 	setUpdateInterval(int interval)

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QTimeLine::userData(unsigned int id) const

fn value_for_time(&self, msec: c_int) -> c_double

C++ method: virtual double QTimeLine::valueForTime(int msec) const

C++ documentation:

Returns the timeline value for the time msec. The returned value, which varies depending on the curve shape, is always between 0 and 1. If msec is 0, the default implementation always returns 0.

Reimplement this function to provide a custom curve shape for your timeline.

See also CurveShape and frameForTime().