Module qt_core::time_zone::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

TimeZoneAbbreviationArgs
This trait represents a set of arguments accepted by TimeZone::abbreviation method.

TimeZoneAvailableTimeZoneIdsArgs
This trait represents a set of arguments accepted by TimeZone::available_time_zone_ids method.

TimeZoneCommentArgs
This trait represents a set of arguments accepted by TimeZone::comment method.

TimeZoneDisplayNameArgs
This trait represents a set of arguments accepted by TimeZone::display_name method.

TimeZoneIanaIdToWindowsIdArgs
This trait represents a set of arguments accepted by TimeZone::iana_id_to_windows_id method.

TimeZoneIdArgs
This trait represents a set of arguments accepted by TimeZone::id method.

TimeZoneNewArgs
This trait represents a set of arguments accepted by TimeZone::new method.

TimeZoneNextTransitionArgs
This trait represents a set of arguments accepted by TimeZone::next_transition method.

TimeZoneOffsetDataArgs
This trait represents a set of arguments accepted by TimeZone::offset_data method.

TimeZonePreviousTransitionArgs
This trait represents a set of arguments accepted by TimeZone::previous_transition method.

TimeZoneSystemTimeZoneArgs
This trait represents a set of arguments accepted by TimeZone::system_time_zone method.

TimeZoneSystemTimeZoneIdArgs
This trait represents a set of arguments accepted by TimeZone::system_time_zone_id method.

TimeZoneTransitionsArgs
This trait represents a set of arguments accepted by TimeZone::transitions method.

TimeZoneUtcArgs
This trait represents a set of arguments accepted by TimeZone::utc method.

TimeZoneWindowsIdToDefaultIanaIdArgs
This trait represents a set of arguments accepted by TimeZone::windows_id_to_default_iana_id method.

TimeZoneWindowsIdToIanaIdsArgs
This trait represents a set of arguments accepted by TimeZone::windows_id_to_iana_ids method.