Struct qt_core::time_zone::OffsetData

pub struct OffsetData {
    // some fields omitted
}

C++ type: QTimeZone::OffsetData

C++ documentation:

The time zone offset data for a given moment in time, i.e. the time zone offsets and abbreviation to use at that moment in time.

OffsetData::atUtc The datetime of the offset data in UTC time.

OffsetData::offsetFromUtc The total offset from UTC in effect at the datetime.

OffsetData::standardTimeOffset The standard time offset component of the total offset.

OffsetData::daylightTimeOffset The daylight time offset component of the total offset.

OffsetData::abbreviation The abbreviation in effect at the datetime.

For example, for time zone "Europe/Berlin" the OffsetDate in standard and daylight time might be:

atUtc = QDateTime(QDate(2013, 1, 1), QTime(0, 0, 0), Qt::UTC)

offsetFromUtc = 3600

standardTimeOffset = 3600

daylightTimeOffset = 0

abbreviation = "CET"

atUtc = QDateTime(QDate(2013, 6, 1), QTime(0, 0, 0), Qt::UTC)

offsetFromUtc = 7200

standardTimeOffset = 3600

daylightTimeOffset = 3600

abbreviation = "CEST"

Methods

impl OffsetData

fn abbreviation<'l0>(&'l0 self) -> &'l0 String

C++ method: const QString& QTimeZone::OffsetData::abbreviation() const

fn abbreviation_mut<'l0>(&'l0 mut self) -> &'l0 mut String

C++ method: QString& QTimeZone::OffsetData::abbreviation_mut()

fn at_utc<'l0>(&'l0 self) -> &'l0 DateTime

C++ method: const QDateTime& QTimeZone::OffsetData::atUtc() const

fn at_utc_mut<'l0>(&'l0 mut self) -> &'l0 mut DateTime

C++ method: QDateTime& QTimeZone::OffsetData::atUtc_mut()

fn daylight_time_offset(&self) -> c_int

C++ method: int QTimeZone::OffsetData::daylightTimeOffset() const

fn offset_from_utc(&self) -> c_int

C++ method: int QTimeZone::OffsetData::offsetFromUtc() const

fn set_abbreviation(&mut self, value: &String)

C++ method: void QTimeZone::OffsetData::set_abbreviation(QString value)

fn set_at_utc(&mut self, value: &DateTime)

C++ method: void QTimeZone::OffsetData::set_atUtc(QDateTime value)

fn set_daylight_time_offset(&mut self, value: c_int)

C++ method: void QTimeZone::OffsetData::set_daylightTimeOffset(int value)

fn set_offset_from_utc(&mut self, value: c_int)

C++ method: void QTimeZone::OffsetData::set_offsetFromUtc(int value)

fn set_standard_time_offset(&mut self, value: c_int)

C++ method: void QTimeZone::OffsetData::set_standardTimeOffset(int value)

fn standard_time_offset(&self) -> c_int

C++ method: int QTimeZone::OffsetData::standardTimeOffset() const