Struct qt_core::time_zone::TimeZone

pub struct TimeZone {
    // some fields omitted
}

C++ type: QTimeZone

C++ documentation:

The QTimeZone class converts between between UTC and local time in a specific time zone.

This class provides a stateless calculator for time zone conversions between UTC and the local time in a specific time zone. By default it uses the host system time zone data to perform these conversions.

This class is primarily designed for use in QDateTime; most applications will not need to access this class directly and should instead use QDateTime with a Qt::TimeSpec of Qt::TimeZone.

Note: For consistency with QDateTime, QTimeZone does not account for leap seconds.

IANA Time Zone IDs

QTimeZone uses the IANA time zone IDs as defined in the IANA Time Zone Database (http://www.iana.org/time-zones). This is to ensure a standard ID across all supported platforms. Most platforms support the IANA IDs and the IANA Database natively, but for Windows a mapping is required to the native IDs. See below for more details.

The IANA IDs can and do change on a regular basis, and can vary depending on how recently the host system data was updated. As such you cannot rely on any given ID existing on any host system. You must use availableTimeZoneIds() to determine what IANA IDs are available.

The IANA IDs and database are also know as the Olson IDs and database, named after their creator.

UTC Offset Time Zones

A default UTC time zone backend is provided which is always guaranteed to be available. This provides a set of generic Offset From UTC time zones in the range UTC-14:00 to UTC+14:00. These time zones can be created using either the standard ISO format names "UTC+00:00" as listed by availableTimeZoneIds(), or using the number of offset seconds.

Windows Time Zones

Windows native time zone support is severely limited compared to the standard IANA TZ Database. Windows time zones cover larger geographic areas and are thus less accurate in their conversions. They also do not support as much historic conversion data and so may only be accurate for the current year.

QTimeZone uses a conversion table derived form the Unicode CLDR data to map between IANA IDs and Windows IDs. Depending on your version of Windows and Qt, this table may not be able to provide a valid conversion, in which "UTC" will be returned.

QTimeZone provides a public API to use this conversion table. The Windows ID used is the Windows Registry Key for the time zone which is also the MS Exchange EWS ID as well, but is different to the Time Zone Name (TZID) and COD code used by MS Exchange in versions before 2007.

System Time Zone

QTimeZone does not support any concept of a system or default time zone. If you require a QDateTime that uses the current system time zone at any given moment then you should use a Qt::TimeSpec of Qt::LocalTime.

The method systemTimeZoneId() returns the current system IANA time zone ID which on Unix-like systems will always be correct. On Windows this ID is translated from the Windows system ID using an internal translation table and the user's selected country. As a consequence there is a small chance any Windows install may have IDs not known by Qt, in which case "UTC" will be returned.

Creating a new QTimeZone instance using the system time zone ID will only produce a fixed named copy of the time zone, it will not change if the system time zone changes.

Time Zone Offsets

The difference between UTC and the local time in a time zone is expressed as an offset in seconds from UTC, i.e. the number of seconds to add to UTC to obtain the local time. The total offset is comprised of two component parts, the standard time offset and the daylight time offset. The standard time offset is the number of seconds to add to UTC to obtain standard time in the time zone. The daylight time offset is the number of seconds to add to the standard time offset to obtain daylight time in the time zone.

Note that the standard and daylight offsets for a time zone may change over time as countries have changed daylight time laws or even their standard time offset.

License

This class includes data obtained from the CLDR data files under the terms of the Unicode license.

COPYRIGHT AND PERMISSION NOTICE

Copyright © 1991-2012 Unicode, Inc. All rights reserved. Distributed under the Terms of Use in http://www.unicode.org/copyright.html.

Permission is hereby granted, free of charge, to any person obtaining a copy of the Unicode data files and any associated documentation (the "Data Files") or Unicode software and any associated documentation (the "Software") to deal in the Data Files or Software without restriction, including without limitation the rights to use, copy, modify, merge, publish, distribute, and/or sell copies of the Data Files or Software, and to permit persons to whom the Data Files or Software are furnished to do so, provided that (a) the above copyright notice(s) and this permission notice appear with all copies of the Data Files or Software, (b) both the above copyright notice(s) and this permission notice appear in associated documentation, and (c) there is clear notice in each modified Data File or in the Software as well as in the documentation associated with the Data File(s) or Software that the data or software has been modified.

Methods

impl TimeZone

fn abbreviation<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZoneAbbreviationArgs<'largs>

C++ method: QTimeZone::abbreviation

This is an overloaded function. Available variants:

Rust arguments:
1) fn abbreviation(&self, (&::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn abbreviation(&self, (&::date_time::DateTime, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTimeZone::abbreviation(const QDateTime& atDateTime) const

C++ documentation:

Returns the time zone abbreviation at the given atDateTime. The abbreviation may change depending on daylight time or even historical events.

Note that the abbreviation is not guaranteed to be unique to this time zone and should not be used in place of the ID or display name.

See also displayName().

fn available_time_zone_ids<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneAvailableTimeZoneIdsArgs

C++ method: QTimeZone::availableTimeZoneIds

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn available_time_zone_ids(cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn available_time_zone_ids(cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: static QList<QByteArray> QTimeZone::availableTimeZoneIds()

C++ documentation:

Returns a list of all available IANA time zone IDs on this system.

See also isTimeZoneIdAvailable().

Variant 2

Rust arguments:
1) fn available_time_zone_ids((::locale::Country, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn available_time_zone_ids((::locale::Country, cpp_utils::AsStruct)) -> ::list::ListByteArrayRef
C++ method: static QList<QByteArray> QTimeZone::availableTimeZoneIds(QLocale::Country country)

C++ documentation:

Returns a list of all available IANA time zone IDs for a given country.

As a special case, a country of Qt::AnyCountry returns those time zones that do not have any country related to them, such as UTC. If you require a list of all time zone IDs for all countries then use the standard availableTimeZoneIds() method.

See also isTimeZoneIdAvailable().

Variant 3

Rust arguments:
1) fn available_time_zone_ids((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn available_time_zone_ids((libc::c_int, cpp_utils::AsStruct)) -> ::list::ListByteArrayRef
C++ method: static QList<QByteArray> QTimeZone::availableTimeZoneIds(int offsetSeconds)

C++ documentation:

Returns a list of all available IANA time zone IDs with a given standard time offset of offsetSeconds.

See also isTimeZoneIdAvailable().

fn comment<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZoneCommentArgs<'largs>

C++ method: QTimeZone::comment

This is an overloaded function. Available variants:

Rust arguments:
1) fn comment(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn comment(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QTimeZone::comment() const

C++ documentation:

Returns any comment for the time zone.

A comment may be provided by the host platform to assist users in choosing the correct time zone. Depending on the platform this may not be localized.

fn country(&self) -> Country

C++ method: QLocale::Country QTimeZone::country() const

C++ documentation:

Returns the country for the time zone.

fn daylight_time_offset(&self, at_date_time: &DateTime) -> c_int

C++ method: int QTimeZone::daylightTimeOffset(const QDateTime& atDateTime) const

C++ documentation:

Returns the daylight time offset at the given atDateTime, i.e. the number of seconds to add to the standard time offset to obtain the local daylight time.

For example, for the time zone "Europe/Berlin" the daylight time offset is +3600 seconds. During standard time daylightTimeOffset() will return 0, and during daylight time it will return +3600.

See also offsetFromUtc() and standardTimeOffset().

fn display_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZoneDisplayNameArgs<'largs>

C++ method: QTimeZone::displayName

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn display_name(&self, (::time_zone::TimeType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn display_name(&self, (::time_zone::TimeType, ::time_zone::NameType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn display_name(&self, (::time_zone::TimeType, ::time_zone::NameType, &::locale::Locale, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
4) fn display_name(&self, (::time_zone::TimeType, cpp_utils::AsStruct)) -> ::string::String
5) fn display_name(&self, (::time_zone::TimeType, ::time_zone::NameType, cpp_utils::AsStruct)) -> ::string::String
6) fn display_name(&self, (::time_zone::TimeType, ::time_zone::NameType, &::locale::Locale, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTimeZone::displayName(QTimeZone::TimeType timeType, QTimeZone::NameType nameType = ?, const QLocale& locale = ?) const

C++ documentation:

Returns the localized time zone display name for the given timeType and nameType in the given locale. The nameType and locale requested may not be supported on all platforms, in which case the best available option will be returned.

If the locale is not provided then the application default locale will be used.

Where the time zone display names have changed over time then the most recent names will be used.

See also abbreviation().

Variant 2

Rust arguments:
1) fn display_name(&self, (&::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn display_name(&self, (&::date_time::DateTime, ::time_zone::NameType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
3) fn display_name(&self, (&::date_time::DateTime, ::time_zone::NameType, &::locale::Locale, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
4) fn display_name(&self, (&::date_time::DateTime, cpp_utils::AsStruct)) -> ::string::String
5) fn display_name(&self, (&::date_time::DateTime, ::time_zone::NameType, cpp_utils::AsStruct)) -> ::string::String
6) fn display_name(&self, (&::date_time::DateTime, ::time_zone::NameType, &::locale::Locale, cpp_utils::AsStruct)) -> ::string::String
C++ method: QString QTimeZone::displayName(const QDateTime& atDateTime, QTimeZone::NameType nameType = ?, const QLocale& locale = ?) const

C++ documentation:

Returns the localized time zone display name at the given atDateTime for the given nameType in the given locale. The nameType and locale requested may not be supported on all platforms, in which case the best available option will be returned.

If the locale is not provided then the application default locale will be used.

The display name may change depending on daylight time or historical events.

See also abbreviation().

fn has_daylight_time(&self) -> bool

C++ method: bool QTimeZone::hasDaylightTime() const

C++ documentation:

Returns true if the time zone has observed daylight time at any time.

See also isDaylightTime() and daylightTimeOffset().

fn has_transitions(&self) -> bool

C++ method: bool QTimeZone::hasTransitions() const

C++ documentation:

Returns true if the system backend supports obtaining transitions.

fn iana_id_to_windows_id<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneIanaIdToWindowsIdArgs

C++ method: QTimeZone::ianaIdToWindowsId

This is an overloaded function. Available variants:

Rust arguments:
1) fn iana_id_to_windows_id((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn iana_id_to_windows_id((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: static QByteArray QTimeZone::ianaIdToWindowsId(const QByteArray& ianaId)

C++ documentation:

Returns the Windows ID equivalent to the given ianaId.

See also windowsIdToDefaultIanaId() and windowsIdToIanaIds().

fn id<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZoneIdArgs<'largs>

C++ method: QTimeZone::id

This is an overloaded function. Available variants:

Rust arguments:
1) fn id(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn id(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QTimeZone::id() const

C++ documentation:

Returns the IANA ID for the time zone.

IANA IDs are used on all platforms. On Windows these are translated from the Windows ID into the closest IANA ID for the time zone and country.

fn is_daylight_time(&self, at_date_time: &DateTime) -> bool

C++ method: bool QTimeZone::isDaylightTime(const QDateTime& atDateTime) const

C++ documentation:

Returns true if the given atDateTime is in daylight time.

See also hasDaylightTime() and daylightTimeOffset().

fn is_time_zone_id_available(iana_id: &ByteArray) -> bool

C++ method: static bool QTimeZone::isTimeZoneIdAvailable(const QByteArray& ianaId)

C++ documentation:

Returns true if a given time zone ianaId is available on this system.

See also availableTimeZoneIds().

fn is_valid(&self) -> bool

C++ method: bool QTimeZone::isValid() const

C++ documentation:

Returns true if this time zone is valid.

fn new<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneNewArgs

C++ method: QTimeZone::QTimeZone

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::time_zone::TimeZone
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::time_zone::TimeZone>
C++ method: [constructor] void QTimeZone::QTimeZone()

Variant 2

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::time_zone::TimeZone
2) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::TimeZone>
C++ method: [constructor] void QTimeZone::QTimeZone(const QByteArray& ianaId)

Variant 3

Rust arguments:
1) fn new((&::byte_array::ByteArray, libc::c_int, &::string::String, &::string::String, cpp_utils::AsStruct)) -> ::time_zone::TimeZone
2) fn new((&::byte_array::ByteArray, libc::c_int, &::string::String, &::string::String, ::locale::Country, cpp_utils::AsStruct)) -> ::time_zone::TimeZone
3) fn new((&::byte_array::ByteArray, libc::c_int, &::string::String, &::string::String, ::locale::Country, &::string::String, cpp_utils::AsStruct)) -> ::time_zone::TimeZone
4) fn new((&::byte_array::ByteArray, libc::c_int, &::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::TimeZone>
5) fn new((&::byte_array::ByteArray, libc::c_int, &::string::String, &::string::String, ::locale::Country, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::TimeZone>
6) fn new((&::byte_array::ByteArray, libc::c_int, &::string::String, &::string::String, ::locale::Country, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::TimeZone>
C++ method: [constructor] void QTimeZone::QTimeZone(const QByteArray& zoneId, int offsetSeconds, const QString& name, const QString& abbreviation, QLocale::Country country = ?, const QString& comment = ?)

Variant 4

Rust arguments:
1) fn new((&::time_zone::TimeZone, cpp_utils::AsStruct)) -> ::time_zone::TimeZone
2) fn new((&::time_zone::TimeZone, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::TimeZone>
C++ method: [constructor] void QTimeZone::QTimeZone(const QTimeZone& other)

Variant 5

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::time_zone::TimeZone
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::TimeZone>
C++ method: [constructor] void QTimeZone::QTimeZone(int offsetSeconds)

fn next_transition<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZoneNextTransitionArgs<'largs>

C++ method: QTimeZone::nextTransition

This is an overloaded function. Available variants:

Rust arguments:
1) fn next_transition(&self, (&::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::OffsetData>
2) fn next_transition(&self, (&::date_time::DateTime, cpp_utils::AsStruct)) -> ::time_zone::OffsetData
C++ method: QTimeZone::OffsetData QTimeZone::nextTransition(const QDateTime& afterDateTime) const

C++ documentation:

Returns the first time zone Transition after the given afterDateTime. This is most useful when you have a Transition time and wish to find the Transition after it.

If there is no transition after the given afterDateTime then an invalid OffsetData will be returned with an invalid QDateTime.

The given afterDateTime is exclusive.

See also hasTransitions(), previousTransition(), and transitions().

fn offset_data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZoneOffsetDataArgs<'largs>

C++ method: QTimeZone::offsetData

This is an overloaded function. Available variants:

Rust arguments:
1) fn offset_data(&self, (&::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::OffsetData>
2) fn offset_data(&self, (&::date_time::DateTime, cpp_utils::AsStruct)) -> ::time_zone::OffsetData
C++ method: QTimeZone::OffsetData QTimeZone::offsetData(const QDateTime& forDateTime) const

C++ documentation:

Returns the effective offset details at the given forDateTime. This is the equivalent of calling offsetFromUtc(), abbreviation(), etc individually but is more efficient.

See also offsetFromUtc(), standardTimeOffset(), daylightTimeOffset(), and abbreviation().

fn offset_from_utc(&self, at_date_time: &DateTime) -> c_int

C++ method: int QTimeZone::offsetFromUtc(const QDateTime& atDateTime) const

C++ documentation:

Returns the total effective offset at the given atDateTime, i.e. the number of seconds to add to UTC to obtain the local time. This includes any daylight time offset that may be in effect, i.e. it is the sum of standardTimeOffset() and daylightTimeOffset() for the given datetime.

For example, for the time zone "Europe/Berlin" the standard time offset is +3600 seconds and the daylight time offset is +3600 seconds. During standard time offsetFromUtc() will return +3600 (UTC+01:00), and during daylight time it will return +7200 (UTC+02:00).

See also standardTimeOffset() and daylightTimeOffset().

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 TimeZone) -> &'l0 mut TimeZone

C++ method: QTimeZone& QTimeZone::operator=(const QTimeZone& other)

C++ documentation:

Assignment operator, assign other to this.

fn op_eq(&self, other: &TimeZone) -> bool

C++ method: bool QTimeZone::operator==(const QTimeZone& other) const

C++ documentation:

Returns true if this time zone is equal to the other time zone.

fn op_neq(&self, other: &TimeZone) -> bool

C++ method: bool QTimeZone::operator!=(const QTimeZone& other) const

C++ documentation:

Returns true if this time zone is not equal to the other time zone.

fn previous_transition<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZonePreviousTransitionArgs<'largs>

C++ method: QTimeZone::previousTransition

This is an overloaded function. Available variants:

Rust arguments:
1) fn previous_transition(&self, (&::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::time_zone::OffsetData>
2) fn previous_transition(&self, (&::date_time::DateTime, cpp_utils::AsStruct)) -> ::time_zone::OffsetData
C++ method: QTimeZone::OffsetData QTimeZone::previousTransition(const QDateTime& beforeDateTime) const

C++ documentation:

Returns the first time zone Transition before the given beforeDateTime. This is most useful when you have a Transition time and wish to find the Transition before it.

If there is no transition before the given beforeDateTime then an invalid OffsetData will be returned with an invalid QDateTime.

The given beforeDateTime is exclusive.

See also hasTransitions(), nextTransition(), and transitions().

fn standard_time_offset(&self, at_date_time: &DateTime) -> c_int

C++ method: int QTimeZone::standardTimeOffset(const QDateTime& atDateTime) const

C++ documentation:

Returns the standard time offset at the given atDateTime, i.e. the number of seconds to add to UTC to obtain the local Standard Time. This excludes any daylight time offset that may be in effect.

For example, for the time zone "Europe/Berlin" the standard time offset is +3600 seconds. During both standard and daylight time offsetFromUtc() will return +3600 (UTC+01:00).

See also offsetFromUtc() and daylightTimeOffset().

fn swap(&mut self, other: &mut TimeZone)

C++ method: void QTimeZone::swap(QTimeZone& other)

C++ documentation:

Swaps this time zone instance with other. This function is very fast and never fails.

fn system_time_zone<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneSystemTimeZoneArgs

C++ method: QTimeZone::systemTimeZone

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_time_zone(cpp_utils::AsBox) -> cpp_utils::CppBox<::time_zone::TimeZone>
2) fn system_time_zone(cpp_utils::AsStruct) -> ::time_zone::TimeZone
C++ method: static QTimeZone QTimeZone::systemTimeZone()

C++ documentation:

Returns a QTimeZone object that refers to the local system time, as specified by systemTimeZoneId().

This function was introduced in Qt 5.5.

See also utc().

fn system_time_zone_id<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneSystemTimeZoneIdArgs

C++ method: QTimeZone::systemTimeZoneId

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_time_zone_id(cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn system_time_zone_id(cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: static QByteArray QTimeZone::systemTimeZoneId()

C++ documentation:

Returns the current system time zone IANA ID.

On Windows this ID is translated from the Windows ID using an internal translation table and the user's selected country. As a consequence there is a small chance any Windows install may have IDs not known by Qt, in which case "UTC" will be returned.

fn transitions<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: TimeZoneTransitionsArgs<'largs>

C++ method: QTimeZone::transitions

This is an overloaded function. Available variants:

Rust arguments:
1) fn transitions(&self, (&::date_time::DateTime, &::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorOffsetDataRef>
2) fn transitions(&self, (&::date_time::DateTime, &::date_time::DateTime, cpp_utils::AsStruct)) -> ::vector::VectorOffsetDataRef
C++ method: QVector<QTimeZone::OffsetData> QTimeZone::transitions(const QDateTime& fromDateTime, const QDateTime& toDateTime) const

C++ documentation:

Returns a list of all time zone transitions between the given datetimes.

The given fromDateTime and toDateTime are inclusive.

See also hasTransitions(), nextTransition(), and previousTransition().

fn utc<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneUtcArgs

C++ method: QTimeZone::utc

This is an overloaded function. Available variants:

Rust arguments:
1) fn utc(cpp_utils::AsBox) -> cpp_utils::CppBox<::time_zone::TimeZone>
2) fn utc(cpp_utils::AsStruct) -> ::time_zone::TimeZone
C++ method: static QTimeZone QTimeZone::utc()

C++ documentation:

Returns a QTimeZone object that refers to UTC (Universal Time Coordinated).

This function was introduced in Qt 5.5.

See also systemTimeZone().

fn windows_id_to_default_iana_id<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneWindowsIdToDefaultIanaIdArgs

C++ method: QTimeZone::windowsIdToDefaultIanaId

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn windows_id_to_default_iana_id((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn windows_id_to_default_iana_id((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: static QByteArray QTimeZone::windowsIdToDefaultIanaId(const QByteArray& windowsId)

C++ documentation:

Returns the default IANA ID for a given windowsId.

Because a Windows ID can cover several IANA IDs in several different countries, this function returns the most frequently used IANA ID with no regard for the country and should thus be used with care. It is usually best to request the default for a specific country.

See also ianaIdToWindowsId() and windowsIdToIanaIds().

Variant 2

Rust arguments:
1) fn windows_id_to_default_iana_id((&::byte_array::ByteArray, ::locale::Country, cpp_utils::AsBox)) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn windows_id_to_default_iana_id((&::byte_array::ByteArray, ::locale::Country, cpp_utils::AsStruct)) -> ::byte_array::ByteArray
C++ method: static QByteArray QTimeZone::windowsIdToDefaultIanaId(const QByteArray& windowsId, QLocale::Country country)

C++ documentation:

Returns the default IANA ID for a given windowsId and country.

Because a Windows ID can cover several IANA IDs within a given country, the most frequently used IANA ID in that country is returned.

As a special case, QLocale::AnyCountry returns the default of those IANA IDs that do not have any specific country.

See also ianaIdToWindowsId() and windowsIdToIanaIds().

fn windows_id_to_iana_ids<Args>(args: Args) -> Args::ReturnType where Args: TimeZoneWindowsIdToIanaIdsArgs

C++ method: QTimeZone::windowsIdToIanaIds

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn windows_id_to_iana_ids((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn windows_id_to_iana_ids((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::list::ListByteArrayRef
C++ method: static QList<QByteArray> QTimeZone::windowsIdToIanaIds(const QByteArray& windowsId)

C++ documentation:

Returns all the IANA IDs for a given windowsId.

The returned list is sorted alphabetically.

See also ianaIdToWindowsId() and windowsIdToDefaultIanaId().

Variant 2

Rust arguments:
1) fn windows_id_to_iana_ids((&::byte_array::ByteArray, ::locale::Country, cpp_utils::AsBox)) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn windows_id_to_iana_ids((&::byte_array::ByteArray, ::locale::Country, cpp_utils::AsStruct)) -> ::list::ListByteArrayRef
C++ method: static QList<QByteArray> QTimeZone::windowsIdToIanaIds(const QByteArray& windowsId, QLocale::Country country)

C++ documentation:

Returns all the IANA IDs for a given windowsId and country.

As a special case QLocale::AnyCountry returns those IANA IDs that do not have any specific country.

The returned list is in order of frequency of usage, i.e. larger zones within a country are listed first.

See also ianaIdToWindowsId() and windowsIdToDefaultIanaId().