Module qt_core::translator::overloading

Traits

TranslatorConnectArgs
This trait represents a set of arguments accepted by Translator::connect method.

TranslatorConnectStaticArgs
This trait represents a set of arguments accepted by Translator::connect_static method.

TranslatorDestroyedArgs
This trait represents a set of arguments accepted by Translator::destroyed method.

TranslatorDisconnectArgs
This trait represents a set of arguments accepted by Translator::disconnect method.

TranslatorDisconnectStaticArgs
This trait represents a set of arguments accepted by Translator::disconnect_static method.

TranslatorDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by Translator::dynamic_property_names method.

TranslatorLoadArgs
This trait represents a set of arguments accepted by Translator::load method.

TranslatorNewArgs
This trait represents a set of arguments accepted by Translator::new method.

TranslatorObjectNameArgs
This trait represents a set of arguments accepted by Translator::object_name method.

TranslatorPropertyArgs
This trait represents a set of arguments accepted by Translator::property method.

TranslatorStartTimerArgs
This trait represents a set of arguments accepted by Translator::start_timer method.

TranslatorTrArgs
This trait represents a set of arguments accepted by Translator::tr method.

TranslatorTrUtf8Args
This trait represents a set of arguments accepted by Translator::tr_utf8 method.

TranslatorTranslateArgs
This trait represents a set of arguments accepted by Translator::translate method.