Enum qt_core::url::UrlFormattingOption

pub enum UrlFormattingOption {
    None,
    RemoveScheme,
    RemovePassword,
    RemoveUserInfo,
    RemovePort,
    RemoveAuthority,
    RemovePath,
    RemoveQuery,
    RemoveFragment,
    PreferLocalFile,
    StripTrailingSlash,
    RemoveFilename,
    NormalizePathSegments,
}

C++ type: QUrl::UrlFormattingOption

C++ documentation:

The formatting options define how the URL is formatted when written out as text.

Note that the case folding rules in Nameprep, which QUrl conforms to, require host names to always be converted to lower case, regardless of the Qt::FormattingOptions used.

The options from QUrl::ComponentFormattingOptions are also possible.

The FormattingOptions type is a typedef for QFlags<UrlFormattingOption>. It stores an OR combination of UrlFormattingOption values.

See also QUrl::ComponentFormattingOptions.

Variants
None

The format of the URL is unchanged. (C++ enum variant: None = 0)

RemoveScheme

The scheme is removed from the URL. (C++ enum variant: RemoveScheme = 1)

RemovePassword

Any password in the URL is removed. (C++ enum variant: RemovePassword = 2)

RemoveUserInfo

Any user information in the URL is removed. (C++ enum variant: RemoveUserInfo = 6)

RemovePort

Any specified port is removed from the URL. (C++ enum variant: RemovePort = 8)

RemoveAuthority

(C++ enum variant: RemoveAuthority = 30)

RemovePath

The URL's path is removed, leaving only the scheme, host address, and port (if present). (C++ enum variant: RemovePath = 32)

RemoveQuery

The query part of the URL (following a '?' character) is removed. (C++ enum variant: RemoveQuery = 64)

RemoveFragment

(C++ enum variant: RemoveFragment = 128)

PreferLocalFile

If the URL is a local file according to isLocalFile() and contains no query or fragment, a local file path is returned. (C++ enum variant: PreferLocalFile = 512)

StripTrailingSlash

The trailing slash is removed if one is present. (C++ enum variant: StripTrailingSlash = 1024)

RemoveFilename

The filename (i.e. everything after the last '/' in the path) is removed. The trailing '/' is kept, unless StripTrailingSlash is set. Only valid if RemovePath is not set. (C++ enum variant: RemoveFilename = 2048)

NormalizePathSegments

Modifies the path to remove redundant directory separators, and to resolve "."s and ".."s (as far as possible). (C++ enum variant: NormalizePathSegments = 4096)