Module qt_core::url::overloading

Traits

HashArgs
This trait represents a set of arguments accepted by hash method.

OpBitOrArgs
This trait represents a set of arguments accepted by op_bit_or method.

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

UrlAdjustedArgs
This trait represents a set of arguments accepted by Url::adjusted method.

UrlAuthorityArgs
This trait represents a set of arguments accepted by Url::authority method.

UrlErrorStringArgs
This trait represents a set of arguments accepted by Url::error_string method.

UrlFileNameArgs
This trait represents a set of arguments accepted by Url::file_name method.

UrlFragmentArgs
This trait represents a set of arguments accepted by Url::fragment method.

UrlFromAceArgs
This trait represents a set of arguments accepted by Url::from_ace method.

UrlFromEncodedArgs
This trait represents a set of arguments accepted by Url::from_encoded method.

UrlFromLocalFileArgs
This trait represents a set of arguments accepted by Url::from_local_file method.

UrlFromPercentEncodingArgs
This trait represents a set of arguments accepted by Url::from_percent_encoding method.

UrlFromStringListArgs
This trait represents a set of arguments accepted by Url::from_string_list method.

UrlFromUserInputArgs
This trait represents a set of arguments accepted by Url::from_user_input method.

UrlHostArgs
This trait represents a set of arguments accepted by Url::host method.

UrlIdnWhitelistArgs
This trait represents a set of arguments accepted by Url::idn_whitelist method.

UrlNewArgs
This trait represents a set of arguments accepted by Url::new method.

UrlOpAssignArgs
This trait represents a set of arguments accepted by Url::op_assign method.

UrlPasswordArgs
This trait represents a set of arguments accepted by Url::password method.

UrlPathArgs
This trait represents a set of arguments accepted by Url::path method.

UrlPortArgs
This trait represents a set of arguments accepted by Url::port method.

UrlQueryArgs
This trait represents a set of arguments accepted by Url::query method.

UrlResolvedArgs
This trait represents a set of arguments accepted by Url::resolved method.

UrlSchemeArgs
This trait represents a set of arguments accepted by Url::scheme method.

UrlSetAuthorityArgs
This trait represents a set of arguments accepted by Url::set_authority method.

UrlSetFragmentArgs
This trait represents a set of arguments accepted by Url::set_fragment method.

UrlSetHostArgs
This trait represents a set of arguments accepted by Url::set_host method.

UrlSetPasswordArgs
This trait represents a set of arguments accepted by Url::set_password method.

UrlSetPathArgs
This trait represents a set of arguments accepted by Url::set_path method.

UrlSetQueryArgs
This trait represents a set of arguments accepted by Url::set_query method.

UrlSetUrlArgs
This trait represents a set of arguments accepted by Url::set_url method.

UrlSetUserInfoArgs
This trait represents a set of arguments accepted by Url::set_user_info method.

UrlSetUserNameArgs
This trait represents a set of arguments accepted by Url::set_user_name method.

UrlToAceArgs
This trait represents a set of arguments accepted by Url::to_ace method.

UrlToDisplayStringArgs
This trait represents a set of arguments accepted by Url::to_display_string method.

UrlToEncodedArgs
This trait represents a set of arguments accepted by Url::to_encoded method.

UrlToLocalFileArgs
This trait represents a set of arguments accepted by Url::to_local_file method.

UrlToPercentEncodingArgs
This trait represents a set of arguments accepted by Url::to_percent_encoding method.

UrlToStringArgs
This trait represents a set of arguments accepted by Url::to_string method.

UrlToStringListArgs
This trait represents a set of arguments accepted by Url::to_string_list method.

UrlTopLevelDomainArgs
This trait represents a set of arguments accepted by Url::top_level_domain method.

UrlUrlArgs
This trait represents a set of arguments accepted by Url::url method.

UrlUserInfoArgs
This trait represents a set of arguments accepted by Url::user_info method.

UrlUserNameArgs
This trait represents a set of arguments accepted by Url::user_name method.