Enum qt_core::uuid::Version

pub enum Version {
    VerUnknown,
    Time,
    EmbeddedPOSIX,
    Md5,
    Random,
    Sha1,
}

C++ type: QUuid::Version

C++ documentation:

This enum defines the values used in the version field of the UUID. The version field is meaningful only if the value in the variant field is QUuid::DCE.

Variants
VerUnknown

Version is unknown (C++ enum variant: VerUnknown = -1)

Time

Time-based, by using timestamp, clock sequence, and MAC network card address (if available) for the node sections (C++ enum variant: Time = 1)

EmbeddedPOSIX

DCE Security version, with embedded POSIX UUIDs (C++ enum variant: EmbeddedPOSIX = 2)

Md5

This variant corresponds to multiple C++ enum variants with the same value:

Md5 = 3: Alias for Name

Name = 3: Name-based, by using values from a name for all sections

Random

Random-based, by using random numbers for all sections (C++ enum variant: Random = 4)

Sha1

(C++ enum variant: Sha1 = 5)