Struct qt_core::uuid::Uuid

pub struct Uuid {
    // some fields omitted
}

C++ type: QUuid

C++ documentation:

The QUuid class stores a Universally Unique Identifier (UUID).

Using Universally Unique IDentifiers (UUID) is a standard way to uniquely identify entities in a distributed computing environment. A UUID is a 16-byte (128-bit) number generated by some algorithm that is meant to guarantee that the UUID will be unique in the distributed computing environment where it is used. The acronym GUID is often used instead, Globally Unique IDentifiers, but it refers to the same thing.

Actually, the GUID is one variant of UUID. Multiple variants are in use. Each UUID contains a bit field that specifies which type (variant) of UUID it is. Call variant() to discover which type of UUID an instance of QUuid contains. It extracts the three most significant bits of byte 8 of the 16 bytes. In QUuid, byte 8 is QUuid::data4[0]. If you create instances of QUuid using the constructor that accepts all the numeric values as parameters, use the following table to set the three most significant bits of parameter b1, which becomes QUuid::data4[0] and contains the variant field in its three most significant bits. In the table, 'x' means don't care.

msb0	msb1	msb2	Variant

0	x	x	NCS (Network Computing System)

1	0	x	DCE (Distributed Computing Environment)

1	1	0	Microsoft (GUID)

1	1	1	Reserved for future expansion

If variant() returns QUuid::DCE, the UUID also contains a version field in the four most significant bits of QUuid::data3, and you can call version() to discover which version your QUuid contains. If you create instances of QUuid using the constructor that accepts all the numeric values as parameters, use the following table to set the four most significant bits of parameter w2, which becomes QUuid::data3 and contains the version field in its four most significant bits.

msb0	msb1	msb2	msb3	Version

0	0	0	1	Time

0	0	1	0	Embedded POSIX

0	0	1	1	Md5(Name)

0	1	0	0	Random

0	1	0	1	Sha1

The field layouts for the DCE versions listed in the table above are specified in the Network Working Group UUID Specification.

Most platforms provide a tool for generating new UUIDs, e.g. uuidgen and guidgen. You can also use createUuid(). UUIDs generated by createUuid() are of the random type. Their QUuid::Version bits are set to QUuid::Random, and their QUuid::Variant bits are set to QUuid::DCE. The rest of the UUID is composed of random numbers. Theoretically, this means there is a small chance that a UUID generated by createUuid() will not be unique. But it is a very small chance.

UUIDs can be constructed from numeric values or from strings, or using the static createUuid() function. They can be converted to a string with toString(). UUIDs have a variant() and a version(), and null UUIDs return true from isNull().

Methods

impl Uuid

fn create_uuid<Args>(args: Args) -> Args::ReturnType where Args: UuidCreateUuidArgs

C++ method: QUuid::createUuid

This is an overloaded function. Available variants:

Rust arguments:
1) fn create_uuid(cpp_utils::AsBox) -> cpp_utils::CppBox<::uuid::Uuid>
2) fn create_uuid(cpp_utils::AsStruct) -> ::uuid::Uuid
C++ method: static QUuid QUuid::createUuid()

C++ documentation:

On any platform other than Windows, this function returns a new UUID with variant QUuid::DCE and version QUuid::Random. If the /dev/urandom device exists, then the numbers used to construct the UUID will be of cryptographic quality, which will make the UUID unique. Otherwise, the numbers of the UUID will be obtained from the local pseudo-random number generator (qrand(), which is seeded by qsrand()) which is usually not of cryptograhic quality, which means that the UUID can't be guaranteed to be unique.

On a Windows platform, a GUID is generated, which almost certainly will be unique, on this or any other system, networked or not.

See also variant() and version().

fn create_uuid_v3<Args>(args: Args) -> Args::ReturnType where Args: UuidCreateUuidV3Args

C++ method: QUuid::createUuidV3

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn create_uuid_v3((&::uuid::Uuid, &::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
2) fn create_uuid_v3((&::uuid::Uuid, &::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::uuid::Uuid
C++ method: static QUuid QUuid::createUuidV3(const QUuid& ns, const QByteArray& baseData)

C++ documentation:

This function returns a new UUID with variant QUuid::DCE and version QUuid::Md5. ns is the namespace and baseData is the basic data as described by RFC 4122.

This function was introduced in Qt 5.0.

See also variant(), version(), and createUuidV5().

Variant 2

Rust arguments:
1) fn create_uuid_v3((&::uuid::Uuid, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
2) fn create_uuid_v3((&::uuid::Uuid, &::string::String, cpp_utils::AsStruct)) -> ::uuid::Uuid
C++ method: static QUuid QUuid::createUuidV3(const QUuid& ns, const QString& baseData)

C++ documentation:

This function returns a new UUID with variant QUuid::DCE and version QUuid::Md5. ns is the namespace and baseData is the basic data as described by RFC 4122.

This function was introduced in Qt 5.0.

See also variant(), version(), and createUuidV5().

fn create_uuid_v5<Args>(args: Args) -> Args::ReturnType where Args: UuidCreateUuidV5Args

C++ method: QUuid::createUuidV5

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn create_uuid_v5((&::uuid::Uuid, &::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
2) fn create_uuid_v5((&::uuid::Uuid, &::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::uuid::Uuid
C++ method: static QUuid QUuid::createUuidV5(const QUuid& ns, const QByteArray& baseData)

C++ documentation:

This function returns a new UUID with variant QUuid::DCE and version QUuid::Sha1. ns is the namespace and baseData is the basic data as described by RFC 4122.

This function was introduced in Qt 5.0.

See also variant(), version(), and createUuidV3().

Variant 2

Rust arguments:
1) fn create_uuid_v5((&::uuid::Uuid, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
2) fn create_uuid_v5((&::uuid::Uuid, &::string::String, cpp_utils::AsStruct)) -> ::uuid::Uuid
C++ method: static QUuid QUuid::createUuidV5(const QUuid& ns, const QString& baseData)

C++ documentation:

This function returns a new UUID with variant QUuid::DCE and version QUuid::Sha1. ns is the namespace and baseData is the basic data as described by RFC 4122.

This function was introduced in Qt 5.0.

See also variant(), version(), and createUuidV3().

fn data1(&self) -> c_uint

C++ method: unsigned int QUuid::data1() const

fn data2(&self) -> c_ushort

C++ method: unsigned short QUuid::data2() const

fn data3(&self) -> c_ushort

C++ method: unsigned short QUuid::data3() const

fn from_rfc4122<Args>(args: Args) -> Args::ReturnType where Args: UuidFromRfc4122Args

C++ method: QUuid::fromRfc4122

This is an overloaded function. Available variants:

Rust arguments:
1) fn from_rfc4122((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
2) fn from_rfc4122((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::uuid::Uuid
C++ method: static QUuid QUuid::fromRfc4122(const QByteArray& arg1)

C++ documentation:

Creates a QUuid object from the binary representation of the UUID, as specified by RFC 4122 section 4.1.2. See toRfc4122() for a further explanation of the order of bytes required.

The byte array accepted is NOT a human readable format.

If the conversion fails, a null UUID is created.

This function was introduced in Qt 4.8.

See also toRfc4122() and QUuid().

fn is_null(&self) -> bool

C++ method: bool QUuid::isNull() const

C++ documentation:

Returns true if this is the null UUID {00000000-0000-0000-0000-000000000000}; otherwise returns false.

fn new<Args>(args: Args) -> Args::ReturnType where Args: UuidNewArgs

C++ method: QUuid::QUuid

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::uuid::Uuid
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::uuid::Uuid>
C++ method: [constructor] void QUuid::QUuid()

Variant 2

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::uuid::Uuid
2) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
C++ method: [constructor] void QUuid::QUuid(const QByteArray& arg1)

Variant 3

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::uuid::Uuid
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
C++ method: [constructor] void QUuid::QUuid(const QString& arg1)

Variant 4

Rust arguments:
1) fn new((*const libc::c_char, cpp_utils::AsStruct)) -> ::uuid::Uuid
2) fn new((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
C++ method: [constructor] void QUuid::QUuid(const char* arg1)

Variant 5

Rust arguments:
1) fn new((libc::c_uint, libc::c_ushort, libc::c_ushort, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, cpp_utils::AsStruct)) -> ::uuid::Uuid
2) fn new((libc::c_uint, libc::c_ushort, libc::c_ushort, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, libc::c_uchar, cpp_utils::AsBox)) -> cpp_utils::CppBox<::uuid::Uuid>
C++ method: [constructor] void QUuid::QUuid(unsigned int l, unsigned short w1, unsigned short w2, unsigned char b1, unsigned char b2, unsigned char b3, unsigned char b4, unsigned char b5, unsigned char b6, unsigned char b7, unsigned char b8)

fn op_eq(&self, orig: &Uuid) -> bool

C++ method: bool QUuid::operator==(const QUuid& orig) const

C++ documentation:

Returns true if this QUuid and the other QUuid are identical; otherwise returns false.

fn op_gt(&self, other: &Uuid) -> bool

C++ method: bool QUuid::operator>(const QUuid& other) const

C++ documentation:

Returns true if this QUuid has the same variant field as the other QUuid and is lexicographically after the other QUuid. If the other QUuid has a different variant field, the return value is determined by comparing the two variants.

See also variant().

fn op_lt(&self, other: &Uuid) -> bool

C++ method: bool QUuid::operator<(const QUuid& other) const

C++ documentation:

Returns true if this QUuid has the same variant field as the other QUuid and is lexicographically before the other QUuid. If the other QUuid has a different variant field, the return value is determined by comparing the two variants.

See also variant().

fn op_neq(&self, orig: &Uuid) -> bool

C++ method: bool QUuid::operator!=(const QUuid& orig) const

C++ documentation:

Returns true if this QUuid and the other QUuid are different; otherwise returns false.

fn set_data1(&mut self, value: c_uint)

C++ method: void QUuid::set_data1(unsigned int value)

fn set_data2(&mut self, value: c_ushort)

C++ method: void QUuid::set_data2(unsigned short value)

fn set_data3(&mut self, value: c_ushort)

C++ method: void QUuid::set_data3(unsigned short value)

fn to_byte_array<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: UuidToByteArrayArgs<'largs>

C++ method: QUuid::toByteArray

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_byte_array(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_byte_array(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QUuid::toByteArray() const

C++ documentation:

Returns the binary representation of this QUuid. The byte array is formatted as five hex fields separated by '-' and enclosed in curly braces, i.e., "{xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx}" where 'x' is a hex digit. From left to right, the five hex fields are obtained from the four public data members in QUuid as follows:

Field #	Source

1	data1

2	data2

3	data3

4	data4[0] .. data4[1]

5	data4[2] .. data4[7]

This function was introduced in Qt 4.8.

fn to_rfc4122<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: UuidToRfc4122Args<'largs>

C++ method: QUuid::toRfc4122

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_rfc4122(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_rfc4122(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QUuid::toRfc4122() const

C++ documentation:

Returns the binary representation of this QUuid. The byte array is in big endian format, and formatted according to RFC 4122, section 4.1.2 - "Layout and byte order".

The order is as follows:

Field #	Source

1	data1

2	data2

3	data3

4	data4[0] .. data4[7]

This function was introduced in Qt 4.8.

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: UuidToStringArgs<'largs>

C++ method: QUuid::toString

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QUuid::toString() const

C++ documentation:

Returns the string representation of this QUuid. The string is formatted as five hex fields separated by '-' and enclosed in curly braces, i.e., "{xxxxxxxx-xxxx-xxxx-xxxx-xxxxxxxxxxxx}" where 'x' is a hex digit. From left to right, the five hex fields are obtained from the four public data members in QUuid as follows:

Field #	Source

1	data1

2	data2

3	data3

4	data4[0] .. data4[1]

5	data4[2] .. data4[7]

fn variant(&self) -> Variant

C++ method: QUuid::Variant QUuid::variant() const

C++ documentation:

Returns the value in the variant field of the UUID. If the return value is QUuid::DCE, call version() to see which layout it uses. The null UUID is considered to be of an unknown variant.

See also version().

fn version(&self) -> Version

C++ method: QUuid::Version QUuid::version() const

C++ documentation:

Returns the version field of the UUID, if the UUID's variant field is QUuid::DCE. Otherwise it returns QUuid::VerUnknown.

See also variant().