Module qt_core::variant::overloading

Traits

OpShlArgs
This trait represents a set of arguments accepted by op_shl method.

OpShrArgs
This trait represents a set of arguments accepted by op_shr method.

PrivateNewArgs
This trait represents a set of arguments accepted by Private::new method.

PrivateSharedNewArgs
This trait represents a set of arguments accepted by PrivateShared::new method.

VariantNewArgs
This trait represents a set of arguments accepted by Variant::new method.

VariantToBitArrayArgs
This trait represents a set of arguments accepted by Variant::to_bit_array method.

VariantToByteArrayArgs
This trait represents a set of arguments accepted by Variant::to_byte_array method.

VariantToCharArgs
This trait represents a set of arguments accepted by Variant::to_char method.

VariantToDateArgs
This trait represents a set of arguments accepted by Variant::to_date method.

VariantToDateTimeArgs
This trait represents a set of arguments accepted by Variant::to_date_time method.

VariantToDoubleArgs
This trait represents a set of arguments accepted by Variant::to_double method.

VariantToEasingCurveArgs
This trait represents a set of arguments accepted by Variant::to_easing_curve method.

VariantToFloatArgs
This trait represents a set of arguments accepted by Variant::to_float method.

VariantToHashArgs
This trait represents a set of arguments accepted by Variant::to_hash method.

VariantToIntArgs
This trait represents a set of arguments accepted by Variant::to_int method.

VariantToJsonArrayArgs
This trait represents a set of arguments accepted by Variant::to_json_array method.

VariantToJsonDocumentArgs
This trait represents a set of arguments accepted by Variant::to_json_document method.

VariantToJsonObjectArgs
This trait represents a set of arguments accepted by Variant::to_json_object method.

VariantToJsonValueArgs
This trait represents a set of arguments accepted by Variant::to_json_value method.

VariantToLineArgs
This trait represents a set of arguments accepted by Variant::to_line method.

VariantToLineFArgs
This trait represents a set of arguments accepted by Variant::to_line_f method.

VariantToListArgs
This trait represents a set of arguments accepted by Variant::to_list method.

VariantToLocaleArgs
This trait represents a set of arguments accepted by Variant::to_locale method.

VariantToLongLongArgs
This trait represents a set of arguments accepted by Variant::to_long_long method.

VariantToMapArgs
This trait represents a set of arguments accepted by Variant::to_map method.

VariantToModelIndexArgs
This trait represents a set of arguments accepted by Variant::to_model_index method.

VariantToPersistentModelIndexArgs
This trait represents a set of arguments accepted by Variant::to_persistent_model_index method.

VariantToPointArgs
This trait represents a set of arguments accepted by Variant::to_point method.

VariantToPointFArgs
This trait represents a set of arguments accepted by Variant::to_point_f method.

VariantToRealArgs
This trait represents a set of arguments accepted by Variant::to_real method.

VariantToRectArgs
This trait represents a set of arguments accepted by Variant::to_rect method.

VariantToRectFArgs
This trait represents a set of arguments accepted by Variant::to_rect_f method.

VariantToRegExpArgs
This trait represents a set of arguments accepted by Variant::to_reg_exp method.

VariantToRegularExpressionArgs
This trait represents a set of arguments accepted by Variant::to_regular_expression method.

VariantToSizeArgs
This trait represents a set of arguments accepted by Variant::to_size method.

VariantToSizeFArgs
This trait represents a set of arguments accepted by Variant::to_size_f method.

VariantToStringArgs
This trait represents a set of arguments accepted by Variant::to_string method.

VariantToStringListArgs
This trait represents a set of arguments accepted by Variant::to_string_list method.

VariantToTimeArgs
This trait represents a set of arguments accepted by Variant::to_time method.

VariantToUIntArgs
This trait represents a set of arguments accepted by Variant::to_u_int method.

VariantToULongLongArgs
This trait represents a set of arguments accepted by Variant::to_u_long_long method.

VariantToUrlArgs
This trait represents a set of arguments accepted by Variant::to_url method.

VariantToUuidArgs
This trait represents a set of arguments accepted by Variant::to_uuid method.