Struct qt_core::variant::Variant

pub struct Variant {
    // some fields omitted
}

C++ type: QVariant

C++ documentation:

The QVariant class acts like a union for the most common Qt data types.

Because C++ forbids unions from including types that have non-default constructors or destructors, most interesting Qt classes cannot be used in unions. Without QVariant, this would be a problem for QObject::property() and for database work, etc.

A QVariant object holds a single value of a single type() at a time. (Some type()s are multi-valued, for example a string list.) You can find out what type, T, the variant holds, convert it to a different type using convert(), get its value using one of the toT() functions (e.g., toSize()) and check whether the type can be converted to a particular type using canConvert().

The methods named toT() (e.g., toInt(), toString()) are const. If you ask for the stored type, they return a copy of the stored object. If you ask for a type that can be generated from the stored type, toT() copies and converts and leaves the object itself unchanged. If you ask for a type that cannot be generated from the stored type, the result depends on the type; see the function documentation for details.

Here is some example code to demonstrate the use of QVariant:

QDataStream out(...);
QVariant v(123);                // The variant now contains an int
int x = v.toInt();              // x = 123
out << v;                       // Writes a type tag and an int to out
v = QVariant("hello");          // The variant now contains a QByteArray
v = QVariant(tr("hello"));      // The variant now contains a QString
int y = v.toInt();              // y = 0 since v cannot be converted to an int
QString s = v.toString();       // s = tr("hello")  (see QObject::tr())
out << v;                       // Writes a type tag and a QString to out
...
QDataStream in(...);            // (opening the previously written stream)
in >> v;                        // Reads an Int variant
int z = v.toInt();              // z = 123
qDebug("Type is %s",            // prints "Type is int"
        v.typeName());
v = v.toInt() + 100;            // The variant now hold the value 223
v = QVariant(QStringList());

You can even store QList<QVariant> and QMap<QString, QVariant> values in a variant, so you can easily construct arbitrarily complex data structures of arbitrary types. This is very powerful and versatile, but may prove less memory and speed efficient than storing specific types in standard data structures.

QVariant also supports the notion of null values, where you can have a defined type with no value set. However, note that QVariant types can only be cast when they have had a value set.

QVariant x, y(QString()), z(QString(""));
x.convert(QVariant::Int);
// x.isNull() == true
// y.isNull() == true, z.isNull() == false

QVariant can be extended to support other types than those mentioned in the Type enum. See the QMetaType documentation for details.

Methods

impl Variant

fn can_convert(&self, target_type_id: c_int) -> bool

C++ method: bool QVariant::canConvert(int targetTypeId) const

C++ documentation:

Returns true if the variant's type can be cast to the requested type, targetTypeId. Such casting is done automatically when calling the toInt(), toBool(), ... methods.

The following casts are done automatically:

Type	Automatically Cast To

QMetaType::Bool	QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, QMetaType::ULongLong

QMetaType::QByteArray	QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, QMetaType::ULongLong

QMetaType::QChar	QMetaType::Bool, QMetaType::Int, QMetaType::UInt, QMetaType::LongLong, QMetaType::ULongLong

QMetaType::QColor	QMetaType::QString

QMetaType::QDate	QMetaType::QDateTime, QMetaType::QString

QMetaType::QDateTime	QMetaType::QDate, QMetaType::QString, QMetaType::QTime

QMetaType::Double	QMetaType::Bool, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, QMetaType::ULongLong

QMetaType::QFont	QMetaType::QString

QMetaType::Int	QMetaType::Bool, QMetaType::QChar, QMetaType::Double, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, QMetaType::ULongLong

QMetaType::QKeySequence	QMetaType::Int, QMetaType::QString

QMetaType::QVariantList	QMetaType::QStringList (if the list's items can be converted to QStrings)

QMetaType::LongLong	QMetaType::Bool, QMetaType::QByteArray, QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::QString, QMetaType::UInt, QMetaType::ULongLong

QMetaType::QPoint	QMetaType::QPointF

QMetaType::QRect	QMetaType::QRectF

QMetaType::QString	QMetaType::Bool, QMetaType::QByteArray, QMetaType::QChar, QMetaType::QColor, QMetaType::QDate, QMetaType::QDateTime, QMetaType::Double, QMetaType::QFont, QMetaType::Int, QMetaType::QKeySequence, QMetaType::LongLong, QMetaType::QStringList, QMetaType::QTime, QMetaType::UInt, QMetaType::ULongLong

QMetaType::QStringList	QMetaType::QVariantList, QMetaType::QString (if the list contains exactly one item)

QMetaType::QTime	QMetaType::QString

QMetaType::UInt	QMetaType::Bool, QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::ULongLong

QMetaType::ULongLong	QMetaType::Bool, QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt

A QVariant containing a pointer to a type derived from QObject will also return true for this function if a qobject_cast to the type described by targetTypeId would succeed. Note that this only works for QObject subclasses which use the Q_OBJECT macro.

A QVariant containing a sequential container will also return true for this function if the targetTypeId is QVariantList. It is possible to iterate over the contents of the container without extracting it as a (copied) QVariantList:

QList<int> intList;
intList.push_back(7);
intList.push_back(11);
intList.push_back(42);

QVariant variant = QVariant::fromValue(intList); if (variant.canConvert<QVariantList>()) { QSequentialIterable iterable = variant.value<QSequentialIterable>(); // Can use foreach: foreach (const QVariant &v, iterable) { qDebug() << v; } // Can use C++11 range-for: for (const QVariant &v : iterable) { qDebug() << v; } // Can use iterators: QSequentialIterable::const_iterator it = iterable.begin(); const QSequentialIterable::const_iterator end = iterable.end(); for ( ; it != end; ++it) { qDebug() << *it; } }
This requires that the value_type of the container is itself a metatype.

Similarly, a QVariant containing a sequential container will also return true for this function the targetTypeId is QVariantHash or QVariantMap. It is possible to iterate over the contents of the container without extracting it as a (copied) QVariantHash or QVariantMap:

QHash<int, QString> mapping;
mapping.insert(7, "Seven");
mapping.insert(11, "Eleven");
mapping.insert(42, "Forty-two");

QVariant variant = QVariant::fromValue(mapping); if (variant.canConvert<QVariantHash>()) { QAssociativeIterable iterable = variant.value<QAssociativeIterable>(); // Can use foreach over the values: foreach (const QVariant &v, iterable) { qDebug() << v; } // Can use C++11 range-for over the values: for (const QVariant &v : iterable) { qDebug() << v; } // Can use iterators: QAssociativeIterable::const_iterator it = iterable.begin(); const QAssociativeIterable::const_iterator end = iterable.end(); for ( ; it != end; ++it) { qDebug() << *it; // The current value qDebug() << it.key(); qDebug() << it.value(); } }
See also convert(), QSequentialIterable, Q_DECLARE_SEQUENTIAL_CONTAINER_METATYPE(), QAssociativeIterable, and Q_DECLARE_ASSOCIATIVE_CONTAINER_METATYPE().

fn clear(&mut self)

C++ method: void QVariant::clear()

C++ documentation:

Convert this variant to type QMetaType::UnknownType and free up any resources used.

fn const_data(&self) -> *const c_void

C++ method: const void* QVariant::constData() const

fn convert(&mut self, target_type_id: c_int) -> bool

C++ method: bool QVariant::convert(int targetTypeId)

C++ documentation:

Casts the variant to the requested type, targetTypeId. If the cast cannot be done, the variant is cleared. Returns true if the current type of the variant was successfully cast; otherwise returns false.

A QVariant containing a pointer to a type derived from QObject will also convert and return true for this function if a qobject_cast to the type described by targetTypeId would succeed. Note that this only works for QObject subclasses which use the Q_OBJECT macro.

Warning: For historical reasons, converting a null QVariant results in a null value of the desired type (e.g., an empty string for QString) and a result of false.

See also canConvert() and clear().

fn data(&self) -> *const c_void

C++ method: const void* QVariant::data() const

fn data_mut(&mut self) -> *mut c_void

C++ method: void* QVariant::data()

fn data_ptr<'l0>(&'l0 self) -> &'l0 Private

C++ method: const QVariant::Private& QVariant::data_ptr() const

fn data_ptr_mut<'l0>(&'l0 mut self) -> &'l0 mut Private

C++ method: QVariant::Private& QVariant::data_ptr()

fn detach(&mut self)

C++ method: void QVariant::detach()

fn is_detached(&self) -> bool

C++ method: bool QVariant::isDetached() const

fn is_null(&self) -> bool

C++ method: bool QVariant::isNull() const

C++ documentation:

Returns true if this is a null variant, false otherwise. A variant is considered null if it contains a default constructed value or a built-in type instance that has an isNull method, in which case the result would be the same as calling isNull on the wrapped object.

Warning: The result of the function doesn't affect == operator, which means that two values can be equal even if one of them is null and another is not.

fn is_valid(&self) -> bool

C++ method: bool QVariant::isValid() const

C++ documentation:

Returns true if the storage type of this variant is not QMetaType::UnknownType; otherwise returns false.

fn load(&mut self, ds: &mut DataStream)

C++ method: void QVariant::load(QDataStream& ds)

fn name_to_type(name: *const c_char) -> Type

C++ method: static QVariant::Type QVariant::nameToType(const char* name)

C++ documentation:

Converts the string representation of the storage type given in name, to its enum representation.

If the string representation cannot be converted to any enum representation, the variant is set to Invalid.

fn new<Args>(args: Args) -> Args::ReturnType where Args: VariantNewArgs

C++ method: QVariant::QVariant

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::variant::Variant
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant()

Variant 2

Rust arguments:
1) fn new((&::char::Char, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::char::Char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(QChar qchar)

Variant 3

Rust arguments:
1) fn new((&mut ::data_stream::DataStream, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&mut ::data_stream::DataStream, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(QDataStream& s)

Variant 4

Rust arguments:
1) fn new((&::latin1_string::Latin1String, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::latin1_string::Latin1String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(QLatin1String string)

Variant 5

Rust arguments:
1) fn new((::variant::Type, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((::variant::Type, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(QVariant::Type type)

Variant 6

Rust arguments:
1) fn new((bool, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((bool, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(bool b)

Variant 7

Rust arguments:
1) fn new((&::bit_array::BitArray, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::bit_array::BitArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QBitArray& bitarray)

Variant 8

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QByteArray& bytearray)

Variant 9

Rust arguments:
1) fn new((&::date::Date, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::date::Date, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QDate& date)

Variant 10

Rust arguments:
1) fn new((&::date_time::DateTime, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::date_time::DateTime, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QDateTime& datetime)

Variant 11

Rust arguments:
1) fn new((&::easing_curve::EasingCurve, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::easing_curve::EasingCurve, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QEasingCurve& easing)

Variant 12

Rust arguments:
1) fn new((&::hash::HashStringRefVariantRef, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::hash::HashStringRefVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QHash<QString, QVariant>& hash)

Variant 13

Rust arguments:
1) fn new((&::json_array::JsonArray, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::json_array::JsonArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QJsonArray& jsonArray)

Variant 14

Rust arguments:
1) fn new((&::json_document::JsonDocument, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::json_document::JsonDocument, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QJsonDocument& jsonDocument)

Variant 15

Rust arguments:
1) fn new((&::json_object::JsonObject, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::json_object::JsonObject, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QJsonObject& jsonObject)

Variant 16

Rust arguments:
1) fn new((&::json_value::JsonValue, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::json_value::JsonValue, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QJsonValue& jsonValue)

Variant 17

Rust arguments:
1) fn new((&::line::Line, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::line::Line, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QLine& line)

Variant 18

Rust arguments:
1) fn new((&::line_f::LineF, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::line_f::LineF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QLineF& line)

Variant 19

Rust arguments:
1) fn new((&::list::ListVariantRef, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::list::ListVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QList<QVariant>& list)

Variant 20

Rust arguments:
1) fn new((&::locale::Locale, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::locale::Locale, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QLocale& locale)

Variant 21

Rust arguments:
1) fn new((&::map::MapStringRefVariantRef, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::map::MapStringRefVariantRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QMap<QString, QVariant>& map)

Variant 22

Rust arguments:
1) fn new((&::model_index::ModelIndex, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::model_index::ModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QModelIndex& modelIndex)

Variant 23

Rust arguments:
1) fn new((&::persistent_model_index::PersistentModelIndex, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::persistent_model_index::PersistentModelIndex, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QPersistentModelIndex& modelIndex)

Variant 24

Rust arguments:
1) fn new((&::point::Point, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::point::Point, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QPoint& pt)

Variant 25

Rust arguments:
1) fn new((&::point_f::PointF, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::point_f::PointF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QPointF& pt)

Variant 26

Rust arguments:
1) fn new((&::rect::Rect, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::rect::Rect, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QRect& rect)

Variant 27

Rust arguments:
1) fn new((&::rect_f::RectF, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::rect_f::RectF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QRectF& rect)

Variant 28

Rust arguments:
1) fn new((&::reg_exp::RegExp, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::reg_exp::RegExp, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QRegExp& regExp)

Variant 29

Rust arguments:
1) fn new((&::regular_expression::RegularExpression, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::regular_expression::RegularExpression, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QRegularExpression& re)

Variant 30

Rust arguments:
1) fn new((&::size::Size, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::size::Size, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QSize& size)

Variant 31

Rust arguments:
1) fn new((&::size_f::SizeF, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::size_f::SizeF, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QSizeF& size)

Variant 32

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QString& string)

Variant 33

Rust arguments:
1) fn new((&::string_list::StringList, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::string_list::StringList, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QStringList& stringlist)

Variant 34

Rust arguments:
1) fn new((&::time::Time, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::time::Time, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QTime& time)

Variant 35

Rust arguments:
1) fn new((&::url::Url, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::url::Url, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QUrl& url)

Variant 36

Rust arguments:
1) fn new((&::uuid::Uuid, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::uuid::Uuid, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QUuid& uuid)

Variant 37

Rust arguments:
1) fn new((&::variant::Variant, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((&::variant::Variant, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const QVariant& other)

Variant 38

Rust arguments:
1) fn new((*const libc::c_char, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(const char* str)

Variant 39

Rust arguments:
1) fn new((libc::c_double, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(double d)

Variant 40

Rust arguments:
1) fn new((libc::c_float, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((libc::c_float, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(float f)

Variant 41

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(int i)

Variant 42

Rust arguments:
1) fn new((libc::c_int, *const libc::c_void, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((libc::c_int, *const libc::c_void, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(int typeId, const void* copy)

Variant 43

Rust arguments:
1) fn new((libc::c_int, *const libc::c_void, libc::c_uint, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((libc::c_int, *const libc::c_void, libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(int typeId, const void* copy, unsigned int flags)

Variant 44

Rust arguments:
1) fn new((i64, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((i64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(qlonglong ll)

Variant 45

Rust arguments:
1) fn new((u64, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((u64, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(qulonglong ull)

Variant 46

Rust arguments:
1) fn new((libc::c_uint, cpp_utils::AsStruct)) -> ::variant::Variant
2) fn new((libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
C++ method: [constructor] void QVariant::QVariant(unsigned int ui)

fn op_assign<'l0, 'l1>(&'l0 mut self, other: &'l1 Variant) -> &'l0 mut Variant

C++ method: QVariant& QVariant::operator=(const QVariant& other)

C++ documentation:

Assigns the value of the variant variant to this variant.

fn op_eq(&self, v: &Variant) -> bool

C++ method: bool QVariant::operator==(const QVariant& v) const

C++ documentation:

Compares this QVariant with v and returns true if they are equal; otherwise returns false.

QVariant uses the equality operator of the type() it contains to check for equality. QVariant will try to convert() v if its type is not the same as this variant's type. See canConvert() for a list of possible conversions.

Warning: To make this function work with a custom type registered with qRegisterMetaType(), its comparison operator must be registered using QMetaType::registerComparators().

fn op_ge(&self, v: &Variant) -> bool

C++ method: bool QVariant::operator>=(const QVariant& v) const

C++ documentation:

Compares this QVariant with v and returns true if this is larger or equal than v.

Note: Comparability might not be available for the type stored in this QVariant or in v.

Warning: To make this function work with a custom type registered with qRegisterMetaType(), its comparison operator must be registered using QMetaType::registerComparators().

fn op_gt(&self, v: &Variant) -> bool

C++ method: bool QVariant::operator>(const QVariant& v) const

C++ documentation:

Compares this QVariant with v and returns true if this is larger than v.

Note: Comparability might not be available for the type stored in this QVariant or in v.

Warning: To make this function work with a custom type registered with qRegisterMetaType(), its comparison operator must be registered using QMetaType::registerComparators().

fn op_le(&self, v: &Variant) -> bool

C++ method: bool QVariant::operator<=(const QVariant& v) const

C++ documentation:

Compares this QVariant with v and returns true if this is less or equal than v.

Note: Comparability might not be available for the type stored in this QVariant or in v.

Warning: To make this function work with a custom type registered with qRegisterMetaType(), its comparison operator must be registered using QMetaType::registerComparators().

fn op_lt(&self, v: &Variant) -> bool

C++ method: bool QVariant::operator<(const QVariant& v) const

C++ documentation:

Compares this QVariant with v and returns true if this is less than v.

Note: Comparability might not be availabe for the type stored in this QVariant or in v.

Warning: To make this function work with a custom type registered with qRegisterMetaType(), its comparison operator must be registered using QMetaType::registerComparators().

fn op_neq(&self, v: &Variant) -> bool

C++ method: bool QVariant::operator!=(const QVariant& v) const

C++ documentation:

Compares this QVariant with v and returns true if they are not equal; otherwise returns false.

Warning: To make this function work with a custom type registered with qRegisterMetaType(), its comparison operator must be registered using QMetaType::registerComparators().

fn save(&self, ds: &mut DataStream)

C++ method: void QVariant::save(QDataStream& ds) const

fn swap(&mut self, other: &mut Variant)

C++ method: void QVariant::swap(QVariant& other)

C++ documentation:

Swaps variant other with this variant. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn to_bit_array<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToBitArrayArgs<'largs>

C++ method: QVariant::toBitArray

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_bit_array(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::bit_array::BitArray>
2) fn to_bit_array(&self, cpp_utils::AsStruct) -> ::bit_array::BitArray
C++ method: QBitArray QVariant::toBitArray() const

C++ documentation:

Returns the variant as a QBitArray if the variant has userType() QMetaType::QBitArray; otherwise returns an empty bit array.

See also canConvert() and convert().

fn to_bool(&self) -> bool

C++ method: bool QVariant::toBool() const

C++ documentation:

Returns the variant as a bool if the variant has userType() Bool.

Returns true if the variant has userType() QMetaType::Bool, QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::UInt, or QMetaType::ULongLong and the value is non-zero, or if the variant has type QMetaType::QString or QMetaType::QByteArray and its lower-case content is not one of the following: empty, "0" or "false"; otherwise returns false.

See also canConvert() and convert().

fn to_byte_array<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToByteArrayArgs<'largs>

C++ method: QVariant::toByteArray

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_byte_array(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::byte_array::ByteArray>
2) fn to_byte_array(&self, cpp_utils::AsStruct) -> ::byte_array::ByteArray
C++ method: QByteArray QVariant::toByteArray() const

C++ documentation:

Returns the variant as a QByteArray if the variant has userType() QMetaType::QByteArray or QMetaType::QString (converted using QString::fromUtf8()); otherwise returns an empty byte array.

See also canConvert() and convert().

fn to_char<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToCharArgs<'largs>

C++ method: QVariant::toChar

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_char(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::char::Char>
2) fn to_char(&self, cpp_utils::AsStruct) -> ::char::Char
C++ method: QChar QVariant::toChar() const

C++ documentation:

Returns the variant as a QChar if the variant has userType() QMetaType::QChar, QMetaType::Int, or QMetaType::UInt; otherwise returns an invalid QChar.

See also canConvert() and convert().

fn to_date<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToDateArgs<'largs>

C++ method: QVariant::toDate

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_date(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date::Date>
2) fn to_date(&self, cpp_utils::AsStruct) -> ::date::Date
C++ method: QDate QVariant::toDate() const

C++ documentation:

Returns the variant as a QDate if the variant has userType() QMetaType::QDate, QMetaType::QDateTime, or QMetaType::QString; otherwise returns an invalid date.

If the type() is QMetaType::QString, an invalid date will be returned if the string cannot be parsed as a Qt::ISODate format date.

See also canConvert() and convert().

fn to_date_time<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToDateTimeArgs<'largs>

C++ method: QVariant::toDateTime

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_date_time(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::date_time::DateTime>
2) fn to_date_time(&self, cpp_utils::AsStruct) -> ::date_time::DateTime
C++ method: QDateTime QVariant::toDateTime() const

C++ documentation:

Returns the variant as a QDateTime if the variant has userType() QMetaType::QDateTime, QMetaType::QDate, or QMetaType::QString; otherwise returns an invalid date/time.

If the type() is QMetaType::QString, an invalid date/time will be returned if the string cannot be parsed as a Qt::ISODate format date/time.

See also canConvert() and convert().

fn to_double<'largs, Args>(&'largs self, args: Args) -> c_double where Args: VariantToDoubleArgs<'largs>

C++ method: QVariant::toDouble

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_double(&self, ()) -> libc::c_double
2) fn to_double(&self, *mut bool) -> libc::c_double
C++ method: double QVariant::toDouble(bool* ok = ?) const

C++ documentation:

Returns the variant as a double if the variant has userType() QMetaType::Double, QMetaType::Float, QMetaType::Bool, QMetaType::QByteArray, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, or QMetaType::ULongLong; otherwise returns 0.0.

If ok is non-null: ok is set to true if the value could be converted to a double; otherwise ok is set to false.

See also canConvert() and convert().

fn to_easing_curve<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToEasingCurveArgs<'largs>

C++ method: QVariant::toEasingCurve

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_easing_curve(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::easing_curve::EasingCurve>
2) fn to_easing_curve(&self, cpp_utils::AsStruct) -> ::easing_curve::EasingCurve
C++ method: QEasingCurve QVariant::toEasingCurve() const

C++ documentation:

Returns the variant as a QEasingCurve if the variant has userType() QMetaType::QEasingCurve; otherwise returns a default easing curve.

This function was introduced in Qt 4.7.

See also canConvert() and convert().

fn to_float<'largs, Args>(&'largs self, args: Args) -> c_float where Args: VariantToFloatArgs<'largs>

C++ method: QVariant::toFloat

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_float(&self, ()) -> libc::c_float
2) fn to_float(&self, *mut bool) -> libc::c_float
C++ method: float QVariant::toFloat(bool* ok = ?) const

C++ documentation:

Returns the variant as a float if the variant has userType() QMetaType::Double, QMetaType::Float, QMetaType::Bool, QMetaType::QByteArray, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, or QMetaType::ULongLong; otherwise returns 0.0.

If ok is non-null: ok is set to true if the value could be converted to a double; otherwise ok is set to false.

This function was introduced in Qt 4.6.

See also canConvert() and convert().

fn to_hash<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToHashArgs<'largs>

C++ method: QVariant::toHash

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_hash(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::hash::HashStringRefVariantRef>
2) fn to_hash(&self, cpp_utils::AsStruct) -> ::hash::HashStringRefVariantRef
C++ method: QHash<QString, QVariant> QVariant::toHash() const

C++ documentation:

Returns the variant as a QHash<QString, QVariant> if the variant has type() QMetaType::QVariantHash; otherwise returns an empty map.

See also canConvert() and convert().

fn to_int<'largs, Args>(&'largs self, args: Args) -> c_int where Args: VariantToIntArgs<'largs>

C++ method: QVariant::toInt

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_int(&self, ()) -> libc::c_int
2) fn to_int(&self, *mut bool) -> libc::c_int
C++ method: int QVariant::toInt(bool* ok = ?) const

C++ documentation:

Returns the variant as an int if the variant has userType() QMetaType::Int, QMetaType::Bool, QMetaType::QByteArray, QMetaType::QChar, QMetaType::Double, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, or QMetaType::ULongLong; otherwise returns 0.

If ok is non-null: ok is set to true if the value could be converted to an int; otherwise ok is set to false.

Warning: If the value is convertible to a QMetaType::LongLong but is too large to be represented in an int, the resulting arithmetic overflow will not be reflected in ok. A simple workaround is to use QString::toInt().

See also canConvert() and convert().

fn to_json_array<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToJsonArrayArgs<'largs>

C++ method: QVariant::toJsonArray

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_json_array(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_array::JsonArray>
2) fn to_json_array(&self, cpp_utils::AsStruct) -> ::json_array::JsonArray
C++ method: QJsonArray QVariant::toJsonArray() const

C++ documentation:

Returns the variant as a QJsonArray if the variant has userType() QJsonArray; otherwise returns a default constructed QJsonArray.

This function was introduced in Qt 5.0.

See also canConvert() and convert().

fn to_json_document<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToJsonDocumentArgs<'largs>

C++ method: QVariant::toJsonDocument

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_json_document(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_document::JsonDocument>
2) fn to_json_document(&self, cpp_utils::AsStruct) -> ::json_document::JsonDocument
C++ method: QJsonDocument QVariant::toJsonDocument() const

C++ documentation:

Returns the variant as a QJsonDocument if the variant has userType() QJsonDocument; otherwise returns a default constructed QJsonDocument.

This function was introduced in Qt 5.0.

See also canConvert() and convert().

fn to_json_object<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToJsonObjectArgs<'largs>

C++ method: QVariant::toJsonObject

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_json_object(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_object::JsonObject>
2) fn to_json_object(&self, cpp_utils::AsStruct) -> ::json_object::JsonObject
C++ method: QJsonObject QVariant::toJsonObject() const

C++ documentation:

Returns the variant as a QJsonObject if the variant has userType() QJsonObject; otherwise returns a default constructed QJsonObject.

This function was introduced in Qt 5.0.

See also canConvert() and convert().

fn to_json_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToJsonValueArgs<'largs>

C++ method: QVariant::toJsonValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_json_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::json_value::JsonValue>
2) fn to_json_value(&self, cpp_utils::AsStruct) -> ::json_value::JsonValue
C++ method: QJsonValue QVariant::toJsonValue() const

C++ documentation:

Returns the variant as a QJsonValue if the variant has userType() QJsonValue; otherwise returns a default constructed QJsonValue.

This function was introduced in Qt 5.0.

See also canConvert() and convert().

fn to_line<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToLineArgs<'largs>

C++ method: QVariant::toLine

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_line(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::line::Line>
2) fn to_line(&self, cpp_utils::AsStruct) -> ::line::Line
C++ method: QLine QVariant::toLine() const

C++ documentation:

Returns the variant as a QLine if the variant has userType() QMetaType::QLine; otherwise returns an invalid QLine.

See also canConvert() and convert().

fn to_line_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToLineFArgs<'largs>

C++ method: QVariant::toLineF

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_line_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::line_f::LineF>
2) fn to_line_f(&self, cpp_utils::AsStruct) -> ::line_f::LineF
C++ method: QLineF QVariant::toLineF() const

C++ documentation:

Returns the variant as a QLineF if the variant has userType() QMetaType::QLineF; otherwise returns an invalid QLineF.

See also canConvert() and convert().

fn to_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToListArgs<'largs>

C++ method: QVariant::toList

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListVariantRef>
2) fn to_list(&self, cpp_utils::AsStruct) -> ::list::ListVariantRef
C++ method: QList<QVariant> QVariant::toList() const

C++ documentation:

Returns the variant as a QVariantList if the variant has userType() QMetaType::QVariantList or QMetaType::QStringList; otherwise returns an empty list.

See also canConvert() and convert().

fn to_locale<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToLocaleArgs<'largs>

C++ method: QVariant::toLocale

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_locale(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::locale::Locale>
2) fn to_locale(&self, cpp_utils::AsStruct) -> ::locale::Locale
C++ method: QLocale QVariant::toLocale() const

C++ documentation:

Returns the variant as a QLocale if the variant has userType() QMetaType::QLocale; otherwise returns an invalid QLocale.

See also canConvert() and convert().

fn to_long_long<'largs, Args>(&'largs self, args: Args) -> i64 where Args: VariantToLongLongArgs<'largs>

C++ method: QVariant::toLongLong

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_long_long(&self, ()) -> i64
2) fn to_long_long(&self, *mut bool) -> i64
C++ method: qlonglong QVariant::toLongLong(bool* ok = ?) const

C++ documentation:

Returns the variant as a long long int if the variant has userType() QMetaType::LongLong, QMetaType::Bool, QMetaType::QByteArray, QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::QString, QMetaType::UInt, or QMetaType::ULongLong; otherwise returns 0.

If ok is non-null: ok is set to true if the value could be converted to an int; otherwise ok is set to false.

See also canConvert() and convert().

fn to_map<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToMapArgs<'largs>

C++ method: QVariant::toMap

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_map(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::map::MapStringRefVariantRef>
2) fn to_map(&self, cpp_utils::AsStruct) -> ::map::MapStringRefVariantRef
C++ method: QMap<QString, QVariant> QVariant::toMap() const

C++ documentation:

Returns the variant as a QMap<QString, QVariant> if the variant has type() QMetaType::QVariantMap; otherwise returns an empty map.

See also canConvert() and convert().

fn to_model_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToModelIndexArgs<'largs>

C++ method: QVariant::toModelIndex

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_model_index(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::model_index::ModelIndex>
2) fn to_model_index(&self, cpp_utils::AsStruct) -> ::model_index::ModelIndex
C++ method: QModelIndex QVariant::toModelIndex() const

C++ documentation:

Returns the variant as a QModelIndex if the variant has userType() QModelIndex; otherwise returns a default constructed QModelIndex.

This function was introduced in Qt 5.0.

See also canConvert(), convert(), and toPersistentModelIndex().

fn to_persistent_model_index<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToPersistentModelIndexArgs<'largs>

C++ method: QVariant::toPersistentModelIndex

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_persistent_model_index(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::persistent_model_index::PersistentModelIndex>
2) fn to_persistent_model_index(&self, cpp_utils::AsStruct) -> ::persistent_model_index::PersistentModelIndex
C++ method: QPersistentModelIndex QVariant::toPersistentModelIndex() const

C++ documentation:

Returns the variant as a QPersistentModelIndex if the variant has userType() QPersistentModelIndex; otherwise returns a default constructed QPersistentModelIndex.

This function was introduced in Qt 5.5.

See also canConvert(), convert(), and toModelIndex().

fn to_point<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToPointArgs<'largs>

C++ method: QVariant::toPoint

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::point::Point>
2) fn to_point(&self, cpp_utils::AsStruct) -> ::point::Point
C++ method: QPoint QVariant::toPoint() const

C++ documentation:

Returns the variant as a QPoint if the variant has userType() QMetaType::QPointF or QMetaType::QPointF; otherwise returns a null QPoint.

See also canConvert() and convert().

fn to_point_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToPointFArgs<'largs>

C++ method: QVariant::toPointF

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_point_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::point_f::PointF>
2) fn to_point_f(&self, cpp_utils::AsStruct) -> ::point_f::PointF
C++ method: QPointF QVariant::toPointF() const

C++ documentation:

Returns the variant as a QPointF if the variant has userType() QMetaType::QPoint or QMetaType::QPointF; otherwise returns a null QPointF.

See also canConvert() and convert().

fn to_real<'largs, Args>(&'largs self, args: Args) -> c_double where Args: VariantToRealArgs<'largs>

C++ method: QVariant::toReal

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_real(&self, ()) -> libc::c_double
2) fn to_real(&self, *mut bool) -> libc::c_double
C++ method: double QVariant::toReal(bool* ok = ?) const

C++ documentation:

Returns the variant as a qreal if the variant has userType() QMetaType::Double, QMetaType::Float, QMetaType::Bool, QMetaType::QByteArray, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, QMetaType::UInt, or QMetaType::ULongLong; otherwise returns 0.0.

If ok is non-null: ok is set to true if the value could be converted to a double; otherwise ok is set to false.

This function was introduced in Qt 4.6.

See also canConvert() and convert().

fn to_rect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToRectArgs<'largs>

C++ method: QVariant::toRect

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_rect(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::rect::Rect>
2) fn to_rect(&self, cpp_utils::AsStruct) -> ::rect::Rect
C++ method: QRect QVariant::toRect() const

C++ documentation:

Returns the variant as a QRect if the variant has userType() QMetaType::QRect; otherwise returns an invalid QRect.

See also canConvert() and convert().

fn to_rect_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToRectFArgs<'largs>

C++ method: QVariant::toRectF

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_rect_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::rect_f::RectF>
2) fn to_rect_f(&self, cpp_utils::AsStruct) -> ::rect_f::RectF
C++ method: QRectF QVariant::toRectF() const

C++ documentation:

Returns the variant as a QRectF if the variant has userType() QMetaType::QRect or QMetaType::QRectF; otherwise returns an invalid QRectF.

See also canConvert() and convert().

fn to_reg_exp<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToRegExpArgs<'largs>

C++ method: QVariant::toRegExp

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_reg_exp(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::reg_exp::RegExp>
2) fn to_reg_exp(&self, cpp_utils::AsStruct) -> ::reg_exp::RegExp
C++ method: QRegExp QVariant::toRegExp() const

C++ documentation:

Returns the variant as a QRegExp if the variant has userType() QMetaType::QRegExp; otherwise returns an empty QRegExp.

This function was introduced in Qt 4.1.

See also canConvert() and convert().

fn to_regular_expression<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToRegularExpressionArgs<'largs>

C++ method: QVariant::toRegularExpression

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_regular_expression(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::regular_expression::RegularExpression>
2) fn to_regular_expression(&self, cpp_utils::AsStruct) -> ::regular_expression::RegularExpression
C++ method: QRegularExpression QVariant::toRegularExpression() const

C++ documentation:

Returns the variant as a QRegularExpression if the variant has userType() QRegularExpression; otherwise returns an empty QRegularExpression.

This function was introduced in Qt 5.0.

See also canConvert() and convert().

fn to_size<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToSizeArgs<'largs>

C++ method: QVariant::toSize

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_size(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size::Size>
2) fn to_size(&self, cpp_utils::AsStruct) -> ::size::Size
C++ method: QSize QVariant::toSize() const

C++ documentation:

Returns the variant as a QSize if the variant has userType() QMetaType::QSize; otherwise returns an invalid QSize.

See also canConvert() and convert().

fn to_size_f<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToSizeFArgs<'largs>

C++ method: QVariant::toSizeF

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_size_f(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::size_f::SizeF>
2) fn to_size_f(&self, cpp_utils::AsStruct) -> ::size_f::SizeF
C++ method: QSizeF QVariant::toSizeF() const

C++ documentation:

Returns the variant as a QSizeF if the variant has userType() QMetaType::QSizeF; otherwise returns an invalid QSizeF.

See also canConvert() and convert().

fn to_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToStringArgs<'largs>

C++ method: QVariant::toString

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn to_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QVariant::toString() const

C++ documentation:

Returns the variant as a QString if the variant has userType() QMetaType::QString, QMetaType::Bool, QMetaType::QByteArray, QMetaType::QChar, QMetaType::QDate, QMetaType::QDateTime, QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::QStringList, QMetaType::QTime, QMetaType::UInt, or QMetaType::ULongLong; otherwise returns an empty string.

See also canConvert() and convert().

fn to_string_list<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToStringListArgs<'largs>

C++ method: QVariant::toStringList

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_string_list(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_list::StringList>
2) fn to_string_list(&self, cpp_utils::AsStruct) -> ::string_list::StringList
C++ method: QStringList QVariant::toStringList() const

C++ documentation:

Returns the variant as a QStringList if the variant has userType() QMetaType::QStringList, QMetaType::QString, or QMetaType::QVariantList of a type that can be converted to QString; otherwise returns an empty list.

See also canConvert() and convert().

fn to_time<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToTimeArgs<'largs>

C++ method: QVariant::toTime

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_time(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::time::Time>
2) fn to_time(&self, cpp_utils::AsStruct) -> ::time::Time
C++ method: QTime QVariant::toTime() const

C++ documentation:

Returns the variant as a QTime if the variant has userType() QMetaType::QTime, QMetaType::QDateTime, or QMetaType::QString; otherwise returns an invalid time.

If the type() is QMetaType::QString, an invalid time will be returned if the string cannot be parsed as a Qt::ISODate format time.

See also canConvert() and convert().

fn to_u_int<'largs, Args>(&'largs self, args: Args) -> c_uint where Args: VariantToUIntArgs<'largs>

C++ method: QVariant::toUInt

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_u_int(&self, ()) -> libc::c_uint
2) fn to_u_int(&self, *mut bool) -> libc::c_uint
C++ method: unsigned int QVariant::toUInt(bool* ok = ?) const

C++ documentation:

Returns the variant as an unsigned int if the variant has userType() QMetaType::UInt, QMetaType::Bool, QMetaType::QByteArray, QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, or QMetaType::ULongLong; otherwise returns 0.

If ok is non-null: ok is set to true if the value could be converted to an unsigned int; otherwise ok is set to false.

Warning: If the value is convertible to a QMetaType::ULongLong but is too large to be represented in an unsigned int, the resulting arithmetic overflow will not be reflected in ok. A simple workaround is to use QString::toUInt().

See also canConvert() and convert().

fn to_u_long_long<'largs, Args>(&'largs self, args: Args) -> u64 where Args: VariantToULongLongArgs<'largs>

C++ method: QVariant::toULongLong

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_u_long_long(&self, ()) -> u64
2) fn to_u_long_long(&self, *mut bool) -> u64
C++ method: qulonglong QVariant::toULongLong(bool* ok = ?) const

C++ documentation:

Returns the variant as as an unsigned long long int if the variant has type() QMetaType::ULongLong, QMetaType::Bool, QMetaType::QByteArray, QMetaType::QChar, QMetaType::Double, QMetaType::Int, QMetaType::LongLong, QMetaType::QString, or QMetaType::UInt; otherwise returns 0.

If ok is non-null: ok is set to true if the value could be converted to an int; otherwise ok is set to false.

See also canConvert() and convert().

fn to_url<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToUrlArgs<'largs>

C++ method: QVariant::toUrl

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_url(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::url::Url>
2) fn to_url(&self, cpp_utils::AsStruct) -> ::url::Url
C++ method: QUrl QVariant::toUrl() const

C++ documentation:

Returns the variant as a QUrl if the variant has userType() QMetaType::QUrl; otherwise returns an invalid QUrl.

See also canConvert() and convert().

fn to_uuid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantToUuidArgs<'largs>

C++ method: QVariant::toUuid

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_uuid(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::uuid::Uuid>
2) fn to_uuid(&self, cpp_utils::AsStruct) -> ::uuid::Uuid
C++ method: QUuid QVariant::toUuid() const

C++ documentation:

Returns the variant as a QUuid if the variant has userType() QUuid; otherwise returns a default constructed QUuid.

This function was introduced in Qt 5.0.

See also canConvert() and convert().

fn type_(&self) -> Type

C++ method: QVariant::Type QVariant::type() const

C++ documentation:

Returns the storage type of the value stored in the variant. Although this function is declared as returning QVariant::Type, the return value should be interpreted as QMetaType::Type. In particular, QVariant::UserType is returned here only if the value is equal or greater than QMetaType::User.

Note that return values in the ranges QVariant::Char through QVariant::RegExp and QVariant::Font through QVariant::Transform correspond to the values in the ranges QMetaType::QChar through QMetaType::QRegExp and QMetaType::QFont through QMetaType::QQuaternion.

Pay particular attention when working with char and QChar variants. Note that there is no QVariant constructor specifically for type char, but there is one for QChar. For a variant of type QChar, this function returns QVariant::Char, which is the same as QMetaType::QChar, but for a variant of type char, this function returns QMetaType::Char, which is not the same as QVariant::Char.

Also note that the types void*, long, short, unsigned long, unsigned short, unsigned char, float, QObject*, and QWidget* are represented in QMetaType::Type but not in QVariant::Type, and they can be returned by this function. However, they are considered to be user defined types when tested against QVariant::Type.

To test whether an instance of QVariant contains a data type that is compatible with the data type you are interested in, use canConvert().

fn type_name(&self) -> *const c_char

C++ method: const char* QVariant::typeName() const

C++ documentation:

Returns the name of the type stored in the variant. The returned strings describe the C++ datatype used to store the data: for example, "QFont", "QString", or "QVariantList". An Invalid variant returns 0.

fn type_to_name(type_id: c_int) -> *const c_char

C++ method: static const char* QVariant::typeToName(int typeId)

C++ documentation:

Converts the int representation of the storage type, typeId, to its string representation.

Returns a null pointer if the type is QMetaType::UnknownType or doesn't exist.

fn user_type(&self) -> c_int

C++ method: int QVariant::userType() const

C++ documentation:

Returns the storage type of the value stored in the variant. For non-user types, this is the same as type().

See also type().