Module qt_core::variant_animation::overloading

Traits

VariantAnimationConnectArgs
This trait represents a set of arguments accepted by VariantAnimation::connect method.

VariantAnimationConnectStaticArgs
This trait represents a set of arguments accepted by VariantAnimation::connect_static method.

VariantAnimationCurrentValueArgs
This trait represents a set of arguments accepted by VariantAnimation::current_value method.

VariantAnimationDestroyedArgs
This trait represents a set of arguments accepted by VariantAnimation::destroyed method.

VariantAnimationDisconnectArgs
This trait represents a set of arguments accepted by VariantAnimation::disconnect method.

VariantAnimationDisconnectStaticArgs
This trait represents a set of arguments accepted by VariantAnimation::disconnect_static method.

VariantAnimationDynamicPropertyNamesArgs
This trait represents a set of arguments accepted by VariantAnimation::dynamic_property_names method.

VariantAnimationEasingCurveArgs
This trait represents a set of arguments accepted by VariantAnimation::easing_curve method.

VariantAnimationEndValueArgs
This trait represents a set of arguments accepted by VariantAnimation::end_value method.

VariantAnimationKeyValueAtArgs
This trait represents a set of arguments accepted by VariantAnimation::key_value_at method.

VariantAnimationKeyValuesArgs
This trait represents a set of arguments accepted by VariantAnimation::key_values method.

VariantAnimationNewArgs
This trait represents a set of arguments accepted by VariantAnimation::new method.

VariantAnimationObjectNameArgs
This trait represents a set of arguments accepted by VariantAnimation::object_name method.

VariantAnimationPropertyArgs
This trait represents a set of arguments accepted by VariantAnimation::property method.

VariantAnimationStartArgs
This trait represents a set of arguments accepted by VariantAnimation::start method.

VariantAnimationStartTimerArgs
This trait represents a set of arguments accepted by VariantAnimation::start_timer method.

VariantAnimationStartValueArgs
This trait represents a set of arguments accepted by VariantAnimation::start_value method.

VariantAnimationTrArgs
This trait represents a set of arguments accepted by VariantAnimation::tr method.

VariantAnimationTrUtf8Args
This trait represents a set of arguments accepted by VariantAnimation::tr_utf8 method.