Struct qt_core::variant_animation::VariantAnimation

pub struct VariantAnimation {
    // some fields omitted
}

C++ type: QVariantAnimation

C++ documentation:

The QVariantAnimation class provides an abstract base class for animations.

This class is part of The Animation Framework. It serves as a base class for property and item animations, with functions for shared functionality.

QVariantAnimation cannot be used directly as it is an abstract class; it has a pure virtual method called updateCurrentValue(). The class performs interpolation over QVariants, but leaves using the interpolated values to its subclasses. Currently, Qt provides QPropertyAnimation, which animates Qt properties. See the QPropertyAnimation class description if you wish to animate such properties.

You can then set start and end values for the property by calling setStartValue() and setEndValue(), and finally call start() to start the animation. QVariantAnimation will interpolate the property of the target object and emit valueChanged(). To react to a change in the current value you have to reimplement the updateCurrentValue() virtual function.

It is also possible to set values at specified steps situated between the start and end value. The interpolation will then touch these points at the specified steps. Note that the start and end values are defined as the key values at 0.0 and 1.0.

There are two ways to affect how QVariantAnimation interpolates the values. You can set an easing curve by calling setEasingCurve(), and configure the duration by calling setDuration(). You can change how the QVariants are interpolated by creating a subclass of QVariantAnimation, and reimplementing the virtual interpolated() function.

Subclassing QVariantAnimation can be an alternative if you have QVariants that you do not wish to declare as Qt properties. Note, however, that you in most cases will be better off declaring your QVariant as a property.

Not all QVariant types are supported. Below is a list of currently supported QVariant types:

Int

UInt

Double

Float

QLine

QLineF

QPoint

QPointF

QSize

QSizeF

QRect

QRectF

QColor

If you need to interpolate other variant types, including custom types, you have to implement interpolation for these yourself. To do this, you can register an interpolator function for a given type. This function takes 3 parameters: the start value, the end value, and the current progress.

Example:

QVariant myColorInterpolator(const QColor &start, const QColor &end, qreal progress)
{
    ...
    return QColor(...);
}
...
qRegisterAnimationInterpolator<QColor>(myColorInterpolator);

Another option is to reimplement interpolated(), which returns interpolation values for the value being interpolated.

Methods

impl VariantAnimation

fn block_signals(&mut self, b: bool) -> bool

C++ method: bool QVariantAnimation::blockSignals(bool b)

C++ documentation:

If block is true, signals emitted by this object are blocked (i.e., emitting a signal will not invoke anything connected to it). If block is false, no such blocking will occur.

The return value is the previous value of signalsBlocked().

Note that the destroyed() signal will be emitted even if the signals for this object have been blocked.

See also signalsBlocked().

fn children<'l0>(&'l0 self) -> &'l0 ListObjectMutPtr

C++ method: const QList<QObject*>& QVariantAnimation::children() const

C++ documentation:

Returns a list of child objects. The QObjectList class is defined in the <QObject> header file as the following:

typedef QList<QObject*> QObjectList;

The first child added is the first object in the list and the last child added is the last object in the list, i.e. new children are appended at the end.

Note that the list order changes when QWidget children are raised or lowered. A widget that is raised becomes the last object in the list, and a widget that is lowered becomes the first object in the list.

See also findChild(), findChildren(), parent(), and setParent().

fn connect<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationConnectArgs<'largs>

C++ method: QVariantAnimation::connect

This is an overloaded function. Available variants:

Rust arguments:
1) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect(&self, (*const ::object::Object, *const libc::c_char, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: QMetaObject::Connection QVariantAnimation::connect(const QObject* sender, const char* signal, const char* member, Qt::ConnectionType type = ?) const

C++ documentation:

This function overloads connect().

Connects signal from the sender object to this object's method.

Equivalent to connect(sender, signal, this, method, type).

Every connection you make emits a signal, so duplicate connections emit two signals. You can break a connection using disconnect().

Note: This function is thread-safe.

See also disconnect().

fn connect_static<Args>(args: Args) -> Args::ReturnType where Args: VariantAnimationConnectStaticArgs

C++ method: QVariantAnimation::connect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QVariantAnimation::connect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& method, Qt::ConnectionType type = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

The Connection handle will be invalid if it cannot create the connection, for example, the parameters were invalid. You can check if the QMetaObject::Connection is valid by casting it to a bool.

This function works in the same way as connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type) but it uses QMetaMethod to specify signal and method.

This function was introduced in Qt 4.8.

See also connect(const QObject *sender, const char *signal, const QObject *receiver, const char *method, Qt::ConnectionType type).

Variant 2

Rust arguments:
1) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
2) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsBox)) -> cpp_utils::CppBox<::meta_object::Connection>
3) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, cpp_utils::AsStruct)) -> ::meta_object::Connection
4) fn connect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char, ::qt::ConnectionType, cpp_utils::AsStruct)) -> ::meta_object::Connection
C++ method: static QMetaObject::Connection QVariantAnimation::connect(const QObject* sender, const char* signal, const QObject* receiver, const char* member, Qt::ConnectionType arg5 = ?)

C++ documentation:

Creates a connection of the given type from the signal in the sender object to the method in the receiver object. Returns a handle to the connection that can be used to disconnect it later.

You must use the SIGNAL() and SLOT() macros when specifying the signal and the method, for example:

QLabel label = new QLabel;
QScrollBar scrollBar = new QScrollBar;
QObject::connect(scrollBar, SIGNAL(valueChanged(int)),
                 label,  SLOT(setNum(int)));

This example ensures that the label always displays the current scroll bar value. Note that the signal and slots parameters must not contain any variable names, only the type. E.g. the following would not work and return false:

// WRONG
QObject::connect(scrollBar, SIGNAL(valueChanged(int value)),
                 label, SLOT(setNum(int value)));

A signal can also be connected to another signal:

class MyWidget : public QWidget
{
    Q_OBJECT

public: MyWidget();

signals: void buttonClicked();

private: QPushButton *myButton; };

MyWidget::MyWidget() { myButton = new QPushButton(this); connect(myButton, SIGNAL(clicked()), this, SIGNAL(buttonClicked())); }
In this example, the MyWidget constructor relays a signal from a private member variable, and makes it available under a name that relates to MyWidget.

A signal can be connected to many slots and signals. Many signals can be connected to one slot.

If a signal is connected to several slots, the slots are activated in the same order in which the connections were made, when the signal is emitted.

The function returns a QMetaObject::Connection that represents a handle to a connection if it successfully connects the signal to the slot. The connection handle will be invalid if it cannot create the connection, for example, if QObject is unable to verify the existence of either signal or method, or if their signatures aren't compatible. You can check if the handle is valid by casting it to a bool.

By default, a signal is emitted for every connection you make; two signals are emitted for duplicate connections. You can break all of these connections with a single disconnect() call. If you pass the Qt::UniqueConnection type, the connection will only be made if it is not a duplicate. If there is already a duplicate (exact same signal to the exact same slot on the same objects), the connection will fail and connect will return an invalid QMetaObject::Connection.

The optional type parameter describes the type of connection to establish. In particular, it determines whether a particular signal is delivered to a slot immediately or queued for delivery at a later time. If the signal is queued, the parameters must be of types that are known to Qt's meta-object system, because Qt needs to copy the arguments to store them in an event behind the scenes. If you try to use a queued connection and get the error message

QObject::connect: Cannot queue arguments of type 'MyType'
(Make sure 'MyType' is registered using qRegisterMetaType().)

call qRegisterMetaType() to register the data type before you establish the connection.

Note: This function is thread-safe.

See also disconnect(), sender(), qRegisterMetaType(), and Q_DECLARE_METATYPE().

fn current_loop(&self) -> c_int

C++ method: int QVariantAnimation::currentLoop() const

C++ documentation:

This property holds the current loop of the animation.

This property describes the current loop of the animation. By default, the animation's loop count is 1, and so the current loop will always be 0. If the loop count is 2 and the animation runs past its duration, it will automatically rewind and restart at current time 0, and current loop 1, and so on.

When the current loop changes, QAbstractAnimation emits the currentLoopChanged() signal.

Access functions:

int 	currentLoop() const

Notifier signal:

void 	currentLoopChanged(int currentLoop)

fn current_loop_changed(&mut self, current_loop: c_int)

C++ method: void QVariantAnimation::currentLoopChanged(int currentLoop)

C++ documentation:

This property holds the current loop of the animation.

This property describes the current loop of the animation. By default, the animation's loop count is 1, and so the current loop will always be 0. If the loop count is 2 and the animation runs past its duration, it will automatically rewind and restart at current time 0, and current loop 1, and so on.

When the current loop changes, QAbstractAnimation emits the currentLoopChanged() signal.

Access functions:

int 	currentLoop() const

Notifier signal:

void 	currentLoopChanged(int currentLoop)

fn current_loop_time(&self) -> c_int

C++ method: int QVariantAnimation::currentLoopTime() const

C++ documentation:

Returns the current time inside the current loop. It can go from 0 to duration().

See also duration() and currentTime.

fn current_time(&self) -> c_int

C++ method: int QVariantAnimation::currentTime() const

C++ documentation:

This property holds the current time and progress of the animation.

This property describes the animation's current time. You can change the current time by calling setCurrentTime, or you can call start() and let the animation run, setting the current time automatically as the animation progresses.

The animation's current time starts at 0, and ends at totalDuration().

Access functions:

int 	currentTime() const

void 	setCurrentTime(int msecs)

See also loopCount and currentLoopTime().

fn current_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationCurrentValueArgs<'largs>

C++ method: QVariantAnimation::currentValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn current_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn current_value(&self, cpp_utils::AsStruct) -> ::variant::Variant
C++ method: QVariant QVariantAnimation::currentValue() const

C++ documentation:

This property holds the current value of the animation.

This property describes the current value; an interpolated value between the start value and the end value, using the current time for progress. The value itself is obtained from interpolated(), which is called repeatedly as the animation is running.

QVariantAnimation calls the virtual updateCurrentValue() function when the current value changes. This is particularly useful for subclasses that need to track updates. For example, QPropertyAnimation uses this function to animate Qt properties.

Access functions:

QVariant 	currentValue() const

Notifier signal:

void 	valueChanged(const QVariant & value)

See also startValue and endValue.

fn delete_later(&mut self)

C++ method: void QVariantAnimation::deleteLater()

C++ documentation:

Schedules this object for deletion.

The object will be deleted when control returns to the event loop. If the event loop is not running when this function is called (e.g. deleteLater() is called on an object before QCoreApplication::exec()), the object will be deleted once the event loop is started. If deleteLater() is called after the main event loop has stopped, the object will not be deleted. Since Qt 4.8, if deleteLater() is called on an object that lives in a thread with no running event loop, the object will be destroyed when the thread finishes.

Note that entering and leaving a new event loop (e.g., by opening a modal dialog) will not perform the deferred deletion; for the object to be deleted, the control must return to the event loop from which deleteLater() was called.

Note: It is safe to call this function more than once; when the first deferred deletion event is delivered, any pending events for the object are removed from the event queue.

See also destroyed() and QPointer.

fn destroyed<'largs, Args>(&'largs mut self, args: Args) where Args: VariantAnimationDestroyedArgs<'largs>

C++ method: QVariantAnimation::destroyed

This is an overloaded function. Available variants:

Rust arguments:
1) fn destroyed(&mut self, *mut ::object::Object) -> ()
2) fn destroyed(&mut self, ()) -> ()
C++ method: void QVariantAnimation::destroyed(QObject* arg1 = ?)

C++ documentation:

This signal is emitted immediately before the object obj is destroyed, and can not be blocked.

All the objects's children are destroyed immediately after this signal is emitted.

See also deleteLater() and QPointer.

fn direction(&self) -> Direction

C++ method: QAbstractAnimation::Direction QVariantAnimation::direction() const

C++ documentation:

This property holds the direction of the animation when it is in Running state.

This direction indicates whether the time moves from 0 towards the animation duration, or from the value of the duration and towards 0 after start() has been called.

By default, this property is set to Forward.

Access functions:

Direction 	direction() const

void 	setDirection(Direction direction)

Notifier signal:

void 	directionChanged(QAbstractAnimation::Direction newDirection)

fn direction_changed(&mut self, arg1: Direction)

C++ method: void QVariantAnimation::directionChanged(QAbstractAnimation::Direction arg1)

C++ documentation:

This property holds the direction of the animation when it is in Running state.

This direction indicates whether the time moves from 0 towards the animation duration, or from the value of the duration and towards 0 after start() has been called.

By default, this property is set to Forward.

Access functions:

Direction 	direction() const

void 	setDirection(Direction direction)

Notifier signal:

void 	directionChanged(QAbstractAnimation::Direction newDirection)

fn disconnect<'largs, Args>(&'largs self, args: Args) -> bool where Args: VariantAnimationDisconnectArgs<'largs>

C++ method: QVariantAnimation::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn disconnect(&self, *const ::object::Object) -> bool
2) fn disconnect(&self, (*const ::object::Object, *const libc::c_char)) -> bool
C++ method: bool QVariantAnimation::disconnect(const QObject* receiver, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects all signals in this object from receiver's method.

A signal-slot connection is removed when either of the objects involved are destroyed.

Variant 2

Rust arguments:
1) fn disconnect(&self, *const libc::c_char) -> bool
2) fn disconnect(&self, (*const libc::c_char, *const ::object::Object)) -> bool
3) fn disconnect(&self, (*const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
4) fn disconnect(&self, ()) -> bool
C++ method: bool QVariantAnimation::disconnect(const char* signal = ?, const QObject* receiver = ?, const char* member = ?) const

C++ documentation:

This function overloads disconnect().

Disconnects signal from method of receiver.

A signal-slot connection is removed when either of the objects involved are destroyed.

Note: This function is thread-safe.

fn disconnect_static<Args>(args: Args) -> bool where Args: VariantAnimationDisconnectStaticArgs

C++ method: QVariantAnimation::disconnect

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn disconnect_static(&::meta_object::Connection) -> bool
C++ method: static bool QVariantAnimation::disconnect(const QMetaObject::Connection& arg1)

C++ documentation:

Disconnect a connection.

If the connection is invalid or has already been disconnected, do nothing and return false.

See also connect().

Variant 2

Rust arguments: fn disconnect_static((*const ::object::Object, &::meta_method::MetaMethod, *const ::object::Object, &::meta_method::MetaMethod)) -> bool
C++ method: static bool QVariantAnimation::disconnect(const QObject* sender, const QMetaMethod& signal, const QObject* receiver, const QMetaMethod& member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

This function provides the same possibilities like disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method) but uses QMetaMethod to represent the signal and the method to be disconnected.

Additionally this function returnsfalse and no signals and slots disconnected if:

signal is not a member of sender class or one of its parent classes.

method is not a member of receiver class or one of its parent classes.

signal instance represents not a signal.

QMetaMethod() may be used as wildcard in the meaning "any signal" or "any slot in receiving object". In the same way 0 can be used for receiver in the meaning "any receiving object". In this case method should also be QMetaMethod(). sender parameter should be never 0.

This function was introduced in Qt 4.8.

See also disconnect(const QObject *sender, const char *signal, const QObject *receiver, const char *method).

Variant 3

Rust arguments: fn disconnect_static((*const ::object::Object, *const libc::c_char, *const ::object::Object, *const libc::c_char)) -> bool
C++ method: static bool QVariantAnimation::disconnect(const QObject* sender, const char* signal, const QObject* receiver, const char* member)

C++ documentation:

Disconnects signal in object sender from method in object receiver. Returns true if the connection is successfully broken; otherwise returns false.

A signal-slot connection is removed when either of the objects involved are destroyed.

disconnect() is typically used in three ways, as the following examples demonstrate.

Disconnect everything connected to an object's signals:
disconnect(myObject, 0, 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect();

Disconnect everything connected to a specific signal:
disconnect(myObject, SIGNAL(mySignal()), 0, 0);

equivalent to the non-static overloaded function

myObject->disconnect(SIGNAL(mySignal()));

Disconnect a specific receiver:
disconnect(myObject, 0, myReceiver, 0);

equivalent to the non-static overloaded function

myObject->disconnect(myReceiver);

0 may be used as a wildcard, meaning "any signal", "any receiving object", or "any slot in the receiving object", respectively.

The sender may never be 0. (You cannot disconnect signals from more than one object in a single call.)

If signal is 0, it disconnects receiver and method from any signal. If not, only the specified signal is disconnected.

If receiver is 0, it disconnects anything connected to signal. If not, slots in objects other than receiver are not disconnected.

If method is 0, it disconnects anything that is connected to receiver. If not, only slots named method will be disconnected, and all other slots are left alone. The method must be 0 if receiver is left out, so you cannot disconnect a specifically-named slot on all objects.

Note: This function is thread-safe.

See also connect().

fn dump_object_info(&mut self)

C++ method: void QVariantAnimation::dumpObjectInfo()

C++ documentation:

Dumps information about signal connections, etc. for this object to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectTree().

fn dump_object_tree(&mut self)

C++ method: void QVariantAnimation::dumpObjectTree()

C++ documentation:

Dumps a tree of children to the debug output.

This function is useful for debugging, but does nothing if the library has been compiled in release mode (i.e. without debugging information).

See also dumpObjectInfo().

fn duration(&self) -> c_int

C++ method: virtual int QVariantAnimation::duration() const

C++ documentation:

This property holds the duration of the animation.

This property describes the duration in milliseconds of the animation. The default duration is 250 milliseconds.

Access functions:

virtual int 	duration() const

void 	setDuration(int msecs)

See also QAbstractAnimation::duration().

fn dynamic_property_names<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationDynamicPropertyNamesArgs<'largs>

C++ method: QVariantAnimation::dynamicPropertyNames

This is an overloaded function. Available variants:

Rust arguments:
1) fn dynamic_property_names(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::list::ListByteArrayRef>
2) fn dynamic_property_names(&self, cpp_utils::AsStruct) -> ::list::ListByteArrayRef
C++ method: QList<QByteArray> QVariantAnimation::dynamicPropertyNames() const

C++ documentation:

Returns the names of all properties that were dynamically added to the object using setProperty().

This function was introduced in Qt 4.2.

fn easing_curve<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationEasingCurveArgs<'largs>

C++ method: QVariantAnimation::easingCurve

This is an overloaded function. Available variants:

Rust arguments:
1) fn easing_curve(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::easing_curve::EasingCurve>
2) fn easing_curve(&self, cpp_utils::AsStruct) -> ::easing_curve::EasingCurve
C++ method: QEasingCurve QVariantAnimation::easingCurve() const

C++ documentation:

This property holds the easing curve of the animation.

This property defines the easing curve of the animation. By default, a linear easing curve is used, resulting in linear interpolation. Other curves are provided, for instance, QEasingCurve::InCirc, which provides a circular entry curve. Another example is QEasingCurve::InOutElastic, which provides an elastic effect on the values of the interpolated variant.

QVariantAnimation will use the QEasingCurve::valueForProgress() to transform the "normalized progress" (currentTime / totalDuration) of the animation into the effective progress actually used by the animation. It is this effective progress that will be the progress when interpolated() is called. Also, the steps in the keyValues are referring to this effective progress.

The easing curve is used with the interpolator, the interpolated() virtual function, and the animation's duration to control how the current value changes as the animation progresses.

Access functions:

QEasingCurve 	easingCurve() const

void 	setEasingCurve(const QEasingCurve & easing)

fn end_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationEndValueArgs<'largs>

C++ method: QVariantAnimation::endValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn end_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn end_value(&self, cpp_utils::AsStruct) -> ::variant::Variant
C++ method: QVariant QVariantAnimation::endValue() const

C++ documentation:

This property holds the end value of the animation.

This property describes the end value of the animation.

Access functions:

QVariant 	endValue() const

void 	setEndValue(const QVariant & value)

See also startValue.

fn event_filter(&mut self, arg1: *mut Object, arg2: *mut Event) -> bool

C++ method: virtual bool QVariantAnimation::eventFilter(QObject* arg1, QEvent* arg2)

C++ documentation:

Filters events if this object has been installed as an event filter for the watched object.

In your reimplementation of this function, if you want to filter the event out, i.e. stop it being handled further, return true; otherwise return false.

Example:

class MainWindow : public QMainWindow
{
public:
    MainWindow();

protected: bool eventFilter(QObject obj, QEvent ev);

private: QTextEdit *textEdit; };

MainWindow::MainWindow() { textEdit = new QTextEdit; setCentralWidget(textEdit);

textEdit->installEventFilter(this); }

bool MainWindow::eventFilter(QObject obj, QEvent event) { if (obj == textEdit) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent>(event); qDebug() << "Ate key press" << keyEvent->key(); return true; } else { return false; } } else { // pass the event on to the parent class return QMainWindow::eventFilter(obj, event); } }
Notice in the example above that unhandled events are passed to the base class's eventFilter() function, since the base class might have reimplemented eventFilter() for its own internal purposes.

Warning: If you delete the receiver object in this function, be sure to return true. Otherwise, Qt will forward the event to the deleted object and the program might crash.

See also installEventFilter().

fn finished(&mut self)

C++ method: void QVariantAnimation::finished()

C++ documentation:

QAbstractAnimation emits this signal after the animation has stopped and has reached the end.

This signal is emitted after stateChanged().

See also stateChanged().

fn group(&self) -> *mut AnimationGroup

C++ method: QAnimationGroup* QVariantAnimation::group() const

C++ documentation:

If this animation is part of a QAnimationGroup, this function returns a pointer to the group; otherwise, it returns 0.

See also QAnimationGroup::addAnimation().

fn inherits(&self, classname: *const c_char) -> bool

C++ method: bool QVariantAnimation::inherits(const char* classname) const

C++ documentation:

Returns true if this object is an instance of a class that inherits className or a QObject subclass that inherits className; otherwise returns false.

A class is considered to inherit itself.

Example:

QTimer *timer = new QTimer;         // QTimer inherits QObject
timer->inherits("QTimer");          // returns true
timer->inherits("QObject");         // returns true
timer->inherits("QAbstractButton"); // returns false

// QVBoxLayout inherits QObject and QLayoutItem QVBoxLayout *layout = new QVBoxLayout; layout->inherits("QObject"); // returns true layout->inherits("QLayoutItem"); // returns true (even though QLayoutItem is not a QObject)
If you need to determine whether an object is an instance of a particular class for the purpose of casting it, consider using qobject_cast<Type *>(object) instead.

See also metaObject() and qobject_cast().

fn install_event_filter(&mut self, arg1: *mut Object)

C++ method: void QVariantAnimation::installEventFilter(QObject* arg1)

C++ documentation:

Installs an event filter filterObj on this object. For example:

monitoredObj->installEventFilter(filterObj);

An event filter is an object that receives all events that are sent to this object. The filter can either stop the event or forward it to this object. The event filter filterObj receives events via its eventFilter() function. The eventFilter() function must return true if the event should be filtered, (i.e. stopped); otherwise it must return false.

If multiple event filters are installed on a single object, the filter that was installed last is activated first.

Here's a KeyPressEater class that eats the key presses of its monitored objects:

class KeyPressEater : public QObject
{
    Q_OBJECT
    ...

protected: bool eventFilter(QObject obj, QEvent event); };

bool KeyPressEater::eventFilter(QObject obj, QEvent event) { if (event->type() == QEvent::KeyPress) { QKeyEvent keyEvent = static_cast<QKeyEvent >(event); qDebug("Ate key press %d", keyEvent->key()); return true; } else { // standard event processing return QObject::eventFilter(obj, event); } }
And here's how to install it on two widgets:

KeyPressEater keyPressEater = new KeyPressEater(this);
QPushButton pushButton = new QPushButton(this);
QListView *listView = new QListView(this);

pushButton->installEventFilter(keyPressEater); listView->installEventFilter(keyPressEater);
The QShortcut class, for example, uses this technique to intercept shortcut key presses.

Warning: If you delete the receiver object in your eventFilter() function, be sure to return true. If you return false, Qt sends the event to the deleted object and the program will crash.

Note that the filtering object must be in the same thread as this object. If filterObj is in a different thread, this function does nothing. If either filterObj or this object are moved to a different thread after calling this function, the event filter will not be called until both objects have the same thread affinity again (it is not removed).

See also removeEventFilter(), eventFilter(), and event().

fn is_widget_type(&self) -> bool

C++ method: bool QVariantAnimation::isWidgetType() const

C++ documentation:

Returns true if the object is a widget; otherwise returns false.

Calling this function is equivalent to calling inherits("QWidget"), except that it is much faster.

fn is_window_type(&self) -> bool

C++ method: bool QVariantAnimation::isWindowType() const

C++ documentation:

Returns true if the object is a window; otherwise returns false.

Calling this function is equivalent to calling inherits("QWindow"), except that it is much faster.

fn key_value_at<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationKeyValueAtArgs<'largs>

C++ method: QVariantAnimation::keyValueAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_value_at(&self, (libc::c_double, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn key_value_at(&self, (libc::c_double, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QVariantAnimation::keyValueAt(double step) const

C++ documentation:

Returns the key frame value for the given step. The given step must be in the range 0 to 1. If there is no KeyValue for step, it returns an invalid QVariant.

See also keyValues() and setKeyValueAt().

fn key_values<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationKeyValuesArgs<'largs>

C++ method: QVariantAnimation::keyValues

This is an overloaded function. Available variants:

Rust arguments:
1) fn key_values(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorPairCDoubleVariantRefRef>
2) fn key_values(&self, cpp_utils::AsStruct) -> ::vector::VectorPairCDoubleVariantRefRef
C++ method: QVector<QPair<double, QVariant>> QVariantAnimation::keyValues() const

C++ documentation:

Returns the key frames of this animation.

See also keyValueAt() and setKeyValues().

fn kill_timer(&mut self, id: c_int)

C++ method: void QVariantAnimation::killTimer(int id)

C++ documentation:

Kills the timer with timer identifier, id.

The timer identifier is returned by startTimer() when a timer event is started.

See also timerEvent() and startTimer().

fn loop_count(&self) -> c_int

C++ method: int QVariantAnimation::loopCount() const

C++ documentation:

This property holds the loop count of the animation.

This property describes the loop count of the animation as an integer. By default this value is 1, indicating that the animation should run once only, and then stop. By changing it you can let the animation loop several times. With a value of 0, the animation will not run at all, and with a value of -1, the animation will loop forever until stopped. It is not supported to have loop on an animation that has an undefined duration. It will only run once.

Access functions:

int 	loopCount() const

void 	setLoopCount(int loopCount)

fn meta_object(&self) -> *const MetaObject

C++ method: virtual const QMetaObject* QVariantAnimation::metaObject() const

fn move_to_thread(&mut self, thread: *mut Thread)

C++ method: void QVariantAnimation::moveToThread(QThread* thread)

C++ documentation:

Changes the thread affinity for this object and its children. The object cannot be moved if it has a parent. Event processing will continue in the targetThread.

To move an object to the main thread, use QApplication::instance() to retrieve a pointer to the current application, and then use QApplication::thread() to retrieve the thread in which the application lives. For example:

myObject->moveToThread(QApplication::instance()->thread());

If targetThread is zero, all event processing for this object and its children stops.

Note that all active timers for the object will be reset. The timers are first stopped in the current thread and restarted (with the same interval) in the targetThread. As a result, constantly moving an object between threads can postpone timer events indefinitely.

A QEvent::ThreadChange event is sent to this object just before the thread affinity is changed. You can handle this event to perform any special processing. Note that any new events that are posted to this object will be handled in the targetThread.

Warning: This function is not thread-safe; the current thread must be same as the current thread affinity. In other words, this function can only "push" an object from the current thread to another thread, it cannot "pull" an object from any arbitrary thread to the current thread.

See also thread().

fn new<Args>(args: Args) -> Args::ReturnType where Args: VariantAnimationNewArgs

C++ method: QVariantAnimation::QVariantAnimation

This is an overloaded function. Available variants:

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::variant_animation::VariantAnimation
2) fn new((*mut ::object::Object, cpp_utils::AsStruct)) -> ::variant_animation::VariantAnimation
3) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::variant_animation::VariantAnimation>
4) fn new((*mut ::object::Object, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant_animation::VariantAnimation>
C++ method: [constructor] void QVariantAnimation::QVariantAnimation(QObject* parent = ?)

fn object_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationObjectNameArgs<'largs>

C++ method: QVariantAnimation::objectName

This is an overloaded function. Available variants:

Rust arguments:
1) fn object_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn object_name(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QVariantAnimation::objectName() const

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn parent(&self) -> *mut Object

C++ method: QObject* QVariantAnimation::parent() const

C++ documentation:

Returns a pointer to the parent object.

See also setParent() and children().

fn pause(&mut self)

C++ method: void QVariantAnimation::pause()

C++ documentation:

Pauses the animation. When the animation is paused, state() returns Paused. The value of currentTime will remain unchanged until resume() or start() is called. If you want to continue from the current time, call resume().

See also start(), state(), and resume().

fn property<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationPropertyArgs<'largs>

C++ method: QVariantAnimation::property

This is an overloaded function. Available variants:

Rust arguments:
1) fn property(&self, (*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::variant::Variant>
2) fn property(&self, (*const libc::c_char, cpp_utils::AsStruct)) -> ::variant::Variant
C++ method: QVariant QVariantAnimation::property(const char* name) const

C++ documentation:

Returns the value of the object's name property.

If no such property exists, the returned variant is invalid.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

See also setProperty(), QVariant::isValid(), metaObject(), and dynamicPropertyNames().

fn qt_metacall(&mut self, arg1: Call, arg2: c_int, arg3: *mut *mut c_void) -> c_int

C++ method: virtual int QVariantAnimation::qt_metacall(QMetaObject::Call arg1, int arg2, void** arg3)

fn qt_metacast(&mut self, arg1: *const c_char) -> *mut c_void

C++ method: virtual void* QVariantAnimation::qt_metacast(const char* arg1)

fn register_user_data() -> c_uint

C++ method: static unsigned int QVariantAnimation::registerUserData()

fn remove_event_filter(&mut self, arg1: *mut Object)

C++ method: void QVariantAnimation::removeEventFilter(QObject* arg1)

C++ documentation:

Removes an event filter object obj from this object. The request is ignored if such an event filter has not been installed.

All event filters for this object are automatically removed when this object is destroyed.

It is always safe to remove an event filter, even during event filter activation (i.e. from the eventFilter() function).

See also installEventFilter(), eventFilter(), and event().

fn resume(&mut self)

C++ method: void QVariantAnimation::resume()

C++ documentation:

Resumes the animation after it was paused. When the animation is resumed, it emits the resumed() and stateChanged() signals. The currenttime is not changed.

See also start(), pause(), and state().

fn set_current_time(&mut self, msecs: c_int)

C++ method: void QVariantAnimation::setCurrentTime(int msecs)

C++ documentation:

This property holds the current time and progress of the animation.

This property describes the animation's current time. You can change the current time by calling setCurrentTime, or you can call start() and let the animation run, setting the current time automatically as the animation progresses.

The animation's current time starts at 0, and ends at totalDuration().

Access functions:

int 	currentTime() const

void 	setCurrentTime(int msecs)

See also loopCount and currentLoopTime().

fn set_direction(&mut self, direction: Direction)

C++ method: void QVariantAnimation::setDirection(QAbstractAnimation::Direction direction)

C++ documentation:

This property holds the direction of the animation when it is in Running state.

This direction indicates whether the time moves from 0 towards the animation duration, or from the value of the duration and towards 0 after start() has been called.

By default, this property is set to Forward.

Access functions:

Direction 	direction() const

void 	setDirection(Direction direction)

Notifier signal:

void 	directionChanged(QAbstractAnimation::Direction newDirection)

fn set_duration(&mut self, msecs: c_int)

C++ method: void QVariantAnimation::setDuration(int msecs)

C++ documentation:

This property holds the duration of the animation.

This property describes the duration in milliseconds of the animation. The default duration is 250 milliseconds.

Access functions:

virtual int 	duration() const

void 	setDuration(int msecs)

See also QAbstractAnimation::duration().

fn set_easing_curve(&mut self, easing: &EasingCurve)

C++ method: void QVariantAnimation::setEasingCurve(const QEasingCurve& easing)

C++ documentation:

This property holds the easing curve of the animation.

This property defines the easing curve of the animation. By default, a linear easing curve is used, resulting in linear interpolation. Other curves are provided, for instance, QEasingCurve::InCirc, which provides a circular entry curve. Another example is QEasingCurve::InOutElastic, which provides an elastic effect on the values of the interpolated variant.

QVariantAnimation will use the QEasingCurve::valueForProgress() to transform the "normalized progress" (currentTime / totalDuration) of the animation into the effective progress actually used by the animation. It is this effective progress that will be the progress when interpolated() is called. Also, the steps in the keyValues are referring to this effective progress.

The easing curve is used with the interpolator, the interpolated() virtual function, and the animation's duration to control how the current value changes as the animation progresses.

Access functions:

QEasingCurve 	easingCurve() const

void 	setEasingCurve(const QEasingCurve & easing)

fn set_end_value(&mut self, value: &Variant)

C++ method: void QVariantAnimation::setEndValue(const QVariant& value)

C++ documentation:

This property holds the end value of the animation.

This property describes the end value of the animation.

Access functions:

QVariant 	endValue() const

void 	setEndValue(const QVariant & value)

See also startValue.

fn set_key_value_at(&mut self, step: c_double, value: &Variant)

C++ method: void QVariantAnimation::setKeyValueAt(double step, const QVariant& value)

C++ documentation:

Creates a key frame at the given step with the given value. The given step must be in the range 0 to 1.

See also setKeyValues() and keyValueAt().

fn set_key_values(&mut self, values: &VectorPairCDoubleVariantRefRef)

C++ method: void QVariantAnimation::setKeyValues(const QVector<QPair<double, QVariant>>& values)

C++ documentation:

Replaces the current set of key frames with the given keyValues. the step of the key frames must be in the range 0 to 1.

See also keyValues() and keyValueAt().

fn set_loop_count(&mut self, loop_count: c_int)

C++ method: void QVariantAnimation::setLoopCount(int loopCount)

C++ documentation:

This property holds the loop count of the animation.

This property describes the loop count of the animation as an integer. By default this value is 1, indicating that the animation should run once only, and then stop. By changing it you can let the animation loop several times. With a value of 0, the animation will not run at all, and with a value of -1, the animation will loop forever until stopped. It is not supported to have loop on an animation that has an undefined duration. It will only run once.

Access functions:

int 	loopCount() const

void 	setLoopCount(int loopCount)

fn set_object_name(&mut self, name: &String)

C++ method: void QVariantAnimation::setObjectName(const QString& name)

C++ documentation:

This property holds the name of this object.

You can find an object by name (and type) using findChild(). You can find a set of objects with findChildren().

qDebug("MyClass::setPrecision(): (%s) invalid precision %f",
       qPrintable(objectName()), newPrecision);

By default, this property contains an empty string.

Access functions:

QString 	objectName() const

void 	setObjectName(const QString & name)

Notifier signal:

void 	objectNameChanged(const QString & objectName)	[see note below]

Note: This is a private signal. It can be used in signal connections but cannot be emitted by the user.

See also metaObject() and QMetaObject::className().

fn set_parent(&mut self, arg1: *mut Object)

C++ method: void QVariantAnimation::setParent(QObject* arg1)

C++ documentation:

Makes the object a child of parent.

See also parent() and children().

fn set_paused(&mut self, arg1: bool)

C++ method: void QVariantAnimation::setPaused(bool arg1)

C++ documentation:

If paused is true, the animation is paused. If paused is false, the animation is resumed.

See also state(), pause(), and resume().

fn set_property(&mut self, name: *const c_char, value: &Variant) -> bool

C++ method: bool QVariantAnimation::setProperty(const char* name, const QVariant& value)

C++ documentation:

Sets the value of the object's name property to value.

If the property is defined in the class using Q_PROPERTY then true is returned on success and false otherwise. If the property is not defined using Q_PROPERTY, and therefore not listed in the meta-object, it is added as a dynamic property and false is returned.

Information about all available properties is provided through the metaObject() and dynamicPropertyNames().

Dynamic properties can be queried again using property() and can be removed by setting the property value to an invalid QVariant. Changing the value of a dynamic property causes a QDynamicPropertyChangeEvent to be sent to the object.

Note: Dynamic properties starting with "_q_" are reserved for internal purposes.

See also property(), metaObject(), and dynamicPropertyNames().

fn set_start_value(&mut self, value: &Variant)

C++ method: void QVariantAnimation::setStartValue(const QVariant& value)

C++ documentation:

This property holds the optional start value of the animation.

This property describes the optional start value of the animation. If omitted, or if a null QVariant is assigned as the start value, the animation will use the current position of the end when the animation is started.

Access functions:

QVariant 	startValue() const

void 	setStartValue(const QVariant & value)

See also endValue.

fn set_user_data(&mut self, id: c_uint, data: *mut ObjectUserData)

C++ method: void QVariantAnimation::setUserData(unsigned int id, QObjectUserData* data)

fn signals_blocked(&self) -> bool

C++ method: bool QVariantAnimation::signalsBlocked() const

C++ documentation:

Returns true if signals are blocked; otherwise returns false.

Signals are not blocked by default.

See also blockSignals().

fn start<'largs, Args>(&'largs mut self, args: Args) where Args: VariantAnimationStartArgs<'largs>

C++ method: QVariantAnimation::start

This is an overloaded function. Available variants:

Rust arguments:
1) fn start(&mut self, ()) -> ()
2) fn start(&mut self, ::abstract_animation::DeletionPolicy) -> ()
C++ method: void QVariantAnimation::start(QAbstractAnimation::DeletionPolicy policy = ?)

C++ documentation:

Starts the animation. The policy argument says whether or not the animation should be deleted when it's done. When the animation starts, the stateChanged() signal is emitted, and state() returns Running. When control reaches the event loop, the animation will run by itself, periodically calling updateCurrentTime() as the animation progresses.

If the animation is currently stopped or has already reached the end, calling start() will rewind the animation and start again from the beginning. When the animation reaches the end, the animation will either stop, or if the loop level is more than 1, it will rewind and continue from the beginning.

If the animation is already running, this function does nothing.

See also stop() and state().

fn start_timer<'largs, Args>(&'largs mut self, args: Args) -> c_int where Args: VariantAnimationStartTimerArgs<'largs>

C++ method: QVariantAnimation::startTimer

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_timer(&mut self, libc::c_int) -> libc::c_int
2) fn start_timer(&mut self, (libc::c_int, ::qt::TimerType)) -> libc::c_int
C++ method: int QVariantAnimation::startTimer(int interval, Qt::TimerType timerType = ?)

C++ documentation:

Starts a timer and returns a timer identifier, or returns zero if it could not start a timer.

A timer event will occur every interval milliseconds until killTimer() is called. If interval is 0, then the timer event occurs once every time there are no more window system events to process.

The virtual timerEvent() function is called with the QTimerEvent event parameter class when a timer event occurs. Reimplement this function to get timer events.

If multiple timers are running, the QTimerEvent::timerId() can be used to find out which timer was activated.

Example:

class MyObject : public QObject
{
    Q_OBJECT

public: MyObject(QObject *parent = 0);

protected: void timerEvent(QTimerEvent *event); };

MyObject::MyObject(QObject *parent) : QObject(parent) { startTimer(50); // 50-millisecond timer startTimer(1000); // 1-second timer startTimer(60000); // 1-minute timer }

void MyObject::timerEvent(QTimerEvent *event) { qDebug() << "Timer ID:" << event->timerId(); }
Note that QTimer's accuracy depends on the underlying operating system and hardware. The timerType argument allows you to customize the accuracy of the timer. See Qt::TimerType for information on the different timer types. Most platforms support an accuracy of 20 milliseconds; some provide more. If Qt is unable to deliver the requested number of timer events, it will silently discard some.

The QTimer class provides a high-level programming interface with single-shot timers and timer signals instead of events. There is also a QBasicTimer class that is more lightweight than QTimer and less clumsy than using timer IDs directly.

See also timerEvent(), killTimer(), and QTimer::singleShot().

fn start_value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VariantAnimationStartValueArgs<'largs>

C++ method: QVariantAnimation::startValue

This is an overloaded function. Available variants:

Rust arguments:
1) fn start_value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::variant::Variant>
2) fn start_value(&self, cpp_utils::AsStruct) -> ::variant::Variant
C++ method: QVariant QVariantAnimation::startValue() const

C++ documentation:

This property holds the optional start value of the animation.

This property describes the optional start value of the animation. If omitted, or if a null QVariant is assigned as the start value, the animation will use the current position of the end when the animation is started.

Access functions:

QVariant 	startValue() const

void 	setStartValue(const QVariant & value)

See also endValue.

fn state(&self) -> State

C++ method: QAbstractAnimation::State QVariantAnimation::state() const

C++ documentation:

This property holds state of the animation.

This property describes the current state of the animation. When the animation state changes, QAbstractAnimation emits the stateChanged() signal.

Access functions:

State 	state() const

Notifier signal:

void 	stateChanged(QAbstractAnimation::State newState, QAbstractAnimation::State oldState)

fn state_changed(&mut self, new_state: State, old_state: State)

C++ method: void QVariantAnimation::stateChanged(QAbstractAnimation::State newState, QAbstractAnimation::State oldState)

C++ documentation:

This property holds state of the animation.

This property describes the current state of the animation. When the animation state changes, QAbstractAnimation emits the stateChanged() signal.

Access functions:

State 	state() const

Notifier signal:

void 	stateChanged(QAbstractAnimation::State newState, QAbstractAnimation::State oldState)

fn stop(&mut self)

C++ method: void QVariantAnimation::stop()

C++ documentation:

Stops the animation. When the animation is stopped, it emits the stateChanged() signal, and state() returns Stopped. The current time is not changed.

If the animation stops by itself after reaching the end (i.e., currentLoopTime() == duration() and currentLoop() > loopCount() - 1), the finished() signal is emitted.

See also start() and state().

fn thread(&self) -> *mut Thread

C++ method: QThread* QVariantAnimation::thread() const

C++ documentation:

Returns the thread in which the object lives.

See also moveToThread().

fn total_duration(&self) -> c_int

C++ method: int QVariantAnimation::totalDuration() const

C++ documentation:

Returns the total and effective duration of the animation, including the loop count.

See also duration() and currentTime.

fn tr<Args>(args: Args) -> Args::ReturnType where Args: VariantAnimationTrArgs

C++ method: QVariantAnimation::tr

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QVariantAnimation::tr(const char* s, const char* c, int n)

fn tr_utf8<Args>(args: Args) -> Args::ReturnType where Args: VariantAnimationTrUtf8Args

C++ method: QVariantAnimation::trUtf8

This is an overloaded function. Available variants:

Rust arguments:
1) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn tr_utf8((*const libc::c_char, *const libc::c_char, libc::c_int, cpp_utils::AsStruct)) -> ::string::String
C++ method: static QString QVariantAnimation::trUtf8(const char* s, const char* c, int n)

fn user_data(&self, id: c_uint) -> *mut ObjectUserData

C++ method: QObjectUserData* QVariantAnimation::userData(unsigned int id) const

fn value_changed(&mut self, value: &Variant)

C++ method: void QVariantAnimation::valueChanged(const QVariant& value)

C++ documentation:

This property holds the current value of the animation.

This property describes the current value; an interpolated value between the start value and the end value, using the current time for progress. The value itself is obtained from interpolated(), which is called repeatedly as the animation is running.

QVariantAnimation calls the virtual updateCurrentValue() function when the current value changes. This is particularly useful for subclasses that need to track updates. For example, QPropertyAnimation uses this function to animate Qt properties.

Access functions:

QVariant 	currentValue() const

Notifier signal:

void 	valueChanged(const QVariant & value)

See also startValue and endValue.