Struct qt_core::vector::VectorCUint

pub struct VectorCUint {
    // some fields omitted
}

C++ type: QVector<unsigned int>

Methods

impl VectorCUint

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: VectorCUintAppendArgs<'largs>

C++ method: QVector<unsigned int>::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &::vector::VectorCUint) -> ()
C++ method: void QVector<unsigned int>::append(const QVector<unsigned int>& l)

C++ documentation:

This is an overloaded function.

Appends the items of the value vector to this vector.

This function was introduced in Qt 5.5.

See also operator<<() and operator+=().

Variant 2

Rust arguments: fn append(&mut self, &libc::c_uint) -> ()
C++ method: void QVector<unsigned int>::append(const unsigned int& t)

C++ documentation:

Inserts value at the end of the vector.

Example:

QVector<QString> vector(0);
vector.append("one");
vector.append("two");
vector.append("three");
// vector: ["one", "two", "three"]

This is the same as calling resize(size() + 1) and assigning value to the new last element in the vector.

This operation is relatively fast, because QVector typically allocates more memory than necessary, so it can grow without reallocating the entire vector each time.

See also operator<<(), prepend(), and insert().

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 c_uint

C++ method: const unsigned int& QVector<unsigned int>::at(int i) const

C++ documentation:

Returns the item at index position i in the vector.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 c_uint

C++ method: const unsigned int& QVector<unsigned int>::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut c_uint

C++ method: unsigned int& QVector<unsigned int>::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last().

fn capacity(&self) -> c_int

C++ method: int QVector<unsigned int>::capacity() const

C++ documentation:

Returns the maximum number of items that can be stored in the vector without forcing a reallocation.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the vector, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QVector<unsigned int>::clear()

C++ documentation:

Removes all the elements from the vector and releases the memory used by the vector.

fn const_data(&self) -> *const c_uint

C++ method: const unsigned int* QVector<unsigned int>::constData() const

C++ documentation:

Returns a const pointer to the data stored in the vector. The pointer can be used to access the items in the vector. The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also data() and operator[]().

fn contains(&self, t: &c_uint) -> bool

C++ method: bool QVector<unsigned int>::contains(const unsigned int& t) const

C++ documentation:

Returns true if the vector contains an occurrence of value; otherwise returns false.

This function requires the value type to have an implementation of operator==().

See also indexOf() and count().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: VectorCUintCountArgs<'largs>

C++ method: QVector<unsigned int>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QVector<unsigned int>::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &libc::c_uint) -> libc::c_int
C++ method: int QVector<unsigned int>::count(const unsigned int& t) const

C++ documentation:

Returns the number of occurrences of value in the vector.

This function requires the value type to have an implementation of operator==().

See also contains() and indexOf().

fn data(&self) -> *const c_uint

C++ method: const unsigned int* QVector<unsigned int>::data() const

C++ documentation:

This is an overloaded function.

fn data_mut(&mut self) -> *mut c_uint

C++ method: unsigned int* QVector<unsigned int>::data()

C++ documentation:

Returns a pointer to the data stored in the vector. The pointer can be used to access and modify the items in the vector.

Example:

QVector<int> vector(10);
int data = vector.data();
for (int i = 0; i < 10; ++i)
    data[i] = 2  i;

The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also constData() and operator[]().

fn detach(&mut self)

C++ method: void QVector<unsigned int>::detach()

fn empty(&self) -> bool

C++ method: bool QVector<unsigned int>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the vector is empty; otherwise returns false.

fn ends_with(&self, t: &c_uint) -> bool

C++ method: bool QVector<unsigned int>::endsWith(const unsigned int& t) const

C++ documentation:

Returns true if this vector is not empty and its last item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and last().

fn fill<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorCUint where Args: VectorCUintFillArgs<'largs>

C++ method: QVector<unsigned int>::fill

This is an overloaded function. Available variants:

Rust arguments:
1) fn fill(&mut self, &'l1 libc::c_uint) -> &'l0 mut ::vector::VectorCUint
2) fn fill(&mut self, (&'l1 libc::c_uint, libc::c_int)) -> &'l0 mut ::vector::VectorCUint
C++ method: QVector<unsigned int>& QVector<unsigned int>::fill(const unsigned int& t, int size = ?)

C++ documentation:

Assigns value to all items in the vector. If size is different from -1 (the default), the vector is resized to size size beforehand.

Example:

QVector<QString> vector(3);
vector.fill("Yes");
// vector: ["Yes", "Yes", "Yes"]

vector.fill("oh", 5); // vector: ["oh", "oh", "oh", "oh", "oh"]
See also resize().

fn first<'l0>(&'l0 self) -> &'l0 c_uint

C++ method: const unsigned int& QVector<unsigned int>::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut c_uint

C++ method: unsigned int& QVector<unsigned int>::first()

C++ documentation:

Returns a reference to the first item in the vector. This function assumes that the vector isn't empty.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 c_uint

C++ method: const unsigned int& QVector<unsigned int>::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut c_uint

C++ method: unsigned int& QVector<unsigned int>::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first().

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: VectorCUintIndexOfArgs<'largs>

C++ method: QVector<unsigned int>::indexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn index_of(&self, &libc::c_uint) -> libc::c_int
2) fn index_of(&self, (&libc::c_uint, libc::c_int)) -> libc::c_int
C++ method: int QVector<unsigned int>::indexOf(const unsigned int& t, int from = ?) const

C++ documentation:

Returns the index position of the first occurrence of value in the vector, searching forward from index position from. Returns -1 if no item matched.

Example:

QVector<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.indexOf("B");            // returns 1
vector.indexOf("B", 1);         // returns 1
vector.indexOf("B", 2);         // returns 3
vector.indexOf("X");            // returns -1

This function requires the value type to have an implementation of operator==().

See also lastIndexOf() and contains().

fn insert<'largs, Args>(&'largs mut self, args: Args) where Args: VectorCUintInsertArgs<'largs>

C++ method: QVector<unsigned int>::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert(&mut self, (libc::c_int, &libc::c_uint)) -> ()
C++ method: void QVector<unsigned int>::insert(int i, const unsigned int& t)

C++ documentation:

Inserts value at index position i in the vector. If i is 0, the value is prepended to the vector. If i is size(), the value is appended to the vector.

Example:

QVector<QString> vector;
vector << "alpha" << "beta" << "delta";
vector.insert(2, "gamma");
// vector: ["alpha", "beta", "gamma", "delta"]

For large vectors, this operation can be slow (linear time), because it requires moving all the items at indexes i and above by one position further in memory. If you want a container class that provides a fast insert() function, use QLinkedList instead.

See also append(), prepend(), and remove().

Variant 2

Rust arguments: fn insert(&mut self, (libc::c_int, libc::c_int, &libc::c_uint)) -> ()
C++ method: void QVector<unsigned int>::insert(int i, int n, const unsigned int& t)

C++ documentation:

This is an overloaded function.

Inserts count copies of value at index position i in the vector.

Example:

QVector<double> vector;
vector << 2.718 << 1.442 << 0.4342;
vector.insert(1, 3, 9.9);
// vector: [2.718, 9.9, 9.9, 9.9, 1.442, 0.4342]

fn is_detached(&self) -> bool

C++ method: bool QVector<unsigned int>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QVector<unsigned int>::isEmpty() const

C++ documentation:

Returns true if the vector has size 0; otherwise returns false.

See also size() and resize().

fn is_shared_with(&self, other: &VectorCUint) -> bool

C++ method: bool QVector<unsigned int>::isSharedWith(const QVector<unsigned int>& other) const

fn last<'l0>(&'l0 self) -> &'l0 c_uint

C++ method: const unsigned int& QVector<unsigned int>::last() const

C++ documentation:

This is an overloaded function.

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: VectorCUintLastIndexOfArgs<'largs>

C++ method: QVector<unsigned int>::lastIndexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_index_of(&self, &libc::c_uint) -> libc::c_int
2) fn last_index_of(&self, (&libc::c_uint, libc::c_int)) -> libc::c_int
C++ method: int QVector<unsigned int>::lastIndexOf(const unsigned int& t, int from = ?) const

C++ documentation:

Returns the index position of the last occurrence of the value value in the vector, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

Example:

QList<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.lastIndexOf("B");        // returns 3
vector.lastIndexOf("B", 3);     // returns 3
vector.lastIndexOf("B", 2);     // returns 1
vector.lastIndexOf("X");        // returns -1

This function requires the value type to have an implementation of operator==().

See also indexOf().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut c_uint

C++ method: unsigned int& QVector<unsigned int>::last()

C++ documentation:

Returns a reference to the last item in the vector. This function assumes that the vector isn't empty.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QVector<unsigned int>::length() const

C++ documentation:

Same as size() and count().

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also size(), count(), and QList::length().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorCUintMidArgs<'largs>

C++ method: QVector<unsigned int>::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorCUint>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorCUint>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorCUint
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorCUint
C++ method: QVector<unsigned int> QVector<unsigned int>::mid(int pos, int len = ?) const

C++ documentation:

Returns a sub-vector which contains elements from this vector, starting at position pos. If length is -1 (the default), all elements after pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn new<Args>(args: Args) -> Args::ReturnType where Args: VectorCUintNewArgs

C++ method: QVector<unsigned int>::QVector

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::vector::VectorCUint
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorCUint>
C++ method: [constructor] void QVector<unsigned int>::QVector()

Variant 2

Rust arguments:
1) fn new((&::vector::VectorCUint, cpp_utils::AsStruct)) -> ::vector::VectorCUint
2) fn new((&::vector::VectorCUint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorCUint>
C++ method: [constructor] void QVector<unsigned int>::QVector(const QVector<unsigned int>& v)

Variant 3

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorCUint
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorCUint>
C++ method: [constructor] void QVector<unsigned int>::QVector(int size)

Variant 4

Rust arguments:
1) fn new((libc::c_int, &libc::c_uint, cpp_utils::AsStruct)) -> ::vector::VectorCUint
2) fn new((libc::c_int, &libc::c_uint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorCUint>
C++ method: [constructor] void QVector<unsigned int>::QVector(int size, const unsigned int& t)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorCUintOpAddArgs<'largs>

C++ method: QVector<unsigned int>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::vector::VectorCUint, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorCUint>
2) fn op_add(&self, (&::vector::VectorCUint, cpp_utils::AsStruct)) -> ::vector::VectorCUint
C++ method: QVector<unsigned int> QVector<unsigned int>::operator+(const QVector<unsigned int>& l) const

C++ documentation:

Returns a vector that contains all the items in this vector followed by all the items in the other vector.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorCUint where Args: VectorCUintOpAddAssignArgs<'largs>

C++ method: QVector<unsigned int>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::vector::VectorCUint) -> &'l0 mut ::vector::VectorCUint
C++ method: QVector<unsigned int>& QVector<unsigned int>::operator+=(const QVector<unsigned int>& l)

C++ documentation:

Appends the items of the other vector to this vector and returns a reference to this vector.

See also operator+() and append().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 libc::c_uint) -> &'l0 mut ::vector::VectorCUint
C++ method: QVector<unsigned int>& QVector<unsigned int>::operator+=(const unsigned int& t)

C++ documentation:

This is an overloaded function.

Appends value to the vector.

See also append() and operator<<().

fn op_assign<'l0, 'l1>(&'l0 mut self, v: &'l1 VectorCUint) -> &'l0 mut VectorCUint

C++ method: QVector<unsigned int>& QVector<unsigned int>::operator=(const QVector<unsigned int>& v)

C++ documentation:

Assigns other to this vector and returns a reference to this vector.

fn op_eq(&self, v: &VectorCUint) -> bool

C++ method: bool QVector<unsigned int>::operator==(const QVector<unsigned int>& v) const

C++ documentation:

Returns true if other is equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator!=().

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 c_uint

C++ method: const unsigned int& QVector<unsigned int>::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(i).

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut c_uint

C++ method: unsigned int& QVector<unsigned int>::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

Note that using non-const operators can cause QVector to do a deep copy.

See also at() and value().

fn op_neq(&self, v: &VectorCUint) -> bool

C++ method: bool QVector<unsigned int>::operator!=(const QVector<unsigned int>& v) const

C++ documentation:

Returns true if other is not equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator==().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorCUint where Args: VectorCUintOpShlArgs<'largs>

C++ method: QVector<unsigned int>::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::vector::VectorCUint) -> &'l0 mut ::vector::VectorCUint
C++ method: QVector<unsigned int>& QVector<unsigned int>::operator<<(const QVector<unsigned int>& l)

C++ documentation:

Appends other to the vector and returns a reference to the vector.

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 libc::c_uint) -> &'l0 mut ::vector::VectorCUint
C++ method: QVector<unsigned int>& QVector<unsigned int>::operator<<(const unsigned int& t)

C++ documentation:

Appends value to the vector and returns a reference to this vector.

See also append() and operator+=().

fn pop_back(&mut self)

C++ method: void QVector<unsigned int>::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast().

fn pop_front(&mut self)

C++ method: void QVector<unsigned int>::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst().

fn prepend(&mut self, t: &c_uint)

C++ method: void QVector<unsigned int>::prepend(const unsigned int& t)

C++ documentation:

Inserts value at the beginning of the vector.

Example:

QVector<QString> vector;
vector.prepend("one");
vector.prepend("two");
vector.prepend("three");
// vector: ["three", "two", "one"]

This is the same as vector.insert(0, value).

For large vectors, this operation can be slow (linear time), because it requires moving all the items in the vector by one position further in memory. If you want a container class that provides a fast prepend() function, use QList or QLinkedList instead.

See also append() and insert().

fn push_back(&mut self, t: &c_uint)

C++ method: void QVector<unsigned int>::push_back(const unsigned int& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &c_uint)

C++ method: void QVector<unsigned int>::push_front(const unsigned int& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove<'largs, Args>(&'largs mut self, args: Args) where Args: VectorCUintRemoveArgs<'largs>

C++ method: QVector<unsigned int>::remove

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn remove(&mut self, libc::c_int) -> ()
C++ method: void QVector<unsigned int>::remove(int i)

C++ documentation:

This is an overloaded function.

Removes the element at index position i.

See also insert(), replace(), and fill().

Variant 2

Rust arguments: fn remove(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QVector<unsigned int>::remove(int i, int n)

C++ documentation:

This is an overloaded function.

Removes count elements from the middle of the vector, starting at index position i.

See also insert(), replace(), and fill().

fn remove_all(&mut self, t: &c_uint) -> c_int

C++ method: int QVector<unsigned int>::removeAll(const unsigned int& t)

C++ documentation:

Removes all elements that compare equal to t from the vector. Returns the number of elements removed, if any.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeOne() and QList::removeAll().

fn remove_at(&mut self, i: c_int)

C++ method: void QVector<unsigned int>::removeAt(int i)

C++ documentation:

Removes the element at index position i. Equivalent to

remove(i);

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also remove() and QList::removeAt().

fn remove_first(&mut self)

C++ method: void QVector<unsigned int>::removeFirst()

C++ documentation:

Removes the first item in the vector. Calling this function is equivalent to calling remove(0). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeFirst(), and isEmpty().

fn remove_last(&mut self)

C++ method: void QVector<unsigned int>::removeLast()

C++ documentation:

Removes the last item in the vector. Calling this function is equivalent to calling remove(size() - 1). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeLast(), removeFirst(), and isEmpty().

fn remove_one(&mut self, t: &c_uint) -> bool

C++ method: bool QVector<unsigned int>::removeOne(const unsigned int& t)

C++ documentation:

Removes the first element that compares equal to t from the vector. Returns whether an element was, in fact, removed.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeAll() and QList::removeOne().

fn replace(&mut self, i: c_int, t: &c_uint)

C++ method: void QVector<unsigned int>::replace(int i, const unsigned int& t)

C++ documentation:

Replaces the item at index position i with value.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also operator[]() and remove().

fn reserve(&mut self, size: c_int)

C++ method: void QVector<unsigned int>::reserve(int size)

C++ documentation:

Attempts to allocate memory for at least size elements. If you know in advance how large the vector will be, you can call this function, and if you call resize() often you are likely to get better performance. If size is an underestimate, the worst that will happen is that the QVector will be a bit slower.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to change the size of the vector, call resize().

See also squeeze() and capacity().

fn resize(&mut self, size: c_int)

C++ method: void QVector<unsigned int>::resize(int size)

C++ documentation:

Sets the size of the vector to size. If size is greater than the current size, elements are added to the end; the new elements are initialized with a default-constructed value. If size is less than the current size, elements are removed from the end.

See also size().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QVector<unsigned int>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QVector<unsigned int>::size() const

C++ documentation:

Returns the number of items in the vector.

See also isEmpty() and resize().

fn squeeze(&mut self)

C++ method: void QVector<unsigned int>::squeeze()

C++ documentation:

Releases any memory not required to store the items.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn starts_with(&self, t: &c_uint) -> bool

C++ method: bool QVector<unsigned int>::startsWith(const unsigned int& t) const

C++ documentation:

Returns true if this vector is not empty and its first item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and first().

fn swap(&mut self, other: &mut VectorCUint)

C++ method: void QVector<unsigned int>::swap(QVector<unsigned int>& other)

C++ documentation:

Swaps vector other with this vector. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take_at(&mut self, i: c_int) -> c_uint

C++ method: unsigned int QVector<unsigned int>::takeAt(int i)

C++ documentation:

Removes the element at index position i and returns it.

Equivalent to

T t = at(i);
remove(i);
return t;

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also takeFirst(), takeLast(), and QList::takeAt().

fn take_first(&mut self) -> c_uint

C++ method: unsigned int QVector<unsigned int>::takeFirst()

C++ documentation:

Removes the first item in the vector and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also takeLast() and removeFirst().

fn take_last(&mut self) -> c_uint

C++ method: unsigned int QVector<unsigned int>::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

If you don't use the return value, removeLast() is more efficient.

This function was introduced in Qt 5.1.

See also takeFirst() and removeLast().

fn value<'largs, Args>(&'largs self, args: Args) -> c_uint where Args: VectorCUintValueArgs<'largs>

C++ method: QVector<unsigned int>::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn value(&self, libc::c_int) -> libc::c_uint
C++ method: unsigned int QVector<unsigned int>::value(int i) const

C++ documentation:

Returns the value at index position i in the vector.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that i is within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments: fn value(&self, (libc::c_int, &libc::c_uint)) -> libc::c_uint
C++ method: unsigned int QVector<unsigned int>::value(int i, const unsigned int& defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.