Struct qt_core::vector::VectorXmlStreamEntityDeclarationRef

pub struct VectorXmlStreamEntityDeclarationRef {
    // some fields omitted
}

C++ type: QVector<QXmlStreamEntityDeclaration>

Methods

impl VectorXmlStreamEntityDeclarationRef

fn append<'largs, Args>(&'largs mut self, args: Args) where Args: VectorXmlStreamEntityDeclarationRefAppendArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::append

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn append(&mut self, &::vector::VectorXmlStreamEntityDeclarationRef) -> ()
C++ method: void QVector<QXmlStreamEntityDeclaration>::append(const QVector<QXmlStreamEntityDeclaration>& l)

C++ documentation:

This is an overloaded function.

Appends the items of the value vector to this vector.

This function was introduced in Qt 5.5.

See also operator<<() and operator+=().

Variant 2

Rust arguments: fn append(&mut self, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration) -> ()
C++ method: void QVector<QXmlStreamEntityDeclaration>::append(const QXmlStreamEntityDeclaration& t)

C++ documentation:

Inserts value at the end of the vector.

Example:

QVector<QString> vector(0);
vector.append("one");
vector.append("two");
vector.append("three");
// vector: ["one", "two", "three"]

This is the same as calling resize(size() + 1) and assigning value to the new last element in the vector.

This operation is relatively fast, because QVector typically allocates more memory than necessary, so it can grow without reallocating the entire vector each time.

See also operator<<(), prepend(), and insert().

fn at<'l0>(&'l0 self, i: c_int) -> &'l0 XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::at(int i) const

C++ documentation:

Returns the item at index position i in the vector.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also value() and operator[]().

fn back<'l0>(&'l0 self) -> &'l0 XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::back() const

C++ documentation:

This is an overloaded function.

fn back_mut<'l0>(&'l0 mut self) -> &'l0 mut XmlStreamEntityDeclaration

C++ method: QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to last().

fn capacity(&self) -> c_int

C++ method: int QVector<QXmlStreamEntityDeclaration>::capacity() const

C++ documentation:

Returns the maximum number of items that can be stored in the vector without forcing a reallocation.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to know how many items are in the vector, call size().

See also reserve() and squeeze().

fn clear(&mut self)

C++ method: void QVector<QXmlStreamEntityDeclaration>::clear()

C++ documentation:

Removes all the elements from the vector and releases the memory used by the vector.

fn const_data(&self) -> *const XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration* QVector<QXmlStreamEntityDeclaration>::constData() const

C++ documentation:

Returns a const pointer to the data stored in the vector. The pointer can be used to access the items in the vector. The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also data() and operator[]().

fn contains(&self, t: &XmlStreamEntityDeclaration) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::contains(const QXmlStreamEntityDeclaration& t) const

C++ documentation:

Returns true if the vector contains an occurrence of value; otherwise returns false.

This function requires the value type to have an implementation of operator==().

See also indexOf() and count().

fn count<'largs, Args>(&'largs self, args: Args) -> c_int where Args: VectorXmlStreamEntityDeclarationRefCountArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::count

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn count(&self, ()) -> libc::c_int
C++ method: int QVector<QXmlStreamEntityDeclaration>::count() const

C++ documentation:

This is an overloaded function.

Same as size().

Variant 2

Rust arguments: fn count(&self, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration) -> libc::c_int
C++ method: int QVector<QXmlStreamEntityDeclaration>::count(const QXmlStreamEntityDeclaration& t) const

C++ documentation:

Returns the number of occurrences of value in the vector.

This function requires the value type to have an implementation of operator==().

See also contains() and indexOf().

fn data(&self) -> *const XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration* QVector<QXmlStreamEntityDeclaration>::data() const

C++ documentation:

This is an overloaded function.

fn data_mut(&mut self) -> *mut XmlStreamEntityDeclaration

C++ method: QXmlStreamEntityDeclaration* QVector<QXmlStreamEntityDeclaration>::data()

C++ documentation:

Returns a pointer to the data stored in the vector. The pointer can be used to access and modify the items in the vector.

Example:

QVector<int> vector(10);
int data = vector.data();
for (int i = 0; i < 10; ++i)
    data[i] = 2  i;

The pointer remains valid as long as the vector isn't reallocated.

This function is mostly useful to pass a vector to a function that accepts a plain C++ array.

See also constData() and operator[]().

fn detach(&mut self)

C++ method: void QVector<QXmlStreamEntityDeclaration>::detach()

fn empty(&self) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::empty() const

C++ documentation:

This function is provided for STL compatibility. It is equivalent to isEmpty(), returning true if the vector is empty; otherwise returns false.

fn ends_with(&self, t: &XmlStreamEntityDeclaration) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::endsWith(const QXmlStreamEntityDeclaration& t) const

C++ documentation:

Returns true if this vector is not empty and its last item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and last().

fn fill<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorXmlStreamEntityDeclarationRef where Args: VectorXmlStreamEntityDeclarationRefFillArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::fill

This is an overloaded function. Available variants:

Rust arguments:
1) fn fill(&mut self, &'l1 ::xml_stream_entity_declaration::XmlStreamEntityDeclaration) -> &'l0 mut ::vector::VectorXmlStreamEntityDeclarationRef
2) fn fill(&mut self, (&'l1 ::xml_stream_entity_declaration::XmlStreamEntityDeclaration, libc::c_int)) -> &'l0 mut ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration>& QVector<QXmlStreamEntityDeclaration>::fill(const QXmlStreamEntityDeclaration& t, int size = ?)

C++ documentation:

Assigns value to all items in the vector. If size is different from -1 (the default), the vector is resized to size size beforehand.

Example:

QVector<QString> vector(3);
vector.fill("Yes");
// vector: ["Yes", "Yes", "Yes"]

vector.fill("oh", 5); // vector: ["oh", "oh", "oh", "oh", "oh"]
See also resize().

fn first<'l0>(&'l0 self) -> &'l0 XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::first() const

C++ documentation:

This is an overloaded function.

fn first_mut<'l0>(&'l0 mut self) -> &'l0 mut XmlStreamEntityDeclaration

C++ method: QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::first()

C++ documentation:

Returns a reference to the first item in the vector. This function assumes that the vector isn't empty.

See also last() and isEmpty().

fn front<'l0>(&'l0 self) -> &'l0 XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::front() const

C++ documentation:

This is an overloaded function.

fn front_mut<'l0>(&'l0 mut self) -> &'l0 mut XmlStreamEntityDeclaration

C++ method: QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to first().

fn index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: VectorXmlStreamEntityDeclarationRefIndexOfArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::indexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn index_of(&self, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration) -> libc::c_int
2) fn index_of(&self, (&::xml_stream_entity_declaration::XmlStreamEntityDeclaration, libc::c_int)) -> libc::c_int
C++ method: int QVector<QXmlStreamEntityDeclaration>::indexOf(const QXmlStreamEntityDeclaration& t, int from = ?) const

C++ documentation:

Returns the index position of the first occurrence of value in the vector, searching forward from index position from. Returns -1 if no item matched.

Example:

QVector<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.indexOf("B");            // returns 1
vector.indexOf("B", 1);         // returns 1
vector.indexOf("B", 2);         // returns 3
vector.indexOf("X");            // returns -1

This function requires the value type to have an implementation of operator==().

See also lastIndexOf() and contains().

fn insert<'largs, Args>(&'largs mut self, args: Args) where Args: VectorXmlStreamEntityDeclarationRefInsertArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::insert

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn insert(&mut self, (libc::c_int, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration)) -> ()
C++ method: void QVector<QXmlStreamEntityDeclaration>::insert(int i, const QXmlStreamEntityDeclaration& t)

C++ documentation:

Inserts value at index position i in the vector. If i is 0, the value is prepended to the vector. If i is size(), the value is appended to the vector.

Example:

QVector<QString> vector;
vector << "alpha" << "beta" << "delta";
vector.insert(2, "gamma");
// vector: ["alpha", "beta", "gamma", "delta"]

For large vectors, this operation can be slow (linear time), because it requires moving all the items at indexes i and above by one position further in memory. If you want a container class that provides a fast insert() function, use QLinkedList instead.

See also append(), prepend(), and remove().

Variant 2

Rust arguments: fn insert(&mut self, (libc::c_int, libc::c_int, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration)) -> ()
C++ method: void QVector<QXmlStreamEntityDeclaration>::insert(int i, int n, const QXmlStreamEntityDeclaration& t)

C++ documentation:

This is an overloaded function.

Inserts count copies of value at index position i in the vector.

Example:

QVector<double> vector;
vector << 2.718 << 1.442 << 0.4342;
vector.insert(1, 3, 9.9);
// vector: [2.718, 9.9, 9.9, 9.9, 1.442, 0.4342]

fn is_detached(&self) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::isDetached() const

fn is_empty(&self) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::isEmpty() const

C++ documentation:

Returns true if the vector has size 0; otherwise returns false.

See also size() and resize().

fn is_shared_with(&self, other: &VectorXmlStreamEntityDeclarationRef) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::isSharedWith(const QVector<QXmlStreamEntityDeclaration>& other) const

fn last<'l0>(&'l0 self) -> &'l0 XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::last() const

C++ documentation:

This is an overloaded function.

fn last_index_of<'largs, Args>(&'largs self, args: Args) -> c_int where Args: VectorXmlStreamEntityDeclarationRefLastIndexOfArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::lastIndexOf

This is an overloaded function. Available variants:

Rust arguments:
1) fn last_index_of(&self, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration) -> libc::c_int
2) fn last_index_of(&self, (&::xml_stream_entity_declaration::XmlStreamEntityDeclaration, libc::c_int)) -> libc::c_int
C++ method: int QVector<QXmlStreamEntityDeclaration>::lastIndexOf(const QXmlStreamEntityDeclaration& t, int from = ?) const

C++ documentation:

Returns the index position of the last occurrence of the value value in the vector, searching backward from index position from. If from is -1 (the default), the search starts at the last item. Returns -1 if no item matched.

Example:

QList<QString> vector;
vector << "A" << "B" << "C" << "B" << "A";
vector.lastIndexOf("B");        // returns 3
vector.lastIndexOf("B", 3);     // returns 3
vector.lastIndexOf("B", 2);     // returns 1
vector.lastIndexOf("X");        // returns -1

This function requires the value type to have an implementation of operator==().

See also indexOf().

fn last_mut<'l0>(&'l0 mut self) -> &'l0 mut XmlStreamEntityDeclaration

C++ method: QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::last()

C++ documentation:

Returns a reference to the last item in the vector. This function assumes that the vector isn't empty.

See also first() and isEmpty().

fn length(&self) -> c_int

C++ method: int QVector<QXmlStreamEntityDeclaration>::length() const

C++ documentation:

Same as size() and count().

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also size(), count(), and QList::length().

fn mid<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorXmlStreamEntityDeclarationRefMidArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::mid

This is an overloaded function. Available variants:

Rust arguments:
1) fn mid(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
2) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
3) fn mid(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorXmlStreamEntityDeclarationRef
4) fn mid(&self, (libc::c_int, libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration> QVector<QXmlStreamEntityDeclaration>::mid(int pos, int len = ?) const

C++ documentation:

Returns a sub-vector which contains elements from this vector, starting at position pos. If length is -1 (the default), all elements after pos are included; otherwise length elements (or all remaining elements if there are less than length elements) are included.

fn new<Args>(args: Args) -> Args::ReturnType where Args: VectorXmlStreamEntityDeclarationRefNewArgs

C++ method: QVector<QXmlStreamEntityDeclaration>::QVector

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::vector::VectorXmlStreamEntityDeclarationRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
C++ method: [constructor] void QVector<QXmlStreamEntityDeclaration>::QVector()

Variant 2

Rust arguments:
1) fn new((&::vector::VectorXmlStreamEntityDeclarationRef, cpp_utils::AsStruct)) -> ::vector::VectorXmlStreamEntityDeclarationRef
2) fn new((&::vector::VectorXmlStreamEntityDeclarationRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
C++ method: [constructor] void QVector<QXmlStreamEntityDeclaration>::QVector(const QVector<QXmlStreamEntityDeclaration>& v)

Variant 3

Rust arguments:
1) fn new((libc::c_int, cpp_utils::AsStruct)) -> ::vector::VectorXmlStreamEntityDeclarationRef
2) fn new((libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
C++ method: [constructor] void QVector<QXmlStreamEntityDeclaration>::QVector(int size)

Variant 4

Rust arguments:
1) fn new((libc::c_int, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration, cpp_utils::AsStruct)) -> ::vector::VectorXmlStreamEntityDeclarationRef
2) fn new((libc::c_int, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
C++ method: [constructor] void QVector<QXmlStreamEntityDeclaration>::QVector(int size, const QXmlStreamEntityDeclaration& t)

fn op_add<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorXmlStreamEntityDeclarationRefOpAddArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::operator+

This is an overloaded function. Available variants:

Rust arguments:
1) fn op_add(&self, (&::vector::VectorXmlStreamEntityDeclarationRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
2) fn op_add(&self, (&::vector::VectorXmlStreamEntityDeclarationRef, cpp_utils::AsStruct)) -> ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration> QVector<QXmlStreamEntityDeclaration>::operator+(const QVector<QXmlStreamEntityDeclaration>& l) const

C++ documentation:

Returns a vector that contains all the items in this vector followed by all the items in the other vector.

See also operator+=().

fn op_add_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorXmlStreamEntityDeclarationRef where Args: VectorXmlStreamEntityDeclarationRefOpAddAssignArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::operator+=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_add_assign(&mut self, &'l1 ::vector::VectorXmlStreamEntityDeclarationRef) -> &'l0 mut ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration>& QVector<QXmlStreamEntityDeclaration>::operator+=(const QVector<QXmlStreamEntityDeclaration>& l)

C++ documentation:

Appends the items of the other vector to this vector and returns a reference to this vector.

See also operator+() and append().

Variant 2

Rust arguments: fn op_add_assign(&mut self, &'l1 ::xml_stream_entity_declaration::XmlStreamEntityDeclaration) -> &'l0 mut ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration>& QVector<QXmlStreamEntityDeclaration>::operator+=(const QXmlStreamEntityDeclaration& t)

C++ documentation:

This is an overloaded function.

Appends value to the vector.

See also append() and operator<<().

fn op_assign<'l0, 'l1>(&'l0 mut self, v: &'l1 VectorXmlStreamEntityDeclarationRef) -> &'l0 mut VectorXmlStreamEntityDeclarationRef

C++ method: QVector<QXmlStreamEntityDeclaration>& QVector<QXmlStreamEntityDeclaration>::operator=(const QVector<QXmlStreamEntityDeclaration>& v)

C++ documentation:

Assigns other to this vector and returns a reference to this vector.

fn op_eq(&self, v: &VectorXmlStreamEntityDeclarationRef) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::operator==(const QVector<QXmlStreamEntityDeclaration>& v) const

C++ documentation:

Returns true if other is equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator!=().

fn op_index<'l0>(&'l0 self, i: c_int) -> &'l0 XmlStreamEntityDeclaration

C++ method: const QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::operator[](int i) const

C++ documentation:

This is an overloaded function.

Same as at(i).

fn op_index_mut<'l0>(&'l0 mut self, i: c_int) -> &'l0 mut XmlStreamEntityDeclaration

C++ method: QXmlStreamEntityDeclaration& QVector<QXmlStreamEntityDeclaration>::operator[](int i)

C++ documentation:

Returns the item at index position i as a modifiable reference.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

Note that using non-const operators can cause QVector to do a deep copy.

See also at() and value().

fn op_neq(&self, v: &VectorXmlStreamEntityDeclarationRef) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::operator!=(const QVector<QXmlStreamEntityDeclaration>& v) const

C++ documentation:

Returns true if other is not equal to this vector; otherwise returns false.

Two vectors are considered equal if they contain the same values in the same order.

This function requires the value type to have an implementation of operator==().

See also operator==().

fn op_shl<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut VectorXmlStreamEntityDeclarationRef where Args: VectorXmlStreamEntityDeclarationRefOpShlArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::operator<<

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_shl(&mut self, &'l1 ::vector::VectorXmlStreamEntityDeclarationRef) -> &'l0 mut ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration>& QVector<QXmlStreamEntityDeclaration>::operator<<(const QVector<QXmlStreamEntityDeclaration>& l)

C++ documentation:

Appends other to the vector and returns a reference to the vector.

Variant 2

Rust arguments: fn op_shl(&mut self, &'l1 ::xml_stream_entity_declaration::XmlStreamEntityDeclaration) -> &'l0 mut ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration>& QVector<QXmlStreamEntityDeclaration>::operator<<(const QXmlStreamEntityDeclaration& t)

C++ documentation:

Appends value to the vector and returns a reference to this vector.

See also append() and operator+=().

fn pop_back(&mut self)

C++ method: void QVector<QXmlStreamEntityDeclaration>::pop_back()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeLast().

fn pop_front(&mut self)

C++ method: void QVector<QXmlStreamEntityDeclaration>::pop_front()

C++ documentation:

This function is provided for STL compatibility. It is equivalent to removeFirst().

fn prepend(&mut self, t: &XmlStreamEntityDeclaration)

C++ method: void QVector<QXmlStreamEntityDeclaration>::prepend(const QXmlStreamEntityDeclaration& t)

C++ documentation:

Inserts value at the beginning of the vector.

Example:

QVector<QString> vector;
vector.prepend("one");
vector.prepend("two");
vector.prepend("three");
// vector: ["three", "two", "one"]

This is the same as vector.insert(0, value).

For large vectors, this operation can be slow (linear time), because it requires moving all the items in the vector by one position further in memory. If you want a container class that provides a fast prepend() function, use QList or QLinkedList instead.

See also append() and insert().

fn push_back(&mut self, t: &XmlStreamEntityDeclaration)

C++ method: void QVector<QXmlStreamEntityDeclaration>::push_back(const QXmlStreamEntityDeclaration& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to append(value).

fn push_front(&mut self, t: &XmlStreamEntityDeclaration)

C++ method: void QVector<QXmlStreamEntityDeclaration>::push_front(const QXmlStreamEntityDeclaration& t)

C++ documentation:

This function is provided for STL compatibility. It is equivalent to prepend(value).

fn remove<'largs, Args>(&'largs mut self, args: Args) where Args: VectorXmlStreamEntityDeclarationRefRemoveArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::remove

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn remove(&mut self, libc::c_int) -> ()
C++ method: void QVector<QXmlStreamEntityDeclaration>::remove(int i)

C++ documentation:

This is an overloaded function.

Removes the element at index position i.

See also insert(), replace(), and fill().

Variant 2

Rust arguments: fn remove(&mut self, (libc::c_int, libc::c_int)) -> ()
C++ method: void QVector<QXmlStreamEntityDeclaration>::remove(int i, int n)

C++ documentation:

This is an overloaded function.

Removes count elements from the middle of the vector, starting at index position i.

See also insert(), replace(), and fill().

fn remove_all(&mut self, t: &XmlStreamEntityDeclaration) -> c_int

C++ method: int QVector<QXmlStreamEntityDeclaration>::removeAll(const QXmlStreamEntityDeclaration& t)

C++ documentation:

Removes all elements that compare equal to t from the vector. Returns the number of elements removed, if any.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeOne() and QList::removeAll().

fn remove_at(&mut self, i: c_int)

C++ method: void QVector<QXmlStreamEntityDeclaration>::removeAt(int i)

C++ documentation:

Removes the element at index position i. Equivalent to

remove(i);

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also remove() and QList::removeAt().

fn remove_first(&mut self)

C++ method: void QVector<QXmlStreamEntityDeclaration>::removeFirst()

C++ documentation:

Removes the first item in the vector. Calling this function is equivalent to calling remove(0). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeFirst(), and isEmpty().

fn remove_last(&mut self)

C++ method: void QVector<QXmlStreamEntityDeclaration>::removeLast()

C++ documentation:

Removes the last item in the vector. Calling this function is equivalent to calling remove(size() - 1). The vector must not be empty. If the vector can be empty, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also remove(), takeLast(), removeFirst(), and isEmpty().

fn remove_one(&mut self, t: &XmlStreamEntityDeclaration) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::removeOne(const QXmlStreamEntityDeclaration& t)

C++ documentation:

Removes the first element that compares equal to t from the vector. Returns whether an element was, in fact, removed.

Provided for compatibility with QList.

This function was introduced in Qt 5.4.

See also removeAll() and QList::removeOne().

fn replace(&mut self, i: c_int, t: &XmlStreamEntityDeclaration)

C++ method: void QVector<QXmlStreamEntityDeclaration>::replace(int i, const QXmlStreamEntityDeclaration& t)

C++ documentation:

Replaces the item at index position i with value.

i must be a valid index position in the vector (i.e., 0 <= i < size()).

See also operator[]() and remove().

fn reserve(&mut self, size: c_int)

C++ method: void QVector<QXmlStreamEntityDeclaration>::reserve(int size)

C++ documentation:

Attempts to allocate memory for at least size elements. If you know in advance how large the vector will be, you can call this function, and if you call resize() often you are likely to get better performance. If size is an underestimate, the worst that will happen is that the QVector will be a bit slower.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function. If you want to change the size of the vector, call resize().

See also squeeze() and capacity().

fn resize(&mut self, size: c_int)

C++ method: void QVector<QXmlStreamEntityDeclaration>::resize(int size)

C++ documentation:

Sets the size of the vector to size. If size is greater than the current size, elements are added to the end; the new elements are initialized with a default-constructed value. If size is less than the current size, elements are removed from the end.

See also size().

fn set_sharable(&mut self, sharable: bool)

C++ method: void QVector<QXmlStreamEntityDeclaration>::setSharable(bool sharable)

fn size(&self) -> c_int

C++ method: int QVector<QXmlStreamEntityDeclaration>::size() const

C++ documentation:

Returns the number of items in the vector.

See also isEmpty() and resize().

fn squeeze(&mut self)

C++ method: void QVector<QXmlStreamEntityDeclaration>::squeeze()

C++ documentation:

Releases any memory not required to store the items.

The sole purpose of this function is to provide a means of fine tuning QVector's memory usage. In general, you will rarely ever need to call this function.

See also reserve() and capacity().

fn starts_with(&self, t: &XmlStreamEntityDeclaration) -> bool

C++ method: bool QVector<QXmlStreamEntityDeclaration>::startsWith(const QXmlStreamEntityDeclaration& t) const

C++ documentation:

Returns true if this vector is not empty and its first item is equal to value; otherwise returns false.

This function was introduced in Qt 4.5.

See also isEmpty() and first().

fn swap(&mut self, other: &mut VectorXmlStreamEntityDeclarationRef)

C++ method: void QVector<QXmlStreamEntityDeclaration>::swap(QVector<QXmlStreamEntityDeclaration>& other)

C++ documentation:

Swaps vector other with this vector. This operation is very fast and never fails.

This function was introduced in Qt 4.8.

fn take_at<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: VectorXmlStreamEntityDeclarationRefTakeAtArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::takeAt

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_at(&mut self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_entity_declaration::XmlStreamEntityDeclaration>
2) fn take_at(&mut self, (libc::c_int, cpp_utils::AsStruct)) -> ::xml_stream_entity_declaration::XmlStreamEntityDeclaration
C++ method: QXmlStreamEntityDeclaration QVector<QXmlStreamEntityDeclaration>::takeAt(int i)

C++ documentation:

Removes the element at index position i and returns it.

Equivalent to

T t = at(i);
remove(i);
return t;

Provided for compatibility with QList.

This function was introduced in Qt 5.2.

See also takeFirst(), takeLast(), and QList::takeAt().

fn take_first<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: VectorXmlStreamEntityDeclarationRefTakeFirstArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::takeFirst

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_first(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_entity_declaration::XmlStreamEntityDeclaration>
2) fn take_first(&mut self, cpp_utils::AsStruct) -> ::xml_stream_entity_declaration::XmlStreamEntityDeclaration
C++ method: QXmlStreamEntityDeclaration QVector<QXmlStreamEntityDeclaration>::takeFirst()

C++ documentation:

Removes the first item in the vector and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

This function was introduced in Qt 5.1.

See also takeLast() and removeFirst().

fn take_last<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: VectorXmlStreamEntityDeclarationRefTakeLastArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::takeLast

This is an overloaded function. Available variants:

Rust arguments:
1) fn take_last(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_entity_declaration::XmlStreamEntityDeclaration>
2) fn take_last(&mut self, cpp_utils::AsStruct) -> ::xml_stream_entity_declaration::XmlStreamEntityDeclaration
C++ method: QXmlStreamEntityDeclaration QVector<QXmlStreamEntityDeclaration>::takeLast()

C++ documentation:

Removes the last item in the list and returns it. This function assumes the vector is not empty. To avoid failure, call isEmpty() before calling this function.

If you don't use the return value, removeLast() is more efficient.

This function was introduced in Qt 5.1.

See also takeFirst() and removeLast().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: VectorXmlStreamEntityDeclarationRefValueArgs<'largs>

C++ method: QVector<QXmlStreamEntityDeclaration>::value

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn value(&self, (libc::c_int, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_entity_declaration::XmlStreamEntityDeclaration>
2) fn value(&self, (libc::c_int, cpp_utils::AsStruct)) -> ::xml_stream_entity_declaration::XmlStreamEntityDeclaration
C++ method: QXmlStreamEntityDeclaration QVector<QXmlStreamEntityDeclaration>::value(int i) const

C++ documentation:

Returns the value at index position i in the vector.

If the index i is out of bounds, the function returns a default-constructed value. If you are certain that i is within bounds, you can use at() instead, which is slightly faster.

See also at() and operator[]().

Variant 2

Rust arguments:
1) fn value(&self, (libc::c_int, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_entity_declaration::XmlStreamEntityDeclaration>
2) fn value(&self, (libc::c_int, &::xml_stream_entity_declaration::XmlStreamEntityDeclaration, cpp_utils::AsStruct)) -> ::xml_stream_entity_declaration::XmlStreamEntityDeclaration
C++ method: QXmlStreamEntityDeclaration QVector<QXmlStreamEntityDeclaration>::value(int i, const QXmlStreamEntityDeclaration& defaultValue) const

C++ documentation:

This is an overloaded function.

If the index i is out of bounds, the function returns defaultValue.