Struct qt_core::weak_pointer::WeakPointerObjectRef

pub struct WeakPointerObjectRef {
    // some fields omitted
}

C++ type: QWeakPointer<QObject>

Methods

impl WeakPointerObjectRef

fn clear(&mut self)

C++ method: void QWeakPointer<QObject>::clear()

C++ documentation:

Clears this QWeakPointer object, dropping the reference that it may have had to the pointer.

fn data(&self) -> *mut Object

C++ method: QObject* QWeakPointer<QObject>::data() const

C++ documentation:

Returns the value of the pointer being tracked by this QWeakPointer, without ensuring that it cannot get deleted. To have that guarantee, use toStrongRef(), which returns a QSharedPointer object. If this function can determine that the pointer has already been deleted, it returns 0.

It is ok to obtain the value of the pointer and using that value itself, like for example in debugging statements:

qDebug("Tracking %p", weakref.data());

However, dereferencing the pointer is only allowed if you can guarantee by external means that the pointer does not get deleted. For example, if you can be certain that no other thread can delete it, nor the functions that you may call.

If that is the case, then the following code is valid:

// this pointer cannot be used in another thread
// so other threads cannot delete it
QWeakPointer<int> weakref = obtainReference();

Object *obj = weakref.data(); if (obj) { // if the pointer wasn't deleted yet, we know it can't get // deleted by our own code here nor the functions we call otherFunction(obj); }
Use this function with care.

This function was introduced in Qt 4.6.

See also isNull() and toStrongRef().

fn is_null(&self) -> bool

C++ method: bool QWeakPointer<QObject>::isNull() const

C++ documentation:

Returns true if this object is holding a reference to a null pointer.

Note that, due to the nature of weak references, the pointer that QWeakPointer references can become null at any moment, so the value returned from this function can change from false to true from one call to the next.

fn lock<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WeakPointerObjectRefLockArgs<'largs>

C++ method: QWeakPointer<QObject>::lock

This is an overloaded function. Available variants:

Rust arguments:
1) fn lock(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::shared_pointer::SharedPointerObjectRef>
2) fn lock(&self, cpp_utils::AsStruct) -> ::shared_pointer::SharedPointerObjectRef
C++ method: QSharedPointer<QObject> QWeakPointer<QObject>::lock() const

C++ documentation:

Same as toStrongRef().

This function is provided for API compatibility with std::weak_ptr.

This function was introduced in Qt 5.4.

fn new<Args>(args: Args) -> Args::ReturnType where Args: WeakPointerObjectRefNewArgs

C++ method: QWeakPointer<QObject>::QWeakPointer

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::weak_pointer::WeakPointerObjectRef
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::weak_pointer::WeakPointerObjectRef>
C++ method: [constructor] void QWeakPointer<QObject>::QWeakPointer()

Variant 2

Rust arguments:
1) fn new((&::shared_pointer::SharedPointerObjectRef, cpp_utils::AsStruct)) -> ::weak_pointer::WeakPointerObjectRef
2) fn new((&::shared_pointer::SharedPointerObjectRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::weak_pointer::WeakPointerObjectRef>
C++ method: [constructor] void QWeakPointer<QObject>::QWeakPointer(const QSharedPointer<QObject>& o)

Variant 3

Rust arguments:
1) fn new((&::weak_pointer::WeakPointerObjectRef, cpp_utils::AsStruct)) -> ::weak_pointer::WeakPointerObjectRef
2) fn new((&::weak_pointer::WeakPointerObjectRef, cpp_utils::AsBox)) -> cpp_utils::CppBox<::weak_pointer::WeakPointerObjectRef>
C++ method: [constructor] void QWeakPointer<QObject>::QWeakPointer(const QWeakPointer<QObject>& o)

fn op_assign<'largs, Args>(&'largs mut self, args: Args) -> &'largs mut WeakPointerObjectRef where Args: WeakPointerObjectRefOpAssignArgs<'largs>

C++ method: QWeakPointer<QObject>::operator=

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn op_assign(&mut self, &'l1 ::shared_pointer::SharedPointerObjectRef) -> &'l0 mut ::weak_pointer::WeakPointerObjectRef
C++ method: QWeakPointer<QObject>& QWeakPointer<QObject>::operator=(const QSharedPointer<QObject>& o)

C++ documentation:

Makes this object share other's pointer. The current pointer reference is discarded but is not deleted.

If T is a derived type of the template parameter of this class, QWeakPointer will perform an automatic cast. Otherwise, you will get a compiler error.

Variant 2

Rust arguments: fn op_assign(&mut self, &'l1 ::weak_pointer::WeakPointerObjectRef) -> &'l0 mut ::weak_pointer::WeakPointerObjectRef
C++ method: QWeakPointer<QObject>& QWeakPointer<QObject>::operator=(const QWeakPointer<QObject>& o)

fn op_not(&self) -> bool

C++ method: bool QWeakPointer<QObject>::operator!() const

C++ documentation:

Returns true if this object is null. This function is suitable for use in if-constructs, like:

if (!weakref) { ... }

Note that, due to the nature of weak references, the pointer that QWeakPointer references can become null at any moment, so the value returned from this function can change from false to true from one call to the next.

See also isNull().

fn swap(&mut self, other: &mut WeakPointerObjectRef)

C++ method: void QWeakPointer<QObject>::swap(QWeakPointer<QObject>& other)

Warning: no exact match found in C++ documentation.Below is the C++ documentation for void QWeakPointer::swap(QWeakPointer & other):

Swaps this weak pointer instance with other. This function is very fast and never fails.

This function was introduced in Qt 5.4.

fn to_strong_ref<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: WeakPointerObjectRefToStrongRefArgs<'largs>

C++ method: QWeakPointer<QObject>::toStrongRef

This is an overloaded function. Available variants:

Rust arguments:
1) fn to_strong_ref(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::shared_pointer::SharedPointerObjectRef>
2) fn to_strong_ref(&self, cpp_utils::AsStruct) -> ::shared_pointer::SharedPointerObjectRef
C++ method: QSharedPointer<QObject> QWeakPointer<QObject>::toStrongRef() const

C++ documentation:

Promotes this weak reference to a strong one and returns a QSharedPointer object holding that reference. When promoting to QSharedPointer, this function verifies if the object has been deleted already or not. If it hasn't, this function increases the reference count to the shared object, thus ensuring that it will not get deleted.

Since this function can fail to obtain a valid strong reference to the shared object, you should always verify if the conversion succeeded, by calling QSharedPointer::isNull() on the returned object.

For example, the following code promotes a QWeakPointer that was held to a strong reference and, if it succeeded, it prints the value of the integer that was held:

QWeakPointer<int> weakref;

// ...

QSharedPointer<int> strong = weakref.toStrongRef(); if (strong) qDebug() << "The value is:" << *strong; else qDebug() << "The value has already been deleted";
See also QSharedPointer::QSharedPointer().