Struct qt_core::xml_stream_attribute::XmlStreamAttribute

pub struct XmlStreamAttribute {
    // some fields omitted
}

C++ type: QXmlStreamAttribute

C++ documentation:

The QXmlStreamAttribute class represents a single XML attribute

An attribute consists of an optionally empty namespaceUri(), a name(), a value(), and an isDefault() attribute.

The raw XML attribute name is returned as qualifiedName().

Methods

impl XmlStreamAttribute

fn is_default(&self) -> bool

C++ method: bool QXmlStreamAttribute::isDefault() const

C++ documentation:

Returns true if the parser added this attribute with a default value following an ATTLIST declaration in the DTD; otherwise returns false.

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamAttributeNameArgs<'largs>

C++ method: QXmlStreamAttribute::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamAttribute::name() const

C++ documentation:

Returns the attribute's local name.

fn namespace_uri<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamAttributeNamespaceUriArgs<'largs>

C++ method: QXmlStreamAttribute::namespaceUri

This is an overloaded function. Available variants:

Rust arguments:
1) fn namespace_uri(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn namespace_uri(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamAttribute::namespaceUri() const

C++ documentation:

Returns the attribute's resolved namespaceUri, or an empty string reference if the attribute does not have a defined namespace.

fn new<Args>(args: Args) -> Args::ReturnType where Args: XmlStreamAttributeNewArgs

C++ method: QXmlStreamAttribute::QXmlStreamAttribute

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::xml_stream_attribute::XmlStreamAttribute
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_attribute::XmlStreamAttribute>
C++ method: [constructor] void QXmlStreamAttribute::QXmlStreamAttribute()

Variant 2

Rust arguments:
1) fn new((&::string::String, &::string::String, &::string::String, cpp_utils::AsStruct)) -> ::xml_stream_attribute::XmlStreamAttribute
2) fn new((&::string::String, &::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_attribute::XmlStreamAttribute>
C++ method: [constructor] void QXmlStreamAttribute::QXmlStreamAttribute(const QString& namespaceUri, const QString& name, const QString& value)

Variant 3

Rust arguments:
1) fn new((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::xml_stream_attribute::XmlStreamAttribute
2) fn new((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_attribute::XmlStreamAttribute>
C++ method: [constructor] void QXmlStreamAttribute::QXmlStreamAttribute(const QString& qualifiedName, const QString& value)

Variant 4

Rust arguments:
1) fn new((&::xml_stream_attribute::XmlStreamAttribute, cpp_utils::AsStruct)) -> ::xml_stream_attribute::XmlStreamAttribute
2) fn new((&::xml_stream_attribute::XmlStreamAttribute, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_attribute::XmlStreamAttribute>
C++ method: [constructor] void QXmlStreamAttribute::QXmlStreamAttribute(const QXmlStreamAttribute& arg1)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 XmlStreamAttribute) -> &'l0 mut XmlStreamAttribute

C++ method: QXmlStreamAttribute& QXmlStreamAttribute::operator=(const QXmlStreamAttribute& arg1)

C++ documentation:

Assigns other to this attribute.

fn op_eq(&self, other: &XmlStreamAttribute) -> bool

C++ method: bool QXmlStreamAttribute::operator==(const QXmlStreamAttribute& other) const

C++ documentation:

Compares this attribute with other and returns true if they are equal; otherwise returns false.

fn op_neq(&self, other: &XmlStreamAttribute) -> bool

C++ method: bool QXmlStreamAttribute::operator!=(const QXmlStreamAttribute& other) const

C++ documentation:

Compares this attribute with other and returns true if they are not equal; otherwise returns false.

fn prefix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamAttributePrefixArgs<'largs>

C++ method: QXmlStreamAttribute::prefix

This is an overloaded function. Available variants:

Rust arguments:
1) fn prefix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn prefix(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamAttribute::prefix() const

C++ documentation:

Returns the attribute's namespace prefix.

This function was introduced in Qt 4.4.

See also name() and qualifiedName().

fn qualified_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamAttributeQualifiedNameArgs<'largs>

C++ method: QXmlStreamAttribute::qualifiedName

This is an overloaded function. Available variants:

Rust arguments:
1) fn qualified_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn qualified_name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamAttribute::qualifiedName() const

C++ documentation:

Returns the attribute's qualified name.

A qualified name is the raw name of an attribute in the XML data. It consists of the namespace prefix(), followed by colon, followed by the attribute's local name(). Since the namespace prefix is not unique (the same prefix can point to different namespaces and different prefixes can point to the same namespace), you shouldn't use qualifiedName(), but the resolved namespaceUri() and the attribute's local name().

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamAttributeValueArgs<'largs>

C++ method: QXmlStreamAttribute::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn value(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamAttribute::value() const

C++ documentation:

Returns the attribute's value.