Module qt_core::xml_stream_entity_declaration::overloading

Traits

XmlStreamEntityDeclarationNameArgs
This trait represents a set of arguments accepted by XmlStreamEntityDeclaration::name method.

XmlStreamEntityDeclarationNewArgs
This trait represents a set of arguments accepted by XmlStreamEntityDeclaration::new method.

XmlStreamEntityDeclarationNotationNameArgs
This trait represents a set of arguments accepted by XmlStreamEntityDeclaration::notation_name method.

XmlStreamEntityDeclarationPublicIdArgs
This trait represents a set of arguments accepted by XmlStreamEntityDeclaration::public_id method.

XmlStreamEntityDeclarationSystemIdArgs
This trait represents a set of arguments accepted by XmlStreamEntityDeclaration::system_id method.

XmlStreamEntityDeclarationValueArgs
This trait represents a set of arguments accepted by XmlStreamEntityDeclaration::value method.