Struct qt_core::xml_stream_entity_declaration::XmlStreamEntityDeclaration

pub struct XmlStreamEntityDeclaration {
    // some fields omitted
}

C++ type: QXmlStreamEntityDeclaration

C++ documentation:

The QXmlStreamEntityDeclaration class represents a DTD entity declaration.

An entity declaration consists of a name(), a notationName(), a systemId(), a publicId(), and a value().

Methods

impl XmlStreamEntityDeclaration

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamEntityDeclarationNameArgs<'largs>

C++ method: QXmlStreamEntityDeclaration::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamEntityDeclaration::name() const

C++ documentation:

Returns the entity name.

fn new<Args>(args: Args) -> Args::ReturnType where Args: XmlStreamEntityDeclarationNewArgs

C++ method: QXmlStreamEntityDeclaration::QXmlStreamEntityDeclaration

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::xml_stream_entity_declaration::XmlStreamEntityDeclaration
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_entity_declaration::XmlStreamEntityDeclaration>
C++ method: [constructor] void QXmlStreamEntityDeclaration::QXmlStreamEntityDeclaration()

Variant 2

Rust arguments:
1) fn new((&::xml_stream_entity_declaration::XmlStreamEntityDeclaration, cpp_utils::AsStruct)) -> ::xml_stream_entity_declaration::XmlStreamEntityDeclaration
2) fn new((&::xml_stream_entity_declaration::XmlStreamEntityDeclaration, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_entity_declaration::XmlStreamEntityDeclaration>
C++ method: [constructor] void QXmlStreamEntityDeclaration::QXmlStreamEntityDeclaration(const QXmlStreamEntityDeclaration& arg1)

fn notation_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamEntityDeclarationNotationNameArgs<'largs>

C++ method: QXmlStreamEntityDeclaration::notationName

This is an overloaded function. Available variants:

Rust arguments:
1) fn notation_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn notation_name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamEntityDeclaration::notationName() const

C++ documentation:

Returns the notation name.

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 XmlStreamEntityDeclaration) -> &'l0 mut XmlStreamEntityDeclaration

C++ method: QXmlStreamEntityDeclaration& QXmlStreamEntityDeclaration::operator=(const QXmlStreamEntityDeclaration& arg1)

C++ documentation:

Assigns other to this entity declaration.

fn op_eq(&self, other: &XmlStreamEntityDeclaration) -> bool

C++ method: bool QXmlStreamEntityDeclaration::operator==(const QXmlStreamEntityDeclaration& other) const

C++ documentation:

Compares this entity declaration with other and returns true if they are equal; otherwise returns false.

fn op_neq(&self, other: &XmlStreamEntityDeclaration) -> bool

C++ method: bool QXmlStreamEntityDeclaration::operator!=(const QXmlStreamEntityDeclaration& other) const

C++ documentation:

Compares this entity declaration with other and returns true if they are not equal; otherwise returns false.

fn public_id<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamEntityDeclarationPublicIdArgs<'largs>

C++ method: QXmlStreamEntityDeclaration::publicId

This is an overloaded function. Available variants:

Rust arguments:
1) fn public_id(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn public_id(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamEntityDeclaration::publicId() const

C++ documentation:

Returns the public identifier.

fn system_id<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamEntityDeclarationSystemIdArgs<'largs>

C++ method: QXmlStreamEntityDeclaration::systemId

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_id(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn system_id(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamEntityDeclaration::systemId() const

C++ documentation:

Returns the system identifier.

fn value<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamEntityDeclarationValueArgs<'largs>

C++ method: QXmlStreamEntityDeclaration::value

This is an overloaded function. Available variants:

Rust arguments:
1) fn value(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn value(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamEntityDeclaration::value() const

C++ documentation:

Returns the entity's value.