Struct qt_core::xml_stream_entity_resolver::XmlStreamEntityResolver

pub struct XmlStreamEntityResolver {
    // some fields omitted
}

C++ type: QXmlStreamEntityResolver

C++ documentation:

The QXmlStreamEntityResolver class provides an entity resolver for a QXmlStreamReader.

Methods

impl XmlStreamEntityResolver

fn resolve_entity<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: XmlStreamEntityResolverResolveEntityArgs<'largs>

C++ method: QXmlStreamEntityResolver::resolveEntity

This is an overloaded function. Available variants:

Rust arguments:
1) fn resolve_entity(&mut self, (&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn resolve_entity(&mut self, (&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: virtual QString QXmlStreamEntityResolver::resolveEntity(const QString& publicId, const QString& systemId)

fn resolve_undeclared_entity<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: XmlStreamEntityResolverResolveUndeclaredEntityArgs<'largs>

C++ method: QXmlStreamEntityResolver::resolveUndeclaredEntity

This is an overloaded function. Available variants:

Rust arguments:
1) fn resolve_undeclared_entity(&mut self, (&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn resolve_undeclared_entity(&mut self, (&::string::String, cpp_utils::AsStruct)) -> ::string::String
C++ method: virtual QString QXmlStreamEntityResolver::resolveUndeclaredEntity(const QString& name)

C++ documentation:

Resolves the undeclared entity name and returns its replacement text. If the entity is also unknown to the entity resolver, it returns an empty string.

The default implementation always returns an empty string.