Struct qt_core::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration

pub struct XmlStreamNamespaceDeclaration {
    // some fields omitted
}

C++ type: QXmlStreamNamespaceDeclaration

C++ documentation:

The QXmlStreamNamespaceDeclaration class represents a namespace declaration.

An namespace declaration consists of a prefix() and a namespaceUri().

Methods

impl XmlStreamNamespaceDeclaration

fn namespace_uri<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamNamespaceDeclarationNamespaceUriArgs<'largs>

C++ method: QXmlStreamNamespaceDeclaration::namespaceUri

This is an overloaded function. Available variants:

Rust arguments:
1) fn namespace_uri(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn namespace_uri(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamNamespaceDeclaration::namespaceUri() const

C++ documentation:

Returns the namespaceUri.

fn new<Args>(args: Args) -> Args::ReturnType where Args: XmlStreamNamespaceDeclarationNewArgs

C++ method: QXmlStreamNamespaceDeclaration::QXmlStreamNamespaceDeclaration

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration>
C++ method: [constructor] void QXmlStreamNamespaceDeclaration::QXmlStreamNamespaceDeclaration()

Variant 2

Rust arguments:
1) fn new((&::string::String, &::string::String, cpp_utils::AsStruct)) -> ::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration
2) fn new((&::string::String, &::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration>
C++ method: [constructor] void QXmlStreamNamespaceDeclaration::QXmlStreamNamespaceDeclaration(const QString& prefix, const QString& namespaceUri)

Variant 3

Rust arguments:
1) fn new((&::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration, cpp_utils::AsStruct)) -> ::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration
2) fn new((&::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_namespace_declaration::XmlStreamNamespaceDeclaration>
C++ method: [constructor] void QXmlStreamNamespaceDeclaration::QXmlStreamNamespaceDeclaration(const QXmlStreamNamespaceDeclaration& arg1)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 XmlStreamNamespaceDeclaration) -> &'l0 mut XmlStreamNamespaceDeclaration

C++ method: QXmlStreamNamespaceDeclaration& QXmlStreamNamespaceDeclaration::operator=(const QXmlStreamNamespaceDeclaration& arg1)

C++ documentation:

Assigns other to this namespace declaration.

fn op_eq(&self, other: &XmlStreamNamespaceDeclaration) -> bool

C++ method: bool QXmlStreamNamespaceDeclaration::operator==(const QXmlStreamNamespaceDeclaration& other) const

C++ documentation:

Compares this namespace declaration with other and returns true if they are equal; otherwise returns false.

fn op_neq(&self, other: &XmlStreamNamespaceDeclaration) -> bool

C++ method: bool QXmlStreamNamespaceDeclaration::operator!=(const QXmlStreamNamespaceDeclaration& other) const

C++ documentation:

Compares this namespace declaration with other and returns true if they are not equal; otherwise returns false.

fn prefix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamNamespaceDeclarationPrefixArgs<'largs>

C++ method: QXmlStreamNamespaceDeclaration::prefix

This is an overloaded function. Available variants:

Rust arguments:
1) fn prefix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn prefix(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamNamespaceDeclaration::prefix() const

C++ documentation:

Returns the prefix.