Struct qt_core::xml_stream_notation_declaration::XmlStreamNotationDeclaration

pub struct XmlStreamNotationDeclaration {
    // some fields omitted
}

C++ type: QXmlStreamNotationDeclaration

C++ documentation:

The QXmlStreamNotationDeclaration class represents a DTD notation declaration.

An notation declaration consists of a name(), a systemId(), and a publicId().

Methods

impl XmlStreamNotationDeclaration

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamNotationDeclarationNameArgs<'largs>

C++ method: QXmlStreamNotationDeclaration::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamNotationDeclaration::name() const

C++ documentation:

Returns the notation name.

fn new<Args>(args: Args) -> Args::ReturnType where Args: XmlStreamNotationDeclarationNewArgs

C++ method: QXmlStreamNotationDeclaration::QXmlStreamNotationDeclaration

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::xml_stream_notation_declaration::XmlStreamNotationDeclaration
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_notation_declaration::XmlStreamNotationDeclaration>
C++ method: [constructor] void QXmlStreamNotationDeclaration::QXmlStreamNotationDeclaration()

Variant 2

Rust arguments:
1) fn new((&::xml_stream_notation_declaration::XmlStreamNotationDeclaration, cpp_utils::AsStruct)) -> ::xml_stream_notation_declaration::XmlStreamNotationDeclaration
2) fn new((&::xml_stream_notation_declaration::XmlStreamNotationDeclaration, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_notation_declaration::XmlStreamNotationDeclaration>
C++ method: [constructor] void QXmlStreamNotationDeclaration::QXmlStreamNotationDeclaration(const QXmlStreamNotationDeclaration& arg1)

fn op_assign<'l0, 'l1>(&'l0 mut self, arg1: &'l1 XmlStreamNotationDeclaration) -> &'l0 mut XmlStreamNotationDeclaration

C++ method: QXmlStreamNotationDeclaration& QXmlStreamNotationDeclaration::operator=(const QXmlStreamNotationDeclaration& arg1)

C++ documentation:

Assigns other to this notation declaration.

fn op_eq(&self, other: &XmlStreamNotationDeclaration) -> bool

C++ method: bool QXmlStreamNotationDeclaration::operator==(const QXmlStreamNotationDeclaration& other) const

C++ documentation:

Compares this notation declaration with other and returns true if they are equal; otherwise returns false.

fn op_neq(&self, other: &XmlStreamNotationDeclaration) -> bool

C++ method: bool QXmlStreamNotationDeclaration::operator!=(const QXmlStreamNotationDeclaration& other) const

C++ documentation:

Compares this notation declaration with other and returns true if they are not equal; otherwise returns false.

fn public_id<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamNotationDeclarationPublicIdArgs<'largs>

C++ method: QXmlStreamNotationDeclaration::publicId

This is an overloaded function. Available variants:

Rust arguments:
1) fn public_id(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn public_id(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamNotationDeclaration::publicId() const

C++ documentation:

Returns the public identifier.

fn system_id<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamNotationDeclarationSystemIdArgs<'largs>

C++ method: QXmlStreamNotationDeclaration::systemId

This is an overloaded function. Available variants:

Rust arguments:
1) fn system_id(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn system_id(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamNotationDeclaration::systemId() const

C++ documentation:

Returns the system identifier.