Enum qt_core::xml_stream_reader::TokenType

pub enum TokenType {
    NoToken,
    Invalid,
    StartDocument,
    EndDocument,
    StartElement,
    EndElement,
    Characters,
    Comment,
    Dtd,
    EntityReference,
    ProcessingInstruction,
}

C++ type: QXmlStreamReader::TokenType

C++ documentation:

This enum specifies the type of token the reader just read.

Variants
NoToken

The reader has not yet read anything. (C++ enum variant: NoToken = 0)

Invalid

An error has occurred, reported in error() and errorString(). (C++ enum variant: Invalid = 1)

StartDocument

The reader reports the XML version number in documentVersion(), and the encoding as specified in the XML document in documentEncoding(). If the document is declared standalone, isStandaloneDocument() returns true; otherwise it returns false. (C++ enum variant: StartDocument = 2)

EndDocument

The reader reports the end of the document. (C++ enum variant: EndDocument = 3)

StartElement

The reader reports the start of an element with namespaceUri() and name(). Empty elements are also reported as StartElement, followed directly by EndElement. The convenience function readElementText() can be called to concatenate all content until the corresponding EndElement. Attributes are reported in attributes(), namespace declarations in namespaceDeclarations(). (C++ enum variant: StartElement = 4)

EndElement

The reader reports the end of an element with namespaceUri() and name(). (C++ enum variant: EndElement = 5)

Characters

The reader reports characters in text(). If the characters are all white-space, isWhitespace() returns true. If the characters stem from a CDATA section, isCDATA() returns true. (C++ enum variant: Characters = 6)

Comment

The reader reports a comment in text(). (C++ enum variant: Comment = 7)

Dtd

The reader reports a DTD in text(), notation declarations in notationDeclarations(), and entity declarations in entityDeclarations(). Details of the DTD declaration are reported in in dtdName(), dtdPublicId(), and dtdSystemId(). (C++ enum variant: DTD = 8)

EntityReference

The reader reports an entity reference that could not be resolved. The name of the reference is reported in name(), the replacement text in text(). (C++ enum variant: EntityReference = 9)

ProcessingInstruction

The reader reports a processing instruction in processingInstructionTarget() and processingInstructionData(). (C++ enum variant: ProcessingInstruction = 10)