Module qt_core::xml_stream_reader::overloading

Traits

XmlStreamReaderAddDataArgs
This trait represents a set of arguments accepted by XmlStreamReader::add_data method.

XmlStreamReaderAttributesArgs
This trait represents a set of arguments accepted by XmlStreamReader::attributes method.

XmlStreamReaderDocumentEncodingArgs
This trait represents a set of arguments accepted by XmlStreamReader::document_encoding method.

XmlStreamReaderDocumentVersionArgs
This trait represents a set of arguments accepted by XmlStreamReader::document_version method.

XmlStreamReaderDtdNameArgs
This trait represents a set of arguments accepted by XmlStreamReader::dtd_name method.

XmlStreamReaderDtdPublicIdArgs
This trait represents a set of arguments accepted by XmlStreamReader::dtd_public_id method.

XmlStreamReaderDtdSystemIdArgs
This trait represents a set of arguments accepted by XmlStreamReader::dtd_system_id method.

XmlStreamReaderEntityDeclarationsArgs
This trait represents a set of arguments accepted by XmlStreamReader::entity_declarations method.

XmlStreamReaderErrorStringArgs
This trait represents a set of arguments accepted by XmlStreamReader::error_string method.

XmlStreamReaderNameArgs
This trait represents a set of arguments accepted by XmlStreamReader::name method.

XmlStreamReaderNamespaceDeclarationsArgs
This trait represents a set of arguments accepted by XmlStreamReader::namespace_declarations method.

XmlStreamReaderNamespaceUriArgs
This trait represents a set of arguments accepted by XmlStreamReader::namespace_uri method.

XmlStreamReaderNewArgs
This trait represents a set of arguments accepted by XmlStreamReader::new method.

XmlStreamReaderNotationDeclarationsArgs
This trait represents a set of arguments accepted by XmlStreamReader::notation_declarations method.

XmlStreamReaderPrefixArgs
This trait represents a set of arguments accepted by XmlStreamReader::prefix method.

XmlStreamReaderProcessingInstructionDataArgs
This trait represents a set of arguments accepted by XmlStreamReader::processing_instruction_data method.

XmlStreamReaderProcessingInstructionTargetArgs
This trait represents a set of arguments accepted by XmlStreamReader::processing_instruction_target method.

XmlStreamReaderQualifiedNameArgs
This trait represents a set of arguments accepted by XmlStreamReader::qualified_name method.

XmlStreamReaderRaiseErrorArgs
This trait represents a set of arguments accepted by XmlStreamReader::raise_error method.

XmlStreamReaderReadElementTextArgs
This trait represents a set of arguments accepted by XmlStreamReader::read_element_text method.

XmlStreamReaderTextArgs
This trait represents a set of arguments accepted by XmlStreamReader::text method.

XmlStreamReaderTokenStringArgs
This trait represents a set of arguments accepted by XmlStreamReader::token_string method.