Struct qt_core::xml_stream_reader::XmlStreamReader

pub struct XmlStreamReader {
    // some fields omitted
}

C++ type: QXmlStreamReader

C++ documentation:

The QXmlStreamReader class provides a fast parser for reading well-formed XML via a simple streaming API.

QXmlStreamReader is a faster and more convenient replacement for Qt's own SAX parser (see QXmlSimpleReader). In some cases it might also be a faster and more convenient alternative for use in applications that would otherwise use a DOM tree (see QDomDocument). QXmlStreamReader reads data either from a QIODevice (see setDevice()), or from a raw QByteArray (see addData()).

Qt provides QXmlStreamWriter for writing XML.

The basic concept of a stream reader is to report an XML document as a stream of tokens, similar to SAX. The main difference between QXmlStreamReader and SAX is how these XML tokens are reported. With SAX, the application must provide handlers (callback functions) that receive so-called XML events from the parser at the parser's convenience. With QXmlStreamReader, the application code itself drives the loop and pulls tokens from the reader, one after another, as it needs them. This is done by calling readNext(), where the reader reads from the input stream until it completes the next token, at which point it returns the tokenType(). A set of convenient functions including isStartElement() and text() can then be used to examine the token to obtain information about what has been read. The big advantage of this pulling approach is the possibility to build recursive descent parsers with it, meaning you can split your XML parsing code easily into different methods or classes. This makes it easy to keep track of the application's own state when parsing XML.

A typical loop with QXmlStreamReader looks like this:

  QXmlStreamReader xml;
  ...
  while (!xml.atEnd()) {
        xml.readNext();
        ... // do processing
  }
  if (xml.hasError()) {
        ... // do error handling
  }

QXmlStreamReader is a well-formed XML 1.0 parser that does not include external parsed entities. As long as no error occurs, the application code can thus be assured that the data provided by the stream reader satisfies the W3C's criteria for well-formed XML. For example, you can be certain that all tags are indeed nested and closed properly, that references to internal entities have been replaced with the correct replacement text, and that attributes have been normalized or added according to the internal subset of the DTD.

If an error occurs while parsing, atEnd() and hasError() return true, and error() returns the error that occurred. The functions errorString(), lineNumber(), columnNumber(), and characterOffset() are for constructing an appropriate error or warning message. To simplify application code, QXmlStreamReader contains a raiseError() mechanism that lets you raise custom errors that trigger the same error handling described.

The QXmlStream Bookmarks Example illustrates how to use the recursive descent technique to read an XML bookmark file (XBEL) with a stream reader.

Methods

impl XmlStreamReader

fn add_data<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamReaderAddDataArgs<'largs>

C++ method: QXmlStreamReader::addData

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn add_data(&mut self, &::byte_array::ByteArray) -> ()
C++ method: void QXmlStreamReader::addData(const QByteArray& data)

C++ documentation:

Adds more data for the reader to read. This function does nothing if the reader has a device().

See also readNext() and clear().

Variant 2

Rust arguments: fn add_data(&mut self, &::string::String) -> ()
C++ method: void QXmlStreamReader::addData(const QString& data)

C++ documentation:

Adds more data for the reader to read. This function does nothing if the reader has a device().

See also readNext() and clear().

Variant 3

Rust arguments: fn add_data(&mut self, *const libc::c_char) -> ()
C++ method: void QXmlStreamReader::addData(const char* data)

C++ documentation:

Adds more data for the reader to read. This function does nothing if the reader has a device().

See also readNext() and clear().

fn add_extra_namespace_declaration(&mut self, extra_namespace_declaraction: &XmlStreamNamespaceDeclaration)

C++ method: void QXmlStreamReader::addExtraNamespaceDeclaration(const QXmlStreamNamespaceDeclaration& extraNamespaceDeclaraction)

C++ documentation:

Adds an extraNamespaceDeclaration. The declaration will be valid for children of the current element, or - should the function be called before any elements are read - for the entire XML document.

This function was introduced in Qt 4.4.

See also namespaceDeclarations(), addExtraNamespaceDeclarations(), and setNamespaceProcessing().

fn add_extra_namespace_declarations(&mut self, extra_namespace_declaractions: &VectorXmlStreamNamespaceDeclarationRef)

C++ method: void QXmlStreamReader::addExtraNamespaceDeclarations(const QVector<QXmlStreamNamespaceDeclaration>& extraNamespaceDeclaractions)

C++ documentation:

Adds a vector of declarations specified by extraNamespaceDeclarations.

This function was introduced in Qt 4.4.

See also namespaceDeclarations() and addExtraNamespaceDeclaration().

fn at_end(&self) -> bool

C++ method: bool QXmlStreamReader::atEnd() const

C++ documentation:

Returns true if the reader has read until the end of the XML document, or if an error() has occurred and reading has been aborted. Otherwise, it returns false.

When atEnd() and hasError() return true and error() returns PrematureEndOfDocumentError, it means the XML has been well-formed so far, but a complete XML document has not been parsed. The next chunk of XML can be added with addData(), if the XML is being read from a QByteArray, or by waiting for more data to arrive if the XML is being read from a QIODevice. Either way, atEnd() will return false once more data is available.

See also hasError(), error(), device(), and QIODevice::atEnd().

fn attributes<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderAttributesArgs<'largs>

C++ method: QXmlStreamReader::attributes

This is an overloaded function. Available variants:

Rust arguments:
1) fn attributes(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_attributes::XmlStreamAttributes>
2) fn attributes(&self, cpp_utils::AsStruct) -> ::xml_stream_attributes::XmlStreamAttributes
C++ method: QXmlStreamAttributes QXmlStreamReader::attributes() const

C++ documentation:

Returns the attributes of a StartElement.

fn character_offset(&self) -> i64

C++ method: qint64 QXmlStreamReader::characterOffset() const

C++ documentation:

Returns the current character offset, starting with 0.

See also lineNumber() and columnNumber().

fn clear(&mut self)

C++ method: void QXmlStreamReader::clear()

C++ documentation:

Removes any device() or data from the reader and resets its internal state to the initial state.

See also addData().

fn column_number(&self) -> i64

C++ method: qint64 QXmlStreamReader::columnNumber() const

C++ documentation:

Returns the current column number, starting with 0.

See also lineNumber() and characterOffset().

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QXmlStreamReader::device() const

C++ documentation:

Returns the current device associated with the QXmlStreamReader, or 0 if no device has been assigned.

See also setDevice().

fn document_encoding<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderDocumentEncodingArgs<'largs>

C++ method: QXmlStreamReader::documentEncoding

This is an overloaded function. Available variants:

Rust arguments:
1) fn document_encoding(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn document_encoding(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::documentEncoding() const

C++ documentation:

If the tokenType() is StartDocument, this function returns the encoding string as specified in the XML declaration. Otherwise an empty string is returned.

This function was introduced in Qt 4.4.

fn document_version<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderDocumentVersionArgs<'largs>

C++ method: QXmlStreamReader::documentVersion

This is an overloaded function. Available variants:

Rust arguments:
1) fn document_version(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn document_version(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::documentVersion() const

C++ documentation:

If the tokenType() is StartDocument, this function returns the version string as specified in the XML declaration. Otherwise an empty string is returned.

This function was introduced in Qt 4.4.

fn dtd_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderDtdNameArgs<'largs>

C++ method: QXmlStreamReader::dtdName

This is an overloaded function. Available variants:

Rust arguments:
1) fn dtd_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn dtd_name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::dtdName() const

C++ documentation:

If the tokenType() is DTD, this function returns the DTD's name. Otherwise an empty string is returned.

This function was introduced in Qt 4.4.

fn dtd_public_id<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderDtdPublicIdArgs<'largs>

C++ method: QXmlStreamReader::dtdPublicId

This is an overloaded function. Available variants:

Rust arguments:
1) fn dtd_public_id(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn dtd_public_id(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::dtdPublicId() const

C++ documentation:

If the tokenType() is DTD, this function returns the DTD's public identifier. Otherwise an empty string is returned.

This function was introduced in Qt 4.4.

fn dtd_system_id<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderDtdSystemIdArgs<'largs>

C++ method: QXmlStreamReader::dtdSystemId

This is an overloaded function. Available variants:

Rust arguments:
1) fn dtd_system_id(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn dtd_system_id(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::dtdSystemId() const

C++ documentation:

If the tokenType() is DTD, this function returns the DTD's system identifier. Otherwise an empty string is returned.

This function was introduced in Qt 4.4.

fn entity_declarations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderEntityDeclarationsArgs<'largs>

C++ method: QXmlStreamReader::entityDeclarations

This is an overloaded function. Available variants:

Rust arguments:
1) fn entity_declarations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorXmlStreamEntityDeclarationRef>
2) fn entity_declarations(&self, cpp_utils::AsStruct) -> ::vector::VectorXmlStreamEntityDeclarationRef
C++ method: QVector<QXmlStreamEntityDeclaration> QXmlStreamReader::entityDeclarations() const

C++ documentation:

If the tokenType() is DTD, this function returns the DTD's unparsed (external) entity declarations. Otherwise an empty vector is returned.

The QXmlStreamEntityDeclarations class is defined to be a QVector of QXmlStreamEntityDeclaration.

fn entity_resolver(&self) -> *mut XmlStreamEntityResolver

C++ method: QXmlStreamEntityResolver* QXmlStreamReader::entityResolver() const

C++ documentation:

Returns the entity resolver, or 0 if there is no entity resolver.

This function was introduced in Qt 4.4.

See also setEntityResolver().

fn error(&self) -> Error

C++ method: QXmlStreamReader::Error QXmlStreamReader::error() const

C++ documentation:

Returns the type of the current error, or NoError if no error occurred.

See also errorString() and raiseError().

fn error_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderErrorStringArgs<'largs>

C++ method: QXmlStreamReader::errorString

This is an overloaded function. Available variants:

Rust arguments:
1) fn error_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn error_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QXmlStreamReader::errorString() const

C++ documentation:

Returns the error message that was set with raiseError().

See also error(), lineNumber(), columnNumber(), and characterOffset().

fn has_error(&self) -> bool

C++ method: bool QXmlStreamReader::hasError() const

C++ documentation:

Returns true if an error has occurred, otherwise false.

See also errorString() and error().

fn is_cdata(&self) -> bool

C++ method: bool QXmlStreamReader::isCDATA() const

C++ documentation:

Returns true if the reader reports characters that stem from a CDATA section; otherwise returns false.

See also isCharacters() and text().

fn is_characters(&self) -> bool

C++ method: bool QXmlStreamReader::isCharacters() const

C++ documentation:

Returns true if tokenType() equals Characters; otherwise returns false.

See also isWhitespace() and isCDATA().

fn is_comment(&self) -> bool

C++ method: bool QXmlStreamReader::isComment() const

C++ documentation:

Returns true if tokenType() equals Comment; otherwise returns false.

fn is_dtd(&self) -> bool

C++ method: bool QXmlStreamReader::isDTD() const

C++ documentation:

Returns true if tokenType() equals DTD; otherwise returns false.

fn is_end_document(&self) -> bool

C++ method: bool QXmlStreamReader::isEndDocument() const

C++ documentation:

Returns true if tokenType() equals EndDocument; otherwise returns false.

fn is_end_element(&self) -> bool

C++ method: bool QXmlStreamReader::isEndElement() const

C++ documentation:

Returns true if tokenType() equals EndElement; otherwise returns false.

fn is_entity_reference(&self) -> bool

C++ method: bool QXmlStreamReader::isEntityReference() const

C++ documentation:

Returns true if tokenType() equals EntityReference; otherwise returns false.

fn is_processing_instruction(&self) -> bool

C++ method: bool QXmlStreamReader::isProcessingInstruction() const

C++ documentation:

Returns true if tokenType() equals ProcessingInstruction; otherwise returns false.

fn is_standalone_document(&self) -> bool

C++ method: bool QXmlStreamReader::isStandaloneDocument() const

C++ documentation:

Returns true if this document has been declared standalone in the XML declaration; otherwise returns false.

If no XML declaration has been parsed, this function returns false.

fn is_start_document(&self) -> bool

C++ method: bool QXmlStreamReader::isStartDocument() const

C++ documentation:

Returns true if tokenType() equals StartDocument; otherwise returns false.

fn is_start_element(&self) -> bool

C++ method: bool QXmlStreamReader::isStartElement() const

C++ documentation:

Returns true if tokenType() equals StartElement; otherwise returns false.

fn is_whitespace(&self) -> bool

C++ method: bool QXmlStreamReader::isWhitespace() const

C++ documentation:

Returns true if the reader reports characters that only consist of white-space; otherwise returns false.

See also isCharacters() and text().

fn line_number(&self) -> i64

C++ method: qint64 QXmlStreamReader::lineNumber() const

C++ documentation:

Returns the current line number, starting with 1.

See also columnNumber() and characterOffset().

fn name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderNameArgs<'largs>

C++ method: QXmlStreamReader::name

This is an overloaded function. Available variants:

Rust arguments:
1) fn name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::name() const

C++ documentation:

Returns the local name of a StartElement, EndElement, or an EntityReference.

See also namespaceUri() and qualifiedName().

fn namespace_declarations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderNamespaceDeclarationsArgs<'largs>

C++ method: QXmlStreamReader::namespaceDeclarations

This is an overloaded function. Available variants:

Rust arguments:
1) fn namespace_declarations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorXmlStreamNamespaceDeclarationRef>
2) fn namespace_declarations(&self, cpp_utils::AsStruct) -> ::vector::VectorXmlStreamNamespaceDeclarationRef
C++ method: QVector<QXmlStreamNamespaceDeclaration> QXmlStreamReader::namespaceDeclarations() const

C++ documentation:

If the tokenType() is StartElement, this function returns the element's namespace declarations. Otherwise an empty vector is returned.

The QXmlStreamNamespaceDeclarations class is defined to be a QVector of QXmlStreamNamespaceDeclaration.

See also addExtraNamespaceDeclaration() and addExtraNamespaceDeclarations().

fn namespace_processing(&self) -> bool

C++ method: bool QXmlStreamReader::namespaceProcessing() const

C++ documentation:

The namespace-processing flag of the stream reader

This property controls whether or not the stream reader processes namespaces. If enabled, the reader processes namespaces, otherwise it does not.

By default, namespace-processing is enabled.

Access functions:

bool 	namespaceProcessing() const

void 	setNamespaceProcessing(bool)

fn namespace_uri<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderNamespaceUriArgs<'largs>

C++ method: QXmlStreamReader::namespaceUri

This is an overloaded function. Available variants:

Rust arguments:
1) fn namespace_uri(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn namespace_uri(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::namespaceUri() const

C++ documentation:

Returns the namespaceUri of a StartElement or EndElement.

See also name() and qualifiedName().

fn new<Args>(args: Args) -> Args::ReturnType where Args: XmlStreamReaderNewArgs

C++ method: QXmlStreamReader::QXmlStreamReader

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::xml_stream_reader::XmlStreamReader
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_reader::XmlStreamReader>
C++ method: [constructor] void QXmlStreamReader::QXmlStreamReader()

Variant 2

Rust arguments:
1) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::xml_stream_reader::XmlStreamReader
2) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_reader::XmlStreamReader>
C++ method: [constructor] void QXmlStreamReader::QXmlStreamReader(QIODevice* device)

Variant 3

Rust arguments:
1) fn new((&::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::xml_stream_reader::XmlStreamReader
2) fn new((&::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_reader::XmlStreamReader>
C++ method: [constructor] void QXmlStreamReader::QXmlStreamReader(const QByteArray& data)

Variant 4

Rust arguments:
1) fn new((&::string::String, cpp_utils::AsStruct)) -> ::xml_stream_reader::XmlStreamReader
2) fn new((&::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_reader::XmlStreamReader>
C++ method: [constructor] void QXmlStreamReader::QXmlStreamReader(const QString& data)

Variant 5

Rust arguments:
1) fn new((*const libc::c_char, cpp_utils::AsStruct)) -> ::xml_stream_reader::XmlStreamReader
2) fn new((*const libc::c_char, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_reader::XmlStreamReader>
C++ method: [constructor] void QXmlStreamReader::QXmlStreamReader(const char* data)

fn notation_declarations<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderNotationDeclarationsArgs<'largs>

C++ method: QXmlStreamReader::notationDeclarations

This is an overloaded function. Available variants:

Rust arguments:
1) fn notation_declarations(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::vector::VectorXmlStreamNotationDeclarationRef>
2) fn notation_declarations(&self, cpp_utils::AsStruct) -> ::vector::VectorXmlStreamNotationDeclarationRef
C++ method: QVector<QXmlStreamNotationDeclaration> QXmlStreamReader::notationDeclarations() const

C++ documentation:

If the tokenType() is DTD, this function returns the DTD's notation declarations. Otherwise an empty vector is returned.

The QXmlStreamNotationDeclarations class is defined to be a QVector of QXmlStreamNotationDeclaration.

fn prefix<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderPrefixArgs<'largs>

C++ method: QXmlStreamReader::prefix

This is an overloaded function. Available variants:

Rust arguments:
1) fn prefix(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn prefix(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::prefix() const

C++ documentation:

Returns the prefix of a StartElement or EndElement.

This function was introduced in Qt 4.4.

See also name() and qualifiedName().

fn processing_instruction_data<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderProcessingInstructionDataArgs<'largs>

C++ method: QXmlStreamReader::processingInstructionData

This is an overloaded function. Available variants:

Rust arguments:
1) fn processing_instruction_data(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn processing_instruction_data(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::processingInstructionData() const

C++ documentation:

Returns the data of a ProcessingInstruction.

fn processing_instruction_target<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderProcessingInstructionTargetArgs<'largs>

C++ method: QXmlStreamReader::processingInstructionTarget

This is an overloaded function. Available variants:

Rust arguments:
1) fn processing_instruction_target(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn processing_instruction_target(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::processingInstructionTarget() const

C++ documentation:

Returns the target of a ProcessingInstruction.

fn qualified_name<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderQualifiedNameArgs<'largs>

C++ method: QXmlStreamReader::qualifiedName

This is an overloaded function. Available variants:

Rust arguments:
1) fn qualified_name(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn qualified_name(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::qualifiedName() const

C++ documentation:

Returns the qualified name of a StartElement or EndElement;

A qualified name is the raw name of an element in the XML data. It consists of the namespace prefix, followed by colon, followed by the element's local name. Since the namespace prefix is not unique (the same prefix can point to different namespaces and different prefixes can point to the same namespace), you shouldn't use qualifiedName(), but the resolved namespaceUri() and the attribute's local name().

See also name(), prefix(), and namespaceUri().

fn raise_error<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamReaderRaiseErrorArgs<'largs>

C++ method: QXmlStreamReader::raiseError

This is an overloaded function. Available variants:

Rust arguments:
1) fn raise_error(&mut self, &::string::String) -> ()
2) fn raise_error(&mut self, ()) -> ()
C++ method: void QXmlStreamReader::raiseError(const QString& message = ?)

C++ documentation:

Raises a custom error with an optional error message.

See also error() and errorString().

fn read_element_text<'largs, Args>(&'largs mut self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderReadElementTextArgs<'largs>

C++ method: QXmlStreamReader::readElementText

This is an overloaded function. Available variants:

Rust arguments:
1) fn read_element_text(&mut self, (::xml_stream_reader::ReadElementTextBehaviour, cpp_utils::AsBox)) -> cpp_utils::CppBox<::string::String>
2) fn read_element_text(&mut self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
3) fn read_element_text(&mut self, (::xml_stream_reader::ReadElementTextBehaviour, cpp_utils::AsStruct)) -> ::string::String
4) fn read_element_text(&mut self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QXmlStreamReader::readElementText(QXmlStreamReader::ReadElementTextBehaviour behaviour = ?)

C++ documentation:

Convenience function to be called in case a StartElement was read. Reads until the corresponding EndElement and returns all text in-between. In case of no error, the current token (see tokenType()) after having called this function is EndElement.

The function concatenates text() when it reads either Characters or EntityReference tokens, but skips ProcessingInstruction and Comment. If the current token is not StartElement, an empty string is returned.

The behaviour defines what happens in case anything else is read before reaching EndElement. The function can include the text from child elements (useful for example for HTML), ignore child elements, or raise an UnexpectedElementError and return what was read so far (default).

This function was introduced in Qt 4.6.

fn read_next(&mut self) -> TokenType

C++ method: QXmlStreamReader::TokenType QXmlStreamReader::readNext()

C++ documentation:

Reads the next token and returns its type.

With one exception, once an error() is reported by readNext(), further reading of the XML stream is not possible. Then atEnd() returns true, hasError() returns true, and this function returns QXmlStreamReader::Invalid.

The exception is when error() returns PrematureEndOfDocumentError. This error is reported when the end of an otherwise well-formed chunk of XML is reached, but the chunk doesn't represent a complete XML document. In that case, parsing can be resumed by calling addData() to add the next chunk of XML, when the stream is being read from a QByteArray, or by waiting for more data to arrive when the stream is being read from a device().

See also tokenType() and tokenString().

fn read_next_start_element(&mut self) -> bool

C++ method: bool QXmlStreamReader::readNextStartElement()

C++ documentation:

Reads until the next start element within the current element. Returns true when a start element was reached. When the end element was reached, or when an error occurred, false is returned.

The current element is the element matching the most recently parsed start element of which a matching end element has not yet been reached. When the parser has reached the end element, the current element becomes the parent element.

This is a convenience function for when you're only concerned with parsing XML elements. The QXmlStream Bookmarks Example makes extensive use of this function.

This function was introduced in Qt 4.6.

See also readNext().

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QXmlStreamReader::setDevice(QIODevice* device)

C++ documentation:

Sets the current device to device. Setting the device resets the stream to its initial state.

See also device() and clear().

fn set_entity_resolver(&mut self, resolver: *mut XmlStreamEntityResolver)

C++ method: void QXmlStreamReader::setEntityResolver(QXmlStreamEntityResolver* resolver)

C++ documentation:

Makes resolver the new entityResolver().

The stream reader does not take ownership of the resolver. It's the callers responsibility to ensure that the resolver is valid during the entire life-time of the stream reader object, or until another resolver or 0 is set.

This function was introduced in Qt 4.4.

See also entityResolver().

fn set_namespace_processing(&mut self, arg1: bool)

C++ method: void QXmlStreamReader::setNamespaceProcessing(bool arg1)

C++ documentation:

The namespace-processing flag of the stream reader

This property controls whether or not the stream reader processes namespaces. If enabled, the reader processes namespaces, otherwise it does not.

By default, namespace-processing is enabled.

Access functions:

bool 	namespaceProcessing() const

void 	setNamespaceProcessing(bool)

fn skip_current_element(&mut self)

C++ method: void QXmlStreamReader::skipCurrentElement()

C++ documentation:

Reads until the end of the current element, skipping any child nodes. This function is useful for skipping unknown elements.

The current element is the element matching the most recently parsed start element of which a matching end element has not yet been reached. When the parser has reached the end element, the current element becomes the parent element.

This function was introduced in Qt 4.6.

fn text<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderTextArgs<'largs>

C++ method: QXmlStreamReader::text

This is an overloaded function. Available variants:

Rust arguments:
1) fn text(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string_ref::StringRef>
2) fn text(&self, cpp_utils::AsStruct) -> ::string_ref::StringRef
C++ method: QStringRef QXmlStreamReader::text() const

C++ documentation:

Returns the text of Characters, Comment, DTD, or EntityReference.

fn token_string<'largs, Args>(&'largs self, args: Args) -> Args::ReturnType where Args: XmlStreamReaderTokenStringArgs<'largs>

C++ method: QXmlStreamReader::tokenString

This is an overloaded function. Available variants:

Rust arguments:
1) fn token_string(&self, cpp_utils::AsBox) -> cpp_utils::CppBox<::string::String>
2) fn token_string(&self, cpp_utils::AsStruct) -> ::string::String
C++ method: QString QXmlStreamReader::tokenString() const

C++ documentation:

Returns the reader's current token as string.

See also tokenType().

fn token_type(&self) -> TokenType

C++ method: QXmlStreamReader::TokenType QXmlStreamReader::tokenType() const

C++ documentation:

Returns the type of the current token.

The current token can also be queried with the convenience functions isStartDocument(), isEndDocument(), isStartElement(), isEndElement(), isCharacters(), isComment(), isDTD(), isEntityReference(), and isProcessingInstruction().

See also tokenString().