Struct qt_core::xml_stream_writer::XmlStreamWriter

pub struct XmlStreamWriter {
    // some fields omitted
}

C++ type: QXmlStreamWriter

C++ documentation:

The QXmlStreamWriter class provides an XML writer with a simple streaming API.

QXmlStreamWriter is the counterpart to QXmlStreamReader for writing XML. Like its related class, it operates on a QIODevice specified with setDevice(). The API is simple and straightforward: for every XML token or event you want to write, the writer provides a specialized function.

You start a document with writeStartDocument() and end it with writeEndDocument(). This will implicitly close all remaining open tags.

Element tags are opened with writeStartElement() followed by writeAttribute() or writeAttributes(), element content, and then writeEndElement(). A shorter form writeEmptyElement() can be used to write empty elements, followed by writeAttributes().

Element content consists of either characters, entity references or nested elements. It is written with writeCharacters(), which also takes care of escaping all forbidden characters and character sequences, writeEntityReference(), or subsequent calls to writeStartElement(). A convenience method writeTextElement() can be used for writing terminal elements that contain nothing but text.

The following abridged code snippet shows the basic use of the class to write formatted XML with indentation:

    QXmlStreamWriter stream(&output);
    stream.setAutoFormatting(true);
    stream.writeStartDocument();
    ...
    stream.writeStartElement("bookmark");
    stream.writeAttribute("href", "http://qt-project.org/");
    stream.writeTextElement("title", "Qt Project");
    stream.writeEndElement(); // bookmark
    ...
    stream.writeEndDocument();

QXmlStreamWriter takes care of prefixing namespaces, all you have to do is specify the namespaceUri when writing elements or attributes. If you must conform to certain prefixes, you can force the writer to use them by declaring the namespaces manually with either writeNamespace() or writeDefaultNamespace(). Alternatively, you can bypass the stream writer's namespace support and use overloaded methods that take a qualified name instead. The namespace http://www.w3.org/XML/1998/namespace is implicit and mapped to the prefix xml.

The stream writer can automatically format the generated XML data by adding line-breaks and indentation to empty sections between elements, making the XML data more readable for humans and easier to work with for most source code management systems. The feature can be turned on with the autoFormatting property, and customized with the autoFormattingIndent property.

Other functions are writeCDATA(), writeComment(), writeProcessingInstruction(), and writeDTD(). Chaining of XML streams is supported with writeCurrentToken().

By default, QXmlStreamWriter encodes XML in UTF-8. Different encodings can be enforced using setCodec().

If an error occurs while writing to the underlying device, hasError() starts returning true and subsequent writes are ignored.

The QXmlStream Bookmarks Example illustrates how to use a stream writer to write an XML bookmark file (XBEL) that was previously read in by a QXmlStreamReader.

Methods

impl XmlStreamWriter

fn auto_formatting(&self) -> bool

C++ method: bool QXmlStreamWriter::autoFormatting() const

C++ documentation:

The auto-formatting flag of the stream writer

This property controls whether or not the stream writer automatically formats the generated XML data. If enabled, the writer automatically adds line-breaks and indentation to empty sections between elements (ignorable whitespace). The main purpose of auto-formatting is to split the data into several lines, and to increase readability for a human reader. The indentation depth can be controlled through the autoFormattingIndent property.

By default, auto-formatting is disabled.

This property was introduced in Qt 4.4.

Access functions:

bool 	autoFormatting() const

void 	setAutoFormatting(bool enable)

fn auto_formatting_indent(&self) -> c_int

C++ method: int QXmlStreamWriter::autoFormattingIndent() const

C++ documentation:

This property holds the number of spaces or tabs used for indentation when auto-formatting is enabled. Positive numbers indicate spaces, negative numbers tabs.

The default indentation is 4.

This property was introduced in Qt 4.4.

Access functions:

int 	autoFormattingIndent() const

void 	setAutoFormattingIndent(int spacesOrTabs)

See also autoFormatting.

fn codec(&self) -> *mut TextCodec

C++ method: QTextCodec* QXmlStreamWriter::codec() const

C++ documentation:

Returns the codec that is currently assigned to the stream.

See also setCodec().

fn device(&self) -> *mut QiodEvice

C++ method: QIODevice* QXmlStreamWriter::device() const

C++ documentation:

Returns the current device associated with the QXmlStreamWriter, or 0 if no device has been assigned.

See also setDevice().

fn has_error(&self) -> bool

C++ method: bool QXmlStreamWriter::hasError() const

C++ documentation:

Returns true if the stream failed to write to the underlying device.

The error status is never reset. Writes happening after the error occurred are ignored, even if the error condition is cleared.

fn new<Args>(args: Args) -> Args::ReturnType where Args: XmlStreamWriterNewArgs

C++ method: QXmlStreamWriter::QXmlStreamWriter

This is an overloaded function. Available variants:

Variant 1

Rust arguments:
1) fn new(cpp_utils::AsStruct) -> ::xml_stream_writer::XmlStreamWriter
2) fn new(cpp_utils::AsBox) -> cpp_utils::CppBox<::xml_stream_writer::XmlStreamWriter>
C++ method: [constructor] void QXmlStreamWriter::QXmlStreamWriter()

Variant 2

Rust arguments:
1) fn new((*mut ::byte_array::ByteArray, cpp_utils::AsStruct)) -> ::xml_stream_writer::XmlStreamWriter
2) fn new((*mut ::byte_array::ByteArray, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_writer::XmlStreamWriter>
C++ method: [constructor] void QXmlStreamWriter::QXmlStreamWriter(QByteArray* array)

Variant 3

Rust arguments:
1) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsStruct)) -> ::xml_stream_writer::XmlStreamWriter
2) fn new((*mut ::qiod_evice::QiodEvice, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_writer::XmlStreamWriter>
C++ method: [constructor] void QXmlStreamWriter::QXmlStreamWriter(QIODevice* device)

Variant 4

Rust arguments:
1) fn new((*mut ::string::String, cpp_utils::AsStruct)) -> ::xml_stream_writer::XmlStreamWriter
2) fn new((*mut ::string::String, cpp_utils::AsBox)) -> cpp_utils::CppBox<::xml_stream_writer::XmlStreamWriter>
C++ method: [constructor] void QXmlStreamWriter::QXmlStreamWriter(QString* string)

fn set_auto_formatting(&mut self, arg1: bool)

C++ method: void QXmlStreamWriter::setAutoFormatting(bool arg1)

C++ documentation:

The auto-formatting flag of the stream writer

This property controls whether or not the stream writer automatically formats the generated XML data. If enabled, the writer automatically adds line-breaks and indentation to empty sections between elements (ignorable whitespace). The main purpose of auto-formatting is to split the data into several lines, and to increase readability for a human reader. The indentation depth can be controlled through the autoFormattingIndent property.

By default, auto-formatting is disabled.

This property was introduced in Qt 4.4.

Access functions:

bool 	autoFormatting() const

void 	setAutoFormatting(bool enable)

fn set_auto_formatting_indent(&mut self, spaces_or_tabs: c_int)

C++ method: void QXmlStreamWriter::setAutoFormattingIndent(int spacesOrTabs)

C++ documentation:

This property holds the number of spaces or tabs used for indentation when auto-formatting is enabled. Positive numbers indicate spaces, negative numbers tabs.

The default indentation is 4.

This property was introduced in Qt 4.4.

Access functions:

int 	autoFormattingIndent() const

void 	setAutoFormattingIndent(int spacesOrTabs)

See also autoFormatting.

fn set_codec<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterSetCodecArgs<'largs>

C++ method: QXmlStreamWriter::setCodec

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn set_codec(&mut self, *mut ::text_codec::TextCodec) -> ()
C++ method: void QXmlStreamWriter::setCodec(QTextCodec* codec)

C++ documentation:

Sets the codec for this stream to codec. The codec is used for encoding any data that is written. By default, QXmlStreamWriter uses UTF-8.

The encoding information is stored in the initial xml tag which gets written when you call writeStartDocument(). Call this function before calling writeStartDocument().

See also codec().

Variant 2

Rust arguments: fn set_codec(&mut self, *const libc::c_char) -> ()
C++ method: void QXmlStreamWriter::setCodec(const char* codecName)

C++ documentation:

Sets the codec for this stream to the QTextCodec for the encoding specified by codecName. Common values for codecName include "ISO 8859-1", "UTF-8", and "UTF-16". If the encoding isn't recognized, nothing happens.

See also QTextCodec::codecForName().

fn set_device(&mut self, device: *mut QiodEvice)

C++ method: void QXmlStreamWriter::setDevice(QIODevice* device)

C++ documentation:

Sets the current device to device. If you want the stream to write into a QByteArray, you can create a QBuffer device.

See also device().

fn write_attribute<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterWriteAttributeArgs<'largs>

C++ method: QXmlStreamWriter::writeAttribute

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn write_attribute(&mut self, (&::string::String, &::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeAttribute(const QString& namespaceUri, const QString& name, const QString& value)

C++ documentation:

Writes an attribute with name and value, prefixed for the specified namespaceUri. If the namespace has not been declared yet, QXmlStreamWriter will generate a namespace declaration for it.

This function can only be called after writeStartElement() before any content is written, or after writeEmptyElement().

Variant 2

Rust arguments: fn write_attribute(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeAttribute(const QString& qualifiedName, const QString& value)

C++ documentation:

This is an overloaded function.

Writes an attribute with qualifiedName and value.

This function can only be called after writeStartElement() before any content is written, or after writeEmptyElement().

Variant 3

Rust arguments: fn write_attribute(&mut self, &::xml_stream_attribute::XmlStreamAttribute) -> ()
C++ method: void QXmlStreamWriter::writeAttribute(const QXmlStreamAttribute& attribute)

C++ documentation:

This is an overloaded function.

Writes the attribute.

This function can only be called after writeStartElement() before any content is written, or after writeEmptyElement().

fn write_attributes(&mut self, attributes: &XmlStreamAttributes)

C++ method: void QXmlStreamWriter::writeAttributes(const QXmlStreamAttributes& attributes)

C++ documentation:

Writes the attribute vector attributes. If a namespace referenced in an attribute not been declared yet, QXmlStreamWriter will generate a namespace declaration for it.

This function can only be called after writeStartElement() before any content is written, or after writeEmptyElement().

See also writeAttribute() and writeNamespace().

fn write_cdata(&mut self, text: &String)

C++ method: void QXmlStreamWriter::writeCDATA(const QString& text)

C++ documentation:

Writes text as CDATA section. If text contains the forbidden character sequence "]]>", it is split into different CDATA sections.

This function mainly exists for completeness. Normally you should not need use it, because writeCharacters() automatically escapes all non-content characters.

fn write_characters(&mut self, text: &String)

C++ method: void QXmlStreamWriter::writeCharacters(const QString& text)

C++ documentation:

Writes text. The characters "<", "&", and """ are escaped as entity references "&lt;", "&amp;, and "&quot;". To avoid the forbidden sequence "]]>", ">" is also escaped as "&gt;".

See also writeEntityReference().

fn write_comment(&mut self, text: &String)

C++ method: void QXmlStreamWriter::writeComment(const QString& text)

C++ documentation:

Writes text as XML comment, where text must not contain the forbidden sequence "--" or end with "-". Note that XML does not provide any way to escape "-" in a comment.

fn write_current_token(&mut self, reader: &XmlStreamReader)

C++ method: void QXmlStreamWriter::writeCurrentToken(const QXmlStreamReader& reader)

C++ documentation:

Writes the current state of the reader. All possible valid states are supported.

The purpose of this function is to support chained processing of XML data.

See also QXmlStreamReader::tokenType().

fn write_default_namespace(&mut self, namespace_uri: &String)

C++ method: void QXmlStreamWriter::writeDefaultNamespace(const QString& namespaceUri)

C++ documentation:

Writes a default namespace declaration for namespaceUri.

If writeStartElement() or writeEmptyElement() was called, the declaration applies to the current element; otherwise it applies to the next child element.

Note that the namespaces http://www.w3.org/XML/1998/namespace (bound to xmlns) and http://www.w3.org/2000/xmlns/ (bound to xml) by definition cannot be declared as default.

fn write_dtd(&mut self, dtd: &String)

C++ method: void QXmlStreamWriter::writeDTD(const QString& dtd)

C++ documentation:

Writes a DTD section. The dtd represents the entire doctypedecl production from the XML 1.0 specification.

fn write_empty_element<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterWriteEmptyElementArgs<'largs>

C++ method: QXmlStreamWriter::writeEmptyElement

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn write_empty_element(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeEmptyElement(const QString& namespaceUri, const QString& name)

C++ documentation:

Writes an empty element with name, prefixed for the specified namespaceUri. If the namespace has not been declared, QXmlStreamWriter will generate a namespace declaration for it. Subsequent calls to writeAttribute() will add attributes to this element.

See also writeNamespace().

Variant 2

Rust arguments: fn write_empty_element(&mut self, &::string::String) -> ()
C++ method: void QXmlStreamWriter::writeEmptyElement(const QString& qualifiedName)

C++ documentation:

This is an overloaded function.

Writes an empty element with qualified name qualifiedName. Subsequent calls to writeAttribute() will add attributes to this element.

fn write_end_document(&mut self)

C++ method: void QXmlStreamWriter::writeEndDocument()

C++ documentation:

Closes all remaining open start elements and writes a newline.

See also writeStartDocument().

fn write_end_element(&mut self)

C++ method: void QXmlStreamWriter::writeEndElement()

C++ documentation:

Closes the previous start element.

See also writeStartElement().

fn write_entity_reference(&mut self, name: &String)

C++ method: void QXmlStreamWriter::writeEntityReference(const QString& name)

C++ documentation:

Writes the entity reference name to the stream, as "&name;".

fn write_namespace<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterWriteNamespaceArgs<'largs>

C++ method: QXmlStreamWriter::writeNamespace

This is an overloaded function. Available variants:

Rust arguments:
1) fn write_namespace(&mut self, &::string::String) -> ()
2) fn write_namespace(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeNamespace(const QString& namespaceUri, const QString& prefix = ?)

C++ documentation:

Writes a namespace declaration for namespaceUri with prefix. If prefix is empty, QXmlStreamWriter assigns a unique prefix consisting of the letter 'n' followed by a number.

If writeStartElement() or writeEmptyElement() was called, the declaration applies to the current element; otherwise it applies to the next child element.

Note that the prefix xml is both predefined and reserved for http://www.w3.org/XML/1998/namespace, which in turn cannot be bound to any other prefix. The prefix xmlns and its URI http://www.w3.org/2000/xmlns/ are used for the namespace mechanism itself and thus completely forbidden in declarations.

fn write_processing_instruction<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterWriteProcessingInstructionArgs<'largs>

C++ method: QXmlStreamWriter::writeProcessingInstruction

This is an overloaded function. Available variants:

Rust arguments:
1) fn write_processing_instruction(&mut self, &::string::String) -> ()
2) fn write_processing_instruction(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeProcessingInstruction(const QString& target, const QString& data = ?)

C++ documentation:

Writes an XML processing instruction with target and data, where data must not contain the sequence "?>".

fn write_start_document<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterWriteStartDocumentArgs<'largs>

C++ method: QXmlStreamWriter::writeStartDocument

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn write_start_document(&mut self, ()) -> ()
C++ method: void QXmlStreamWriter::writeStartDocument()

C++ documentation:

This is an overloaded function.

Writes a document start with XML version number "1.0". This also writes the encoding information.

This function was introduced in Qt 4.5.

See also writeEndDocument() and setCodec().

Variant 2

Rust arguments: fn write_start_document(&mut self, &::string::String) -> ()
C++ method: void QXmlStreamWriter::writeStartDocument(const QString& version)

C++ documentation:

Writes a document start with the XML version number version.

See also writeEndDocument().

Variant 3

Rust arguments: fn write_start_document(&mut self, (&::string::String, bool)) -> ()
C++ method: void QXmlStreamWriter::writeStartDocument(const QString& version, bool standalone)

C++ documentation:

Writes a document start with the XML version number version and a standalone attribute standalone.

This function was introduced in Qt 4.5.

See also writeEndDocument().

fn write_start_element<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterWriteStartElementArgs<'largs>

C++ method: QXmlStreamWriter::writeStartElement

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn write_start_element(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeStartElement(const QString& namespaceUri, const QString& name)

C++ documentation:

Writes a start element with name, prefixed for the specified namespaceUri. If the namespace has not been declared yet, QXmlStreamWriter will generate a namespace declaration for it. Subsequent calls to writeAttribute() will add attributes to this element.

See also writeNamespace(), writeEndElement(), and writeEmptyElement().

Variant 2

Rust arguments: fn write_start_element(&mut self, &::string::String) -> ()
C++ method: void QXmlStreamWriter::writeStartElement(const QString& qualifiedName)

C++ documentation:

This is an overloaded function.

Writes a start element with qualifiedName. Subsequent calls to writeAttribute() will add attributes to this element.

See also writeEndElement() and writeEmptyElement().

fn write_text_element<'largs, Args>(&'largs mut self, args: Args) where Args: XmlStreamWriterWriteTextElementArgs<'largs>

C++ method: QXmlStreamWriter::writeTextElement

This is an overloaded function. Available variants:

Variant 1

Rust arguments: fn write_text_element(&mut self, (&::string::String, &::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeTextElement(const QString& namespaceUri, const QString& name, const QString& text)

C++ documentation:

Writes a text element with name, prefixed for the specified namespaceUri, and text. If the namespace has not been declared, QXmlStreamWriter will generate a namespace declaration for it.

This is a convenience function equivalent to:

        writeStartElement(namespaceUri, name);
        writeCharacters(text);
        writeEndElement();

Variant 2

Rust arguments: fn write_text_element(&mut self, (&::string::String, &::string::String)) -> ()
C++ method: void QXmlStreamWriter::writeTextElement(const QString& qualifiedName, const QString& text)

C++ documentation:

This is an overloaded function.

Writes a text element with qualifiedName and text.

This is a convenience function equivalent to:

        writeStartElement(qualifiedName);
        writeCharacters(text);
        writeEndElement();